Generate data points along a conic spiral curve with optional noise.

This function generates data points along a conic spiral curve with optional noise.

Usage

conic_spiral_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

conic_spiral_3d(n = 100, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>                [,1]         [,2]         [,3]          [,4]          [,5]
#>   [1,]  0.100317268  0.242442293  0.594455021  0.0079563795  0.0449367228
#>   [2,] -0.111312460 -0.082918088  0.794614502 -0.0080621674 -0.0278348752
#>   [3,] -0.119355013 -0.058095218  0.271566426 -0.0177530236  0.0262463843
#>   [4,]  0.464251538  0.149837823  0.029523984  0.0442266062  0.0465917273
#>   [5,] -0.072426930 -0.168328260  0.323200014 -0.0240833007  0.0008668442
#>   [6,] -0.040015308 -0.148726406  0.851644918  0.0294055375  0.0083156769
#>   [7,]  0.094446329 -0.051682790  0.958887776 -0.0112767190 -0.0261415791
#>   [8,] -0.104841272 -0.017265445  0.759243171  0.0444708498 -0.0127400472
#>   [9,]  0.097305760 -0.034505923  0.972091343  0.0474726310  0.0026967967
#>  [10,] -0.236270452 -0.231791849  0.295635511  0.0469105998  0.0447065806
#>  [11,] -0.019130677 -0.108423131  0.857692010  0.0027951926  0.0469071049
#>  [12,]  0.125032465  0.395240024  0.081640026 -0.0141814547  0.0476171049
#>  [13,] -0.049877072 -0.113910188  0.837265941  0.0476895383  0.0270145549
#>  [14,] -0.018957336 -0.145278321  0.349034609 -0.0395413799  0.0230932452
#>  [15,]  0.228792578 -0.031306631  0.473575910 -0.0048207024  0.0332899704
#>  [16,]  0.084497546 -0.053566523  0.955183067 -0.0435353255 -0.0494913912
#>  [17,]  0.240732807  0.112913159  0.065315875 -0.0168435012  0.0180584875
#>  [18,]  0.081612836 -0.068638447  0.945975309 -0.0016846273  0.0057810330
#>  [19,] -0.096668701 -0.338510305  0.359458678 -0.0055237127  0.0268198846
#>  [20,]  0.402031225  0.130108373  0.001426694  0.0344739319  0.0090309300
#>  [21,] -0.341496406 -0.067295672  0.282436889 -0.0337072589  0.0408864923
#>  [22,]  0.019851080  0.186769389  0.604498153 -0.0151076644 -0.0475052497
#>  [23,]  0.148537012  0.062937235  0.517907756  0.0112710816  0.0408270949
#>  [24,] -0.047058302  0.101123489  0.149920444  0.0469012581 -0.0322113955
#>  [25,]  0.330401132  0.017538840 -0.027071995 -0.0454881808  0.0473972037
#>  [26,]  0.004070408 -0.111967894  0.881242057  0.0202110410 -0.0405113815
#>  [27,] -0.126153550 -0.114952228  0.812029620 -0.0198924854  0.0175357757
#>  [28,] -0.070679300  0.072618739  0.684373542 -0.0282906733 -0.0358552554
#>  [29,]  0.021834629 -0.311100532  0.366695015  0.0181393915 -0.0168537469
#>  [30,]  0.051660155 -0.341273261  0.386361552 -0.0488985675 -0.0313194686
#>  [31,] -0.078098846  0.222488353  0.655312564  0.0150532481  0.0449730973
#>  [32,]  0.065418125 -0.091021618  0.416417966 -0.0406153001  0.0115737197
#>  [33,] -0.288205527  0.031548210  0.217951721  0.0013131405 -0.0047517328
#>  [34,] -0.134457607  0.136708784  0.163543441 -0.0406045414  0.0243731138
#>  [35,] -0.071263471  0.083845074  0.190044910  0.0444074168  0.0001357735
#>  [36,]  0.119665341 -0.313702637  0.407976757  0.0273640122  0.0313657785
#>  [37,]  0.138097176  0.036877019  0.508046034  0.0448621313 -0.0113197442
#>  [38,]  0.015991665  0.098759714  0.612650647 -0.0154920605 -0.0195295999
#>  [39,] -0.045942082 -0.090011537  0.838754007  0.0232113268  0.0342531075
#>  [40,]  0.078941089  0.077884070  0.555229666  0.0309820558 -0.0367499343
#>  [41,] -0.025728218 -0.132402062  0.855396007  0.0165879672 -0.0435704472
#>  [42,] -0.116245972  0.094984997  0.705022671  0.0408845174 -0.0443234814
#>  [43,]  0.289515717  0.050629944  0.512089514 -0.0483968466  0.0360059981
#>  [44,]  0.097491093 -0.022831041  0.981540794 -0.0220638282  0.0138027464
#>  [45,] -0.011607177 -0.130326709  0.354762179 -0.0416369024  0.0168353574
#>  [46,] -0.030415383 -0.133321952  0.861507851 -0.0152918071 -0.0004815067
#>  [47,] -0.055765912 -0.091005164  0.828032243  0.0336906504 -0.0417606037
#>  [48,]  0.009658850 -0.178018974  0.397673087 -0.0039921657  0.0302830352
#>  [49,] -0.124400850 -0.053019373  0.297054627 -0.0405394587 -0.0240311141
#>  [50,] -0.122308465 -0.066476311  0.782936528  0.0014166133  0.0036603304
#>  [51,] -0.069594039  0.085370595  0.188466371 -0.0145192811 -0.0124674802
#>  [52,]  0.238117306 -0.021203164  0.507919844  0.0263793958 -0.0481497042
#>  [53,]  0.096826377  0.203670071  0.117228120  0.0441354186 -0.0495487981
#>  [54,] -0.157015984  0.186143162  0.206637407  0.0433810569 -0.0356491636
#>  [55,]  0.183027079  0.094250377  0.523287492  0.0295955809 -0.0156919445
#>  [56,]  0.286006475 -0.001332064  0.491467458 -0.0255480983  0.0277329674
#>  [57,] -0.260509159 -0.103803191  0.258008563 -0.0351625924  0.0475231444
#>  [58,]  0.097007556 -0.060281432  0.954633517 -0.0452888449 -0.0434655228
#>  [59,] -0.221778157 -0.286421565  0.314954888  0.0212880229 -0.0318070989
#>  [60,] -0.093328032 -0.129579103  0.821406328  0.0217189781 -0.0073697052
#>  [61,] -0.060628232  0.133011876  0.669305488 -0.0176879408  0.0091343700
#>  [62,] -0.134115805 -0.044532543  0.782634729  0.0360345612 -0.0259445542
#>  [63,]  0.053324136 -0.142916750  0.419155709 -0.0177663640  0.0360602752
#>  [64,]  0.093865306 -0.050405149  0.959552946 -0.0482342542 -0.0011025476
#>  [65,] -0.057168796  0.082047872  0.173499665 -0.0221804218 -0.0346860921
#>  [66,] -0.139325068  0.413966779  0.155793725  0.0183542342 -0.0201842599
#>  [67,]  0.175183098 -0.267701571  0.413063428 -0.0413902960  0.0475916094
#>  [68,]  0.099089260  0.148980491  0.102361923 -0.0341554393  0.0339980856
#>  [69,]  0.066909774 -0.090105167  0.923504333 -0.0093898495  0.0192077778
#>  [70,] -0.153073347 -0.072466839  0.790580209  0.0001329356 -0.0051438112
#>  [71,]  0.175536277 -0.125668721  0.433093138  0.0403530250  0.0324792245
#>  [72,] -0.077907800  0.106240361  0.190834379 -0.0251450060 -0.0035597055
#>  [73,] -0.055020668 -0.142494920  0.362386790 -0.0458892580 -0.0166545776
#>  [74,]  0.252341796  0.176652492  0.018450977 -0.0274770753 -0.0443946753
#>  [75,] -0.173173051  0.117533035  0.697369740 -0.0428423730  0.0406402306
#>  [76,]  0.053520830  0.089265337  0.586126010  0.0391007241  0.0159685715
#>  [77,] -0.185204122  0.002424082  0.743231004  0.0410606227  0.0284260285
#>  [78,]  0.085737490 -0.074855641  0.944661665 -0.0157256222  0.0442108643
#>  [79,] -0.175447685 -0.055985249  0.295536929  0.0003133077  0.0152067449
#>  [80,]  0.181020694  0.425846802  0.092214367 -0.0170834569  0.0452647741
#>  [81,] -0.035008131 -0.347960064  0.362332190  0.0260949676  0.0210558357
#>  [82,] -0.127712905  0.067027069  0.194109256  0.0394221349 -0.0025384415
#>  [83,]  0.100767460 -0.019024124  0.984683302 -0.0422183104  0.0190003939
#>  [84,] -0.141940005  0.101309061  0.691427130 -0.0497511781 -0.0441884734
#>  [85,]  0.043439300 -0.113703010  0.907068117 -0.0482168708 -0.0443951389
#>  [86,] -0.090517381  0.370539187  0.164908970 -0.0415272179  0.0395159463
#>  [87,] -0.007081518  0.438577944  0.116637163 -0.0022500239 -0.0125995015
#>  [88,] -0.067765589  0.086719216  0.186797211  0.0081601578  0.0131761827
#>  [89,] -0.166414405  0.310785613  0.186998473 -0.0289747416 -0.0439675216
#>  [90,] -0.102091391 -0.019134798  0.270077338  0.0380159040 -0.0010329881
#>  [91,]  0.062701734 -0.098422667  0.917599210  0.0076414786  0.0331853610
#>  [92,]  0.102097245 -0.017873300  0.985793381 -0.0329458751 -0.0297326259
#>  [93,]  0.025844932  0.115541653  0.615524856  0.0047987403 -0.0361330229
#>  [94,]  0.171973862  0.169141509  0.550802984  0.0019111664 -0.0433877075
#>  [95,]  0.095286488  0.050138906  0.544409451  0.0095294249 -0.0456942514
#>  [96,]  0.003276466 -0.181179479  0.395040742  0.0316277596  0.0301101064
#>  [97,] -0.152265440  0.075539026  0.234177966  0.0100920168 -0.0138789925
#>  [98,] -0.081005809  0.436299947  0.137741230 -0.0104793709  0.0413999083
#>  [99,] -0.015788683 -0.099039761  0.863054333  0.0082689498  0.0336278187
#> [100,]  0.095949385  0.051560868  0.030027123 -0.0498556386 -0.0183629357
#>                [,6]          [,7]
#>   [1,]  0.037981665  0.0073714942
#>   [2,] -0.032132551 -0.0235729843
#>   [3,]  0.015411746  0.0241295215
#>   [4,]  0.022073596 -0.0333915436
#>   [5,] -0.001155149  0.0049097516
#>   [6,]  0.048256301 -0.0348574166
#>   [7,]  0.042928767 -0.0211250156
#>   [8,]  0.018086913 -0.0379800418
#>   [9,] -0.033277073 -0.0164364322
#>  [10,]  0.017236554 -0.0343726723
#>  [11,]  0.028272863  0.0422407936
#>  [12,] -0.025719619 -0.0262933123
#>  [13,] -0.039568099 -0.0291287732
#>  [14,]  0.030060228  0.0199809818
#>  [15,]  0.047963398 -0.0251283602
#>  [16,]  0.026767539 -0.0261952245
#>  [17,] -0.046555025  0.0469465774
#>  [18,] -0.017717235 -0.0156533482
#>  [19,]  0.041424722 -0.0320100896
#>  [20,] -0.032523568 -0.0194646583
#>  [21,] -0.014504452 -0.0004838257
#>  [22,] -0.017693657 -0.0005229040
#>  [23,] -0.027412361 -0.0368053711
#>  [24,] -0.014125699  0.0247531708
#>  [25,]  0.030069399  0.0106406637
#>  [26,] -0.013336544 -0.0132545890
#>  [27,]  0.042985390  0.0434689051
#>  [28,]  0.009149651 -0.0434673291
#>  [29,]  0.005064318 -0.0223461513
#>  [30,]  0.037205702 -0.0230567977
#>  [31,]  0.004501776 -0.0446839707
#>  [32,] -0.024617790 -0.0312252070
#>  [33,] -0.046456399 -0.0415263098
#>  [34,]  0.031941963 -0.0262051034
#>  [35,]  0.032024879  0.0383513908
#>  [36,] -0.043774793  0.0257319342
#>  [37,] -0.012780146  0.0046353613
#>  [38,]  0.009862374 -0.0266905197
#>  [39,] -0.045813180  0.0078141792
#>  [40,] -0.027905769 -0.0438466760
#>  [41,] -0.023192027 -0.0437529095
#>  [42,]  0.044519212 -0.0313916142
#>  [43,]  0.009619498  0.0010618086
#>  [44,] -0.012024411 -0.0157101776
#>  [45,] -0.011949455  0.0141670325
#>  [46,] -0.040830292 -0.0188702296
#>  [47,]  0.006202705 -0.0250899901
#>  [48,]  0.020384668 -0.0294831059
#>  [49,] -0.024742598 -0.0150545330
#>  [50,]  0.031899737  0.0159904186
#>  [51,]  0.048394445 -0.0083468215
#>  [52,]  0.008649500  0.0043365644
#>  [53,]  0.028397370 -0.0236846178
#>  [54,] -0.032524540 -0.0273124617
#>  [55,] -0.030408853  0.0031710098
#>  [56,] -0.043664925  0.0181591777
#>  [57,]  0.013860082  0.0094306486
#>  [58,] -0.038309993 -0.0398439076
#>  [59,] -0.007382033 -0.0195515747
#>  [60,] -0.035033281 -0.0389921789
#>  [61,] -0.015021379 -0.0294075003
#>  [62,] -0.017953356  0.0465446624
#>  [63,] -0.022757822 -0.0432874428
#>  [64,]  0.038275962 -0.0477833519
#>  [65,] -0.019920682 -0.0060525641
#>  [66,] -0.023121862  0.0191674542
#>  [67,] -0.044606399  0.0406359599
#>  [68,] -0.039688599 -0.0092019238
#>  [69,] -0.043682621 -0.0281114249
#>  [70,]  0.040073105 -0.0404606683
#>  [71,] -0.011246936  0.0374543541
#>  [72,] -0.042626149  0.0239540484
#>  [73,] -0.033309770  0.0373056878
#>  [74,]  0.024815485 -0.0350661323
#>  [75,] -0.001662185 -0.0131150192
#>  [76,] -0.033646943  0.0206161962
#>  [77,]  0.007400034  0.0235541194
#>  [78,] -0.019733465 -0.0468740655
#>  [79,]  0.041166826  0.0268530738
#>  [80,] -0.033860789  0.0240528282
#>  [81,]  0.032976219  0.0464231196
#>  [82,] -0.043617474  0.0483057057
#>  [83,]  0.040018513 -0.0374614776
#>  [84,]  0.021202758 -0.0061107646
#>  [85,] -0.041128956  0.0388080163
#>  [86,] -0.048581713 -0.0373334902
#>  [87,] -0.003485879 -0.0249674416
#>  [88,]  0.037123881  0.0198674103
#>  [89,]  0.021456078 -0.0375325317
#>  [90,] -0.031398711 -0.0035511570
#>  [91,] -0.021700694 -0.0309108103
#>  [92,] -0.014044181 -0.0379896699
#>  [93,] -0.032038997 -0.0022331275
#>  [94,]  0.047170254 -0.0218949194
#>  [95,] -0.038150050 -0.0241546585
#>  [96,] -0.045091664 -0.0004549618
#>  [97,] -0.003582499  0.0105465141
#>  [98,]  0.006372761  0.0307745221
#>  [99,]  0.042276604  0.0136739163
#> [100,]  0.026322946  0.0069974528