Generate data points along a conic spiral curve with optional noise.

This function generates data points along a conic spiral curve with optional noise.

Usage

conic_spiral_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

conic_spiral_with_noise(sample_size = 100, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>                 [,1]         [,2]          [,3]         [,4]          [,5]
#>   [1,] -0.0207510344 -0.104240813  0.8565423349 -0.040931028 -0.0166046375
#>   [2,] -0.1897456848 -0.153165003  0.3261834884 -0.031813508 -0.0477462039
#>   [3,] -0.1143007122 -0.238729076  0.3190896370  0.013744115 -0.0347793166
#>   [4,] -0.1181911266  0.011985674  0.2303204236  0.007345806  0.0237506682
#>   [5,]  0.0936446329  0.046165714  0.5319536472  0.004771700  0.0283607942
#>   [6,] -0.1736094196 -0.253496995  0.3091928313  0.025875320 -0.0392013968
#>   [7,] -0.0202558489 -0.105323362  0.8628744091 -0.035268743  0.0050412531
#>   [8,] -0.0509061132 -0.140855742  0.8511714180  0.014293427  0.0146856452
#>   [9,]  0.0844611734 -0.287802336  0.3833104242 -0.040356171  0.0025612342
#>  [10,] -0.0977533422  0.079629714  0.7035529646  0.026405083 -0.0012461972
#>  [11,] -0.1368858991  0.031220103  0.7238915232 -0.021452677  0.0239702933
#>  [12,]  0.0206820358  0.101406267  0.6053109085  0.033733608 -0.0493029877
#>  [13,] -0.1949940914 -0.039564701  0.2438396855 -0.013635186  0.0036622147
#>  [14,]  0.1847525264  0.153583430  0.5436209538 -0.019939320 -0.0343417981
#>  [15,]  0.2992513885  0.102976891 -0.0044238195 -0.008601024  0.0357280403
#>  [16,]  0.0285792536 -0.100764988  0.4023204640 -0.022914698  0.0420856933
#>  [17,] -0.0922436934  0.209666098  0.6529501173  0.036285604  0.0108949394
#>  [18,] -0.0353903685  0.340045096  0.1079105155  0.015457681  0.0179547363
#>  [19,] -0.3881403830  0.024112243  0.2548485238  0.014558589  0.0122834559
#>  [20,] -0.0705605545  0.157659972  0.6693351256  0.041608703 -0.0008436582
#>  [21,] -0.0983858086 -0.039369079  0.2868194365 -0.007250921  0.0364216801
#>  [22,]  0.0758128601 -0.069242251  0.4372337578 -0.016189654  0.0360680232
#>  [23,]  0.0863155389  0.106181104  0.5594642800  0.003030169  0.0363069149
#>  [24,] -0.1179893830  0.097122442  0.7048106941 -0.005471833 -0.0362884380
#>  [25,] -0.0442495627 -0.121312338  0.8520263852  0.020322568  0.0485623369
#>  [26,]  0.0749787177 -0.104765762  0.9234217379 -0.039003743 -0.0034347890
#>  [27,] -0.0094250376 -0.105119468  0.8653073284  0.036990595 -0.0393978570
#>  [28,] -0.0391375577  0.155413611  0.6558118381 -0.049432276 -0.0467532965
#>  [29,]  0.1150048046 -0.305982271  0.3954184439  0.047076767  0.0019607651
#>  [30,] -0.0409613807 -0.209908526  0.3799208949  0.042625021  0.0145789130
#>  [31,]  0.1603522208  0.185534476  0.0393569516  0.017738804  0.0193060695
#>  [32,]  0.2320314575  0.139328772  0.5357896381 -0.013839810 -0.0043427410
#>  [33,] -0.0271065134  0.425067054  0.1165777828  0.034145119  0.0453300655
#>  [34,] -0.1762240245 -0.064180096  0.7795144070  0.026294504 -0.0041827540
#>  [35,] -0.0785955201  0.117102130  0.6817192891  0.047181112 -0.0081302271
#>  [36,] -0.1413355355  0.112686375  0.6873163998 -0.028553091  0.0122232551
#>  [37,] -0.1131517581 -0.108788186  0.2931728342 -0.020200717 -0.0168733340
#>  [38,]  0.0812929781  0.247056912  0.5993497488  0.006907545 -0.0099013192
#>  [39,]  0.0054567782  0.103040301  0.6242508078  0.001849898  0.0051537106
#>  [40,]  0.3084933922  0.187887130  0.0151933319  0.020957683  0.0399808738
#>  [41,] -0.1820643836  0.054552746  0.7202142267 -0.046946775  0.0086850211
#>  [42,]  0.4933407216  0.033592614 -0.0003778863  0.014940627  0.0036506869
#>  [43,] -0.0582192903  0.117644906  0.6524890165 -0.034076899  0.0397402511
#>  [44,] -0.2834242770 -0.115406203  0.3014162360  0.016994147 -0.0025732327
#>  [45,]  0.0043549958 -0.130942869  0.8813984699  0.003986424  0.0191004558
#>  [46,] -0.1878736166  0.036518240  0.2124275530 -0.041816719 -0.0366099922
#>  [47,]  0.1958612497  0.073825079  0.0564893513 -0.012845770 -0.0099098236
#>  [48,]  0.1838460121  0.102819694  0.5262634252 -0.035709347 -0.0186177733
#>  [49,] -0.0315737786  0.108513293  0.1579299517 -0.020558316 -0.0214005019
#>  [50,] -0.1721295318  0.280490422  0.1932942820 -0.013372063  0.0148630873
#>  [51,]  0.0788930509  0.233502261  0.5918898853 -0.023581988 -0.0269665238
#>  [52,]  0.1694746893 -0.149592876  0.4600479651  0.004148794  0.0278712874
#>  [53,] -0.1847249782 -0.275992943  0.3427454146 -0.017790378 -0.0391516257
#>  [54,]  0.0892903413  0.106431130  0.0875795412 -0.036711160  0.0028530608
#>  [55,]  0.1645865229  0.369605296  0.0703729169  0.031528962  0.0459047012
#>  [56,] -0.1504657718 -0.099165182  0.7979626454 -0.047778206 -0.0427526213
#>  [57,] -0.0344767793  0.235312666  0.6417089871 -0.020413165 -0.0177211324
#>  [58,] -0.2142710700  0.320668947  0.1900372363  0.049922821 -0.0332402042
#>  [59,] -0.3006092328 -0.108657671  0.2972390053  0.005551628  0.0373407168
#>  [60,] -0.0386599964  0.129484628  0.1645809073  0.003388784  0.0380526948
#>  [61,] -0.1282222951  0.308885058  0.1315751715  0.014668755  0.0306826322
#>  [62,] -0.0006501423 -0.102529206  0.8727471090 -0.048575067 -0.0436701143
#>  [63,]  0.0176619014 -0.259057689  0.3993007277  0.006334455  0.0490626896
#>  [64,]  0.1299246533  0.059083418  0.5214930050 -0.027062034 -0.0108800165
#>  [65,]  0.0958308666 -0.056450272  0.4652848015  0.029108361 -0.0397776565
#>  [66,] -0.1297398466  0.011551467  0.2573860312 -0.016626037  0.0134091606
#>  [67,]  0.0164506261  0.437899112  0.1325599119 -0.023652252 -0.0336862096
#>  [68,]  0.2447696416  0.041203906  0.5000619148  0.008318395  0.0317193163
#>  [69,] -0.0108711168  0.104407131  0.6374807026  0.009927869 -0.0266226169
#>  [70,] -0.1297792438 -0.212311034  0.3101906556  0.038506698  0.0008334388
#>  [71,] -0.0529094738 -0.187773727  0.3729977629 -0.029403563 -0.0343323008
#>  [72,]  0.0084603408 -0.107956268  0.8840930695  0.002679979 -0.0015254343
#>  [73,] -0.3829944128  0.052062855  0.2278876711 -0.016571605  0.0038303378
#>  [74,]  0.0559559254  0.433386783  0.1268646537  0.015887003 -0.0345719742
#>  [75,]  0.2216785679  0.009417408  0.0326095714  0.033088708  0.0091628966
#>  [76,] -0.0615413790  0.387791568  0.1572226907 -0.026307247  0.0440838093
#>  [77,]  0.0410575835 -0.091390362  0.4096653366  0.047253790  0.0450331292
#>  [78,] -0.0262710353 -0.127099269  0.8632684635  0.002374215  0.0490162657
#>  [79,] -0.0776551452 -0.071562970  0.8124199381  0.016724799 -0.0397617172
#>  [80,]  0.3888382297  0.108443188  0.0476115656 -0.017373281  0.0439050785
#>  [81,]  0.1357549598  0.149367412  0.0398247867  0.035518221  0.0315915325
#>  [82,]  0.0540559989 -0.104221499  0.4032856691 -0.044805381 -0.0239363016
#>  [83,]  0.0492343745 -0.089248197  0.9138656132 -0.006721843  0.0038314486
#>  [84,]  0.2359720686  0.160597931  0.0778600291 -0.041630979 -0.0303803701
#>  [85,] -0.0515612919  0.101682700  0.1516719274  0.046272723 -0.0183495949
#>  [86,]  0.0939973131 -0.042339121  0.9672236489 -0.012299518 -0.0470291610
#>  [87,]  0.1040000357 -0.087639851  0.4312430230 -0.001424167  0.0249677304
#>  [88,]  0.2695428462  0.046531530  0.5039892468 -0.045294641 -0.0070858952
#>  [89,]  0.2962495377  0.285824760  0.0386123707  0.033631678 -0.0094294017
#>  [90,] -0.0087333011  0.179379999  0.6170887788  0.025637897 -0.0481999118
#>  [91,]  0.2379382419 -0.213885036  0.4460636535 -0.039511600 -0.0094946513
#>  [92,] -0.0905929700 -0.042616128  0.2840734843  0.039075641  0.0253177606
#>  [93,] -0.1706739229  0.060549401  0.7150527081  0.037107969  0.0161232012
#>  [94,] -0.1360831914 -0.033029253  0.2849462431  0.045537271  0.0454102998
#>  [95,] -0.1008578029 -0.033352669  0.7790467544 -0.048247618  0.0063011053
#>  [96,] -0.0385805443  0.093954275  0.6583300483  0.034860211 -0.0048989431
#>  [97,] -0.0117020517  0.106078272  0.6386217548  0.031360758  0.0311486642
#>  [98,]  0.0948303088 -0.031769684  0.9743277327  0.035160720 -0.0444968065
#>  [99,] -0.0372878187  0.093342343  0.6565817882 -0.039983402  0.0338841403
#> [100,]  0.1380934928  0.119572152  0.0816276323  0.004460126 -0.0494551719
#>                 [,6]          [,7]
#>   [1,] -0.0110229467  0.0280678025
#>   [2,]  0.0326227301  0.0004336454
#>   [3,] -0.0325215473  0.0229241136
#>   [4,]  0.0068436270 -0.0292481343
#>   [5,]  0.0423433360  0.0054062088
#>   [6,] -0.0485084417  0.0456845059
#>   [7,] -0.0462298115  0.0381087259
#>   [8,]  0.0140330959  0.0279816974
#>   [9,]  0.0095162804 -0.0488384327
#>  [10,]  0.0483097026  0.0260525496
#>  [11,]  0.0056718194  0.0321702650
#>  [12,]  0.0258407044  0.0480019712
#>  [13,] -0.0454658790  0.0013045795
#>  [14,]  0.0369926601  0.0180185257
#>  [15,] -0.0422524688 -0.0086155087
#>  [16,] -0.0139214773 -0.0418833989
#>  [17,] -0.0429054409  0.0409441932
#>  [18,]  0.0354668033 -0.0329860271
#>  [19,] -0.0006494006 -0.0349944814
#>  [20,]  0.0468205287 -0.0090962249
#>  [21,] -0.0320514872 -0.0482844761
#>  [22,]  0.0037088885 -0.0326037734
#>  [23,] -0.0296484943 -0.0213717327
#>  [24,]  0.0304009301  0.0271356672
#>  [25,]  0.0288791345 -0.0199058671
#>  [26,]  0.0427317785 -0.0067966312
#>  [27,] -0.0173631754 -0.0227669221
#>  [28,]  0.0426639666  0.0259825175
#>  [29,] -0.0119858269 -0.0466565744
#>  [30,]  0.0001357275 -0.0166667101
#>  [31,] -0.0394884266 -0.0436671678
#>  [32,]  0.0428039743 -0.0390939416
#>  [33,]  0.0490397459  0.0292753518
#>  [34,]  0.0136631798  0.0245399739
#>  [35,] -0.0158443863 -0.0438238688
#>  [36,]  0.0291489437 -0.0189134000
#>  [37,]  0.0095626371 -0.0306682875
#>  [38,]  0.0155196347  0.0183028130
#>  [39,] -0.0203011862  0.0451175238
#>  [40,] -0.0387987648 -0.0298013397
#>  [41,] -0.0382789159 -0.0352702702
#>  [42,] -0.0178381897 -0.0475535302
#>  [43,]  0.0015231580  0.0486317588
#>  [44,] -0.0451707703 -0.0141817606
#>  [45,] -0.0315645338  0.0394274886
#>  [46,] -0.0321344066 -0.0425195831
#>  [47,]  0.0199425494  0.0148866777
#>  [48,]  0.0495089726  0.0361127549
#>  [49,]  0.0271934611 -0.0157036036
#>  [50,] -0.0075116255  0.0462665078
#>  [51,]  0.0132151721 -0.0145536509
#>  [52,]  0.0457658006 -0.0125137276
#>  [53,] -0.0375053124 -0.0002467945
#>  [54,]  0.0291964776 -0.0161143983
#>  [55,] -0.0238335003 -0.0264748553
#>  [56,]  0.0370033091 -0.0390767797
#>  [57,]  0.0328116854  0.0096893140
#>  [58,] -0.0454200732 -0.0098315520
#>  [59,]  0.0055158303 -0.0325448072
#>  [60,]  0.0031474810  0.0409176209
#>  [61,]  0.0294143023  0.0076397389
#>  [62,]  0.0228433051  0.0296082929
#>  [63,] -0.0462891246 -0.0299848676
#>  [64,] -0.0066806800 -0.0334942108
#>  [65,]  0.0169367977  0.0310398541
#>  [66,]  0.0444890694  0.0366358405
#>  [67,]  0.0252456949 -0.0383800159
#>  [68,] -0.0399235833 -0.0017355042
#>  [69,]  0.0236036867 -0.0294951793
#>  [70,] -0.0077346393  0.0392258976
#>  [71,] -0.0394991875  0.0337399670
#>  [72,]  0.0397848870  0.0329826276
#>  [73,]  0.0295114224  0.0441331063
#>  [74,] -0.0172039648  0.0319469451
#>  [75,]  0.0487822811  0.0029284820
#>  [76,] -0.0165812812  0.0264666552
#>  [77,] -0.0068373466 -0.0416067633
#>  [78,] -0.0342513664 -0.0107712212
#>  [79,] -0.0347471705  0.0151850512
#>  [80,]  0.0485392787 -0.0047690683
#>  [81,] -0.0349301712 -0.0050426709
#>  [82,]  0.0166795245  0.0311217160
#>  [83,] -0.0329281465 -0.0444642481
#>  [84,] -0.0354682682  0.0006159852
#>  [85,] -0.0059760903 -0.0431443637
#>  [86,] -0.0266563027 -0.0276904524
#>  [87,] -0.0023973350  0.0206516870
#>  [88,] -0.0339686426  0.0191506107
#>  [89,]  0.0467595117 -0.0058095852
#>  [90,]  0.0203179860  0.0220340129
#>  [91,] -0.0009649931  0.0196805966
#>  [92,] -0.0186282847 -0.0347807297
#>  [93,] -0.0322077814  0.0182979421
#>  [94,]  0.0299500553 -0.0246130962
#>  [95,] -0.0055100260 -0.0390620256
#>  [96,]  0.0312982025  0.0029677907
#>  [97,]  0.0341319787  0.0313811888
#>  [98,] -0.0423835285  0.0491580483
#>  [99,] -0.0055301218  0.0393861586
#> [100,] -0.0246701091  0.0439624317