Generate data representing small spheres within a larger encompassing sphere with added noise.

This function generates data points representing small spheres within a larger encompassing sphere and adds noise to the data if specified.

Usage

diff_sphere(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate, should be a multiple of 13.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

diff_sphere(n = 390, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>                [,1]         [,2]          [,3]         [,4]          [,5]
#>   [1,]   2.78088833   2.33996313  7.756832e+00  -2.28756152  2.119752e-02
#>   [2,]   3.32103033  -0.86843469  4.373069e+00  -4.16986674  4.407204e-03
#>   [3,]   0.45684204   1.02324355  5.238316e+00  -3.73315615 -4.109340e-02
#>   [4,]   0.89983437   0.72306896  4.830636e+00  -4.38277590  1.425837e-02
#>   [5,]   5.15944620  -0.61704856  6.254079e+00  -4.47453585 -4.164189e-02
#>   [6,]   2.54872579  -2.93952267  5.957923e+00  -3.14162232  4.320562e-02
#>   [7,]   1.67424357  -1.53473004  8.100425e+00  -6.06000378 -9.122338e-03
#>   [8,]   2.94599408  -0.90306603  6.334628e+00  -0.89471573 -7.998826e-03
#>   [9,]   2.52333494  -1.95690185  9.566363e+00  -3.70342638 -2.149705e-02
#>  [10,]   3.03119295  -1.47114349  8.686979e+00  -1.51117135 -3.295747e-02
#>  [11,]  -0.07515339  -1.85423756  7.359036e+00  -3.72087577 -1.539029e-03
#>  [12,]   0.36389217   1.16449063  5.492215e+00  -3.27933413 -1.607065e-02
#>  [13,]   0.18349537  -0.14355168  5.517153e+00  -2.49074182 -6.052583e-03
#>  [14,]   0.63627708  -0.92312188  6.382434e+00  -5.84135289  3.711165e-02
#>  [15,]   3.62216018   2.39120633  6.541552e+00  -3.02663770 -1.956939e-02
#>  [16,]   2.95886205   0.58875588  9.379882e+00  -5.37923079  2.220976e-02
#>  [17,]   3.10982945   1.73317211  7.034864e+00  -5.83015331  1.107535e-02
#>  [18,]   2.36896308   2.25760823  7.467758e+00  -2.01648679 -3.592733e-02
#>  [19,]   1.44278377   0.99873376  9.621032e+00  -4.17443089 -2.348383e-02
#>  [20,]   3.10442957  -2.02525155  8.447442e+00  -5.54462141 -9.677381e-04
#>  [21,]   2.00805980  -0.45660214  5.820367e+00  -1.00158215  8.094144e-03
#>  [22,]   4.82145458   1.44390950  6.887884e+00  -2.99137394  1.142912e-03
#>  [23,]   4.17006872  -0.57302599  9.181373e+00  -2.39659017 -6.317902e-04
#>  [24,]   3.31489956  -1.04604549  4.482319e+00  -4.43470821  3.448119e-02
#>  [25,]   2.72927300  -1.22213525  5.380183e+00  -5.87518109 -5.200418e-03
#>  [26,]   4.40755931   1.23179347  7.133556e+00  -1.93825370 -4.962016e-02
#>  [27,]   5.05400867  -1.42500385  7.945883e+00  -3.64033361 -2.542628e-02
#>  [28,]   3.04386296   0.11056112  4.662266e+00  -5.23705633 -4.550321e-02
#>  [29,]   2.22140836   1.92934171  7.941034e+00  -1.75545406  2.060716e-03
#>  [30,]   3.72478060  -0.05593981  7.748657e+00  -6.28996441  2.425135e-02
#>  [31,]   1.44757500   1.00528708  4.810415e+00   7.98394487 -4.435086e-02
#>  [32,]   1.98771701  -2.20311074  1.426653e+00   6.10163964 -3.957419e-02
#>  [33,]  -0.87647128  -0.31143250  2.291899e+00   6.53835024  4.814540e-02
#>  [34,]  -0.43347896  -0.61160709  1.884220e+00   5.88873048 -1.139712e-02
#>  [35,]   3.82613287  -1.95172461  3.307663e+00   5.79697054 -3.070283e-02
#>  [36,]   1.21541247  -4.27419872  3.011506e+00   7.12988406 -1.004428e-02
#>  [37,]   0.34093025  -2.86940609  5.154008e+00   4.21150260  4.311327e-02
#>  [38,]   1.61268076  -2.23774208  3.388211e+00   9.37679066  1.734305e-02
#>  [39,]   1.19002161  -3.29157790  6.619946e+00   6.56808000 -1.427448e-02
#>  [40,]   1.69787963  -2.80581953  5.740563e+00   8.76033503 -3.566219e-03
#>  [41,]  -1.40846672  -3.18891361  4.412620e+00   6.55063061 -2.170764e-03
#>  [42,]  -0.96942116  -0.17018542  2.545799e+00   6.99217225  2.559562e-02
#>  [43,]  -1.14981795  -1.47822773  2.570737e+00   7.78076456  3.950962e-02
#>  [44,]  -0.69703624  -2.25779793  3.436018e+00   4.43015349 -4.606274e-02
#>  [45,]   2.28884685   1.05653028  3.595136e+00   7.24486868 -4.043523e-02
#>  [46,]   1.62554872  -0.74592017  6.433466e+00   4.89227560  4.265123e-02
#>  [47,]   1.77651613   0.39849606  4.088448e+00   4.44135307  1.519863e-02
#>  [48,]   1.03564976   0.92293218  4.521342e+00   8.25501959 -2.671261e-02
#>  [49,]   0.10947045  -0.33594228  6.674615e+00   6.09707549  1.131513e-03
#>  [50,]   1.77111625  -3.35992760  5.501026e+00   4.72688497 -2.650298e-02
#>  [51,]   0.67474647  -1.79127819  2.873951e+00   9.26992423 -8.124357e-03
#>  [52,]   3.48814125   0.10923345  3.941468e+00   7.28013244  3.844069e-02
#>  [53,]   2.83675540  -1.90770204  6.234956e+00   7.87491621  1.426528e-02
#>  [54,]   1.98158624  -2.38072153  1.535903e+00   5.83679817  1.395655e-02
#>  [55,]   1.39595967  -2.55681130  2.433766e+00   4.39632530 -2.688330e-02
#>  [56,]   3.07424598  -0.10288258  4.187140e+00   8.33325268  3.181555e-02
#>  [57,]   3.72069534  -2.75967990  4.999467e+00   6.63117278 -1.072513e-02
#>  [58,]   1.71054964  -1.22411493  1.715850e+00   5.03445006  2.000969e-03
#>  [59,]   0.88809504   0.59466567  4.994617e+00   8.51605233  1.156971e-02
#>  [60,]   2.39146727  -1.39061585  4.802240e+00   3.98154197  1.686676e-02
#>  [61,]   2.83257555   7.26519111  7.246169e+00   3.85102336 -2.791214e-02
#>  [62,]   3.37271756   4.05679329  3.862407e+00   1.96871813  6.160927e-03
#>  [63,]   0.50852926   5.94847153  4.727654e+00   2.40542873  2.125418e-03
#>  [64,]   0.95152159   5.64829694  4.319974e+00   1.75580897  2.047032e-02
#>  [65,]   5.21113342   4.30817942  5.743417e+00   1.66404903 -2.087757e-02
#>  [66,]   2.60041301   1.98570531  5.447261e+00   2.99696255 -2.689253e-02
#>  [67,]   1.72593080   3.39049794  7.589762e+00   0.07858110 -4.156664e-02
#>  [68,]   2.99768131   4.02216195  5.823965e+00   5.24386915 -3.113350e-03
#>  [69,]   2.57502216   2.96832613  9.055701e+00   2.43515849  1.225311e-02
#>  [70,]   3.08288017   3.45408450  8.176317e+00   4.62741352 -9.255104e-03
#>  [71,]  -0.02346617   3.07099042  6.848374e+00   2.41770910 -2.728511e-04
#>  [72,]   0.41557939   6.08971861  4.981553e+00   2.85925074 -3.555190e-02
#>  [73,]   0.23518260   4.78167630  5.006491e+00   3.64784306 -3.454367e-02
#>  [74,]   0.68796430   4.00210610  5.871772e+00   0.29723198  1.519724e-03
#>  [75,]   3.67384740   7.31643431  6.030890e+00   3.11194717  4.914756e-02
#>  [76,]   3.01054927   5.51398386  8.869220e+00   0.75935409 -3.495483e-02
#>  [77,]   3.16151667   6.65840009  6.524202e+00   0.30843156  1.688714e-02
#>  [78,]   2.42065030   7.18283621  6.957096e+00   4.12209808 -1.884202e-02
#>  [79,]   1.49447100   5.92396175  9.110370e+00   1.96415399 -8.854334e-03
#>  [80,]   3.15611680   2.89997643  7.936780e+00   0.59396346 -3.923091e-02
#>  [81,]   2.05974702   4.46862584  5.309705e+00   5.13700272 -1.409715e-02
#>  [82,]   4.87314180   6.36913748  6.377222e+00   3.14721093 -2.392911e-02
#>  [83,]   4.22175594   4.35220199  8.670711e+00   3.74199471  6.578788e-03
#>  [84,]   3.36658679   3.87918250  3.971657e+00   1.70387666  4.784717e-02
#>  [85,]   2.78096022   3.70309273  4.869521e+00   0.26340379 -4.371274e-02
#>  [86,]   4.45924653   6.15702145  6.622894e+00   4.20033117  2.616970e-03
#>  [87,]   5.10569589   3.50022413  7.435221e+00   2.49825127 -8.407147e-04
#>  [88,]   3.09555018   5.03578910  4.151604e+00   0.90152855 -4.906835e-02
#>  [89,]   2.27309558   6.85456970  7.430371e+00   4.38313082  2.320366e-02
#>  [90,]   3.77646782   4.86928818  7.237995e+00  -0.15137954 -4.976807e-02
#>  [91,]  -0.84909490   4.82061915 -8.911128e+00  11.02871101  2.188159e-03
#>  [92,]  14.06405455   2.46551720  4.369364e+00  -1.42557087 -2.770147e-02
#>  [93,]  -6.68796665  -6.58344707  7.933624e+00   8.60156557  2.282448e-02
#>  [94,]   7.08727512   4.57838367 -1.208516e+01  -2.78527555  2.526256e-02
#>  [95,]   5.15325766   4.49826227 -3.543960e+00  12.87050585 -1.929894e-02
#>  [96,]  -1.87358020  -4.58735991 -3.531904e-01 -14.15348306 -2.057966e-02
#>  [97,]  10.78482104   3.47761218 -8.222802e+00  -5.38324907  8.523359e-03
#>  [98,]  -5.84589698 -13.00165985  4.807473e-01   4.64232826 -2.497946e-02
#>  [99,]   5.69056445  -8.07706921  3.009031e+00 -10.87769103  3.678720e-02
#> [100,]   7.83998131  -9.36188643 -8.681732e+00  -0.71923866 -3.312055e-02
#> [101,] -11.70967257   7.10565234 -6.114948e+00   0.02622236  3.714656e-02
#> [102,]  -0.68262355   6.25962808  1.334683e+01  -2.68572139 -1.496564e-02
#> [103,]   4.45628930 -10.87383272 -7.471218e+00   5.57513624  7.110611e-03
#> [104,]  -9.33547250   6.00410158  5.341532e+00   8.55965848  4.416161e-02
#> [105,] -11.19419006   2.86417291  7.086694e+00   6.42381452 -2.221742e-02
#> [106,]   7.18704158  -3.26928583 -1.199843e+01   4.32388034  4.160060e-02
#> [107,]  -8.60949249  -8.70298165 -8.337912e+00   2.36938321  4.942870e-02
#> [108,]  -9.11790159  -9.18371283  3.485149e+00  -6.73624736  2.812343e-02
#> [109,] -13.87470074  -4.70343324 -2.651979e+00  -1.82685597  1.233664e-02
#> [110,]   7.87421132  -2.15046238 -1.257853e+01   0.39089729  2.587666e-03
#> [111,]   4.20917080  -3.67880827  1.215939e+01  -6.77484605  4.334489e-02
#> [112,]   8.24902968   2.80931866 -1.045212e+01   6.30987164 -2.238674e-02
#> [113,] -13.78064644  -2.08814724 -5.142934e+00  -2.06970023  1.035265e-02
#> [114,]   1.93807868 -14.01736547 -2.761786e+00   4.13882259  1.938353e-02
#> [115,] -13.23418941  -2.83364029 -7.965623e-02   6.46686697  2.516532e-02
#> [116,]  10.20714913  -9.31750685 -3.481761e+00   4.67712628  2.846954e-02
#> [117,] -12.56150617   3.98780470 -6.624454e+00  -2.72444195  3.558132e-02
#> [118,]   7.10135008 -10.33482027 -5.561223e+00   6.06919396 -4.653549e-02
#> [119,]   8.92111815  -2.67343138 -1.128594e+01  -3.30059827 -4.397912e-02
#> [120,]  -9.12567834   1.38918887 -1.117993e+01  -3.84725210 -9.837862e-03
#> [121,] -12.02653902   5.99763082  2.994981e+00  -5.95154380  1.502584e-02
#> [122,]   4.61562234   9.02811174  8.902368e+00   6.55263879 -2.768992e-02
#> [123,]   1.82827207  12.36953200  5.755515e+00  -5.96038162 -2.038591e-02
#> [124,]  -1.48649732   7.46522351  1.446728e+00 -12.84397684  4.164279e-03
#> [125,]  -4.01897469   5.37888999  9.705661e-01  13.37809352 -2.942143e-03
#> [126,]   0.01281682   7.48973979 -5.705337e+00 -11.67701853 -4.200428e-02
#> [127,]  -3.65061413  10.87196804 -7.154086e+00   6.50325874  1.614731e-02
#> [128,]   6.00206974  12.69336417 -2.957437e+00   4.37118158 -3.910330e-02
#> [129,]  10.74543620  -5.31727398  3.181131e+00   8.43460739 -4.759995e-02
#> [130,]  -8.55564129  -2.84631303  1.191724e+01  -1.29573661  1.271391e-03
#> [131,]   4.48229254  -0.51270739  1.034662e+01  -9.87894956 -2.682387e-02
#> [132,]   2.07231755  14.70501708 -1.289456e+00   1.67489566 -4.068318e-02
#> [133,]  10.98577946  -7.69388518  4.028438e+00   5.37479921  2.545696e-02
#> [134,]  -4.24963750 -10.02635432  2.910224e+00  -9.89663576 -4.834636e-02
#> [135,]  -9.48967386   5.17655089 -8.926955e+00  -5.33468757 -4.049201e-02
#> [136,]   8.30789802   7.07958303 -9.571668e+00   3.77379237 -2.050693e-02
#> [137,]   0.08004273  -2.20814498  1.367535e+01  -5.75347902 -3.924001e-02
#> [138,]   7.99317236   2.23323306 -9.119598e+00  -8.54135815  6.726639e-03
#> [139,]   4.71958034   2.95097124  3.758144e+00  13.41244521 -1.627577e-02
#> [140,]   1.00187571   3.97974522  7.215899e+00 -12.49354543 -3.529330e-02
#> [141,]   7.48017648   1.49985198  3.731753e-01 -12.90961441  4.564999e-02
#> [142,]  14.08493228  -0.19093788  4.131359e+00  -3.08384430  3.806127e-02
#> [143,]   3.60874800   7.77875007 -1.230674e+01  -0.11045650 -9.438175e-03
#> [144,]  -3.48317194 -11.77533594 -5.579703e+00  -6.56322278  4.568874e-02
#> [145,]  -4.32714962  -9.02116989 -4.307944e+00 -10.31192939  9.076306e-03
#> [146,]   5.07344638 -12.25878698 -1.622742e+00   6.80800943 -7.311449e-03
#> [147,]   1.82108150   9.62133191  7.960405e+00  -8.10836549  1.199882e-03
#> [148,]  10.64334431   9.06718538  4.027280e-01  -5.41693469 -4.364454e-02
#> [149,]  -5.11794706   0.24900933  1.409727e+01   0.10812122 -2.062866e-02
#> [150,]  12.41691281   7.99666493 -1.421187e+00  -2.20314635  4.274155e-02
#> [151,]   2.54319891   3.66398681 -1.406208e+01  -2.71392045  2.467376e-03
#> [152,]   5.65678905   0.64274664 -5.774427e+00  12.61917633  3.234391e-02
#> [153,]  -5.92319267   6.13257729 -1.099820e+01   5.59882974  2.938651e-02
#> [154,]  -8.55656555   5.43272411 -9.018918e+00   6.39764153 -3.631522e-02
#> [155,]  -2.56432828   6.88213724 -1.021692e+01  -8.16546815  1.149788e-02
#> [156,]   2.26005071   1.52032962  5.096777e+00  13.84209641 -4.396800e-02
#> [157,]   8.94122367   4.86890860 -1.016303e+01   4.24983958  4.825960e-02
#> [158,]   0.71554126   8.71436071 -1.161578e+01   3.69073238  4.995151e-02
#> [159,]   3.63263507   6.22793917 -6.662920e+00 -11.34117450 -2.920059e-02
#> [160,]  -3.70918716   5.40018097 -7.103908e+00 -11.47233487  2.500942e-02
#> [161,]   1.14035696   3.96088105  2.823761e+00  14.14345711 -3.697154e-02
#> [162,]  -4.83316681   3.89404417  1.363207e+01  -0.80226503  4.152219e-02
#> [163,]  11.58973045   0.66607138 -9.417585e+00  -1.24241206  1.895963e-02
#> [164,]  -7.83685494 -10.39876168 -1.797911e-01   7.44426861 -3.878539e-02
#> [165,]   9.00076739   5.70866077  1.055442e+01   0.04049619  2.841763e-02
#> [166,]   3.32213302 -12.91415528 -6.720878e+00   1.42050199 -3.690353e-02
#> [167,]  -2.04753066   3.95188916  4.154302e+00 -13.70882813 -2.794753e-04
#> [168,]   6.04112256   1.14729421  4.157351e-01 -13.67536904 -3.719877e-02
#> [169,]  -8.11789964  -0.71107335 -4.003459e+00  11.94011721 -1.607803e-02
#> [170,]   8.76418758  -8.02785173 -5.422697e+00  -7.37136175  9.830789e-03
#> [171,]  -5.16858551   8.39984315 -8.422616e+00   7.53577463  5.660332e-03
#> [172,]  -6.58148496  11.08846697 -7.250854e+00  -2.48094161 -4.339424e-02
#> [173,]  -6.79430718  -2.35781184 -1.024652e+01   8.26358703  4.127370e-03
#> [174,] -13.22910868   0.79404750  5.054980e+00  -4.87927797  2.912364e-02
#> [175,]  -2.20878992   7.65180139 -5.209106e+00 -11.59467092 -4.891797e-02
#> [176,]   2.67404211   6.08039908  6.183172e+00 -11.94347658  1.799517e-02
#> [177,]   3.68758095  -8.70439101 -4.799938e+00  10.61112257  3.651148e-02
#> [178,]   8.92521286   4.93805807  1.056185e+01  -3.06649729 -4.843824e-03
#> [179,]  -0.93520345   7.46262141 -9.801248e+00  -8.50706893 -2.179660e-02
#> [180,]   7.19996173  -0.79990673 -1.269719e+01  -3.36185350  1.485232e-02
#> [181,]  -1.52157888  13.94432421 -5.313879e+00   0.05755983 -1.007181e-02
#> [182,]  -0.44242495  13.43752706 -5.471918e+00   3.78090466  2.219363e-02
#> [183,]  -0.18219003   4.31235536 -1.387914e+01   3.70674833  2.136460e-02
#> [184,]  -3.14054774 -13.29144371 -2.455856e+00  -5.69589832 -6.395668e-04
#> [185,]   2.67623444 -13.34178953  6.303346e+00  -0.31975877 -2.953338e-02
#> [186,]  -2.35823312  -2.05330124  1.173430e+01   8.80505449  2.085189e-02
#> [187,]  -5.00646958  -2.17096255  1.283183e+00  13.91314579 -1.336674e-02
#> [188,]   5.60728896   1.34872380 -1.037289e+01   9.17291998 -1.248787e-02
#> [189,]   1.14203744  -7.96993475 -1.228844e+01  -3.02821551  1.586938e-02
#> [190,]  -2.78107014   7.70301784 -5.950136e+00  11.06910317 -3.877860e-02
#> [191,]   6.16794809   5.12095516  9.783605e+00   8.06308362 -6.385443e-03
#> [192,]  10.59483329   5.81720900 -4.931199e+00   7.38869856  2.024609e-02
#> [193,]  -7.03933327  -9.70035018 -8.975049e+00   0.89413936 -3.618420e-02
#> [194,]   3.79146975   8.57040949  4.848151e+00  10.66153205 -1.515928e-02
#> [195,]  -2.28450812   9.56126295 -1.076601e+01   3.52935165  4.991564e-04
#> [196,]   1.10462203   9.09562829  1.032481e+01  -5.86921375  4.780127e-02
#> [197,]  -0.81476402   7.86138133  1.259302e+01  -1.98764028 -9.849798e-03
#> [198,]  -5.84524703   2.50182030 -4.362416e+00  12.86636337 -3.725998e-03
#> [199,]  12.16132976   6.18074582 -9.378789e-01   6.16610273  4.002719e-03
#> [200,]  -6.55140978  10.36952710  4.544906e+00   7.34137371  2.499095e-02
#> [201,]   0.74696658  -5.22270568  4.711116e+00 -13.22765185 -5.971168e-05
#> [202,]   6.26149739   8.96087764  5.857786e+00  -8.43698171  4.778684e-02
#> [203,]  -8.25163406  -4.99037692 -2.635511e+00  11.18305663  4.884081e-02
#> [204,]  10.54381127   9.12520776  1.251322e+00   5.38449815 -3.081416e-04
#> [205,] -14.62898031   1.17424741 -3.051567e+00   0.54956132 -1.157775e-02
#> [206,]  -5.96447861  -6.73759103  7.724950e+00  -9.18449792  1.140528e-02
#> [207,] -11.59420519   8.47015778  4.274960e+00   0.74535277 -2.588702e-02
#> [208,] -11.78118339   6.47092126 -6.119366e+00  -2.62378641 -4.723779e-03
#> [209,]   3.35247509   3.74739220  6.680260e+00  12.45359733  1.481597e-03
#> [210,]  -1.56528921  -9.45712776 -9.636461e+00   6.34438447 -4.744274e-02
#> [211,]   8.33374423   5.10319002 -2.409072e+00  11.12216389  2.334709e-03
#> [212,]  10.78378863   1.78474595  1.025335e+01   0.62729371  2.957095e-02
#> [213,]  14.18424258   3.90942273 -2.745220e+00   0.99370178 -3.022071e-03
#> [214,]  -5.49021285  -2.91032610  1.173668e+01   6.97409625 -2.955389e-02
#> [215,]  -8.56548526   9.59497583  3.688905e+00  -6.77944524  3.516651e-02
#> [216,]   8.97259619  -3.81647445  4.943162e+00  10.27093902  2.309665e-02
#> [217,]   8.55626339  -3.01253472  9.275927e+00   7.52809176 -1.678585e-02
#> [218,]   7.75432793  -5.50664690  5.731194e+00 -10.08467411 -3.970648e-02
#> [219,]   7.95751183  -5.46643290  1.665555e+00 -11.35878706  2.288596e-02
#> [220,]  -9.09784686   7.80843478 -6.379159e-01  -8.99169573  1.882477e-02
#> [221,]  -6.22042015   3.37948479  1.277563e+01   3.41593854 -4.588844e-02
#> [222,]   3.70181807 -14.23242018  2.931654e+00   0.37438814  3.444220e-02
#> [223,]  -5.02249108   0.85157595  1.019281e+01   9.75479908 -6.757220e-03
#> [224,]   7.15940945   9.85475957  8.252042e+00   2.92068101 -1.470961e-02
#> [225,]   0.48633759  10.70452723  6.122139e+00  -8.52619380  2.290608e-02
#> [226,]  -3.88676657  -3.80143088 -4.619115e+00 -13.19492128  4.690872e-02
#> [227,]   0.86215471  -6.32748659  6.300798e+00  12.02162852  3.119051e-02
#> [228,]  12.77293838  -2.17198335  4.368455e+00   6.16856010 -2.992585e-02
#> [229,]   3.09358398   4.38880432  3.069056e+00  13.66561482  4.988285e-02
#> [230,]  -3.39337725  -6.76856621 -9.572133e+00   8.72042305  1.310880e-02
#> [231,] -12.50990154  -7.66477596  2.956142e+00  -1.00737110  2.897276e-02
#> [232,]  -1.53636565  -3.62634597  3.124906e+00  14.13237966 -4.884262e-02
#> [233,]  13.55154117   6.30726834 -1.220892e+00   0.28900074 -4.314493e-02
#> [234,]  -4.59193932  -4.57223216  4.469871e+00  12.76828249 -1.916850e-02
#> [235,]   9.63313972  -7.81948814  6.291120e+00  -5.61070673  2.920335e-02
#> [236,]   2.24201849  -7.30273294 -9.115003e+00   9.14112517  2.193589e-02
#> [237,]  -2.53018740   5.43476838  1.371099e+01  -1.03456639  3.378874e-02
#> [238,]  -2.98544857  -7.21957565  1.059715e+01   7.18784770  3.349941e-02
#> [239,] -12.42164820  -5.85233140  3.932671e+00   4.58115441 -1.281368e-02
#> [240,]   8.50562021 -12.11296889  1.369204e-01  -2.43139104  3.606427e-02
#> [241,]  -6.13781227   6.22873492  1.155686e+01  -3.86898538  3.735948e-02
#> [242,]   3.05174517  10.83698203 -7.672519e+00  -6.27527891  2.248131e-02
#> [243,]   6.70814191  -9.24771074 -9.049511e+00   3.54781910 -3.676185e-02
#> [244,]  -3.24178758 -13.84766250  3.145496e+00  -3.58314270  4.042590e-03
#> [245,]   0.89931454  -5.82588198  1.296423e+00  13.73206538 -6.437020e-03
#> [246,] -11.90533269   6.16931167 -3.117420e+00   5.95687343 -2.625819e-03
#> [247,]   6.38427102   3.66346111  1.028620e+01   8.06313540  4.278188e-02
#> [248,]   5.13801212   8.73192758 -1.096312e+01  -1.47117542 -3.382415e-03
#> [249,]   8.52923525 -11.55129281  1.700301e+00   3.99108458 -3.115502e-02
#> [250,] -11.99275394   7.55217050  4.909413e+00   0.19036165 -1.687215e-02
#> [251,]   7.58637099  -3.42027306  4.968165e+00 -11.44840799  3.293582e-02
#> [252,]   7.78424779  -9.78272652  5.209801e+00  -6.44683827  4.259266e-02
#> [253,]  -0.97780138   0.34168021  3.086058e+00  14.64252038 -1.871295e-02
#> [254,] -11.83534767  -8.00454269  3.837829e+00  -2.47445142 -4.298311e-02
#> [255,] -11.05439617  -3.50054450  7.506840e+00  -5.84755237 -2.297925e-02
#> [256,]   1.54508509  -6.14051786 -1.351307e+01  -1.51777445 -2.238484e-02
#> [257,]  -0.10545042  -1.91638054  1.233227e+01   8.32055091  2.481259e-02
#> [258,]  12.39087282  -4.46750349 -6.169083e+00  -3.66743807 -1.953493e-02
#> [259,]  11.89068836  -0.06980639  8.208648e+00  -4.02799727  1.248788e-02
#> [260,]  -0.37242572   9.87507779  4.012041e+00 -10.54740070  1.395239e-03
#> [261,]  -0.09190045  10.87093924 -1.025925e+01  -1.24982564  3.772188e-02
#> [262,]  -1.10776451  -1.42903156  1.456364e+01  -3.10340732  4.284242e-02
#> [263,]  -2.52768733   8.81799516  1.173904e+01   1.74607606 -2.257583e-02
#> [264,]   3.04240377   4.39118428 -1.705135e+00  13.91236118 -2.504987e-02
#> [265,]   2.05663240 -12.20623333 -8.435658e+00   0.78600668  4.321080e-02
#> [266,]  -9.21785370   0.87533054  3.823157e+00 -11.16460675 -3.126673e-02
#> [267,]   5.78584068  -1.68351886  1.339951e+01  -3.02373176  3.543568e-02
#> [268,]   4.08558517 -11.76877895  4.616417e+00   6.96365785 -3.644009e-02
#> [269,]   1.31782128  -2.61619627  3.538359e+00 -14.27931660 -4.734126e-03
#> [270,]  12.66376324  -8.02986319 -1.177347e-01  -0.36950828 -3.691742e-02
#> [271,]  -2.26316612 -11.14401737 -9.090558e+00  -3.61257621  1.018689e-02
#> [272,]   6.48085474   4.42844586 -8.997203e+00  -9.07952287  4.300803e-02
#> [273,]   7.03464083  -1.63249495 -7.383728e+00  10.87792939  3.014390e-02
#> [274,]   3.56656194   1.38959143  1.433854e+01   2.18058042  6.829785e-03
#> [275,]  11.59882729  -6.02171279 -6.686075e+00  -3.08262470 -4.809596e-02
#> [276,]  -9.83345716   5.79307711 -5.773716e+00   7.83629870  1.844139e-02
#> [277,]   5.61150495  -5.13916884 -6.260878e+00  11.30934839  2.285096e-02
#> [278,]   6.29415355  -4.91030400  1.267149e+01   0.84010426  2.077131e-02
#> [279,]  -1.47385512 -13.35079297  2.737520e+00  -6.09016097 -2.832918e-02
#> [280,]  -8.06716882  -7.57654600 -3.029070e+00   9.66133916 -1.420498e-03
#> [281,]  -4.97785900  -3.54909924 -7.152741e+00  11.68174255 -1.744025e-02
#> [282,]  -5.16291959  -4.54972132 -1.018513e+01   8.59694426  3.571248e-02
#> [283,]   4.13183018  -6.47625343 -1.218097e+01   4.19644323  3.546354e-02
#> [284,]   6.90065538   5.52899891  3.556020e+00 -11.58299831  4.747915e-02
#> [285,]   8.94681939  -8.32208686  8.220444e+00  -2.84984008  1.182741e-02
#> [286,]  -0.56924898   6.14608559  9.351635e+00   9.97238751 -3.595941e-03
#> [287,]  -4.22709769  -5.83323965 -1.199082e+01   5.41528262 -8.067243e-03
#> [288,]   5.88725395   2.90093927  6.347844e+00  11.90082640  1.791996e-02
#> [289,] -11.98473069  -2.84888997  4.414520e+00   7.33226234  4.133530e-02
#> [290,]   0.07326659   3.17026231  4.452528e+00 -13.96850239 -1.960712e-02
#> [291,] -10.79189730   7.84325961  5.397867e+00  -4.22862469 -4.418129e-03
#> [292,]  12.35641487   3.84504610 -7.582886e+00  -0.18568331  5.997418e-03
#> [293,]  -0.25288890  -9.21314048  8.685775e-01 -11.80252782  3.679707e-02
#> [294,]  12.86480635  -2.97321493 -7.093505e+00  -0.58218631  3.331468e-02
#> [295,]   8.95412210 -11.70770671 -1.810225e+00   2.11574726 -4.519436e-02
#> [296,]  -6.40127092   4.61117435 -9.724279e+00  -8.25828064  3.670291e-02
#> [297,]  10.70083017   9.68481046  2.158257e+00   3.46966949 -2.395298e-03
#> [298,]   3.43517953  -8.07877451 -2.347245e+00  11.93412696 -3.058459e-02
#> [299,]  -1.72491199  12.12538267 -7.959761e+00  -3.41203551 -4.129787e-02
#> [300,] -11.14105768   4.02149285  1.769218e+00   9.03184905 -3.444815e-02
#> [301,]  -2.13470750  10.92115621  8.828981e+00   4.81876174 -3.728426e-02
#> [302,]  -1.82100394  -0.68586277 -1.470508e+01   2.23028275  2.255143e-02
#> [303,]  -3.89125339   8.94637642  1.104781e+01  -2.78681123 -3.509616e-02
#> [304,]   6.49299877  -4.67810793 -4.930047e+00 -11.68977783 -1.419003e-02
#> [305,]  -5.44781104  -6.19970764 -3.412423e-04 -12.52537345 -4.256971e-02
#> [306,]   3.44885214  -0.83451595  1.338386e+01   5.76900479  3.014323e-02
#> [307,]  11.03532248   9.65799105 -8.245749e-01   3.04383686  3.632622e-02
#> [308,]  -9.61715167   5.36161569 -6.599920e+00  -7.75915755 -4.061169e-02
#> [309,] -12.73489859   0.83113811  7.452572e+00   2.56724352  8.135768e-03
#> [310,]   2.96463524   7.55558311  2.468346e+00  12.37058490  4.070070e-02
#> [311,]   1.20901702 -14.18087063  4.080688e+00   2.40606977  9.050935e-03
#> [312,]   1.29463806   6.48903807 -1.067646e+01  -8.19935877 -1.097815e-02
#> [313,]   5.63725912   9.89354135 -3.704516e+00  -9.03414136  3.555734e-02
#> [314,]   7.05598727   0.97302655  3.274902e+00 -12.78832594  2.122623e-02
#> [315,]  -8.11562194  11.19229758  4.854813e+00   3.20935258  4.223197e-02
#> [316,]   4.67346969  -4.88398813 -5.490016e-01  13.37923534  4.392982e-02
#> [317,]  -3.47944045 -11.04478087 -6.376686e+00   7.08831305 -4.473538e-02
#> [318,]   4.34660766   8.04034241 -7.860022e+00   8.92636268 -4.779740e-03
#> [319,]  -1.67727021   6.73525944 -3.211314e+00  12.90389511 -2.015911e-03
#> [320,] -13.66548547  -1.83469611  5.530180e+00   2.07497112  2.170264e-02
#> [321,]  10.18682172   5.08514292  2.465432e+00   9.44942489  2.675968e-02
#> [322,] -11.50980434   1.40717189 -1.832113e+00   9.33743187  2.549521e-02
#> [323,]   1.09359308 -11.30293817 -4.336155e+00   8.78893658  8.990197e-04
#> [324,] -12.46158889  -0.54707182 -8.326339e+00  -0.28565759  1.518470e-02
#> [325,]   6.44876187   1.15433212  4.224948e+00 -12.81525655 -3.892260e-02
#> [326,]  -8.16588354   2.05755897 -5.664906e+00  11.04507301  6.933350e-04
#> [327,]   6.51438196   0.37603179  8.563220e+00 -10.44474480  4.043211e-02
#> [328,]  -6.86230242  -7.57831254  9.578653e+00   5.35979336 -2.176775e-02
#> [329,]   5.71687574   2.60803554 -8.809916e-01  13.59188494 -2.395148e-02
#> [330,]  -3.45093944  -0.73775055 -7.783708e+00 -12.32723133  5.037838e-03
#> [331,]  -1.40914726  -3.66080982  6.063270e+00 -13.14722540  8.084562e-03
#> [332,]  -2.39861909   9.64825290 -2.923782e+00  10.84478412  3.412655e-03
#> [333,]  -3.47444621   2.15294229  1.364897e+01   4.69027290 -1.746116e-02
#> [334,]  -6.31563973  -3.93297194 -3.162185e+00  12.63507074 -1.482479e-03
#> [335,]  -5.00498026  10.66853608  8.164872e+00   4.41218533  2.922647e-02
#> [336,] -12.44703113   6.82459173  1.527517e+00   4.60033203  2.402587e-02
#> [337,]   6.48880475  -4.50745535 -9.147728e-02  12.75028985  3.686412e-02
#> [338,] -11.00217657  -2.82100771  8.249341e-02   9.79730683  5.877332e-03
#> [339,]   6.10502583   7.96561559  2.958726e+00 -10.74818903 -2.322444e-02
#> [340,]  -6.99853392  -7.01431611 -7.946760e+00   7.97927925  4.252679e-03
#> [341,]  -3.92921643   4.15519026  1.368701e+01   2.22742091  7.354630e-03
#> [342,] -13.01029966   4.18931531 -6.170189e+00   0.33241871  2.542894e-03
#> [343,]   7.14342384  -2.63001791 -1.273583e+01  -2.20297477 -1.006610e-02
#> [344,]   0.02616510   4.36674164  1.431285e+01  -1.03600334 -4.508139e-03
#> [345,]  -6.47885498   9.13838142 -3.505751e+00  -9.33938611  4.493796e-03
#> [346,]   6.56984030  10.40076772  8.448154e+00   1.51324788 -4.195812e-02
#> [347,]  -2.22311030  -1.29103777  9.489005e+00 -11.32915638 -4.230538e-03
#> [348,]   1.18808662 -11.92954530 -2.739186e+00  -8.58901970 -4.107033e-02
#> [349,]   5.81243624  13.60898253 -5.307336e-01   2.39363764 -3.984716e-02
#> [350,]   6.76329327  -1.49468429  1.294153e+01  -3.08877287 -2.366010e-03
#> [351,]   0.66514503  -8.86400462  3.109906e+00 -11.67542241  2.425938e-02
#> [352,]   7.96121636  -9.41823741  5.322351e+00   6.67745575 -4.524561e-03
#> [353,]   1.20896833  -2.25333876 -1.006170e+01 -10.82696223 -2.335574e-03
#> [354,]  12.94693093  -0.18412670  2.502497e+00  -7.14706825 -2.571968e-02
#> [355,]  14.95411943  -0.97733792 -1.071765e-01   0.63846379 -2.063218e-02
#> [356,]  -3.85826192  -8.46784210 -1.914911e+00 -11.60786721  4.673031e-02
#> [357,]   9.12381439   6.72669139  9.791971e+00   0.79052909 -4.009734e-02
#> [358,]   5.10850896  -2.02378451 -1.336808e+01   4.01272948  1.495716e-02
#> [359,]   8.20368887   1.83903317  1.169934e+01   4.17648124  3.895489e-02
#> [360,]  -1.94728975   7.02845221 -8.067737e+00  10.33056380 -4.867993e-02
#> [361,]  -9.42635936  11.05289474  3.729192e+00  -0.26532423  1.026676e-02
#> [362,]  -5.91712922  -4.99479704  1.234931e+01   3.54034257 -3.737688e-02
#> [363,]  14.12277754  -2.51610639  1.916118e+00  -3.94269633 -4.699990e-02
#> [364,]   6.81995537  -0.01745716  6.756557e+00 -11.52548681  1.615831e-02
#> [365,]   2.90310580   0.72173612 -1.006175e+01  10.71505102 -3.342405e-02
#> [366,]   4.11304969   4.65374439 -1.269676e+01   5.02173112  1.630481e-02
#> [367,] -12.11087695  -6.11464354  3.159175e+00  -5.56393825 -4.429824e-03
#> [368,]  -1.63540297  13.45174471  5.498126e+00  -3.33865725 -2.928207e-02
#> [369,]  -3.24736405  11.07207646 -9.114329e+00  -2.96525857  4.423066e-02
#> [370,]   4.31902059  -1.68477683  4.008720e+00  13.69079066  3.273544e-03
#> [371,] -10.48887895   7.89895910  1.263846e+00   7.14090723 -4.040187e-02
#> [372,]  12.24129447  -8.15428302 -6.690358e-01  -2.86544396  3.172538e-02
#> [373,]   1.22426893 -13.77441927  5.778179e+00   0.61578336 -2.822688e-02
#> [374,]  10.65373401  -6.57415903  7.024117e+00   4.35203035  4.992910e-02
#> [375,]  -3.37672461  -9.38388004 -7.202926e+00   8.58244632  1.916741e-02
#> [376,]   0.34996776  -9.31813713 -2.240368e+00  11.53388900 -1.407647e-02
#> [377,]   5.94294504  -7.94346722  3.161247e-01  11.24645712 -2.473050e-02
#> [378,]  12.28008221   6.92292571  4.504180e+00   2.44643488 -3.923576e-02
#> [379,]  10.60871207   1.74246065  6.899515e+00  -7.86229970 -2.137634e-02
#> [380,]  13.38013650  -4.39647871 -5.119367e+00  -0.65955017  2.458692e-02
#> [381,] -14.48944140   2.72516832 -1.469769e-01   2.75825004 -1.804617e-02
#> [382,]  -8.11686647  -3.93288755 -1.014824e+01   6.37668642 -2.617696e-02
#> [383,]   7.27446733  -6.48262412 -8.678741e+00  -7.39845686  4.641356e-02
#> [384,]  -8.60271170   9.96380044 -7.101968e+00   1.13052584  2.116173e-02
#> [385,]   9.37399846   6.62095264 -7.278573e+00  -6.34929192  3.116627e-02
#> [386,]  -3.23255302  -2.93593054 -9.019753e+00 -11.16131554 -2.020980e-02
#> [387,]  -3.46745723   5.93403535  1.967168e+00 -13.18689564  4.257285e-03
#> [388,]   4.99663777  -8.82586646  8.957187e+00  -6.47352261 -4.714652e-02
#> [389,]  -4.02984834  -1.55057277 -1.114072e+01   9.06865343 -1.788693e-02
#> [390,]  10.39367743   0.70119438 -9.505353e-01 -10.75064085  4.725287e-02
#>                 [,6]          [,7]          [,8]
#>   [1,] -2.725589e-02  4.807062e-02  2.143221e-02
#>   [2,] -1.974407e-02  2.619058e-02  1.400685e-02
#>   [3,]  3.499050e-02  3.819498e-02  3.159583e-02
#>   [4,] -8.270626e-03 -4.797017e-02 -1.806822e-02
#>   [5,]  1.790180e-02  4.466554e-02 -9.069514e-03
#>   [6,] -1.211530e-02 -3.759834e-02  1.622277e-02
#>   [7,]  1.074600e-02 -3.186154e-02 -1.492865e-02
#>   [8,] -2.867444e-02 -4.238409e-02  2.788791e-02
#>   [9,]  4.398592e-02  4.589386e-02  2.266046e-02
#>  [10,]  3.245707e-02 -2.147497e-02  6.136849e-03
#>  [11,] -3.499937e-02  2.742632e-02 -2.495755e-03
#>  [12,]  3.670242e-02  1.177007e-02 -4.814793e-03
#>  [13,]  1.686723e-02  5.833409e-03 -2.457666e-02
#>  [14,] -2.105961e-03  1.148629e-02  2.200348e-02
#>  [15,]  3.177886e-03 -2.272011e-02 -8.636339e-03
#>  [16,] -8.612931e-03  4.067989e-02  2.984439e-02
#>  [17,]  4.452852e-02 -4.220154e-03 -2.966622e-02
#>  [18,]  1.227049e-02  1.734245e-02  7.913500e-03
#>  [19,] -3.011600e-02 -1.419315e-02 -2.888944e-03
#>  [20,]  3.939988e-02  4.983522e-02 -1.360331e-04
#>  [21,] -4.479618e-02  4.891290e-02  2.114406e-02
#>  [22,]  2.617135e-02 -3.944035e-02  3.816315e-03
#>  [23,] -5.317230e-03  4.142519e-02  3.081273e-02
#>  [24,] -9.454503e-03  2.573322e-03 -2.941447e-02
#>  [25,]  7.281610e-04  2.110304e-02  4.864052e-02
#>  [26,]  4.654679e-02 -4.490643e-03 -1.133078e-02
#>  [27,] -1.948486e-02  1.798541e-02 -6.850344e-03
#>  [28,]  5.425305e-03  4.487544e-02  1.959002e-02
#>  [29,]  1.173622e-02 -3.264018e-02  4.913614e-02
#>  [30,] -3.609064e-02 -1.224003e-02 -3.475536e-02
#>  [31,] -2.619458e-02 -4.791761e-02  1.339228e-02
#>  [32,]  4.741295e-02 -1.490193e-02 -4.561594e-02
#>  [33,] -2.685661e-02 -3.747365e-02 -4.290220e-02
#>  [34,] -2.898023e-02  3.705284e-02 -4.110060e-02
#>  [35,]  2.552986e-03 -2.457692e-02 -1.541174e-02
#>  [36,] -1.673233e-03 -6.395966e-03 -4.505088e-02
#>  [37,] -1.092127e-02 -4.841870e-02 -1.246421e-02
#>  [38,]  1.058497e-02  2.090099e-02  7.646789e-03
#>  [39,] -2.232070e-02 -2.787061e-02 -1.795463e-02
#>  [40,]  2.467705e-02 -7.634402e-03 -4.922956e-02
#>  [41,]  2.552546e-02 -3.530192e-02  1.245028e-02
#>  [42,]  1.612519e-02  3.904941e-02  4.748682e-02
#>  [43,] -1.799118e-02 -1.103279e-03  1.983683e-03
#>  [44,] -4.585439e-02  1.502409e-02 -1.628076e-02
#>  [45,]  2.116401e-02 -2.007950e-02 -4.252936e-03
#>  [46,]  3.968487e-02  4.727226e-02 -4.494739e-02
#>  [47,] -2.446568e-02  3.427478e-02  2.612574e-02
#>  [48,]  3.824736e-02  1.312548e-02  6.096628e-03
#>  [49,]  3.179635e-02 -4.860042e-02  2.515916e-02
#>  [50,] -3.849697e-02  1.966940e-02 -1.865056e-02
#>  [51,]  2.109216e-02  1.170285e-02  3.987971e-02
#>  [52,]  4.520037e-02 -1.326909e-03 -2.647774e-02
#>  [53,] -4.645264e-02  3.776715e-02 -4.402300e-02
#>  [54,] -7.009590e-03 -2.198494e-02 -6.902738e-03
#>  [55,] -5.001243e-03  4.407308e-02 -4.605067e-02
#>  [56,]  3.848919e-02 -3.642492e-03  3.408927e-02
#>  [57,]  3.743089e-02 -3.902755e-03 -4.585515e-02
#>  [58,]  3.946946e-02 -4.812149e-02 -3.777249e-02
#>  [59,]  4.481940e-02 -4.081186e-02 -4.723902e-02
#>  [60,]  2.865133e-02 -1.260736e-02 -1.906833e-02
#>  [61,]  1.098665e-03  3.666297e-02 -3.025915e-02
#>  [62,] -5.044512e-03  4.325693e-02 -2.447077e-02
#>  [63,] -4.643550e-02 -3.437925e-02 -2.141784e-02
#>  [64,] -2.988956e-02  3.527072e-03 -4.616009e-02
#>  [65,]  4.858749e-02 -3.574238e-02 -1.579805e-02
#>  [66,]  3.736948e-02 -1.944347e-02  1.153011e-02
#>  [67,]  6.019995e-03  3.824583e-02  4.132720e-02
#>  [68,]  4.661930e-02 -4.324336e-02 -4.736118e-02
#>  [69,] -3.123819e-02 -2.608264e-02  1.861307e-02
#>  [70,]  2.564111e-02  1.384417e-02 -3.318702e-02
#>  [71,] -4.366576e-02  1.167986e-02  4.815551e-02
#>  [72,]  2.854976e-02 -9.400068e-03 -2.699375e-02
#>  [73,]  2.529068e-03 -1.928401e-02  1.804418e-02
#>  [74,] -4.268323e-02  4.864314e-02 -8.653480e-03
#>  [75,]  4.251163e-02 -1.660545e-02  3.828025e-02
#>  [76,]  1.546555e-02  1.039096e-02 -2.217652e-02
#>  [77,] -1.512697e-02 -1.551473e-03  2.621478e-02
#>  [78,] -1.461642e-02 -2.208109e-03  2.919746e-02
#>  [79,]  6.571608e-03  3.980417e-02  4.684597e-02
#>  [80,] -3.290225e-02  4.126378e-02  4.062745e-02
#>  [81,] -1.347212e-02 -1.317601e-02  4.014081e-02
#>  [82,]  1.172697e-02 -3.625839e-02 -3.284147e-02
#>  [83,]  2.242406e-02  7.287593e-03  3.385059e-03
#>  [84,] -4.208867e-02  1.721251e-02  4.768724e-02
#>  [85,] -3.747746e-02  2.732871e-02 -4.739880e-03
#>  [86,] -2.062796e-02  1.308532e-02  7.857894e-03
#>  [87,] -2.875362e-02 -4.002686e-03  4.675842e-02
#>  [88,]  4.285986e-02 -4.042479e-02 -2.390856e-02
#>  [89,] -1.012597e-02  4.800204e-02  2.656806e-02
#>  [90,]  3.914128e-02 -4.534588e-02 -1.227546e-02
#>  [91,] -3.226247e-02 -4.644043e-02 -1.901055e-02
#>  [92,]  3.818125e-02  1.089123e-02  4.100406e-02
#>  [93,]  6.584840e-03 -4.795007e-02 -3.633561e-02
#>  [94,] -3.034217e-02  6.306673e-04  4.793660e-02
#>  [95,]  3.071103e-02 -8.746058e-03  4.084911e-02
#>  [96,] -1.385804e-02  3.087143e-02  2.803676e-02
#>  [97,]  6.941074e-03 -1.616561e-02  3.685599e-02
#>  [98,] -3.756205e-02  2.121655e-02  2.170734e-02
#>  [99,]  3.061010e-02  3.715122e-02  1.423864e-02
#> [100,]  3.544337e-03  3.793681e-02  3.524280e-02
#> [101,]  2.206068e-02  4.126534e-02 -4.610020e-02
#> [102,] -4.379954e-02 -4.625990e-04  2.138170e-02
#> [103,] -1.296058e-02  2.568265e-02  1.212397e-02
#> [104,]  3.806896e-02 -3.626186e-03  2.955113e-02
#> [105,]  1.201140e-02  3.337015e-03 -1.797240e-02
#> [106,]  1.661078e-03 -6.658151e-03 -4.074148e-02
#> [107,] -1.294834e-03  4.456276e-02 -4.882002e-02
#> [108,]  3.263347e-02  2.172690e-02 -4.670028e-02
#> [109,]  4.977925e-02  2.400892e-02  3.258605e-02
#> [110,]  6.884424e-03 -3.494105e-02 -3.057911e-02
#> [111,]  1.887094e-02 -2.781710e-02  8.622365e-03
#> [112,]  4.420158e-02  4.653214e-02  3.449147e-02
#> [113,]  1.669742e-02  4.415947e-02 -1.773614e-02
#> [114,] -8.713724e-03 -4.162183e-02  2.827927e-04
#> [115,]  1.212043e-02 -1.140853e-02 -2.650054e-02
#> [116,] -9.779638e-03  4.884074e-02  2.228075e-05
#> [117,]  2.922469e-03 -8.051722e-03 -4.100980e-02
#> [118,] -9.109209e-03 -4.787244e-02  1.117810e-02
#> [119,]  2.971583e-02  2.633604e-02  1.613483e-02
#> [120,]  3.464228e-02 -4.615733e-02 -2.529527e-03
#> [121,]  2.769897e-02  1.905935e-02  3.513072e-02
#> [122,] -1.901520e-02 -4.389006e-02  8.502391e-03
#> [123,] -5.261057e-03 -1.584381e-02 -8.119921e-03
#> [124,] -6.383280e-03 -4.123246e-02  1.470308e-02
#> [125,]  1.302685e-02 -1.491246e-02  7.398958e-03
#> [126,] -3.862736e-02  4.716835e-02  4.783583e-02
#> [127,] -4.296679e-02  3.409848e-02 -3.497061e-02
#> [128,]  4.372068e-02  1.067399e-02  4.707591e-02
#> [129,] -4.052221e-03 -3.186797e-02 -2.127369e-02
#> [130,]  4.911200e-02  4.382131e-02  1.983361e-02
#> [131,] -4.570955e-02 -4.732346e-02  3.872776e-02
#> [132,] -1.964631e-02  1.170020e-02  3.216397e-02
#> [133,] -2.746672e-02  1.562213e-03  1.727323e-02
#> [134,]  4.185904e-02  5.055529e-03 -3.242316e-02
#> [135,] -2.249232e-02 -7.871584e-03 -4.605358e-02
#> [136,]  1.639770e-02  3.136875e-02  9.298701e-03
#> [137,]  1.758761e-02  1.336365e-02 -6.624417e-03
#> [138,] -6.155541e-03  3.058177e-02 -2.303775e-02
#> [139,] -3.704825e-02  2.994403e-02 -3.278399e-02
#> [140,] -2.737787e-03 -3.375107e-04  3.586759e-02
#> [141,]  4.180194e-02 -4.568854e-02 -3.335092e-02
#> [142,] -1.783016e-02 -2.113561e-02 -4.164210e-02
#> [143,] -3.379796e-02  5.847620e-03  1.178902e-02
#> [144,]  2.417640e-02 -2.900195e-02 -3.817588e-02
#> [145,]  9.768575e-03  8.136354e-03  5.068860e-03
#> [146,]  3.614771e-02  9.310758e-03  5.418557e-04
#> [147,] -1.837491e-02 -1.015232e-02 -2.638436e-02
#> [148,]  2.611815e-02  4.532068e-02  2.992840e-03
#> [149,] -1.998018e-02  1.109912e-02 -4.379735e-02
#> [150,] -2.024701e-02 -4.258413e-02  1.261036e-02
#> [151,]  2.649941e-03  2.129604e-02  3.888246e-02
#> [152,] -2.194230e-02  7.990350e-03  5.089465e-03
#> [153,]  4.772476e-02 -4.247868e-02  5.430477e-03
#> [154,]  4.437355e-02  1.285577e-02  4.846235e-02
#> [155,] -1.540275e-02  2.970316e-02 -2.648440e-02
#> [156,]  4.233039e-02  8.860278e-04 -6.484847e-04
#> [157,] -4.473026e-02  1.092218e-02 -3.828916e-02
#> [158,] -6.977729e-03 -9.432422e-03 -1.500894e-02
#> [159,] -2.084294e-02  2.562565e-02 -2.123017e-02
#> [160,]  5.512118e-03 -4.452194e-02 -4.117599e-02
#> [161,]  2.276993e-02 -4.229982e-02 -1.389544e-03
#> [162,] -1.871739e-02 -4.171410e-02 -4.974725e-02
#> [163,]  3.504196e-02  1.852533e-02 -3.341643e-02
#> [164,] -4.436111e-02 -3.495846e-02 -3.582812e-02
#> [165,] -3.854164e-02 -4.564808e-03 -2.524288e-02
#> [166,] -3.921128e-02 -1.919149e-02 -4.851032e-02
#> [167,] -6.966321e-03 -3.654839e-02 -4.512681e-02
#> [168,]  4.095344e-02 -9.981636e-03 -3.546060e-02
#> [169,]  2.610865e-02  1.464549e-02  4.121911e-02
#> [170,] -3.710713e-02 -2.891023e-02  2.707564e-02
#> [171,] -2.022473e-02 -3.056576e-02 -4.100202e-02
#> [172,] -3.195373e-02 -3.536555e-02  1.425405e-03
#> [173,] -5.382900e-03 -1.091225e-02 -3.023971e-02
#> [174,] -1.020972e-02  1.542141e-02  7.740475e-03
#> [175,] -3.540508e-03 -1.322659e-02 -4.290614e-02
#> [176,] -3.532034e-02  4.930141e-02 -4.937397e-02
#> [177,] -4.804960e-03  4.924562e-02  4.604663e-03
#> [178,] -2.526271e-02 -7.720514e-03 -3.681332e-02
#> [179,]  4.235090e-02 -4.202032e-02  5.732808e-03
#> [180,]  3.585660e-02 -2.263474e-02  2.162300e-02
#> [181,] -1.181461e-02  3.126249e-02  1.180698e-03
#> [182,] -2.282957e-02 -8.191626e-03  2.212503e-02
#> [183,]  1.334986e-02 -1.705948e-02  4.375009e-02
#> [184,] -2.388796e-02 -8.869767e-03 -1.424972e-02
#> [185,]  4.649789e-02  2.087201e-02  2.310508e-02
#> [186,] -2.190878e-03 -2.762339e-02  1.281107e-02
#> [187,]  1.582505e-03  2.972161e-02 -1.667323e-02
#> [188,]  2.131408e-02 -3.227708e-02  3.863988e-02
#> [189,]  4.399981e-03  2.748716e-02  2.905965e-02
#> [190,] -2.131633e-02 -7.773402e-03 -3.142536e-02
#> [191,] -9.338042e-03 -7.985421e-03  4.060201e-02
#> [192,]  1.832726e-02 -4.922473e-05  4.267584e-02
#> [193,] -4.609976e-02  2.379258e-02  4.266960e-02
#> [194,]  4.960093e-03  2.703206e-02  4.091820e-02
#> [195,] -3.081706e-02  2.507608e-02  3.898391e-02
#> [196,] -1.480377e-02 -4.380573e-02 -4.424041e-03
#> [197,]  1.856049e-02  2.087009e-03  1.100037e-02
#> [198,] -3.663626e-02 -1.378298e-02 -4.091158e-02
#> [199,] -2.829155e-02 -5.703729e-03 -1.316473e-02
#> [200,]  1.502215e-02  4.482685e-03 -1.216111e-02
#> [201,] -4.884982e-02  3.604764e-02 -3.378338e-02
#> [202,] -2.388762e-02 -4.219787e-02 -9.403151e-03
#> [203,] -1.703887e-02  2.644203e-02 -2.069080e-02
#> [204,]  1.839338e-02  4.120643e-02 -4.178408e-02
#> [205,]  4.242527e-02  2.180283e-02  2.376548e-02
#> [206,]  4.701599e-02 -3.711441e-02 -2.743162e-03
#> [207,]  1.237921e-02  2.771799e-02  2.677250e-02
#> [208,] -1.030449e-02  4.642117e-02  2.035060e-02
#> [209,] -3.014730e-02 -2.253875e-02 -1.767049e-02
#> [210,]  3.932330e-02 -4.500128e-02 -1.886774e-02
#> [211,] -1.187100e-02  3.574746e-02  2.503968e-02
#> [212,]  4.405928e-02 -3.661884e-02 -1.449979e-02
#> [213,] -3.818355e-02 -5.883514e-03  1.848628e-02
#> [214,]  4.530209e-03  4.100838e-02  4.957704e-03
#> [215,]  5.035120e-03  2.494344e-02  3.082978e-02
#> [216,]  3.780648e-02 -2.707321e-02 -1.665491e-02
#> [217,]  1.638959e-02 -1.884452e-02 -4.507167e-02
#> [218,] -2.069770e-03  1.890988e-02 -2.898287e-02
#> [219,] -6.613755e-03 -1.594761e-02  1.311545e-02
#> [220,]  2.465702e-03  1.373652e-02 -1.839970e-02
#> [221,] -3.678217e-02 -7.223809e-03  2.867599e-02
#> [222,]  2.914576e-02  7.489637e-04 -2.108199e-02
#> [223,] -2.616702e-03  1.162703e-02  7.245934e-03
#> [224,]  4.682208e-02 -3.568980e-02 -4.886211e-02
#> [225,]  2.815896e-03  2.008323e-02 -4.954357e-02
#> [226,] -4.245028e-02 -1.686014e-03  9.590369e-03
#> [227,] -3.971664e-03 -2.948010e-02 -2.342326e-02
#> [228,]  1.130804e-02 -3.987959e-02 -1.608384e-02
#> [229,] -3.927886e-02  3.106717e-02 -3.234523e-02
#> [230,]  2.705013e-02  4.185940e-02 -1.014159e-02
#> [231,] -3.408329e-02 -4.456875e-02 -1.013016e-02
#> [232,]  2.478185e-02  3.819052e-02 -3.583204e-02
#> [233,] -2.727476e-02  2.152861e-02  3.250030e-02
#> [234,]  3.763288e-02 -3.988933e-02 -4.302959e-03
#> [235,]  3.783229e-02  1.078674e-02 -4.236852e-02
#> [236,]  3.359219e-02 -4.810175e-02 -4.226161e-02
#> [237,]  7.675450e-04  2.038927e-02  4.258053e-02
#> [238,] -2.609848e-02 -2.093066e-03  4.912805e-02
#> [239,]  1.558530e-02  5.625620e-03  1.391565e-02
#> [240,]  7.904729e-03  1.182754e-02 -3.251177e-02
#> [241,]  2.784167e-03  1.420231e-02  2.063219e-02
#> [242,]  3.098927e-02 -2.521595e-02 -2.288509e-02
#> [243,] -4.862907e-02 -4.347875e-02  1.377779e-02
#> [244,]  2.715853e-02  3.608984e-02  2.436158e-02
#> [245,]  1.786937e-02  4.766815e-02 -3.947925e-02
#> [246,] -1.885507e-02 -1.196497e-02  2.656437e-02
#> [247,] -4.143527e-03  3.265071e-03 -2.455074e-02
#> [248,]  4.917720e-02 -4.288611e-02 -4.197362e-03
#> [249,] -1.450641e-02 -4.063235e-02  4.422980e-02
#> [250,]  1.816296e-02  3.558421e-02 -2.323364e-02
#> [251,]  1.477305e-02 -1.291501e-02 -7.057840e-03
#> [252,]  4.864182e-02 -2.092708e-03 -1.052499e-02
#> [253,] -4.775126e-02 -3.143464e-03  1.107906e-02
#> [254,]  4.339868e-02 -3.260895e-02  2.283127e-02
#> [255,]  3.676155e-02  1.343232e-03 -4.009115e-02
#> [256,]  2.293199e-02  1.123712e-02 -1.872719e-02
#> [257,]  4.466721e-02  1.660072e-02 -2.736435e-04
#> [258,] -3.396570e-02  6.197027e-05 -1.370389e-02
#> [259,]  2.151560e-02  2.786381e-02  4.452903e-02
#> [260,]  1.425739e-02  5.189422e-03 -3.099186e-02
#> [261,]  2.034776e-02  6.181410e-03 -1.552135e-02
#> [262,] -4.115935e-02  1.353840e-02 -1.106985e-02
#> [263,]  3.587362e-02 -3.434293e-02  3.881360e-02
#> [264,] -4.560117e-02  1.632359e-03 -2.574894e-02
#> [265,] -3.196337e-02  2.461167e-02 -3.115364e-02
#> [266,]  2.885990e-03  1.868749e-02 -4.557294e-02
#> [267,] -3.842689e-02 -4.435388e-02 -3.796714e-02
#> [268,] -2.889133e-02 -1.144072e-02 -2.126384e-02
#> [269,] -3.534522e-02  1.527425e-02  2.715702e-03
#> [270,]  2.262125e-02  1.854447e-02  4.940489e-02
#> [271,] -4.499565e-02 -3.000983e-02  7.134886e-03
#> [272,]  4.139166e-02 -7.072922e-04 -2.566294e-02
#> [273,]  1.330767e-02  2.682834e-02  2.259698e-03
#> [274,] -3.209451e-02  1.281095e-02  4.709004e-02
#> [275,]  2.956887e-02 -4.723828e-02 -8.078028e-03
#> [276,] -4.824844e-02 -1.675434e-02  1.182078e-04
#> [277,] -3.029433e-02 -3.613299e-02 -1.956226e-02
#> [278,]  1.128091e-02 -4.571636e-03  3.814493e-02
#> [279,]  2.461710e-02  9.791854e-03  2.375841e-02
#> [280,] -1.729320e-02 -4.817020e-02 -3.200577e-02
#> [281,]  4.448131e-02 -4.544395e-02 -1.587258e-02
#> [282,]  2.855272e-02  1.895141e-02 -1.865741e-02
#> [283,] -1.373032e-02  2.536128e-02 -3.084635e-02
#> [284,]  7.853709e-03 -4.886969e-02  8.512360e-03
#> [285,] -2.366332e-02  2.529614e-02 -3.185079e-02
#> [286,] -3.484619e-02 -8.285256e-03 -1.461843e-03
#> [287,]  3.554706e-02  3.213798e-02 -3.360342e-02
#> [288,] -9.335084e-03 -6.048105e-03  1.587659e-02
#> [289,]  3.062414e-02 -4.263925e-02  2.450961e-03
#> [290,] -3.307997e-02  4.487490e-02 -1.444882e-02
#> [291,] -4.032828e-02 -2.789397e-02  3.462528e-02
#> [292,]  2.490283e-02 -1.866193e-02 -2.127793e-02
#> [293,]  8.523030e-03 -4.110925e-02 -4.279755e-03
#> [294,]  4.053561e-02 -3.028193e-03  6.198328e-03
#> [295,] -3.437938e-02  3.456477e-02 -1.665889e-02
#> [296,]  3.671138e-02 -6.589102e-03  2.056259e-02
#> [297,]  4.639248e-02 -1.023920e-02 -3.770201e-02
#> [298,]  1.645124e-02  5.751756e-03 -1.727585e-02
#> [299,] -1.732230e-02  2.485927e-02  2.550048e-03
#> [300,]  2.166417e-02  4.593561e-02 -5.529376e-05
#> [301,]  2.029266e-02  3.487521e-02 -2.631702e-02
#> [302,]  3.260987e-02  1.642178e-03 -3.352288e-02
#> [303,]  1.182291e-02  3.807066e-02  4.734411e-02
#> [304,] -1.463314e-02 -3.252297e-02 -3.176392e-02
#> [305,] -1.767846e-02  3.500738e-02 -3.639882e-02
#> [306,]  1.075657e-02  4.568412e-02 -4.066014e-02
#> [307,] -3.928738e-02 -3.255387e-03 -1.906455e-02
#> [308,] -3.541125e-02 -2.259572e-02  4.035473e-02
#> [309,]  4.985527e-02 -1.534011e-02 -4.271516e-02
#> [310,]  2.731939e-02 -2.728353e-02  3.592010e-02
#> [311,] -1.732747e-02  3.627704e-02 -4.898808e-02
#> [312,] -4.025894e-02 -1.061754e-02  3.049091e-02
#> [313,] -7.029882e-03 -2.241939e-03 -1.751526e-02
#> [314,]  7.403566e-03 -6.638449e-03 -7.552895e-03
#> [315,]  3.722304e-02 -1.613537e-02 -1.123437e-02
#> [316,] -2.441549e-02 -8.572557e-03  1.494616e-02
#> [317,]  2.448337e-02  4.338024e-02  2.653173e-02
#> [318,]  6.084009e-03  2.903153e-02 -2.537598e-02
#> [319,]  3.866092e-03  4.366263e-02 -3.987975e-02
#> [320,] -4.750504e-02 -3.348053e-02  3.213177e-02
#> [321,]  4.547663e-02 -2.608331e-02  1.554278e-02
#> [322,]  3.574544e-02 -2.814322e-02  4.425110e-02
#> [323,] -4.365287e-02  2.647301e-02  1.898923e-02
#> [324,] -4.579859e-03 -7.266336e-03 -3.787728e-02
#> [325,]  2.050217e-02 -4.516555e-02  1.091412e-02
#> [326,] -4.951090e-02 -2.114896e-02 -4.525274e-02
#> [327,] -3.717518e-02 -6.774864e-03  2.675659e-02
#> [328,] -2.589756e-02 -2.744436e-02 -3.728959e-02
#> [329,]  2.557691e-02  2.167935e-02  1.180949e-02
#> [330,]  4.798051e-02 -1.020665e-02  1.307036e-02
#> [331,] -2.853539e-02 -1.335403e-02 -4.229729e-02
#> [332,] -3.937567e-02 -3.439182e-03  3.883527e-02
#> [333,]  1.582387e-03  3.496937e-02 -3.963409e-02
#> [334,] -1.876026e-02  3.188722e-02  3.402146e-02
#> [335,]  2.102924e-02  2.060423e-02 -3.662956e-02
#> [336,]  1.615249e-03 -1.158720e-03 -4.199903e-03
#> [337,]  1.056797e-02 -3.651801e-02  1.137668e-02
#> [338,] -1.525636e-02 -6.299387e-03  2.372651e-02
#> [339,]  4.810808e-02  1.073806e-02 -2.254642e-02
#> [340,]  1.724775e-02  1.715036e-02 -2.248460e-02
#> [341,] -1.847390e-03 -3.577608e-02 -3.128070e-02
#> [342,]  5.665619e-04 -3.816478e-02  5.454054e-04
#> [343,]  5.052608e-03  2.883960e-02 -1.454746e-02
#> [344,]  3.977282e-02 -4.005051e-02 -1.260494e-02
#> [345,] -3.215835e-02  1.177821e-02  2.663990e-03
#> [346,]  7.491998e-04 -1.587932e-02  2.613329e-02
#> [347,] -4.575949e-02  4.207828e-02  3.856339e-02
#> [348,]  1.186934e-02 -1.622793e-02 -1.269780e-03
#> [349,]  4.040296e-02  1.612619e-02 -5.779088e-03
#> [350,]  5.508977e-03 -2.470288e-02  2.624131e-02
#> [351,] -3.054242e-05 -3.713111e-02  3.018606e-02
#> [352,] -2.138218e-02  3.866756e-02  1.924001e-02
#> [353,]  6.579628e-03  3.347056e-02  1.202980e-02
#> [354,] -1.061075e-02  1.325918e-03 -1.455381e-02
#> [355,] -2.325734e-03 -4.065287e-02 -1.713367e-02
#> [356,] -1.855520e-02 -1.453756e-02  4.694662e-02
#> [357,]  1.495453e-02 -3.704367e-02  4.199889e-02
#> [358,]  4.644040e-02  3.508030e-02  3.576476e-02
#> [359,] -2.858782e-03 -4.250741e-02  4.407493e-02
#> [360,]  2.577911e-02 -3.073339e-02 -8.582146e-03
#> [361,]  1.242353e-02  2.325568e-02 -4.365471e-03
#> [362,] -2.189300e-02 -1.648876e-02  4.769539e-03
#> [363,] -1.352081e-02  9.168390e-03 -4.007377e-02
#> [364,]  2.502761e-02 -2.038251e-04  1.561628e-03
#> [365,] -2.843135e-02 -2.261575e-02 -6.880374e-03
#> [366,] -3.455203e-02 -1.635096e-02  2.151551e-03
#> [367,] -2.937913e-02  2.113843e-02  3.185524e-03
#> [368,]  4.585348e-03  3.917336e-02  8.232308e-03
#> [369,]  1.447537e-02  4.087960e-02  3.845836e-02
#> [370,] -4.331193e-02  2.159362e-02  1.742698e-02
#> [371,] -2.860388e-02 -3.862429e-02  4.211198e-02
#> [372,]  3.834314e-02  3.984587e-02 -3.831749e-02
#> [373,] -1.351494e-02 -4.023831e-02  5.544495e-03
#> [374,] -7.631979e-03  3.678611e-02  1.990505e-02
#> [375,] -4.673404e-03  2.140101e-02  3.811369e-02
#> [376,]  4.429489e-03  2.975745e-02 -2.907702e-02
#> [377,] -2.630584e-02  4.033101e-02  2.591043e-02
#> [378,] -2.359990e-02 -2.800371e-02  3.950066e-02
#> [379,] -3.421408e-02  3.799038e-02  2.514272e-02
#> [380,] -1.094985e-02  4.289779e-02 -4.880420e-03
#> [381,]  4.909621e-02  2.482924e-02  3.650728e-02
#> [382,] -1.235896e-02 -1.198488e-02  2.881561e-02
#> [383,]  3.169943e-02 -8.853319e-04  1.974665e-02
#> [384,] -2.495931e-02 -4.684870e-02  3.732917e-02
#> [385,]  4.514168e-02  3.388019e-02  1.536821e-02
#> [386,]  2.614879e-02  1.895642e-02 -2.261426e-02
#> [387,] -3.759031e-02 -3.561941e-02  4.788671e-03
#> [388,] -1.096353e-02 -1.827350e-02 -1.125323e-02
#> [389,]  2.866972e-02 -2.329762e-02 -4.053687e-02
#> [390,] -3.160759e-02  3.028335e-02  2.860664e-02