Generate points sampled from the Dini surface with optional noise.

This function generates points sampled from the Dini surface along with optional noise.

Usage

dini_surface_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

dini_surface_3d(n = 100, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>                 [,1]         [,2]        [,3]          [,4]         [,5]
#>   [1,]  0.2025669690  0.382692011  6.78893642  0.0289538688 -0.048208714
#>   [2,]  0.7833219149  0.617479639  6.95061832  0.0249310469 -0.036588946
#>   [3,]  0.4917925145  0.870048112  7.33949439  0.0250675706  0.035378303
#>   [4,]  0.7824550738  0.059007258 -0.02967529 -0.0187072770 -0.041413508
#>   [5,] -0.2453530867 -0.968543798 10.74744794 -0.0328095513  0.013023477
#>   [6,]  0.7850394826 -0.605702838  5.62529406  0.0371262625 -0.025296864
#>   [7,]  0.3907826417 -0.380566168 11.41749642 -0.0151232299 -0.036826613
#>   [8,]  0.3200622743  0.595930720  7.15482985  0.0259679128 -0.041953378
#>   [9,] -0.3934691958 -0.118203444  9.09077477 -0.0260726815  0.047345051
#>  [10,] -0.1161992444  0.500528778  7.65469689  0.0315386858 -0.040925190
#>  [11,]  0.8552846196  0.315533519  6.61080159 -0.0257912662 -0.027928377
#>  [12,]  0.1432528680 -0.096531988 10.51681521  0.0207393543  0.036522632
#>  [13,] -0.3885714467  0.883651865  1.99126832 -0.0200284075  0.034622799
#>  [14,] -0.9869647617  0.101911968  3.03934956 -0.0339800407 -0.047521955
#>  [15,]  0.4422362973  0.881796311  7.39061242  0.0126332500 -0.008645263
#>  [16,]  0.8420793170  0.386975387  6.69460654 -0.0002564987 -0.021950474
#>  [17,]  0.3597568058 -0.675689635 11.36424323 -0.0267581337 -0.005070444
#>  [18,]  0.0750447666 -0.912719326 11.05364871 -0.0258414693  0.003983883
#>  [19,]  0.0434930207  0.418091664  0.86211026  0.0310556170 -0.009466871
#>  [20,] -0.8298497481  0.273839898  2.77799400  0.0188306865  0.013397903
#>  [21,] -0.2441340767 -0.408484860  3.67960219  0.0112175181 -0.035932845
#>  [22,]  0.2615008684 -0.344202260 11.06554291 -0.0040347483 -0.017732823
#>  [23,]  0.5081023277  0.130598508 -0.15806990 -0.0098363018 -0.046313200
#>  [24,]  0.8524711615  0.498834705  0.53073256  0.0078864069 -0.041911192
#>  [25,] -0.5787600997 -0.398398757  3.56569285 -0.0141214118 -0.038217667
#>  [26,] -0.1247561455  0.138639860  0.92236811 -0.0405885955 -0.006629414
#>  [27,]  0.3948602655 -0.360343932 11.43292632 -0.0108659976  0.003698734
#>  [28,]  0.1604138659  0.986787600  1.40964039  0.0282023039  0.038469664
#>  [29,]  0.7475160194 -0.661078149  5.55916574  0.0471984961 -0.033575053
#>  [30,]  0.2194133018 -0.953627422  4.93554380 -0.0295681397  0.011001009
#>  [31,]  0.0959584208 -0.377093715 10.56911000  0.0205608217  0.003114196
#>  [32,]  0.8818996099  0.449800838  0.47069499  0.0402536208 -0.043724129
#>  [33,]  0.7749215058 -0.274214476 12.14898925 -0.0286334296  0.024198333
#>  [34,] -0.9329405171  0.227129447  9.19359488 -0.0211556719 -0.013565047
#>  [35,] -0.4764686082  0.183499543  8.62430222 -0.0110196688  0.047533788
#>  [36,]  0.1133811329  0.215962837 -0.03171646 -0.0239600501 -0.043491758
#>  [37,]  0.0430668300  0.958235851  1.51796697 -0.0244596597 -0.036100640
#>  [38,] -0.9052807510 -0.422195996  9.86119652  0.0374066290  0.015842541
#>  [39,]  0.0787282784 -0.126403872  3.66635001 -0.0013236231 -0.027295435
#>  [40,]  0.5965460537  0.688531213  7.16613666 -0.0129789079 -0.029380586
#>  [41,]  0.1246888948 -0.013636026  4.40118040  0.0004161482 -0.049538155
#>  [42,] -0.8956663503  0.441770625  8.96660479 -0.0417437030  0.027833847
#>  [43,]  0.1271428581 -0.305185203  4.27945585 -0.0213083774 -0.013198720
#>  [44,]  0.8961511661  0.248184719  6.57144419 -0.0170261841  0.019999169
#>  [45,] -0.9734219354  0.050176962  3.08625652 -0.0185976432 -0.017755566
#>  [46,] -0.3255177628  0.205420731  1.89300612  0.0180423740  0.023764167
#>  [47,] -0.8226525727 -0.425851411  3.63877168 -0.0120025624  0.049152932
#>  [48,]  0.8968734945 -0.432215202  5.83383452  0.0221162780  0.040687143
#>  [49,]  0.1173465170  0.293072834  0.31857490 -0.0152938433  0.019915336
#>  [50,]  0.5132193143  0.702411657  7.17590506 -0.0418044343  0.021595296
#>  [51,] -0.6343866494  0.714263624  2.28794457 -0.0445466048  0.001475257
#>  [52,]  0.2687440960  0.554556067  0.84212855  0.0329350983  0.014334228
#>  [53,]  0.0006634475 -0.300553205  3.79647498  0.0070030334 -0.044913175
#>  [54,] -0.8430402956 -0.495112206  3.67579336 -0.0395700088  0.025478976
#>  [55,]  0.4898428256  0.521338755  6.93361161  0.0113723217 -0.018850810
#>  [56,] -0.4044278959  0.807258946  2.00537218  0.0027182969  0.042455095
#>  [57,] -0.9714809271  0.022692908  9.40595877 -0.0498214228  0.031537010
#>  [58,] -0.8165472896  0.538639112  2.55539615  0.0003180930  0.007971454
#>  [59,] -0.2173297149  0.007652705  8.15881912  0.0446259994  0.025873130
#>  [60,]  0.9345212080  0.355905776  0.36388336 -0.0371489430  0.000538149
#>  [61,]  0.0055285708 -0.021140557  7.73460874  0.0337070487 -0.049052259
#>  [62,] -0.5348255479 -0.832248131  4.14226550  0.0066368950  0.006429928
#>  [63,]  0.9024130282 -0.430811354 12.12096635  0.0195208239 -0.003842061
#>  [64,] -0.8471073158 -0.298109281  3.44775750  0.0066462308 -0.020392211
#>  [65,] -0.5849463624 -0.594339616 10.14892993  0.0231982947  0.011250298
#>  [66,]  0.1389908661  0.182747778  6.02591556 -0.0458731051 -0.026048588
#>  [67,]  0.9706178071 -0.100180047 12.46711659  0.0372809324  0.039798548
#>  [68,]  0.5758632161  0.470055476  0.54472755 -0.0278635119  0.029293599
#>  [69,] -0.7133260219 -0.351698622  9.78615296  0.0204128070 -0.046374129
#>  [70,] -0.2987136646 -0.242404173  3.13713076 -0.0168384430  0.010126506
#>  [71,] -0.0386825592  0.197796832  0.45882321 -0.0307768832 -0.020115174
#>  [72,]  0.5644653317  0.206644552  6.33697515  0.0043486173 -0.048476763
#>  [73,]  0.9278000473 -0.350165737  5.92158077 -0.0399409109  0.022483305
#>  [74,]  0.9482959339  0.096774658  6.37513798  0.0060741224  0.003782398
#>  [75,]  0.8372440006 -0.545721584  5.70555481  0.0294829949  0.049708932
#>  [76,]  0.0418813272  0.281927224  0.45446049  0.0302195185 -0.026583413
#>  [77,] -0.0116060501  0.102046464 -0.28764592 -0.0294904219  0.043673694
#>  [78,] -0.5122585443 -0.215180990  9.46112228 -0.0095565879  0.019950624
#>  [79,]  0.0931790606  0.900996512  7.72243685  0.0122791571  0.047473604
#>  [80,]  0.9813338797  0.157706326  0.15979126  0.0273459295 -0.016258715
#>  [81,]  0.2339299768 -0.971658315  4.94863302 -0.0085205612 -0.024696356
#>  [82,] -0.2416496259 -0.130127096  8.91362987  0.0158290173 -0.046178393
#>  [83,] -0.1463198513  0.986548299  1.71790839  0.0133208312 -0.049280752
#>  [84,] -0.4638167022  0.873822730  2.05772151 -0.0094936829 -0.042501217
#>  [85,]  0.3285726683 -0.918573851  5.05226562  0.0416282046 -0.038749807
#>  [86,]  0.9825490328  0.036259946  6.32213790  0.0373934465  0.030311456
#>  [87,] -0.5645451322 -0.791979489  4.09743626 -0.0433607815  0.015231343
#>  [88,] -0.0159512492  0.047683473 -0.79029420  0.0136171688 -0.020498237
#>  [89,]  0.2044677441 -0.189853064 10.82873622 -0.0278542275  0.008382909
#>  [90,] -0.4536517319  0.280170681  8.47563030  0.0477666195 -0.012574826
#>  [91,] -0.8091154594  0.512057710  2.56894729 -0.0038518402 -0.038457466
#>  [92,] -0.0590776939 -0.276667386  9.80919329  0.0027436481 -0.002317173
#>  [93,]  0.5996293603  0.750896524  0.88952906  0.0439309014 -0.019904646
#>  [94,]  0.3350687497 -0.938001460  5.05571180 -0.0217186566 -0.038748971
#>  [95,]  0.6933509025  0.651209764  7.04764166 -0.0338975338 -0.003963952
#>  [96,]  0.5051256335 -0.706300081 11.56843218 -0.0266213184 -0.041294155
#>  [97,]  0.8044022693 -0.335456533 12.12511448  0.0311338457  0.006507690
#>  [98,]  0.3407668010 -0.641115363  5.04510969  0.0152318791  0.049079527
#>  [99,] -0.6147913066  0.447810674  2.38732778 -0.0367786056 -0.046282508
#> [100,]  0.1642238702  0.802210993  7.57258426  0.0408764165 -0.029677003
#>                 [,6]          [,7]
#>   [1,]  1.999831e-02  2.235994e-02
#>   [2,]  3.582538e-02  6.616277e-03
#>   [3,]  4.051422e-02  4.438931e-03
#>   [4,] -4.454861e-02 -2.020315e-02
#>   [5,]  6.443084e-03 -2.641509e-02
#>   [6,]  2.358491e-02 -2.135582e-03
#>   [7,]  4.285176e-02  2.580471e-02
#>   [8,] -4.891276e-02  2.972650e-02
#>   [9,]  3.034251e-02  4.120593e-02
#>  [10,] -7.144477e-03  4.837785e-02
#>  [11,] -2.279981e-02  3.714580e-02
#>  [12,]  3.572393e-02  4.906680e-03
#>  [13,] -1.564727e-02  3.666606e-02
#>  [14,] -1.689092e-02  9.038719e-03
#>  [15,] -1.023518e-02 -4.051108e-03
#>  [16,]  3.138686e-03  2.555157e-02
#>  [17,] -3.051500e-02 -1.808335e-02
#>  [18,]  4.784915e-02 -1.544150e-02
#>  [19,] -4.434506e-02 -4.807873e-02
#>  [20,] -7.183638e-03 -2.923316e-02
#>  [21,] -4.375579e-02  3.540885e-02
#>  [22,] -2.133556e-03 -4.662534e-02
#>  [23,] -2.504063e-05  9.025088e-03
#>  [24,] -4.525426e-02 -3.263804e-02
#>  [25,]  1.010355e-02 -4.559712e-02
#>  [26,]  2.764944e-02  2.500738e-03
#>  [27,] -6.511780e-03 -3.763024e-02
#>  [28,]  2.703825e-02  4.264267e-03
#>  [29,]  6.209660e-03  9.977285e-03
#>  [30,]  2.675452e-02  2.866356e-02
#>  [31,]  9.612660e-03  1.776844e-02
#>  [32,] -2.349478e-02 -2.814070e-02
#>  [33,] -3.209831e-02  9.476694e-03
#>  [34,] -1.897569e-02 -4.051988e-02
#>  [35,]  4.812319e-02  1.080925e-02
#>  [36,]  2.876011e-02  9.105336e-04
#>  [37,] -4.257016e-02  2.102593e-02
#>  [38,]  2.974453e-02 -4.650655e-02
#>  [39,]  8.339206e-03  3.685860e-02
#>  [40,] -3.183314e-02 -2.462998e-02
#>  [41,] -2.957064e-02  2.214194e-02
#>  [42,] -7.037850e-03 -3.582744e-02
#>  [43,]  1.081558e-02  4.926137e-02
#>  [44,]  4.133719e-02  2.405475e-02
#>  [45,] -9.529967e-03 -3.365858e-02
#>  [46,] -4.801223e-02  1.552200e-02
#>  [47,] -2.185665e-02 -9.119884e-03
#>  [48,]  1.808537e-02  5.028880e-03
#>  [49,] -3.769387e-02  1.685119e-05
#>  [50,]  2.598746e-02 -2.000412e-02
#>  [51,] -3.237856e-02 -2.470780e-02
#>  [52,]  2.811257e-02  4.537560e-02
#>  [53,]  2.239914e-03 -4.620594e-03
#>  [54,]  1.616154e-02 -9.118684e-03
#>  [55,]  4.070864e-02 -7.568073e-03
#>  [56,] -3.444297e-02 -1.288406e-02
#>  [57,]  4.220291e-04  1.473192e-02
#>  [58,] -1.403511e-02  7.183425e-03
#>  [59,] -3.108639e-02 -2.843822e-02
#>  [60,] -4.519986e-02  1.073646e-02
#>  [61,]  4.962913e-02 -2.450281e-02
#>  [62,] -1.237712e-02 -1.292449e-02
#>  [63,] -2.870431e-02  3.899559e-02
#>  [64,]  4.391893e-02 -1.993900e-02
#>  [65,]  1.965875e-02  3.378709e-02
#>  [66,] -7.865232e-03  1.626223e-02
#>  [67,]  4.988115e-02  4.717706e-02
#>  [68,]  3.832464e-04  3.250426e-03
#>  [69,] -2.814099e-02 -2.003141e-02
#>  [70,] -2.073059e-02  4.206260e-02
#>  [71,]  4.427533e-02  2.163429e-02
#>  [72,]  3.424311e-02 -3.467058e-02
#>  [73,] -2.530377e-02  1.766185e-02
#>  [74,]  7.093283e-03 -3.786482e-02
#>  [75,]  1.296294e-03  1.167185e-02
#>  [76,] -8.648069e-03 -3.136679e-02
#>  [77,] -3.778791e-02 -3.313074e-02
#>  [78,] -1.833535e-02 -6.848199e-03
#>  [79,] -7.361179e-03  1.165584e-02
#>  [80,] -1.852726e-02  2.060201e-04
#>  [81,]  3.670981e-02  2.162830e-03
#>  [82,]  2.866782e-02  4.878575e-02
#>  [83,] -2.894180e-02 -4.571071e-02
#>  [84,]  4.278989e-02 -4.356302e-02
#>  [85,] -8.257557e-03  4.748533e-02
#>  [86,]  1.630694e-02 -2.342946e-02
#>  [87,] -1.377157e-02 -4.068626e-02
#>  [88,] -3.526644e-02 -6.861787e-03
#>  [89,]  2.046354e-02 -4.403167e-02
#>  [90,] -3.436556e-02 -4.679952e-03
#>  [91,]  1.759290e-02  4.289359e-02
#>  [92,]  6.553589e-03 -1.799178e-02
#>  [93,]  4.772930e-02 -1.486225e-03
#>  [94,] -4.866735e-02 -2.458046e-02
#>  [95,] -4.015985e-02  4.295555e-03
#>  [96,] -1.480816e-02 -2.687500e-02
#>  [97,] -2.757303e-02  2.814726e-02
#>  [98,] -3.028548e-02 -1.551136e-02
#>  [99,] -4.294311e-03 -3.214686e-02
#> [100,]  2.849970e-02  1.880855e-02