Generate points sampled from the Dini surface with optional noise.

This function generates points sampled from the Dini surface along with optional noise.

Usage

dini_surface_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

dini_surface_with_noise(sample_size = 100, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>                [,1]        [,2]        [,3]          [,4]          [,5]
#>   [1,]  0.699092962 -0.65511203  5.52200160 -0.0162277209  0.0060418914
#>   [2,] -0.374434107  0.91808272  8.24197273  0.0484597159 -0.0191798187
#>   [3,]  0.284601309  0.50737717  0.73606615 -0.0459075569  0.0050158339
#>   [4,] -0.342721855 -0.42163785  9.93307435 -0.0334925848  0.0389458635
#>   [5,] -0.793895839 -0.16749763  3.26473933  0.0377066206 -0.0143328433
#>   [6,]  0.257649608  0.28818288  0.15958364 -0.0395732186 -0.0402970245
#>   [7,] -0.240195496  0.96504863  8.09753330  0.0337654296 -0.0216920764
#>   [8,]  0.573188948  0.72570361  7.16531677  0.0218800192 -0.0335653397
#>   [9,] -0.109793317  0.97563431  7.96838098 -0.0388236875 -0.0224314824
#>  [10,] -0.846449722  0.35284705  9.05324265 -0.0199245394 -0.0214177442
#>  [11,]  0.874189549  0.12307149  0.09987056  0.0412999706 -0.0124897638
#>  [12,] -0.990201701  0.10584151  9.31854666 -0.0067402777 -0.0479673292
#>  [13,] -0.651263166  0.57808605  2.36908064  0.0006591435  0.0094347040
#>  [14,]  0.237777619  0.04682834 -0.93087973 -0.0398559760  0.0412415796
#>  [15,] -0.185972931 -0.10666646  8.70087148 -0.0198713040  0.0087055276
#>  [16,]  0.064883143 -0.07731992  9.70477613 -0.0141472039 -0.0092912858
#>  [17,] -0.100763165 -0.71571161  4.41357271  0.0352118721  0.0096569608
#>  [18,]  0.051605453  0.05656992  4.85030678 -0.0079735630 -0.0032210261
#>  [19,] -0.972407882  0.21130345  2.92794391 -0.0272157980  0.0109461431
#>  [20,]  0.044675975 -0.09287627  9.47570564  0.0465477508  0.0049074243
#>  [21,]  0.026058834 -0.72529191 10.87538600  0.0281944202  0.0441229291
#>  [22,] -0.317135291 -0.07178834  8.80400866  0.0079305935  0.0046589536
#>  [23,] -0.356668717  0.47163331  8.19113792  0.0332555604 -0.0223319527
#>  [24,]  0.744597364  0.65007916  0.71890777 -0.0172750928  0.0057234509
#>  [25,] -0.016187285 -0.31831298 10.08245847  0.0227653674  0.0245447804
#>  [26,] -0.440732527 -0.82680298  4.23802172  0.0063066037  0.0020481335
#>  [27,]  0.001930333  0.32416982  0.71859420  0.0139974989 -0.0447922603
#>  [28,]  0.278671024 -0.16858562  4.89730581 -0.0178559134 -0.0302934072
#>  [29,]  0.095674337 -0.70259139  4.67542660 -0.0006271232  0.0415414709
#>  [30,] -0.550589899 -0.82984035  4.12634010  0.0100844382 -0.0057582209
#>  [31,]  0.724662242  0.68794714  0.75943425  0.0107934259  0.0170204326
#>  [32,] -0.844896490  0.51623546  2.59219175 -0.0089848436 -0.0341853175
#>  [33,] -0.238598519 -0.20588472  2.98051269  0.0471387753  0.0140931498
#>  [34,] -0.069890699 -0.06031583  1.77579628  0.0110489386 -0.0086917707
#>  [35,] -0.669495583 -0.44691172  3.64073412  0.0244097535  0.0001630595
#>  [36,] -0.702972227  0.30243905  8.89771460 -0.0419164541 -0.0093407636
#>  [37,] -0.589923098  0.17831851  2.57069678  0.0323652630  0.0404280977
#>  [38,]  0.889218026 -0.31690674 12.21119508  0.0169541707 -0.0093717973
#>  [39,] -0.652541208 -0.70045794 10.25405567  0.0285713033  0.0424148160
#>  [40,] -0.756748858  0.06891379  9.20861796 -0.0476668029  0.0063516749
#>  [41,]  0.079503494  0.28879850  6.66400655  0.0117054416  0.0126243548
#>  [42,] -0.583015858  0.78103585  2.21588277  0.0282385771 -0.0183351835
#>  [43,] -0.103940983  0.43013108  7.53241120  0.0237279037 -0.0445817056
#>  [44,] -0.002445968 -0.15540866  9.41905835  0.0061557579 -0.0422443861
#>  [45,] -0.285270863  0.60380303  8.07955104  0.0282824257  0.0146294602
#>  [46,] -0.757160494 -0.65202006 10.13568216 -0.0181531919  0.0074056810
#>  [47,]  0.329004284 -0.08902195  5.21994886  0.0320352582 -0.0314484064
#>  [48,] -0.028480790  0.70059691  1.43117823 -0.0024434826  0.0337394231
#>  [49,]  0.084103028 -0.98449656 11.07956158 -0.0462410185  0.0498855758
#>  [50,]  0.745031839 -0.26651772  5.83984774  0.0249521357 -0.0193995566
#>  [51,]  0.147321232  0.93700942  1.40357118  0.0325814311  0.0202686697
#>  [52,]  0.461427007 -0.61172235 11.52186623 -0.0285212245  0.0132212917
#>  [53,]  0.153748084  0.45119844  7.03294269 -0.0451294365 -0.0398065699
#>  [54,] -0.437557985 -0.14917710  9.23366961  0.0203104993 -0.0225772933
#>  [55,] -0.408625640  0.89431386  8.28463159  0.0239448926 -0.0396199872
#>  [56,]  0.946335848 -0.08754474 12.48478701 -0.0196615033  0.0460981871
#>  [57,] -0.480213247  0.86955579  2.07587980  0.0245240715 -0.0319413923
#>  [58,]  0.947587554  0.28377846  0.28990658  0.0170367325 -0.0451131897
#>  [59,] -0.008460198  0.91179440  7.83753545  0.0193450200  0.0097942993
#>  [60,]  0.281277560 -0.95865905 11.28094513 -0.0439215184  0.0116099892
#>  [61,]  0.068396612 -0.13988858  9.89137142 -0.0409448875  0.0496068916
#>  [62,]  0.550550617 -0.64937416  5.35754833  0.0400419420 -0.0014850126
#>  [63,] -0.252963865 -0.94935914  4.45418507 -0.0370169153 -0.0052995480
#>  [64,] -0.080699625  0.02695708  0.66011739 -0.0176829793 -0.0481359621
#>  [65,] -0.183892780 -0.19704598  9.22211918  0.0186770425  0.0178595704
#>  [66,]  0.883632052  0.42352906  0.44966835 -0.0321597103 -0.0409791842
#>  [67,] -0.301979112 -0.95164881  4.40517685  0.0087555014 -0.0067884965
#>  [68,]  0.497305309 -0.05657797 12.00304469  0.0182474675 -0.0159242862
#>  [69,]  0.292207822 -0.66393162 11.25372703 -0.0049072516 -0.0269608588
#>  [70,] -0.920282193 -0.38333442  3.53643670  0.0331201559 -0.0299704217
#>  [71,] -0.458017227  0.49169950  2.10956677  0.0164907773 -0.0348179629
#>  [72,] -0.062652321 -0.07737214  2.02908458  0.0100751778 -0.0338285447
#>  [73,] -0.586891360  0.49714557  2.32126755  0.0261379774 -0.0419435456
#>  [74,] -0.921438405  0.37731172  2.75267301  0.0421313571 -0.0461354145
#>  [75,] -0.785765768 -0.18197158  3.28088383  0.0452762598  0.0168495948
#>  [76,]  0.459414265  0.06622420  5.91025277 -0.0208375141 -0.0001259127
#>  [77,]  0.294732107  0.95519677  7.55468690  0.0154112210  0.0156917039
#>  [78,] -0.409489320 -0.26477768  9.52605675  0.0491431011  0.0095616273
#>  [79,] -0.712383673 -0.59879075 10.14155462  0.0331097768  0.0099626242
#>  [80,] -0.272269823 -0.31922841  3.39949810  0.0032137258  0.0103170313
#>  [81,]  0.616064492 -0.65781710  5.43437268 -0.0241745994 -0.0041123882
#>  [82,]  0.568862423  0.39552462  0.41874747 -0.0108795706 -0.0461574898
#>  [83,] -0.426760102 -0.07092872  2.72727555 -0.0033740966  0.0284061805
#>  [84,]  0.779403480 -0.14930854 12.27907704  0.0047070357  0.0314620565
#>  [85,] -0.422903923  0.59782616  2.03649994  0.0089358447  0.0183065254
#>  [86,]  0.274814052  0.52454289  0.77904671 -0.0472702682 -0.0411002299
#>  [87,] -0.034075885  0.88885597  1.57262330  0.0126192685 -0.0477070170
#>  [88,] -0.479250683 -0.39484553  9.84248861 -0.0387640618 -0.0026511545
#>  [89,] -0.104943849 -0.24900736  9.57661742 -0.0300083662  0.0040045442
#>  [90,]  0.821032715  0.52787623  0.56787309 -0.0129118889  0.0109562132
#>  [91,]  0.105181481  0.33546232  6.78024288  0.0322294819  0.0153415834
#>  [92,]  0.102298018 -0.01715647  4.15518468 -0.0168839773  0.0174701914
#>  [93,]  0.330106101  0.56668589  7.09714326 -0.0162145063  0.0385697496
#>  [94,] -0.511198018  0.65435170  2.16241104  0.0001356485 -0.0460913025
#>  [95,]  0.469771167  0.28551312  0.17589639 -0.0229241776  0.0423819362
#>  [96,] -0.340508934  0.73612388  1.91904580 -0.0169863004  0.0356609766
#>  [97,] -0.639919866  0.35976883  2.48115025 -0.0409264283  0.0377503688
#>  [98,]  0.705833531 -0.65938331 11.82101011 -0.0099313898  0.0040915332
#>  [99,] -0.914458598  0.27848112  2.83708067 -0.0190295386 -0.0420256294
#> [100,]  0.914092379 -0.01768551 12.52238624 -0.0226834745  0.0407121711
#>                 [,6]          [,7]
#>   [1,] -0.0247835588 -0.0179297998
#>   [2,] -0.0291449046 -0.0186733560
#>   [3,] -0.0419486718  0.0484900970
#>   [4,] -0.0259754745  0.0071866210
#>   [5,]  0.0417307753 -0.0122555044
#>   [6,]  0.0245025961  0.0185985072
#>   [7,] -0.0462076291  0.0099559400
#>   [8,]  0.0461586950 -0.0121719220
#>   [9,]  0.0031082865 -0.0412589109
#>  [10,]  0.0399917101  0.0307866731
#>  [11,] -0.0342379403  0.0473428499
#>  [12,] -0.0460008409 -0.0290710928
#>  [13,]  0.0233293395  0.0358799386
#>  [14,] -0.0074213206 -0.0438829132
#>  [15,] -0.0117619061  0.0097305555
#>  [16,]  0.0334684087 -0.0394715721
#>  [17,] -0.0293025211  0.0381387482
#>  [18,] -0.0015684485 -0.0477320825
#>  [19,] -0.0290819821 -0.0432082476
#>  [20,]  0.0322903698  0.0472968488
#>  [21,]  0.0170338519 -0.0017855093
#>  [22,] -0.0089122050  0.0340600477
#>  [23,] -0.0044701922 -0.0199805643
#>  [24,]  0.0228978785  0.0415052952
#>  [25,]  0.0015217240 -0.0278199562
#>  [26,] -0.0087552177  0.0467150789
#>  [27,]  0.0315082168 -0.0322540076
#>  [28,] -0.0382459324 -0.0416060831
#>  [29,]  0.0135661706  0.0202375573
#>  [30,] -0.0065753211 -0.0170134119
#>  [31,] -0.0354548863 -0.0283356632
#>  [32,] -0.0023627899 -0.0035127251
#>  [33,]  0.0406844551 -0.0026691965
#>  [34,] -0.0202380276  0.0260852291
#>  [35,]  0.0129092838  0.0388146388
#>  [36,]  0.0417667701  0.0339427894
#>  [37,]  0.0391168787 -0.0153835058
#>  [38,]  0.0272624169  0.0492188953
#>  [39,]  0.0355979941 -0.0309391920
#>  [40,]  0.0005671144  0.0429588423
#>  [41,]  0.0449339391 -0.0149408118
#>  [42,] -0.0016237431 -0.0178790121
#>  [43,]  0.0122192391 -0.0328574287
#>  [44,]  0.0106427161  0.0145855378
#>  [45,] -0.0017422932  0.0310074142
#>  [46,] -0.0282969272 -0.0085242157
#>  [47,] -0.0205847530 -0.0368709176
#>  [48,]  0.0067314286 -0.0063704733
#>  [49,]  0.0018338934  0.0485404771
#>  [50,]  0.0082189079 -0.0350358319
#>  [51,]  0.0003823774 -0.0328537597
#>  [52,]  0.0483877727 -0.0484171387
#>  [53,] -0.0052650960  0.0245394473
#>  [54,] -0.0082508259  0.0035061127
#>  [55,] -0.0431251653 -0.0411622626
#>  [56,] -0.0243078477 -0.0248105018
#>  [57,]  0.0444022915  0.0058300460
#>  [58,] -0.0204011699 -0.0286415770
#>  [59,] -0.0277264541  0.0096196144
#>  [60,] -0.0039783272  0.0482262683
#>  [61,]  0.0190108571 -0.0343061295
#>  [62,]  0.0328069537 -0.0130194174
#>  [63,] -0.0049575569 -0.0130231560
#>  [64,] -0.0205270803  0.0277021285
#>  [65,]  0.0201687511  0.0171359950
#>  [66,] -0.0294866548 -0.0410016794
#>  [67,]  0.0274991157 -0.0436537053
#>  [68,] -0.0427972180 -0.0208453951
#>  [69,] -0.0233123783 -0.0166789675
#>  [70,]  0.0301500687  0.0467434797
#>  [71,]  0.0215750773  0.0209211191
#>  [72,]  0.0493816933 -0.0007117688
#>  [73,]  0.0184374433  0.0029358712
#>  [74,]  0.0385404728  0.0076141684
#>  [75,] -0.0096014414  0.0303298179
#>  [76,] -0.0453389097 -0.0243408122
#>  [77,]  0.0216569343  0.0453203099
#>  [78,]  0.0081911560  0.0371353919
#>  [79,] -0.0085186577  0.0335493435
#>  [80,]  0.0241911409  0.0247125781
#>  [81,]  0.0293384820 -0.0191103565
#>  [82,] -0.0017504400  0.0361984506
#>  [83,]  0.0410792979 -0.0015809973
#>  [84,] -0.0126740734 -0.0071589286
#>  [85,]  0.0131146887  0.0150105994
#>  [86,]  0.0281264512 -0.0422230177
#>  [87,]  0.0116592989 -0.0320363989
#>  [88,]  0.0424530120  0.0003174423
#>  [89,]  0.0149552828 -0.0067882733
#>  [90,] -0.0150328351 -0.0409338249
#>  [91,] -0.0432485052 -0.0448285777
#>  [92,] -0.0284293052  0.0145013161
#>  [93,] -0.0090045543 -0.0063322747
#>  [94,]  0.0367839109  0.0086621583
#>  [95,]  0.0471732917  0.0208860335
#>  [96,] -0.0248816428 -0.0066309534
#>  [97,]  0.0302545738  0.0104186096
#>  [98,]  0.0194909660  0.0498341559
#>  [99,] -0.0080567336 -0.0078010959
#> [100,] -0.0276187985  0.0133110690