Generate synthetic data with Gaussian clusters

Generate Gaussian Clusters

Usage

gau_clust(
  n,
  num_clust,
  mean_matrix,
  var_vec,
  num_dims,
  num_noise,
  min_n,
  max_n
)

Arguments

n

The total number of data points to be generated.

num_clust

The number of clusters to generate.

mean_matrix

A matrix where each row represents the mean vector for a cluster.

var_vec

A vector specifying the variance for each cluster.

num_dims

The number of effective dimensions for the data points.

num_noise

The number of additional noise dimensions to be generated.

min_n

The minimum value for the noise added to the data points.

max_n

The maximum value for the noise added to the data points.

Value

A matrix containing the generated Gaussian clusters.

Details

This function generates Gaussian clusters with specified parameters.

Examples

set.seed(20240412)
gau_clust(
  n = 300, num_clust = 5,
  mean_matrix = rbind(
    c(1, 0, 0, 0), c(0, 1, 0, 0), c(0, 0, 1, 0),
    c(0, 0, 0, 1), c(0, 0, 0, 0)
  ), var_vec = c(0.05, 0.05, 0.05, 0.05, 0.05),
  num_dims = 4, num_noise = 2, min_n = -0.05, max_n = 0.05
)
#>                [,1]          [,2]          [,3]         [,4]          [,5]
#>   [1,]  1.065888271  1.155893e-02 -0.0028248202  0.043355399 -1.636529e-02
#>   [2,]  0.967892068  2.700506e-02 -0.0565672727  0.050669874  2.224537e-03
#>   [3,]  0.935397183  9.359796e-03 -0.0219227720  0.003017822 -1.341185e-02
#>   [4,]  1.064240596 -1.794435e-02  0.0903411412 -0.073652551 -1.857520e-02
#>   [5,]  1.018875050 -6.293476e-02  0.0034800095 -0.093499215 -4.569429e-02
#>   [6,]  0.992292799  2.508501e-02 -0.0018301121 -0.055169416  4.582746e-02
#>   [7,]  0.923393696  7.850385e-02  0.0087584391 -0.040002441 -3.455254e-02
#>   [8,]  1.019534151  2.847716e-02  0.0573456865 -0.076501464  2.582430e-02
#>   [9,]  0.976776618  4.040263e-02 -0.0583115651 -0.040563073  1.039657e-02
#>  [10,]  1.148961970  3.300531e-02 -0.0181647854 -0.025479244 -1.376949e-02
#>  [11,]  1.018104413 -2.484500e-02 -0.0858407160 -0.012305240 -4.073165e-02
#>  [12,]  0.998910960 -1.740868e-02  0.0294708627  0.055490959  4.847071e-02
#>  [13,]  1.012690907  4.146446e-02 -0.0094031884 -0.006772876 -2.970373e-02
#>  [14,]  0.996382333  1.453406e-02  0.0530385340  0.007672083  6.991699e-05
#>  [15,]  0.974518984  5.850721e-02 -0.0280024721  0.024472207  2.383065e-02
#>  [16,]  1.019005573 -6.346488e-02  0.1130664959 -0.070043785 -4.223552e-02
#>  [17,]  0.984418166  1.368462e-02  0.0820531079  0.013932211  1.261618e-02
#>  [18,]  1.017046870  4.016927e-03 -0.0050552601 -0.043019505  4.623033e-02
#>  [19,]  1.020397969  6.587756e-02  0.0476367021 -0.005109306 -2.674662e-02
#>  [20,]  1.049006888  2.323483e-02  0.0541729784  0.049163190  4.767000e-02
#>  [21,]  1.066875366  4.023786e-02 -0.0554606694  0.018712370 -2.139238e-02
#>  [22,]  0.950148980 -7.048070e-02 -0.0531451202 -0.035472627  6.776200e-03
#>  [23,]  0.992438663  4.125534e-03 -0.0301178809 -0.134900536  2.578724e-02
#>  [24,]  0.933618083  1.640820e-02  0.0043618660  0.030354472  4.627391e-02
#>  [25,]  1.017382570  7.113662e-03 -0.0456492318 -0.053700618  4.988278e-03
#>  [26,]  1.018612289 -4.095574e-02 -0.1440496245 -0.013833392  1.933860e-02
#>  [27,]  1.005519074 -2.866340e-02 -0.0073935120 -0.043701769 -2.677462e-02
#>  [28,]  1.006784814 -1.764408e-02 -0.0375083402 -0.038932020 -1.975191e-02
#>  [29,]  0.956074595  9.754566e-02 -0.0234700700 -0.005816252  4.798282e-02
#>  [30,]  1.062728980  8.289040e-02  0.0951455402  0.017404684 -7.112537e-03
#>  [31,]  0.990215162 -4.228563e-02 -0.0632394018  0.008840101  8.021953e-03
#>  [32,]  0.955226389 -4.462242e-02  0.0710563835  0.017283165  4.239841e-03
#>  [33,]  1.018130919  5.021090e-02 -0.1627563806 -0.001642179  2.394034e-02
#>  [34,]  1.033394898 -2.111408e-02  0.0019603475  0.042149772  3.145418e-03
#>  [35,]  1.013544829 -4.440800e-02 -0.0393466260 -0.032863752 -2.163410e-02
#>  [36,]  0.978769870 -2.592817e-02 -0.0525648991  0.094399599 -2.126563e-03
#>  [37,]  0.989160924 -1.402369e-02 -0.0065310537  0.049113611  2.195558e-02
#>  [38,]  1.050015410 -5.825633e-02 -0.0109372832  0.023850501  3.284489e-03
#>  [39,]  1.027336129 -2.664706e-02 -0.0330496008 -0.065924058  4.787364e-02
#>  [40,]  0.956166376  6.359599e-02  0.0302401232  0.005595998  4.729554e-02
#>  [41,]  0.968458179 -9.268763e-02  0.0203299410  0.008310896 -1.139702e-02
#>  [42,]  0.937503847  6.142306e-02  0.0274065517 -0.076233553 -1.521704e-02
#>  [43,]  1.095309062  9.809635e-03  0.0493576298  0.050248125  2.190798e-02
#>  [44,]  1.063367075 -1.217379e-02  0.0809974346 -0.009373328 -1.470256e-02
#>  [45,]  0.936650953  1.400371e-01 -0.0207557859 -0.072934136 -2.105489e-02
#>  [46,]  1.040364057 -5.531727e-02  0.0232940068 -0.080225199  4.948568e-02
#>  [47,]  0.991113598  4.471806e-02  0.0046997946 -0.015824682 -4.643196e-02
#>  [48,]  1.054926921  4.021514e-02  0.0080229677  0.003814820  2.906671e-02
#>  [49,]  1.016496610 -2.970837e-03 -0.0751120596 -0.011829216 -4.218292e-02
#>  [50,]  1.035293299  2.399767e-02  0.0539979096 -0.008991393  4.991574e-02
#>  [51,]  1.125937962 -5.140767e-02 -0.1342892042  0.004305581 -4.200190e-02
#>  [52,]  1.012076002  4.376590e-02 -0.0203078396 -0.028233408 -1.333662e-02
#>  [53,]  1.054386192 -6.094579e-02  0.0694048412 -0.046806292 -2.654866e-02
#>  [54,]  0.969709993  6.975678e-03 -0.0183803388  0.011389586 -2.783990e-02
#>  [55,]  1.057468567 -4.143359e-02 -0.0063490057 -0.008893490 -4.780630e-02
#>  [56,]  0.984413780  8.128411e-03  0.0009034067  0.034771926  4.970827e-02
#>  [57,]  1.008786265  3.386463e-02 -0.0504333384  0.035220816  3.006407e-03
#>  [58,]  0.927964859  5.269478e-02 -0.0022130705 -0.076881494  8.176925e-03
#>  [59,]  0.949959366  5.061137e-02 -0.0197629483  0.005587806  4.956267e-02
#>  [60,]  0.994509457  8.428144e-02 -0.0008822339  0.091037622 -1.627479e-02
#>  [61,] -0.035328177  9.845086e-01  0.0101642891 -0.030884556 -1.695537e-03
#>  [62,]  0.081242007  9.438558e-01  0.0315458345  0.056394725 -4.294849e-02
#>  [63,] -0.015459419  9.156110e-01 -0.0893369602 -0.067097950 -1.343644e-02
#>  [64,]  0.049102472  1.056988e+00  0.0053537407  0.043075394 -2.834701e-02
#>  [65,] -0.013601442  1.051932e+00 -0.0759207484  0.050856686  2.343479e-02
#>  [66,]  0.089069756  9.030271e-01  0.0397527114  0.004273613 -3.167909e-02
#>  [67,]  0.026646567  1.085056e+00 -0.0677997246  0.013914283  1.233606e-02
#>  [68,]  0.028615787  9.804797e-01  0.0650981813  0.028779522 -1.595814e-02
#>  [69,]  0.006484737  9.973788e-01 -0.0249915490 -0.078733837 -2.958210e-03
#>  [70,] -0.076551231  1.013356e+00 -0.0086995766  0.129898720  3.158711e-02
#>  [71,] -0.012405546  9.744679e-01 -0.0515312185 -0.019954947  1.725153e-02
#>  [72,] -0.047259734  9.859045e-01 -0.0189481113 -0.021100988  3.596430e-03
#>  [73,]  0.043332304  9.791187e-01  0.0441696990 -0.103633123  4.273322e-02
#>  [74,]  0.023935603  1.037689e+00  0.0164879688  0.009061329 -4.458161e-02
#>  [75,] -0.016304191  9.509659e-01 -0.0556936046  0.032238194 -3.275242e-02
#>  [76,]  0.012158230  1.026462e+00 -0.0581533129 -0.071623861  1.450788e-02
#>  [77,]  0.069628186  1.024652e+00 -0.0135085123  0.026074108 -2.351118e-03
#>  [78,] -0.024522988  9.371883e-01 -0.0055377701 -0.030985479 -1.977617e-02
#>  [79,] -0.009947305  1.015071e+00 -0.0083628364  0.045792442  3.313871e-02
#>  [80,] -0.039924050  9.310522e-01  0.1308436567  0.044552065 -2.347922e-02
#>  [81,] -0.038412713  9.453104e-01 -0.0124491338  0.019507663 -8.599617e-03
#>  [82,] -0.006513210  9.769655e-01  0.0641481260 -0.028414707 -4.361080e-02
#>  [83,] -0.026618393  9.795593e-01 -0.0071573339 -0.060567727  3.581735e-02
#>  [84,]  0.003585330  1.040487e+00  0.0101857967 -0.033919966 -2.593632e-02
#>  [85,]  0.021150459  1.010176e+00  0.0059945335 -0.026149274  1.919897e-02
#>  [86,] -0.026690883  9.646388e-01 -0.0548312377  0.065786268 -3.999030e-02
#>  [87,] -0.022174158  9.167271e-01 -0.0304713105 -0.079585738  1.818160e-03
#>  [88,]  0.044440286  1.060115e+00  0.0538962754  0.069110477  3.022244e-02
#>  [89,]  0.035299429  1.029633e+00  0.0072891433  0.010903015  3.515036e-02
#>  [90,] -0.045311453  1.016051e+00 -0.0128642537 -0.019286849  4.743709e-02
#>  [91,] -0.018331998  9.683970e-01  0.0146515473  0.053480095 -4.547735e-02
#>  [92,] -0.094385202  1.001547e+00  0.0104417030 -0.037115981 -4.286151e-02
#>  [93,]  0.055425265  1.007535e+00 -0.0700005702 -0.054029262 -3.121946e-02
#>  [94,]  0.009886321  1.088194e+00 -0.0070790340 -0.023645125 -4.569175e-03
#>  [95,] -0.048431027  1.059902e+00 -0.0188307080  0.004806090  5.641874e-03
#>  [96,]  0.001844253  9.225677e-01 -0.0028216008 -0.046747572  4.250327e-02
#>  [97,] -0.052253870  1.023900e+00  0.0313699754  0.053110714  3.357234e-03
#>  [98,]  0.036317374  1.026081e+00  0.0240660247 -0.081378484 -5.549375e-04
#>  [99,] -0.053909328  9.288535e-01  0.0513185334 -0.054248121 -2.555933e-02
#> [100,] -0.112129809  9.720851e-01  0.0495688098 -0.004564382  2.165986e-02
#> [101,]  0.006152877  1.025086e+00 -0.0178401442 -0.042970657  2.600802e-02
#> [102,] -0.011023788  9.300220e-01 -0.0375745154 -0.055678032 -4.266945e-02
#> [103,]  0.026504690  1.079617e+00 -0.0336614738 -0.011563494  1.312419e-02
#> [104,] -0.042589877  9.712218e-01  0.0888387485 -0.038883701 -3.457974e-02
#> [105,]  0.073421730  9.331954e-01  0.0321521821 -0.034831600  3.813075e-02
#> [106,] -0.008393931  9.226315e-01 -0.0070047787  0.006453957  4.583308e-02
#> [107,] -0.041801438  9.290396e-01  0.0762643633 -0.014738532  3.153494e-02
#> [108,]  0.014590984  9.263375e-01  0.0302586227  0.026008770 -1.656454e-02
#> [109,]  0.051739319  1.033900e+00  0.0807670195 -0.006674470  4.074052e-02
#> [110,]  0.014064551  9.968784e-01 -0.0321887529 -0.025340024 -5.229591e-03
#> [111,] -0.002157658  9.592049e-01 -0.1209826633 -0.033523287  4.289947e-02
#> [112,]  0.011984284  9.570450e-01  0.0081503673  0.032009804 -4.693646e-02
#> [113,]  0.056586780  9.424975e-01 -0.0507292523 -0.024659246  4.079205e-02
#> [114,] -0.017708015  1.045866e+00  0.0308173838  0.085945969  4.593775e-02
#> [115,] -0.017830810  1.036551e+00 -0.0002154313 -0.072935760 -3.207291e-02
#> [116,] -0.017083588  1.063053e+00 -0.0317360569 -0.013007695 -4.309148e-02
#> [117,] -0.015148545  9.530852e-01 -0.1052470497 -0.023119542  4.208677e-02
#> [118,] -0.004780701  1.074676e+00  0.0254802286 -0.041544169  3.084040e-02
#> [119,]  0.004572058  9.706006e-01  0.0330457975 -0.029306130 -3.804917e-03
#> [120,] -0.035086328  9.179110e-01  0.0027499280  0.050364521 -1.082145e-02
#> [121,] -0.085051285 -3.646887e-02  0.9913488762  0.065507046 -8.248837e-03
#> [122,] -0.038304918 -7.514657e-02  1.0309606354 -0.030699977 -2.394960e-02
#> [123,]  0.027365206 -9.882285e-02  1.0086437045  0.032775520 -4.812042e-02
#> [124,] -0.070349805 -5.758031e-02  1.0366386790  0.019225349  4.655695e-02
#> [125,]  0.094911105  5.822307e-02  1.0052424965 -0.039231787  4.746066e-02
#> [126,]  0.046516860  2.127034e-03  1.0456405149  0.124252592 -1.690880e-02
#> [127,]  0.090292831 -1.342696e-02  1.0524021492  0.009034555 -3.542551e-02
#> [128,]  0.027905964  1.160132e-02  1.0135764592 -0.045794038 -2.490102e-02
#> [129,]  0.019896665 -3.575892e-02  1.0843572643  0.064397906 -4.686855e-02
#> [130,]  0.036286100  2.427904e-02  1.0021911341 -0.019133318  4.568508e-02
#> [131,]  0.075293256  4.788805e-02  1.0854777401  0.041652120 -1.694252e-03
#> [132,]  0.007520230  1.626115e-02  1.0203360159 -0.024212534 -2.421517e-02
#> [133,] -0.015931137 -3.392143e-03  0.8875897644  0.001678849  1.448532e-02
#> [134,] -0.067476743  3.456905e-02  1.0405605506  0.003389206  3.271302e-02
#> [135,]  0.075811192 -2.818549e-02  1.0226414428 -0.008037322 -6.234291e-03
#> [136,]  0.005799540 -5.139786e-02  1.0454434408  0.051888451 -4.804818e-02
#> [137,]  0.007472601 -9.474133e-03  1.0477151938  0.074478811 -8.559206e-05
#> [138,] -0.042908611  2.821401e-02  1.0029540332  0.052806595 -2.051229e-03
#> [139,] -0.042941278  2.243489e-02  1.0796736358 -0.049254594 -4.483748e-02
#> [140,] -0.014508897  6.993325e-02  1.0636535924 -0.033788673  1.087038e-02
#> [141,]  0.049537581 -4.418934e-02  1.0321791773  0.034933338 -3.384040e-03
#> [142,] -0.057312195  4.831921e-02  0.9564683310  0.113990073 -4.725878e-02
#> [143,]  0.039380137 -5.962204e-02  0.9827821454  0.033399117 -4.406941e-02
#> [144,]  0.019387236 -1.052856e-02  0.9924997858 -0.081011613  1.295361e-02
#> [145,] -0.036733093  4.679421e-05  1.0524449260 -0.058728599  2.167122e-02
#> [146,]  0.046460424 -2.470802e-02  0.9836576710 -0.010591476 -4.919681e-02
#> [147,]  0.012895380  3.327731e-02  0.9718063500  0.023302257 -3.184544e-02
#> [148,] -0.011800993 -2.657109e-03  1.0329088012 -0.076696813 -1.948746e-02
#> [149,]  0.018250073 -1.248055e-01  1.0054484150 -0.057110985  2.529822e-02
#> [150,] -0.031515216 -1.334599e-02  0.9993426530 -0.058468094  9.170931e-03
#> [151,]  0.029537093 -6.120481e-02  0.9466947882  0.093905192 -1.527652e-02
#> [152,] -0.006164119  5.225142e-02  1.0635787843 -0.011065773  2.237138e-02
#> [153,] -0.002062637 -3.657979e-02  1.0566798190  0.039304840  9.025664e-03
#> [154,] -0.087763808 -9.619617e-03  1.0031589816 -0.046286908  3.841776e-02
#> [155,]  0.065962556  2.794777e-02  1.0567137403 -0.004441563  1.986252e-02
#> [156,] -0.001731481  1.720317e-02  1.0377840659 -0.054198474 -2.024381e-03
#> [157,] -0.003173674 -3.075895e-02  1.0431402105  0.115522815 -3.638958e-03
#> [158,]  0.141941562  1.059424e-01  0.9639641934  0.036643496  4.160269e-02
#> [159,] -0.023132103  5.632095e-02  0.9702460004 -0.009707397 -2.210917e-02
#> [160,]  0.057900830  6.296028e-03  0.9869420090 -0.073802097  9.798289e-03
#> [161,] -0.027307835  4.939952e-02  1.1036113161  0.034831906  2.813212e-02
#> [162,] -0.045629893 -5.731689e-02  0.9617371497  0.079376741  1.161731e-02
#> [163,]  0.062077555 -1.829916e-03  1.0005571253  0.039799586 -7.181504e-03
#> [164,] -0.062091335 -6.468131e-02  0.9552088772  0.010320968  3.274150e-02
#> [165,]  0.018645948  1.920006e-02  1.0015175737  0.001259680  2.195759e-02
#> [166,] -0.027451644  9.566163e-02  1.0549364143 -0.015001235  1.017579e-02
#> [167,]  0.061981499  8.618565e-03  0.9921736882 -0.130953851 -4.184624e-02
#> [168,]  0.054056296 -4.890321e-02  0.9917980615 -0.013644478  3.240644e-02
#> [169,] -0.102643525 -7.166324e-02  0.9457158525  0.053846197 -4.793698e-02
#> [170,] -0.048779839 -2.143125e-02  0.9439841358 -0.095366078  4.696704e-02
#> [171,]  0.034705332  2.186236e-02  0.8925351215 -0.108009644  2.402314e-02
#> [172,] -0.036241895 -4.011914e-02  0.9902054351  0.015871560 -2.001644e-02
#> [173,]  0.027195571 -7.526343e-02  0.9236040238  0.023616750  2.770050e-02
#> [174,]  0.006614102  8.751798e-03  0.9726617384  0.035486768 -3.930738e-03
#> [175,] -0.034897837 -4.177927e-02  1.0349706476  0.026585118  2.764713e-02
#> [176,]  0.004805214  5.544970e-02  1.0465242665  0.012394845 -4.944980e-02
#> [177,]  0.020929503 -1.049143e-01  1.0386930950  0.070753895 -8.236953e-03
#> [178,] -0.007020809  1.543353e-02  1.0187138799  0.011947896  3.023053e-02
#> [179,] -0.062183903  1.604099e-02  1.0101677714  0.084470774  1.955693e-02
#> [180,]  0.031289563 -1.982588e-02  1.0118758251  0.055373978 -4.117295e-02
#> [181,] -0.025004467 -6.854813e-02  0.0088456820  1.033170555  4.636580e-02
#> [182,] -0.037815282  1.533315e-02  0.0176415935  0.988574398  3.212397e-02
#> [183,]  0.070560582  6.540502e-02 -0.0514690571  0.982218060  4.390859e-03
#> [184,]  0.053204854 -8.435665e-02 -0.0175375648  1.035581167 -1.128168e-02
#> [185,] -0.062389703 -2.063838e-02  0.0436012616  0.947427527  1.274856e-02
#> [186,]  0.020450508  4.899892e-02 -0.0021643047  0.975043543 -3.872981e-02
#> [187,]  0.006515197  1.784136e-01 -0.0626222650  0.912643216  1.738305e-02
#> [188,]  0.026798336 -4.691122e-02 -0.0595100526  0.965388409  2.541680e-03
#> [189,]  0.034122991  6.978465e-02  0.0434399467  0.952015291  2.832790e-02
#> [190,] -0.022302746 -4.048773e-02 -0.0046278578  0.984575770 -3.933804e-02
#> [191,] -0.046786574  3.407150e-02  0.0610936800  1.008117408  1.680248e-02
#> [192,]  0.050388224  4.858652e-02 -0.0146097465  1.028112496 -8.446755e-04
#> [193,]  0.060399431 -4.302518e-02 -0.0026907982  1.065304104 -4.111465e-02
#> [194,]  0.049194402 -5.855265e-02  0.0574684521  0.980605101  2.642885e-02
#> [195,] -0.013676694 -1.811536e-06  0.0329736827  1.030177059  4.722976e-02
#> [196,] -0.075047545  2.319719e-02 -0.0462866495  1.000140259  1.143594e-02
#> [197,]  0.008930212  5.200887e-03 -0.0656088831  0.885741927 -3.369414e-02
#> [198,] -0.021882870  9.247778e-02  0.0019448508  0.980292962 -4.448640e-02
#> [199,] -0.066826471  3.196754e-02 -0.0683709701  1.027335393 -5.295939e-03
#> [200,]  0.031680588 -3.148871e-02  0.0987452739  1.073749799  8.587251e-03
#> [201,]  0.029341928 -1.773538e-02  0.0976863595  1.001224804  6.016336e-03
#> [202,] -0.055350980 -1.550837e-02  0.0954221327  1.011831571 -4.331658e-02
#> [203,] -0.023143877  5.148335e-03  0.0053010777  1.092407045  4.619585e-02
#> [204,] -0.073795722 -1.290719e-03  0.0750704176  0.988968632  4.661723e-02
#> [205,]  0.005632855 -2.947375e-02 -0.0172387773  0.991460821  1.180746e-02
#> [206,]  0.045011143 -4.815667e-02  0.0823018959  1.004790721 -1.000135e-02
#> [207,] -0.084695723  6.047729e-03 -0.0186731552  0.997291392  2.395831e-02
#> [208,]  0.024663892 -6.139822e-02  0.0554785693  1.030240182  7.926267e-03
#> [209,]  0.018400954 -3.514917e-02  0.0079262878  1.010967693 -2.957340e-02
#> [210,] -0.011178146 -5.127387e-02  0.0566327808  1.080633430  1.588908e-02
#> [211,] -0.085526163 -1.780797e-02  0.0371214636  0.944811028  1.216441e-02
#> [212,]  0.011276256  7.079383e-02  0.0483696903  0.999562710 -1.406608e-02
#> [213,] -0.027269451  7.344809e-02 -0.0449244698  0.988264573  1.813453e-02
#> [214,]  0.069128644  1.536621e-02 -0.0908334108  0.965478080 -1.180972e-02
#> [215,] -0.084961332  4.844615e-02 -0.0214513098  0.996366012 -4.236208e-02
#> [216,] -0.022526661  7.296277e-02 -0.0493970838  0.988889320 -4.797558e-02
#> [217,] -0.073909763 -2.429347e-02 -0.0308548276  0.938955373 -4.800309e-02
#> [218,] -0.025315988  1.149620e-02  0.0649383407  0.934430594 -3.240149e-03
#> [219,] -0.024921286 -8.221553e-04 -0.0229995939  0.978701300  2.024942e-02
#> [220,]  0.059823280 -2.605054e-02 -0.0052613580  1.038173868  1.567020e-03
#> [221,]  0.005467867 -2.531808e-02  0.0468938859  0.959217629  3.064822e-02
#> [222,] -0.061943042  2.946785e-02 -0.0295610118  0.977064691 -3.081816e-02
#> [223,] -0.005528384 -4.180945e-02  0.0328881580  1.001333149 -4.520892e-02
#> [224,]  0.038009622 -1.075326e-02  0.0337872596  1.007614938 -4.436592e-02
#> [225,]  0.063794022  1.764721e-02  0.1634499849  1.008037110  1.480271e-02
#> [226,] -0.037220993  5.884071e-02  0.0348127970  0.985855986 -2.395036e-02
#> [227,] -0.088508754  4.055146e-02  0.0009757282  0.965605725 -1.574961e-03
#> [228,] -0.039349350 -5.698671e-03  0.0161654156  0.989508558 -3.050621e-02
#> [229,] -0.001058472  4.169142e-02  0.0709972647  1.024666101 -2.813635e-02
#> [230,] -0.017274673  5.430432e-02  0.0378631809  1.062377770 -6.375915e-03
#> [231,] -0.035672273  2.389937e-02 -0.0320540428  1.064179209 -1.282854e-02
#> [232,] -0.064560172  1.252148e-01 -0.0442479772  0.955517796 -3.236524e-03
#> [233,]  0.018686669 -8.615176e-02 -0.0208755300  1.004833626  2.683606e-03
#> [234,] -0.041160238  2.919070e-02 -0.1316827998  0.989896544 -3.180316e-02
#> [235,]  0.082912389  2.250583e-02  0.0085812667  0.970226877  1.196132e-02
#> [236,]  0.058361864  3.296542e-03  0.0128481908  1.027696498  3.349630e-03
#> [237,]  0.069347984  9.098899e-02 -0.0203641536  0.992026204  9.458663e-03
#> [238,]  0.041417802  4.781403e-02  0.0182236500  1.005152653 -4.785613e-02
#> [239,] -0.065464324  2.979647e-02  0.0462102469  1.031617472 -4.000756e-02
#> [240,] -0.012242552  1.782238e-02 -0.0327549587  0.978930635 -2.342327e-02
#> [241,] -0.051023469 -1.645813e-02  0.0403722683 -0.039515922 -1.398042e-02
#> [242,]  0.003520198  1.133095e-01 -0.0295298624 -0.071508165 -7.630326e-03
#> [243,]  0.012077697 -5.300695e-02  0.0933441878  0.086607177 -2.130766e-03
#> [244,] -0.011947524  6.783689e-02  0.0872742441  0.032133589 -1.424578e-02
#> [245,] -0.085228757 -2.962083e-02  0.0393768853  0.025479659 -3.703061e-02
#> [246,] -0.024065745 -4.272705e-02 -0.0557676580  0.041296509  4.368189e-02
#> [247,]  0.098640209 -4.677501e-02  0.0046611213  0.015759818 -3.408686e-02
#> [248,] -0.065313628 -7.361018e-02 -0.0298238115  0.048355803 -1.782028e-03
#> [249,] -0.060560525 -3.149467e-02 -0.0815921077 -0.006029976  4.574857e-02
#> [250,]  0.051566721 -5.378725e-02  0.0565617543 -0.030635498 -3.758198e-02
#> [251,] -0.059766355 -1.186272e-02 -0.0290438491 -0.065732163 -3.224220e-02
#> [252,] -0.066743554  5.788596e-02  0.0455488536  0.119860146 -3.329519e-02
#> [253,]  0.133910329  2.995389e-02  0.0110365788  0.018548741 -2.614621e-02
#> [254,]  0.045942689 -2.748146e-02 -0.0015928557  0.046058316  1.047185e-02
#> [255,] -0.056951891  1.612262e-03  0.0915512652 -0.023849177  9.539057e-03
#> [256,] -0.006212290 -1.205934e-02  0.1032428248 -0.051800400  3.582947e-02
#> [257,] -0.034652692  4.777534e-02 -0.0200609260  0.001708344  4.787654e-02
#> [258,]  0.037964380  2.029821e-02 -0.0271131852 -0.071954990 -3.923594e-02
#> [259,]  0.050665292  4.188587e-02 -0.0642508104 -0.001426933 -5.881421e-04
#> [260,] -0.038819220  1.790239e-02 -0.0210832666  0.067543685 -3.738044e-02
#> [261,] -0.038584770 -2.412940e-02  0.0278597261  0.123618765 -1.258585e-02
#> [262,] -0.079763748  4.590490e-02  0.0140863448 -0.073285769 -1.155495e-02
#> [263,]  0.011165224  9.537624e-03  0.0381762398  0.060812520 -2.969666e-03
#> [264,]  0.009889027 -2.928470e-02  0.0423537801  0.013625957  3.450479e-02
#> [265,] -0.029535471  8.560527e-03  0.0809393650 -0.039835557  1.701318e-02
#> [266,] -0.093585810 -2.686145e-04 -0.0553283750  0.077407011  1.606255e-03
#> [267,]  0.022960008  5.963625e-02 -0.0456825985 -0.067327280 -1.515195e-02
#> [268,] -0.051158848 -2.033465e-02 -0.0122634806  0.014061360 -1.504552e-02
#> [269,]  0.019020667 -5.083106e-02  0.0261610153 -0.079967014 -3.916970e-03
#> [270,]  0.065442085  1.226494e-02 -0.0027647388 -0.008133330 -3.101151e-02
#> [271,] -0.050261167  1.478564e-02  0.0975837598 -0.006872937  1.392381e-02
#> [272,] -0.029754475 -6.984689e-02  0.0405115001  0.043968470 -9.808695e-03
#> [273,] -0.025896543  5.692452e-02  0.0983943383 -0.028408856 -1.376565e-03
#> [274,] -0.010721953 -2.513373e-02 -0.0143028486  0.016877241  4.721620e-02
#> [275,]  0.025733296  6.519772e-02 -0.0707464729 -0.021474269  4.137618e-02
#> [276,]  0.010067666 -4.176965e-02 -0.0258703679 -0.014524172  1.398072e-02
#> [277,] -0.071225088  3.064358e-03  0.0458304420  0.111855153  2.432093e-03
#> [278,] -0.014910645 -9.148821e-02 -0.0524737475  0.053054988 -2.182356e-02
#> [279,] -0.006589292  7.978325e-02 -0.0559548271  0.042426881 -2.312348e-04
#> [280,]  0.018638038  4.162003e-02  0.0186291596 -0.141462897 -4.676903e-02
#> [281,] -0.057174292  1.687270e-02 -0.0661718583  0.022036265 -4.664230e-02
#> [282,]  0.064021272 -4.651792e-02  0.0006311524 -0.034355763 -1.697943e-02
#> [283,] -0.064983429 -5.250410e-02 -0.0208466554  0.034416921  1.247595e-02
#> [284,]  0.010702211  5.641132e-02  0.0608869042 -0.034134170 -1.270305e-02
#> [285,]  0.063051308  6.863249e-02  0.0158142529  0.026989699  1.831584e-03
#> [286,] -0.055526451 -9.890627e-02 -0.0399054752  0.051868476 -2.295712e-02
#> [287,] -0.031248480  4.276347e-02  0.0370374705 -0.022847399  7.475937e-03
#> [288,] -0.004767215 -2.780439e-02 -0.0409426372  0.102872190 -2.954348e-02
#> [289,] -0.065111417  4.075295e-02 -0.0057974740  0.092543343  4.756168e-02
#> [290,]  0.055796032 -1.843437e-02 -0.0894643958 -0.039375038 -1.390950e-02
#> [291,] -0.049035690 -3.257938e-02  0.0201573346 -0.050223139  5.781383e-03
#> [292,]  0.080950748 -2.300040e-02  0.0437423885  0.060183493  2.104691e-02
#> [293,] -0.055262586  5.617661e-02 -0.0272454826  0.062086237  4.288171e-02
#> [294,]  0.043448962 -1.109332e-02  0.0012179523 -0.001378103  2.265591e-02
#> [295,]  0.028089992  2.005366e-03  0.0587989148 -0.048767750  3.854656e-02
#> [296,] -0.015242788  5.884112e-02 -0.0233267787  0.001121090  3.538895e-02
#> [297,]  0.029254875 -5.022800e-03  0.0269199956  0.051411022  1.012173e-02
#> [298,] -0.119236809 -2.666068e-02  0.0518458671 -0.011527267  3.307285e-02
#> [299,]  0.042202753  9.931677e-02  0.0669995448  0.075419321 -3.944466e-02
#> [300,] -0.059586431  3.463283e-02 -0.0172457637 -0.036218692  2.121629e-02
#>                 [,6]
#>   [1,]  2.486890e-02
#>   [2,] -7.464143e-05
#>   [3,] -3.914577e-02
#>   [4,] -2.089173e-02
#>   [5,]  1.894828e-02
#>   [6,]  4.280163e-02
#>   [7,] -4.947531e-02
#>   [8,] -1.891325e-03
#>   [9,] -1.922574e-02
#>  [10,]  2.878356e-02
#>  [11,] -2.562035e-02
#>  [12,] -4.041058e-02
#>  [13,] -2.578040e-02
#>  [14,]  4.699994e-02
#>  [15,] -3.792324e-02
#>  [16,] -3.249542e-02
#>  [17,] -4.670734e-02
#>  [18,]  4.094721e-02
#>  [19,]  2.815265e-02
#>  [20,] -2.635475e-02
#>  [21,] -4.490325e-02
#>  [22,] -1.117406e-02
#>  [23,] -2.813520e-02
#>  [24,]  4.569216e-02
#>  [25,] -1.555211e-02
#>  [26,] -3.139715e-02
#>  [27,]  2.859112e-02
#>  [28,] -1.182557e-02
#>  [29,]  1.707904e-02
#>  [30,]  3.909184e-02
#>  [31,]  4.284323e-02
#>  [32,] -1.697390e-03
#>  [33,] -1.653476e-02
#>  [34,] -4.995363e-02
#>  [35,] -1.641861e-02
#>  [36,] -5.542426e-03
#>  [37,] -2.505893e-02
#>  [38,] -3.475889e-02
#>  [39,] -6.820234e-03
#>  [40,] -8.336465e-03
#>  [41,]  2.106332e-02
#>  [42,] -6.247264e-04
#>  [43,]  3.697544e-02
#>  [44,]  2.549011e-02
#>  [45,]  2.692062e-03
#>  [46,]  4.806504e-02
#>  [47,]  2.019991e-02
#>  [48,]  3.063711e-03
#>  [49,] -4.393639e-02
#>  [50,] -1.201383e-02
#>  [51,]  4.244301e-02
#>  [52,]  3.218874e-02
#>  [53,] -8.492386e-03
#>  [54,]  2.604337e-02
#>  [55,] -1.912451e-02
#>  [56,]  6.823914e-03
#>  [57,] -7.189237e-04
#>  [58,]  2.087572e-02
#>  [59,]  4.265772e-02
#>  [60,]  2.790086e-02
#>  [61,] -3.074746e-02
#>  [62,] -4.285247e-02
#>  [63,]  5.767070e-04
#>  [64,]  2.300305e-02
#>  [65,] -2.029363e-02
#>  [66,]  4.745385e-02
#>  [67,]  1.705410e-02
#>  [68,] -1.890293e-02
#>  [69,]  1.236857e-02
#>  [70,]  3.340483e-02
#>  [71,] -1.148428e-02
#>  [72,]  1.961427e-02
#>  [73,]  1.563474e-02
#>  [74,]  2.543063e-02
#>  [75,] -1.997720e-02
#>  [76,] -2.749464e-02
#>  [77,] -2.690204e-02
#>  [78,] -3.704098e-02
#>  [79,] -2.934588e-02
#>  [80,]  2.540732e-02
#>  [81,] -5.279571e-03
#>  [82,] -3.807583e-02
#>  [83,]  4.963515e-02
#>  [84,] -4.806631e-02
#>  [85,] -1.766202e-02
#>  [86,] -4.904982e-04
#>  [87,]  3.309283e-02
#>  [88,] -2.918599e-02
#>  [89,] -2.563736e-02
#>  [90,]  1.180737e-02
#>  [91,]  1.580594e-02
#>  [92,] -4.122989e-02
#>  [93,] -4.302583e-03
#>  [94,] -2.874878e-02
#>  [95,] -3.216237e-02
#>  [96,]  1.234710e-02
#>  [97,] -6.071669e-03
#>  [98,]  3.965713e-02
#>  [99,] -3.225420e-02
#> [100,] -2.275280e-02
#> [101,]  4.379803e-02
#> [102,]  3.650069e-03
#> [103,]  2.254789e-04
#> [104,] -2.091612e-02
#> [105,]  4.435780e-02
#> [106,] -3.647009e-02
#> [107,]  3.068151e-02
#> [108,] -3.726166e-02
#> [109,]  2.222062e-02
#> [110,]  1.180547e-02
#> [111,]  4.439201e-02
#> [112,]  3.776503e-02
#> [113,]  3.226880e-02
#> [114,]  2.295346e-02
#> [115,] -6.655193e-03
#> [116,] -3.767581e-02
#> [117,] -4.138408e-02
#> [118,]  3.749861e-04
#> [119,] -2.554590e-02
#> [120,] -4.587207e-02
#> [121,] -3.865607e-02
#> [122,]  3.888366e-02
#> [123,] -2.585429e-02
#> [124,] -3.986343e-02
#> [125,]  1.334919e-02
#> [126,] -1.321781e-02
#> [127,]  3.843992e-02
#> [128,] -8.204084e-03
#> [129,]  4.651237e-02
#> [130,]  3.535699e-02
#> [131,]  2.727869e-02
#> [132,]  7.070095e-03
#> [133,] -1.396984e-02
#> [134,] -4.017218e-02
#> [135,]  2.136979e-02
#> [136,]  1.134227e-02
#> [137,]  4.025717e-02
#> [138,]  1.231432e-02
#> [139,]  4.593196e-02
#> [140,]  1.280777e-02
#> [141,] -1.659563e-02
#> [142,]  2.951029e-02
#> [143,]  3.957487e-02
#> [144,]  4.261148e-02
#> [145,]  3.593054e-02
#> [146,] -8.075021e-03
#> [147,]  3.040882e-02
#> [148,] -1.110145e-03
#> [149,] -3.424502e-02
#> [150,]  8.308103e-03
#> [151,] -1.395520e-02
#> [152,] -3.907406e-02
#> [153,]  4.091924e-04
#> [154,] -3.423203e-02
#> [155,]  4.820397e-02
#> [156,] -2.182712e-02
#> [157,]  1.844037e-02
#> [158,] -2.035590e-02
#> [159,]  9.199083e-04
#> [160,]  4.349411e-02
#> [161,]  1.195387e-02
#> [162,] -1.584331e-02
#> [163,]  3.097637e-02
#> [164,] -1.821427e-02
#> [165,] -1.864395e-03
#> [166,] -2.359111e-03
#> [167,] -1.646955e-02
#> [168,] -2.074100e-02
#> [169,] -1.843426e-02
#> [170,]  1.768110e-02
#> [171,] -2.695115e-02
#> [172,] -4.962866e-02
#> [173,] -3.323244e-02
#> [174,] -4.900636e-03
#> [175,]  7.924833e-03
#> [176,] -1.121300e-02
#> [177,] -3.590300e-02
#> [178,]  1.361298e-02
#> [179,]  2.054805e-02
#> [180,] -1.574631e-02
#> [181,]  2.745160e-02
#> [182,] -6.646649e-03
#> [183,] -3.757337e-02
#> [184,]  1.374055e-02
#> [185,] -1.313869e-02
#> [186,] -4.040160e-02
#> [187,]  2.814390e-02
#> [188,]  2.463870e-02
#> [189,] -3.547120e-02
#> [190,] -3.341442e-02
#> [191,]  2.429026e-02
#> [192,]  1.993372e-02
#> [193,]  1.338825e-02
#> [194,]  4.514059e-02
#> [195,] -3.350452e-02
#> [196,] -1.250297e-02
#> [197,] -1.774048e-02
#> [198,] -4.320934e-02
#> [199,]  2.429277e-02
#> [200,]  3.501206e-02
#> [201,]  4.045311e-02
#> [202,] -3.129198e-02
#> [203,] -1.935798e-02
#> [204,]  1.037647e-03
#> [205,]  4.306777e-02
#> [206,]  3.460671e-02
#> [207,]  4.177815e-02
#> [208,] -8.869496e-03
#> [209,] -3.471472e-02
#> [210,] -3.079085e-02
#> [211,]  2.586799e-02
#> [212,]  1.473175e-02
#> [213,] -1.238128e-02
#> [214,] -2.117140e-02
#> [215,] -1.160142e-02
#> [216,] -3.765488e-02
#> [217,]  6.447285e-03
#> [218,]  3.134232e-02
#> [219,] -3.567554e-03
#> [220,] -4.305765e-02
#> [221,] -3.287951e-02
#> [222,]  2.894529e-02
#> [223,] -3.999188e-02
#> [224,] -3.823197e-02
#> [225,] -3.094601e-02
#> [226,]  4.190498e-02
#> [227,] -4.235367e-02
#> [228,]  3.708997e-02
#> [229,] -2.700184e-03
#> [230,] -2.154518e-02
#> [231,]  3.521341e-02
#> [232,] -4.113018e-02
#> [233,]  2.685297e-02
#> [234,] -2.965704e-02
#> [235,]  1.825985e-02
#> [236,] -7.389881e-03
#> [237,] -2.757974e-02
#> [238,] -4.603738e-03
#> [239,]  8.885216e-04
#> [240,] -3.778235e-02
#> [241,] -3.089138e-02
#> [242,]  3.559450e-03
#> [243,] -4.653795e-02
#> [244,]  2.192818e-02
#> [245,]  2.271448e-02
#> [246,]  1.962436e-02
#> [247,] -3.174975e-02
#> [248,] -4.505909e-02
#> [249,]  1.530852e-02
#> [250,]  4.522111e-02
#> [251,]  2.363636e-02
#> [252,] -6.830867e-04
#> [253,]  3.961406e-02
#> [254,]  1.772474e-02
#> [255,] -4.463025e-02
#> [256,] -1.259472e-02
#> [257,] -5.567049e-04
#> [258,]  8.778661e-03
#> [259,] -1.659580e-03
#> [260,] -1.191862e-02
#> [261,] -3.202104e-02
#> [262,]  2.628882e-03
#> [263,]  2.443555e-02
#> [264,]  2.574609e-02
#> [265,] -4.289312e-02
#> [266,]  7.759684e-03
#> [267,]  3.906299e-02
#> [268,] -2.608614e-02
#> [269,]  4.270295e-02
#> [270,] -1.387756e-02
#> [271,] -1.974681e-02
#> [272,] -1.468811e-02
#> [273,]  3.571725e-02
#> [274,] -2.849413e-02
#> [275,] -1.622426e-02
#> [276,]  4.297363e-02
#> [277,] -2.848153e-02
#> [278,] -1.093619e-02
#> [279,] -2.283338e-02
#> [280,] -3.871867e-02
#> [281,]  1.886178e-02
#> [282,]  3.420622e-02
#> [283,] -1.994977e-02
#> [284,]  3.164221e-02
#> [285,] -3.328469e-02
#> [286,] -3.784973e-02
#> [287,] -2.611218e-02
#> [288,]  2.860999e-02
#> [289,] -2.366030e-02
#> [290,] -2.996860e-02
#> [291,]  7.427214e-03
#> [292,]  3.970157e-02
#> [293,]  2.195209e-02
#> [294,]  3.600596e-02
#> [295,]  4.273254e-02
#> [296,]  4.676937e-02
#> [297,] -3.858613e-02
#> [298,]  2.336692e-02
#> [299,]  1.325866e-02
#> [300,]  2.264760e-02