Generate synthetic data with Gaussian clusters

Generate Gaussian Clusters

Usage

gaussian_clusters(
  sample_size,
  num_clusters,
  mean_matrix,
  var_vec,
  num_dims,
  num_noise_dims,
  min_noise,
  max_noise
)

Arguments

sample_size

The total number of data points to be generated.

num_clusters

The number of clusters to generate.

mean_matrix

A matrix where each row represents the mean vector for a cluster.

var_vec

A vector specifying the variance for each cluster.

num_dims

The number of effective dimensions for the data points.

num_noise_dims

The number of additional noise dimensions to be generated.

min_noise

The minimum value for the noise added to the data points.

max_noise

The maximum value for the noise added to the data points.

Value

A matrix containing the generated Gaussian clusters.

Details

This function generates Gaussian clusters with specified parameters.

Examples

gaussian_clusters(sample_size = 300, num_clusters = 5,
mean_matrix = rbind(c(1,0,0,0), c(0,1,0,0), c(0,0,1,0),
c(0,0,0,1), c(0,0,0,0)), var_vec = c(0.05, 0.05, 0.05, 0.05, 0.05),
num_dims = 4, num_noise_dims = 2, min_noise = -0.05, max_noise = 0.05)
#>                 [,1]          [,2]         [,3]          [,4]          [,5]
#>   [1,]  9.706239e-01 -5.666241e-03 -0.006919228  0.0098598752  0.0268809722
#>   [2,]  9.386164e-01 -8.216005e-02  0.018332288  0.0462507502  0.0144801217
#>   [3,]  1.076572e+00  2.214494e-03 -0.045036921 -0.0341759047 -0.0059184681
#>   [4,]  1.059068e+00  4.853941e-02  0.100309750  0.0012510689 -0.0425121390
#>   [5,]  9.694301e-01  1.021059e-01 -0.100519890 -0.0529118756  0.0491911710
#>   [6,]  9.599125e-01  5.874332e-02  0.042919194  0.0471887229 -0.0184156843
#>   [7,]  1.048732e+00  4.186154e-02  0.079317803  0.0043423620  0.0376564745
#>   [8,]  1.018659e+00  3.576727e-02 -0.038717794  0.0667804802  0.0337217029
#>   [9,]  9.437627e-01  1.326787e-02  0.006106019  0.0151690665 -0.0055717256
#>  [10,]  9.918151e-01 -5.491040e-02  0.029714698  0.0881071081 -0.0219485987
#>  [11,]  1.021258e+00  5.364312e-03 -0.022010519  0.0586730073 -0.0058045116
#>  [12,]  1.004109e+00 -2.880923e-03  0.064027300 -0.0046078485 -0.0302668328
#>  [13,]  8.926842e-01  8.665387e-02  0.048725768  0.0757281150  0.0222657867
#>  [14,]  1.105630e+00  5.071989e-02  0.102252513 -0.0188758546  0.0142517534
#>  [15,]  1.068305e+00 -9.290020e-02  0.029073393 -0.0009509086 -0.0048338986
#>  [16,]  1.031739e+00 -3.115538e-02 -0.061162491 -0.0537743384 -0.0448315877
#>  [17,]  1.017583e+00 -5.967709e-02 -0.098780894  0.0587711132  0.0109261744
#>  [18,]  1.043566e+00  2.449332e-02 -0.016866132  0.0644411387  0.0074092948
#>  [19,]  1.010683e+00  3.655469e-02  0.044534935 -0.0249387531  0.0315089982
#>  [20,]  9.376236e-01  2.115573e-02 -0.018519138  0.0075051933  0.0181185631
#>  [21,]  1.018468e+00 -3.636601e-02 -0.032944830  0.0636450732  0.0104012838
#>  [22,]  1.044856e+00 -3.486607e-02  0.062582175 -0.1223867885  0.0282545492
#>  [23,]  9.605127e-01  5.372814e-02  0.003427788  0.0650750919  0.0257462068
#>  [24,]  1.044389e+00  7.419358e-02 -0.109107811  0.0084597603  0.0352274368
#>  [25,]  1.035295e+00 -7.089526e-02  0.107517496  0.0343681254 -0.0195516608
#>  [26,]  9.439231e-01  1.813623e-02  0.013306247  0.0120796163  0.0130255657
#>  [27,]  1.016384e+00 -6.390094e-03  0.028854973  0.0045415346  0.0370371234
#>  [28,]  9.178356e-01  3.850189e-02 -0.051062061 -0.0374847844 -0.0059450605
#>  [29,]  1.036762e+00  1.123298e-02  0.034518753 -0.0319845077  0.0409017809
#>  [30,]  9.764528e-01  9.421216e-03 -0.071689059  0.0210843236  0.0294616356
#>  [31,]  1.002895e+00  3.487598e-02  0.011670511 -0.0051083506 -0.0347174336
#>  [32,]  1.044650e+00  3.332221e-02 -0.021770389  0.0283571496  0.0165013298
#>  [33,]  9.847193e-01 -4.428657e-02 -0.011276651  0.0129551138  0.0020600955
#>  [34,]  1.037752e+00  1.822831e-02 -0.059330130  0.0560197882  0.0291315723
#>  [35,]  1.001272e+00 -6.942771e-02 -0.009982426  0.0199860426 -0.0241734581
#>  [36,]  1.099768e+00  2.368616e-02  0.006138647  0.0375230706 -0.0043311558
#>  [37,]  9.056984e-01  7.161460e-03  0.069259590  0.0333017849  0.0197050641
#>  [38,]  8.766409e-01  1.049174e-01  0.055923013  0.0072925632  0.0111815475
#>  [39,]  8.943924e-01 -8.133101e-02 -0.032593031  0.0190475179  0.0195326750
#>  [40,]  9.969616e-01 -4.929460e-02 -0.077313554  0.0044283904 -0.0029131754
#>  [41,]  1.007877e+00 -3.546488e-02  0.018222629 -0.0721403928  0.0086491222
#>  [42,]  1.055745e+00  4.917377e-02 -0.001484604  0.0215271596 -0.0103846712
#>  [43,]  9.563664e-01  2.694232e-02  0.015092222 -0.0554927870 -0.0316298332
#>  [44,]  1.044703e+00 -2.660902e-02  0.031483043 -0.0363353259  0.0449931754
#>  [45,]  9.996213e-01  1.430640e-01  0.021346997  0.0218734170 -0.0336844461
#>  [46,]  9.879833e-01 -4.939372e-03 -0.058050457 -0.0281062325  0.0295185160
#>  [47,]  9.615116e-01  2.560870e-02 -0.112381996  0.0436388727  0.0109642183
#>  [48,]  1.087455e+00  4.497203e-02 -0.023556520  0.0004430964  0.0450857915
#>  [49,]  1.015554e+00  6.906215e-02 -0.075606058 -0.0517023393  0.0418632961
#>  [50,]  9.683398e-01 -1.355888e-02  0.052259106 -0.0239929554  0.0161078225
#>  [51,]  9.501099e-01  2.089700e-02 -0.056323582 -0.0793583652 -0.0408706952
#>  [52,]  1.014604e+00  9.478986e-03 -0.030503620  0.0022711592 -0.0379863535
#>  [53,]  9.770445e-01 -1.057155e-01  0.019974736  0.0576619939 -0.0256054351
#>  [54,]  9.764408e-01  3.503143e-02  0.029945581 -0.0105580563  0.0403680203
#>  [55,]  1.066064e+00 -5.941855e-02  0.023029248  0.0669891798  0.0322701490
#>  [56,]  9.791256e-01 -5.974826e-02  0.022273538  0.0184385562 -0.0094683617
#>  [57,]  1.045529e+00  1.062824e-02  0.026994233 -0.0288864001  0.0421421865
#>  [58,]  9.135719e-01  2.059069e-02  0.068615734  0.0549376726 -0.0435855846
#>  [59,]  1.034612e+00  3.685297e-02  0.071647269  0.0684048054 -0.0016118542
#>  [60,]  9.852723e-01  1.088140e-01 -0.057633372 -0.0201061646 -0.0311812602
#>  [61,] -1.404679e-02  1.071320e+00  0.071022677 -0.0115637086 -0.0317744627
#>  [62,] -2.482471e-02  9.507502e-01  0.018140348  0.0018069272  0.0283873527
#>  [63,] -7.991271e-03  9.455399e-01  0.003201641 -0.0732612526  0.0355877011
#>  [64,]  7.301485e-02  9.708355e-01  0.046061964  0.0930701099 -0.0408561724
#>  [65,] -4.138949e-02  9.828116e-01 -0.059434387  0.0395176597  0.0192152658
#>  [66,] -8.144664e-02  1.044193e+00  0.027556799 -0.0153592070  0.0053714332
#>  [67,] -3.432438e-02  1.079028e+00 -0.099458361 -0.0458528881 -0.0471178949
#>  [68,] -2.106250e-02  9.764330e-01  0.060254802 -0.0176601389 -0.0151959421
#>  [69,] -3.329283e-02  1.041406e+00 -0.024419993 -0.0897898832 -0.0242929839
#>  [70,]  7.882760e-03  9.169890e-01 -0.021741924 -0.0493191545  0.0023604543
#>  [71,]  6.417091e-02  1.011793e+00 -0.073823758  0.0596840476 -0.0336597935
#>  [72,]  1.126119e-01  9.489572e-01  0.013625051 -0.0359978249  0.0292309778
#>  [73,]  5.688297e-02  1.044005e+00  0.034685279  0.1297735614 -0.0048298348
#>  [74,]  8.103603e-02  9.686354e-01  0.142031410  0.0270262188  0.0458123848
#>  [75,]  1.430033e-01  9.762041e-01 -0.060994720 -0.0380039631  0.0315432504
#>  [76,]  4.370683e-02  1.017541e+00  0.062136843 -0.0285369116  0.0143149123
#>  [77,]  6.315710e-02  1.069409e+00  0.090740723 -0.0350094459  0.0359524664
#>  [78,] -6.236192e-02  1.009733e+00 -0.027258199  0.0023445357 -0.0309190253
#>  [79,]  2.262419e-02  1.015253e+00  0.050340282  0.0088209636 -0.0276809559
#>  [80,]  1.795000e-02  1.054030e+00  0.044172302  0.0334931971 -0.0186468510
#>  [81,]  1.989532e-02  1.042008e+00  0.012784212 -0.0379221147  0.0025774861
#>  [82,] -8.920482e-03  1.050093e+00  0.020571342 -0.0187804575  0.0385901623
#>  [83,] -7.448073e-02  9.422825e-01  0.012020095  0.0135785867 -0.0419593446
#>  [84,] -4.742426e-02  1.083824e+00  0.047655523  0.0070850918  0.0003885070
#>  [85,] -4.016994e-02  9.716700e-01 -0.048529892 -0.0666469222  0.0186002290
#>  [86,] -9.678213e-03  1.061643e+00  0.061771298  0.0502261081  0.0296849377
#>  [87,]  5.407201e-02  9.733461e-01 -0.004071969 -0.0345455839 -0.0478539778
#>  [88,] -4.322187e-02  9.787008e-01 -0.077566255 -0.0496747040 -0.0312170489
#>  [89,] -6.371242e-02  1.028192e+00 -0.059431106 -0.0288484460  0.0116718386
#>  [90,]  2.550981e-02  9.366358e-01 -0.024095397  0.0105587761  0.0491407896
#>  [91,] -2.517884e-02  9.913367e-01 -0.043859345 -0.0722915526  0.0436075462
#>  [92,]  5.359646e-02  1.055006e+00 -0.059111941  0.0010200977 -0.0343158830
#>  [93,]  3.326932e-02  9.873188e-01 -0.046946099  0.0053971947 -0.0264792681
#>  [94,]  3.878983e-02  9.565274e-01 -0.071567671  0.0125488018 -0.0195171470
#>  [95,] -6.662070e-02  9.954391e-01  0.030506652  0.0234515423  0.0446470994
#>  [96,] -7.814284e-02  9.434974e-01 -0.021277412  0.0359377758  0.0061668147
#>  [97,] -6.256933e-02  9.105657e-01  0.013371673  0.0227906466  0.0266115719
#>  [98,]  2.444078e-02  1.025721e+00 -0.003382534  0.0097610078 -0.0087039553
#>  [99,] -9.647690e-03  1.032885e+00  0.027175337  0.0232228949 -0.0263440029
#> [100,] -4.601015e-02  9.954213e-01 -0.016649081 -0.1026469254 -0.0331091805
#> [101,]  3.168728e-03  1.024180e+00  0.012820450 -0.0473838424  0.0422211939
#> [102,]  2.473857e-02  1.015870e+00  0.003077915 -0.0697735191 -0.0315983411
#> [103,]  3.112527e-02  1.071093e+00 -0.059854913 -0.0462527250  0.0052642121
#> [104,]  3.951170e-02  9.683598e-01 -0.087381294 -0.0410862670  0.0454919790
#> [105,] -5.530448e-02  1.032353e+00  0.030610406 -0.0764324079  0.0093741347
#> [106,] -5.177414e-02  1.059351e+00  0.050899999 -0.0191696517 -0.0093655701
#> [107,] -2.932732e-02  1.089793e+00 -0.141585798  0.0266554238  0.0419065030
#> [108,]  4.669760e-02  9.505704e-01 -0.119531755 -0.0215109525 -0.0107712290
#> [109,] -2.861672e-02  9.709549e-01  0.059149173 -0.0349928424  0.0382439618
#> [110,] -9.117097e-02  9.764843e-01 -0.094730598  0.0066727121  0.0274786165
#> [111,] -4.906415e-02  9.818971e-01  0.067546751 -0.0322386892  0.0455234236
#> [112,] -8.832277e-02  1.058231e+00  0.081469705 -0.0062782443 -0.0443170598
#> [113,] -7.746377e-02  1.029500e+00 -0.073583151  0.0577860981 -0.0325108052
#> [114,] -6.988675e-02  9.687376e-01 -0.033105998 -0.0218670105 -0.0283417728
#> [115,]  5.072506e-02  9.815237e-01 -0.096954067 -0.0034117833 -0.0423423398
#> [116,] -3.823305e-02  1.014300e+00 -0.066648908  0.0161519928 -0.0462612609
#> [117,]  2.332887e-02  9.614096e-01  0.108053094  0.0729481583  0.0481902373
#> [118,] -7.594761e-02  1.095070e+00  0.022029199  0.0044385187 -0.0367731982
#> [119,] -6.524894e-02  1.033973e+00 -0.022488929 -0.0531785375 -0.0409848387
#> [120,] -5.558886e-02  1.116629e+00  0.004812795  0.0088575461  0.0220249811
#> [121,] -5.439330e-02 -2.904947e-02  0.992133270  0.0056383454 -0.0202425378
#> [122,] -5.353079e-02  2.014852e-03  1.088125167  0.0242716868  0.0121788213
#> [123,]  2.439506e-02 -4.041357e-02  1.006894869  0.0160429490 -0.0068942426
#> [124,]  4.982097e-02 -1.251761e-02  0.956071418 -0.0230194759  0.0307231782
#> [125,]  5.025494e-02  4.238976e-02  1.040319776 -0.0397145376  0.0217060344
#> [126,] -4.371259e-02  1.541714e-02  1.094592014 -0.0350346881 -0.0187386144
#> [127,]  6.745822e-02  2.977251e-02  1.041560064  0.0362520600  0.0085297864
#> [128,] -8.138126e-02 -1.087678e-02  1.009239397 -0.0742789690 -0.0342787080
#> [129,]  4.882022e-02  3.064535e-02  0.972075629 -0.0154030329 -0.0130825861
#> [130,] -3.437824e-02  6.851839e-02  0.981399421  0.0077907664 -0.0296558761
#> [131,]  1.109899e-01 -3.513248e-02  1.073218990  0.0186078459 -0.0087859962
#> [132,] -3.041284e-02  7.981891e-02  1.055222022 -0.0389071362 -0.0067248370
#> [133,]  5.163125e-02 -8.484606e-02  1.028950753 -0.0064397067 -0.0195859058
#> [134,]  5.972390e-03 -1.589652e-02  1.040060077  0.0402337558 -0.0140192374
#> [135,]  4.564006e-02  2.005788e-02  1.031701166 -0.0217850857  0.0276155784
#> [136,]  2.401332e-02 -5.399579e-02  1.017978964  0.0380767587 -0.0496757670
#> [137,]  3.717132e-02 -5.373481e-02  0.968251014  0.0821087963  0.0077413448
#> [138,] -4.993533e-02 -1.098516e-01  1.006850113  0.0198603976  0.0181666031
#> [139,] -1.470553e-02  4.399873e-02  1.104219033 -0.0179246870 -0.0222881468
#> [140,]  3.028521e-02 -2.385073e-02  0.889154723 -0.0783303824 -0.0250797663
#> [141,] -2.452089e-02 -1.530007e-02  0.994264854 -0.0198099695 -0.0416851049
#> [142,]  6.693719e-03  7.628935e-03  0.974975676  0.0005067693 -0.0041878419
#> [143,]  5.015304e-02  2.745494e-02  1.026585064  0.0088430929 -0.0210783483
#> [144,] -7.661652e-03 -6.038015e-02  1.019087505 -0.1017744513 -0.0210225819
#> [145,]  1.468848e-02  7.134090e-02  0.925457554 -0.0865261489  0.0195216516
#> [146,]  6.328190e-03 -2.956168e-02  1.056703346 -0.0115226526 -0.0365325848
#> [147,]  2.262804e-02  3.120253e-02  0.898642268  0.0540672223 -0.0318755670
#> [148,]  5.676704e-02  3.863884e-02  1.071243439  0.0047384305 -0.0255422524
#> [149,]  2.837036e-02  9.060712e-02  0.898235458 -0.0142564201  0.0324034995
#> [150,] -9.977014e-02  8.586908e-02  1.134106925  0.0445054251  0.0222529342
#> [151,]  1.062357e-02 -4.848236e-02  0.933983442 -0.0723746521  0.0125231730
#> [152,]  7.500273e-02 -5.176670e-02  0.935272024 -0.0345037859 -0.0305580688
#> [153,]  5.838374e-03 -3.827191e-02  1.025388065  0.0345056912 -0.0230701996
#> [154,]  6.950161e-02  1.232583e-01  0.937180080  0.0651091996 -0.0334526993
#> [155,] -3.154082e-02  4.310128e-02  1.022382306  0.0225334434  0.0363401029
#> [156,]  5.529703e-02 -5.995251e-02  1.065723089 -0.0454559297  0.0047297754
#> [157,] -1.142571e-02 -8.492666e-02  0.994256585  0.0514806216  0.0262276742
#> [158,] -5.399920e-03  2.010555e-04  1.053520580  0.0652780605  0.0063406537
#> [159,] -2.187578e-02  5.782868e-02  1.040334372 -0.0382856053 -0.0356266348
#> [160,]  1.005093e-01  8.615330e-03  0.947689160  0.0364983852 -0.0054299544
#> [161,] -2.314528e-02  1.775168e-02  1.115810316 -0.0287697924  0.0198119192
#> [162,] -7.548377e-02 -2.604965e-02  0.904693265 -0.0013375131  0.0187066480
#> [163,] -1.368746e-02 -5.098503e-02  0.997002993 -0.0166307001  0.0266646778
#> [164,] -3.038485e-02  9.587830e-02  0.993543218  0.0238510182 -0.0155541293
#> [165,]  1.746846e-02 -2.163558e-02  0.898459327  0.0038823381  0.0103463094
#> [166,]  3.895018e-02  4.430548e-02  0.993139422  0.0172332318  0.0243805559
#> [167,]  7.689784e-02 -2.419413e-02  0.950496999  0.0667778867  0.0022564473
#> [168,]  5.355672e-02  6.317491e-02  1.006009199 -0.0053782841 -0.0147251505
#> [169,] -2.621420e-02  7.147334e-02  1.046424602 -0.0450851719  0.0023531646
#> [170,]  2.189296e-02  2.350628e-02  0.948637622 -0.1166821848 -0.0141204971
#> [171,]  1.097076e-01  4.483503e-02  1.031040921 -0.0415092363  0.0222251430
#> [172,] -7.081844e-03 -8.139374e-02  1.040679047  0.0889069272 -0.0424317487
#> [173,]  3.523279e-02 -1.379264e-02  0.984440356  0.0186638580  0.0273736841
#> [174,] -2.646456e-02  1.685363e-02  1.023541949 -0.0246225426 -0.0093505156
#> [175,]  4.257794e-02 -1.621866e-01  0.992730268  0.0702001391 -0.0136715992
#> [176,]  7.865294e-02  3.847925e-02  1.077999404 -0.0454359365 -0.0163987746
#> [177,]  2.951921e-02 -6.620630e-02  0.949011415  0.0357406026  0.0420799294
#> [178,] -1.646297e-02 -3.237755e-02  1.034321804 -0.0918634846  0.0427381800
#> [179,]  2.620223e-02 -1.556966e-02  0.999584335 -0.0787017013 -0.0356263854
#> [180,] -4.880291e-02  4.651519e-02  1.078884571 -0.0242121972 -0.0174747279
#> [181,]  7.680291e-02  4.731849e-02  0.006500028  0.9321472764  0.0082457735
#> [182,] -4.474467e-03 -2.936473e-02  0.026132387  0.9276357912  0.0017619994
#> [183,]  5.355738e-02 -6.278718e-02  0.060667832  0.9692158696  0.0152316551
#> [184,]  3.513423e-02  7.229771e-02 -0.037616523  0.9435628392 -0.0154368999
#> [185,] -1.404724e-02  1.834060e-02  0.017222189  1.0023711274  0.0435169203
#> [186,] -5.854322e-02 -4.798564e-02  0.065176556  0.9805935668 -0.0174501113
#> [187,]  6.777906e-03 -4.881480e-02 -0.028234492  0.9957012820  0.0032279018
#> [188,] -5.667419e-03  9.050711e-02  0.033125518  0.9762385957  0.0372528795
#> [189,]  3.115096e-02 -1.071612e-01 -0.015776349  0.9982530177  0.0076119278
#> [190,] -2.270491e-02  6.861678e-02 -0.043919202  0.9211221391 -0.0187623257
#> [191,]  8.212885e-02 -7.698832e-02 -0.053847986  0.9214616755 -0.0177276352
#> [192,] -2.391387e-02  4.011983e-02  0.003838076  1.0042791663  0.0223837977
#> [193,] -1.247279e-02  1.027020e-01  0.024918599  0.8977376772  0.0371830637
#> [194,] -3.448340e-02  5.714744e-03 -0.076842273  1.0207340153 -0.0045476012
#> [195,] -2.331378e-02  7.074237e-02  0.029188378  1.0401636961 -0.0005303696
#> [196,] -1.333130e-01 -6.515296e-02  0.023041379  1.0677707153  0.0052830201
#> [197,]  4.668989e-02 -1.624320e-02 -0.027231395  1.0417354823 -0.0143382676
#> [198,]  6.486551e-03 -4.168339e-02  0.024475880  1.0107606390  0.0105669356
#> [199,] -3.250978e-02  8.987848e-02 -0.021146632  0.9742013018  0.0176092945
#> [200,]  3.814057e-02  3.683713e-02  0.032961281  0.9162233352  0.0153366521
#> [201,]  5.894569e-02  1.382800e-02  0.037918696  0.9943538362 -0.0225903200
#> [202,] -6.081092e-02  5.337515e-02  0.051849229  0.9550362334 -0.0222332434
#> [203,] -8.179547e-02 -3.767089e-02  0.035243440  0.9347608681 -0.0181774315
#> [204,] -2.830524e-02 -7.002879e-02  0.001159915  1.0033592485  0.0016266773
#> [205,] -8.707425e-02  1.480121e-03  0.020753103  1.1046016709  0.0445976659
#> [206,] -1.615362e-02  3.084451e-02 -0.089708646  0.9707556739 -0.0230915554
#> [207,]  6.322069e-02 -6.380834e-03 -0.089938402  0.9006356800 -0.0328565649
#> [208,] -7.460424e-02  5.400486e-02 -0.001317745  0.9568429952  0.0237467542
#> [209,]  1.174369e-02 -3.004414e-03 -0.033448099  1.0930545879 -0.0182747348
#> [210,] -1.535774e-02 -6.084941e-02  0.035914870  1.0226374265 -0.0154779445
#> [211,] -8.551288e-02  1.422663e-02  0.040841195  0.9648930996  0.0138439229
#> [212,] -4.681899e-02  7.048120e-02  0.060879511  1.0000012975  0.0490239524
#> [213,]  6.456396e-02  2.029737e-02  0.136762796  1.0259853310 -0.0118590961
#> [214,]  4.199036e-02  5.148546e-02  0.005795314  1.0446402611 -0.0092767853
#> [215,]  1.781612e-02 -9.347645e-02  0.112783303  1.0053389511 -0.0440224341
#> [216,]  3.051735e-02 -4.831409e-02 -0.043001708  0.9935897337  0.0303186050
#> [217,]  1.313692e-02  1.892232e-02 -0.021967978  1.0300154377 -0.0360515487
#> [218,]  4.699315e-02 -1.772302e-02  0.043694655  1.0213545143  0.0052421737
#> [219,] -8.525624e-04 -8.946734e-03 -0.018461997  0.9347619424  0.0132511463
#> [220,] -2.654785e-02 -1.264798e-01  0.056647671  1.1179325144  0.0322365890
#> [221,]  1.458491e-02  1.053677e-01  0.048817034  0.9572983057 -0.0389746529
#> [222,]  3.425518e-02 -2.996827e-02 -0.014054329  1.0290675318 -0.0233045466
#> [223,]  3.671032e-03  5.809672e-02 -0.049556039  0.9992121660 -0.0448459361
#> [224,]  9.258902e-02 -6.917924e-02  0.023705542  0.9236376700 -0.0061127545
#> [225,]  1.202405e-01 -1.439253e-03 -0.008204308  1.0281671807  0.0269400091
#> [226,]  1.620481e-02  8.874528e-03  0.083456679  1.0855345225 -0.0459554137
#> [227,]  4.639101e-02 -4.812345e-02 -0.066917793  1.0005364848  0.0384784768
#> [228,]  3.137575e-02 -1.992793e-02  0.015104041  1.1051193788 -0.0184362920
#> [229,] -5.786785e-02  2.648864e-02 -0.120127818  0.9396445587  0.0368400267
#> [230,] -2.320078e-02  9.010458e-02  0.061714274  0.9818221519  0.0087757826
#> [231,] -3.825227e-02 -7.801763e-02  0.030974799  0.9617468149 -0.0142939562
#> [232,] -3.674249e-02  5.160177e-02  0.053725983  1.0070721743  0.0343425299
#> [233,] -8.065248e-03 -9.199247e-03 -0.039722671  0.9766309500  0.0460298492
#> [234,] -2.785978e-02 -2.471524e-02 -0.057033560  1.0307824025 -0.0454177026
#> [235,]  5.678362e-02  5.049471e-02  0.068513391  1.0259519816  0.0245877968
#> [236,] -9.516918e-02  4.025687e-04  0.021527632  1.0057862191 -0.0329071053
#> [237,]  2.829190e-02  5.871070e-03  0.031952263  1.0435016057  0.0486597701
#> [238,] -1.459151e-01  3.711791e-04  0.050689890  1.0266452837 -0.0096877526
#> [239,] -6.949615e-02 -2.047346e-02  0.033282480  0.9454770770 -0.0022689855
#> [240,]  1.414396e-02  8.509616e-02 -0.018424498  0.9342168055 -0.0114620452
#> [241,] -8.910479e-02 -3.405938e-02 -0.020393007 -0.0102303451 -0.0304706511
#> [242,]  1.956995e-02 -6.046996e-02 -0.064495776 -0.0303460388 -0.0053492002
#> [243,] -5.395147e-02  1.903896e-02  0.081931985 -0.1313420691  0.0468809475
#> [244,] -9.358150e-02  3.629355e-02 -0.104046038 -0.0614619209 -0.0256173287
#> [245,]  1.356139e-02 -3.378786e-03  0.013524387 -0.0053053191 -0.0216864800
#> [246,]  6.538722e-02 -5.349591e-02 -0.027792767 -0.0516180816 -0.0493982766
#> [247,] -1.554934e-02  3.717515e-02 -0.009425225 -0.0470468457 -0.0009889928
#> [248,] -2.529796e-02 -3.069089e-02 -0.078015572 -0.0527940348 -0.0409229299
#> [249,] -2.108103e-02  3.709712e-02  0.024113972 -0.1267070697  0.0201006942
#> [250,] -1.295124e-03  1.356985e-01  0.009124548 -0.0466537728 -0.0275514362
#> [251,] -5.315610e-03 -1.476208e-02  0.069425778  0.0368374858  0.0416743542
#> [252,]  4.397476e-02  4.496651e-02  0.023030768 -0.0676345359  0.0317820466
#> [253,]  3.132315e-03  4.446043e-02 -0.055852406 -0.0043519913  0.0112101452
#> [254,]  4.376295e-02 -2.621917e-02 -0.039291135  0.1169637941 -0.0087697486
#> [255,] -1.000866e-02  1.050985e-01 -0.029523560  0.0152259979 -0.0105108654
#> [256,] -2.531709e-02 -1.829235e-02 -0.042386223 -0.0164222162 -0.0002496808
#> [257,] -8.229471e-03  4.131094e-02 -0.011306674  0.0353683687 -0.0277831663
#> [258,]  2.911283e-02  8.115820e-02 -0.015580315  0.0910682903 -0.0003906720
#> [259,]  6.200991e-02  1.349757e-02 -0.135847208 -0.0434038101  0.0195604800
#> [260,]  1.881034e-02  4.059467e-03 -0.012774668  0.0738664124 -0.0239886094
#> [261,] -1.088074e-02  1.039106e-02  0.009792007  0.0313642148 -0.0050562901
#> [262,] -1.514501e-02 -1.483862e-02 -0.094472825  0.0311427203 -0.0378001003
#> [263,] -4.055210e-02 -6.889766e-02 -0.024024587  0.0086996774  0.0367144670
#> [264,]  9.334168e-02  1.044734e-02  0.017546467 -0.0989056124 -0.0101272510
#> [265,] -1.308026e-02 -7.464007e-02  0.014856660 -0.0276248594  0.0098042526
#> [266,] -2.940206e-04  1.887065e-02  0.033417710  0.0326027606  0.0333451068
#> [267,]  1.413225e-02 -2.090243e-02 -0.026053857 -0.0206163480 -0.0067767795
#> [268,]  2.070643e-02  5.284797e-02 -0.030413959 -0.0437271613  0.0173167662
#> [269,] -3.279963e-02  1.658096e-04 -0.031991587 -0.0066083521 -0.0069795073
#> [270,] -4.070068e-03  9.246388e-03 -0.035001098  0.0664406130  0.0008196496
#> [271,]  4.710670e-02 -4.444100e-04  0.039513488  0.0447594944 -0.0326340943
#> [272,] -2.771292e-02  2.977949e-02  0.018202806 -0.0054689777 -0.0390625400
#> [273,] -6.233749e-03  7.264066e-04  0.071189728 -0.0145608325 -0.0097625059
#> [274,]  5.022399e-02  1.245511e-02  0.051334880 -0.0823732990 -0.0448884657
#> [275,] -5.138476e-02  6.711249e-03  0.062392140 -0.0243220666  0.0088589772
#> [276,]  2.371364e-02  3.916363e-02 -0.004666369  0.0120275460 -0.0206375017
#> [277,] -3.509743e-02 -1.866318e-02 -0.034246323 -0.0940780695  0.0462385566
#> [278,] -1.387688e-02  5.271935e-02 -0.073836865  0.0234836152  0.0439026143
#> [279,]  3.027372e-02 -4.324621e-02 -0.026854792 -0.0057964974  0.0229316475
#> [280,]  1.630715e-02  1.238472e-02 -0.034974444 -0.0053314321 -0.0398464252
#> [281,]  4.529611e-05  1.776670e-02  0.013277788 -0.0235930275  0.0114959900
#> [282,] -9.299589e-03  3.095123e-02  0.073290535 -0.0412704710  0.0492539845
#> [283,]  4.263159e-02  4.604559e-02  0.018804525  0.0754067026 -0.0207321491
#> [284,]  2.977640e-02 -5.143501e-02 -0.084072612  0.0810328999 -0.0200072714
#> [285,] -4.587079e-02 -1.001470e-01 -0.035697811  0.0043085299 -0.0152252437
#> [286,]  7.708054e-03 -3.074016e-02  0.018905540  0.0923526307  0.0303547882
#> [287,] -1.773734e-02  1.893038e-02  0.015899463  0.0363869461 -0.0312556549
#> [288,]  1.209424e-02  3.705338e-02  0.023427998  0.0269941730 -0.0249376538
#> [289,] -9.090383e-03 -8.432559e-02 -0.011323369  0.0534871000  0.0467580548
#> [290,] -5.887823e-02 -5.886942e-02 -0.150334032  0.0541176717 -0.0349366278
#> [291,]  4.608625e-02 -5.376805e-02  0.004341438  0.0107270840  0.0398311000
#> [292,] -2.354559e-02 -6.945583e-02  0.024206066 -0.0519656951 -0.0180907781
#> [293,]  1.034329e-01 -1.450846e-01  0.057547200  0.0744301716  0.0310856001
#> [294,] -2.572789e-02 -3.642133e-05 -0.021272395 -0.0941396655  0.0421827446
#> [295,]  2.413407e-03  4.684612e-02  0.020574087  0.0224670714 -0.0036943723
#> [296,] -2.795592e-02 -6.951729e-02  0.006283277 -0.0416301041 -0.0338218742
#> [297,] -1.900842e-02  4.128185e-03 -0.059999995 -0.0032332465 -0.0153908504
#> [298,] -4.532163e-02  6.629984e-04  0.022806494 -0.0006229115  0.0142070127
#> [299,]  3.119726e-02 -4.506773e-02 -0.057119856 -0.0187585638  0.0182914285
#> [300,] -1.856837e-02 -2.415694e-02  0.032047320  0.0873019064 -0.0158174434
#>                 [,6]
#>   [1,]  4.433065e-02
#>   [2,] -5.159993e-03
#>   [3,]  2.901371e-02
#>   [4,] -3.223141e-02
#>   [5,]  1.842841e-02
#>   [6,]  4.403040e-04
#>   [7,] -2.312806e-02
#>   [8,]  2.717376e-02
#>   [9,]  4.667927e-02
#>  [10,]  2.092192e-02
#>  [11,] -2.812600e-02
#>  [12,]  4.427140e-02
#>  [13,]  2.996948e-02
#>  [14,] -4.255319e-02
#>  [15,] -4.591559e-02
#>  [16,] -3.010941e-02
#>  [17,]  1.014308e-02
#>  [18,]  1.362999e-02
#>  [19,]  2.881875e-03
#>  [20,] -1.328655e-02
#>  [21,]  2.468257e-02
#>  [22,]  3.288870e-02
#>  [23,]  1.244350e-02
#>  [24,] -3.556471e-02
#>  [25,]  3.540252e-02
#>  [26,] -4.737794e-02
#>  [27,] -4.785855e-02
#>  [28,]  4.792521e-02
#>  [29,] -1.288630e-02
#>  [30,]  4.216332e-02
#>  [31,] -1.250278e-02
#>  [32,] -3.138082e-02
#>  [33,] -1.864859e-02
#>  [34,]  3.568861e-02
#>  [35,] -4.488816e-02
#>  [36,]  4.409060e-02
#>  [37,] -2.063931e-02
#>  [38,]  1.760913e-02
#>  [39,]  1.756106e-02
#>  [40,]  3.512731e-02
#>  [41,]  1.650943e-02
#>  [42,]  4.399430e-03
#>  [43,] -4.930807e-02
#>  [44,]  2.089905e-02
#>  [45,]  3.630027e-02
#>  [46,] -2.922415e-02
#>  [47,] -3.058883e-02
#>  [48,]  4.231600e-02
#>  [49,]  9.227199e-05
#>  [50,] -3.810721e-02
#>  [51,] -1.765522e-02
#>  [52,] -3.893825e-02
#>  [53,] -2.256328e-02
#>  [54,]  3.632050e-02
#>  [55,] -4.600038e-02
#>  [56,] -3.797037e-02
#>  [57,] -4.341467e-02
#>  [58,]  9.860106e-03
#>  [59,]  6.518575e-03
#>  [60,]  3.131913e-02
#>  [61,]  4.145770e-02
#>  [62,] -4.459951e-02
#>  [63,]  2.317101e-02
#>  [64,]  4.205974e-02
#>  [65,] -2.672862e-03
#>  [66,] -4.032025e-02
#>  [67,]  1.285056e-02
#>  [68,]  1.709759e-02
#>  [69,] -1.414097e-02
#>  [70,] -8.517538e-03
#>  [71,] -5.966907e-03
#>  [72,]  3.348067e-02
#>  [73,] -2.500628e-02
#>  [74,] -1.689432e-02
#>  [75,] -5.078914e-03
#>  [76,] -8.997007e-03
#>  [77,]  2.136832e-03
#>  [78,] -2.561261e-03
#>  [79,] -4.463761e-02
#>  [80,] -9.813611e-03
#>  [81,] -4.651076e-03
#>  [82,] -3.981410e-02
#>  [83,]  3.017325e-02
#>  [84,]  4.302109e-02
#>  [85,] -2.861633e-02
#>  [86,]  3.657772e-02
#>  [87,] -2.963972e-02
#>  [88,] -2.903036e-03
#>  [89,] -2.347300e-02
#>  [90,]  4.694113e-02
#>  [91,]  3.722468e-02
#>  [92,]  1.976659e-02
#>  [93,]  8.979754e-04
#>  [94,] -4.553279e-02
#>  [95,] -4.187595e-02
#>  [96,] -3.155606e-02
#>  [97,]  3.058134e-02
#>  [98,] -1.286222e-02
#>  [99,] -4.287145e-02
#> [100,]  4.467736e-02
#> [101,] -4.615600e-02
#> [102,]  3.428953e-02
#> [103,] -1.660807e-02
#> [104,] -9.639518e-03
#> [105,]  2.778713e-02
#> [106,] -2.078307e-02
#> [107,]  3.780633e-02
#> [108,] -2.148886e-02
#> [109,] -2.000228e-03
#> [110,]  1.894506e-02
#> [111,]  4.749640e-02
#> [112,]  1.473222e-02
#> [113,]  2.534874e-03
#> [114,]  3.266231e-02
#> [115,]  4.868516e-02
#> [116,] -9.759634e-03
#> [117,]  2.429749e-02
#> [118,] -1.228189e-02
#> [119,] -4.245628e-02
#> [120,] -2.766666e-02
#> [121,] -3.604310e-02
#> [122,] -4.371054e-02
#> [123,] -4.457245e-02
#> [124,] -4.994426e-02
#> [125,] -2.479857e-02
#> [126,] -4.653036e-03
#> [127,] -1.026872e-03
#> [128,]  2.112033e-02
#> [129,]  4.499602e-02
#> [130,]  4.741891e-02
#> [131,] -4.164700e-02
#> [132,]  9.491201e-03
#> [133,]  3.609727e-02
#> [134,]  3.494421e-02
#> [135,] -4.316373e-02
#> [136,]  3.439609e-02
#> [137,] -4.077935e-02
#> [138,] -1.124647e-02
#> [139,]  3.537356e-02
#> [140,] -6.665141e-03
#> [141,] -3.639664e-02
#> [142,]  3.798521e-02
#> [143,] -2.946943e-02
#> [144,] -5.633454e-03
#> [145,] -3.997359e-02
#> [146,]  4.323496e-02
#> [147,] -3.343178e-02
#> [148,] -2.099256e-02
#> [149,]  3.258637e-02
#> [150,] -5.466797e-03
#> [151,] -2.585597e-02
#> [152,]  3.265164e-02
#> [153,] -2.499320e-02
#> [154,]  4.272568e-02
#> [155,]  9.974741e-03
#> [156,] -3.590610e-02
#> [157,] -3.130817e-02
#> [158,] -1.188488e-02
#> [159,] -4.629284e-02
#> [160,]  4.607664e-02
#> [161,]  1.959681e-02
#> [162,] -1.339523e-02
#> [163,] -2.991836e-02
#> [164,]  3.272018e-02
#> [165,] -2.987723e-02
#> [166,] -5.107522e-03
#> [167,]  3.216588e-02
#> [168,]  4.260120e-02
#> [169,] -1.133000e-02
#> [170,]  2.278921e-02
#> [171,] -1.736960e-03
#> [172,]  3.849788e-02
#> [173,]  2.529639e-02
#> [174,] -2.347301e-02
#> [175,] -3.101763e-02
#> [176,] -4.938802e-02
#> [177,]  1.551646e-02
#> [178,] -1.195475e-02
#> [179,]  3.988212e-02
#> [180,] -1.544144e-02
#> [181,] -1.584531e-02
#> [182,] -4.931070e-02
#> [183,] -2.374926e-02
#> [184,] -2.587860e-02
#> [185,]  6.282758e-04
#> [186,]  1.997064e-02
#> [187,]  2.325103e-03
#> [188,]  4.356470e-02
#> [189,]  1.889095e-03
#> [190,]  2.788533e-03
#> [191,] -1.169739e-02
#> [192,]  1.652146e-02
#> [193,] -2.764994e-02
#> [194,]  3.629148e-02
#> [195,]  2.824238e-02
#> [196,] -4.669556e-02
#> [197,] -1.597479e-02
#> [198,] -3.469667e-02
#> [199,] -4.508864e-02
#> [200,] -3.622665e-03
#> [201,] -2.682889e-02
#> [202,]  5.990647e-04
#> [203,]  5.661189e-03
#> [204,]  1.664990e-03
#> [205,]  1.072171e-02
#> [206,] -1.244846e-02
#> [207,] -1.101373e-02
#> [208,]  8.957047e-04
#> [209,]  3.157125e-02
#> [210,] -3.690969e-02
#> [211,]  4.622525e-03
#> [212,]  3.755400e-02
#> [213,]  4.427587e-03
#> [214,] -4.445058e-02
#> [215,]  3.139449e-02
#> [216,] -2.234362e-02
#> [217,]  1.840624e-02
#> [218,]  4.544628e-02
#> [219,] -4.642792e-02
#> [220,]  3.625105e-02
#> [221,] -7.717851e-03
#> [222,] -4.713201e-02
#> [223,] -5.884550e-03
#> [224,] -4.142111e-02
#> [225,] -3.949016e-02
#> [226,] -3.428738e-03
#> [227,] -3.000965e-02
#> [228,]  3.503348e-02
#> [229,] -2.906261e-02
#> [230,] -5.360450e-04
#> [231,]  2.179153e-02
#> [232,] -1.659577e-02
#> [233,]  1.638775e-02
#> [234,]  2.381165e-02
#> [235,] -3.489622e-02
#> [236,] -1.477963e-03
#> [237,] -6.184042e-03
#> [238,] -2.265914e-02
#> [239,] -2.705488e-02
#> [240,] -2.283133e-02
#> [241,]  3.952214e-02
#> [242,] -4.029729e-02
#> [243,] -4.644743e-02
#> [244,]  1.634146e-02
#> [245,]  4.690740e-02
#> [246,]  8.756434e-03
#> [247,] -2.020635e-02
#> [248,]  2.668230e-02
#> [249,]  3.863528e-02
#> [250,]  2.844607e-02
#> [251,]  2.523269e-02
#> [252,] -2.936419e-02
#> [253,]  3.357595e-02
#> [254,]  7.286758e-03
#> [255,]  3.167629e-02
#> [256,]  1.119003e-02
#> [257,]  4.467077e-02
#> [258,]  2.886254e-02
#> [259,] -3.314809e-02
#> [260,] -2.500731e-03
#> [261,] -1.925436e-02
#> [262,] -3.373951e-02
#> [263,]  3.151793e-02
#> [264,] -2.780764e-02
#> [265,]  1.003331e-02
#> [266,]  4.450276e-03
#> [267,]  4.165815e-02
#> [268,]  2.645295e-02
#> [269,] -1.066512e-02
#> [270,] -3.924495e-02
#> [271,] -3.238748e-02
#> [272,] -4.014463e-02
#> [273,]  4.994269e-03
#> [274,]  3.268213e-02
#> [275,] -2.730252e-02
#> [276,] -3.960486e-02
#> [277,] -3.680838e-02
#> [278,]  1.359392e-02
#> [279,] -2.142654e-02
#> [280,]  1.631026e-02
#> [281,]  2.643821e-02
#> [282,] -2.897991e-02
#> [283,]  1.133365e-02
#> [284,]  4.563436e-02
#> [285,]  3.148666e-02
#> [286,]  2.270839e-02
#> [287,] -4.237534e-02
#> [288,] -2.040541e-02
#> [289,] -6.256820e-03
#> [290,]  9.029177e-03
#> [291,] -2.590173e-02
#> [292,] -4.484322e-02
#> [293,]  4.190617e-02
#> [294,]  3.665605e-02
#> [295,]  2.600894e-02
#> [296,]  4.391216e-02
#> [297,]  9.184317e-03
#> [298,] -3.448573e-02
#> [299,] -2.965565e-02
#> [300,]  4.301496e-02