Generate Random Noise Dimensions

This function generates random noise dimensions to be added to the coordinates of a sphere.

Usage

gen_noise_dims(n, num_noise, min_n, max_n)

Arguments

n

The number of observations for which to generate noise dimensions.

num_noise

The number of noise dimensions to generate.

min_n

The minimum value for the random noise.

max_n

The maximum value for the random noise.

Value

A matrix containing the generated random noise dimensions.

Examples

# Generate random noise dimensions with 3 dimensions, minimum value -1, and maximum value 1
gen_noise_dims(n = 50, num_noise = 3, min_n = -1, max_n = 1)
#>               [,1]        [,2]         [,3]
#>  [1,]  0.002396036  0.83580909 -0.578910731
#>  [2,] -0.893942094  0.60560824 -0.978543148
#>  [3,] -0.006773234 -0.23263103 -0.005156463
#>  [4,] -0.620046671  0.33519619  0.553069955
#>  [5,] -0.990496408 -0.17779533 -0.100971378
#>  [6,] -0.564122605 -0.15107915 -0.723382560
#>  [7,] -0.057381272  0.72405313 -0.371234276
#>  [8,]  0.855687753  0.55978903 -0.918383792
#>  [9,] -0.092927045 -0.90379513  0.976283184
#> [10,]  0.574144050 -0.18551002  0.610187978
#> [11,]  0.226306798  0.17841880  0.734068096
#> [12,]  0.994972476  0.64941131 -0.558916259
#> [13,] -0.957191969  0.42914523 -0.400769622
#> [14,] -0.865665521 -0.30508784 -0.709439183
#> [15,] -0.992885789 -0.86807864 -0.985701002
#> [16,] -0.524235060 -0.93217369  0.973404545
#> [17,] -0.896084989 -0.28154998 -0.733379309
#> [18,]  0.511703780 -0.99092622  0.387226871
#> [19,]  0.294298458 -0.34645846 -0.219253907
#> [20,]  0.656989292  0.61630151 -0.867377997
#> [21,]  0.705824707 -0.05430909 -0.068550198
#> [22,] -0.052570233 -0.23396954 -0.693504897
#> [23,] -0.875697910  0.07975076 -0.685383218
#> [24,] -0.647864141 -0.29577292 -0.250996509
#> [25,]  0.764685934 -0.38494234 -0.089520078
#> [26,]  0.699865836  0.81236465 -0.739066688
#> [27,] -0.752188553 -0.56763611  0.690704722
#> [28,] -0.749718754 -0.95469188  0.806173876
#> [29,] -0.946363891  0.15265126 -0.004774246
#> [30,] -0.017520165  0.47629552 -0.218439861
#> [31,] -0.440296996 -0.66095155  0.823820635
#> [32,]  0.216263321  0.56519747 -0.249742787
#> [33,]  0.279245018  0.63737616  0.536642139
#> [34,] -0.614987599  0.70745020 -0.066155461
#> [35,]  0.317092622  0.10829162  0.118381604
#> [36,]  0.368809975 -0.51582342  0.143609721
#> [37,] -0.378938647  0.86186789  0.985436952
#> [38,]  0.780348534  0.90871256 -0.176852305
#> [39,] -0.198448662  0.87042268  0.692409312
#> [40,] -0.499818436 -0.25680456 -0.930780972
#> [41,] -0.610359496  0.82860208 -0.298139533
#> [42,]  0.281246783  0.42589220  0.246184228
#> [43,] -0.380846267 -0.65938770 -0.241269068
#> [44,]  0.173409197  0.15900421 -0.355427863
#> [45,]  0.610234258 -0.11936039 -0.239042516
#> [46,] -0.813868348 -0.53060927  0.156172215
#> [47,]  0.739265137 -0.82049024  0.318853878
#> [48,]  0.124219352  0.63274310 -0.566952964
#> [49,]  0.758723202  0.19827762  0.226365985
#> [50,]  0.824583707 -0.85891642 -0.987762209