Generate points on a plane in 2D space

This function generates points on a plane in 3D space based on the provided coefficients, intercepts, and ranges for the parameters.

Usage

plane(
  sample_size,
  coefficient_x_1,
  coefficient_x_2,
  coefficient_y_1,
  coefficient_y_2,
  intercept_x,
  intercept_y,
  u_min,
  u_max,
  v_min,
  v_max,
  num_noise_dims,
  min_noise,
  max_noise
)

Arguments

sample_size

The number of points to generate.

coefficient_x_1

The coefficient of the first parameter in the x-dimension equation.

coefficient_x_2

The coefficient of the second parameter in the x-dimension equation.

coefficient_y_1

The coefficient of the first parameter in the y-dimension equation.

coefficient_y_2

The coefficient of the second parameter in the y-dimension equation.

intercept_x

The intercept for the x-dimension equation.

intercept_y

The intercept for the y-dimension equation.

u_min

The minimum value for the first parameter (u) range.

u_max

The maximum value for the first parameter (u) range.

v_min

The minimum value for the second parameter (v) range.

v_max

The maximum value for the second parameter (v) range.

num_noise_dims

The number of noise dimensions to add to the generated points.

min_noise

The minimum value for the noise dimensions.

max_noise

The maximum value for the noise dimensions.

Value

A matrix containing the generated points on the plane.

Examples

plane_points <- plane(sample_size = 100, coefficient_x_1 = 1, coefficient_x_2 = 1,
                     coefficient_y_1 = -1, coefficient_y_2 = 1, intercept_x = -10,
                     intercept_y = 8, u_min = 10, u_max = 30, v_min = 10, v_max = 20,
                     num_noise_dims = 2, min_noise = 0, max_noise = 1)