Generate data points on a Roman surface with optional noise.

This function generates data points on a Roman surface with optional noise.

Usage

roman_surface_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
roman_surface_3d(n = 100, num_noise = 2, min_n = -0.05, max_n = 0.05)
#>                    x            y            z                           
#>   [1,] -1.306904e-02 -0.061479760  0.202577449  0.046811273 -0.0397026339
#>   [2,]  9.455417e-02  0.286017746  0.267938117 -0.012496258  0.0290763175
#>   [3,]  1.963898e-01 -0.141707179 -0.444820895 -0.039036306  0.0422360181
#>   [4,] -3.410296e-03  0.032991970 -0.102162045  0.012114244 -0.0214310022
#>   [5,] -1.353286e-02 -0.162011383  0.080697823 -0.007777032 -0.0499433291
#>   [6,]  1.524588e-02  0.268129074  0.052242111 -0.043311664 -0.0356851794
#>   [7,]  1.410966e-01  0.086468327  0.432959032  0.009618158  0.0015637302
#>   [8,] -3.778011e-02 -0.257409606  0.133184389  0.040446538 -0.0123396746
#>   [9,]  4.340756e-03  0.082636506  0.052022578 -0.049745073 -0.0284339315
#>  [10,] -4.294929e-03 -0.001023880  0.225568717 -0.015265306 -0.0350829806
#>  [11,] -3.503332e-01 -0.267234015  0.127145062  0.036571148 -0.0353439124
#>  [12,]  2.702748e-02  0.096184884  0.002626116  0.015919120  0.0423142502
#>  [13,] -2.723116e-01 -0.250279655  0.081479001 -0.031443569 -0.0051962465
#>  [14,]  8.627048e-02  0.201083149  0.018268628  0.044940477 -0.0389469245
#>  [15,]  1.058617e-01 -0.341822671 -0.239982428 -0.028938895 -0.0086575911
#>  [16,] -3.177041e-01  0.362458950 -0.181920406  0.013595432  0.0232040592
#>  [17,]  3.164463e-01  0.264800864  0.107102099  0.002368991 -0.0245691278
#>  [18,] -3.711852e-01  0.036023883 -0.016051251 -0.030572272 -0.0315889175
#>  [19,] -2.616431e-01  0.425691020 -0.231157200 -0.018436977  0.0227344875
#>  [20,] -3.369538e-01  0.201071157 -0.356383637  0.045953311 -0.0366494890
#>  [21,] -1.402386e-01  0.140172058 -0.479502083  0.034806118  0.0157848667
#>  [22,]  1.542301e-03 -0.028997783 -0.052992069  0.004200594  0.0070135517
#>  [23,]  4.798042e-03  0.156381651  0.029883114 -0.030929957  0.0208199039
#>  [24,]  1.944749e-02 -0.073354659 -0.246270760 -0.023326323 -0.0189224581
#>  [25,] -2.444079e-02 -0.224101599  0.101991176  0.038856192  0.0338216087
#>  [26,] -3.744296e-01 -0.201021601  0.098593090  0.036005257 -0.0460992394
#>  [27,] -9.483536e-02  0.455727976 -0.063321673 -0.005547783  0.0447378346
#>  [28,] -1.555794e-01 -0.242244181  0.041470547  0.010797826  0.0272497393
#>  [29,]  4.054346e-01 -0.187346194 -0.275981004  0.029635492 -0.0160971409
#>  [30,] -3.056012e-01 -0.011605060  0.033959753  0.009967849  0.0464448324
#>  [31,]  1.880595e-01  0.385929272  0.095963473 -0.006457079  0.0179349351
#>  [32,]  3.604144e-01  0.225893639  0.343261555 -0.030954550 -0.0207382537
#>  [33,] -6.348793e-02 -0.333834879  0.159087348 -0.025088984  0.0037443788
#>  [34,] -4.839375e-01 -0.125346203  0.123703604  0.031934072 -0.0118583800
#>  [35,] -2.993626e-01  0.177888309 -0.062008709 -0.035403466  0.0063740385
#>  [36,]  1.680193e-01 -0.424341037 -0.241045789 -0.042859625 -0.0297865495
#>  [37,]  1.303370e-01 -0.481883646 -0.133148518 -0.034428535 -0.0433482581
#>  [38,] -3.824052e-01 -0.136053148  0.250115015  0.034162961 -0.0163532625
#>  [39,] -4.395125e-01  0.226251984 -0.172912388 -0.045406635  0.0359919606
#>  [40,]  2.927428e-01  0.271915803  0.399214894  0.012260502 -0.0370901504
#>  [41,]  4.394053e-01  0.237752837  0.217606460  0.002252682 -0.0496382012
#>  [42,]  1.525665e-01 -0.162937322 -0.472681993  0.004865140  0.0114451926
#>  [43,] -7.165520e-02 -0.034859947  0.390875169  0.037104514 -0.0157686300
#>  [44,] -2.167561e-01  0.141821509 -0.425018425  0.019189041 -0.0482725061
#>  [45,]  5.567480e-02 -0.123064995 -0.368559325  0.016947191  0.0417446218
#>  [46,] -3.185190e-03 -0.049502410  0.063920783  0.023474528  0.0490760001
#>  [47,]  3.280915e-03  0.119753034  0.026978086 -0.046460505 -0.0143416769
#>  [48,] -2.628943e-01  0.190380255 -0.418191297 -0.014047195 -0.0389693632
#>  [49,] -8.022732e-02  0.381207244 -0.163932413 -0.002774406 -0.0050521928
#>  [50,] -4.384255e-01  0.223718063 -0.238032048  0.021978159 -0.0338490285
#>  [51,] -1.612843e-03 -0.005218952  0.282086423  0.001459893  0.0007207744
#>  [52,] -8.841976e-03 -0.166460821  0.051452828 -0.032680244  0.0358561296
#>  [53,] -6.734700e-02 -0.160482446  0.345658202 -0.006952650 -0.0343432767
#>  [54,]  4.941666e-01  0.032563935  0.028289169  0.049056897  0.0380912595
#>  [55,] -1.248938e-02 -0.070708496  0.170400315 -0.037429136 -0.0017651984
#>  [56,]  7.282822e-02 -0.377227722 -0.152620557  0.040186682  0.0271230818
#>  [57,] -2.866358e-02  0.333658094 -0.074300955  0.037823953 -0.0153674109
#>  [58,]  7.749127e-02 -0.110507627 -0.461355088 -0.013801423 -0.0435791666
#>  [59,]  7.506567e-02  0.183020443  0.336765382  0.014180800 -0.0007038843
#>  [60,]  1.118937e-01 -0.376397391 -0.052026298  0.016110678 -0.0042624283
#>  [61,] -2.319400e-01  0.338813374 -0.100066286  0.045699345  0.0307058573
#>  [62,] -4.300422e-02  0.228038935 -0.171705803 -0.017281139 -0.0172181919
#>  [63,] -2.013863e-01 -0.297606701  0.070704360 -0.022220930  0.0344564827
#>  [64,]  6.930720e-06  0.003835237  0.001807084  0.048801365 -0.0145525717
#>  [65,]  2.861557e-01 -0.075750977 -0.223377533  0.007458667  0.0024064121
#>  [66,] -3.461152e-01  0.360831116 -0.248650743 -0.033425671  0.0470545179
#>  [67,] -1.460557e-01  0.070315649 -0.380287636 -0.035560294 -0.0429631366
#>  [68,] -2.152882e-01 -0.386984698  0.311044926  0.046699327  0.0314480813
#>  [69,] -3.799672e-01  0.251239088 -0.324884617  0.021227714 -0.0469257004
#>  [70,] -4.812073e-01  0.135710909 -0.131850042  0.028738201  0.0067218661
#>  [71,]  1.047156e-01 -0.345124072 -0.235133315 -0.048813058 -0.0365070718
#>  [72,]  1.438713e-01 -0.441928157 -0.201455190 -0.025414831  0.0148092618
#>  [73,] -4.198529e-01 -0.195994221  0.263669125 -0.044960774 -0.0288158743
#>  [74,] -2.943910e-01  0.287633709 -0.108019482 -0.032921567 -0.0393909458
#>  [75,] -2.488041e-02  0.096025936 -0.240387757  0.004026652 -0.0436995254
#>  [76,]  2.648335e-01 -0.100789831 -0.303159563  0.014284641 -0.0465217978
#>  [77,] -1.369873e-01  0.467823993 -0.164848723 -0.032963758 -0.0291392233
#>  [78,] -3.124783e-02 -0.461281505  0.046553355  0.049464466  0.0114881480
#>  [79,] -1.114733e-01  0.142382394 -0.468963608 -0.036069776 -0.0194828836
#>  [80,] -4.223685e-01  0.206207479 -0.129888998  0.020289299 -0.0485348954
#>  [81,] -1.423431e-01  0.279692940 -0.359457614  0.036633093 -0.0097199444
#>  [82,]  9.374558e-02 -0.120420518 -0.473158829 -0.029914531 -0.0044783109
#>  [83,] -1.765258e-02  0.375470240 -0.038922570  0.035608689  0.0366461525
#>  [84,] -3.314004e-01  0.253939337 -0.108571663  0.030768446 -0.0395055116
#>  [85,] -6.367353e-02  0.474109057 -0.085153841  0.022653194 -0.0053874930
#>  [86,]  4.548659e-02  0.176637905  0.233177284  0.014886355  0.0058034629
#>  [87,]  7.350894e-02  0.277639358  0.224974856 -0.003475857  0.0060975001
#>  [88,]  1.334260e-01 -0.432944387 -0.200268978  0.021506445 -0.0234173421
#>  [89,] -9.328826e-04  0.008609155 -0.107093783  0.031937469  0.0187736356
#>  [90,] -1.467996e-01  0.023752496 -0.154106722 -0.037160310 -0.0442129100
#>  [91,]  1.106733e-03  0.028216455  0.039131464  0.049801787 -0.0419374371
#>  [92,]  8.657199e-02  0.215663220  0.325908924  0.012772074 -0.0498900986
#>  [93,] -4.023227e-01 -0.090663372  0.167856091  0.005877741  0.0109741065
#>  [94,]  1.682362e-01  0.425318734  0.240107256  0.025761254 -0.0074751925
#>  [95,]  4.215350e-01  0.156230462  0.234249521  0.027342288 -0.0305212979
#>  [96,]  3.105447e-02 -0.201198219 -0.144225112 -0.035392174 -0.0172342707
#>  [97,]  2.310118e-02  0.243457796  0.088033115  0.018060863 -0.0040695532
#>  [98,]  1.925985e-01  0.185806049  0.460879339 -0.004448415  0.0392953042
#>  [99,]  3.495120e-01 -0.356432450 -0.264067007 -0.047147391  0.0337261188
#> [100,] -2.947455e-01  0.033603044 -0.101569692 -0.017987990  0.0190098491