Generate data points on a Roman surface with optional noise.

This function generates data points on a Roman surface with optional noise.

Usage

roman_surface_with_noise(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

roman_surface_with_noise(n = 100, num_noise = 8,
min_n = -0.05, max_n = 0.05)
#>                 [,1]         [,2]         [,3]          [,4]          [,5]
#>   [1,] -1.869304e-02  0.329270267 -0.049560856  4.226709e-02  0.0223936454
#>   [2,] -2.466923e-02 -0.492037824  0.029275656 -1.907785e-02 -0.0071506144
#>   [3,] -4.358393e-02 -0.063073226  0.464923817  1.221186e-03 -0.0109671088
#>   [4,] -1.752646e-02  0.054024720 -0.292573598  3.539278e-02  0.0064272638
#>   [5,] -5.327454e-02  0.278226618 -0.168454375 -1.869261e-02 -0.0470498977
#>   [6,] -1.840764e-01  0.146244010 -0.458452788 -2.295105e-02  0.0489539401
#>   [7,] -2.107160e-03 -0.057523165  0.036460805 -1.906750e-02 -0.0297324308
#>   [8,] -3.802841e-01  0.277628872 -0.318239980  2.157296e-03  0.0419777144
#>   [9,]  2.175713e-01  0.042061513  0.176706776 -2.297152e-02  0.0115164994
#>  [10,] -2.765505e-01  0.285074585 -0.098089475  1.346350e-02 -0.0391720870
#>  [11,]  2.219657e-03 -0.015147465 -0.143422507  1.786862e-02  0.0009205266
#>  [12,]  1.885566e-01 -0.254661687 -0.424089694  2.660021e-02 -0.0065964586
#>  [13,]  1.661493e-01 -0.469781539 -0.170136403 -1.127088e-02  0.0428596318
#>  [14,]  1.692916e-01  0.274615460  0.393995656  6.324651e-03  0.0371440677
#>  [15,]  2.199473e-01 -0.418818596 -0.139170471 -1.737827e-04  0.0460846022
#>  [16,] -4.533133e-01 -0.134088575  0.091698133 -8.754926e-05  0.0225111873
#>  [17,]  2.154258e-02  0.151090458  0.136405426 -1.576318e-03  0.0343880526
#>  [18,] -4.081163e-01  0.024620521 -0.047354088 -8.744025e-03  0.0356741199
#>  [19,] -8.476092e-02  0.003477488 -0.040661171  4.237367e-02  0.0419562812
#>  [20,]  3.234095e-01  0.338197112  0.337737677  3.436238e-02  0.0189736202
#>  [21,]  8.279127e-03 -0.065496151 -0.123873655  4.161446e-02  0.0410217005
#>  [22,]  6.446923e-02  0.192157732  0.288614004  3.937608e-02  0.0135789534
#>  [23,] -9.052237e-03  0.080426295 -0.110412324 -2.959415e-02 -0.0374767438
#>  [24,] -2.856067e-02 -0.049942359  0.429507960  1.086969e-02 -0.0383876778
#>  [25,] -4.957928e-01  0.063828264 -0.061959954  4.528240e-03 -0.0405857127
#>  [26,]  1.047129e-01 -0.382582431 -0.049802054 -4.987060e-02 -0.0071278363
#>  [27,] -2.339731e-01 -0.344024157  0.361464098  2.194735e-02  0.0246126576
#>  [28,] -1.686729e-01 -0.426092042  0.239557308  3.317970e-02 -0.0347781395
#>  [29,] -1.290513e-02  0.051133098 -0.236607906  3.887020e-02  0.0043314905
#>  [30,] -1.829791e-02  0.499567841 -0.017928658 -1.921623e-02 -0.0298917597
#>  [31,]  1.228967e-02 -0.172532008 -0.068683723  3.032382e-02 -0.0496032486
#>  [32,]  1.563834e-01 -0.472557357 -0.135035832  4.897828e-02  0.0384484261
#>  [33,]  1.359194e-02  0.043132255  0.286069249  3.076901e-02  0.0232940348
#>  [34,] -1.655363e-01 -0.258594740  0.406221446  5.099417e-03 -0.0256177188
#>  [35,] -2.467215e-03  0.035773733 -0.068552358  1.084563e-02 -0.0334240521
#>  [36,]  3.961581e-01 -0.212414685 -0.299411335 -5.274749e-03  0.0155697784
#>  [37,] -4.041230e-01  0.216250796 -0.291112668 -3.968429e-02 -0.0068328597
#>  [38,] -1.052397e-01 -0.464258691  0.140775255  3.874378e-02 -0.0233134592
#>  [39,]  1.023293e-01  0.429150151  0.165893797 -3.735936e-02  0.0336599418
#>  [40,] -3.482451e-05 -0.021794432  0.001597099 -4.502559e-02 -0.0162945946
#>  [41,]  7.820449e-03  0.384077980  0.016690336  1.340022e-03 -0.0419883305
#>  [42,]  2.572400e-01  0.428737419  0.219071295  1.123429e-02 -0.0031171251
#>  [43,] -3.649395e-02 -0.208626291  0.161741712  2.104798e-02  0.0071959540
#>  [44,]  2.092809e-01  0.369752915  0.327334183  4.098467e-02  0.0244554508
#>  [45,] -3.704264e-01 -0.305672313  0.177169705 -1.327224e-02  0.0171396435
#>  [46,]  1.086745e-02  0.212166025  0.048667023 -3.253563e-02  0.0190867299
#>  [47,]  1.304773e-01  0.114377167  0.480292868  2.845191e-02 -0.0138306181
#>  [48,]  1.594809e-02 -0.261041287 -0.056371859  3.826363e-02  0.0281512015
#>  [49,]  1.057811e-01  0.486435890  0.113539310 -2.702939e-02 -0.0112867876
#>  [50,]  1.191541e-02  0.137609229  0.084272185  1.726371e-02  0.0432607265
#>  [51,] -1.146131e-02 -0.112012733  0.099960824  3.702805e-02 -0.0402172916
#>  [52,]  3.853038e-01 -0.255610656 -0.142679270  1.080385e-03 -0.0160099028
#>  [53,]  3.004467e-01 -0.399505470 -0.234764894  4.370220e-02  0.0263601514
#>  [54,] -2.570390e-02  0.468682628 -0.036758492 -3.925076e-02  0.0420658859
#>  [55,]  1.065325e-01  0.420295087  0.178389420 -4.510350e-02 -0.0414498751
#>  [56,]  7.715912e-02 -0.157742504 -0.382135233  4.851057e-02 -0.0433457989
#>  [57,]  1.482825e-01  0.220397293  0.427765082  6.801806e-03 -0.0282532581
#>  [58,]  7.068091e-02 -0.010053782 -0.138706690 -3.337894e-02  0.0325244784
#>  [59,]  1.912872e-01  0.087426296  0.360531362 -9.471644e-03  0.0379992667
#>  [60,] -6.102955e-02 -0.286883753  0.184174594  2.101121e-02 -0.0167505963
#>  [61,] -5.143732e-02  0.467919839 -0.036001730  1.262462e-02 -0.0409270704
#>  [62,]  1.799763e-01  0.227690829  0.441312440  2.158171e-03  0.0006489908
#>  [63,]  2.604877e-01  0.119706495  0.345141109 -4.922596e-02 -0.0093246567
#>  [64,]  4.250404e-01  0.119089660  0.068881215  3.797090e-02  0.0126763642
#>  [65,]  5.786475e-03 -0.075753481 -0.075501583 -2.569373e-02  0.0401505214
#>  [66,] -4.631548e-01 -0.181632346  0.186930878 -4.687618e-02 -0.0001963264
#>  [67,]  9.913924e-03  0.140164099  0.068960589  4.583331e-02 -0.0258231820
#>  [68,]  4.612670e-02 -0.140457705 -0.289805958 -2.068314e-02 -0.0134059048
#>  [69,]  2.454056e-05 -0.006095035 -0.004026104  1.737092e-02  0.0422499707
#>  [70,]  8.873479e-02 -0.409618070 -0.159875575 -4.104829e-02  0.0410741950
#>  [71,]  2.582104e-01  0.294599155  0.093829951 -4.642680e-02 -0.0256814371
#>  [72,]  1.746940e-01 -0.264690914 -0.407578860  2.582269e-02 -0.0417041842
#>  [73,]  2.532860e-01 -0.137784946 -0.038414428 -4.796091e-02  0.0480310617
#>  [74,] -4.864166e-02 -0.077198430  0.447238418 -2.830332e-02 -0.0189275688
#>  [75,] -2.152384e-04 -0.118373647  0.001792445 -1.761483e-03 -0.0122336936
#>  [76,]  4.189377e-02  0.088142020  0.383982700  1.187724e-02 -0.0356931264
#>  [77,] -1.623248e-01  0.049083072 -0.268834462 -1.828450e-02  0.0431083549
#>  [78,]  3.936116e-01 -0.044775154 -0.090410060 -2.768976e-02  0.0132792585
#>  [79,] -1.552958e-01 -0.252905383  0.402390408 -3.474475e-02  0.0037300746
#>  [80,]  1.134744e-01  0.357014971  0.239936953 -3.679294e-02 -0.0288095554
#>  [81,]  3.080737e-01  0.327751937  0.140564003  1.403010e-02 -0.0140399342
#>  [82,] -2.771426e-01  0.219951102 -0.071438628 -3.778858e-02  0.0067738879
#>  [83,]  9.174278e-02 -0.195189685 -0.366143546 -2.726059e-02  0.0049952163
#>  [84,] -1.808904e-01  0.278755910 -0.398912853  4.308015e-02  0.0450095822
#>  [85,]  1.005881e-01  0.358646205  0.043279458 -1.580515e-03  0.0260545784
#>  [86,]  2.109083e-01 -0.413125636 -0.278080738  1.406075e-02  0.0397649904
#>  [87,] -2.759066e-01  0.293625404 -0.397267095 -3.540137e-02 -0.0324591641
#>  [88,]  4.885925e-01 -0.012184790 -0.009832301 -2.740673e-02 -0.0180726969
#>  [89,] -1.657615e-02 -0.004350246  0.245457169 -3.603089e-02 -0.0377690815
#>  [90,]  1.300756e-01 -0.125959740 -0.482781783  3.893225e-02  0.0035431983
#>  [91,] -2.330830e-01 -0.272736619  0.418950093  4.776698e-03  0.0208210790
#>  [92,] -7.706033e-02 -0.222791092  0.290680884  1.212261e-02 -0.0434058924
#>  [93,] -4.381169e-02 -0.057288553  0.480017733 -1.081943e-02  0.0119114721
#>  [94,] -2.048063e-01 -0.312308446  0.076837472  2.861613e-02 -0.0276892831
#>  [95,] -5.727490e-02 -0.224318980  0.226079257  3.043388e-02 -0.0180413956
#>  [96,] -1.194980e-01 -0.047699892  0.338510635 -2.612150e-02  0.0290381813
#>  [97,]  4.822347e-01 -0.116291905 -0.099669997 -3.278583e-02  0.0446637604
#>  [98,]  2.756459e-01 -0.322861169 -0.116461828 -4.197546e-02  0.0029574167
#>  [99,] -9.577927e-02  0.352439849 -0.212955112  4.256038e-02 -0.0184993402
#> [100,] -2.883591e-01  0.017479044 -0.054843310  2.350844e-02 -0.0016095812
#>                 [,6]          [,7]          [,8]          [,9]         [,10]
#>   [1,]  9.521616e-05 -0.0172073594  0.0307516940  0.0234618294 -0.0401038633
#>   [2,] -1.893956e-02 -0.0223002825  0.0026822003 -0.0040666848 -0.0132228706
#>   [3,]  4.361391e-02 -0.0231287358 -0.0047595450  0.0347266830  0.0385298261
#>   [4,] -1.992240e-02 -0.0101917212  0.0174088389  0.0231895196  0.0400158201
#>   [5,]  1.764004e-02  0.0057753126 -0.0495410092  0.0231923949 -0.0430111898
#>   [6,]  4.132948e-02  0.0197244972  0.0411356451  0.0084781711  0.0017101304
#>   [7,] -5.802554e-03 -0.0307689531  0.0117141113  0.0332241725 -0.0481662693
#>   [8,] -2.166620e-02 -0.0363289408 -0.0331977856 -0.0099579564  0.0142342242
#>   [9,] -1.522336e-02 -0.0228172658  0.0243638615  0.0290661146 -0.0250347239
#>  [10,]  4.651794e-02 -0.0297810500 -0.0277867216  0.0354406696 -0.0021663966
#>  [11,]  1.456180e-02 -0.0477103471 -0.0411975005 -0.0157318188  0.0257814487
#>  [12,] -3.574610e-02 -0.0344292084  0.0353954606 -0.0354974853 -0.0151992977
#>  [13,]  3.959473e-02  0.0204119854  0.0003660906  0.0311661947 -0.0024110409
#>  [14,] -4.783713e-02 -0.0264584337 -0.0369230833 -0.0483883592  0.0499900384
#>  [15,] -2.103648e-02  0.0028616752 -0.0457143753  0.0445671861  0.0145019288
#>  [16,] -8.665278e-04  0.0330531284 -0.0180086509  0.0010109721  0.0083652628
#>  [17,] -1.530778e-02  0.0378845478 -0.0157960910 -0.0425921815  0.0294790995
#>  [18,] -6.782656e-03 -0.0177045155  0.0483799116 -0.0464208638 -0.0035670719
#>  [19,]  2.324877e-02 -0.0337753196  0.0138730677 -0.0351645889  0.0373078809
#>  [20,]  1.053327e-02 -0.0189064131  0.0429092847  0.0037568555  0.0341527650
#>  [21,] -3.786034e-02  0.0387770105 -0.0418528705 -0.0076332285 -0.0431773406
#>  [22,]  2.771146e-02 -0.0207285941 -0.0148885951 -0.0420274923 -0.0376229295
#>  [23,] -4.074711e-02 -0.0079270235 -0.0388904460  0.0028938612 -0.0355245161
#>  [24,] -4.319066e-02  0.0015550741  0.0083328215  0.0008152405  0.0326189054
#>  [25,] -1.122707e-02  0.0178151881  0.0202727830 -0.0444145594  0.0399738909
#>  [26,] -4.268610e-02 -0.0384385654 -0.0323776991 -0.0245736291 -0.0253501086
#>  [27,]  2.283705e-02  0.0011409885  0.0353496553 -0.0366529968 -0.0037875813
#>  [28,]  4.889322e-02 -0.0444714993  0.0479728599 -0.0494076489 -0.0002357048
#>  [29,]  3.582931e-02  0.0219036604 -0.0368332097 -0.0192902023  0.0308989903
#>  [30,]  2.579276e-02 -0.0043650464 -0.0348450779 -0.0398451325 -0.0261858245
#>  [31,] -3.661453e-02  0.0034961618 -0.0325548516  0.0326397044 -0.0093042359
#>  [32,]  2.352589e-02 -0.0490032254  0.0200867542 -0.0422794077  0.0248394853
#>  [33,]  4.164040e-02 -0.0427478649  0.0160743825  0.0302672950 -0.0343459307
#>  [34,]  3.933539e-02  0.0299405253  0.0467719716  0.0194971071  0.0077628783
#>  [35,] -4.495990e-02  0.0321191312  0.0103122525 -0.0136373684 -0.0094798177
#>  [36,] -2.231202e-02 -0.0161784599  0.0115688592  0.0382314543  0.0190921098
#>  [37,]  7.432453e-03 -0.0082737287  0.0463264088 -0.0016079857  0.0315385019
#>  [38,] -4.171506e-02 -0.0407160239  0.0253210499 -0.0305146767 -0.0062556417
#>  [39,] -1.469917e-02 -0.0046219188 -0.0307449190 -0.0222338091  0.0280707784
#>  [40,] -4.183133e-02  0.0216060758 -0.0071683184  0.0330714962 -0.0094769391
#>  [41,]  1.254448e-02  0.0226711328  0.0360957523  0.0115472567 -0.0357489867
#>  [42,] -2.620543e-02 -0.0412582674 -0.0271281819 -0.0483312341  0.0230836401
#>  [43,] -2.503487e-02  0.0318188905  0.0125189194 -0.0195997853 -0.0090781627
#>  [44,] -4.331636e-02 -0.0442896327  0.0074606275 -0.0216131426 -0.0282419795
#>  [45,]  2.719116e-02 -0.0063633443 -0.0138390964 -0.0195462710  0.0432821232
#>  [46,] -3.578799e-02  0.0267206956 -0.0241253724  0.0284186496  0.0286807350
#>  [47,]  1.450617e-02 -0.0368153896 -0.0409164861  0.0386915315 -0.0440817680
#>  [48,] -2.105688e-02  0.0146264939  0.0185878391  0.0459210987 -0.0143894460
#>  [49,] -3.809925e-02  0.0424700768 -0.0171436717 -0.0460873031 -0.0198992772
#>  [50,] -3.630901e-02 -0.0178925464  0.0241327165 -0.0253418096 -0.0195603144
#>  [51,]  2.110960e-02 -0.0073440419  0.0070315686 -0.0085353440 -0.0412865476
#>  [52,] -4.024541e-02  0.0311158936 -0.0156992736  0.0307900076 -0.0192870727
#>  [53,]  3.352398e-02  0.0474344181 -0.0347195798 -0.0469534871  0.0312743022
#>  [54,] -3.968051e-02  0.0232392969 -0.0244759848 -0.0163375882  0.0193799291
#>  [55,]  4.071049e-02 -0.0156027002  0.0340756334  0.0239952798 -0.0015812704
#>  [56,] -4.713577e-02 -0.0171846927 -0.0412143918  0.0388280467 -0.0285622859
#>  [57,] -7.900039e-03  0.0381772376 -0.0091738202  0.0280978021  0.0396219824
#>  [58,] -4.415749e-02 -0.0435139775  0.0495074975 -0.0148177539 -0.0352514074
#>  [59,] -3.527622e-02  0.0258826403 -0.0294796787  0.0259473208  0.0279465789
#>  [60,] -1.392980e-02  0.0134433061 -0.0210169895 -0.0165952451  0.0214841192
#>  [61,] -6.786625e-03  0.0169973451 -0.0188552435 -0.0028494145 -0.0095472151
#>  [62,] -1.361176e-02  0.0428563133 -0.0353263823 -0.0227259549 -0.0464742140
#>  [63,]  2.405228e-04 -0.0354614842 -0.0333354607  0.0463121660 -0.0050839094
#>  [64,] -4.909863e-02 -0.0361828258  0.0469660298  0.0133380719  0.0056709421
#>  [65,]  2.629593e-02 -0.0299343970 -0.0433029271 -0.0331625406  0.0265413817
#>  [66,] -8.610424e-03 -0.0419908367 -0.0415557840  0.0428700109 -0.0325969992
#>  [67,] -5.624744e-03  0.0289580954  0.0370256638  0.0138360330 -0.0089809836
#>  [68,]  2.530408e-02  0.0336261392  0.0367632700 -0.0070837569 -0.0122859175
#>  [69,]  2.293321e-02  0.0143133345  0.0184526358 -0.0156154521  0.0270166170
#>  [70,]  4.829068e-02 -0.0168624699  0.0444686700  0.0022942668  0.0348576669
#>  [71,]  2.489589e-02 -0.0185596225  0.0115633949  0.0045842411  0.0298342893
#>  [72,]  1.231946e-02 -0.0283408355 -0.0157634551  0.0117807796 -0.0185022906
#>  [73,]  3.373204e-02  0.0155424506 -0.0240182868 -0.0205344067 -0.0366329697
#>  [74,]  2.028203e-02 -0.0078990056 -0.0282700221 -0.0392575591  0.0120214755
#>  [75,]  4.299481e-02 -0.0389050137  0.0173535326  0.0032986777  0.0176842314
#>  [76,]  1.529263e-02 -0.0393282842  0.0464151531 -0.0059147818  0.0158300434
#>  [77,]  2.242068e-02  0.0270081083 -0.0175588437  0.0211108564 -0.0265946368
#>  [78,]  1.893371e-02 -0.0342164542 -0.0419447914  0.0026276361 -0.0041189420
#>  [79,]  3.296047e-02 -0.0278535092 -0.0429412813 -0.0033157368 -0.0236673514
#>  [80,] -1.663092e-02 -0.0034390963 -0.0232101000 -0.0095035827  0.0475180184
#>  [81,]  1.799231e-02 -0.0102707196 -0.0160370285  0.0009629260  0.0143179732
#>  [82,] -4.152410e-02 -0.0462536138  0.0300367596  0.0384218578 -0.0025411272
#>  [83,]  1.335996e-02 -0.0208506366  0.0370825181  0.0081344341  0.0399965454
#>  [84,] -3.572740e-02 -0.0300771132  0.0381656251  0.0465914276 -0.0001760205
#>  [85,]  1.366818e-03 -0.0040198287  0.0137281560  0.0127549055 -0.0353530779
#>  [86,] -4.547348e-02  0.0009411381  0.0251597907 -0.0420259516 -0.0435326782
#>  [87,]  4.343821e-02 -0.0241386232 -0.0211132850 -0.0332060132  0.0213847220
#>  [88,] -3.386441e-02 -0.0462775002 -0.0029209384 -0.0048058668 -0.0144955840
#>  [89,]  6.082831e-03 -0.0114114023 -0.0384433763  0.0371581825  0.0444819321
#>  [90,] -3.824451e-02  0.0002874756  0.0366361377  0.0029651299  0.0462432905
#>  [91,]  3.160291e-02  0.0187196399  0.0450994452 -0.0218857854 -0.0343918403
#>  [92,] -3.196001e-02  0.0169161956  0.0049369015  0.0480088063  0.0110005761
#>  [93,] -4.022213e-02  0.0236343998  0.0221250155  0.0213965237 -0.0080495462
#>  [94,] -2.301282e-02  0.0237382195 -0.0427354569 -0.0360540837 -0.0317908633
#>  [95,]  2.078378e-02  0.0413819076 -0.0217138006  0.0010926298 -0.0043027752
#>  [96,]  1.417774e-02 -0.0003530915 -0.0069523322  0.0395389074  0.0235061679
#>  [97,]  6.001035e-03  0.0092663842 -0.0152474643  0.0164348182 -0.0424865043
#>  [98,]  4.566157e-02  0.0292030083  0.0057981252 -0.0064327515  0.0179289135
#>  [99,] -4.577350e-02  0.0151751673 -0.0439960615  0.0345864706  0.0320648373
#> [100,]  3.204834e-02 -0.0272947988 -0.0064159404 -0.0367117811 -0.0435004009
#>                [,11]
#>   [1,]  0.0009348501
#>   [2,] -0.0054105347
#>   [3,]  0.0264372552
#>   [4,]  0.0180847583
#>   [5,]  0.0457994778
#>   [6,] -0.0055335815
#>   [7,] -0.0186649161
#>   [8,]  0.0130545829
#>   [9,]  0.0460881645
#>  [10,] -0.0043422894
#>  [11,] -0.0237134074
#>  [12,]  0.0425501091
#>  [13,]  0.0338479123
#>  [14,]  0.0127845748
#>  [15,]  0.0294204812
#>  [16,]  0.0145647840
#>  [17,] -0.0032317718
#>  [18,]  0.0184520660
#>  [19,] -0.0497543893
#>  [20,] -0.0413771939
#>  [21,]  0.0436538680
#>  [22,] -0.0387757104
#>  [23,] -0.0370841580
#>  [24,]  0.0220069072
#>  [25,]  0.0235231907
#>  [26,]  0.0262789725
#>  [27,]  0.0291862053
#>  [28,]  0.0020364895
#>  [29,]  0.0409814705
#>  [30,] -0.0048118972
#>  [31,]  0.0061563396
#>  [32,] -0.0384689309
#>  [33,] -0.0024159052
#>  [34,]  0.0403071044
#>  [35,]  0.0217459274
#>  [36,] -0.0143887536
#>  [37,]  0.0257435456
#>  [38,] -0.0441222281
#>  [39,] -0.0201434591
#>  [40,] -0.0431365239
#>  [41,]  0.0195870253
#>  [42,]  0.0103439874
#>  [43,] -0.0042058643
#>  [44,] -0.0172630761
#>  [45,]  0.0112077414
#>  [46,] -0.0499552988
#>  [47,] -0.0205677158
#>  [48,] -0.0274273557
#>  [49,] -0.0152336852
#>  [50,]  0.0188353597
#>  [51,]  0.0166639086
#>  [52,] -0.0383014771
#>  [53,] -0.0028754394
#>  [54,] -0.0316228552
#>  [55,]  0.0076716572
#>  [56,] -0.0059632291
#>  [57,] -0.0324870453
#>  [58,] -0.0102576839
#>  [59,]  0.0403943449
#>  [60,] -0.0257778027
#>  [61,] -0.0147510694
#>  [62,]  0.0309416761
#>  [63,]  0.0298743446
#>  [64,]  0.0295923552
#>  [65,] -0.0155748448
#>  [66,]  0.0461055982
#>  [67,]  0.0032778464
#>  [68,] -0.0124241151
#>  [69,] -0.0207130924
#>  [70,] -0.0216261991
#>  [71,]  0.0085332769
#>  [72,] -0.0292874601
#>  [73,]  0.0346820571
#>  [74,]  0.0411153909
#>  [75,] -0.0444066453
#>  [76,]  0.0255811164
#>  [77,] -0.0395144009
#>  [78,] -0.0367564047
#>  [79,]  0.0391642334
#>  [80,]  0.0482323285
#>  [81,] -0.0048800559
#>  [82,] -0.0201071579
#>  [83,] -0.0493286729
#>  [84,] -0.0093631394
#>  [85,]  0.0373909755
#>  [86,]  0.0167994725
#>  [87,] -0.0470255262
#>  [88,] -0.0423687363
#>  [89,] -0.0257875062
#>  [90,]  0.0282214087
#>  [91,]  0.0224509608
#>  [92,] -0.0061513941
#>  [93,]  0.0215357641
#>  [94,] -0.0264615026
#>  [95,]  0.0101999008
#>  [96,] -0.0493766953
#>  [97,] -0.0458203715
#>  [98,]  0.0108395814
#>  [99,]  0.0249532758
#> [100,]  0.0444567442