Generate data representing small spheres within a larger encompassing sphere with added noise.

This function generates data points representing small spheres within a larger encompassing sphere and adds noise to the data if specified.

Usage

small_big_sphere_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate, should be a multiple of 13.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

small_big_sphere_with_noise(sample_size = 390, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>                 [,1]        [,2]         [,3]         [,4]          [,5]
#>   [1,] -2.789258e+00   7.9559840   1.75863393  -5.29338200 -0.0027409989
#>   [2,] -3.755919e+00   5.3105791   2.21345275  -5.20496143 -0.0009818338
#>   [3,] -4.045390e+00   4.2776694   0.82205983  -8.06796440  0.0143159085
#>   [4,] -4.088544e+00   6.5594368   1.24894101  -9.15023511 -0.0288522211
#>   [5,]  4.138040e-01   7.5583489   2.24224535  -8.35518223 -0.0411806517
#>   [6,] -3.466967e+00   5.3952107   2.65744748  -5.11100567  0.0331097727
#>   [7,] -2.706765e+00   6.2834357   1.04501125 -10.10636552  0.0034899638
#>   [8,] -1.405975e+00   3.5601653   3.10018661  -6.41733859 -0.0370544043
#>   [9,] -4.243428e+00   4.6359328   1.38296006  -6.17552934  0.0292880963
#>  [10,] -2.342293e+00   5.1094422   2.78702478  -4.74552621  0.0290518005
#>  [11,] -2.465405e-01   8.3748631   0.90864218  -7.81009039  0.0173076425
#>  [12,] -1.882541e-01   4.2604204   0.45223411  -5.86906006  0.0081820370
#>  [13,] -3.567060e+00   4.7695401   1.06429535  -5.26196861 -0.0022542704
#>  [14,] -3.501971e+00   3.7222177   1.39138868  -8.35197466 -0.0008428099
#>  [15,] -1.412790e+00   7.6555660  -0.82401487  -7.04437187 -0.0061960225
#>  [16,] -8.624990e-01   5.5829691   0.60645710  -4.68112065  0.0425813718
#>  [17,] -3.664612e+00   3.6067353   1.53549772  -7.42581202 -0.0188001706
#>  [18,]  7.450763e-01   4.6276501   1.93218528  -6.59071185 -0.0420870892
#>  [19,]  7.589631e-01   6.5738827   0.27009993  -7.75843045  0.0136882657
#>  [20,]  2.006301e-01   5.7131491  -0.58182921  -7.56103496 -0.0165171196
#>  [21,]  8.303503e-01   5.1226969   1.50367802  -8.43475734 -0.0136907826
#>  [22,] -2.647485e+00   5.1791799   0.48933943  -4.77663067  0.0088853210
#>  [23,] -2.319913e+00   3.3293443   1.74027233  -5.98444666  0.0393947933
#>  [24,] -4.555704e-01   4.7420308  -0.16964697  -8.88845490 -0.0123166470
#>  [25,] -2.214787e+00   8.8609489   0.92017022  -7.35533175 -0.0192548504
#>  [26,] -2.155464e+00   5.7741390  -1.32869472  -7.78882364 -0.0472971624
#>  [27,] -1.311183e+00   8.2853253   1.33418598  -5.49838923 -0.0275381119
#>  [28,] -3.527637e+00   3.7839462   0.47566764  -7.14721598 -0.0014641122
#>  [29,] -3.872663e+00   7.0964612   2.99124275  -6.05106429  0.0440712923
#>  [30,] -1.666217e+00   4.7322316   4.33705386  -7.27668604  0.0431698710
#>  [31,]  2.671725e+00   1.1566627  -0.30995618   8.05147814 -0.0401073257
#>  [32,]  1.705064e+00  -1.4887422   0.14486264   8.13989871  0.0295935417
#>  [33,]  1.415593e+00  -2.5216519  -1.24653028   5.27689573 -0.0035549731
#>  [34,]  1.372439e+00  -0.2398845  -0.81964910   4.19462503  0.0470787340
#>  [35,]  5.874787e+00   0.7590276   0.17365524   4.98967791 -0.0039680616
#>  [36,]  1.994015e+00  -1.4041106   0.58885737   8.23385447 -0.0117731159
#>  [37,]  2.754218e+00  -0.5158856  -1.02357886   3.23849461  0.0396301861
#>  [38,]  4.055008e+00  -3.2391560   1.03159650   6.92752155 -0.0148805361
#>  [39,]  1.217555e+00  -2.1633885  -0.68563005   7.16933080 -0.0033798450
#>  [40,]  3.118689e+00  -1.6898791   0.71843467   8.59933392  0.0499992079
#>  [41,]  5.214442e+00   1.5755418  -1.15994792   5.53476975 -0.0432841314
#>  [42,]  5.272729e+00  -2.5389009  -1.61635600   7.47580008 -0.0166666994
#>  [43,]  1.893922e+00  -2.0297812  -1.00429476   8.08289153 -0.0002536011
#>  [44,]  1.959011e+00  -3.0771036  -0.67720143   4.99288547 -0.0034222506
#>  [45,]  4.048193e+00   0.8562447  -2.89260497   6.30048827 -0.0220524735
#>  [46,]  4.598484e+00  -1.2163522  -1.46213301   8.66373949  0.0193484914
#>  [47,]  1.796370e+00  -3.1925860  -0.53309239   5.91904811  0.0249574766
#>  [48,]  6.206059e+00  -2.1716712  -0.13640483   6.75414828  0.0233004792
#>  [49,]  6.219946e+00  -0.2254386  -1.79849017   5.58642969 -0.0492604014
#>  [50,]  5.661613e+00  -1.0861722  -2.65041932   5.78382518  0.0287831886
#>  [51,]  6.291333e+00  -1.6766244  -0.56491209   4.91010279  0.0351324049
#>  [52,]  2.813498e+00  -1.6201414  -1.57925068   8.56822947 -0.0323690788
#>  [53,]  3.141070e+00  -3.4699770  -0.32831778   7.36041348  0.0072392003
#>  [54,]  5.005412e+00  -2.0572905  -2.23823708   4.45640524  0.0017862579
#>  [55,]  3.246196e+00   2.0616276  -1.14841989   5.98952838 -0.0118683703
#>  [56,]  3.305519e+00  -1.0251823  -3.39728483   5.55603649  0.0399624557
#>  [57,]  4.149800e+00   1.4860040  -0.73440412   7.84647091 -0.0374529202
#>  [58,]  1.933345e+00  -3.0153751  -1.59292247   6.19764416  0.0172401139
#>  [59,]  1.588320e+00   0.2971399   0.92265265   7.29379585 -0.0067165892
#>  [60,]  3.794766e+00  -2.0670897   2.26846375   6.06817410  0.0051688562
#>  [61,]  6.855684e-01 -11.4697289  -4.83890193  10.44185515  0.0461753666
#>  [62,] -2.810927e-01 -14.1151337  -4.38408312  10.53027572  0.0379448387
#>  [63,] -5.705636e-01 -15.1480434  -5.77547604   7.66727274 -0.0114364368
#>  [64,] -6.137172e-01 -12.8662760  -5.34859485   6.58500204  0.0355006760
#>  [65,]  3.888631e+00 -11.8673639  -4.35529052   7.38005492  0.0267504568
#>  [66,]  7.859131e-03 -14.0305021  -3.94008838  10.62423147 -0.0080116730
#>  [67,]  7.680617e-01 -13.1422772  -5.55252462   5.62887162 -0.0363195139
#>  [68,]  2.068852e+00 -15.8655475  -3.49734926   9.31789855  0.0466556178
#>  [69,] -7.686016e-01 -14.7897800  -5.21457580   9.55970780  0.0345957809
#>  [70,]  1.132533e+00 -14.3162706  -3.81051109  10.98971093  0.0092369739
#>  [71,]  3.228286e+00 -11.0508498  -5.68889368   7.92514676 -0.0356611551
#>  [72,]  3.286572e+00 -15.1652924  -6.14530175   9.86617709 -0.0381525940
#>  [73,] -9.223394e-02 -14.6561728  -5.53324051  10.47326854  0.0055163404
#>  [74,] -2.714484e-02 -15.7034951  -5.20614719   7.38326248 -0.0406530660
#>  [75,]  2.062037e+00 -11.7701468  -7.42155073   8.69086528 -0.0498011990
#>  [76,]  2.612328e+00 -13.8427437  -5.99107877  11.05411649  0.0058121504
#>  [77,] -1.897860e-01 -15.8189775  -5.06203814   8.30942512 -0.0477030213
#>  [78,]  4.219903e+00 -14.7980627  -4.66535058   9.14452529 -0.0489644559
#>  [79,]  4.233790e+00 -12.8518302  -6.32743593   7.97680669  0.0477977195
#>  [80,]  3.675457e+00 -13.7125638  -7.17936508   8.17420219 -0.0341426316
#>  [81,]  4.305177e+00 -14.3030159  -5.09385784   7.30047980 -0.0230897730
#>  [82,]  8.273417e-01 -14.2465329  -6.10819644  10.95860647  0.0435775504
#>  [83,]  1.154913e+00 -16.0963685  -4.85726353   9.75079048  0.0333234228
#>  [84,]  3.019256e+00 -14.6836820  -6.76718283   6.84678225 -0.0286111622
#>  [85,]  1.260040e+00 -10.5647639  -5.67736564   8.37990539  0.0076730531
#>  [86,]  1.319363e+00 -13.6515739  -7.92623058   7.94641350 -0.0306368138
#>  [87,]  2.163644e+00 -11.1403875  -5.26334988  10.23684792 -0.0304929947
#>  [88,] -5.281089e-02 -15.6417666  -6.12186823   8.58802116 -0.0334831300
#>  [89,] -3.978362e-01 -12.3292516  -3.60629311   9.68417286 -0.0436278134
#>  [90,]  1.808610e+00 -14.6934812  -2.26048200   8.45855110  0.0127554375
#>  [91,]  8.970830e+00 -10.4923491  -3.76413837  -4.50178560  0.0438709634
#>  [92,] -1.219548e+00   1.1580734 -14.87490067   0.95336195 -0.0190674186
#>  [93,]  9.291564e+00  -9.4025046   7.08859646   0.10745554  0.0259475975
#>  [94,]  1.002518e+01   5.6855402  -3.55487370   8.91814167 -0.0127910370
#>  [95,]  2.120202e+00   7.5109037   8.07785586   9.94179628  0.0420616323
#>  [96,] -1.342531e-01  12.6406898  -2.36489952  -7.72024536 -0.0290283532
#>  [97,]  6.517274e-01 -11.3948830   8.81356352   4.12952680  0.0381300780
#>  [98,]  9.540911e+00  -6.6291105  -9.24757937   2.12324984  0.0473881890
#>  [99,] -9.600716e+00  -1.1714683  -7.17836906  -8.94007462 -0.0223718336
#> [100,] -1.092833e+01  -3.4135010  -9.58931979   1.40163039  0.0496897068
#> [101,]  1.369448e+01   1.8621162  -4.04525347  -4.19877936  0.0418166999
#> [102,]  8.392240e+00   8.3248200   4.38300798  -8.12754057  0.0025830809
#> [103,]  3.190486e+00   0.4001064 -14.63501785  -0.69062943 -0.0073573437
#> [104,] -1.217542e+01   4.8095795  -5.85252094  -4.40171351  0.0445457743
#> [105,]  2.778291e+00   3.1423888 -13.30943123  -5.50141196  0.0089797237
#> [106,] -5.077444e+00  -3.7502475   2.89967734 -13.29462579 -0.0465552265
#> [107,] -5.174517e+00  -1.7521784 -13.95327385  -0.67852085  0.0082594251
#> [108,]  5.079660e+00  11.8774318   7.50872557  -1.32011863 -0.0114868055
#> [109,]  4.864642e+00   2.8206769 -10.00954995   9.65339064  0.0039424574
#> [110,] -4.677560e+00  11.5999018  -7.74732020  -2.92262535  0.0204892196
#> [111,]  1.213649e+01   1.2237000   6.48215759   5.84720221 -0.0295103819
#> [112,]  7.446904e+00   9.4894180   7.74567447   4.41577825 -0.0299411011
#> [113,] -1.025956e+01   8.4818403   4.54875327  -5.20659857 -0.0271450755
#> [114,] -1.671821e+00   3.5388453  14.01056489  -3.65864162  0.0345005997
#> [115,]  1.857591e+00  14.8723398   0.59680097   0.08180880 -0.0169374186
#> [116,]  8.642611e+00  -9.8246518  -6.87063081  -2.56435498 -0.0150691135
#> [117,] -1.178021e+01   4.6266948  -3.61681742   7.19298426  0.0348874428
#> [118,] -1.362217e+01  -1.1017634  -2.17527122  -5.78712798  0.0225739281
#> [119,] -5.567425e+00   6.3104554  10.58177445  -6.49676699 -0.0121233527
#> [120,]  1.643380e+00  -5.1823583   3.46526942 -13.54379461 -0.0497840195
#> [121,] -3.451204e+00   8.5372253  11.28325042  -3.59071494  0.0017397061
#> [122,] -1.245261e+01   1.1069862  -0.52560310   8.27229748  0.0349155360
#> [123,]  5.553088e+00  -3.0199074  12.12243074  -6.17171363 -0.0079226844
#> [124,]  1.508742e+00  14.1957760  -0.64273017   4.55966434 -0.0216042877
#> [125,] -8.443711e-01  12.9289666   0.11647363  -7.55746606  0.0493799857
#> [126,]  3.040668e+00  -3.8490839  10.03798617  10.00888246 -0.0203905018
#> [127,]  1.295716e+01   1.4283372  -6.42481148  -3.71396977  0.0381571715
#> [128,]  1.369483e+01  -2.3991752   3.47840833  -4.42677436  0.0183561302
#> [129,] -1.189586e+01  -5.5496048  -4.87532685  -5.37787049  0.0128347773
#> [130,]  1.570652e+00   2.4881742   1.43843941  14.63806451 -0.0417610256
#> [131,] -3.842867e+00 -11.2744171  -2.15153057   8.85950374  0.0277707363
#> [132,]  1.304515e+01  -0.5887936   4.36346654   5.95293423  0.0039500390
#> [133,] -2.422868e+00   6.8539193 -11.01483617   7.12929768 -0.0488838205
#> [134,] -2.312309e+00  11.6244592  -3.44417955  -8.52424795  0.0177850027
#> [135,]  1.365011e+01   3.0607056   3.48004128   4.14677869  0.0111809357
#> [136,]  9.142978e+00  -9.1632219   6.66630518   3.60578482  0.0384782864
#> [137,] -9.379370e+00   5.8370099   6.79125865  -7.53893544  0.0073563515
#> [138,] -8.354017e+00  -6.9240328 -10.09469785   2.31630021  0.0128151776
#> [139,] -2.486705e+00  14.7342399   1.07419385  -0.75138493 -0.0066297252
#> [140,]  1.095188e+01  -0.9633666   7.45846435   6.96416384 -0.0435744314
#> [141,] -7.533783e-01   2.9837341 -11.10213601   9.60584862 -0.0223847083
#> [142,] -1.531496e+00 -13.7828193  -1.12037454   5.60652969 -0.0415841853
#> [143,]  1.154009e+00   0.4745054 -14.87226688   1.50292538  0.0393458884
#> [144,]  1.097601e+01  -3.2931671  -9.19976845  -3.00776042  0.0101729277
#> [145,] -4.213885e+00   1.3706771  14.07693036   2.68411079  0.0351967489
#> [146,]  1.071302e+01   4.2735083   6.84660019  -6.71508157 -0.0216529948
#> [147,] -3.603658e+00  -8.0307956  -1.13592394 -12.09254501 -0.0051659320
#> [148,] -2.580168e+00 -11.8338117   7.23308687  -5.09765513  0.0093801327
#> [149,] -5.876400e-01   3.5408649 -11.45291047  -8.99709936  0.0210738510
#> [150,]  2.179483e+00  -2.6791729  14.52804817   1.41693362  0.0374932231
#> [151,]  7.428223e+00   9.9763187  -8.15571836   1.94391892  0.0032604345
#> [152,]  1.379257e+00  -9.5212549  -1.45863473  11.41559203  0.0180750810
#> [153,]  7.268556e+00  -0.6651727 -12.69260498   3.25935802  0.0341655283
#> [154,]  3.545248e+00  13.6340701   4.68052416  -2.15314679 -0.0262935215
#> [155,]  7.842902e+00 -10.4141328   0.93389743  -7.35952163 -0.0384207920
#> [156,]  5.523396e+00   9.8117250   1.92794331  -9.72137786 -0.0495829426
#> [157,] -1.351017e+01   3.0669858   5.75055486  -0.01103607  0.0261582428
#> [158,]  1.632099e+00   6.0047831  13.40361949  -2.57328887  0.0480949732
#> [159,] -1.154044e+01   2.0989624   1.12568737   9.28145157 -0.0190551373
#> [160,] -7.379426e+00  -6.9815273 -10.76786508   2.41979916 -0.0452986259
#> [161,] -1.093696e+01  -5.8279055   4.10762249   7.38551162  0.0280814481
#> [162,]  9.884038e+00  -3.0238540   3.49693671 -10.29240220  0.0495720570
#> [163,]  2.511732e+00   2.1775462  -8.23582395 -12.08803949 -0.0346997277
#> [164,]  1.431052e+01   2.2811201   3.85386587   0.39142600  0.0475954048
#> [165,] -8.602462e+00  -7.4931341   5.46946937  -8.05825591 -0.0379341354
#> [166,]  1.031140e+01  -8.6527696  -5.38504273   3.84785237 -0.0417005731
#> [167,]  3.425478e+00  -5.8894758   5.14810770 -12.33195697 -0.0178785620
#> [168,] -1.034462e+01 -10.5081165   0.86013690  -2.61312574  0.0265422601
#> [169,] -9.323544e+00   9.3999722  -0.43439841   7.03728248 -0.0330254275
#> [170,] -1.067640e+01  -8.4347501   6.31273251   0.13745887  0.0423137183
#> [171,] -1.061530e+01  10.0883511  -2.93690564   1.38391074 -0.0222706668
#> [172,]  3.169413e+00   9.9693062 -10.43898127  -2.56815645  0.0047309237
#> [173,] -7.091385e+00  12.5315147  -3.50628083  -2.31935202 -0.0348084323
#> [174,] -8.001419e+00 -11.3213322  -2.70448050   5.04881369 -0.0243717068
#> [175,]  1.190161e+00  -0.2615637 -14.02236587  -5.18539850 -0.0146831196
#> [176,]  7.346385e+00 -12.5138557  -3.09457611   2.20400514 -0.0139821905
#> [177,]  7.358388e+00  -9.0871610  -8.50835050  -3.98567482  0.0148778549
#> [178,]  6.020749e+00  -6.5895472  -8.57970645   8.46859428  0.0067748491
#> [179,]  7.884142e+00   7.9415752   3.00898137   9.52458467  0.0363769507
#> [180,] -7.817362e+00  10.1414548  -2.19029297  -7.49949114 -0.0069149152
#> [181,] -6.152102e+00  -9.0924728  -2.09781041  10.00388744 -0.0005955629
#> [182,]  1.261098e+01  -1.3193423   7.46847655  -2.90591149  0.0001354119
#> [183,]  3.055763e+00  -4.7621847   1.75907942 -13.78004177 -0.0058698016
#> [184,]  5.942037e+00  12.8924080   4.16225156   2.48066018  0.0351338839
#> [185,]  1.556011e+00  -9.9612662   4.19394489 -10.28410575 -0.0441461657
#> [186,]  7.453625e+00   5.0697805  -4.61797898 -11.06413423  0.0023779168
#> [187,] -1.797702e+00   3.3815965  -4.89597908  13.65146378  0.0288384800
#> [188,] -1.216963e+01  -6.6828247  -3.91431555   4.11315815  0.0162981322
#> [189,] -5.793123e+00 -12.5376495   3.60226769   4.61202135 -0.0324985641
#> [190,] -3.940021e+00  -1.9799217 -13.95862236  -3.27307278  0.0221179466
#> [191,]  3.295676e+00  13.1080004  -5.67653359  -3.17739024  0.0266337472
#> [192,]  1.237133e+00  13.2869801  -5.53893503   4.03061530 -0.0175042943
#> [193,] -8.484825e+00   4.7401821   1.03757490 -11.37813072 -0.0058369315
#> [194,] -7.546503e+00  -1.9243410  11.76905577  -5.08296412  0.0305234385
#> [195,]  3.902432e+00  -9.2681763  -9.51934308  -5.76663132  0.0117787368
#> [196,]  1.017791e+01  -1.1132920  -6.88033870  -8.53415223  0.0340112844
#> [197,]  1.109157e+01   1.0920916  -9.28106932  -3.82704150 -0.0495549047
#> [198,]  1.411466e+01   2.9397602  -3.72895112  -1.79693480 -0.0440663379
#> [199,]  4.829457e+00   3.4129209 -13.53831030   2.59662730 -0.0366588131
#> [200,] -4.486208e+00   6.2511327  12.66044060  -2.34745059 -0.0423519193
#> [201,] -6.844723e+00  -3.9016853   4.69937535 -11.86770786  0.0179676127
#> [202,] -1.225935e+01  -3.2385413  -5.04148216  -6.22926375 -0.0346720112
#> [203,] -9.147037e+00 -10.3235453  -5.14677152   2.87521751 -0.0105358604
#> [204,]  8.933283e+00  -2.7075767  -9.90330005   6.30794170 -0.0388897476
#> [205,] -9.252580e+00  -5.7729562  -0.47511205  10.28771168  0.0206338906
#> [206,]  1.267178e+01  -6.7067418   0.46354847  -4.38529681 -0.0286781157
#> [207,]  4.989183e+00  -1.1249065   5.12719435 -13.13600065 -0.0029163714
#> [208,] -5.689257e+00  -5.8078820  10.54927241   6.90026923 -0.0339468971
#> [209,] -6.567299e+00  -1.3456757  10.38586706  -8.49667654  0.0146281933
#> [210,]  5.787823e+00  13.0230257  -4.66191527  -0.41043515 -0.0006662173
#> [211,]  1.406655e+01   1.0934177   3.80559292   3.38439376  0.0340562089
#> [212,]  6.747270e+00   2.1903478  -3.79434088  12.66016217  0.0423673738
#> [213,] -4.903225e-01   2.3059747   4.49729107  14.11440532 -0.0377279501
#> [214,] -1.288186e+01   6.1269682   1.27516938   4.46002530  0.0164545753
#> [215,]  2.360657e+00  12.2801559  -4.85289809   6.71375095 -0.0312253740
#> [216,] -2.467914e-01 -12.5212837   0.53778172   8.23816365 -0.0465212737
#> [217,]  1.264390e+00  -1.5776002  10.18122625  10.82844071  0.0079668593
#> [218,]  1.925921e+00   6.1081607  -1.07959230 -13.52093497 -0.0172529313
#> [219,] -1.197266e+01   2.2945044   0.37909059   8.73194715  0.0377933598
#> [220,]  7.963130e+00   4.7706985  -3.48125160 -11.25654819  0.0288388964
#> [221,]  2.813432e+00  14.0397717  -0.69694474  -4.41403201  0.0317813022
#> [222,]  5.802850e+00  11.1298757   4.28586814   7.00600722  0.0252395107
#> [223,] -2.313957e+00  -1.1542448  14.76757254   0.48179363  0.0289578733
#> [224,] -2.863086e-01  -9.2877322   6.31746184  -9.93708879  0.0246742840
#> [225,] -6.260542e+00  -5.2301818   4.53752018  11.74145338  0.0498130434
#> [226,] -9.686265e+00  10.3620209   4.82464221   0.72637628  0.0209366097
#> [227,] -1.729272e+00  -9.5370993  10.76659489  -3.89021723 -0.0439008097
#> [228,] -9.432607e+00  -9.9151479  -1.00719331  -6.05816271  0.0150919481
#> [229,] -2.649804e-01 -11.3567014  -1.07416021  -9.73659581  0.0073373328
#> [230,]  5.146779e+00  13.8105220  -2.16697176   1.75624136  0.0234213716
#> [231,]  2.950395e+00  14.1330896   0.85882767   3.97660209  0.0141660189
#> [232,]  2.933099e+00  10.2068383   7.84049905   7.12347938  0.0067751970
#> [233,]  4.959096e+00  -6.5170095  -4.46089223  11.74888897  0.0005198856
#> [234,]  5.784224e+00  10.8571204  -5.72746259  -6.39232799  0.0084695989
#> [235,] -1.741703e+00   6.0193638  -4.63444779 -12.81622500  0.0148915792
#> [236,] -3.568842e+00  12.8957768   3.97935497   5.48881074  0.0499511852
#> [237,] -1.737129e+00 -11.6381890   4.02364619  -8.38720523  0.0215529104
#> [238,] -7.343435e+00  -8.3696116  -1.57133006  -9.92746094 -0.0224976930
#> [239,]  5.873549e+00 -10.2388429   9.13934103   1.46286188 -0.0293394150
#> [240,] -3.258355e+00 -11.8746077  -5.94651451   6.16569356 -0.0299163764
#> [241,] -2.326852e+00   8.5377975  11.96502789   1.87879804  0.0349481124
#> [242,]  9.844870e+00   0.3790023   9.29610263  -6.44339683 -0.0348385172
#> [243,]  1.153996e+01   1.0240071   6.23789101   7.20204614  0.0155268652
#> [244,]  4.015035e+00   3.1997047   5.83881613  12.82768916 -0.0146783921
#> [245,] -3.616183e+00  -1.2486459  11.13029782  -9.29949339  0.0087095233
#> [246,] -8.849682e+00   5.5241820  -5.34707361   9.35817021 -0.0145297334
#> [247,]  1.576192e+00  -2.2935087  -5.07132447  13.83969311 -0.0396200011
#> [248,] -9.920493e+00   4.2535769  -8.67910074   5.75882848  0.0065742848
#> [249,] -1.127291e+01   7.3586252  -4.98179081   4.35361982  0.0119224151
#> [250,] -1.078669e+01  -7.4303820  -7.14503829  -1.54440822 -0.0481219360
#> [251,]  6.101343e+00   2.6449235   9.46266452   9.55175222  0.0471485959
#> [252,] -4.343488e+00  -4.9995658  -8.47706357  10.45360463 -0.0173415651
#> [253,] -9.899301e+00  -1.2720325   3.90435512  10.49484601 -0.0351459031
#> [254,] -7.795089e+00  -3.3292664   8.37613891   9.11004218  0.0066405958
#> [255,]  7.662999e+00   9.5762619  -0.18169364   8.63369199 -0.0443197683
#> [256,] -5.689410e+00   6.8984352   2.80516529  11.71209871  0.0476516780
#> [257,] -1.922015e+00   8.3377654  -5.45226427 -11.04809212 -0.0063464875
#> [258,]  3.830009e+00   6.2297861  -6.39921233 -11.42676151 -0.0081631593
#> [259,] -1.081935e+01  -5.0949921   0.07379534  -9.05412590  0.0091076988
#> [260,] -1.192374e+01   9.0317243   0.20137630   1.10084331  0.0400294757
#> [261,]  1.334770e+01   1.8321066   6.15337949  -2.37029657  0.0094354750
#> [262,]  1.260624e+01   2.3516991  -0.57733529  -7.76008189  0.0172927913
#> [263,]  7.712801e+00  -3.7232812   0.36878666  12.30909698 -0.0088471429
#> [264,] -1.112173e+00  10.4624738  -4.67768061  -9.61348101 -0.0489781593
#> [265,] -7.773170e+00   8.7279151  -6.81939924  -6.47279810  0.0202286234
#> [266,] -1.095984e+01  -5.1945286   7.50018972  -4.65250918 -0.0271672612
#> [267,]  9.988769e+00 -10.6709125  -3.13322859   1.24056146  0.0030188583
#> [268,]  3.281927e+00  13.5028246  -2.44401167   5.09210058  0.0202359326
#> [269,] -3.609230e+00  11.5993748  -6.54198154  -5.88476347  0.0332122876
#> [270,] -2.041081e+00  11.6966460   8.94700036   1.99340074  0.0120355370
#> [271,] -4.793521e+00  -8.0546047   1.65594273  11.59324618 -0.0487785397
#> [272,] -8.352441e+00 -11.5210749   1.51664209   4.49459183  0.0415468803
#> [273,] -9.662554e-01   6.1352049   2.41468653  13.43856022  0.0226467820
#> [274,] -5.156586e+00  -1.2220370   0.02695507 -14.03265911  0.0479416980
#> [275,] -1.257502e+01   2.4917017   5.92538536  -5.05472136 -0.0275996075
#> [276,]  4.263991e+00   3.5727588   1.92748272  13.79632500 -0.0491144868
#> [277,]  2.429987e+00  -7.7886323   9.50740405   8.24873545  0.0474935835
#> [278,] -6.546434e+00  -1.9981465   5.65090778  12.09209875  0.0343168010
#> [279,]  5.630797e-01 -12.0449938   8.63408132  -2.24804468  0.0389112024
#> [280,]  1.116486e+01   0.3385444  -3.64510565   9.32440472 -0.0258209685
#> [281,]  5.214858e+00   1.2396337  13.76184755  -2.62299756 -0.0121918444
#> [282,] -5.092339e-01  14.3658399   4.17749227  -0.95492601 -0.0420192374
#> [283,] -3.927868e-01   2.4696634 -10.97167339  -9.91810788 -0.0283980956
#> [284,] -8.092701e+00  -1.9250121 -11.03176894  -5.83974207  0.0383471697
#> [285,]  2.373941e-01  10.7309576  -7.18455847   7.62707754  0.0492783289
#> [286,]  8.604068e-01  -1.4729960   9.47561239  11.50229337  0.0054242202
#> [287,]  5.688089e-01 -10.2114724   0.36667921 -10.96666928  0.0221705060
#> [288,]  9.939153e+00   2.3166121  -7.08579014  -8.40464947 -0.0055958019
#> [289,]  1.420468e+00   7.5347762  -4.80773703 -11.96223574 -0.0373771026
#> [290,]  5.680120e+00   6.1578060   7.89684978  -9.61547846  0.0339889998
#> [291,] -1.288804e+01  -3.1411222  -6.99383084  -0.34364973  0.0487770563
#> [292,] -4.808325e+00   1.5816389  -2.38200393 -13.91777605  0.0255775591
#> [293,] -7.658863e+00  -0.5642634  10.80782491  -7.01529337  0.0250430639
#> [294,] -1.270061e+01   7.6247653  -1.18040026  -2.04059076  0.0071043166
#> [295,] -8.516547e-01  -0.6431432  13.13546728  -7.16383629 -0.0003512936
#> [296,] -1.934801e+00  -7.5972493 -12.78752935  -0.13206265  0.0289206893
#> [297,] -1.780937e+00 -11.0887807   8.80577306  -4.61796142 -0.0176823216
#> [298,]  2.363835e+00  -1.8018569  14.64804690  -1.26503705 -0.0175330482
#> [299,] -3.671368e+00  11.1482007  -9.23543509  -1.39478225  0.0177750885
#> [300,]  6.615401e+00   3.4966744 -12.99476404   0.38189480  0.0201823969
#> [301,] -9.616325e-01  -6.6297852 -11.38158095   7.11201980 -0.0324931980
#> [302,]  7.139326e+00   5.6132255   9.48018410   7.25588219 -0.0368393197
#> [303,]  7.001406e+00  12.0696615  -2.45497599   4.92713701  0.0080396352
#> [304,]  4.194039e+00  -5.7474796 -12.86523835   2.97693736  0.0251980245
#> [305,]  1.107671e+01  -0.1375287  -9.89445805  -2.09460102 -0.0443637070
#> [306,]  1.915624e+00 -13.8006838  -1.62205235  -5.31417526  0.0489498000
#> [307,] -1.456086e+01   2.9968474   1.07244645   1.68823160  0.0456054189
#> [308,]  7.382809e+00   7.3004639   6.18409976  -8.88562131 -0.0337738652
#> [309,]  4.729002e+00  -4.7145921   4.17652786 -12.76580486  0.0063907456
#> [310,]  1.639165e+00  -1.7163027   8.81442146 -11.90266435 -0.0121837748
#> [311,]  9.089541e+00   2.9677468   6.89941099   9.27204725  0.0188696034
#> [312,]  2.008099e+00 -10.9926734   3.21897872  -9.47453677 -0.0301792020
#> [313,]  8.591685e+00  10.6389620   4.92183389   3.71092720  0.0013068495
#> [314,] -1.857813e+00   2.6869798   8.04298179 -12.23270671 -0.0135454748
#> [315,]  6.324164e+00  13.4493407   1.67275875   1.14981046 -0.0301305974
#> [316,] -1.010907e+01   4.1864156   8.80739641   5.26406806 -0.0139486994
#> [317,]  7.426178e+00   1.3473405  -8.20246936  10.03773105 -0.0007730641
#> [318,]  8.160435e+00   3.9655093  11.92837062   0.62929337 -0.0167835000
#> [319,] -2.444585e+00  -2.6905832  -3.04709767  14.23024815 -0.0445043142
#> [320,]  5.618025e-01  -8.6944331 -11.34400106   4.51717285  0.0294958376
#> [321,]  5.457828e-01 -13.3449313   3.48782671  -5.86941178  0.0191389521
#> [322,]  6.967787e+00  -0.3928122  -1.39622242  13.20402214  0.0355913641
#> [323,] -6.601998e+00  -9.8862271   4.66512722  -7.86846390  0.0213377306
#> [324,] -9.729482e+00   8.1284167  -4.15064370  -6.85843793 -0.0412513025
#> [325,] -1.090919e+01  -9.7585611   2.10844652   2.51287343 -0.0168443735
#> [326,] -1.015226e+01  -4.8515264  -1.26262148  -9.83870535  0.0117259522
#> [327,]  3.422743e+00  -4.2795408  -9.49259619 -10.24016491  0.0010004979
#> [328,]  5.793852e+00  -5.4438667   3.18892982  12.31366412 -0.0200507615
#> [329,]  1.451151e+01  -1.4261359  -3.51768214  -0.08999197  0.0482584288
#> [330,]  9.947785e+00  -6.4327738   6.13905759  -6.85368321  0.0150716881
#> [331,] -8.989630e+00  -0.5002622   9.26244619   7.62518072  0.0197244098
#> [332,] -3.889283e+00  -1.0987605  13.20717247  -5.85122162 -0.0392193277
#> [333,]  7.027719e+00  -5.7880367   8.65047514  -8.20238198 -0.0110151795
#> [334,] -1.100248e+01  -4.3044644  -8.79589572  -2.83711394  0.0255675613
#> [335,] -5.158345e+00  -6.7870410   8.41649678  -9.02718859  0.0371528264
#> [336,]  9.565078e+00   0.6410726  -7.48928832   8.77546969 -0.0436272402
#> [337,] -8.242143e+00  11.1118246   3.03372756  -4.93871714  0.0230249973
#> [338,] -3.095241e+00   9.9287781   7.40289870   7.87628981  0.0025096398
#> [339,] -1.125682e+01   4.6596843  -0.40141918  -8.74128772  0.0246775269
#> [340,] -6.272322e+00  12.1563488   2.14345315   5.76946843  0.0374241393
#> [341,] -1.383272e+01   2.0611429   5.28786280  -1.20255747  0.0193517477
#> [342,] -1.266973e+00  -5.4616041   1.36795113  13.84537359 -0.0435180277
#> [343,]  1.116233e+01   9.3112266  -1.89831796  -3.17802825  0.0305815461
#> [344,]  4.366893e+00   1.9413864  14.05183559   2.16960362 -0.0244791933
#> [345,] -5.137379e-01  -6.9971518   3.01251435 -12.91126245 -0.0314262797
#> [346,] -5.886666e+00  -6.6264282  10.07804464  -6.69855409 -0.0084417328
#> [347,] -1.223824e+01  -5.2666574  -1.63086028  -6.69538212 -0.0362624110
#> [348,] -9.374832e+00  11.1784592  -3.26631451   1.21892337 -0.0331478010
#> [349,] -5.923815e+00   5.2672756   4.63903513  11.85932457  0.0249309478
#> [350,] -4.878711e-01  -1.5403600  14.72742595  -2.34354381  0.0295108484
#> [351,] -3.163334e+00   1.7255592   5.33199327 -13.54937681 -0.0191200798
#> [352,]  2.944536e+00 -12.4076970   6.69099897   4.19634303  0.0094126848
#> [353,]  2.348976e+00 -13.5433828   0.51319660  -5.98295263  0.0058763724
#> [354,]  1.103218e+01   9.4905496  -3.41476220  -1.24895300  0.0083123195
#> [355,] -5.146886e+00   9.5528756  -9.35051994  -4.45195574 -0.0040908035
#> [356,]  6.233178e+00  -2.8640007 -10.54023617  -8.17608786 -0.0331562624
#> [357,] -5.768914e+00  13.4533154  -3.27367724  -0.10477261 -0.0459406769
#> [358,]  3.325291e+00 -13.4852636   0.20964960  -5.66093191 -0.0263532790
#> [359,] -6.636595e+00  -6.2371978  -7.90681344   8.91825510 -0.0377790165
#> [360,] -8.423993e+00 -12.2412979   1.53866536  -1.34887738 -0.0054815637
#> [361,]  9.456365e+00 -11.3231261   0.51161468   2.66500027 -0.0479714178
#> [362,] -9.569778e+00  -4.6735220 -10.36430031   2.03931781 -0.0390876873
#> [363,] -1.830245e+00   4.8983717  -8.66326397 -11.07266969  0.0026662960
#> [364,]  7.658369e+00  -0.1180079   7.27564025  10.64896784 -0.0421718092
#> [365,] -1.083678e+00   6.6462817   4.09770421  12.76171622  0.0086325987
#> [366,]  1.245455e+00  -5.8922090 -13.15028261   3.97501968 -0.0156443051
#> [367,] -5.686703e+00   3.9409180   7.74605619 -10.82262399  0.0300898795
#> [368,]  9.558439e+00  -3.5360496   0.19833526  11.00423794 -0.0057084480
#> [369,] -9.755008e+00   1.0370209  -5.51211600  -9.91871888 -0.0045738742
#> [370,]  4.377580e+00   8.3656350 -10.89590103   4.13911683  0.0266026248
#> [371,]  1.038034e+01  -7.2027213   5.74751514   5.68641611 -0.0297770931
#> [372,]  4.865054e+00   2.6859710  -2.19506517  13.75857894  0.0069560290
#> [373,]  1.223262e+01  -6.5354364   2.34147043  -5.21234139 -0.0483589163
#> [374,] -5.632518e-01  11.4244058  -6.54340525   7.16585992 -0.0383498276
#> [375,] -8.497250e+00   8.8021081  -4.95480144   7.12527757  0.0288464078
#> [376,]  1.124544e+01  -5.2335342   8.14164935  -2.20540338  0.0343661637
#> [377,] -1.508826e+00  14.6065483   0.06587007  -3.06069475 -0.0493101237
#> [378,]  6.359486e+00   5.5364983 -12.29675017  -1.64135842  0.0211965375
#> [379,]  7.610205e+00  -4.6864353   4.86232140  11.02179393  0.0039784590
#> [380,]  2.055110e+00  -3.6011972 -13.55677689   4.90119376 -0.0197525343
#> [381,]  7.928413e+00  10.0450280   0.07705623  -7.82507090  0.0404901317
#> [382,]  7.597167e+00  -0.4572633  -9.61905517   8.63410302 -0.0066356559
#> [383,] -1.158818e+01   3.6495535  -8.16164635   3.28364379  0.0371381865
#> [384,]  2.948488e-05   0.7565957   5.27714528  14.02067404 -0.0305965141
#> [385,] -7.345857e+00  -0.2857074  -3.70637769 -12.53872086 -0.0169804489
#> [386,]  6.388330e+00 -12.2079169   1.01078581  -5.84245759  0.0315736965
#> [387,] -9.826952e-01   1.4349522  14.78652375  -1.82590742  0.0017287161
#> [388,] -7.438338e+00   6.0670523  -9.84167139  -6.00029203 -0.0411404355
#> [389,] -5.751716e+00  -2.0014867  13.56487861   1.97633191  0.0464021380
#> [390,]  6.545980e+00  -0.7935336   3.63324328  12.97381955  0.0140651580
#>                 [,6]          [,7]          [,8]
#>   [1,] -0.0498393067  1.892525e-02 -4.563111e-02
#>   [2,]  0.0451536968  2.773644e-02 -4.336417e-02
#>   [3,] -0.0280449965 -3.173620e-02 -1.789050e-02
#>   [4,] -0.0056458567 -1.118437e-02 -2.029913e-02
#>   [5,] -0.0398530548 -1.424695e-02 -4.984841e-02
#>   [6,] -0.0207916648  7.818178e-03  1.173335e-02
#>   [7,]  0.0219589043  4.924428e-02 -3.183726e-02
#>   [8,]  0.0124213236 -1.900522e-02  1.706885e-02
#>   [9,]  0.0044157534  1.412843e-02 -5.342047e-03
#>  [10,]  0.0092767084 -4.196112e-04 -3.112296e-02
#>  [11,]  0.0135475572 -1.610557e-02  1.082736e-02
#>  [12,] -0.0041835041 -2.821384e-02 -2.181403e-03
#>  [13,] -0.0391406306  1.222711e-02 -1.912578e-02
#>  [14,]  0.0290602097 -2.007073e-02  1.283292e-02
#>  [15,]  0.0102725765  4.311229e-02 -5.704142e-03
#>  [16,] -0.0481411841  3.963124e-02 -4.272579e-02
#>  [17,] -0.0009434334  9.426901e-03  5.579568e-03
#>  [18,] -0.0251287451 -3.434017e-02  2.995291e-03
#>  [19,] -0.0334533131 -2.814454e-02  2.950437e-02
#>  [20,]  0.0226924400 -1.842447e-02 -1.016878e-02
#>  [21,]  0.0153882534  3.040888e-03  4.225628e-02
#>  [22,] -0.0192061445  4.902254e-02  1.620462e-02
#>  [23,] -0.0001013235  9.129078e-03 -2.439108e-03
#>  [24,]  0.0168872097 -4.472413e-02  2.472372e-02
#>  [25,] -0.0212829936 -3.133268e-02  3.343260e-02
#>  [26,] -0.0296480511  3.067976e-02  3.714496e-02
#>  [27,] -0.0090389605  3.023662e-02  1.980035e-02
#>  [28,]  0.0272115196  3.346905e-02  1.052646e-02
#>  [29,]  0.0213846744  2.747565e-02  8.395954e-03
#>  [30,]  0.0009698119  1.455804e-02 -8.922619e-03
#>  [31,]  0.0480860865 -2.990596e-02  4.706160e-02
#>  [32,] -0.0349666421  2.112739e-02  4.274642e-02
#>  [33,]  0.0098777179  2.228716e-02  4.160035e-02
#>  [34,] -0.0041983687 -2.395582e-02  2.749255e-02
#>  [35,] -0.0130523423 -1.184794e-04  3.034556e-02
#>  [36,] -0.0093946351  3.505941e-02  2.139725e-02
#>  [37,] -0.0121117136 -4.260565e-02 -4.414726e-02
#>  [38,] -0.0022403454 -1.231356e-02 -4.239714e-02
#>  [39,]  0.0369144969  2.576021e-02 -2.322742e-02
#>  [40,] -0.0088323499 -1.289703e-02 -5.148936e-03
#>  [41,]  0.0095764856  7.094609e-04 -3.302802e-02
#>  [42,]  0.0464114999 -2.497024e-02  3.299800e-02
#>  [43,]  0.0477102483 -4.421404e-03 -1.745677e-02
#>  [44,]  0.0474303407  4.587912e-02  1.934359e-02
#>  [45,]  0.0087080968  2.117285e-03  2.298787e-02
#>  [46,]  0.0178621636 -2.639497e-02 -3.891202e-02
#>  [47,] -0.0369462701  1.452840e-02 -2.021059e-02
#>  [48,] -0.0208844455 -2.406938e-02  6.770005e-03
#>  [49,] -0.0376464167  2.438022e-02  3.723678e-02
#>  [50,] -0.0228666042 -4.548817e-02  6.624447e-03
#>  [51,]  0.0485046670 -3.180283e-02  2.918043e-02
#>  [52,]  0.0461939265  2.836666e-02 -3.271358e-02
#>  [53,]  0.0065362056 -1.969667e-02  3.394993e-02
#>  [54,]  0.0114326499  3.775126e-02 -3.910166e-02
#>  [55,]  0.0024644357  3.777332e-02  3.341849e-02
#>  [56,]  0.0395197440  1.230545e-02 -5.263522e-03
#>  [57,]  0.0187534676 -7.668746e-03 -2.872346e-02
#>  [58,]  0.0359935214 -2.169839e-02  4.554617e-02
#>  [59,]  0.0124769656  2.171626e-02  1.509024e-02
#>  [60,] -0.0076072273  3.978454e-02 -2.076385e-02
#>  [61,]  0.0263257338 -3.875901e-02 -1.244687e-03
#>  [62,]  0.0044397798  4.607732e-02 -2.182944e-02
#>  [63,]  0.0267112524  4.860652e-02 -3.853270e-03
#>  [64,]  0.0228135129  1.253276e-02  1.308319e-02
#>  [65,] -0.0231531464  1.519430e-02  4.602874e-02
#>  [66,] -0.0067987888 -1.951061e-02 -3.955683e-02
#>  [67,]  0.0249986364  4.080286e-02  1.328174e-02
#>  [68,] -0.0293081690  1.368922e-02 -3.320095e-02
#>  [69,] -0.0442586954 -3.587512e-02 -2.450472e-02
#>  [70,] -0.0175510955  1.444008e-02 -5.767946e-03
#>  [71,] -0.0458632032 -2.769914e-02  8.453039e-03
#>  [72,] -0.0113224096  1.805635e-02 -4.966143e-02
#>  [73,]  0.0091238019  1.769735e-02  4.189141e-02
#>  [74,] -0.0015369967 -2.628551e-02 -4.459628e-02
#>  [75,]  0.0174324103  3.276527e-02  6.603105e-04
#>  [76,]  0.0096760882  1.013149e-02  2.804644e-02
#>  [77,] -0.0112215067 -1.719283e-02 -4.265122e-02
#>  [78,] -0.0406055247  2.110917e-02 -4.384847e-02
#>  [79,] -0.0376776326  4.910112e-02  4.953826e-02
#>  [80,] -0.0481057837 -1.143816e-02 -4.457870e-02
#>  [81,] -0.0109047671 -2.695412e-02 -1.484462e-02
#>  [82,]  0.0447565397  4.585967e-02  3.077961e-02
#>  [83,]  0.0310495166  2.277257e-03 -3.346440e-02
#>  [84,] -0.0330248760  2.807068e-02  3.969311e-02
#>  [85,] -0.0160083171  1.377502e-02 -3.472715e-02
#>  [86,] -0.0060951419 -3.469323e-02  3.008664e-02
#>  [87,] -0.0059534746 -1.203818e-02  4.103246e-02
#>  [88,] -0.0434921290 -4.940264e-02 -3.403479e-02
#>  [89,] -0.0303488126 -1.145720e-02 -8.310187e-05
#>  [90,] -0.0138705691 -4.293891e-03  9.379722e-03
#>  [91,]  0.0420159218  1.266595e-02  2.771935e-02
#>  [92,] -0.0433659005 -3.613023e-02  4.028728e-02
#>  [93,] -0.0495314758  9.678168e-03  4.560592e-02
#>  [94,]  0.0194746824  2.476464e-02  1.104355e-02
#>  [95,]  0.0065333394  1.782686e-02  4.564350e-02
#>  [96,]  0.0344118395 -8.513175e-04 -2.546260e-02
#>  [97,] -0.0085771780 -4.401581e-02  3.603365e-02
#>  [98,] -0.0363177381  1.642616e-02  4.528823e-02
#>  [99,] -0.0095317049  4.639688e-02  1.849691e-02
#> [100,]  0.0068524925  2.448232e-02  1.168250e-02
#> [101,] -0.0178026940  2.846342e-02  1.174071e-02
#> [102,] -0.0004171018  2.745074e-02  4.716641e-02
#> [103,]  0.0357460975  3.553626e-02  4.543943e-02
#> [104,] -0.0087324518  2.039715e-02  4.970216e-02
#> [105,]  0.0202622945 -2.548602e-03 -1.212756e-02
#> [106,] -0.0229893559  1.679754e-02 -3.217612e-02
#> [107,]  0.0346691798  3.302240e-02  4.040831e-02
#> [108,] -0.0405789726  2.592803e-02  3.446613e-02
#> [109,]  0.0196550015  1.957304e-02 -4.675201e-02
#> [110,]  0.0327333928 -9.092975e-03  2.296887e-02
#> [111,] -0.0396731389  2.582671e-02 -3.910353e-02
#> [112,]  0.0050500424 -1.058305e-02 -2.763894e-02
#> [113,]  0.0222523796 -3.779474e-02  4.141575e-02
#> [114,] -0.0226958374  3.454870e-02 -1.781756e-03
#> [115,] -0.0347799618  4.885079e-02  3.774724e-02
#> [116,] -0.0428095266  5.580309e-03 -4.033636e-03
#> [117,] -0.0191357206 -4.800598e-02 -4.698418e-02
#> [118,] -0.0096166594 -2.040473e-03  8.103914e-04
#> [119,]  0.0061355447 -1.992891e-02 -3.579828e-02
#> [120,] -0.0190324896 -1.517283e-02  5.924443e-03
#> [121,] -0.0297304905  1.621100e-02 -9.544471e-03
#> [122,] -0.0415984048  1.742515e-02 -1.271696e-02
#> [123,] -0.0140887571  3.552806e-02 -8.433280e-03
#> [124,] -0.0008866568 -4.905840e-02  3.532045e-02
#> [125,] -0.0453380776 -4.452618e-02 -4.803023e-02
#> [126,]  0.0385745028  2.603686e-02  2.460874e-02
#> [127,]  0.0264782653  9.833951e-03  6.349383e-03
#> [128,]  0.0056096796 -4.390994e-03  2.763158e-02
#> [129,] -0.0253685348  3.108823e-02 -4.156512e-02
#> [130,]  0.0401472228  3.216536e-02  5.304969e-03
#> [131,]  0.0312097506  2.333433e-02  4.974718e-02
#> [132,] -0.0359511758  1.756335e-02 -4.678972e-02
#> [133,]  0.0175254943 -2.766807e-02  2.210390e-02
#> [134,] -0.0024896817 -1.337244e-02 -1.742619e-03
#> [135,]  0.0114440351  1.404891e-02 -1.098029e-02
#> [136,]  0.0458712738  2.905269e-02  3.168231e-02
#> [137,] -0.0066626399  1.404969e-02  4.072385e-02
#> [138,]  0.0194861766 -8.627668e-03  1.003799e-02
#> [139,] -0.0492242091  4.171956e-02 -2.494683e-02
#> [140,]  0.0062035902  1.135051e-02 -6.233665e-03
#> [141,]  0.0415474713 -8.215701e-03 -3.094194e-02
#> [142,]  0.0102327159  4.274075e-02 -4.832630e-02
#> [143,]  0.0192527417  4.346928e-02  2.898367e-02
#> [144,] -0.0386294509 -4.989754e-02 -5.724461e-04
#> [145,]  0.0192914769 -1.497525e-02  7.214688e-03
#> [146,] -0.0301690357  3.397222e-02 -3.367245e-02
#> [147,] -0.0215733269 -1.529763e-02  7.559647e-03
#> [148,]  0.0061278763  4.191830e-03 -4.338885e-02
#> [149,] -0.0241728410 -3.207837e-03 -2.098144e-02
#> [150,]  0.0474393241 -3.339942e-02  7.317768e-03
#> [151,]  0.0215129375 -8.862964e-03  1.883193e-02
#> [152,] -0.0474182297 -2.996047e-02 -3.021234e-03
#> [153,]  0.0410129749 -2.253775e-02  4.200588e-02
#> [154,]  0.0189748268  6.566355e-03 -1.152966e-02
#> [155,]  0.0042228382  2.228097e-02  4.415447e-02
#> [156,] -0.0273235395 -3.555571e-02  4.058445e-02
#> [157,] -0.0286875791 -3.067091e-03 -1.752085e-02
#> [158,] -0.0423711489 -3.210432e-02 -2.058731e-02
#> [159,] -0.0218472056  6.210665e-03  4.271443e-02
#> [160,] -0.0309671241  2.765883e-02 -2.642306e-02
#> [161,] -0.0040772037 -2.314544e-03 -4.932971e-02
#> [162,] -0.0316639734 -4.910280e-02 -7.128664e-03
#> [163,] -0.0219678072  3.890200e-02  3.287940e-02
#> [164,] -0.0316782717 -4.409502e-03  1.592602e-02
#> [165,] -0.0075809374 -1.024877e-02 -4.146531e-02
#> [166,] -0.0019381610 -4.719092e-02  3.573030e-02
#> [167,]  0.0354225670 -1.622247e-02 -3.574065e-02
#> [168,]  0.0082201671 -3.915616e-02 -3.665961e-02
#> [169,] -0.0030426393  4.194219e-02  9.816264e-03
#> [170,]  0.0073778888 -4.726048e-03  2.768947e-02
#> [171,]  0.0244132529 -1.716937e-02  3.479357e-04
#> [172,] -0.0116682196  3.265369e-02  1.849963e-02
#> [173,] -0.0385070114 -4.315940e-02 -5.350806e-04
#> [174,]  0.0045250302  1.342565e-02  1.583998e-03
#> [175,] -0.0386503571  1.458245e-02 -2.230856e-02
#> [176,]  0.0132987844 -3.361651e-03 -4.395085e-02
#> [177,] -0.0005573324  2.109097e-02 -1.959253e-02
#> [178,] -0.0451539532  1.339186e-02 -2.626117e-02
#> [179,] -0.0035249788  4.177527e-02  4.452689e-02
#> [180,] -0.0113482851 -3.545911e-02  1.592490e-02
#> [181,]  0.0237902335 -4.011782e-02  4.893033e-04
#> [182,]  0.0250188419 -3.510428e-02  1.619489e-02
#> [183,] -0.0168281887 -2.538494e-03  2.557647e-02
#> [184,] -0.0361658940 -3.089631e-02  7.692141e-03
#> [185,] -0.0384284920  1.889410e-02 -2.120145e-02
#> [186,]  0.0482573789 -1.838645e-02  3.803211e-03
#> [187,] -0.0018153150  1.932340e-02  3.410767e-03
#> [188,] -0.0056891972  2.115003e-02  3.970867e-02
#> [189,]  0.0235091108 -4.159286e-02 -4.064302e-02
#> [190,] -0.0433118782  1.438999e-03 -4.146179e-02
#> [191,]  0.0318508651 -1.140734e-02  2.012408e-03
#> [192,] -0.0101559245 -2.942417e-02 -2.566237e-02
#> [193,]  0.0430584712 -3.676060e-02 -1.758677e-02
#> [194,] -0.0461112634  1.501521e-03 -3.856622e-02
#> [195,]  0.0328340234  3.045697e-02  7.863600e-03
#> [196,]  0.0091125147  4.624107e-02 -4.928260e-02
#> [197,]  0.0322095023  3.986439e-03  1.786176e-02
#> [198,] -0.0261450251 -8.180434e-03 -3.539434e-04
#> [199,]  0.0306117061 -4.743464e-02 -3.906098e-02
#> [200,]  0.0483416579  1.344230e-02  3.899200e-02
#> [201,] -0.0001050768  1.985278e-02 -1.316041e-02
#> [202,] -0.0242393107 -2.546511e-02  1.069089e-02
#> [203,] -0.0448149232 -1.282265e-03 -1.638010e-02
#> [204,] -0.0403402146  3.714462e-03  1.578708e-02
#> [205,]  0.0384512077  1.757342e-02 -1.960031e-02
#> [206,] -0.0311754012 -4.093652e-02 -3.857673e-02
#> [207,] -0.0094728988 -1.168775e-02  1.289417e-02
#> [208,] -0.0248839530 -4.919733e-03  1.273894e-02
#> [209,]  0.0084844274  1.621446e-02 -3.451278e-02
#> [210,] -0.0443554433  4.318053e-02 -2.363670e-02
#> [211,] -0.0108276782  4.596753e-02 -2.289036e-03
#> [212,]  0.0470620946 -4.229184e-02  6.148196e-03
#> [213,]  0.0268604745  9.410716e-03  3.492996e-02
#> [214,]  0.0485144481  1.695594e-02 -1.062045e-02
#> [215,] -0.0091415693 -3.825438e-04 -2.128679e-02
#> [216,] -0.0359738877 -1.822379e-02  4.265630e-03
#> [217,] -0.0039780248 -3.873556e-02  3.043189e-02
#> [218,] -0.0294111057  3.494445e-02  2.099480e-02
#> [219,]  0.0288458679  4.866843e-04  6.378680e-03
#> [220,] -0.0143177871 -3.070498e-02 -4.353564e-02
#> [221,]  0.0067108722 -2.495881e-02  4.261111e-02
#> [222,]  0.0100103488 -8.684962e-03 -1.114709e-02
#> [223,] -0.0493065814 -3.320803e-02  1.743370e-02
#> [224,]  0.0463613175 -1.224086e-02  1.208461e-02
#> [225,]  0.0001963110 -6.554595e-03 -1.304712e-02
#> [226,] -0.0114313675  3.721531e-02  4.693469e-02
#> [227,] -0.0096035828 -3.938856e-02 -4.591796e-02
#> [228,]  0.0045797634 -3.610840e-02  2.295358e-02
#> [229,]  0.0490253717 -4.261751e-02  1.247358e-03
#> [230,] -0.0240629866  2.503725e-02 -4.599403e-02
#> [231,]  0.0190400123  4.012388e-02  1.559033e-02
#> [232,]  0.0140088327 -1.574965e-02  2.489630e-02
#> [233,]  0.0141258125  1.204735e-02 -3.298581e-02
#> [234,]  0.0309725679 -6.011913e-05 -2.671830e-02
#> [235,] -0.0186676408 -1.472318e-02  1.606811e-02
#> [236,] -0.0076331608  1.686565e-02  2.180101e-02
#> [237,] -0.0051156343 -3.912067e-02 -2.708413e-02
#> [238,]  0.0163840514 -2.628905e-02  1.975996e-02
#> [239,]  0.0465137657  3.641918e-02 -2.777393e-02
#> [240,] -0.0078086089 -2.166993e-02  1.200812e-02
#> [241,]  0.0016921463 -2.837498e-02 -2.191100e-02
#> [242,]  0.0109608324  4.957882e-02  2.122510e-02
#> [243,]  0.0140914923 -4.504902e-02 -4.161556e-02
#> [244,]  0.0208868657 -4.511219e-02 -2.686163e-03
#> [245,] -0.0469652987 -4.122127e-03  4.278944e-02
#> [246,] -0.0213770616 -3.528759e-02 -2.564461e-02
#> [247,]  0.0049534994  9.541079e-03 -3.326321e-02
#> [248,] -0.0284848131 -2.644935e-02 -1.211097e-02
#> [249,] -0.0250173339 -2.102582e-02 -1.203359e-03
#> [250,] -0.0412070681  2.878174e-02 -3.231948e-02
#> [251,] -0.0223791803  3.820502e-02  4.877509e-02
#> [252,] -0.0026508256  6.878073e-03 -1.694598e-02
#> [253,]  0.0022635695 -4.660010e-02 -3.051479e-02
#> [254,] -0.0310788018  2.459840e-02 -2.716632e-03
#> [255,] -0.0392412572 -9.454527e-03 -1.079546e-02
#> [256,] -0.0122624468 -4.348477e-02 -2.849122e-02
#> [257,] -0.0392029011  3.673580e-02  2.401387e-02
#> [258,]  0.0131748912 -2.546252e-02  7.680290e-03
#> [259,] -0.0052440108 -4.086519e-02 -1.989539e-02
#> [260,]  0.0097453754  6.335734e-03  3.004749e-02
#> [261,]  0.0298328090  2.280405e-02 -1.201624e-02
#> [262,]  0.0242993464 -4.427036e-02 -3.776454e-02
#> [263,]  0.0467767697 -2.386628e-02 -4.335864e-02
#> [264,] -0.0444113135 -1.157363e-02  1.344331e-02
#> [265,]  0.0199388383  1.145387e-02  4.899851e-03
#> [266,]  0.0222760492 -2.663361e-02 -2.956107e-02
#> [267,]  0.0482019972 -5.546521e-03  1.598557e-02
#> [268,] -0.0062034270 -3.283619e-02 -2.899234e-02
#> [269,]  0.0179884742  3.434410e-02  2.661705e-02
#> [270,] -0.0016666973  2.807737e-02  7.585121e-03
#> [271,] -0.0369509361 -7.697356e-03  4.281208e-02
#> [272,] -0.0389772057 -3.596493e-02  4.217377e-03
#> [273,]  0.0217582253  4.144051e-02 -4.411354e-04
#> [274,] -0.0039795617 -6.714028e-03 -2.648989e-03
#> [275,] -0.0040027744 -4.147008e-02  3.994924e-02
#> [276,] -0.0252100952 -3.179607e-02  4.228412e-02
#> [277,] -0.0260198543  1.715398e-02  2.641614e-02
#> [278,]  0.0036594979 -1.740466e-02  4.165274e-02
#> [279,] -0.0153834191 -6.154371e-03 -8.826705e-05
#> [280,] -0.0153936563  3.352938e-02  3.228130e-02
#> [281,]  0.0066760563  3.554813e-03  3.932674e-02
#> [282,]  0.0428217589 -2.321333e-02  1.731362e-02
#> [283,]  0.0359568012 -2.806371e-02 -4.156599e-02
#> [284,]  0.0284178685 -1.125129e-02 -2.535150e-02
#> [285,]  0.0157947525  4.774582e-03  4.162187e-02
#> [286,]  0.0067502912  1.031262e-02  1.381409e-02
#> [287,] -0.0227620619 -8.767949e-03  1.945924e-02
#> [288,]  0.0356345365 -2.336387e-02 -1.548107e-02
#> [289,]  0.0194568714 -5.986653e-03 -2.720664e-03
#> [290,] -0.0082367483 -2.527578e-02  2.783154e-02
#> [291,]  0.0256606100  5.684780e-03 -1.167344e-02
#> [292,] -0.0289314437  2.386014e-02  4.063968e-02
#> [293,]  0.0167165421  2.264079e-02 -3.348219e-02
#> [294,]  0.0246609866 -1.745989e-03 -9.001625e-03
#> [295,]  0.0442919880 -4.697743e-02  1.940950e-02
#> [296,] -0.0351362764 -6.190349e-03  2.624619e-02
#> [297,] -0.0394620901  2.282712e-02  3.206745e-02
#> [298,] -0.0095372504 -4.750421e-02 -1.577218e-02
#> [299,] -0.0060365051 -3.391552e-02 -3.005664e-02
#> [300,]  0.0245002729 -1.960559e-02 -2.276955e-02
#> [301,] -0.0341258698  3.612065e-03  1.077191e-02
#> [302,]  0.0378875782 -2.879768e-03 -2.502405e-03
#> [303,]  0.0493627586 -1.160917e-03  3.772787e-02
#> [304,]  0.0438200423 -4.224137e-02 -4.972048e-02
#> [305,]  0.0479409811  3.014354e-02  3.874291e-02
#> [306,] -0.0372170412  1.189752e-03  4.220027e-02
#> [307,] -0.0445425876  1.177683e-02  2.713626e-02
#> [308,]  0.0347888250 -3.570988e-02 -2.770005e-02
#> [309,]  0.0230775382 -3.005946e-02  3.941449e-02
#> [310,]  0.0490263238 -1.363559e-02  3.332537e-03
#> [311,] -0.0122115753  3.585409e-03 -1.424389e-02
#> [312,]  0.0250296060 -3.249844e-02  4.219774e-02
#> [313,] -0.0332818319  7.494461e-03 -3.780077e-04
#> [314,] -0.0356689838 -1.561549e-02  2.618552e-02
#> [315,]  0.0437616044  2.748106e-02  3.602699e-02
#> [316,]  0.0409066749  7.252112e-03 -3.802825e-02
#> [317,]  0.0299123898  3.835307e-02  3.304868e-02
#> [318,]  0.0381772182 -4.017485e-02  2.830639e-02
#> [319,] -0.0396578943  2.082661e-02 -2.356037e-02
#> [320,]  0.0055896225 -4.536841e-02 -3.251191e-02
#> [321,]  0.0104021395  3.983337e-02 -7.482458e-03
#> [322,]  0.0268479078 -2.046103e-02  1.728910e-02
#> [323,]  0.0187878912  9.955562e-03 -3.866947e-02
#> [324,] -0.0159617424 -4.162090e-02 -2.992925e-02
#> [325,]  0.0133452800  1.262877e-02 -3.301808e-02
#> [326,] -0.0274410672  1.807809e-02  6.645374e-03
#> [327,]  0.0219067287  3.839458e-02 -1.500121e-02
#> [328,] -0.0177262414  1.920623e-02  1.136378e-02
#> [329,]  0.0114701733 -2.973645e-03 -1.682999e-02
#> [330,] -0.0276317285 -3.627890e-02 -3.086636e-02
#> [331,]  0.0189099878  2.531115e-02 -4.466704e-02
#> [332,]  0.0124666015 -2.165974e-02  1.600684e-02
#> [333,]  0.0169812954  4.020626e-02 -5.458934e-03
#> [334,] -0.0378901855 -1.255083e-02  3.393768e-03
#> [335,] -0.0161097038 -7.836413e-03  1.585014e-03
#> [336,] -0.0067666897  5.436135e-03 -2.092009e-02
#> [337,]  0.0092469468 -3.670947e-02  4.643310e-02
#> [338,] -0.0331507916 -2.513314e-02  2.690083e-02
#> [339,] -0.0006146387 -3.756081e-02  2.811939e-02
#> [340,]  0.0245606462  3.980431e-02 -3.309001e-02
#> [341,] -0.0378229271 -3.459597e-02 -1.708869e-03
#> [342,] -0.0122767020  4.271604e-02 -3.511011e-02
#> [343,] -0.0453672324 -7.897625e-03  6.592443e-03
#> [344,]  0.0061189950 -3.571250e-02 -1.999755e-02
#> [345,] -0.0219731480  7.030366e-03  4.611906e-02
#> [346,] -0.0215335082 -1.379323e-02 -9.842820e-03
#> [347,]  0.0077105954 -4.455672e-03 -1.869424e-02
#> [348,]  0.0047630362  4.320028e-03 -4.127390e-02
#> [349,] -0.0277419638  1.244168e-02 -1.062900e-02
#> [350,]  0.0471964790  9.374141e-03  2.048466e-02
#> [351,] -0.0136764089 -1.342988e-02 -3.895319e-02
#> [352,] -0.0402014994 -1.611467e-02  4.605984e-02
#> [353,] -0.0420707073 -4.028547e-02  1.419915e-03
#> [354,] -0.0485590968 -4.579203e-02  3.075363e-02
#> [355,] -0.0122644665  3.170862e-02 -7.554341e-03
#> [356,] -0.0261378356  9.502372e-03 -3.761549e-02
#> [357,] -0.0376238571 -9.332806e-03 -1.057898e-02
#> [358,] -0.0292970977 -3.493982e-02  4.681717e-03
#> [359,]  0.0023666020 -1.795966e-02  6.884452e-03
#> [360,]  0.0211908803 -1.894733e-02 -2.317482e-02
#> [361,] -0.0398905383 -1.402784e-02  3.112008e-02
#> [362,]  0.0249748567  1.290207e-02 -3.618474e-02
#> [363,] -0.0396012254  2.612731e-02  4.721989e-02
#> [364,]  0.0380786851 -4.482053e-02  2.466123e-02
#> [365,] -0.0058208612  4.855641e-02 -2.402000e-02
#> [366,]  0.0122504012  2.384517e-02  3.829150e-02
#> [367,] -0.0200945648  4.360736e-02 -4.258246e-02
#> [368,] -0.0227118198  4.735805e-02 -2.860262e-02
#> [369,]  0.0399813341 -2.473133e-02 -9.667856e-03
#> [370,]  0.0317369873  1.732120e-02  4.902979e-02
#> [371,] -0.0437110557 -1.853077e-02 -1.638077e-02
#> [372,] -0.0101894685  1.144685e-02 -1.339599e-02
#> [373,]  0.0390691745  2.782121e-02 -1.936421e-02
#> [374,]  0.0103816213  2.604611e-02 -4.428517e-02
#> [375,]  0.0136437392 -2.413758e-02 -3.808414e-02
#> [376,] -0.0006388123 -1.497372e-02 -4.703322e-02
#> [377,] -0.0441887991 -4.220667e-02 -2.573769e-02
#> [378,] -0.0337064095 -4.258247e-02  1.076627e-02
#> [379,] -0.0465702948  2.790292e-02 -4.291149e-02
#> [380,] -0.0255896111  2.044601e-02 -1.358779e-02
#> [381,] -0.0260629578 -3.661121e-02  2.472590e-02
#> [382,]  0.0449200979  2.642213e-02 -3.467630e-02
#> [383,] -0.0323711666 -3.552169e-02 -2.137445e-02
#> [384,] -0.0232043326 -1.040866e-02  1.737053e-02
#> [385,] -0.0444635089 -4.030726e-02 -4.218691e-02
#> [386,]  0.0014560674 -4.051596e-02  1.548752e-02
#> [387,] -0.0259912651  2.916617e-02 -4.877649e-03
#> [388,] -0.0486039246 -2.406548e-02  4.718884e-02
#> [389,]  0.0147767596  3.517648e-02  2.539332e-02
#> [390,]  0.0463814764 -2.586840e-02 -4.847059e-02