Generate Coordinates for a Sphere

This function generates the coordinates for a sphere in three-dimensional space.

Usage

sphere(radius, resolution, num_noise, min_n, max_n)

Arguments

radius

The radius of the sphere.

resolution

The number of points used to approximate the surface of the sphere.

num_noise

The number of additional noise dimensions to add to the coordinates.

min_n

The minimum value for the random noise added to the coordinates.

max_n

The maximum value for the random noise added to the coordinates.

Value

A matrix containing the Cartesian coordinates of the points on the sphere.

Examples

# Generate coordinates for a sphere with radius 1 and resolution 100
sphere(radius = 1, resolution = 20, num_noise = 3, min_n = -0.05,
max_n = 0.05)
#>                 [,1]          [,2]        [,3]          [,4]          [,5]
#>   [1,]  0.000000e+00  0.000000e+00  1.00000000 -4.153424e-02  0.0298458060
#>   [2,]  0.000000e+00  0.000000e+00  1.00000000  1.330453e-03 -0.0456850784
#>   [3,]  0.000000e+00  0.000000e+00  1.00000000  3.014497e-02  0.0302826036
#>   [4,]  0.000000e+00  0.000000e+00  1.00000000  4.636038e-02  0.0475745450
#>   [5,]  0.000000e+00  0.000000e+00  1.00000000 -3.132190e-02 -0.0293971367
#>   [6,]  0.000000e+00  0.000000e+00  1.00000000 -2.821166e-02  0.0379506500
#>   [7,]  0.000000e+00  0.000000e+00  1.00000000  2.075588e-02  0.0304768508
#>   [8,]  0.000000e+00  0.000000e+00  1.00000000  3.211691e-02 -0.0344444978
#>   [9,]  0.000000e+00  0.000000e+00  1.00000000  3.301465e-02 -0.0313493622
#>  [10,]  0.000000e+00  0.000000e+00  1.00000000 -2.460401e-02 -0.0411413224
#>  [11,]  0.000000e+00  0.000000e+00  1.00000000 -1.970888e-02 -0.0259513479
#>  [12,]  0.000000e+00  0.000000e+00  1.00000000  8.650882e-03  0.0134880611
#>  [13,]  0.000000e+00  0.000000e+00  1.00000000  5.319524e-03  0.0241032335
#>  [14,]  0.000000e+00  0.000000e+00  1.00000000  3.055015e-02  0.0198172767
#>  [15,]  0.000000e+00  0.000000e+00  1.00000000 -2.034798e-02  0.0022318049
#>  [16,]  0.000000e+00  0.000000e+00  1.00000000 -4.585429e-02 -0.0011847168
#>  [17,]  0.000000e+00  0.000000e+00  1.00000000  2.401503e-02 -0.0351867679
#>  [18,]  0.000000e+00  0.000000e+00  1.00000000 -2.929335e-03  0.0342258775
#>  [19,]  0.000000e+00  0.000000e+00  1.00000000 -7.021707e-03  0.0427423366
#>  [20,]  0.000000e+00  0.000000e+00  1.00000000  1.424632e-02  0.0356535020
#>  [21,]  1.645946e-01  0.000000e+00  0.98636130  4.286386e-03  0.0023285276
#>  [22,]  1.556764e-01  5.344378e-02  0.98636130 -4.251831e-02  0.0147887924
#>  [23,]  1.298883e-01  1.010961e-01  0.98636130  4.794494e-02 -0.0196194176
#>  [24,]  9.002471e-02  1.377931e-01  0.98636130  4.837821e-02 -0.0241319104
#>  [25,]  4.040558e-02  1.595580e-01  0.98636130 -1.632427e-02 -0.0213098814
#>  [26,] -1.359211e-02  1.640324e-01  0.98636130  1.650358e-02 -0.0199147605
#>  [27,] -6.611689e-02  1.507313e-01  0.98636130  2.276357e-02 -0.0382905800
#>  [28,] -1.114769e-01  1.210962e-01  0.98636130  6.451721e-03  0.0129945291
#>  [29,] -1.447566e-01  7.833837e-02  0.98636130 -3.542232e-02  0.0123733278
#>  [30,] -1.623497e-01  2.709138e-02  0.98636130 -3.759895e-03 -0.0454226929
#>  [31,] -1.623497e-01 -2.709138e-02  0.98636130 -3.011540e-02  0.0272900101
#>  [32,] -1.447566e-01 -7.833837e-02  0.98636130  4.962271e-02 -0.0128050873
#>  [33,] -1.114769e-01 -1.210962e-01  0.98636130 -4.935850e-02  0.0256310399
#>  [34,] -6.611689e-02 -1.507313e-01  0.98636130  4.097835e-02  0.0313089143
#>  [35,] -1.359211e-02 -1.640324e-01  0.98636130 -3.314469e-02  0.0457906537
#>  [36,]  4.040558e-02 -1.595580e-01  0.98636130  2.008178e-02  0.0281507588
#>  [37,]  9.002471e-02 -1.377931e-01  0.98636130  4.583084e-02 -0.0210759732
#>  [38,]  1.298883e-01 -1.010961e-01  0.98636130  4.068251e-02  0.0253460482
#>  [39,]  1.556764e-01 -5.344378e-02  0.98636130 -3.215369e-02 -0.0425790761
#>  [40,]  1.645946e-01 -4.031405e-17  0.98636130 -4.660487e-02  0.0274465992
#>  [41,]  3.246995e-01  0.000000e+00  0.94581724  3.455330e-02  0.0287640764
#>  [42,]  3.071064e-01  1.054297e-01  0.94581724 -1.450290e-02  0.0461553073
#>  [43,]  2.562335e-01  1.994345e-01  0.94581724 -1.407460e-02 -0.0467022258
#>  [44,]  1.775938e-01  2.718275e-01  0.94581724 -1.293291e-02 -0.0397807753
#>  [45,]  7.970901e-02  3.147638e-01  0.94581724 -3.028897e-02 -0.0340000546
#>  [46,] -2.681347e-02  3.235905e-01  0.94581724 -3.294745e-02  0.0009033159
#>  [47,] -1.304303e-01  2.973511e-01  0.94581724  2.861514e-03  0.0081764339
#>  [48,] -2.199130e-01  2.388892e-01  0.94581724  2.666948e-02 -0.0076837518
#>  [49,] -2.855647e-01  1.545399e-01  0.94581724 -1.328228e-02  0.0262243232
#>  [50,] -3.202710e-01  5.344378e-02  0.94581724  4.785917e-02  0.0095507334
#>  [51,] -3.202710e-01 -5.344378e-02  0.94581724 -2.735554e-02  0.0076660234
#>  [52,] -2.855647e-01 -1.545399e-01  0.94581724 -4.558347e-02  0.0344668614
#>  [53,] -2.199130e-01 -2.388892e-01  0.94581724 -1.382789e-02 -0.0308570714
#>  [54,] -1.304303e-01 -2.973511e-01  0.94581724 -2.062849e-02 -0.0235655772
#>  [55,] -2.681347e-02 -3.235905e-01  0.94581724  8.794027e-03  0.0247192561
#>  [56,]  7.970901e-02 -3.147638e-01  0.94581724  2.130544e-02  0.0467817533
#>  [57,]  1.775938e-01 -2.718275e-01  0.94581724 -4.944957e-02  0.0395030997
#>  [58,]  2.562335e-01 -1.994345e-01  0.94581724 -4.859961e-02  0.0144045302
#>  [59,]  3.071064e-01 -1.054297e-01  0.94581724 -1.811404e-02 -0.0436296127
#>  [60,]  3.246995e-01 -7.952843e-17  0.94581724  4.537571e-02 -0.0103706606
#>  [61,]  4.759474e-01  0.000000e+00  0.87947375 -3.147408e-02  0.0401667228
#>  [62,]  4.501593e-01  1.545399e-01  0.87947375 -1.064770e-03 -0.0478651644
#>  [63,]  3.755894e-01  2.923329e-01  0.87947375  6.351092e-03  0.0214335615
#>  [64,]  2.603185e-01  3.984472e-01  0.87947375  4.566801e-02  0.0384782605
#>  [65,]  1.168382e-01  4.613835e-01  0.87947375  1.642138e-02  0.0289755498
#>  [66,] -3.930342e-02  4.743218e-01  0.87947375 -2.285188e-02  0.0092323404
#>  [67,] -1.911859e-01  4.358599e-01  0.87947375 -2.293776e-02  0.0184466463
#>  [68,] -3.223504e-01  3.501659e-01  0.87947375  5.704539e-03 -0.0080032000
#>  [69,] -4.185832e-01  2.265259e-01  0.87947375 -3.116350e-02 -0.0027796258
#>  [70,] -4.694561e-01  7.833837e-02  0.87947375  1.080589e-02 -0.0486443652
#>  [71,] -4.694561e-01 -7.833837e-02  0.87947375 -2.030990e-02  0.0348341506
#>  [72,] -4.185832e-01 -2.265259e-01  0.87947375  2.638654e-02  0.0106109110
#>  [73,] -3.223504e-01 -3.501659e-01  0.87947375 -2.207317e-02  0.0387561368
#>  [74,] -1.911859e-01 -4.358599e-01  0.87947375 -3.189402e-02 -0.0395617449
#>  [75,] -3.930342e-02 -4.743218e-01  0.87947375 -2.539848e-03  0.0325523722
#>  [76,]  1.168382e-01 -4.613835e-01  0.87947375  6.054655e-03  0.0065564088
#>  [77,]  2.603185e-01 -3.984472e-01  0.87947375  7.853773e-03  0.0460813394
#>  [78,]  3.755894e-01 -2.923329e-01  0.87947375  1.539729e-02  0.0468133635
#>  [79,]  4.501593e-01 -1.545399e-01  0.87947375  3.038408e-02  0.0432129578
#>  [80,]  4.759474e-01 -1.165735e-16  0.87947375  1.545190e-02 -0.0257025630
#>  [81,]  6.142127e-01  0.000000e+00  0.78914051  4.727311e-02  0.0373341333
#>  [82,]  5.809330e-01  1.994345e-01  0.78914051 -2.729729e-02  0.0373845061
#>  [83,]  4.847001e-01  3.772573e-01  0.78914051  2.225150e-02 -0.0456392032
#>  [84,]  3.359425e-01  5.141983e-01  0.78914051 -2.782474e-02  0.0354060582
#>  [85,]  1.507803e-01  5.954180e-01  0.78914051 -9.544710e-04 -0.0370039008
#>  [86,] -5.072128e-02  6.121149e-01  0.78914051 -3.715066e-02 -0.0037995307
#>  [87,] -2.467264e-01  5.624796e-01  0.78914051  3.340833e-02 -0.0295620600
#>  [88,] -4.159950e-01  4.518910e-01  0.78914051 -3.819117e-02  0.0073244504
#>  [89,] -5.401840e-01  2.923329e-01  0.78914051 -1.258900e-02  0.0083263904
#>  [90,] -6.058357e-01  1.010961e-01  0.78914051  2.994507e-02  0.0466277180
#>  [91,] -6.058357e-01 -1.010961e-01  0.78914051 -5.099938e-03 -0.0109634040
#>  [92,] -5.401840e-01 -2.923329e-01  0.78914051 -3.882695e-02  0.0100374365
#>  [93,] -4.159950e-01 -4.518910e-01  0.78914051  4.227793e-02 -0.0318201166
#>  [94,] -2.467264e-01 -5.624796e-01  0.78914051 -1.627820e-02 -0.0301418320
#>  [95,] -5.072128e-02 -6.121149e-01  0.78914051  3.315091e-02  0.0466431932
#>  [96,]  1.507803e-01 -5.954180e-01  0.78914051  3.135840e-02  0.0306173817
#>  [97,]  3.359425e-01 -5.141983e-01  0.78914051 -3.449499e-02 -0.0261853690
#>  [98,]  4.847001e-01 -3.772573e-01  0.78914051  4.924605e-03  0.0220818508
#>  [99,]  5.809330e-01 -1.994345e-01  0.78914051 -1.913869e-02 -0.0417062872
#> [100,]  6.142127e-01 -1.504387e-16  0.78914051  3.439463e-02  0.0112439436
#> [101,]  7.357239e-01  0.000000e+00  0.67728157  4.580957e-02 -0.0175971042
#> [102,]  6.958604e-01  2.388892e-01  0.67728157  3.605648e-02 -0.0110068649
#> [103,]  5.805895e-01  4.518910e-01  0.67728157 -3.767396e-02  0.0062516510
#> [104,]  4.024028e-01  6.159234e-01  0.67728157  1.726209e-02 -0.0297679845
#> [105,]  1.806095e-01  7.132110e-01  0.67728157 -1.469516e-03  0.0183193847
#> [106,] -6.075560e-02  7.332110e-01  0.67728157 -4.162037e-02  0.0191119038
#> [107,] -2.955369e-01  6.737563e-01  0.67728157  3.341443e-02 -0.0382497837
#> [108,] -4.982922e-01  5.412897e-01  0.67728157 -3.280134e-02  0.0238259561
#> [109,] -6.470499e-01  3.501659e-01  0.67728157  4.754606e-02  0.0102588527
#> [110,] -7.256896e-01  1.210962e-01  0.67728157 -4.634827e-02  0.0456703967
#> [111,] -7.256896e-01 -1.210962e-01  0.67728157 -2.702787e-02  0.0382774559
#> [112,] -6.470499e-01 -3.501659e-01  0.67728157 -2.148088e-02 -0.0340229558
#> [113,] -4.982922e-01 -5.412897e-01  0.67728157  1.149406e-02  0.0344506862
#> [114,] -2.955369e-01 -6.737563e-01  0.67728157 -2.104307e-02  0.0261127813
#> [115,] -6.075560e-02 -7.332110e-01  0.67728157 -1.734355e-02  0.0362474316
#> [116,]  1.806095e-01 -7.132110e-01  0.67728157  8.395652e-03  0.0050193269
#> [117,]  4.024028e-01 -6.159234e-01  0.67728157  1.631517e-02 -0.0419351479
#> [118,]  5.805895e-01 -4.518910e-01  0.67728157 -4.902594e-02 -0.0440424554
#> [119,]  6.958604e-01 -2.388892e-01  0.67728157  1.553425e-02 -0.0093545675
#> [120,]  7.357239e-01 -1.802004e-16  0.67728157  2.061496e-02  0.0496871467
#> [121,]  8.371665e-01  0.000000e+00  0.54694816 -1.175866e-02  0.0375300080
#> [122,]  7.918065e-01  2.718275e-01  0.54694816  4.089815e-02 -0.0242959486
#> [123,]  6.606420e-01  5.141983e-01  0.54694816 -3.342855e-04  0.0387470879
#> [124,]  4.578867e-01  7.008477e-01  0.54694816  3.002607e-02 -0.0364508789
#> [125,]  2.055122e-01  8.115494e-01  0.54694816  4.322579e-02  0.0481919708
#> [126,] -6.913266e-02  8.343071e-01  0.54694816 -3.640539e-02 -0.0236516769
#> [127,] -3.362859e-01  7.666547e-01  0.54694816 -3.468724e-02 -0.0180516620
#> [128,] -5.669974e-01  6.159234e-01  0.54694816  3.009169e-02 -0.0480905774
#> [129,] -7.362659e-01  3.984472e-01  0.54694816  4.997195e-02  0.0272297627
#> [130,] -8.257486e-01  1.377931e-01  0.54694816  1.696164e-02 -0.0239540783
#> [131,] -8.257486e-01 -1.377931e-01  0.54694816 -2.506859e-04 -0.0092006936
#> [132,] -7.362659e-01 -3.984472e-01  0.54694816  2.830742e-02  0.0168698957
#> [133,] -5.669974e-01 -6.159234e-01  0.54694816  2.489782e-02 -0.0024392179
#> [134,] -3.362859e-01 -7.666547e-01  0.54694816  3.334827e-02 -0.0399071505
#> [135,] -6.913266e-02 -8.343071e-01  0.54694816  2.358594e-02 -0.0153525516
#> [136,]  2.055122e-01 -8.115494e-01  0.54694816 -2.905393e-02  0.0218136192
#> [137,]  4.578867e-01 -7.008477e-01  0.54694816 -1.973666e-02 -0.0107856375
#> [138,]  6.606420e-01 -5.141983e-01  0.54694816 -3.136568e-02  0.0447329359
#> [139,]  7.918065e-01 -2.718275e-01  0.54694816  1.777513e-02  0.0321098123
#> [140,]  8.371665e-01 -2.050466e-16  0.54694816 -3.344153e-02  0.0261745216
#> [141,]  9.157733e-01  0.000000e+00  0.40169542 -5.866051e-03  0.0156568203
#> [142,]  8.661542e-01  2.973511e-01  0.40169542  1.158090e-02  0.0346658462
#> [143,]  7.226738e-01  5.624796e-01  0.40169542 -3.328644e-02 -0.0447070410
#> [144,]  5.008805e-01  7.666547e-01  0.40169542 -1.494194e-02  0.0019044608
#> [145,]  2.248091e-01  8.877509e-01  0.40169542  2.708762e-02 -0.0165922410
#> [146,] -7.562396e-02  9.126455e-01  0.40169542  1.393676e-02  0.0392769223
#> [147,] -3.678620e-01  8.386408e-01  0.40169542 -4.287820e-02 -0.0438142309
#> [148,] -6.202364e-01  6.737563e-01  0.40169542 -4.788454e-02  0.0111335540
#> [149,] -8.053986e-01  4.358599e-01  0.40169542 -1.391358e-02 -0.0387945331
#> [150,] -9.032834e-01  1.507313e-01  0.40169542  1.662831e-02 -0.0349643391
#> [151,] -9.032834e-01 -1.507313e-01  0.40169542  2.998335e-02 -0.0017789456
#> [152,] -8.053986e-01 -4.358599e-01  0.40169542  1.648803e-02  0.0014424122
#> [153,] -6.202364e-01 -6.737563e-01  0.40169542 -2.474672e-03 -0.0444907812
#> [154,] -3.678620e-01 -8.386408e-01  0.40169542  3.568368e-03  0.0147553958
#> [155,] -7.562396e-02 -9.126455e-01  0.40169542 -5.090316e-04 -0.0205053264
#> [156,]  2.248091e-01 -8.877509e-01  0.40169542 -7.511063e-04 -0.0236016335
#> [157,]  5.008805e-01 -7.666547e-01  0.40169542 -3.880757e-02 -0.0045184170
#> [158,]  7.226738e-01 -5.624796e-01  0.40169542  1.369997e-02 -0.0339612277
#> [159,]  8.661542e-01 -2.973511e-01  0.40169542 -3.619602e-02 -0.0311791904
#> [160,]  9.157733e-01 -2.242998e-16  0.40169542  2.462692e-02  0.0335771727
#> [161,]  9.694003e-01  0.000000e+00  0.24548549  3.592238e-02 -0.0011223412
#> [162,]  9.168755e-01  3.147638e-01  0.24548549 -1.604586e-02  0.0467129162
#> [163,]  7.649930e-01  5.954180e-01  0.24548549 -8.215983e-03 -0.0294062203
#> [164,]  5.302117e-01  8.115494e-01  0.24548549  2.754420e-02 -0.0284008076
#> [165,]  2.379737e-01  9.397369e-01  0.24548549 -1.569917e-02 -0.0057685219
#> [166,] -8.005244e-02  9.660893e-01  0.24548549 -2.627042e-02  0.0046525453
#> [167,] -3.894037e-01  8.877509e-01  0.24548549  3.815845e-02 -0.0064275829
#> [168,] -6.565569e-01  7.132110e-01  0.24548549 -1.593673e-02  0.0375833918
#> [169,] -8.525621e-01  4.613835e-01  0.24548549 -6.824344e-03  0.0021057146
#> [170,] -9.561789e-01  1.595580e-01  0.24548549 -6.530918e-03  0.0270223118
#> [171,] -9.561789e-01 -1.595580e-01  0.24548549  2.173514e-02  0.0424054142
#> [172,] -8.525621e-01 -4.613835e-01  0.24548549 -2.653340e-03  0.0433021687
#> [173,] -6.565569e-01 -7.132110e-01  0.24548549  3.487736e-02  0.0373980915
#> [174,] -3.894037e-01 -8.877509e-01  0.24548549 -1.347819e-02  0.0260932069
#> [175,] -8.005244e-02 -9.660893e-01  0.24548549  2.473084e-03  0.0130606692
#> [176,]  2.379737e-01 -9.397369e-01  0.24548549  9.591275e-03  0.0008777126
#> [177,]  5.302117e-01 -8.115494e-01  0.24548549  2.972725e-02  0.0181250342
#> [178,]  7.649930e-01 -5.954180e-01  0.24548549  1.980184e-02 -0.0095539575
#> [179,]  9.168755e-01 -3.147638e-01  0.24548549  2.362762e-02 -0.0148243677
#> [180,]  9.694003e-01 -2.374346e-16  0.24548549 -1.599358e-02 -0.0025192920
#> [181,]  9.965845e-01  0.000000e+00  0.08257935  4.891926e-02 -0.0086863666
#> [182,]  9.425868e-01  3.235905e-01  0.08257935  2.766279e-02  0.0471169198
#> [183,]  7.864452e-01  6.121149e-01  0.08257935  5.090469e-03  0.0078140313
#> [184,]  5.450801e-01  8.343071e-01  0.08257935  4.161484e-02  0.0408082086
#> [185,]  2.446470e-01  9.660893e-01  0.08257935  4.979444e-03 -0.0410311182
#> [186,] -8.229730e-02  9.931807e-01  0.08257935 -2.626307e-02 -0.0110679008
#> [187,] -4.003234e-01  9.126455e-01  0.08257935 -4.404330e-02 -0.0495208575
#> [188,] -6.749683e-01  7.332110e-01  0.08257935 -1.739481e-02 -0.0132237768
#> [189,] -8.764699e-01  4.743218e-01  0.08257935  1.545180e-02 -0.0478583822
#> [190,] -9.829924e-01  1.640324e-01  0.08257935  2.621291e-02  0.0412595252
#> [191,] -9.829924e-01 -1.640324e-01  0.08257935 -1.620092e-02  0.0300587652
#> [192,] -8.764699e-01 -4.743218e-01  0.08257935  8.400344e-03 -0.0145161603
#> [193,] -6.749683e-01 -7.332110e-01  0.08257935 -1.259315e-02 -0.0209458076
#> [194,] -4.003234e-01 -9.126455e-01  0.08257935 -1.540671e-02 -0.0062117375
#> [195,] -8.229730e-02 -9.931807e-01  0.08257935  2.656813e-02  0.0121138282
#> [196,]  2.446470e-01 -9.660893e-01  0.08257935  3.028532e-02 -0.0496420466
#> [197,]  5.450801e-01 -8.343071e-01  0.08257935  2.122273e-02 -0.0430230580
#> [198,]  7.864452e-01 -6.121149e-01  0.08257935  4.613793e-02  0.0177699051
#> [199,]  9.425868e-01 -3.235905e-01  0.08257935 -3.906601e-04  0.0074530185
#> [200,]  9.965845e-01 -2.440928e-16  0.08257935  4.656304e-02 -0.0081905388
#> [201,]  9.965845e-01  0.000000e+00 -0.08257935  1.054974e-02  0.0294309496
#> [202,]  9.425868e-01  3.235905e-01 -0.08257935  4.972672e-02 -0.0461243640
#> [203,]  7.864452e-01  6.121149e-01 -0.08257935 -2.664957e-02  0.0349964348
#> [204,]  5.450801e-01  8.343071e-01 -0.08257935 -1.026829e-02 -0.0281297139
#> [205,]  2.446470e-01  9.660893e-01 -0.08257935  3.542493e-02  0.0018720842
#> [206,] -8.229730e-02  9.931807e-01 -0.08257935 -3.153367e-02 -0.0057906110
#> [207,] -4.003234e-01  9.126455e-01 -0.08257935  4.449511e-02  0.0494631143
#> [208,] -6.749683e-01  7.332110e-01 -0.08257935 -6.851563e-05  0.0127077777
#> [209,] -8.764699e-01  4.743218e-01 -0.08257935  2.390672e-02  0.0148093513
#> [210,] -9.829924e-01  1.640324e-01 -0.08257935 -4.668089e-02 -0.0261160227
#> [211,] -9.829924e-01 -1.640324e-01 -0.08257935  1.496274e-02 -0.0307416708
#> [212,] -8.764699e-01 -4.743218e-01 -0.08257935 -6.166075e-03 -0.0052810666
#> [213,] -6.749683e-01 -7.332110e-01 -0.08257935 -2.089087e-02  0.0397020969
#> [214,] -4.003234e-01 -9.126455e-01 -0.08257935 -4.223115e-02 -0.0322848648
#> [215,] -8.229730e-02 -9.931807e-01 -0.08257935  1.690153e-02 -0.0316314426
#> [216,]  2.446470e-01 -9.660893e-01 -0.08257935  2.671759e-03 -0.0169825869
#> [217,]  5.450801e-01 -8.343071e-01 -0.08257935 -2.569338e-02 -0.0356294045
#> [218,]  7.864452e-01 -6.121149e-01 -0.08257935  2.756341e-02 -0.0168258445
#> [219,]  9.425868e-01 -3.235905e-01 -0.08257935  2.975427e-02  0.0083503170
#> [220,]  9.965845e-01 -2.440928e-16 -0.08257935 -3.278267e-02 -0.0319002165
#> [221,]  9.694003e-01  0.000000e+00 -0.24548549 -3.745897e-02 -0.0299223191
#> [222,]  9.168755e-01  3.147638e-01 -0.24548549  2.496602e-02 -0.0034050244
#> [223,]  7.649930e-01  5.954180e-01 -0.24548549  1.342573e-02 -0.0486215314
#> [224,]  5.302117e-01  8.115494e-01 -0.24548549  1.044195e-02  0.0318824032
#> [225,]  2.379737e-01  9.397369e-01 -0.24548549  2.915857e-02  0.0448122117
#> [226,] -8.005244e-02  9.660893e-01 -0.24548549 -4.143900e-02  0.0308863278
#> [227,] -3.894037e-01  8.877509e-01 -0.24548549  3.804411e-02 -0.0006181329
#> [228,] -6.565569e-01  7.132110e-01 -0.24548549  4.396678e-02 -0.0038565366
#> [229,] -8.525621e-01  4.613835e-01 -0.24548549 -3.654322e-02 -0.0380501115
#> [230,] -9.561789e-01  1.595580e-01 -0.24548549 -4.773197e-02 -0.0485416156
#> [231,] -9.561789e-01 -1.595580e-01 -0.24548549 -2.834478e-02  0.0256864625
#> [232,] -8.525621e-01 -4.613835e-01 -0.24548549  9.218093e-03 -0.0369596384
#> [233,] -6.565569e-01 -7.132110e-01 -0.24548549 -2.130571e-02 -0.0364870833
#> [234,] -3.894037e-01 -8.877509e-01 -0.24548549 -4.745588e-02 -0.0113334192
#> [235,] -8.005244e-02 -9.660893e-01 -0.24548549 -1.942000e-02 -0.0318635189
#> [236,]  2.379737e-01 -9.397369e-01 -0.24548549  4.339826e-02  0.0222340727
#> [237,]  5.302117e-01 -8.115494e-01 -0.24548549  3.079799e-02 -0.0211734026
#> [238,]  7.649930e-01 -5.954180e-01 -0.24548549 -3.659461e-02  0.0259589361
#> [239,]  9.168755e-01 -3.147638e-01 -0.24548549 -1.845321e-02 -0.0338443739
#> [240,]  9.694003e-01 -2.374346e-16 -0.24548549 -3.736770e-02  0.0331535453
#> [241,]  9.157733e-01  0.000000e+00 -0.40169542 -1.757218e-02  0.0383407873
#> [242,]  8.661542e-01  2.973511e-01 -0.40169542  2.381306e-02 -0.0306462098
#> [243,]  7.226738e-01  5.624796e-01 -0.40169542  3.585541e-02 -0.0221329492
#> [244,]  5.008805e-01  7.666547e-01 -0.40169542 -2.276861e-02  0.0089671019
#> [245,]  2.248091e-01  8.877509e-01 -0.40169542 -2.092906e-02  0.0317656433
#> [246,] -7.562396e-02  9.126455e-01 -0.40169542  4.405842e-02  0.0159642034
#> [247,] -3.678620e-01  8.386408e-01 -0.40169542  4.785016e-03 -0.0441006338
#> [248,] -6.202364e-01  6.737563e-01 -0.40169542  6.018797e-03  0.0153438816
#> [249,] -8.053986e-01  4.358599e-01 -0.40169542 -2.188560e-02  0.0319480885
#> [250,] -9.032834e-01  1.507313e-01 -0.40169542  1.001763e-02  0.0363429497
#> [251,] -9.032834e-01 -1.507313e-01 -0.40169542 -3.583888e-02 -0.0067770916
#> [252,] -8.053986e-01 -4.358599e-01 -0.40169542  2.352574e-03  0.0048531985
#> [253,] -6.202364e-01 -6.737563e-01 -0.40169542 -8.313727e-03  0.0499829204
#> [254,] -3.678620e-01 -8.386408e-01 -0.40169542 -3.338476e-02 -0.0294000644
#> [255,] -7.562396e-02 -9.126455e-01 -0.40169542 -5.500066e-03 -0.0145081357
#> [256,]  2.248091e-01 -8.877509e-01 -0.40169542 -3.441749e-02 -0.0484665609
#> [257,]  5.008805e-01 -7.666547e-01 -0.40169542  1.126366e-02  0.0152674341
#> [258,]  7.226738e-01 -5.624796e-01 -0.40169542 -4.020236e-02 -0.0081914325
#> [259,]  8.661542e-01 -2.973511e-01 -0.40169542 -1.728321e-02 -0.0478213581
#> [260,]  9.157733e-01 -2.242998e-16 -0.40169542 -4.400895e-02 -0.0171233777
#> [261,]  8.371665e-01  0.000000e+00 -0.54694816  1.867386e-02 -0.0224149135
#> [262,]  7.918065e-01  2.718275e-01 -0.54694816  3.628577e-02  0.0180084690
#> [263,]  6.606420e-01  5.141983e-01 -0.54694816  4.817311e-02 -0.0244204663
#> [264,]  4.578867e-01  7.008477e-01 -0.54694816  4.432656e-02 -0.0328797022
#> [265,]  2.055122e-01  8.115494e-01 -0.54694816  2.477080e-02  0.0062860237
#> [266,] -6.913266e-02  8.343071e-01 -0.54694816  3.478985e-02 -0.0281579070
#> [267,] -3.362859e-01  7.666547e-01 -0.54694816  3.304903e-02  0.0440832254
#> [268,] -5.669974e-01  6.159234e-01 -0.54694816 -1.232729e-02  0.0063462410
#> [269,] -7.362659e-01  3.984472e-01 -0.54694816 -4.115079e-03 -0.0055930366
#> [270,] -8.257486e-01  1.377931e-01 -0.54694816  6.191502e-04  0.0035797982
#> [271,] -8.257486e-01 -1.377931e-01 -0.54694816 -1.104245e-02 -0.0412336068
#> [272,] -7.362659e-01 -3.984472e-01 -0.54694816  2.025362e-02  0.0307503221
#> [273,] -5.669974e-01 -6.159234e-01 -0.54694816 -3.266529e-02  0.0143153314
#> [274,] -3.362859e-01 -7.666547e-01 -0.54694816  3.345635e-02 -0.0167631278
#> [275,] -6.913266e-02 -8.343071e-01 -0.54694816 -9.999954e-03 -0.0271173385
#> [276,]  2.055122e-01 -8.115494e-01 -0.54694816  1.175155e-02  0.0311200062
#> [277,]  4.578867e-01 -7.008477e-01 -0.54694816  8.293923e-03 -0.0028888069
#> [278,]  6.606420e-01 -5.141983e-01 -0.54694816 -2.586233e-02 -0.0420256688
#> [279,]  7.918065e-01 -2.718275e-01 -0.54694816  4.109525e-02 -0.0158099427
#> [280,]  8.371665e-01 -2.050466e-16 -0.54694816  4.561637e-03  0.0061113670
#> [281,]  7.357239e-01  0.000000e+00 -0.67728157  3.217740e-02 -0.0432061140
#> [282,]  6.958604e-01  2.388892e-01 -0.67728157  1.035375e-02  0.0208390571
#> [283,]  5.805895e-01  4.518910e-01 -0.67728157  2.804943e-05 -0.0007536862
#> [284,]  4.024028e-01  6.159234e-01 -0.67728157  2.084375e-02  0.0350761891
#> [285,]  1.806095e-01  7.132110e-01 -0.67728157  1.104190e-02 -0.0240887148
#> [286,] -6.075560e-02  7.332110e-01 -0.67728157  3.424133e-02  0.0265929442
#> [287,] -2.955369e-01  6.737563e-01 -0.67728157 -2.876319e-02  0.0363146341
#> [288,] -4.982922e-01  5.412897e-01 -0.67728157  3.954651e-02 -0.0391774012
#> [289,] -6.470499e-01  3.501659e-01 -0.67728157 -1.180963e-02 -0.0110796209
#> [290,] -7.256896e-01  1.210962e-01 -0.67728157  1.596681e-02 -0.0391312560
#> [291,] -7.256896e-01 -1.210962e-01 -0.67728157 -2.852978e-02  0.0247845218
#> [292,] -6.470499e-01 -3.501659e-01 -0.67728157 -2.221966e-02  0.0132450915
#> [293,] -4.982922e-01 -5.412897e-01 -0.67728157  4.757209e-02  0.0110336004
#> [294,] -2.955369e-01 -6.737563e-01 -0.67728157  3.828022e-02 -0.0338210255
#> [295,] -6.075560e-02 -7.332110e-01 -0.67728157  4.401323e-02 -0.0390122443
#> [296,]  1.806095e-01 -7.132110e-01 -0.67728157  4.172324e-02 -0.0377009066
#> [297,]  4.024028e-01 -6.159234e-01 -0.67728157 -1.725838e-03  0.0431310005
#> [298,]  5.805895e-01 -4.518910e-01 -0.67728157  8.183228e-03  0.0139698634
#> [299,]  6.958604e-01 -2.388892e-01 -0.67728157  4.648709e-02  0.0440206643
#> [300,]  7.357239e-01 -1.802004e-16 -0.67728157  2.053483e-02 -0.0167696636
#> [301,]  6.142127e-01  0.000000e+00 -0.78914051 -6.900971e-03  0.0199386400
#> [302,]  5.809330e-01  1.994345e-01 -0.78914051 -4.778647e-02 -0.0049957491
#> [303,]  4.847001e-01  3.772573e-01 -0.78914051  4.867959e-02 -0.0447288041
#> [304,]  3.359425e-01  5.141983e-01 -0.78914051  4.669521e-02 -0.0008609129
#> [305,]  1.507803e-01  5.954180e-01 -0.78914051  1.518210e-02  0.0479532950
#> [306,] -5.072128e-02  6.121149e-01 -0.78914051  8.454541e-03  0.0302088394
#> [307,] -2.467264e-01  5.624796e-01 -0.78914051 -2.166283e-02 -0.0459652385
#> [308,] -4.159950e-01  4.518910e-01 -0.78914051 -2.645764e-02 -0.0349139353
#> [309,] -5.401840e-01  2.923329e-01 -0.78914051 -9.912760e-04  0.0272638658
#> [310,] -6.058357e-01  1.010961e-01 -0.78914051  3.708082e-02 -0.0305758360
#> [311,] -6.058357e-01 -1.010961e-01 -0.78914051 -2.996518e-03  0.0267885437
#> [312,] -5.401840e-01 -2.923329e-01 -0.78914051  1.397936e-02  0.0346473492
#> [313,] -4.159950e-01 -4.518910e-01 -0.78914051  1.128797e-02  0.0466495983
#> [314,] -2.467264e-01 -5.624796e-01 -0.78914051 -1.179246e-02 -0.0288113103
#> [315,] -5.072128e-02 -6.121149e-01 -0.78914051  1.467378e-02 -0.0341454009
#> [316,]  1.507803e-01 -5.954180e-01 -0.78914051  2.462333e-02  0.0021935805
#> [317,]  3.359425e-01 -5.141983e-01 -0.78914051 -2.707538e-02  0.0177649498
#> [318,]  4.847001e-01 -3.772573e-01 -0.78914051  1.011142e-02  0.0462015493
#> [319,]  5.809330e-01 -1.994345e-01 -0.78914051  3.360671e-02  0.0217324858
#> [320,]  6.142127e-01 -1.504387e-16 -0.78914051 -2.986767e-02  0.0213668778
#> [321,]  4.759474e-01  0.000000e+00 -0.87947375 -1.254553e-02 -0.0032292174
#> [322,]  4.501593e-01  1.545399e-01 -0.87947375  1.029722e-02 -0.0002915311
#> [323,]  3.755894e-01  2.923329e-01 -0.87947375  1.745579e-02 -0.0151492735
#> [324,]  2.603185e-01  3.984472e-01 -0.87947375 -2.045029e-02  0.0375246133
#> [325,]  1.168382e-01  4.613835e-01 -0.87947375 -2.870391e-02 -0.0078970414
#> [326,] -3.930342e-02  4.743218e-01 -0.87947375  6.322420e-03  0.0195966444
#> [327,] -1.911859e-01  4.358599e-01 -0.87947375 -2.638304e-02  0.0489806940
#> [328,] -3.223504e-01  3.501659e-01 -0.87947375 -3.345348e-02 -0.0337444095
#> [329,] -4.185832e-01  2.265259e-01 -0.87947375  3.892730e-02  0.0393842889
#> [330,] -4.694561e-01  7.833837e-02 -0.87947375 -4.231115e-02 -0.0103240116
#> [331,] -4.694561e-01 -7.833837e-02 -0.87947375 -2.888918e-02 -0.0238302188
#> [332,] -4.185832e-01 -2.265259e-01 -0.87947375 -1.858484e-03  0.0078905101
#> [333,] -3.223504e-01 -3.501659e-01 -0.87947375  1.240992e-02  0.0226593223
#> [334,] -1.911859e-01 -4.358599e-01 -0.87947375  4.275351e-02 -0.0317766289
#> [335,] -3.930342e-02 -4.743218e-01 -0.87947375  4.045569e-02  0.0100143955
#> [336,]  1.168382e-01 -4.613835e-01 -0.87947375  1.259700e-02 -0.0262488032
#> [337,]  2.603185e-01 -3.984472e-01 -0.87947375 -1.599677e-02  0.0415151648
#> [338,]  3.755894e-01 -2.923329e-01 -0.87947375  4.923139e-03 -0.0179416317
#> [339,]  4.501593e-01 -1.545399e-01 -0.87947375  3.546447e-02 -0.0440736821
#> [340,]  4.759474e-01 -1.165735e-16 -0.87947375 -7.066186e-03  0.0094478199
#> [341,]  3.246995e-01  0.000000e+00 -0.94581724  4.731613e-02 -0.0303168807
#> [342,]  3.071064e-01  1.054297e-01 -0.94581724  2.142990e-02 -0.0492198647
#> [343,]  2.562335e-01  1.994345e-01 -0.94581724 -4.683163e-02 -0.0047453457
#> [344,]  1.775938e-01  2.718275e-01 -0.94581724  3.910496e-02  0.0390314454
#> [345,]  7.970901e-02  3.147638e-01 -0.94581724 -2.389168e-02 -0.0174445834
#> [346,] -2.681347e-02  3.235905e-01 -0.94581724 -1.293732e-02 -0.0283883862
#> [347,] -1.304303e-01  2.973511e-01 -0.94581724 -1.490081e-02  0.0084610309
#> [348,] -2.199130e-01  2.388892e-01 -0.94581724 -3.993390e-02 -0.0262848433
#> [349,] -2.855647e-01  1.545399e-01 -0.94581724  1.904374e-02  0.0398271208
#> [350,] -3.202710e-01  5.344378e-02 -0.94581724  1.942438e-02 -0.0470006421
#> [351,] -3.202710e-01 -5.344378e-02 -0.94581724  2.324793e-02  0.0027409989
#> [352,] -2.855647e-01 -1.545399e-01 -0.94581724  4.681659e-02  0.0009818338
#> [353,] -2.199130e-01 -2.388892e-01 -0.94581724 -7.698080e-03 -0.0143159085
#> [354,] -1.304303e-01 -2.973511e-01 -0.94581724 -2.847562e-02  0.0288522211
#> [355,] -2.681347e-02 -3.235905e-01 -0.94581724  4.764334e-02  0.0411806517
#> [356,]  7.970901e-02 -3.147638e-01 -0.94581724  3.659892e-03 -0.0331097727
#> [357,]  1.775938e-01 -2.718275e-01 -0.94581724  2.486526e-02 -0.0034899638
#> [358,]  2.562335e-01 -1.994345e-01 -0.94581724 -4.436575e-02  0.0370544043
#> [359,]  3.071064e-01 -1.054297e-01 -0.94581724  2.125700e-02 -0.0292880963
#> [360,]  3.246995e-01 -7.952843e-17 -0.94581724  2.901051e-02 -0.0290518005
#> [361,]  1.645946e-01  0.000000e+00 -0.98636130  3.876741e-02 -0.0173076425
#> [362,]  1.556764e-01  5.344378e-02 -0.98636130 -2.309413e-02 -0.0081820370
#> [363,]  1.298883e-01  1.010961e-01 -0.98636130  6.145974e-03  0.0022542704
#> [364,]  9.002471e-02  1.377931e-01 -0.98636130  8.619033e-03  0.0008428099
#> [365,]  4.040558e-02  1.595580e-01 -0.98636130  1.425498e-02  0.0061960225
#> [366,] -1.359211e-02  1.640324e-01 -0.98636130  2.099092e-03 -0.0425813718
#> [367,] -6.611689e-02  1.507313e-01 -0.98636130 -1.380505e-02  0.0188001706
#> [368,] -1.114769e-01  1.210962e-01 -0.98636130  2.822685e-03  0.0420870892
#> [369,] -1.447566e-01  7.833837e-02 -0.98636130  1.313254e-02 -0.0136882657
#> [370,] -1.623497e-01  2.709138e-02 -0.98636130 -2.781010e-02  0.0165171196
#> [371,] -1.623497e-01 -2.709138e-02 -0.98636130  4.885434e-02  0.0136907826
#> [372,] -1.447566e-01 -7.833837e-02 -0.98636130 -3.401122e-02 -0.0088853210
#> [373,] -1.114769e-01 -1.210962e-01 -0.98636130  3.116246e-02 -0.0393947933
#> [374,] -6.611689e-02 -1.507313e-01 -0.98636130  2.795724e-02  0.0123166470
#> [375,] -1.359211e-02 -1.640324e-01 -0.98636130 -6.305437e-03  0.0192548504
#> [376,]  4.040558e-02 -1.595580e-01 -0.98636130 -3.743024e-02  0.0472971624
#> [377,]  9.002471e-02 -1.377931e-01 -0.98636130 -3.410341e-02  0.0275381119
#> [378,]  1.298883e-01 -1.010961e-01 -0.98636130 -3.526346e-02  0.0014641122
#> [379,]  1.556764e-01 -5.344378e-02 -0.98636130 -4.808007e-02 -0.0440712923
#> [380,]  1.645946e-01 -4.031405e-17 -0.98636130  3.032708e-03 -0.0431698710
#> [381,]  1.224647e-16  0.000000e+00 -1.00000000  2.728857e-02  0.0401073257
#> [382,]  1.158292e-16  3.976422e-17 -1.00000000  2.019192e-03 -0.0295935417
#> [383,]  9.664184e-17  7.521936e-17 -1.00000000  4.246936e-02  0.0035549731
#> [384,]  6.698183e-17  1.025233e-16 -1.00000000  1.880247e-02 -0.0470787340
#> [385,]  3.006330e-17  1.187173e-16 -1.00000000 -1.577043e-02  0.0039680616
#> [386,] -1.011305e-17  1.220464e-16 -1.00000000  4.068286e-03  0.0117731159
#> [387,] -4.919350e-17  1.121499e-16 -1.00000000  3.531341e-02 -0.0396301861
#> [388,] -8.294307e-17  9.010019e-17 -1.00000000  4.121293e-02  0.0148805361
#> [389,] -1.077045e-16  5.828675e-17 -1.00000000  1.444615e-03  0.0033798450
#> [390,] -1.207944e-16  2.015702e-17 -1.00000000  4.472277e-02 -0.0499992079
#> [391,] -1.207944e-16 -2.015702e-17 -1.00000000 -2.223595e-02  0.0432841314
#> [392,] -1.077045e-16 -5.828675e-17 -1.00000000  3.461254e-02  0.0166666994
#> [393,] -8.294307e-17 -9.010019e-17 -1.00000000 -4.416205e-02  0.0002536011
#> [394,] -4.919350e-17 -1.121499e-16 -1.00000000 -3.065043e-02  0.0034222506
#> [395,] -1.011305e-17 -1.220464e-16 -1.00000000 -4.787478e-02  0.0220524735
#> [396,]  3.006330e-17 -1.187173e-16 -1.00000000  3.136327e-02 -0.0193484914
#> [397,]  6.698183e-17 -1.025233e-16 -1.00000000 -1.714964e-02 -0.0249574766
#> [398,]  9.664184e-17 -7.521936e-17 -1.00000000 -2.061045e-02 -0.0233004792
#> [399,]  1.158292e-16 -3.976422e-17 -1.00000000 -3.069324e-02  0.0492604014
#> [400,]  1.224647e-16 -2.999520e-32 -1.00000000 -1.261886e-02 -0.0287831886
#>                 [,6]
#>   [1,]  0.0351324049
#>   [2,] -0.0323690788
#>   [3,]  0.0072392003
#>   [4,]  0.0017862579
#>   [5,] -0.0118683703
#>   [6,]  0.0399624557
#>   [7,] -0.0374529202
#>   [8,]  0.0172401139
#>   [9,] -0.0067165892
#>  [10,]  0.0051688562
#>  [11,]  0.0461753666
#>  [12,]  0.0379448387
#>  [13,] -0.0114364368
#>  [14,]  0.0355006760
#>  [15,]  0.0267504568
#>  [16,] -0.0080116730
#>  [17,] -0.0363195139
#>  [18,]  0.0466556178
#>  [19,]  0.0345957809
#>  [20,]  0.0092369739
#>  [21,] -0.0356611551
#>  [22,] -0.0381525940
#>  [23,]  0.0055163404
#>  [24,] -0.0406530660
#>  [25,] -0.0498011990
#>  [26,]  0.0058121504
#>  [27,] -0.0477030213
#>  [28,] -0.0489644559
#>  [29,]  0.0477977195
#>  [30,] -0.0341426316
#>  [31,] -0.0230897730
#>  [32,]  0.0435775504
#>  [33,]  0.0333234228
#>  [34,] -0.0286111622
#>  [35,]  0.0076730531
#>  [36,] -0.0306368138
#>  [37,] -0.0304929947
#>  [38,] -0.0334831300
#>  [39,] -0.0436278134
#>  [40,]  0.0127554375
#>  [41,]  0.0438709634
#>  [42,] -0.0190674186
#>  [43,]  0.0259475975
#>  [44,] -0.0127910370
#>  [45,]  0.0420616323
#>  [46,] -0.0290283532
#>  [47,]  0.0381300780
#>  [48,]  0.0473881890
#>  [49,] -0.0223718336
#>  [50,]  0.0496897068
#>  [51,]  0.0418166999
#>  [52,]  0.0025830809
#>  [53,] -0.0073573437
#>  [54,]  0.0445457743
#>  [55,]  0.0089797237
#>  [56,] -0.0465552265
#>  [57,]  0.0082594251
#>  [58,] -0.0114868055
#>  [59,]  0.0039424574
#>  [60,]  0.0204892196
#>  [61,] -0.0295103819
#>  [62,] -0.0299411011
#>  [63,] -0.0271450755
#>  [64,]  0.0345005997
#>  [65,] -0.0169374186
#>  [66,] -0.0150691135
#>  [67,]  0.0348874428
#>  [68,]  0.0225739281
#>  [69,] -0.0121233527
#>  [70,] -0.0497840195
#>  [71,]  0.0017397061
#>  [72,]  0.0349155360
#>  [73,] -0.0079226844
#>  [74,] -0.0216042877
#>  [75,]  0.0493799857
#>  [76,] -0.0203905018
#>  [77,]  0.0381571715
#>  [78,]  0.0183561302
#>  [79,]  0.0128347773
#>  [80,] -0.0417610256
#>  [81,]  0.0277707363
#>  [82,]  0.0039500390
#>  [83,] -0.0488838205
#>  [84,]  0.0177850027
#>  [85,]  0.0111809357
#>  [86,]  0.0384782864
#>  [87,]  0.0073563515
#>  [88,]  0.0128151776
#>  [89,] -0.0066297252
#>  [90,] -0.0435744314
#>  [91,] -0.0223847083
#>  [92,] -0.0415841853
#>  [93,]  0.0393458884
#>  [94,]  0.0101729277
#>  [95,]  0.0351967489
#>  [96,] -0.0216529948
#>  [97,] -0.0051659320
#>  [98,]  0.0093801327
#>  [99,]  0.0210738510
#> [100,]  0.0374932231
#> [101,]  0.0032604345
#> [102,]  0.0180750810
#> [103,]  0.0341655283
#> [104,] -0.0262935215
#> [105,] -0.0384207920
#> [106,] -0.0495829426
#> [107,]  0.0261582428
#> [108,]  0.0480949732
#> [109,] -0.0190551373
#> [110,] -0.0452986259
#> [111,]  0.0280814481
#> [112,]  0.0495720570
#> [113,] -0.0346997277
#> [114,]  0.0475954048
#> [115,] -0.0379341354
#> [116,] -0.0417005731
#> [117,] -0.0178785620
#> [118,]  0.0265422601
#> [119,] -0.0330254275
#> [120,]  0.0423137183
#> [121,] -0.0222706668
#> [122,]  0.0047309237
#> [123,] -0.0348084323
#> [124,] -0.0243717068
#> [125,] -0.0146831196
#> [126,] -0.0139821905
#> [127,]  0.0148778549
#> [128,]  0.0067748491
#> [129,]  0.0363769507
#> [130,] -0.0069149152
#> [131,] -0.0005955629
#> [132,]  0.0001354119
#> [133,] -0.0058698016
#> [134,]  0.0351338839
#> [135,] -0.0441461657
#> [136,]  0.0023779168
#> [137,]  0.0288384800
#> [138,]  0.0162981322
#> [139,] -0.0324985641
#> [140,]  0.0221179466
#> [141,]  0.0266337472
#> [142,] -0.0175042943
#> [143,] -0.0058369315
#> [144,]  0.0305234385
#> [145,]  0.0117787368
#> [146,]  0.0340112844
#> [147,] -0.0495549047
#> [148,] -0.0440663379
#> [149,] -0.0366588131
#> [150,] -0.0423519193
#> [151,]  0.0179676127
#> [152,] -0.0346720112
#> [153,] -0.0105358604
#> [154,] -0.0388897476
#> [155,]  0.0206338906
#> [156,] -0.0286781157
#> [157,] -0.0029163714
#> [158,] -0.0339468971
#> [159,]  0.0146281933
#> [160,] -0.0006662173
#> [161,]  0.0340562089
#> [162,]  0.0423673738
#> [163,] -0.0377279501
#> [164,]  0.0164545753
#> [165,] -0.0312253740
#> [166,] -0.0465212737
#> [167,]  0.0079668593
#> [168,] -0.0172529313
#> [169,]  0.0377933598
#> [170,]  0.0288388964
#> [171,]  0.0317813022
#> [172,]  0.0252395107
#> [173,]  0.0289578733
#> [174,]  0.0246742840
#> [175,]  0.0498130434
#> [176,]  0.0209366097
#> [177,] -0.0439008097
#> [178,]  0.0150919481
#> [179,]  0.0073373328
#> [180,]  0.0234213716
#> [181,]  0.0141660189
#> [182,]  0.0067751970
#> [183,]  0.0005198856
#> [184,]  0.0084695989
#> [185,]  0.0148915792
#> [186,]  0.0499511852
#> [187,]  0.0215529104
#> [188,] -0.0224976930
#> [189,] -0.0293394150
#> [190,] -0.0299163764
#> [191,]  0.0349481124
#> [192,] -0.0348385172
#> [193,]  0.0155268652
#> [194,] -0.0146783921
#> [195,]  0.0087095233
#> [196,] -0.0145297334
#> [197,] -0.0396200011
#> [198,]  0.0065742848
#> [199,]  0.0119224151
#> [200,] -0.0481219360
#> [201,]  0.0471485959
#> [202,] -0.0173415651
#> [203,] -0.0351459031
#> [204,]  0.0066405958
#> [205,] -0.0443197683
#> [206,]  0.0476516780
#> [207,] -0.0063464875
#> [208,] -0.0081631593
#> [209,]  0.0091076988
#> [210,]  0.0400294757
#> [211,]  0.0094354750
#> [212,]  0.0172927913
#> [213,] -0.0088471429
#> [214,] -0.0489781593
#> [215,]  0.0202286234
#> [216,] -0.0271672612
#> [217,]  0.0030188583
#> [218,]  0.0202359326
#> [219,]  0.0332122876
#> [220,]  0.0120355370
#> [221,] -0.0487785397
#> [222,]  0.0415468803
#> [223,]  0.0226467820
#> [224,]  0.0479416980
#> [225,] -0.0275996075
#> [226,] -0.0491144868
#> [227,]  0.0474935835
#> [228,]  0.0343168010
#> [229,]  0.0389112024
#> [230,] -0.0258209685
#> [231,] -0.0121918444
#> [232,] -0.0420192374
#> [233,] -0.0283980956
#> [234,]  0.0383471697
#> [235,]  0.0492783289
#> [236,]  0.0054242202
#> [237,]  0.0221705060
#> [238,] -0.0055958019
#> [239,] -0.0373771026
#> [240,]  0.0339889998
#> [241,]  0.0487770563
#> [242,]  0.0255775591
#> [243,]  0.0250430639
#> [244,]  0.0071043166
#> [245,] -0.0003512936
#> [246,]  0.0289206893
#> [247,] -0.0176823216
#> [248,] -0.0175330482
#> [249,]  0.0177750885
#> [250,]  0.0201823969
#> [251,] -0.0324931980
#> [252,] -0.0368393197
#> [253,]  0.0080396352
#> [254,]  0.0251980245
#> [255,] -0.0443637070
#> [256,]  0.0489498000
#> [257,]  0.0456054189
#> [258,] -0.0337738652
#> [259,]  0.0063907456
#> [260,] -0.0121837748
#> [261,]  0.0188696034
#> [262,] -0.0301792020
#> [263,]  0.0013068495
#> [264,] -0.0135454748
#> [265,] -0.0301305974
#> [266,] -0.0139486994
#> [267,] -0.0007730641
#> [268,] -0.0167835000
#> [269,] -0.0445043142
#> [270,]  0.0294958376
#> [271,]  0.0191389521
#> [272,]  0.0355913641
#> [273,]  0.0213377306
#> [274,] -0.0412513025
#> [275,] -0.0168443735
#> [276,]  0.0117259522
#> [277,]  0.0010004979
#> [278,] -0.0200507615
#> [279,]  0.0482584288
#> [280,]  0.0150716881
#> [281,]  0.0197244098
#> [282,] -0.0392193277
#> [283,] -0.0110151795
#> [284,]  0.0255675613
#> [285,]  0.0371528264
#> [286,] -0.0436272402
#> [287,]  0.0230249973
#> [288,]  0.0025096398
#> [289,]  0.0246775269
#> [290,]  0.0374241393
#> [291,]  0.0193517477
#> [292,] -0.0435180277
#> [293,]  0.0305815461
#> [294,] -0.0244791933
#> [295,] -0.0314262797
#> [296,] -0.0084417328
#> [297,] -0.0362624110
#> [298,] -0.0331478010
#> [299,]  0.0249309478
#> [300,]  0.0295108484
#> [301,] -0.0191200798
#> [302,]  0.0094126848
#> [303,]  0.0058763724
#> [304,]  0.0083123195
#> [305,] -0.0040908035
#> [306,] -0.0331562624
#> [307,] -0.0459406769
#> [308,] -0.0263532790
#> [309,] -0.0377790165
#> [310,] -0.0054815637
#> [311,] -0.0479714178
#> [312,] -0.0390876873
#> [313,]  0.0026662960
#> [314,] -0.0421718092
#> [315,]  0.0086325987
#> [316,] -0.0156443051
#> [317,]  0.0300898795
#> [318,] -0.0057084480
#> [319,] -0.0045738742
#> [320,]  0.0266026248
#> [321,] -0.0297770931
#> [322,]  0.0069560290
#> [323,] -0.0483589163
#> [324,] -0.0383498276
#> [325,]  0.0288464078
#> [326,]  0.0343661637
#> [327,] -0.0493101237
#> [328,]  0.0211965375
#> [329,]  0.0039784590
#> [330,] -0.0197525343
#> [331,]  0.0404901317
#> [332,] -0.0066356559
#> [333,]  0.0371381865
#> [334,] -0.0305965141
#> [335,] -0.0169804489
#> [336,]  0.0315736965
#> [337,]  0.0017287161
#> [338,] -0.0411404355
#> [339,]  0.0464021380
#> [340,]  0.0140651580
#> [341,]  0.0498393067
#> [342,] -0.0451536968
#> [343,]  0.0280449965
#> [344,]  0.0056458567
#> [345,]  0.0398530548
#> [346,]  0.0207916648
#> [347,] -0.0219589043
#> [348,] -0.0124213236
#> [349,] -0.0044157534
#> [350,] -0.0092767084
#> [351,] -0.0135475572
#> [352,]  0.0041835041
#> [353,]  0.0391406306
#> [354,] -0.0290602097
#> [355,] -0.0102725765
#> [356,]  0.0481411841
#> [357,]  0.0009434334
#> [358,]  0.0251287451
#> [359,]  0.0334533131
#> [360,] -0.0226924400
#> [361,] -0.0153882534
#> [362,]  0.0192061445
#> [363,]  0.0001013235
#> [364,] -0.0168872097
#> [365,]  0.0212829936
#> [366,]  0.0296480511
#> [367,]  0.0090389605
#> [368,] -0.0272115196
#> [369,] -0.0213846744
#> [370,] -0.0009698119
#> [371,] -0.0480860865
#> [372,]  0.0349666421
#> [373,] -0.0098777179
#> [374,]  0.0041983687
#> [375,]  0.0130523423
#> [376,]  0.0093946351
#> [377,]  0.0121117136
#> [378,]  0.0022403454
#> [379,] -0.0369144969
#> [380,]  0.0088323499
#> [381,] -0.0095764856
#> [382,] -0.0464114999
#> [383,] -0.0477102483
#> [384,] -0.0474303407
#> [385,] -0.0087080968
#> [386,] -0.0178621636
#> [387,]  0.0369462701
#> [388,]  0.0208844455
#> [389,]  0.0376464167
#> [390,]  0.0228666042
#> [391,] -0.0485046670
#> [392,] -0.0461939265
#> [393,] -0.0065362056
#> [394,] -0.0114326499
#> [395,] -0.0024644357
#> [396,] -0.0395197440
#> [397,] -0.0187534676
#> [398,] -0.0359935214
#> [399,] -0.0124769656
#> [400,]  0.0076072273