Generate a spiral dataset with optional noise.

This function generates a dataset arranged in a spiral pattern with optional noise.

Usage

spiral_3d(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_dims

Number of effective dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

spiral_3d(n = 100, num_dims = 10, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>                [,1]         [,2]          [,3]          [,4]          [,5]
#>   [1,] -1.745023665 -1.462364406 -1.225490e+00 -1.026985e+00 -8.606339e-01
#>   [2,] -1.339970159  0.860563866 -5.526766e-01  3.549434e-01 -2.279539e-01
#>   [3,] -1.647253633 -1.432296745 -1.245390e+00 -1.082874e+00 -9.415656e-01
#>   [4,] -1.407507550 -0.601989446 -2.574702e-01 -1.101197e-01 -4.709809e-02
#>   [5,]  0.561333911 -0.266797117  1.268063e-01 -6.026995e-02  2.864578e-02
#>   [6,]  0.912119862  0.799213113  7.002825e-01  6.135981e-01  5.376438e-01
#>   [7,] -1.874965737 -1.325516969 -9.370812e-01 -6.624746e-01 -4.683399e-01
#>   [8,]  0.111548614 -0.110890583  1.102364e-01 -1.095861e-01  1.089397e-01
#>   [9,]  0.139657808  0.086611027  5.371322e-02  3.331111e-02  2.065842e-02
#>  [10,] -0.564339677 -0.211771894 -7.946869e-02 -2.982111e-02 -1.119055e-02
#>  [11,]  0.099640399  0.079333772  6.316562e-02  5.029252e-02  4.004295e-02
#>  [12,] -0.390256062 -0.381866695 -3.736577e-01 -3.656251e-01 -3.577653e-01
#>  [13,] -0.401750222  0.397951086 -3.941879e-01  3.904603e-01 -3.867679e-01
#>  [14,]  0.562553274  0.134502129  3.215842e-02  7.688828e-03  1.838339e-03
#>  [15,]  1.217089102 -0.477671542  1.874720e-01 -7.357722e-02  2.887689e-02
#>  [16,] -0.151553931  0.151357804 -1.511619e-01  1.509663e-01 -1.507709e-01
#>  [17,] -0.362964912 -0.356304006 -3.497653e-01 -3.433467e-01 -3.370458e-01
#>  [18,]  0.044409079  0.004217630  4.005578e-04  3.804188e-05  3.612923e-06
#>  [19,] -0.205341919  0.073257639 -2.613534e-02  9.324027e-03 -3.326433e-03
#>  [20,]  0.026546478  0.025380020  2.426482e-02  2.319861e-02  2.217926e-02
#>  [21,] -0.274085873 -0.271322504 -2.685870e-01 -2.658791e-01 -2.631984e-01
#>  [22,] -2.361987797  1.605099969 -1.090753e+00  7.412265e-01 -5.037040e-01
#>  [23,]  0.656042615  0.647147807  6.383736e-01  6.297184e-01  6.211805e-01
#>  [24,] -0.345267995  0.307021193 -2.730112e-01  2.427685e-01 -2.158760e-01
#>  [25,] -0.807511389 -0.473789396 -2.779854e-01 -1.631018e-01 -9.569635e-02
#>  [26,]  1.033297174  0.844938140  6.909149e-01  5.649685e-01  4.619808e-01
#>  [27,]  1.052129971  0.846648754  6.812981e-01  5.482404e-01  4.411689e-01
#>  [28,]  0.144780801  0.075495077  3.936645e-02  2.052740e-02  1.070389e-02
#>  [29,] -0.053728085  0.053647457 -5.356695e-02  5.348656e-02 -5.340630e-02
#>  [30,] -0.846971975 -0.695193932 -5.706146e-01 -4.683600e-01 -3.844296e-01
#>  [31,] -1.361432092  0.423942616 -1.320134e-01  4.110828e-02 -1.280090e-02
#>  [32,]  0.144473377  0.072875797  3.676028e-02  1.854276e-02  9.353406e-03
#>  [33,]  0.001717847  0.001713197  1.708560e-03  1.703935e-03  1.699322e-03
#>  [34,]  0.285341685 -0.023518743  1.938487e-03 -1.597761e-04  1.316924e-05
#>  [35,] -1.252499943  1.071344307 -9.163902e-01  7.838478e-01 -6.704758e-01
#>  [36,]  2.074373294  1.595870281  1.227745e+00  9.445369e-01  7.266573e-01
#>  [37,]  0.447631235  0.436497964  4.256416e-01  4.150552e-01  4.047322e-01
#>  [38,] -0.313162089  0.171203173 -9.359539e-02  5.116784e-02 -2.797304e-02
#>  [39,]  0.148305277  0.147904217  1.475042e-01  1.471053e-01  1.467075e-01
#>  [40,] -1.187331906  1.045147893 -9.199905e-01  8.098209e-01 -7.128441e-01
#>  [41,] -0.512558290 -0.173418838 -5.867448e-02 -1.985191e-02 -6.716690e-03
#>  [42,]  0.161964682  0.010604044  6.942610e-04  4.545419e-05  2.975947e-06
#>  [43,]  2.085035112  1.291836127  8.003897e-01  4.959016e-01  3.072484e-01
#>  [44,]  0.400058823  0.392155180  3.844077e-01  3.768132e-01  3.693688e-01
#>  [45,] -0.144580509 -0.005955194 -2.452912e-04 -1.010341e-05 -4.161542e-07
#>  [46,]  0.544971448 -0.246599648  1.115864e-01 -5.049285e-02  2.284802e-02
#>  [47,]  2.105307859  1.563992415  1.161860e+00  8.631230e-01  6.411974e-01
#>  [48,]  0.042603846  0.003869259  3.514041e-04  3.191434e-05  2.898444e-06
#>  [49,]  0.188368719 -0.010203127  5.526597e-04 -2.993521e-05  1.621462e-06
#>  [50,]  1.565744484  1.423625333  1.294406e+00  1.176916e+00  1.070090e+00
#>  [51,] -0.026464150  0.001358158 -6.970157e-05  3.577131e-06 -1.835807e-07
#>  [52,] -1.376080987  0.980654740 -6.988569e-01  4.980356e-01 -3.549217e-01
#>  [53,] -1.174130644 -0.406798357 -1.409425e-01 -4.883202e-02 -1.691872e-02
#>  [54,]  0.511438105 -0.210159298  8.635831e-02 -3.548621e-02  1.458194e-02
#>  [55,] -0.379592196  0.277527327 -2.029057e-01  1.483483e-01 -1.084604e-01
#>  [56,]  0.138851190  0.059307902  2.533235e-02  1.082028e-02  4.621696e-03
#>  [57,] -0.206942989 -0.028008502 -3.790784e-03 -5.130600e-04 -6.943962e-05
#>  [58,]  0.121290454  0.121070630  1.208512e-01  1.206322e-01  1.204135e-01
#>  [59,] -1.315030714  1.080097061 -8.871349e-01  7.286460e-01 -5.984715e-01
#>  [60,] -0.231709522 -0.035078926 -5.310662e-03 -8.039909e-04 -1.217176e-04
#>  [61,]  0.042853685 -0.042815290  4.277693e-02 -4.273860e-02  4.270031e-02
#>  [62,]  0.101095228  0.025116824  6.240204e-03  1.550361e-03  3.851828e-04
#>  [63,] -0.861243434 -0.689635315 -5.522212e-01 -4.421877e-01 -3.540790e-01
#>  [64,]  0.337609088  0.046868883  6.506615e-03  9.032866e-04  1.253996e-04
#>  [65,]  0.115891310  0.086181794  6.408851e-02  4.765899e-02  3.544129e-02
#>  [66,]  0.958848627  0.212912449  4.727723e-02  1.049792e-02  2.331063e-03
#>  [67,] -1.709452985 -1.454490375 -1.237555e+00 -1.052975e+00 -8.959256e-01
#>  [68,] -1.364818488  1.052932264 -8.123178e-01  6.266882e-01 -4.834784e-01
#>  [69,]  0.611664026 -0.456811096  3.411618e-01 -2.547910e-01  1.902864e-01
#>  [70,] -0.104198745  0.002409455 -5.571537e-05  1.288342e-06 -2.979117e-08
#>  [71,]  0.753161468  0.739603885  7.262903e-01  7.132165e-01  7.003779e-01
#>  [72,] -0.335656033  0.199370452 -1.184206e-01  7.033856e-02 -4.177917e-02
#>  [73,]  1.184094564 -0.448748831  1.700671e-01 -6.445212e-02  2.442610e-02
#>  [74,]  0.837144769  0.319107452  1.216391e-01  4.636708e-02  1.767446e-02
#>  [75,]  0.673204739 -0.655960282  6.391575e-01 -6.227852e-01  6.068323e-01
#>  [76,] -2.307555239  1.839045958 -1.465659e+00  1.168083e+00 -9.309235e-01
#>  [77,]  0.938965910  0.812871212  7.037099e-01  6.092080e-01  5.273968e-01
#>  [78,]  0.279199385 -0.022510057  1.814842e-03 -1.463190e-04  1.179677e-05
#>  [79,]  0.004680273  0.004645565  4.611114e-03  4.576919e-03  4.542977e-03
#>  [80,] -1.305835187  1.080444672 -8.939571e-01  7.396578e-01 -6.119910e-01
#>  [81,] -1.069337616  0.257378390 -6.194829e-02  1.491031e-02 -3.588755e-03
#>  [82,]  0.109992079 -0.109918598  1.098452e-01 -1.097718e-01  1.096984e-01
#>  [83,]  0.402718098 -0.047162198  5.523151e-03 -6.468146e-04  7.574827e-05
#>  [84,] -0.118648693  0.025346470 -5.414670e-03  1.156715e-03 -2.471047e-04
#>  [85,]  0.382411992 -0.042472560  4.717212e-03 -5.239168e-04  5.818878e-05
#>  [86,] -0.026315285  0.026306009 -2.629674e-02  2.628747e-02 -2.627820e-02
#>  [87,]  0.337135894 -0.082473399  2.017543e-02 -4.935506e-03  1.207371e-03
#>  [88,] -1.316027778  1.080010965 -8.863215e-01  7.273683e-01 -5.969219e-01
#>  [89,]  0.624000978 -0.380347679  2.318335e-01 -1.413096e-01  8.613254e-02
#>  [90,] -1.755019377  1.614951788 -1.486063e+00  1.367461e+00 -1.258324e+00
#>  [91,] -0.818122638 -0.192603210 -4.534283e-02 -1.067465e-02 -2.513037e-03
#>  [92,]  0.046796927  0.043001140  3.951324e-02  3.630824e-02  3.336321e-02
#>  [93,] -0.711928774  0.112803082 -1.787333e-02  2.831978e-03 -4.487188e-04
#>  [94,]  0.545812610 -0.087514911  1.403203e-02 -2.249878e-03  3.607427e-04
#>  [95,] -1.315368137  0.809921769 -4.986994e-01  3.070680e-01 -1.890734e-01
#>  [96,]  0.422462512 -0.385673907  3.520889e-01 -3.214285e-01  2.934381e-01
#>  [97,]  0.213511058 -0.208990025  2.045647e-01 -2.002331e-01  1.959932e-01
#>  [98,]  0.129292208  0.047050354  1.712196e-02  6.230802e-03  2.267433e-03
#>  [99,]  0.385896584 -0.110406548  3.158775e-02 -9.037382e-03  2.585631e-03
#> [100,] -0.142355407  0.140715902 -1.390953e-01  1.374933e-01 -1.359098e-01
#>                 [,6]          [,7]          [,8]          [,9]       [,10]
#>   [1,] -7.212283e-01 -6.044036e-01 -5.065022e-01 -4.244589e-01  3.19812386
#>   [2,]  1.463980e-01 -9.402060e-02  6.038249e-02 -3.877921e-02 -1.74813249
#>   [3,] -8.186969e-01 -7.118618e-01 -6.189680e-01 -5.381963e-01  3.33504431
#>   [4,] -2.014380e-02 -8.615482e-03 -3.684832e-03 -1.575999e-03  1.55711301
#>   [5,] -1.361509e-02  6.471134e-03 -3.075674e-03  1.461841e-03 -0.63800378
#>   [6,]  4.710916e-01  4.127775e-01  3.616819e-01  3.169111e-01  1.89267748
#>   [7,] -3.310954e-01 -2.340696e-01 -1.654768e-01 -1.169847e-01  2.65103418
#>   [8,] -1.082970e-01  1.076582e-01 -1.070231e-01  1.063918e-01 -1.02848844
#>   [9,]  1.281165e-02  7.945350e-03  4.927436e-03  3.055828e-03  0.17802810
#>  [10,] -4.199323e-03 -1.575821e-03 -5.913366e-04 -2.219026e-04  0.60883243
#>  [11,]  3.188223e-02  2.538466e-02  2.021129e-02  1.609225e-02  0.16468606
#>  [12,] -3.500743e-01 -3.425488e-01 -3.351850e-01 -3.279795e-01  1.89230675
#>  [13,]  3.831104e-01 -3.794876e-01  3.758989e-01 -3.723443e-01 -2.92823499
#>  [14,]  4.395327e-04  1.050889e-04  2.512593e-05  6.007416e-06  0.57935643
#>  [15,] -1.133333e-02  4.447996e-03 -1.745707e-03  6.851384e-04 -1.32326152
#>  [16,]  1.505758e-01 -1.503810e-01  1.501864e-01 -1.499920e-01 -2.97994305
#>  [17,] -3.308605e-01 -3.247888e-01 -3.188285e-01 -3.129775e-01  1.90334273
#>  [18,]  3.431274e-07  3.258758e-08  3.094916e-09  2.939311e-10  0.04461072
#>  [19,]  1.186736e-03 -4.233791e-04  1.510444e-04 -5.388650e-05 -0.21980596
#>  [20,]  2.120470e-02  2.027296e-02  1.938216e-02  1.853050e-02  0.09054920
#>  [21,] -2.605448e-01 -2.579180e-01 -2.553176e-01 -2.527435e-01  1.93505349
#>  [22,]  3.422944e-01 -2.326078e-01  1.580697e-01 -1.074170e-01 -3.21961272
#>  [23,]  6.127583e-01  6.044504e-01  5.962551e-01  5.881709e-01  3.99753123
#>  [24,]  1.919625e-01 -1.706980e-01  1.517891e-01 -1.349748e-01 -0.75473587
#>  [25,] -5.614771e-02 -3.294342e-02 -1.932882e-02 -1.134076e-02  0.99719330
#>  [26,]  3.777666e-01  3.089038e-01  2.525940e-01  2.065488e-01  1.79507358
#>  [27,]  3.550085e-01  2.856753e-01  2.298828e-01  1.849867e-01  1.77221146
#>  [28,]  5.581478e-03  2.910428e-03  1.517625e-03  7.913566e-04  0.16967451
#>  [29,]  5.332615e-02 -5.324613e-02  5.316622e-02 -5.308644e-02 -0.98108487
#>  [30,] -3.155395e-01 -2.589946e-01 -2.125825e-01 -1.744876e-01  1.48275043
#>  [31,]  3.986130e-03 -1.241259e-03  3.865215e-04 -1.203607e-04 -1.43266311
#>  [32,]  4.718080e-03  2.379911e-03  1.200484e-03  6.055525e-04  0.16731985
#>  [33,]  1.694722e-03  1.690135e-03  1.685560e-03  1.680997e-03  0.02336280
#>  [34,] -1.085450e-06  8.946611e-08 -7.374073e-09  6.077939e-10 -0.28631590
#>  [35,]  5.735014e-01 -4.905529e-01  4.196016e-01 -3.589125e-01 -2.41783130
#>  [36,]  5.590367e-01  4.300817e-01  3.308732e-01  2.545495e-01  3.24701312
#>  [37,]  3.946659e-01  3.848499e-01  3.752781e-01  3.659444e-01  2.01962867
#>  [38,]  1.529263e-02 -8.360359e-03  4.570541e-03 -2.498677e-03 -0.37399908
#>  [39,]  1.463108e-01  1.459151e-01  1.455205e-01  1.451270e-01  2.01794286
#>  [40,]  6.274804e-01 -5.523391e-01  4.861960e-01 -4.279736e-01 -2.50221789
#>  [41,] -2.272523e-03 -7.688849e-04 -2.601443e-04 -8.801716e-05  0.54468141
#>  [42,]  1.948392e-07  1.275638e-08  8.351775e-10  5.468019e-11  0.16231293
#>  [43,]  1.903635e-01  1.179445e-01  7.307549e-02  4.527576e-02  2.65630794
#>  [44,]  3.620715e-01  3.549184e-01  3.479065e-01  3.410332e-01  2.02261290
#>  [45,] -1.714117e-08 -7.060356e-10 -2.908123e-11 -1.197840e-12  0.14470331
#>  [46,] -1.033873e-02  4.678277e-03 -2.116921e-03  9.579070e-04 -0.61111599
#>  [47,]  4.763331e-01  3.538586e-01  2.628747e-01  1.952845e-01  3.14495465
#>  [48,]  2.632352e-07  2.390688e-08  2.171211e-09  1.971883e-10  0.04278064
#>  [49,] -8.782766e-08  4.757248e-09 -2.576798e-10  1.395741e-11 -0.18864566
#>  [50,]  9.729600e-01  8.846466e-01  8.043492e-01  7.313402e-01  3.76119332
#>  [51,]  9.421486e-09 -4.835169e-10  2.481441e-11 -1.273492e-12 -0.02649907
#>  [52,]  2.529325e-01 -1.802506e-01  1.284544e-01 -9.154215e-02 -1.96155188
#>  [53,] -5.861790e-03 -2.030921e-03 -7.036485e-04 -2.437915e-04  1.25165583
#>  [54,] -5.991985e-03  2.462217e-03 -1.011770e-03  4.157548e-04 -0.56098906
#>  [55,]  7.929753e-02 -5.797599e-02  4.238738e-02 -3.099025e-02 -0.55638348
#>  [56,]  1.974078e-03  8.431936e-04  3.601557e-04  1.538343e-04  0.15356436
#>  [57,] -9.398240e-06 -1.271996e-06 -1.721571e-07 -2.330043e-08  0.20886482
#>  [58,]  1.201953e-01  1.199775e-01  1.197600e-01  1.195430e-01  2.01550754
#>  [59,]  4.915530e-01 -4.037358e-01  3.316073e-01 -2.723648e-01 -2.30533919
#>  [60,] -1.842705e-05 -2.789705e-06 -4.223386e-07 -6.393861e-08  0.23441139
#>  [61,] -4.266205e-02  4.262383e-02 -4.258564e-02  4.254748e-02 -1.01256743
#>  [62,]  9.569758e-05  2.377579e-05  5.907028e-06  1.467584e-06  0.10436763
#>  [63,] -2.835266e-01 -2.270321e-01 -1.817946e-01 -1.455709e-01  1.43778520
#>  [64,]  1.740871e-05  2.416779e-06  3.355116e-07  4.657769e-08  0.34091019
#>  [65,]  2.635568e-02  1.959922e-02  1.457483e-02  1.083847e-02  0.17334036
#>  [66,]  5.176129e-04  1.149360e-04  2.552155e-05  5.667064e-06  0.98339893
#>  [67,] -7.622995e-01 -6.486035e-01 -5.518652e-01 -4.695552e-01  3.25358175
#>  [68,]  3.729946e-01 -2.877585e-01  2.220003e-01 -1.712692e-01 -2.14509205
#>  [69,] -1.421122e-01  1.061342e-01 -7.926453e-02  5.919739e-02 -0.91977999
#>  [70,]  6.888804e-10 -1.592943e-11  3.683464e-13 -8.517510e-15 -0.10422661
#>  [71,]  6.877705e-01  6.753900e-01  6.632324e-01  6.512936e-01  3.98739511
#>  [72,]  2.481568e-02 -1.473983e-02  8.755053e-03 -5.200261e-03 -0.41723076
#>  [73,] -9.257018e-03  3.508230e-03 -1.329551e-03  5.038740e-04 -1.27954251
#>  [74,]  6.737247e-03  2.568141e-03  9.789381e-04  3.731570e-04  0.90551223
#>  [75,] -5.912880e-01  5.761419e-01 -5.613837e-01  5.470036e-01 -2.99350586
#>  [76,]  7.419156e-01 -5.912824e-01  4.712327e-01 -3.755570e-01 -3.82031328
#>  [77,]  4.565721e-01  3.952586e-01  3.421789e-01  2.962274e-01  1.87587707
#>  [78,] -9.510977e-07  7.668091e-08 -6.182291e-09  4.984385e-10 -0.28011125
#>  [79,]  4.509287e-03  4.475846e-03  4.442654e-03  4.409708e-03  0.03850189
#>  [80,]  5.063598e-01 -4.189608e-01  3.466471e-01 -2.868149e-01 -2.32513529
#>  [81,]  8.637758e-04 -2.079018e-04  5.003979e-05 -1.204406e-05 -1.10172599
#>  [82,] -1.096252e-01  1.095519e-01 -1.094787e-01  1.094056e-01 -3.00962385
#>  [83,] -8.870857e-06  1.038863e-06 -1.216610e-07  1.424768e-08 -0.40550841
#>  [84,]  5.278805e-05 -1.127691e-05  2.409044e-06 -5.146348e-07 -0.12145236
#>  [85,] -6.462733e-06  7.177830e-07 -7.972051e-08  8.854153e-09 -0.38479265
#>  [86,]  2.626894e-02 -2.625968e-02  2.625042e-02 -2.624117e-02 -0.99119871
#>  [87,] -2.953585e-04  7.225342e-05 -1.767532e-05  4.323906e-06 -0.34770020
#>  [88,]  4.898697e-01 -4.020163e-01  3.299186e-01 -2.707509e-01 -2.30309431
#>  [89,] -5.250042e-02  3.200061e-02 -1.950535e-02  1.188911e-02 -0.78712216
#>  [90,]  1.157898e+00 -1.065486e+00  9.804500e-01 -9.022005e-01 -4.48313599
#>  [91,] -5.916215e-04 -1.392801e-04 -3.278945e-05 -7.719324e-06  0.84178219
#>  [92,]  3.065706e-02  2.817041e-02  2.588545e-02  2.378583e-02  0.11861790
#>  [93,]  7.109821e-05 -1.126531e-05  1.784956e-06 -2.828212e-07 -0.72103731
#>  [94,] -5.784103e-05  9.274158e-06 -1.487007e-06  2.384248e-07 -0.55296687
#>  [95,]  1.164196e-01 -7.168394e-02  4.413850e-02 -2.717774e-02 -1.66935232
#>  [96,] -2.678851e-01  2.445573e-01 -2.232609e-01  2.038191e-01 -1.03508818
#>  [97,] -1.918431e-01  1.877809e-01 -1.838047e-01  1.799127e-01 -1.04305683
#>  [98,]  8.251351e-04  3.002725e-04  1.092713e-04  3.976461e-05  0.13880966
#>  [99,] -7.397593e-04  2.116481e-04 -6.055336e-05  1.732456e-05 -0.40273133
#> [100,]  1.343445e-01 -1.327973e-01  1.312679e-01 -1.297561e-01 -0.94068373
#>                [,11]         [,12]         [,13]         [,14]
#>   [1,]  0.0040772037  0.0298328090  0.0398905383  0.0276991382
#>   [2,]  0.0316639734  0.0242993464 -0.0249748567 -0.0180563499
#>   [3,]  0.0219678072  0.0467767697  0.0396012254 -0.0176973455
#>   [4,]  0.0316782717 -0.0444113135 -0.0380786851  0.0262855072
#>   [5,]  0.0075809374  0.0199388383  0.0058208612 -0.0327652673
#>   [6,]  0.0019381610  0.0222760492 -0.0122504012 -0.0101314949
#>   [7,] -0.0354225670  0.0482019972  0.0200945648  0.0171928292
#>   [8,] -0.0082201671 -0.0062034270  0.0227118198 -0.0211091693
#>   [9,]  0.0030426393  0.0179884742 -0.0399813341 -0.0491011232
#>  [10,] -0.0073778888 -0.0016666973 -0.0317369873  0.0114381643
#>  [11,] -0.0244132529 -0.0369509361  0.0437110557  0.0269541200
#>  [12,]  0.0116682196 -0.0389772057  0.0101894685 -0.0458596702
#>  [13,]  0.0385070114  0.0217582253 -0.0390691745 -0.0022772569
#>  [14,] -0.0045250302 -0.0039795617 -0.0103816213 -0.0280706843
#>  [15,]  0.0386503571 -0.0040027744 -0.0136437392 -0.0137750208
#>  [16,] -0.0132987844 -0.0252100952  0.0006388123  0.0346932262
#>  [17,]  0.0005573324 -0.0260198543  0.0441887991  0.0120381791
#>  [18,]  0.0451539532  0.0036594979  0.0337064095  0.0494026359
#>  [19,]  0.0035249788 -0.0153834191  0.0465702948  0.0114571995
#>  [20,]  0.0113482851 -0.0153936563  0.0255896111  0.0042938906
#>  [21,] -0.0237902335  0.0066760563  0.0260629578 -0.0126659537
#>  [22,] -0.0250188419  0.0428217589 -0.0449200979  0.0361302305
#>  [23,]  0.0168281887  0.0359568012  0.0323711666 -0.0096781683
#>  [24,]  0.0361658940  0.0284178685  0.0232043326 -0.0247646410
#>  [25,]  0.0384284920  0.0157947525  0.0444635089 -0.0178268573
#>  [26,] -0.0482573789  0.0067502912 -0.0014560674  0.0008513175
#>  [27,]  0.0018153150 -0.0227620619  0.0259912651  0.0440158128
#>  [28,]  0.0056891972  0.0356345365  0.0486039246 -0.0164261570
#>  [29,] -0.0235091108  0.0194568714 -0.0147767596 -0.0463968792
#>  [30,]  0.0433118782 -0.0082367483 -0.0463814764 -0.0244823189
#>  [31,] -0.0318508651  0.0256606100  0.0189252493 -0.0284634235
#>  [32,]  0.0101559245 -0.0289314437  0.0277364371 -0.0274507443
#>  [33,] -0.0430584712  0.0167165421 -0.0317362003 -0.0355362587
#>  [34,]  0.0461112634  0.0246609866 -0.0111843716 -0.0203971454
#>  [35,] -0.0328340234  0.0442919880 -0.0142469462  0.0025486021
#>  [36,] -0.0091125147 -0.0351362764  0.0078181784 -0.0167975358
#>  [37,] -0.0322095023 -0.0394620901  0.0492442785 -0.0330224035
#>  [38,]  0.0261450251 -0.0095372504 -0.0190052201 -0.0259280276
#>  [39,] -0.0306117061 -0.0060365051  0.0141284323 -0.0195730401
#>  [40,] -0.0483416579  0.0245002729 -0.0004196112  0.0090929748
#>  [41,]  0.0001050768 -0.0341258698 -0.0161055679 -0.0258267094
#>  [42,]  0.0242393107  0.0378875782 -0.0282138446  0.0105830499
#>  [43,]  0.0448149232  0.0493627586  0.0122271145  0.0377947381
#>  [44,]  0.0403402146  0.0438200423 -0.0200707278 -0.0345487010
#>  [45,] -0.0384512077  0.0479409811  0.0431122861 -0.0488507861
#>  [46,]  0.0311754012 -0.0372170412  0.0396312418 -0.0055803093
#>  [47,]  0.0094728988 -0.0445425876  0.0094269007  0.0480059808
#>  [48,]  0.0248839530  0.0347888250 -0.0343401683  0.0020404731
#>  [49,] -0.0084844274  0.0230775382 -0.0281445383  0.0199289078
#>  [50,]  0.0443554433  0.0490263238 -0.0184244748  0.0151728256
#>  [51,]  0.0108276782 -0.0122115753  0.0030408879 -0.0162109953
#>  [52,] -0.0470620946  0.0250296060  0.0490225402 -0.0174251484
#>  [53,] -0.0268604745 -0.0332818319  0.0091290776 -0.0355280585
#>  [54,] -0.0485144481 -0.0356689838 -0.0447241319  0.0490583975
#>  [55,]  0.0091415693  0.0437616044 -0.0313326847  0.0445261759
#>  [56,]  0.0359738877  0.0409066749  0.0306797587 -0.0260368642
#>  [57,]  0.0039780248  0.0299123898  0.0302366201 -0.0098339510
#>  [58,]  0.0294111057  0.0381772182  0.0334690457  0.0043909939
#>  [59,] -0.0288458679 -0.0396578943  0.0274756540 -0.0310882302
#>  [60,]  0.0143177871  0.0055896225  0.0145580373 -0.0321653562
#>  [61,] -0.0067108722  0.0104021395 -0.0299059596 -0.0233343257
#>  [62,] -0.0100103488  0.0268479078  0.0211273911 -0.0175633502
#>  [63,]  0.0493065814  0.0187878912  0.0222871623  0.0276680674
#>  [64,] -0.0463613175 -0.0159617424 -0.0239558213  0.0133724439
#>  [65,] -0.0001963110  0.0133452800 -0.0001184794 -0.0140489068
#>  [66,]  0.0114313675 -0.0274410672  0.0350594132 -0.0290526897
#>  [67,]  0.0096035828  0.0219067287 -0.0426056546 -0.0140496918
#>  [68,] -0.0045797634 -0.0177262414 -0.0123135583  0.0086276682
#>  [69,] -0.0490253717  0.0114701733  0.0257602069 -0.0417195622
#>  [70,]  0.0240629866 -0.0276317285 -0.0128970262 -0.0113505138
#>  [71,] -0.0190400123  0.0189099878  0.0007094609  0.0082157006
#>  [72,] -0.0140088327  0.0124666015 -0.0249702413 -0.0427407475
#>  [73,] -0.0141258125  0.0169812954 -0.0044214039 -0.0434692814
#>  [74,] -0.0309725679 -0.0378901855  0.0458791232  0.0498975376
#>  [75,]  0.0186676408 -0.0161097038  0.0021172850  0.0149752548
#>  [76,]  0.0076331608 -0.0067666897 -0.0263949672 -0.0339722201
#>  [77,]  0.0051156343  0.0092469468  0.0145284006  0.0152976293
#>  [78,] -0.0163840514 -0.0331507916 -0.0240693813 -0.0041918302
#>  [79,] -0.0465137657 -0.0006146387  0.0243802240  0.0032078369
#>  [80,]  0.0078086089  0.0245606462 -0.0454881707  0.0333994237
#>  [81,] -0.0016921463 -0.0378229271 -0.0318028278  0.0088629642
#>  [82,] -0.0109608324 -0.0122767020  0.0283666552  0.0299604747
#>  [83,] -0.0140914923 -0.0453672324 -0.0196966746  0.0225377546
#>  [84,] -0.0208868657  0.0061189950  0.0377512551 -0.0065663550
#>  [85,]  0.0469652987 -0.0219731480  0.0377733182 -0.0222809727
#>  [86,]  0.0213770616 -0.0215335082  0.0123054524  0.0355557095
#>  [87,] -0.0049534994  0.0077105954 -0.0076687456  0.0030670910
#>  [88,]  0.0284848131  0.0047630362 -0.0216983922  0.0321043153
#>  [89,]  0.0250173339 -0.0277419638  0.0217162577 -0.0062106649
#>  [90,]  0.0412070681  0.0471964790  0.0397845391 -0.0276588276
#>  [91,]  0.0223791803 -0.0136764089 -0.0387590126  0.0023145442
#>  [92,]  0.0026508256 -0.0402014994  0.0460773171  0.0491028031
#>  [93,] -0.0022635695 -0.0420707073  0.0486065194 -0.0389019981
#>  [94,]  0.0310788018 -0.0485590968  0.0125327633  0.0044095015
#>  [95,]  0.0392412572 -0.0122644665  0.0151942990  0.0102487721
#>  [96,]  0.0122624468 -0.0261378356 -0.0195106082  0.0471909204
#>  [97,]  0.0392029011 -0.0376238571  0.0408028610  0.0162224739
#>  [98,] -0.0131748912 -0.0292970977  0.0136892191  0.0391561583
#>  [99,]  0.0052440108  0.0023666020 -0.0358751221 -0.0419421863
#> [100,] -0.0097453754  0.0211908803  0.0144400758  0.0047260479