Generate a spiral dataset with optional noise.

This function generates a dataset arranged in a spiral pattern with optional noise.

Usage

spiral_with_noise(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_dims

Number of effective dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

spiral_with_noise(n = 100, num_dims = 10, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,] -0.370670834  3.132640e-01 -2.647479e-01  2.237456e-01 -1.890934e-01
#>   [2,] -1.129755948  5.486004e-01 -2.663960e-01  1.293598e-01 -6.281606e-02
#>   [3,]  1.483331883  1.365668e+00  1.257337e+00  1.157600e+00  1.065774e+00
#>   [4,]  0.153005151 -1.528076e-01  1.526104e-01 -1.524134e-01  1.522167e-01
#>   [5,] -0.663527168 -2.991358e-01 -1.348584e-01 -6.079776e-02 -2.740925e-02
#>   [6,] -1.029283287  9.482623e-01 -8.736189e-01  8.048512e-01 -7.414966e-01
#>   [7,]  0.007380962  7.294199e-03  7.208456e-03  7.123720e-03  7.039981e-03
#>   [8,] -0.920092192  3.486646e-01 -1.321248e-01  5.006809e-02 -1.897307e-02
#>   [9,] -0.333522658 -7.259956e-02 -1.580311e-02 -3.439943e-03 -7.487898e-04
#>  [10,] -0.045715287  8.341561e-04 -1.522065e-05  2.777276e-07 -5.067631e-09
#>  [11,] -1.487186986 -1.345621e+00 -1.217531e+00 -1.101633e+00 -9.967682e-01
#>  [12,] -1.482801388 -6.768741e-01 -3.089817e-01 -1.410450e-01 -6.438468e-02
#>  [13,]  0.121460151  3.956408e-02  1.288749e-02  4.197935e-03  1.367423e-03
#>  [14,] -1.082569703  9.846186e-01 -8.955301e-01  8.145023e-01 -7.408060e-01
#>  [15,]  0.255074835  2.530410e-01  2.510235e-01  2.490220e-01  2.470365e-01
#>  [16,]  0.228889049  2.285162e-01  2.281439e-01  2.277723e-01  2.274012e-01
#>  [17,]  1.619604069 -1.060205e+00  6.940175e-01 -4.543089e-01  2.973939e-01
#>  [18,]  0.676028565 -6.585598e-01  6.415423e-01 -6.249647e-01  6.088154e-01
#>  [19,]  1.894531112  1.597780e+00  1.347511e+00  1.136443e+00  9.584354e-01
#>  [20,]  0.167245710 -1.924297e-02  2.214059e-03 -2.547454e-04  2.931051e-05
#>  [21,] -0.375675722  2.667917e-01 -1.894662e-01  1.345522e-01 -9.555429e-02
#>  [22,] -0.815421969 -4.862814e-01 -2.899966e-01 -1.729411e-01 -1.031344e-01
#>  [23,]  0.308004080  3.044196e-01  3.008769e-01  2.973754e-01  2.939146e-01
#>  [24,] -0.970727935  9.052091e-01 -8.441125e-01  7.871396e-01 -7.340120e-01
#>  [25,]  1.457419406 -1.220431e+00  1.021979e+00 -8.557974e-01  7.166379e-01
#>  [26,] -0.330955098  1.931446e-01 -1.127187e-01  6.578239e-02 -3.839045e-02
#>  [27,] -0.149088827  1.471897e-01 -1.453147e-01  1.434637e-01 -1.416362e-01
#>  [28,] -1.216837059  6.555243e-01 -3.531386e-01  1.902399e-01 -1.024844e-01
#>  [29,] -1.874276862 -1.385415e+00 -1.024062e+00 -7.569591e-01 -5.595239e-01
#>  [30,]  1.384348824  1.291290e+00  1.204487e+00  1.123519e+00  1.047994e+00
#>  [31,]  0.031224124  2.038879e-03  1.331351e-04  8.693484e-06  5.676688e-07
#>  [32,]  1.611942550 -1.188339e+00  8.760551e-01 -6.458362e-01  4.761166e-01
#>  [33,]  0.014060832  1.374185e-02  1.343011e-02  1.312544e-02  1.282769e-02
#>  [34,] -1.326053354 -1.233286e+00 -1.147008e+00 -1.066766e+00 -9.921380e-01
#>  [35,]  0.158970499 -1.570900e-01  1.552317e-01 -1.533955e-01  1.515809e-01
#>  [36,] -0.371423483  3.131672e-01 -2.640483e-01  2.226334e-01 -1.877143e-01
#>  [37,]  1.344349168 -1.174360e+00  1.025866e+00 -8.961488e-01  7.828336e-01
#>  [38,] -2.163868652  1.200817e+00 -6.663807e-01  3.698011e-01 -2.052173e-01
#>  [39,]  1.626677127 -1.132425e+00  7.883465e-01 -5.488138e-01  3.820612e-01
#>  [40,]  0.164867312 -1.868891e-02  2.118525e-03 -2.401502e-04  2.722278e-05
#>  [41,] -0.856637417  2.996936e-01 -1.048475e-01  3.668076e-02 -1.283272e-02
#>  [42,] -0.254564292  2.432856e-01 -2.325066e-01  2.222052e-01 -2.123602e-01
#>  [43,]  1.607051393 -1.014377e+00  6.402786e-01 -4.041463e-01  2.550987e-01
#>  [44,]  1.626935377 -1.123935e+00  7.764471e-01 -5.363924e-01  3.705556e-01
#>  [45,] -1.319531754  1.079625e+00 -8.833366e-01  7.227355e-01 -5.913337e-01
#>  [46,] -0.349500478  3.088484e-01 -2.729248e-01  2.411796e-01 -2.131269e-01
#>  [47,]  0.796424172  7.267829e-01  6.632311e-01  6.052366e-01  5.523132e-01
#>  [48,] -0.054992273  5.490573e-02 -5.481932e-02  5.473305e-02 -5.464691e-02
#>  [49,] -1.363469477  1.054941e+00 -8.162275e-01  6.315302e-01 -4.886265e-01
#>  [50,] -0.867025435 -6.848215e-01 -5.409075e-01 -4.272367e-01 -3.374536e-01
#>  [51,]  0.310687954 -2.966812e-01  2.833059e-01 -2.705337e-01  2.583372e-01
#>  [52,]  0.600463674 -5.883307e-01  5.764430e-01 -5.647954e-01  5.533832e-01
#>  [53,] -0.125367545 -1.253056e-01 -1.252437e-01 -1.251818e-01 -1.251199e-01
#>  [54,] -0.735175517 -3.765632e-01 -1.928789e-01 -9.879421e-02 -5.060324e-02
#>  [55,] -0.104177288 -7.148200e-03 -4.904789e-04 -3.365456e-05 -2.309232e-06
#>  [56,]  1.705205464  7.410056e-01  3.220077e-01  1.399301e-01  6.080732e-02
#>  [57,] -0.336444869  3.026318e-01 -2.722170e-01  2.448589e-01 -2.202503e-01
#>  [58,] -1.043149897  2.446550e-01 -5.738012e-02  1.345764e-02 -3.156285e-03
#>  [59,] -0.134224691  3.215994e-02 -7.705450e-03  1.846209e-03 -4.423475e-04
#>  [60,]  0.627996146  5.960734e-01  5.657733e-01  5.370135e-01  5.097156e-01
#>  [61,]  0.143678467  4.600218e-03  1.472873e-04  4.715763e-06  1.509867e-07
#>  [62,] -0.564783318  1.272772e-01 -2.868266e-02  6.463806e-03 -1.456656e-03
#>  [63,]  1.126858084  7.849606e-01  5.467975e-01  3.808949e-01  2.653285e-01
#>  [64,]  0.280312299  2.776124e-01  2.749385e-01  2.722904e-01  2.696678e-01
#>  [65,]  0.622294803 -4.466627e-01  3.205997e-01 -2.301159e-01  1.651696e-01
#>  [66,] -0.151099991  1.491179e-01 -1.471618e-01  1.452314e-01 -1.433263e-01
#>  [67,]  0.804740307  2.917282e-01  1.057550e-01  3.833748e-02  1.389780e-02
#>  [68,] -0.005403215  5.799651e-05 -6.225174e-07  6.681918e-09 -7.172173e-11
#>  [69,] -0.293832590  1.497207e-01 -7.628928e-02  3.887275e-02 -1.980738e-02
#>  [70,] -0.867365320 -5.916297e-01 -4.035505e-01 -2.752617e-01 -1.877559e-01
#>  [71,] -0.440329967 -1.271089e-01 -3.669222e-02 -1.059185e-02 -3.057522e-03
#>  [72,] -0.236449010 -3.652288e-02 -5.641473e-03 -8.714048e-04 -1.346007e-04
#>  [73,]  0.112722579  3.266648e-02  9.466595e-03  2.743375e-03  7.950175e-04
#>  [74,] -0.578582813 -5.476750e-01 -5.184184e-01 -4.907246e-01 -4.645102e-01
#>  [75,] -1.332930334 -5.340170e-01 -2.139453e-01 -8.571374e-02 -3.433983e-02
#>  [76,] -0.686888943 -6.291383e-01 -5.762431e-01 -5.277950e-01 -4.834203e-01
#>  [77,]  1.424158320 -1.209671e+00  1.027487e+00 -8.727405e-01  7.413001e-01
#>  [78,]  0.540156162 -4.531810e-01  3.802105e-01 -3.189896e-01  2.676263e-01
#>  [79,]  0.606773832 -4.593011e-01  3.476707e-01 -2.631714e-01  1.992092e-01
#>  [80,] -0.716055748 -3.544527e-01 -1.754567e-01 -8.685231e-02 -4.299252e-02
#>  [81,]  0.342870015 -3.406480e-01  3.384405e-01 -3.362472e-01  3.340682e-01
#>  [82,] -1.405298874 -1.290872e+00 -1.185762e+00 -1.089211e+00 -1.000521e+00
#>  [83,] -0.065463892  8.010372e-03 -9.801748e-04  1.199373e-04 -1.467592e-05
#>  [84,]  0.401570026 -3.703504e-01  3.415579e-01 -3.150038e-01  2.905142e-01
#>  [85,] -0.621646815  6.067866e-01 -5.922815e-01  5.781233e-01 -5.643034e-01
#>  [86,] -0.247453561 -2.454394e-01 -2.434416e-01 -2.414601e-01 -2.394948e-01
#>  [87,] -0.276604044  2.753914e-01 -2.741840e-01  2.729819e-01 -2.717851e-01
#>  [88,]  0.061615259  5.477816e-02  4.869973e-02  4.329579e-02  3.849150e-02
#>  [89,]  0.189563870 -2.485876e-02  3.259892e-03 -4.274911e-04  5.605973e-05
#>  [90,] -0.140475474 -1.295789e-02 -1.195277e-03 -1.102560e-04 -1.017036e-05
#>  [91,]  1.603475935 -1.003755e+00  6.283379e-01 -3.933315e-01  2.462204e-01
#>  [92,]  1.974005385  1.077461e+00  5.881051e-01  3.210023e-01  1.752111e-01
#>  [93,] -1.429546737  1.360774e+00 -1.295309e+00  1.232994e+00 -1.173676e+00
#>  [94,] -0.871542922 -6.050049e-01 -4.199804e-01 -2.915406e-01 -2.023807e-01
#>  [95,] -0.861093221  3.029839e-01 -1.066078e-01  3.751098e-02 -1.319860e-02
#>  [96,] -0.315475389  2.899496e-01 -2.664892e-01  2.449271e-01 -2.251095e-01
#>  [97,] -0.008935976  1.579508e-04 -2.791912e-06  4.934936e-08 -8.722909e-10
#>  [98,]  1.977690289  1.614985e+00  1.318800e+00  1.076934e+00  8.794263e-01
#>  [99,]  0.573806780 -4.621358e-01  3.721975e-01 -2.997626e-01  2.414245e-01
#> [100,]  0.032915405  3.109713e-02  2.937930e-02  2.775636e-02  2.622308e-02
#>                 [,6]          [,7]          [,8]          [,9]        [,10]
#>   [1,]  1.598080e-01 -1.350581e-01  1.141412e-01 -9.646384e-02 -0.693405892
#>   [2,]  3.050298e-02 -1.481200e-02  7.192589e-03 -3.492664e-03 -1.292352540
#>   [3,]  9.812324e-01  9.033969e-01  8.317357e-01  7.657589e-01  3.800216717
#>   [4,] -1.520202e-01  1.518239e-01 -1.516280e-01  1.514322e-01 -3.012281730
#>   [5,] -1.235682e-02 -5.570785e-03 -2.511459e-03 -1.132233e-03  0.743354935
#>   [6,]  6.831290e-01 -6.293558e-01  5.798155e-01 -5.341748e-01 -2.646718754
#>   [7,]  6.957226e-03  6.875444e-03  6.794624e-03  6.714753e-03  0.048279977
#>   [8,]  7.189755e-03 -2.724524e-03  1.032446e-03 -3.912404e-04 -0.994243827
#>   [9,] -1.629928e-04 -3.547947e-05 -7.722996e-06 -1.681103e-06  0.341716577
#>  [10,]  9.246788e-11 -1.687240e-12  3.078667e-14 -5.617571e-16 -0.045722899
#>  [11,] -9.018853e-01 -8.160344e-01 -7.383556e-01 -6.680712e-01  3.492552747
#>  [12,] -2.939053e-02 -1.341629e-02 -6.124311e-03 -2.795646e-03  1.666570856
#>  [13,]  4.454206e-04  1.450900e-04  4.726121e-05  1.539473e-05  0.128466677
#>  [14,]  6.737777e-01 -6.128142e-01  5.573666e-01 -5.069360e-01 -2.604455769
#>  [15,]  2.450668e-01  2.431128e-01  2.411744e-01  2.392514e-01  2.023955571
#>  [16,]  2.270308e-01  2.266609e-01  2.262917e-01  2.259230e-01  4.011625712
#>  [17,] -1.946762e-01  1.274364e-01 -8.342081e-02  5.460787e-02 -2.142420802
#>  [18,] -5.930834e-01  5.777579e-01 -5.628284e-01  5.482847e-01 -2.993124397
#>  [19,]  8.083103e-01  6.817002e-01  5.749217e-01  4.848685e-01  3.525745340
#>  [20,] -3.372411e-06  3.880231e-07 -4.464519e-08  5.136789e-09 -0.168363854
#>  [21,]  6.785931e-02 -4.819131e-02  3.422378e-02 -2.430453e-02 -0.533603547
#>  [22,] -6.150474e-02 -3.667869e-02 -2.187354e-02 -1.304440e-02  1.015823098
#>  [23,]  2.904941e-01  2.871134e-01  2.837721e-01  2.804697e-01  2.024768357
#>  [24,]  6.844702e-01 -6.382722e-01  5.951924e-01 -5.550202e-01 -2.687837330
#>  [25,] -6.001068e-01  5.025246e-01 -4.208101e-01  3.523830e-01 -2.666322449
#>  [26,]  2.240458e-02 -1.307526e-02  7.630690e-03 -4.453253e-03 -0.407558603
#>  [27,]  1.398320e-01 -1.380508e-01  1.362923e-01 -1.345561e-01 -0.937048878
#>  [28,]  5.520956e-02 -2.974203e-02  1.602238e-02 -8.631443e-03 -1.444333768
#>  [29,] -4.135850e-01 -3.057110e-01 -2.259734e-01 -1.670335e-01  2.782827103
#>  [30,]  9.775458e-01  9.118333e-01  8.505381e-01  7.933632e-01  3.840599148
#>  [31,]  3.706775e-08  2.420457e-09  1.580515e-10  1.032048e-11  0.031290905
#>  [32,] -3.509977e-01  2.587589e-01 -1.907595e-01  1.406297e-01 -2.385715510
#>  [33,]  1.253668e-02  1.225228e-02  1.197433e-02  1.170269e-02  0.066389366
#>  [34,] -9.227304e-01 -8.581785e-01 -7.981424e-01 -7.423063e-01  3.608778766
#>  [35,] -1.497878e-01  1.480159e-01 -1.462650e-01  1.445348e-01 -1.036599670
#>  [36,]  1.582721e-01 -1.334477e-01  1.125170e-01 -9.486919e-02 -0.690799488
#>  [37,] -6.838468e-01  5.973766e-01 -5.218402e-01  4.558552e-01 -2.762009156
#>  [38,]  1.138832e-01 -6.319830e-02  3.507124e-02 -1.946242e-02 -2.601145659
#>  [39,] -2.659750e-01  1.851607e-01 -1.289011e-01  8.973555e-02 -2.265918987
#>  [40,] -3.085901e-06  3.498094e-07 -3.965345e-08  4.495008e-09 -0.165936893
#>  [41,]  4.489512e-03 -1.570650e-03  5.494902e-04 -1.922385e-04 -0.914423330
#>  [42,]  2.029513e-01 -1.939594e-01  1.853658e-01 -1.771530e-01 -0.864800822
#>  [43,] -1.610193e-01  1.016360e-01 -6.415302e-02  4.049363e-02 -2.071965149
#>  [44,] -2.559907e-01  1.768459e-01 -1.221703e-01  8.439884e-02 -2.250200818
#>  [45,]  4.838222e-01 -3.958576e-01  3.238860e-01 -2.649997e-01 -2.295035194
#>  [46,]  1.883371e-01 -1.664307e-01  1.470724e-01 -1.299657e-01 -0.746665735
#>  [47,]  5.040175e-01  4.599450e-01  4.197262e-01  3.830243e-01  1.947492984
#>  [48,]  5.456091e-02 -5.447505e-02  5.438932e-02 -5.430372e-02 -0.980593983
#>  [49,]  3.780593e-01 -2.925114e-01  2.263214e-01 -1.751091e-01 -2.152179176
#>  [50,] -2.665383e-01 -2.105258e-01 -1.662841e-01 -1.313398e-01  1.413710225
#>  [51,] -2.466906e-01  2.355690e-01 -2.249489e-01  2.148075e-01 -1.046535341
#>  [52,] -5.422016e-01  5.312459e-01 -5.205116e-01  5.099942e-01 -3.002184197
#>  [53,] -1.250580e-01 -1.249962e-01 -1.249345e-01 -1.248727e-01  3.988019959
#>  [54,] -2.591941e-02 -1.327614e-02 -6.800154e-03 -3.483097e-03  0.855988825
#>  [55,] -1.584496e-07 -1.087213e-08 -7.459992e-10 -5.118727e-11  0.104423397
#>  [56,]  2.642413e-02  1.148274e-02  4.989881e-03  2.168378e-03  1.893315266
#>  [57,]  1.981149e-01 -1.782041e-01  1.602944e-01 -1.441847e-01 -0.770032449
#>  [58,]  7.402588e-04 -1.736165e-04  4.071911e-05 -9.550051e-06 -1.073080567
#>  [59,]  1.059855e-04 -2.539389e-05  6.084319e-06 -1.457789e-06 -0.138251652
#>  [60,]  4.838054e-01  4.592122e-01  4.358692e-01  4.137128e-01  1.995081328
#>  [61,]  4.834209e-09  1.547791e-10  4.955631e-12  1.586667e-13  0.143752167
#>  [62,]  3.282660e-04 -7.397666e-05  1.667107e-05 -3.756923e-06 -0.579695090
#>  [63,]  1.848258e-01  1.287482e-01  8.968500e-02  6.247388e-02  1.570605689
#>  [64,]  2.670704e-01  2.644980e-01  2.619505e-01  2.594274e-01  2.024525356
#>  [65,] -1.185533e-01  8.509363e-02 -6.107740e-02  4.383934e-02 -0.893737830
#>  [66,]  1.414462e-01 -1.395908e-01  1.377597e-01 -1.359526e-01 -0.935945244
#>  [67,]  5.038123e-03  1.826381e-03  6.620853e-04  2.400140e-04  0.863474689
#>  [68,]  7.698399e-13 -8.263233e-15  8.869511e-17 -9.520270e-19 -0.005403526
#>  [69,]  1.009273e-02 -5.142693e-03  2.620429e-03 -1.335224e-03 -0.341489282
#>  [70,] -1.280683e-01 -8.735533e-02 -5.958505e-02 -4.064295e-02  1.186125793
#>  [71,] -8.826073e-04 -2.547800e-04 -7.354670e-05 -2.123054e-05  0.459908542
#>  [72,] -2.079098e-05 -3.211459e-06 -4.960551e-07 -7.662269e-08  0.239321246
#>  [73,]  2.303924e-04  6.676665e-05  1.934866e-05  5.607153e-06  0.117776532
#>  [74,] -4.396961e-01 -4.162077e-01 -3.939739e-01 -3.729279e-01  1.794230861
#>  [75,] -1.375770e-02 -5.511800e-03 -2.208214e-03 -8.846853e-04  1.454785762
#>  [76,] -4.427764e-01 -4.055497e-01 -3.714528e-01 -3.402227e-01  1.711441932
#>  [77,] -6.296555e-01  5.348253e-01 -4.542771e-01  3.858601e-01 -2.698496009
#>  [78,] -2.245335e-01  1.883795e-01 -1.580469e-01  1.325984e-01 -0.992642238
#>  [79,] -1.507926e-01  1.141433e-01 -8.640147e-02  6.540211e-02 -0.928547065
#>  [80,] -2.128161e-02 -1.053455e-02 -5.214677e-03 -2.581303e-03  0.824105092
#>  [81,] -3.319032e-01  3.297523e-01 -3.276154e-01  3.254923e-01 -3.016581560
#>  [82,] -9.190534e-01 -8.442191e-01 -7.754782e-01 -7.123345e-01  3.555488645
#>  [83,]  1.795792e-06 -2.197389e-07  2.688795e-08 -3.290096e-09 -0.065959553
#>  [84,] -2.679285e-01  2.470987e-01 -2.278882e-01  2.101713e-01 -1.038777151
#>  [85,]  5.508139e-01 -5.376469e-01  5.247947e-01 -5.122496e-01 -2.860214200
#>  [86,] -2.375454e-01 -2.356119e-01 -2.336941e-01 -2.317920e-01  1.943411182
#>  [87,]  2.705935e-01 -2.694072e-01  2.682261e-01 -2.670501e-01 -2.957160779
#>  [88,]  3.422032e-02  3.042308e-02  2.704720e-02  2.404592e-02  0.134578708
#>  [89,] -7.351481e-06  9.640480e-07 -1.264220e-07  1.657854e-08 -0.191215145
#>  [90,] -9.381455e-07 -8.653746e-08 -7.982484e-09 -7.363292e-10  0.141076954
#>  [91,] -1.541308e-01  9.648392e-02 -6.039769e-02  3.780818e-02 -2.056177625
#>  [92,]  9.563454e-02  5.219971e-02  2.849190e-02  1.555158e-02  2.355898093
#>  [93,]  1.117213e+00 -1.063465e+00  1.012304e+00 -9.636035e-01 -4.665091201
#>  [94,] -1.404880e-01 -9.752352e-02 -6.769857e-02 -4.699477e-02  1.210804997
#>  [95,]  4.644053e-03 -1.634055e-03  5.749578e-04 -2.023044e-04 -0.919919506
#>  [96,]  2.068955e-01 -1.901551e-01  1.747693e-01 -1.606284e-01 -0.800592627
#>  [97,]  1.541847e-11 -2.725342e-13  4.817268e-15 -8.514922e-17 -0.008937372
#>  [98,]  7.181410e-01  5.864352e-01  4.788840e-01  3.910575e-01  3.426342668
#>  [99,] -1.944398e-01  1.565990e-01 -1.261226e-01  1.015773e-01 -0.968039940
#> [100,]  2.477449e-02  2.340593e-02  2.211297e-02  2.089143e-02  0.100423687
#>                [,11]         [,12]        [,13]         [,14]
#>   [1,]  0.0325177282 -0.0496084617 -0.024493821  1.087861e-03
#>   [2,]  0.0197357897  0.0496548860 -0.003834600 -1.753577e-02
#>   [3,] -0.0446901461 -0.0211949528 -0.015610410 -4.892647e-03
#>   [4,]  0.0379975412  0.0109665661 -0.010459521  2.946351e-02
#>   [5,]  0.0127911499 -0.0300021373  0.045321529 -1.338794e-03
#>   [6,] -0.0205977334  0.0489934588  0.032335088 -4.860348e-02
#>   [7,]  0.0084883154  0.0492364852 -0.019114968 -4.768682e-02
#>   [8,]  0.0332598499 -0.0092251071 -0.037005779  4.911674e-03
#>   [9,] -0.0244772696  0.0480642099  0.013104856  1.478762e-03
#>  [10,] -0.0238165887 -0.0427481743  0.046152762  3.396475e-02
#>  [11,] -0.0085745390 -0.0376417662 -0.017530579  2.327500e-02
#>  [12,]  0.0342196446  0.0114574986 -0.027069363  1.529034e-02
#>  [13,]  0.0427801586  0.0050550625  0.042048144  4.116070e-02
#>  [14,] -0.0182862248 -0.0055912322 -0.011510796 -4.600683e-03
#>  [15,]  0.0262164100 -0.0427761449  0.016275571  2.713254e-02
#>  [16,]  0.0262175613  0.0116035474  0.010987490 -3.511799e-03
#>  [17,] -0.0049770458  0.0294428627  0.042168650  4.281406e-02
#>  [18,]  0.0173774109  0.0194357511 -0.021161921  3.955540e-02
#>  [19,] -0.0053953909 -0.0402406576  0.038252271  2.472493e-02
#>  [20,] -0.0333608034 -0.0378762641 -0.001165844 -2.197829e-02
#>  [21,] -0.0237606657  0.0419320840  0.009607756  2.703070e-02
#>  [22,]  0.0489924714 -0.0187297135 -0.024194075 -2.308379e-02
#>  [23,]  0.0281618486  0.0258424798  0.011643830  5.258368e-03
#>  [24,]  0.0383638383 -0.0261992917 -0.035887038  2.934334e-03
#>  [25,] -0.0133779410  0.0371852255  0.037721450 -3.080420e-02
#>  [26,]  0.0091026251 -0.0488955505 -0.015316961  7.382420e-03
#>  [27,]  0.0367521813 -0.0194424135 -0.021548661 -9.274753e-03
#>  [28,]  0.0468709687  0.0464052612  0.022111165 -1.920643e-02
#>  [29,] -0.0197907360  0.0112505255 -0.032859508 -4.270327e-02
#>  [30,] -0.0257531618 -0.0060758382 -0.026268652 -3.467174e-02
#>  [31,] -0.0453562912 -0.0180065034  0.013259805  2.250632e-03
#>  [32,] -0.0188870539 -0.0158157140 -0.021591873  1.559349e-03
#>  [33,]  0.0397880210 -0.0222500099 -0.044628230  1.855286e-02
#>  [34,]  0.0399769354  0.0425925740 -0.041915549 -2.388706e-02
#>  [35,] -0.0452126691  0.0429048963  0.033751151  2.330132e-02
#>  [36,] -0.0102267198 -0.0411874940 -0.023203573 -6.335758e-03
#>  [37,] -0.0162869531  0.0334751221 -0.041832917  4.982678e-02
#>  [38,]  0.0433820173  0.0243977297  0.027931012 -8.388311e-03
#>  [39,]  0.0037209368  0.0120835297  0.030969037 -2.136921e-02
#>  [40,]  0.0235919707  0.0026838267 -0.047234853 -2.955651e-02
#>  [41,] -0.0139762304 -0.0325238462 -0.022037550  1.745432e-02
#>  [42,] -0.0309946286  0.0318230896  0.038791603  1.949603e-02
#>  [43,]  0.0371431192 -0.0314446948 -0.027127304 -7.166690e-03
#>  [44,]  0.0474618278  0.0282636277 -0.001508292 -1.644968e-02
#>  [45,] -0.0424495738  0.0002471352  0.027183734  3.474937e-02
#>  [46,]  0.0189537697  0.0138222694 -0.004212032 -1.836512e-02
#>  [47,] -0.0107207891 -0.0360020735  0.022567658  3.143694e-02
#>  [48,] -0.0442000644  0.0480195849 -0.026007255  2.040336e-02
#>  [49,] -0.0345513066 -0.0059361625  0.024121708  6.414252e-03
#>  [50,] -0.0392247889  0.0224910531 -0.041868537  2.147733e-03
#>  [51,]  0.0403874907 -0.0407675546  0.049753901 -3.569791e-02
#>  [52,]  0.0232506018 -0.0494655325 -0.002271111 -2.269987e-02
#>  [53,]  0.0390127034 -0.0327479591 -0.028198454  1.297259e-02
#>  [54,] -0.0400439808 -0.0495735137  0.014419209  1.150707e-03
#>  [55,] -0.0030734069  0.0073812152 -0.037419659  2.533292e-02
#>  [56,] -0.0001744899  0.0335164269 -0.025770980 -1.741508e-02
#>  [57,]  0.0457450442  0.0204585960 -0.013745672 -3.000111e-05
#>  [58,] -0.0044703108 -0.0361835044 -0.046936385  3.455432e-02
#>  [59,]  0.0459997307  0.0481239486 -0.040933655 -3.752503e-02
#>  [60,] -0.0333375825 -0.0272181874  0.010839902  3.273565e-02
#>  [61,]  0.0451952318  0.0470889839  0.031363228  3.209712e-02
#>  [62,] -0.0308251501  0.0487875284 -0.037055348  1.168393e-02
#>  [63,] -0.0130001391  0.0346080288 -0.019137480  4.180435e-02
#>  [64,]  0.0309405150  0.0022201889  0.034460918  3.993054e-02
#>  [65,]  0.0001373387 -0.0488258399 -0.035508173 -3.756179e-02
#>  [66,]  0.0470024749 -0.0357975025  0.009481941 -2.318035e-02
#>  [67,] -0.0023523403 -0.0093897674 -0.047191443  4.834366e-02
#>  [68,]  0.0003944058 -0.0097790468 -0.023680777  4.321801e-04
#>  [69,]  0.0056659033 -0.0123106665 -0.019174414  3.236065e-02
#>  [70,]  0.0441811061 -0.0388258697  0.033468590 -3.741877e-02
#>  [71,] -0.0109912715 -0.0219804815  0.024846924  3.917383e-02
#>  [72,] -0.0166291791 -0.0255494278  0.001683579 -2.365702e-03
#>  [73,]  0.0350038873  0.0090857357  0.042461604  1.208755e-02
#>  [74,]  0.0409875165 -0.0295774156 -0.049043166  4.913119e-02
#>  [75,]  0.0259902544  0.0154004392  0.035486608 -4.034390e-02
#>  [76,] -0.0257139110 -0.0079552544 -0.037437545  4.821577e-02
#>  [77,]  0.0200969296  0.0258957725  0.041226851  3.810141e-02
#>  [78,]  0.0436321078 -0.0187858912 -0.017548987  2.662370e-03
#>  [79,] -0.0235420116  0.0441380982 -0.025236682 -1.508174e-02
#>  [80,] -0.0308799078  0.0314820318 -0.020403447  4.110233e-02
#>  [81,] -0.0292940410  0.0316921706 -0.038737185 -3.632254e-02
#>  [82,]  0.0087563645 -0.0023766425  0.001785805 -1.393530e-03
#>  [83,]  0.0306671617 -0.0440690644  0.047150404 -3.262982e-02
#>  [84,] -0.0031051864  0.0219186763 -0.033156661 -2.227799e-02
#>  [85,] -0.0069005878 -0.0155982371  0.043388811 -4.340054e-02
#>  [86,] -0.0188523525 -0.0497527860  0.049813196 -4.989288e-02
#>  [87,] -0.0331129279 -0.0008652059 -0.017658273  1.101284e-02
#>  [88,]  0.0402998675  0.0027299128 -0.036755275 -4.740332e-02
#>  [89,] -0.0402387734 -0.0158468477  0.021892301 -4.694237e-02
#>  [90,] -0.0388428311  0.0257605160  0.008420411 -3.467080e-02
#>  [91,] -0.0099897678  0.0313208758 -0.008245978  2.884831e-02
#>  [92,] -0.0407963422 -0.0123781507  0.032968349 -2.572924e-02
#>  [93,]  0.0263680470  0.0143306518  0.013553382 -3.306112e-02
#>  [94,] -0.0189720985 -0.0274483128  0.030463290 -3.407417e-02
#>  [95,]  0.0169448607 -0.0090467601  0.036004622  1.421543e-02
#>  [96,]  0.0339614471 -0.0117790645 -0.001930027 -2.473612e-02
#>  [97,]  0.0490527796 -0.0473529263 -0.029301863  3.000547e-02
#>  [98,] -0.0325274907 -0.0189564774 -0.031035995  4.743072e-02
#>  [99,]  0.0397372108 -0.0199127099 -0.019931361 -4.772346e-02
#> [100,] -0.0388201939 -0.0290106127 -0.024770912  4.928346e-02