Generate three clusters of data points with optional noise.

This function generates three clusters of data points along with optional noise.

Usage

three_clust_diff_dist(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate, should be a multiple of three.

num_dims

Number of dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

three_clust_diff_dist(n = 100, num_dims = 7,
num_noise = 4, min_n = -0.05, max_n = 0.05)
#> Warning: The sample size should be a product of three.
#>               [,1]          [,2]         [,3]         [,4]        [,5]
#>   [1,]  0.05107849 -0.5706557332  1.396191281  0.228640944 -0.36516444
#>   [2,]  0.73455219  1.4340305542  0.019001739 -0.242138761  2.40914137
#>   [3,]  0.22746899 -0.0100174062  0.437678220  0.385072631 -0.15664804
#>   [4,]  1.16737949 -0.2627896150 -1.252312528 -1.826647001 -0.68382145
#>   [5,] -1.56484014  0.5268513787  0.227515519  0.202170846 -0.53256965
#>   [6,] -3.11992261  1.3121561986  0.796731381 -0.696936026 -0.14958814
#>   [7,]  0.36680110 -1.6192418695  0.601094153 -0.891340142 -0.16439611
#>   [8,]  1.38374811  0.4603896981 -0.190181641  0.659691548 -0.25074334
#>   [9,] -0.86925701 -0.3047490332  0.279981490 -1.860999434 -0.60231527
#>  [10,]  0.26648936 -0.6345489601 -0.349599991 -0.808712625  0.06146945
#>  [11,] -2.56874426  1.1275680784  0.950721218 -0.338629252  1.32486985
#>  [12,]  0.03974415  0.9057287153 -0.311088476 -0.028608597  0.46740420
#>  [13,] -0.78204666 -0.5519734859  0.042493298 -0.460202921  0.18094198
#>  [14,]  0.64713929  0.9279144873  0.921157185  0.553627349  0.32013258
#>  [15,]  0.31285798 -0.7929113843  1.273291988 -0.374823343  2.22483319
#>  [16,]  2.52743363 -1.7107729050 -0.269958687  0.581518327 -1.50798112
#>  [17,]  0.03024327 -1.4307619724 -1.688283856  1.197431765  0.48667137
#>  [18,] -0.35556503 -0.0374989464 -0.713859789 -0.520965558 -0.41141535
#>  [19,] -1.11465954 -1.2042251877 -0.105784026 -0.898587890 -0.04955100
#>  [20,]  0.21948983  0.3627661975  1.331266163  0.241294805  1.88477715
#>  [21,] -0.62260214  0.3187188493  1.134008988 -0.773863234 -0.98518197
#>  [22,] -0.32564603 -1.0865313350 -1.082155439 -0.125061857 -0.71928671
#>  [23,] -1.39434684  0.0364855056  0.404298698 -1.151861896  0.60880759
#>  [24,] -0.15738354  0.4415472816 -0.942844984 -1.029627842 -0.05751606
#>  [25,]  0.45850984  0.8998146918 -0.565650623  1.017885321  0.16496128
#>  [26,]  0.28502520 -0.6461940464  0.928922978  1.669493052 -0.42713443
#>  [27,]  0.65576994  0.1628785145 -1.564109940 -0.514460180 -0.68127759
#>  [28,]  0.44584016  0.6612799631 -0.059422893  0.959898501 -1.51316254
#>  [29,]  0.24806878  1.9214425000  0.267780331  0.455814625 -0.53671075
#>  [30,]  0.45843516 -1.9586238131 -2.449721789 -2.413612221  2.26226562
#>  [31,]  0.42650933 -0.0288143897 -0.632147334 -0.233429808 -1.35368174
#>  [32,]  1.40857213 -0.9070640926  0.574742364  0.625826742  0.86749805
#>  [33,]  1.22172918  0.6638327318 -1.434906645 -0.033875771 -0.52721492
#>  [34,]  9.54054776 -0.5140016077 -1.035802714  1.118971969 -0.03036338
#>  [35,] 10.64816889 -1.8536677395 -1.420155767 -1.101960329  0.08651387
#>  [36,]  9.71108576 -0.0752425710  0.752244209  0.438480070  0.76288879
#>  [37,] 10.83894958 -0.9888573710  2.018235222 -1.076441601 -1.11957443
#>  [38,] 10.72748480 -0.0629662814  1.004063481 -0.355402910  1.61670453
#>  [39,]  7.37446366  1.6697624504 -0.688569470  2.864040321  0.95876492
#>  [40,]  9.12872131 -0.2312958255  0.343958647  1.229204774  1.45434102
#>  [41,]  9.31774427  0.0484231725  0.310933692 -1.325316103 -1.36276315
#>  [42,]  9.85722006 -0.1471625077 -1.616915588 -0.256840953 -2.37768682
#>  [43,]  8.94797156 -0.3229538136 -0.251729971  0.121196840 -0.11011205
#>  [44,] 11.62165778  0.0665901625  0.517368945  1.146630359  0.01119210
#>  [45,] 11.18044549  0.1902591603  1.134419194 -0.195356312  1.23330535
#>  [46,] 10.69114397 -0.2350392691 -1.105214753  0.031313562 -0.91613889
#>  [47,]  8.40667699 -1.5477641116  1.222321874  0.269494401 -0.82794679
#>  [48,] 10.35351496 -0.2083532883 -2.801881530 -0.023687262 -0.66632982
#>  [49,]  9.50798610  0.1971703852  0.400664214 -0.100413526 -0.65055806
#>  [50,]  9.90514225  0.5051068989 -0.875249238 -0.469652526  1.01472708
#>  [51,]  9.46535416  1.2761162394 -0.268876434  0.004375355  0.43895125
#>  [52,]  9.96578280 -0.8620180597  0.505270117  1.190676088 -0.59370663
#>  [53,] 10.06161670 -1.1498311441 -0.562046509 -0.947958382 -0.17859584
#>  [54,] 10.59387475 -0.5996713975 -1.303831220  0.790695095  1.09644878
#>  [55,]  9.80583655 -0.8765904577 -0.874671914  2.217933480 -1.81568691
#>  [56,] 10.09711432 -0.2132653944 -1.293569786 -0.364471510 -1.31752823
#>  [57,] 11.82930949 -0.5161271141 -0.920535302 -0.434849186 -1.42180073
#>  [58,] 10.09163989  0.2655730104 -0.391505808  0.644590047 -0.76976460
#>  [59,] 10.80127523 -0.0371743201  1.123819037  0.849385622  0.10908149
#>  [60,] 10.40722059  1.2965055428 -0.452555332  0.100494162 -1.01402001
#>  [61,] 10.70247387 -0.5489671988 -0.155798025 -2.009019680  0.57458258
#>  [62,] 10.98853830 -0.6670030480 -1.867647385  1.218453711  0.17102762
#>  [63,] 11.26551971  1.1525433321 -1.999527523  0.722486972  1.74742188
#>  [64,] 11.15365040  0.5842582941 -0.580639544  0.524095414  1.08643996
#>  [65,] 10.27291495  1.1928090203  0.093359986 -1.199872246 -0.03054596
#>  [66,] 10.61728378 -0.6529439775  0.426219292  0.192777563  1.85883566
#>  [67,] 49.28207182  0.4504953134  0.009934555 -0.106381191  1.37121747
#>  [68,] 49.37225233  0.9702971984  0.693044107  1.479941237  0.33758835
#>  [69,] 49.51357551  0.6022108594 -0.146660900  0.795330302 -0.83521701
#>  [70,] 50.01280125 -0.6728577761  0.719531132  0.898887817 -1.30452105
#>  [71,] 49.89393222 -1.5507049826 -1.030219631 -0.037964995 -0.34638194
#>  [72,] 50.57741898 -0.3711134132  0.581229937 -0.865326868  0.10235545
#>  [73,] 49.17930247  0.6969099052  0.278146356  1.057009505 -1.32298124
#>  [74,] 49.82509911 -0.0573561308 -0.829811122  2.029137492  2.32951820
#>  [75,] 50.94567245  0.3382639570  0.289659457  0.282286140  0.66982941
#>  [76,] 49.95156094  0.1870138567  0.354525326  0.031825299  1.50822551
#>  [77,] 50.42753029 -2.2518225402  0.274679180  1.061164983  1.25006966
#>  [78,] 50.63452841 -0.0628297144  1.144196190  0.840234156  0.23060131
#>  [79,] 50.55433390 -0.1513111863 -1.737042064 -0.172204265 -0.04039720
#>  [80,] 48.24581518 -0.0135371692  0.643852529  0.652269731 -0.08971669
#>  [81,] 51.10729665  0.4121828015 -1.010911480  1.311825990  0.89843734
#>  [82,] 51.81418905 -1.0461893924 -0.700743487 -1.338875531  0.27817987
#>  [83,] 49.62364640  0.7900417955  1.294615828  0.890417786 -0.63357194
#>  [84,] 49.61934518  0.8165088059 -1.513877834  0.682012332 -1.23826488
#>  [85,] 51.55771645 -0.6321010820 -0.059744658 -0.011679090 -0.42136764
#>  [86,] 49.28092939 -3.0976734891 -0.663489034 -1.548446049 -1.34094102
#>  [87,] 51.55843951 -0.2629056325 -0.109329700 -1.171022455 -0.45097641
#>  [88,] 49.02672595 -1.3169177856 -0.397935554  0.640718397 -1.25701098
#>  [89,] 49.07233941  0.0632903851 -0.672246611  0.507290386  0.33623342
#>  [90,] 48.44751874  1.9066136305 -0.177187734  0.083111034  1.88768722
#>  [91,] 49.53241226  0.0244000211 -0.347748958 -0.706524851  1.30063965
#>  [92,] 50.67144372  1.6054124616  1.015111973 -0.568101602 -0.77517059
#>  [93,] 50.91358736  0.5114802183 -0.095445816  0.031723694 -0.39460640
#>  [94,] 51.85756942 -1.1470161135 -1.408402647  2.214408092 -0.20617242
#>  [95,] 50.40017543 -1.3939437245  0.334142356 -0.330512174 -1.22940517
#>  [96,] 48.82270512 -0.0909434690  1.894306158  0.679379352 -1.57842354
#>  [97,] 50.14211799 -0.0001649257  1.537558612 -1.068328170 -0.33369556
#>  [98,] 48.53088007  0.8381700319 -1.542459824  0.081109833  1.38567916
#>  [99,] 49.60422382 -1.7330640192  0.376138559  0.715505293 -0.69578136
#>               [,6]         [,7]          [,8]          [,9]         [,10]
#>   [1,]  1.71302841  1.551287651  1.007984e-02  0.0062276280  0.0490855206
#>   [2,]  0.87558398  0.812695011 -6.041592e-05  0.0296270992 -0.0109389361
#>   [3,]  1.26879400 -0.943698765 -4.436867e-02 -0.0209475275 -0.0110115769
#>   [4,] -0.34965537 -0.551409086  1.565973e-02  0.0472355036  0.0281032638
#>   [5,] -0.96729441 -0.477038487  4.260539e-02  0.0364343969  0.0138222174
#>   [6,] -0.55854451  1.343172979  3.960312e-02  0.0491180022 -0.0295764864
#>   [7,] -2.06289285 -0.169662062  3.987240e-02 -0.0451530931 -0.0080452825
#>   [8,]  0.56089102  0.662825551 -3.429926e-02  0.0106188636  0.0012326364
#>   [9,] -1.14627068 -0.217949341  4.203202e-02  0.0143729872  0.0278222616
#>  [10,]  0.04075364  1.674504559  3.889307e-02  0.0186364058 -0.0294928105
#>  [11,]  1.73095366 -0.825575959 -2.640544e-02  0.0151786698  0.0202620683
#>  [12,] -0.37974406 -0.027666806  4.022155e-03 -0.0171782329 -0.0316898281
#>  [13,]  1.00549826  1.380030576 -2.381379e-02  0.0472077122  0.0066542658
#>  [14,] -1.76226856  0.837891264 -4.964506e-02  0.0358185788 -0.0289293902
#>  [15,]  0.99258491 -1.904879134 -4.347735e-02  0.0281958961 -0.0291462888
#>  [16,]  1.15040181  0.064440987 -2.938788e-02  0.0041599938 -0.0178157664
#>  [17,] -2.98259662 -0.596971183 -2.724562e-02 -0.0383295933 -0.0052497577
#>  [18,]  1.05159302  0.265316901 -4.381072e-02  0.0065389430 -0.0366778420
#>  [19,] -1.06744924 -1.648463796 -6.723346e-03 -0.0434729038 -0.0163714078
#>  [20,]  0.93514459 -0.437055588 -3.317229e-02  0.0193679055  0.0372820590
#>  [21,] -0.25502571 -0.401294812 -3.199412e-03 -0.0172817407  0.0119357234
#>  [22,]  1.08177121 -0.023919249 -1.886010e-02  0.0327387885 -0.0005308995
#>  [23,]  0.93313074  0.340015080 -2.637327e-02  0.0235400109 -0.0331086728
#>  [24,] -0.85020865 -0.013901193  2.428745e-02  0.0189952244 -0.0368468671
#>  [25,] -0.75990220 -0.552491366 -1.718655e-02 -0.0424863736  0.0090609204
#>  [26,] -0.66123264  0.915412890  2.833872e-02  0.0388691817 -0.0448211082
#>  [27,]  1.98024524  0.412776501  2.263685e-02 -0.0074849471 -0.0474256098
#>  [28,] -1.39612652  0.527264453 -6.554542e-03 -0.0357184467 -0.0374919186
#>  [29,]  1.25257444  0.434619480 -9.295457e-03 -0.0426097581  0.0079640294
#>  [30,] -0.73810484  1.169261664 -4.398744e-02  0.0279327920 -0.0202587278
#>  [31,]  0.16219195  0.082537487 -1.609508e-02 -0.0381956547 -0.0019372769
#>  [32,] -1.13991433  0.701883612 -5.396595e-03 -0.0067037379 -0.0276348636
#>  [33,]  0.15048207  1.077753856 -2.016675e-02 -0.0424119496  0.0467467575
#>  [34,]  1.13719716  0.440171987  4.374711e-03 -0.0224602098 -0.0031629815
#>  [35,]  0.52045932  0.450928720 -1.683601e-02 -0.0055907377  0.0462950075
#>  [36,] -0.90670967  0.712366079 -4.694488e-02  0.0103502323  0.0282721765
#>  [37,] -0.02453979  1.510823247 -4.654395e-02  0.0499759899  0.0067053967
#>  [38,] -1.80269128  0.604610236 -4.076701e-02 -0.0210451708  0.0224892192
#>  [39,]  1.58261411 -0.484976451 -3.593369e-02  0.0321713887 -0.0142980804
#>  [40,]  0.14672731 -1.468175694  3.697711e-02 -0.0325611367  0.0340712622
#>  [41,] -0.25322873 -0.393621477  4.349588e-02  0.0472731355 -0.0217796690
#>  [42,]  0.22375696 -1.598178090 -3.586282e-02  0.0330438737  0.0364528202
#>  [43,]  0.59933229  0.604311447 -4.422297e-02 -0.0185132847 -0.0489602479
#>  [44,] -0.46711523  0.184453515  2.622776e-02 -0.0381776393  0.0261356550
#>  [45,]  0.45093984 -0.472365204 -4.435818e-02  0.0491751954 -0.0400020068
#>  [46,] -0.95605313 -0.431793725  9.467699e-03  0.0102211854 -0.0428770204
#>  [47,]  1.70541461  1.060064872 -5.869799e-03  0.0331363229 -0.0431613044
#>  [48,] -0.70285642 -0.002791857  2.560887e-02 -0.0175255578 -0.0126416084
#>  [49,]  1.07658457 -1.402690653 -3.007572e-02  0.0285910801  0.0056867178
#>  [50,]  2.11194282  1.027134359  8.017194e-04  0.0057797917 -0.0230753120
#>  [51,]  1.43075690 -1.569572024 -1.712494e-02  0.0382884510  0.0313909671
#>  [52,] -0.49384819 -0.578353485  4.308434e-02  0.0130134203 -0.0400511595
#>  [53,]  0.53478319  0.359667435 -3.460149e-02  0.0297558629  0.0151861725
#>  [54,]  1.40624775  0.635291835  4.076831e-02 -0.0078153997 -0.0163791305
#>  [55,] -0.08901359 -2.078772925  4.389501e-02  0.0426822144  0.0349516973
#>  [56,] -0.60745861 -0.263327377 -3.763700e-02 -0.0190857294  0.0403903222
#>  [57,] -0.14535292 -1.879535863  5.563707e-03 -0.0187529018  0.0418248882
#>  [58,]  1.18818880 -0.390987016  4.143465e-02  0.0142241365 -0.0264540270
#>  [59,]  0.42845957 -0.471196588  1.116855e-02  0.0013214896 -0.0280683651
#>  [60,] -0.80716359  1.730018224  2.547460e-02 -0.0391262494  0.0244652498
#>  [61,] -0.03928846 -0.322512192 -3.627714e-03 -0.0185954571 -0.0437981037
#>  [62,] -0.36193380 -1.078293493 -1.507992e-02 -0.0428051682  0.0252509293
#>  [63,]  1.36396444 -0.732702402 -4.811062e-02 -0.0215358689  0.0372666656
#>  [64,]  0.23792341  1.092130546  2.066739e-02 -0.0021811160 -0.0113749123
#>  [65,]  0.29963286 -1.287555555  4.474322e-02 -0.0268005346 -0.0461521584
#>  [66,] -1.00816079  1.538550280  2.282539e-02  0.0062437585  0.0107927013
#>  [67,] -1.30297646 -0.047063822 -3.976840e-02 -0.0067620835 -0.0403963021
#>  [68,]  0.05837194  1.400633424 -6.033985e-03 -0.0020524673  0.0196673589
#>  [69,]  0.40751209 -0.915087651 -3.708191e-02  0.0130068716 -0.0245487140
#>  [70,] -0.40804545 -0.773099598 -2.951489e-02 -0.0262611304 -0.0155798627
#>  [71,]  1.37039643 -0.579745671 -2.577447e-02  0.0094330351  0.0032883701
#>  [72,]  0.33087126  0.114169500 -4.805397e-02  0.0306781298  0.0374134157
#>  [73,] -0.47978990  0.102115258  1.106878e-02  0.0336195380  0.0450645224
#>  [74,] -0.02443725  0.437223043  4.040292e-02 -0.0486617747  0.0296831018
#>  [75,] -0.21226143  0.316136695 -3.119832e-02 -0.0271309227  0.0023194010
#>  [76,] -0.65748257  0.334182089  3.937017e-02 -0.0491564754  0.0357904800
#>  [77,]  0.45123094 -0.820789301 -2.372306e-02 -0.0329873063  0.0387855230
#>  [78,] -0.35599343  0.659837242  4.285438e-02  0.0492859304  0.0211234554
#>  [79,]  1.31707856  1.203682902 -2.448120e-02  0.0235203777 -0.0323327085
#>  [80,] -0.92802717 -1.253812906  1.603744e-02 -0.0399727734  0.0496954251
#>  [81,]  0.30832663 -0.419371610  2.753200e-02 -0.0162448523 -0.0291271123
#>  [82,] -0.54853064 -1.454105681 -2.130745e-02 -0.0202701771 -0.0383937900
#>  [83,]  2.37123574  1.105432509 -1.331340e-02  0.0419310278 -0.0302491587
#>  [84,]  1.25752063  1.677583287 -3.272420e-02 -0.0273668953  0.0362525303
#>  [85,] -0.60553635 -1.368500922 -2.113485e-02  0.0418299969 -0.0361935268
#>  [86,]  0.85433109 -0.775372288  4.282494e-02 -0.0442341005 -0.0251071536
#>  [87,]  1.53339163  1.434164315  3.090917e-02  0.0007294259 -0.0237281300
#>  [88,]  2.08390150  0.405021051 -2.202322e-03 -0.0435792620 -0.0227968572
#>  [89,]  0.63604509 -0.858336054 -3.725510e-02 -0.0012864698 -0.0150764172
#>  [90,]  0.24648813  1.467381898 -3.831168e-02  0.0229342554 -0.0160111140
#>  [91,]  1.70672703  0.138611076 -1.476826e-02  0.0447475741  0.0431722310
#>  [92,]  1.27880870 -1.075195394  4.722244e-03 -0.0419997997 -0.0452965989
#>  [93,] -2.72860715 -1.181220828 -1.978069e-02 -0.0174775439 -0.0010931596
#>  [94,] -0.06402423  0.377058200 -4.948597e-02 -0.0231240467  0.0428605371
#>  [95,]  2.59154470 -0.097335154 -2.868999e-02  0.0217224850 -0.0482633718
#>  [96,] -0.66395796  0.003804074  1.823779e-02 -0.0221181043  0.0384968223
#>  [97,] -1.45863511 -0.468343116  3.737257e-02  0.0213373670 -0.0333568781
#>  [98,]  0.49526027  0.492041254 -8.322239e-03  0.0465939296  0.0076114087
#>  [99,] -2.38654228 -0.765528580 -1.702219e-02  0.0340184010 -0.0395966804
#>                [,11]
#>   [1,] -0.0457726171
#>   [2,] -0.0415273239
#>   [3,] -0.0378590331
#>   [4,] -0.0064620207
#>   [5,] -0.0357464936
#>   [6,]  0.0278771799
#>   [7,] -0.0297492649
#>   [8,] -0.0093140276
#>   [9,]  0.0484812086
#>  [10,]  0.0282070826
#>  [11,]  0.0423391420
#>  [12,]  0.0369311152
#>  [13,] -0.0116991268
#>  [14,]  0.0086207635
#>  [15,] -0.0110272372
#>  [16,] -0.0363714692
#>  [17,]  0.0464597169
#>  [18,] -0.0278355452
#>  [19,] -0.0302170455
#>  [20,]  0.0001756774
#>  [21,]  0.0020222628
#>  [22,] -0.0240445717
#>  [23,]  0.0368531077
#>  [24,]  0.0015613105
#>  [25,]  0.0171813812
#>  [26,] -0.0483092913
#>  [27,] -0.0175477125
#>  [28,] -0.0127068096
#>  [29,] -0.0209591254
#>  [30,] -0.0298883930
#>  [31,]  0.0479613789
#>  [32,] -0.0031947945
#>  [33,] -0.0045289649
#>  [34,]  0.0047111596
#>  [35,]  0.0422574815
#>  [36,]  0.0235695945
#>  [37,]  0.0280530693
#>  [38,]  0.0149982235
#>  [39,] -0.0081552007
#>  [40,] -0.0369947617
#>  [41,] -0.0113761966
#>  [42,]  0.0143444609
#>  [43,] -0.0079806234
#>  [44,] -0.0488819213
#>  [45,] -0.0469625079
#>  [46,]  0.0241322211
#>  [47,]  0.0001955472
#>  [48,]  0.0130424777
#>  [49,] -0.0286628637
#>  [50,]  0.0206840465
#>  [51,]  0.0123296750
#>  [52,] -0.0037384278
#>  [53,] -0.0268606083
#>  [54,] -0.0393400292
#>  [55,]  0.0187927836
#>  [56,] -0.0384120806
#>  [57,]  0.0220535899
#>  [58,] -0.0068160862
#>  [59,]  0.0174658340
#>  [60,]  0.0193081265
#>  [61,]  0.0468018752
#>  [62,]  0.0181909968
#>  [63,]  0.0227403991
#>  [64,] -0.0285984506
#>  [65,]  0.0081643233
#>  [66,]  0.0382597720
#>  [67,]  0.0260523763
#>  [68,] -0.0388162149
#>  [69,]  0.0122077952
#>  [70,]  0.0128930290
#>  [71,]  0.0480686878
#>  [72,] -0.0111281753
#>  [73,] -0.0440214353
#>  [74,]  0.0492099294
#>  [75,] -0.0120514749
#>  [76,]  0.0112076391
#>  [77,] -0.0443792011
#>  [78,] -0.0309397108
#>  [79,]  0.0325623300
#>  [80,] -0.0284450867
#>  [81,]  0.0016615998
#>  [82,]  0.0464322751
#>  [83,] -0.0076529551
#>  [84,]  0.0279447603
#>  [85,] -0.0178517729
#>  [86,] -0.0260109813
#>  [87,] -0.0365388458
#>  [88,]  0.0360456272
#>  [89,]  0.0146394616
#>  [90,] -0.0049941087
#>  [91,]  0.0141385597
#>  [92,]  0.0005081309
#>  [93,]  0.0213716947
#>  [94,]  0.0165839344
#>  [95,]  0.0483169795
#>  [96,] -0.0089223376
#>  [97,] -0.0222496789
#>  [98,]  0.0367620483
#>  [99,] -0.0409167815