Generate three clusters of data points with optional noise.

This function generates three clusters of data points along with optional noise.

Usage

three_clusters_data_with_noise(
  sample_size,
  num_dims,
  num_noise_dims,
  min_noise,
  max_noise
)

Arguments

sample_size

Total number of data points to generate, should be a multiple of three.

num_dims

Number of dimensions for each data point.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

three_clusters_data_with_noise(sample_size = 100, num_dims = 7,
num_noise_dims = 4, min_noise = -0.05, max_noise = 0.05)
#> Warning: The sample size should be a product of three.
#>               [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#>   [1,] -2.19314883  0.90630179  1.380969258 -0.03805258 -1.57385502  0.41775847
#>   [2,] -0.10582958 -0.33874867  0.648129677  0.32545462 -0.71557040 -0.31871094
#>   [3,] -0.05757660  0.33003905  0.656409074 -0.57594431 -0.70668125 -0.78400108
#>   [4,] -2.24476178 -1.46939081  0.334765439 -0.13777281 -1.02296262  0.26052864
#>   [5,]  0.71777153 -0.21671085 -1.714935816  1.40151928  0.29486649  0.91885170
#>   [6,]  0.43844893  0.28230361 -0.870314395  0.98611613  0.62580293 -0.13448269
#>   [7,] -0.52061902  1.78688578  1.684569930 -1.09661086 -1.06293213 -0.62154305
#>   [8,]  0.40538549  1.46325751  2.354757201 -0.50834146 -0.86546609 -0.41195810
#>   [9,]  0.69832814 -0.43437912 -0.536674265  2.05718479 -3.14684444 -0.14722458
#>  [10,] -1.03225893 -0.19178881 -2.642450917 -0.71104875 -0.27156150  0.26815832
#>  [11,] -0.08245575  1.18805912 -0.751602749  0.73114007 -0.06244142  0.04168728
#>  [12,] -0.31262335  0.68457766 -0.134013365  2.28625174  0.11465867  1.77709426
#>  [13,] -0.61910957 -0.46915804  1.519195498 -1.57381532  2.20558200 -2.06428276
#>  [14,]  0.51624799  0.43339430 -1.247878620 -1.69120071  1.78916980  0.08339923
#>  [15,] -1.02136451  0.88859425 -0.775842928  0.46046969  0.71614596  0.03196105
#>  [16,]  0.14417026 -0.44415079  0.433968765  0.30839088  0.77822454 -0.57875984
#>  [17,] -0.22273786 -1.05021689  1.383817591  0.72023293  0.29669522  1.98778213
#>  [18,] -1.62012167 -1.49600766 -0.484752664  0.99310755  0.36459430  0.48622746
#>  [19,] -1.42116839 -0.78549513  0.179575195  1.48148191  0.92196799 -0.55196645
#>  [20,] -0.74335497 -0.46933299 -1.722038515  0.26963193 -1.21454123 -1.48730129
#>  [21,]  0.49348896  1.71403915  0.749349460  0.07046417 -1.57816632  1.38274046
#>  [22,] -1.23718221  0.27656809 -1.142122667  0.35800074 -0.43775045 -0.68156648
#>  [23,] -1.42839566  0.66464368 -0.848217940 -0.04823147 -2.53920920  0.47115652
#>  [24,]  0.10522845  0.47869322 -0.228190941 -0.50600313  0.34395273  0.09105672
#>  [25,] -0.02738123 -0.19441088  0.860046410 -0.28500427  0.57156703 -1.39164050
#>  [26,] -0.72837193  0.05085676  1.198040998 -0.45552307 -0.26108849 -0.03480334
#>  [27,]  1.30549327 -0.67453661 -1.322963068 -1.50353368 -0.30348987  0.17337844
#>  [28,] -0.29228466 -0.55565426  2.406269443  0.29289643  0.11438018 -0.70021418
#>  [29,]  0.75969178 -1.00629840 -0.784569409  0.96163093 -0.12011629  0.44755155
#>  [30,] -1.87660075  1.60598805 -1.503845974 -0.33579362 -1.78569781  1.47718534
#>  [31,]  1.04266809  0.22361696 -0.019541596  2.83273229 -2.42117937 -1.53150751
#>  [32,]  0.39983974  0.72721942 -2.975027474 -0.01564961 -0.17132628 -0.63944921
#>  [33,] -0.54829624  0.61232961  0.205869654  0.49399658 -0.08768444  1.00816466
#>  [34,]  9.02332451  1.77285522  1.573647193 -0.75135681  0.23547776  0.63115187
#>  [35,]  9.82144119  0.79569789 -2.491792394 -0.15800248 -0.23391480  0.25884441
#>  [36,]  9.69472559  0.62865756  0.013867083  0.56569175 -1.73717181 -0.94982275
#>  [37,]  9.15905148 -0.26317152  0.255707692 -0.27003152  0.75456043  0.07272108
#>  [38,]  8.81736561 -0.74224317 -1.788046159 -0.26701221 -0.49228317 -1.50865753
#>  [39,]  7.35428281 -0.61719596 -0.001171215  1.93229718  0.04436402 -0.01645098
#>  [40,] 10.57444335  0.45964636 -0.063603651 -0.07799508  0.66179282 -1.28632407
#>  [41,]  9.35007552 -0.03791945 -2.216567972  0.06715127  0.75348936 -1.06566264
#>  [42,]  8.95505678 -0.05720700 -0.433965800  1.64654885 -0.77876600 -0.80901866
#>  [43,]  9.13781585 -0.85219437 -0.174910891  0.08035232 -0.32767425  0.54383477
#>  [44,] 10.15895549  0.05857071  1.197774782 -0.24664449  0.55396019  0.31346565
#>  [45,]  9.38862434  0.31936487 -0.957200429  0.21883616 -1.42404049 -0.63079327
#>  [46,] 10.57863917  0.80937344  1.226253463  1.15733240  0.32768566  0.56537802
#>  [47,] 10.96598901 -0.45374484  0.293641899 -1.64677081 -0.88914941 -1.45336999
#>  [48,] 10.65670620  0.11350631 -0.092744648  0.01357618  0.38573884  0.63148522
#>  [49,]  9.73080275  0.47273722  0.682857709  0.89127114  1.20869909  0.50802111
#>  [50,]  9.85131993 -0.30952228 -1.107403925 -2.11046210  0.24752176  0.77802917
#>  [51,] 11.05372128 -0.11015182  0.380869049 -0.54067764  0.12066406  1.78413184
#>  [52,] 11.09153427  0.55208515  1.161753119 -0.26413408 -0.73134632 -1.58425380
#>  [53,]  9.13897212 -0.52962028 -1.176647849 -0.25054874 -0.44380428 -0.13340667
#>  [54,] 10.30709500 -0.51133625  0.032289268 -0.93956381 -0.09125948 -0.12528420
#>  [55,]  8.95215906 -2.09842604  1.417947366  1.02418263  0.57466467 -1.97368624
#>  [56,]  8.73552121 -0.98290078  1.116219456 -1.31911199  0.21258092 -0.32221529
#>  [57,]  9.68898786 -0.60575819  1.499080292 -0.62522131 -1.43493670  0.67581861
#>  [58,]  9.90405304 -0.32411375  1.303118251 -0.18668081  0.53637913  0.17039987
#>  [59,] 10.36063629 -0.38681257  0.951268577 -1.02416375  0.30771960 -0.13776886
#>  [60,] 10.33966642  1.29923407 -2.345182229  0.20719848 -1.20904556  1.03855443
#>  [61,] 10.15000929 -0.76150018 -1.252802335 -0.83763180 -0.89293989 -0.79939715
#>  [62,]  7.92429143  0.19681458  1.555980063  0.32928997 -0.58448668  0.01831979
#>  [63,]  8.86726029  0.30926194 -0.288725712  0.37346086  1.08480773 -0.43658607
#>  [64,]  9.03252228 -1.68711949  0.725702421  0.51307298  0.77307307 -0.57606358
#>  [65,]  9.39136752 -1.28782799  0.298667897  1.13034236 -0.21024432 -1.23511443
#>  [66,] 10.50488402  0.19515798  0.430335047 -0.58074076 -0.51542316 -0.07316060
#>  [67,] 49.37853101 -0.06350457  1.074907978 -0.05433793 -0.84327375 -0.88017888
#>  [68,] 49.14310409 -0.35414955  1.596855643 -1.23698114  0.05406275  0.30997376
#>  [69,] 49.72524147 -0.12840965  0.505833283  1.53731799 -0.21819288 -0.88581661
#>  [70,] 50.11160827 -1.53970371 -0.202906855 -0.02297115  0.63159555 -0.03993655
#>  [71,] 50.16439350 -1.10522105 -0.895818632 -0.37476117  0.80447572 -0.01955094
#>  [72,] 49.72260938  0.58785269  0.641600576  1.31984976 -0.80543765 -1.05231887
#>  [73,] 50.44975807  1.54416680  0.947031618  0.54846813 -0.63953489  1.59077828
#>  [74,] 48.81863288 -0.49387934 -1.431079992  0.72041532 -0.74258793 -0.21876194
#>  [75,] 50.58919275 -1.58109218 -1.042310172 -1.82439212 -0.43876314  0.07734078
#>  [76,] 51.63273117  0.87117564  0.305218777  0.64826256  0.04905887  0.74410957
#>  [77,] 49.96963582 -1.11997691  0.714244396 -0.87458087 -0.15346926  0.93130357
#>  [78,] 50.57898073 -1.40185127 -0.355163037 -0.85426776  1.70716560  0.19943848
#>  [79,] 51.70757004  0.86617571  1.091601767 -1.10530316  0.24224391 -0.24534840
#>  [80,] 51.89473320  0.36438920 -0.151654105  0.74636131  0.26769474  1.06049460
#>  [81,] 49.22386536  1.73358241 -0.660721206  1.16110455  1.82223650 -0.05057228
#>  [82,] 50.43952610 -0.38073592  0.051558506 -0.69753915 -2.77278313 -0.41563868
#>  [83,] 50.05240483  0.40635815  0.428999266  0.02657036  1.75617721 -1.78090984
#>  [84,] 51.01422526 -0.59384316  0.810462945 -1.80483230  2.75183444  1.35820381
#>  [85,] 49.96047437  0.11154742  0.557180273  0.12081969 -0.10591972 -0.32354376
#>  [86,] 50.32072765 -1.99153088 -0.307176870 -0.14097947  1.34892931  1.42537629
#>  [87,] 48.66000274  0.51588303  0.709729392  0.24696245 -1.96713725  0.63611110
#>  [88,] 50.22010066  1.26005437 -0.664042710  1.26160007 -1.26747441 -1.27839277
#>  [89,] 49.95641765 -2.23933951  1.417106288 -0.93741279 -1.68307518 -1.67845517
#>  [90,] 49.55327105  0.41963987  0.096232880 -0.24470616  0.84659603  0.84464861
#>  [91,] 49.43682764 -0.04047970  0.564049818  0.36280194  0.85517057 -0.85005405
#>  [92,] 49.74994534  0.70874232 -0.083933252  0.47472215  0.97152932 -0.34215600
#>  [93,] 50.20164039  0.98786681 -0.022151098  0.94480883  0.59645022  0.65832243
#>  [94,] 50.84875280 -0.02794175  0.463155161  0.37414965 -0.02187159 -0.33246867
#>  [95,] 49.35451302 -0.49240341  0.488315427 -0.24767760 -0.69623250 -0.71784941
#>  [96,] 49.39973790  0.37776417  0.948299325 -1.50140310  1.12508054 -0.86853673
#>  [97,] 49.01546709  0.02152830 -1.657096056  1.17050783  0.08673689 -0.31713250
#>  [98,] 50.51355291  2.66840566  0.077922960  2.17488162  0.41238564  0.02937130
#>  [99,] 50.19956860  0.48035441 -0.429685292  0.48179432  0.16606044  1.18064402
#>                [,7]          [,8]          [,9]        [,10]         [,11]
#>   [1,]  0.564327616  4.042422e-02 -0.0386524449 -0.037850994  0.0230355620
#>   [2,]  1.284955456  2.706632e-02 -0.0380654131  0.019806420 -0.0464702945
#>   [3,] -0.625058982  2.868065e-02  0.0008586671 -0.025721422  0.0397901131
#>   [4,]  0.161155760  2.536803e-02  0.0152503848  0.011260015  0.0436177553
#>   [5,]  0.645447324 -3.080172e-02 -0.0431983415  0.024827206 -0.0290140109
#>   [6,]  0.686162694  4.772631e-02 -0.0044149122  0.044541804  0.0362886753
#>   [7,] -0.816625837  3.666532e-02 -0.0082004807 -0.040707324  0.0071622534
#>   [8,]  0.831897637  4.191320e-02 -0.0327135013 -0.020245210 -0.0057862737
#>   [9,] -0.165422319  1.741796e-02 -0.0332354474 -0.019267677 -0.0461359359
#>  [10,]  0.881780623 -8.178367e-03  0.0134352890  0.020571771 -0.0498253466
#>  [11,]  0.298690917  4.231087e-02  0.0346757744 -0.045383769  0.0490436703
#>  [12,]  1.239074792 -1.476578e-02  0.0375633751  0.002641426  0.0073355934
#>  [13,]  1.278945293 -3.075290e-02  0.0275703301 -0.036002234  0.0022295303
#>  [14,] -0.545577904  1.736488e-02  0.0318245156 -0.035776996 -0.0295749091
#>  [15,] -0.491791113  4.719218e-03 -0.0414107800 -0.025089814 -0.0417353933
#>  [16,]  0.076394878 -1.282811e-02  0.0182842540  0.006181339 -0.0165278192
#>  [17,]  0.839446864 -2.934835e-02  0.0218355500  0.008597190 -0.0391062590
#>  [18,] -0.487047902 -5.929361e-03  0.0333146504 -0.038968778  0.0067514870
#>  [19,]  0.003444717 -4.520392e-02  0.0362531248 -0.023803585 -0.0015041380
#>  [20,]  0.722430344 -1.069017e-02  0.0069976041 -0.014359351 -0.0090125197
#>  [21,] -0.957981536 -1.256171e-02  0.0122582562 -0.004556570  0.0046156519
#>  [22,]  0.502105588  7.869771e-03 -0.0056410545 -0.001177076 -0.0033800514
#>  [23,] -0.080286491  2.605670e-02 -0.0173053467 -0.001069752  0.0167556488
#>  [24,]  0.241734694  4.831015e-02  0.0120701199  0.046458852  0.0394483536
#>  [25,]  1.310461110 -4.101518e-02 -0.0256430729 -0.024017451 -0.0446575628
#>  [26,]  0.896122608  4.786699e-02 -0.0007919788  0.037455750  0.0469723496
#>  [27,] -0.513487824  8.491204e-03 -0.0304532066  0.001854014 -0.0448599732
#>  [28,]  0.521876441  1.254057e-02  0.0176766022  0.031500933  0.0129582276
#>  [29,]  1.300212710 -3.098916e-02  0.0109321790  0.018941839 -0.0056066030
#>  [30,]  1.746410410  3.649483e-03  0.0139929644 -0.008427741  0.0203856640
#>  [31,] -0.962577048 -1.626492e-02 -0.0319041454  0.024739780 -0.0220490272
#>  [32,]  0.659490364  4.776044e-02  0.0127570285 -0.006233312 -0.0084177858
#>  [33,]  0.620212162 -2.679000e-02  0.0396905482  0.014497841 -0.0486719050
#>  [34,] -0.485527873 -1.688647e-04  0.0132083341  0.007759455  0.0303796712
#>  [35,]  0.171547533  2.911234e-02 -0.0453616493 -0.027210044  0.0182231972
#>  [36,] -1.158670385  4.075171e-02 -0.0276556141 -0.015801341  0.0001404131
#>  [37,]  0.660939962  3.125530e-02  0.0073249945  0.040082700  0.0427050776
#>  [38,] -0.364266566  4.693764e-03  0.0297576727 -0.007479337  0.0212919864
#>  [39,] -0.390424343 -4.285985e-02  0.0106489951 -0.044682368 -0.0140526580
#>  [40,] -1.456147701 -4.491657e-02  0.0245702764  0.008002542  0.0193926302
#>  [41,] -1.989762840  1.857689e-02 -0.0178545069 -0.012294158 -0.0129835188
#>  [42,] -0.429556514  3.706078e-02 -0.0159003206  0.031889731 -0.0022254417
#>  [43,] -0.240403233  3.194818e-02  0.0447471770 -0.035222778  0.0117401400
#>  [44,] -0.511468599  7.227175e-03  0.0468975492  0.037459523  0.0068247873
#>  [45,] -0.264161607  4.379368e-02  0.0281376664 -0.033672730 -0.0096285778
#>  [46,]  1.773762272 -3.026703e-02  0.0241120597  0.049647795 -0.0049848125
#>  [47,]  0.610667978 -4.624916e-03  0.0350392195 -0.046649312 -0.0209032410
#>  [48,]  0.382627411 -1.519407e-02  0.0134902223  0.005776603  0.0178341377
#>  [49,]  0.511784788  1.975296e-02  0.0324176193  0.049607396  0.0323232627
#>  [50,]  0.292209403 -3.607914e-02  0.0392813073  0.029176930 -0.0366059244
#>  [51,] -0.152449819  3.098516e-02 -0.0424384317 -0.031116040 -0.0199414371
#>  [52,]  0.173895834 -2.318124e-02  0.0081701874  0.021782364 -0.0047036141
#>  [53,] -0.139718649  4.098419e-02 -0.0264623580  0.036520940  0.0047313797
#>  [54,] -0.959306991  3.997992e-02 -0.0090396144  0.049074027 -0.0043193338
#>  [55,] -0.585600595  3.493268e-02  0.0496368643  0.014704934  0.0370245442
#>  [56,]  0.590306768  4.828303e-02  0.0123792257 -0.010774099 -0.0162397272
#>  [57,] -1.307490833 -1.186714e-02  0.0228386260  0.026563613 -0.0471876693
#>  [58,] -1.995439504 -3.423708e-02  0.0250458616  0.035769918 -0.0336316298
#>  [59,]  0.958217242  3.059111e-02 -0.0120871730 -0.044496227  0.0208042996
#>  [60,]  0.081612862  1.056960e-03 -0.0248446313  0.021610996 -0.0020275788
#>  [61,]  2.133551115 -4.835289e-02 -0.0025647172  0.035112763 -0.0387384678
#>  [62,]  0.239718258 -1.164140e-02  0.0247115278  0.033675197 -0.0370666194
#>  [63,] -1.267398571  3.041864e-02 -0.0201092135 -0.046766203 -0.0306514346
#>  [64,]  1.728769159 -3.667411e-02  0.0122284327 -0.037164096  0.0202300003
#>  [65,]  1.027258786  4.462335e-02  0.0391451758  0.023033043  0.0142592933
#>  [66,]  1.311992528  2.180849e-02 -0.0121642071  0.033421048  0.0094605768
#>  [67,]  2.400414483  3.829580e-02  0.0436892926  0.005538457 -0.0329641509
#>  [68,] -0.525416481  2.863905e-02  0.0172992665  0.005614134 -0.0371599268
#>  [69,] -1.259249057  4.441746e-02 -0.0419492389 -0.004158913 -0.0249414600
#>  [70,]  0.111049516 -5.719125e-03 -0.0484853951 -0.001159024  0.0319756642
#>  [71,] -1.373715984 -1.485155e-02 -0.0381173596 -0.023118861 -0.0220866299
#>  [72,] -1.519005963 -3.098153e-02  0.0209353369  0.033257312 -0.0205802375
#>  [73,] -0.024220628 -1.316090e-02 -0.0357554186 -0.036162413  0.0357807948
#>  [74,] -0.299527100 -1.517174e-02 -0.0491017628 -0.007956220  0.0248531149
#>  [75,]  0.879736063  2.813185e-02  0.0432806557  0.027702208 -0.0108748715
#>  [76,] -0.984978128 -1.200771e-02 -0.0230415274  0.016699068 -0.0022310298
#>  [77,] -0.814807166 -3.645480e-02  0.0033454647  0.016824624  0.0483912582
#>  [78,]  0.500687707 -2.703718e-02 -0.0091294925  0.032347168 -0.0342942488
#>  [79,] -0.481256885 -4.891962e-02 -0.0082337446  0.021051342  0.0293481530
#>  [80,]  0.489558955  1.250599e-02  0.0433702388  0.013245288 -0.0321273669
#>  [81,]  0.278115377  2.912626e-02  0.0067505342 -0.040422528 -0.0443262044
#>  [82,] -1.024025741 -3.595826e-02  0.0382706701  0.032649443  0.0003696824
#>  [83,]  0.590915342 -3.832016e-03 -0.0125876472  0.046681286 -0.0162919476
#>  [84,] -0.119032020 -1.991772e-02 -0.0211898650  0.048567265  0.0112345540
#>  [85,] -1.071944937  1.876798e-02 -0.0363260629 -0.021976858  0.0301381673
#>  [86,] -0.427702375  2.911250e-02  0.0397010307  0.041231497  0.0053759863
#>  [87,] -1.219065367  4.864002e-02 -0.0414049126  0.001736963  0.0492312837
#>  [88,] -1.552424417  3.012929e-03 -0.0054614848 -0.024974829 -0.0376336952
#>  [89,]  0.321980131 -2.234471e-03  0.0048218299 -0.032078836 -0.0007810471
#>  [90,]  2.016702716 -4.773749e-02  0.0010735424 -0.018514950 -0.0279793007
#>  [91,] -0.322211237 -1.487279e-02  0.0002972984 -0.003404786 -0.0085484510
#>  [92,]  1.604406334 -1.700475e-02  0.0382741228 -0.022321683  0.0133881637
#>  [93,] -0.512531126  2.189897e-02 -0.0041231070  0.015385182  0.0064739871
#>  [94,]  2.709113867 -2.800049e-05 -0.0097657378 -0.044152735  0.0382531768
#>  [95,]  0.187388608  2.853468e-02  0.0497524523  0.013911928 -0.0235968283
#>  [96,]  0.882101393 -1.758749e-02  0.0383343335  0.019511647  0.0027662017
#>  [97,]  0.390939448 -1.821725e-02 -0.0484373735  0.012115257 -0.0284862126
#>  [98,]  1.065135043 -7.992357e-03  0.0490414470  0.019409797 -0.0309171153
#>  [99,] -0.018143104 -2.035458e-02 -0.0254209649 -0.040326768  0.0211005686