Generate a torus-shaped dataset with optional noise.

This function generates a torus-shaped dataset along with optional noise.

Usage

torus_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated torus-shaped data points with or without added noise.

Examples

torus_3d(n = 100, num_noise = 4,
min_n = -0.05, max_n = 0.05)
#>               [,1]        [,2]         [,3]         [,4]          [,5]
#>   [1,]  0.35241497  1.06671867  0.481266335  0.012791100 -2.823347e-02
#>   [2,]  0.08851799 -1.50362177  0.869589721 -0.033238891 -2.983889e-02
#>   [3,]  0.36756211 -2.43569577  0.886215420  0.037457246 -6.905346e-03
#>   [4,] -0.28740069 -1.06948159 -0.450898875  0.014441871  4.229556e-02
#>   [5,] -2.39461124 -0.16151442  0.916492417  0.006232155 -1.444236e-02
#>   [6,] -2.10353105 -0.78832057  0.969169396  0.034290948  2.229104e-02
#>   [7,]  2.42652820 -1.76112692  0.058841135 -0.049897566 -8.772997e-03
#>   [8,]  0.89972198 -0.67150389 -0.479911329  0.022998417 -1.426558e-02
#>   [9,]  0.53548979  2.69753441 -0.661243738  0.049997534  1.872461e-02
#>  [10,]  2.32596822 -0.51893947  0.923684186 -0.046320747  2.853128e-03
#>  [11,]  0.12530417  1.11912941  0.486146102 -0.042006947 -7.265193e-03
#>  [12,] -1.21576386  1.21206723  0.959042510 -0.047650855 -3.024514e-02
#>  [13,] -0.45171854 -1.48664233  0.894910907  0.015200753 -8.268907e-04
#>  [14,]  1.14189868  2.22610988  0.864926333  0.046868184  3.517419e-02
#>  [15,]  0.14799861 -2.50543947 -0.860288879  0.044791340  2.542418e-02
#>  [16,] -2.12269918  1.27999609 -0.877946844 -0.014462608 -3.384913e-02
#>  [17,] -1.27489590  1.13197795  0.955470869  0.043954946 -3.980906e-02
#>  [18,]  0.53668868 -1.86618857  0.998306554 -0.032501633 -4.555869e-02
#>  [19,]  2.58778332  1.31287243  0.432220762 -0.018150632  1.071075e-03
#>  [20,]  2.54744235 -0.54034784 -0.796893745 -0.041723533 -4.319861e-02
#>  [21,] -0.17584101  1.38188224  0.794720911 -0.009682082 -4.960289e-02
#>  [22,]  2.67143846  0.67897571  0.654140709 -0.047822705  7.279688e-03
#>  [23,] -0.32131001  0.94699608 -0.006456841  0.004410062  2.668634e-02
#>  [24,] -2.37505613  0.30522211 -0.918858115  0.004606758 -3.253584e-03
#>  [25,]  0.52996638 -1.04199787 -0.556312612  0.028549424 -2.798182e-02
#>  [26,]  2.79833710  0.67246483 -0.478655698 -0.009686596 -1.850715e-02
#>  [27,] -1.07168533 -0.57049763  0.618321258  0.006652632 -4.998270e-02
#>  [28,]  1.04696195  0.56902053 -0.588635335  0.025314309 -4.251090e-03
#>  [29,] -1.32850070 -2.47101403  0.592597181  0.023762649  3.516564e-02
#>  [30,] -1.24707628 -1.60992111  0.999336224  0.016113764 -1.876324e-02
#>  [31,]  0.54606652 -1.86688195  0.998492181  0.039294575  3.594073e-02
#>  [32,] -0.98765595 -0.22680742  0.162937766  0.025973619  4.907423e-02
#>  [33,] -0.99981525  1.25047073  0.916965410 -0.004434513 -4.354895e-02
#>  [34,] -2.93003807 -0.01296395  0.367390588  0.028882667 -3.549042e-02
#>  [35,]  1.40261877 -2.37522060 -0.651738428 -0.016094256  4.684261e-02
#>  [36,] -1.10138893  1.04342192 -0.875710847  0.011890769  4.632047e-02
#>  [37,]  0.81758396 -2.46909716 -0.799294800  0.038969602  4.521089e-02
#>  [38,]  0.88594320 -0.56156540  0.308970944 -0.015983741 -2.141600e-02
#>  [39,]  0.47665231  2.91012541  0.315568691 -0.042151172  2.434142e-02
#>  [40,] -1.32875138 -2.27657421 -0.771708697 -0.005369851 -4.886025e-02
#>  [41,]  2.20155379 -2.00796174  0.200356543 -0.043929946 -5.466925e-03
#>  [42,]  1.02586808  0.24119006 -0.323698240  0.036793648  2.711679e-02
#>  [43,] -2.18218594  0.76618407  0.949824023 -0.004199096 -2.418931e-02
#>  [44,]  0.11350333  1.22466418  0.637938568 -0.042985806  4.332554e-03
#>  [45,]  1.44765557 -2.17136473  0.792632657 -0.031495725 -3.042768e-02
#>  [46,]  1.03445541 -1.61957470 -0.996933710  0.030987560 -2.087138e-02
#>  [47,]  1.62088042 -0.81142133  0.982291008 -0.004853854 -2.282966e-02
#>  [48,] -1.01384663 -0.55771847  0.538107110  0.049851238 -1.484643e-02
#>  [49,] -2.24049006  0.33833849 -0.964002698  0.035896136 -1.029495e-02
#>  [50,] -0.97132582  0.27339041 -0.134356793  0.017468452 -2.505670e-02
#>  [51,]  0.58041242  1.52703130  0.930463968 -0.037968923 -4.215194e-02
#>  [52,]  0.89449742  0.77090372  0.573586290  0.033466860 -3.046932e-02
#>  [53,]  2.42543752  0.27406779 -0.897569563 -0.007130096 -2.606022e-02
#>  [54,]  0.20831800  1.24372461 -0.673760353  0.023627176  3.158144e-02
#>  [55,]  0.55659897 -1.22200020 -0.753708139 -0.011018267 -4.389890e-02
#>  [56,] -2.54609841 -1.22835945  0.562317674 -0.017789612 -2.013427e-02
#>  [57,] -1.78062960  1.84169444 -0.827318020 -0.002970711  5.298020e-04
#>  [58,]  1.13640215  1.50546147  0.993505915  0.006446102  3.135295e-02
#>  [59,]  0.30867479 -1.03273240  0.386894089  0.046201817  2.244742e-03
#>  [60,]  0.83294129 -0.83645260 -0.572996735 -0.012671070  1.659503e-02
#>  [61,] -2.27427283 -1.53914194  0.665791138 -0.018791009  4.625692e-02
#>  [62,] -1.95630096 -1.08024340  0.972059470 -0.019184058  1.573825e-02
#>  [63,]  1.17377872 -1.73329466 -0.995634416 -0.044407775  2.169226e-02
#>  [64,]  2.17764402 -1.13579688 -0.889955572  0.004695153  2.178904e-02
#>  [65,]  2.81889174 -0.88912656  0.294048946  0.016711449  1.713942e-02
#>  [66,] -1.18210491  0.20856029 -0.600482626  0.046459704  4.740809e-03
#>  [67,] -0.32348123  1.17804565  0.627832099  0.001484640  1.979643e-02
#>  [68,] -2.80030008 -0.09179515 -0.597586808  0.033447429  3.363240e-02
#>  [69,]  0.28355566 -1.39139305 -0.814611066 -0.002840060 -2.173290e-02
#>  [70,] -1.47550237 -0.07760054 -0.852664754 -0.007678419 -4.507030e-02
#>  [71,]  1.28817466  1.27634557  0.982437669  0.046249275  2.572590e-02
#>  [72,]  0.88941792  1.39149706  0.937294827 -0.034185016 -4.570379e-02
#>  [73,]  2.99546922  0.11730732  0.066816023 -0.043605550 -2.062571e-02
#>  [74,] -0.53997054 -2.94789057  0.078218101  0.027679589 -1.468972e-02
#>  [75,]  1.78027216  0.22600849 -0.978669899 -0.002541825  3.833959e-02
#>  [76,]  0.07019122  2.96855821 -0.245534212 -0.016362833 -3.132224e-02
#>  [77,]  0.95434274  1.03179192 -0.804078349 -0.048908887  3.320171e-02
#>  [78,] -1.78173428  1.41493495  0.961381782 -0.040423051  4.181992e-02
#>  [79,] -0.90372516 -0.45061666  0.139931891  0.019021894 -3.738989e-02
#>  [80,] -1.93799130  1.48650586  0.896798431  0.013474451 -3.353056e-02
#>  [81,]  0.52543551 -1.00178581 -0.495197168  0.043497652  9.512925e-03
#>  [82,]  2.26435273  1.56625771 -0.657719575  0.016441964 -4.176536e-02
#>  [83,] -0.16596517  1.80070568 -0.981460940  0.043791908  1.675275e-02
#>  [84,]  1.64739623  1.35276988  0.991297228  0.035049440 -6.748072e-03
#>  [85,] -1.90434760 -2.29956875  0.168360977 -0.040894560  2.684805e-02
#>  [86,]  1.24936568  0.03378652 -0.661236437  0.032098355  1.861722e-02
#>  [87,]  0.66766552  0.81446242  0.321677327 -0.023987112 -1.152453e-02
#>  [88,] -1.84924925 -2.33623102 -0.201217270 -0.035638271 -3.321737e-02
#>  [89,]  1.42555397 -0.95090950  0.958110842 -0.024752458  3.755818e-02
#>  [90,] -0.17660754  2.48866146 -0.868938512 -0.037492534 -4.949380e-02
#>  [91,]  2.95905284 -0.41525686 -0.154145299 -0.027542295  2.619087e-02
#>  [92,]  1.01523236 -0.63372601  0.595695464  0.005158769  2.023108e-02
#>  [93,] -1.24997348 -0.39904393 -0.725828351  0.042106972 -4.429752e-02
#>  [94,] -0.76107286 -1.71980452  0.992855921  0.016799208 -4.586690e-03
#>  [95,] -0.07417931  1.01674022  0.196232571 -0.036330096 -2.607877e-03
#>  [96,]  2.91860103  0.27715884 -0.363148186  0.009382626  8.089365e-03
#>  [97,]  1.75153075 -2.25840006  0.513630168  0.040906155 -4.941682e-02
#>  [98,] -1.92787555 -0.58002747 -0.999912345  0.006382745 -4.968637e-02
#>  [99,]  1.25529213  0.73868151 -0.839412825  0.029974588  3.690385e-02
#> [100,] -1.54101948  1.96032651 -0.869736829 -0.004844667 -6.417651e-05
#>                 [,6]          [,7]
#>   [1,] -4.972993e-02  0.0341249627
#>   [2,] -4.983144e-02 -0.0073011297
#>   [3,]  2.974575e-02 -0.0117109125
#>   [4,]  1.497179e-02 -0.0365131162
#>   [5,]  7.313311e-03 -0.0036725340
#>   [6,] -4.722231e-02  0.0295553001
#>   [7,]  2.959498e-02 -0.0030715918
#>   [8,]  4.692440e-02  0.0382018849
#>   [9,] -3.339180e-02  0.0176522575
#>  [10,]  4.529848e-02  0.0068818477
#>  [11,] -4.853952e-02 -0.0409696253
#>  [12,]  4.585464e-02  0.0229739244
#>  [13,] -1.775798e-02 -0.0316561026
#>  [14,]  4.927120e-02 -0.0358781685
#>  [15,] -3.242121e-02  0.0460893255
#>  [16,]  1.977656e-02 -0.0226696819
#>  [17,] -4.785997e-02 -0.0014900642
#>  [18,] -4.650968e-02  0.0189844785
#>  [19,] -1.136838e-02 -0.0133649752
#>  [20,]  1.318736e-02  0.0407949157
#>  [21,] -4.507323e-02 -0.0261208749
#>  [22,]  1.589389e-02 -0.0152948894
#>  [23,]  2.632600e-02  0.0492130799
#>  [24,] -4.979243e-02 -0.0180194281
#>  [25,]  8.332904e-03  0.0129875111
#>  [26,] -2.558733e-02  0.0047976867
#>  [27,] -9.371038e-03 -0.0058835711
#>  [28,] -4.539305e-02 -0.0496761340
#>  [29,] -7.416861e-03  0.0011246438
#>  [30,]  1.003379e-02  0.0419148930
#>  [31,] -2.413230e-02  0.0219643536
#>  [32,]  4.668068e-02 -0.0393700294
#>  [33,]  2.730897e-02 -0.0261903933
#>  [34,] -5.545237e-03 -0.0184061838
#>  [35,] -4.657803e-03  0.0469244836
#>  [36,] -7.461717e-03  0.0320743180
#>  [37,] -2.917674e-02  0.0092593618
#>  [38,] -1.663250e-02 -0.0431698258
#>  [39,] -1.415156e-02 -0.0138153424
#>  [40,] -3.295290e-02 -0.0330933225
#>  [41,] -3.336156e-02  0.0480378095
#>  [42,] -7.807439e-04 -0.0430530482
#>  [43,] -2.035664e-05 -0.0084600601
#>  [44,] -3.656582e-02 -0.0462478032
#>  [45,]  9.646445e-03  0.0398076522
#>  [46,]  2.624505e-02 -0.0221265692
#>  [47,] -1.255642e-03 -0.0448026431
#>  [48,] -1.556870e-02  0.0412716157
#>  [49,] -3.784196e-02 -0.0391477973
#>  [50,]  6.748624e-03  0.0268867863
#>  [51,] -2.846456e-02 -0.0395524848
#>  [52,] -6.510306e-05  0.0206498849
#>  [53,]  3.066090e-02 -0.0068935606
#>  [54,]  1.188367e-03  0.0211439871
#>  [55,] -2.118296e-02  0.0259801880
#>  [56,] -2.335712e-03  0.0345497238
#>  [57,]  4.258501e-02  0.0365619248
#>  [58,] -2.231296e-02  0.0132344992
#>  [59,] -4.522349e-02  0.0169891869
#>  [60,]  3.283985e-02 -0.0208384463
#>  [61,]  4.611967e-02  0.0192108531
#>  [62,] -2.418985e-02  0.0185938631
#>  [63,]  2.901358e-02 -0.0156636891
#>  [64,] -3.524436e-02  0.0214762757
#>  [65,] -3.019913e-03 -0.0460022151
#>  [66,] -4.225030e-02 -0.0369807220
#>  [67,] -3.380724e-02  0.0001716919
#>  [68,] -1.232655e-02 -0.0315781217
#>  [69,] -4.751083e-03  0.0267148350
#>  [70,] -3.756668e-02  0.0182077345
#>  [71,]  1.617191e-02  0.0265502637
#>  [72,] -2.335840e-02 -0.0474419026
#>  [73,] -2.875084e-02  0.0200243003
#>  [74,] -2.619737e-02  0.0260491568
#>  [75,]  3.711825e-02 -0.0053559335
#>  [76,] -4.061361e-02 -0.0037256106
#>  [77,]  2.122121e-02  0.0190816951
#>  [78,]  3.801389e-02 -0.0481592762
#>  [79,]  1.528288e-02 -0.0014945254
#>  [80,]  4.921162e-02 -0.0226112427
#>  [81,]  2.652462e-02 -0.0374951120
#>  [82,]  2.197040e-02  0.0302090209
#>  [83,] -3.230487e-02 -0.0306259819
#>  [84,] -4.591486e-02  0.0435775331
#>  [85,] -2.748719e-02  0.0079807309
#>  [86,]  7.559957e-03 -0.0477228569
#>  [87,] -6.071685e-03 -0.0352001427
#>  [88,]  7.892396e-03  0.0299622617
#>  [89,]  4.157916e-02 -0.0393320236
#>  [90,] -6.495032e-04  0.0232789922
#>  [91,]  4.092241e-02  0.0076408508
#>  [92,]  4.501126e-02 -0.0266987029
#>  [93,] -2.765750e-02 -0.0223588180
#>  [94,]  4.219702e-02 -0.0380961990
#>  [95,]  3.720175e-02  0.0314252530
#>  [96,] -6.038044e-03  0.0180988424
#>  [97,] -4.429722e-02  0.0285938743
#>  [98,] -1.719134e-03 -0.0140753423
#>  [99,]  3.421598e-02  0.0303626247
#> [100,]  3.758525e-02 -0.0165637269