Generate a torus-shaped dataset with optional noise.

This function generates a torus-shaped dataset along with optional noise.

Usage

torus_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated torus-shaped data points with or without added noise.

Examples

torus_with_noise(sample_size = 100, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>               [,1]        [,2]         [,3]          [,4]          [,5]
#>   [1,]  2.95134715  0.17387318  0.291848932  4.880217e-02 -2.975745e-02
#>   [2,]  2.41254189  0.56938569 -0.877912016 -3.508738e-03 -2.832036e-03
#>   [3,]  2.65671198 -0.01169305  0.754119075 -3.957947e-02 -1.739357e-02
#>   [4,]  0.11370971  1.58848924 -0.913229262  1.389474e-02 -4.792033e-02
#>   [5,]  1.52378847  0.20198199 -0.886419278  1.684525e-02 -9.086172e-03
#>   [6,] -2.84268430 -0.92051280 -0.154396237  2.134029e-02 -4.288819e-02
#>   [7,]  0.81422817 -1.25808502 -0.865204671 -1.401697e-02 -2.083088e-03
#>   [8,] -1.67848295 -2.19673583  0.644517956  4.829316e-02  4.282864e-02
#>   [9,]  1.16382480  1.30022204  0.966944114 -1.391899e-02  3.325126e-02
#>  [10,]  0.56524002 -1.14622722 -0.691913107 -4.913340e-02 -2.226988e-02
#>  [11,]  2.08681507  1.61327676 -0.770285173  3.434210e-02 -3.209194e-03
#>  [12,]  1.45493764  0.94499991  0.964220161  2.785989e-02  2.830486e-02
#>  [13,]  0.94519695  0.35789431  0.145798641 -2.488358e-02  1.969910e-02
#>  [14,] -0.96270968 -0.40966886  0.300600097 -3.759221e-02 -2.465777e-03
#>  [15,] -2.56486244 -0.80402146 -0.725776664 -3.177787e-02  3.195091e-02
#>  [16,]  2.75613475 -0.77513958  0.505099152 -4.379226e-02  2.206003e-02
#>  [17,]  1.86671475 -2.33418840 -0.149098039  2.682114e-02 -3.570364e-02
#>  [18,] -2.00378965 -1.91360815 -0.637134737  3.443469e-02 -2.021334e-02
#>  [19,] -0.76092224 -0.77333140  0.403264716  3.758882e-02 -4.981296e-02
#>  [20,] -2.93475759 -0.60002246 -0.095096203  1.086782e-02 -2.124354e-02
#>  [21,] -1.69653819 -0.52805008  0.974776568 -4.950048e-02  3.341243e-03
#>  [22,] -0.69340432 -2.88543098 -0.252570924  4.775632e-02 -2.966963e-02
#>  [23,] -0.30821167  1.23326484 -0.684721253 -2.048905e-02  3.648157e-04
#>  [24,] -1.05627848  0.91321232 -0.797218982  2.339679e-02 -6.008219e-04
#>  [25,]  0.10886211  1.36010120 -0.772060448 -3.552491e-02 -2.722579e-02
#>  [26,]  0.85788038  0.63260259  0.357013900 -2.206973e-02 -1.485657e-02
#>  [27,]  1.33761856  0.99351710  0.942652121  4.796030e-02  4.345229e-02
#>  [28,] -2.35305695  1.85608496  0.077574572 -2.915045e-02 -3.128554e-02
#>  [29,] -0.92031258 -0.66133867 -0.498810712  6.275896e-03  3.495741e-02
#>  [30,]  1.19064529  0.77544731 -0.815255999 -2.974478e-02 -6.836595e-03
#>  [31,] -0.56321988  2.93781831 -0.131474028 -3.736948e-02 -4.656226e-02
#>  [32,]  0.89984816 -0.50181307  0.244346916  6.094176e-03  2.050010e-02
#>  [33,] -0.89605241  1.95232678  0.988966947 -3.554225e-02  2.024881e-02
#>  [34,]  0.99282507 -0.44416785  0.409415408  2.174160e-03  5.385966e-03
#>  [35,]  1.37141809 -1.62060169 -0.992406425  8.821119e-03  2.951150e-02
#>  [36,] -0.88800012  0.54351598 -0.283849090 -3.886306e-02  2.527886e-02
#>  [37,] -2.12316581  1.32056202  0.865827051 -2.122914e-02  3.425102e-02
#>  [38,] -0.46111486 -1.63700990 -0.954163470  6.643058e-03 -8.550765e-03
#>  [39,]  2.21269502 -2.02439597 -0.043968950 -3.222393e-02 -4.987617e-02
#>  [40,]  0.59319636 -0.83115077  0.204453194  2.762484e-02  4.093110e-02
#>  [41,]  0.10514321  2.21071180 -0.977006234  4.788409e-02 -3.585180e-02
#>  [42,] -0.16502308  1.06068917 -0.376170761  3.150111e-03  1.859722e-03
#>  [43,] -1.41828909 -1.50211644  0.997826953  1.199057e-02  1.746343e-02
#>  [44,]  0.09169235 -1.35761778 -0.768966287 -2.107204e-02 -4.715957e-02
#>  [45,]  0.54238581 -0.94130523 -0.406586138 -3.335159e-02 -2.133331e-02
#>  [46,] -0.18446315  1.02448816 -0.283278673 -3.829153e-03 -1.167491e-02
#>  [47,] -1.48158492 -1.21255622 -0.996339878 -6.764503e-03  5.503356e-03
#>  [48,]  0.57700353  2.59881590  0.749415373  2.616031e-02  2.397076e-02
#>  [49,]  0.24915813  2.86089344 -0.489999593  6.728964e-03 -2.095160e-02
#>  [50,] -2.43398090  0.42173728  0.882534298 -4.289297e-05  4.497229e-02
#>  [51,]  0.88902324 -0.48870667  0.169636179 -4.822358e-02 -2.101542e-02
#>  [52,]  2.38494151 -1.75620852 -0.273779710  1.467119e-02  4.192991e-02
#>  [53,] -0.63268898 -1.02320901  0.604002895  2.729661e-03 -4.816482e-02
#>  [54,] -0.76625532  1.60595040 -0.975361792  4.882831e-02  4.594356e-02
#>  [55,]  1.46396062  0.66243854 -0.919479483  4.360517e-03  4.323544e-02
#>  [56,]  1.19146369  2.75276084 -0.030119856 -7.861015e-04 -2.891505e-02
#>  [57,] -2.00684090 -1.51859580 -0.856194729 -1.098037e-02  2.631412e-02
#>  [58,]  2.40872878 -1.78741286 -0.032559643 -4.971482e-02  3.289054e-03
#>  [59,]  0.11705474  2.62043730 -0.782181680  4.498405e-02 -4.723022e-02
#>  [60,] -0.38842588  0.92381645  0.065591326  3.380627e-02 -3.473974e-02
#>  [61,] -1.30006451 -1.35347117  0.992371066 -1.226316e-02  2.863464e-02
#>  [62,] -1.94116761 -0.39179224 -0.999806159  2.340308e-02  2.109354e-02
#>  [63,]  0.13910776  2.99676540 -0.003923305 -2.922105e-02  1.568831e-02
#>  [64,] -0.68874198  1.41040584 -0.902633099 -3.233031e-02  3.402258e-02
#>  [65,]  1.84387173 -0.06479470  0.987916014  3.272231e-02 -4.149904e-02
#>  [66,]  2.40752342 -1.29385695  0.680042139  1.575659e-02  2.613965e-02
#>  [67,] -0.73696687 -2.17976996  0.953629885  2.605057e-02  5.190094e-03
#>  [68,] -0.81335653  0.87594918  0.593734794 -9.145341e-03 -4.301167e-05
#>  [69,] -1.42722331 -0.25948820  0.835573107 -9.725778e-06  9.491872e-03
#>  [70,]  0.37228343  1.64263265  0.948856143 -3.154991e-02  1.790699e-03
#>  [71,]  1.50336322 -2.59596592  0.016930733 -2.614094e-02 -1.525636e-02
#>  [72,] -0.85388385 -2.44596181  0.806874266 -3.794961e-02  3.204665e-02
#>  [73,]  1.13945332  0.48007498  0.645756962 -3.265003e-03 -4.361246e-02
#>  [74,] -2.18856802 -0.93894413  0.924377172  1.664674e-02  2.539891e-02
#>  [75,]  0.95887946 -1.96024367 -0.983261142  3.976260e-02  3.191270e-02
#>  [76,] -0.02129614  2.64835471  0.761265479 -2.993498e-02  2.362670e-02
#>  [77,]  0.26975806  1.11018339  0.514462308  3.899300e-02 -6.034905e-03
#>  [78,] -0.96345879 -1.21763170 -0.894384387 -2.290516e-02 -1.292873e-02
#>  [79,]  1.66014717 -2.10602460  0.731646865 -2.084781e-02  6.695991e-03
#>  [80,]  0.14714785  1.79880747 -0.980766642 -1.437243e-02 -4.548240e-02
#>  [81,]  1.28933789  1.11070269  0.954496932 -5.538752e-03 -3.474582e-03
#>  [82,]  0.97810533 -0.60674618 -0.528413767 -1.337245e-02  4.332930e-02
#>  [83,]  1.76272203 -2.20701168 -0.565787341 -2.697679e-02  2.843363e-02
#>  [84,] -0.65095171 -2.90762349 -0.200960324  4.674569e-02  9.086946e-03
#>  [85,] -2.28692213  0.13626413 -0.956729708  1.636188e-03  2.853983e-02
#>  [86,]  1.16935115 -0.83556639 -0.826595346  4.056563e-02 -4.917172e-02
#>  [87,]  0.57398496 -1.12420594 -0.675083835 -1.811339e-02  1.327530e-02
#>  [88,]  2.30856945  1.76909993 -0.417946432 -1.259991e-02  3.794350e-02
#>  [89,]  2.80388547 -0.33845682  0.566241846  2.063835e-02  3.337799e-02
#>  [90,] -0.48378518  2.84563406  0.462795276  3.556513e-02 -1.390352e-02
#>  [91,] -0.65364636 -0.76266886  0.094221896  1.170603e-02 -2.059627e-02
#>  [92,] -1.31862617  2.43199562 -0.642275493  2.190612e-03 -1.039211e-02
#>  [93,] -1.44601826 -0.86526559  0.949133846 -1.574797e-03  2.223873e-02
#>  [94,] -0.64333321 -1.40538646  0.890815798 -5.192061e-03  1.727100e-02
#>  [95,] -2.17998244  1.90995645 -0.439342380 -1.648092e-02 -3.421685e-03
#>  [96,] -0.71525791 -2.78444933 -0.484397244  2.697106e-02  4.381683e-02
#>  [97,]  1.27165868 -0.23534611  0.707466481  4.716157e-02  3.132839e-02
#>  [98,] -2.99460752 -0.03446603 -0.101790575  1.104903e-02  3.751982e-02
#>  [99,] -1.30108335  0.76151954 -0.870345034  1.839999e-02  6.358459e-03
#> [100,]  2.74651087 -0.99838742  0.386367670 -3.113110e-02  4.493391e-02
#>                [,6]          [,7]
#>   [1,] -0.016090870 -4.699401e-02
#>   [2,] -0.017780329  1.093699e-02
#>   [3,] -0.027768687  3.724280e-02
#>   [4,]  0.047906418  4.552577e-02
#>   [5,]  0.021696718 -2.982917e-02
#>   [6,]  0.043135066  8.255003e-03
#>   [7,] -0.010555987  1.964640e-02
#>   [8,]  0.011110333 -1.245501e-02
#>   [9,] -0.013501791  9.057025e-03
#>  [10,] -0.039230184 -3.420615e-02
#>  [11,] -0.041803396 -4.427372e-02
#>  [12,] -0.004241541 -3.917940e-02
#>  [13,] -0.029984368  7.334746e-04
#>  [14,]  0.003268838 -3.424701e-02
#>  [15,]  0.010979292 -3.968307e-02
#>  [16,] -0.024520417 -6.871333e-03
#>  [17,] -0.016480547  1.716554e-02
#>  [18,] -0.002208335 -2.610742e-02
#>  [19,]  0.049030443  3.656790e-02
#>  [20,]  0.010715003 -9.712011e-03
#>  [21,]  0.024870809  2.486343e-04
#>  [22,] -0.014151347  1.027410e-02
#>  [23,] -0.039501824 -3.831373e-02
#>  [24,]  0.013186766  1.246819e-02
#>  [25,] -0.005454219  4.814144e-02
#>  [26,]  0.023777201 -3.058403e-02
#>  [27,] -0.045187076 -4.950879e-02
#>  [28,]  0.038252588 -3.492377e-02
#>  [29,] -0.041843885 -4.679127e-02
#>  [30,]  0.033663765  1.129355e-02
#>  [31,] -0.021907293 -3.621819e-02
#>  [32,]  0.022918659  3.430191e-02
#>  [33,]  0.041574036  3.035977e-02
#>  [34,]  0.045960195 -2.569365e-02
#>  [35,]  0.022961718 -6.038688e-03
#>  [36,]  0.003349052 -4.829168e-02
#>  [37,]  0.009554832 -8.983701e-06
#>  [38,]  0.037677806  4.867221e-02
#>  [39,]  0.005395746 -4.461870e-02
#>  [40,]  0.004716351  8.114760e-03
#>  [41,] -0.013417223  2.167777e-02
#>  [42,] -0.045634592 -1.458776e-02
#>  [43,]  0.042779621 -1.816191e-02
#>  [44,]  0.001892221  5.726469e-03
#>  [45,]  0.005268416 -7.377263e-03
#>  [46,]  0.001407830 -2.972492e-02
#>  [47,]  0.030073497 -1.412247e-03
#>  [48,]  0.025624443  3.309990e-02
#>  [49,] -0.026196821  3.203876e-02
#>  [50,]  0.030518061 -4.414504e-02
#>  [51,]  0.024191997  1.190343e-02
#>  [52,] -0.033990481  5.989606e-03
#>  [53,]  0.001826239  1.272558e-02
#>  [54,]  0.027495305 -2.101561e-02
#>  [55,] -0.039167441  3.681496e-03
#>  [56,] -0.037050398  2.590999e-02
#>  [57,]  0.008023091  1.597503e-03
#>  [58,] -0.002694063  3.225110e-02
#>  [59,] -0.042232628 -1.859778e-02
#>  [60,]  0.007195566 -1.155402e-02
#>  [61,] -0.037651404 -3.250797e-02
#>  [62,]  0.005955758 -1.385156e-02
#>  [63,] -0.024714546  2.340686e-02
#>  [64,] -0.017843278  3.286447e-02
#>  [65,] -0.006886643 -5.349512e-03
#>  [66,]  0.004307737  4.965648e-02
#>  [67,] -0.026668750 -8.523085e-03
#>  [68,] -0.006215858 -4.467098e-02
#>  [69,] -0.036019237 -7.912997e-03
#>  [70,] -0.031295885  2.161778e-02
#>  [71,]  0.033524647 -3.787194e-02
#>  [72,]  0.023281929  2.987934e-02
#>  [73,]  0.001166066  2.991376e-02
#>  [74,]  0.019177196 -4.922116e-02
#>  [75,] -0.049316031 -3.614171e-02
#>  [76,]  0.030082012 -2.465299e-02
#>  [77,]  0.032002342 -2.420844e-02
#>  [78,] -0.039560162  4.927322e-02
#>  [79,] -0.006857366 -3.701788e-02
#>  [80,]  0.015416581 -3.688298e-02
#>  [81,] -0.041792513  4.777949e-02
#>  [82,] -0.037657899 -1.868708e-03
#>  [83,] -0.012224513  4.225744e-02
#>  [84,]  0.049788788  6.649602e-03
#>  [85,]  0.012007838  2.181689e-02
#>  [86,]  0.029265750 -4.403794e-02
#>  [87,]  0.025439076 -1.923923e-02
#>  [88,]  0.001943188  1.781122e-02
#>  [89,]  0.046607185 -1.876122e-03
#>  [90,] -0.022978430 -1.206595e-02
#>  [91,]  0.048869907  1.973283e-02
#>  [92,]  0.020235824 -4.823834e-02
#>  [93,] -0.024209797 -4.579645e-03
#>  [94,]  0.022126252  3.292185e-02
#>  [95,] -0.014507758 -1.621343e-02
#>  [96,]  0.027814252 -2.572754e-02
#>  [97,]  0.007538027 -4.929143e-02
#>  [98,] -0.005547903  2.976748e-03
#>  [99,] -0.032723272 -2.678240e-02
#> [100,]  0.009480789 -4.009226e-02