Generate data representing small spheres within a larger encompassing sphere with added noise.

This function generates data points representing small spheres within a larger encompassing sphere and adds noise to the data if specified.

Usage

diff_sphere(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate, should be a multiple of 13.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
diff_sphere(n = 390, num_noise = 2,
min_n = -0.05, max_n = 0.05)
#>                [,1]         [,2]          [,3]         [,4]          [,5]
#>   [1,]   2.15264081  -6.89994741 -2.096558e+00   8.84084278  1.701318e-02
#>   [2,]  -1.59579735  -8.30226886 -1.463587e+00   8.66723218  1.606255e-03
#>   [3,]  -2.62448645  -5.88680200 -2.198440e+00  12.21799524 -1.515195e-02
#>   [4,]   2.15980005  -5.23951048 -3.225834e+00   9.61462398 -1.504552e-02
#>   [5,]   0.37765229  -6.02675098 -4.878207e+00   8.77576025 -3.916970e-03
#>   [6,]  -0.86940891  -8.02818958 -7.633565e-01   8.60420111 -3.101151e-02
#>   [7,]  -2.39456482  -6.82448284 -4.118069e-01  10.05305456  1.392381e-02
#>   [8,]   0.36989578  -4.74958212 -1.516672e+00   8.35579655 -9.808695e-03
#>   [9,]  -1.48269788  -5.17049698 -5.193876e-01   9.10468276 -1.376565e-03
#>  [10,]   2.28701347  -7.30078628 -1.952421e+00  11.54734206  4.721620e-02
#>  [11,]   0.20308189  -7.45391897 -3.257742e+00   7.72206804  4.137618e-02
#>  [12,]  -0.36274894  -8.63026653 -3.115957e+00  12.49410556  1.398072e-02
#>  [13,]   0.03583438  -3.81307206 -1.348623e+00  10.71152276  2.432093e-03
#>  [14,]  -0.49002245  -3.66199795 -1.873192e+00   9.88035835 -2.182356e-02
#>  [15,]  -0.78551818  -7.67379384 -1.476738e+00  12.95608892 -2.312348e-04
#>  [16,]   0.27223069  -5.26091621 -4.529690e+00   8.68996709 -4.676903e-02
#>  [17,]  -1.25946303  -7.06408103 -1.711778e+00  13.11010484 -4.664230e-02
#>  [18,]   0.40136382  -4.18759165 -2.359985e+00  12.14804451 -1.697943e-02
#>  [19,]   0.53605510  -5.83450213  2.532695e-01  10.30611695  1.247595e-02
#>  [20,]   1.80409633  -4.99766301 -1.521430e+00  11.47485039 -1.270305e-02
#>  [21,]   0.95310567  -4.12270428 -1.761696e+00   9.44829682  1.831584e-03
#>  [22,]  -1.85943984  -6.16040689 -4.699184e+00  11.77779547 -2.295712e-02
#>  [23,]  -0.60235595  -4.54539935 -2.380741e+00   8.20568392  7.475937e-03
#>  [24,]  -2.97747131  -7.74160684 -1.876081e+00  10.71030455 -2.954348e-02
#>  [25,]   0.38529625  -4.17960512 -2.240257e+00   8.73579623  4.756168e-02
#>  [26,]   0.82988512  -7.49993471 -5.075312e+00  10.93657741 -1.390950e-02
#>  [27,]   0.03720310  -5.95332339 -4.418671e+00  12.66140138  5.781383e-03
#>  [28,]  -0.10307401  -8.90059021 -3.111622e+00  12.16448292  2.104691e-02
#>  [29,]  -1.35174245  -7.70016329 -2.539498e-01  11.61335921  4.288171e-02
#>  [30,]   1.07885344  -6.65480601 -6.748023e-01  12.31947504  2.265591e-02
#>  [31,]   2.88065982  -0.57944453  1.446203e+00   5.03085584  3.854656e-02
#>  [32,]  -0.86777834  -1.98176597  2.079174e+00   4.85724524  3.538895e-02
#>  [33,]  -1.89646743   0.43370088  1.344321e+00   8.40800831  1.012173e-02
#>  [34,]   2.88781906   1.08099240  3.169271e-01   5.80463704  3.307285e-02
#>  [35,]   1.10567131   0.29375190 -1.335446e+00   4.96577331 -3.944466e-02
#>  [36,]  -0.14138989  -1.70768670  2.779405e+00   4.79421417  2.121629e-02
#>  [37,]  -1.66654580  -0.50397996  3.130954e+00   6.24306763 -2.486890e-02
#>  [38,]   1.09791480   1.57092076  2.026090e+00   4.54580961  7.464143e-05
#>  [39,]  -0.75467887   1.15000590  3.023374e+00   5.29469583  3.914577e-02
#>  [40,]   3.01503249  -0.98028339  1.590340e+00   7.73735513  2.089173e-02
#>  [41,]   0.93110091  -1.13341609  2.850193e-01   3.91208110 -1.894828e-02
#>  [42,]   0.36527008  -2.30976365  4.268044e-01   8.68411862 -4.280163e-02
#>  [43,]   0.76385340   2.50743082  2.194138e+00   6.90153582  4.947531e-02
#>  [44,]   0.23799656   2.65850493  1.669570e+00   6.07037142  1.891325e-03
#>  [45,]  -0.05749916  -1.35329096  2.066023e+00   9.14610199  1.922574e-02
#>  [46,]   1.00024971   1.05958667 -9.869291e-01   4.87998016 -2.878356e-02
#>  [47,]  -0.53144402  -0.74357815  1.830984e+00   9.30011791  2.562035e-02
#>  [48,]   1.12938284   2.13291124  1.182776e+00   8.33805758  4.041058e-02
#>  [49,]   1.26407412   0.48600076  3.796031e+00   6.49613001  2.578040e-02
#>  [50,]   2.53211535   1.32283987  2.021331e+00   7.66486346 -4.699994e-02
#>  [51,]   1.68112469   2.19779860  1.781065e+00   5.63830989  3.792324e-02
#>  [52,]  -1.13142083   0.16009599 -1.156423e+00   7.96780853  3.249542e-02
#>  [53,]   0.12566307   1.77510353  1.162020e+00   4.39569698  4.670734e-02
#>  [54,]  -2.24945229  -1.42110395  1.666680e+00   6.90031761 -4.094721e-02
#>  [55,]   1.11331527   2.14089776  1.302504e+00   4.92580929 -2.815265e-02
#>  [56,]   1.55790413  -1.17943183 -1.532551e+00   7.12659047  2.635475e-02
#>  [57,]   0.76522212   0.36717949 -8.759097e-01   8.85141445  4.490325e-02
#>  [58,]   0.62494500  -2.58008733  4.311392e-01   8.35449598  1.117406e-02
#>  [59,]  -0.62372343  -1.37966041  3.288811e+00   7.80337227  2.813520e-02
#>  [60,]   1.80687246  -0.33430312  2.867959e+00   8.50948810 -4.569216e-02
#>  [61,]  -4.25441662  -2.46561013 -9.477166e+00   1.81214843  1.555211e-02
#>  [62,]  -8.00285478  -3.86793158 -8.844195e+00   1.63853783  3.139715e-02
#>  [63,]  -9.03154388  -1.45246473 -9.579049e+00   5.18930090 -2.859112e-02
#>  [64,]  -4.24725738  -0.80517320 -1.060644e+01   2.58592963  1.182557e-02
#>  [65,]  -6.02940513  -1.59241371 -1.225882e+01   1.74706590 -1.707904e-02
#>  [66,]  -7.27646634  -3.59385230 -8.143965e+00   1.57550677 -3.909184e-02
#>  [67,]  -8.80162224  -2.39014557 -7.792415e+00   3.02436022 -4.284323e-02
#>  [68,]  -6.03716165  -0.31524485 -8.897280e+00   1.32710220  1.697390e-03
#>  [69,]  -7.88975531  -0.73615971 -7.899996e+00   2.07598842  1.653476e-02
#>  [70,]  -4.12004395  -2.86644900 -9.333030e+00   4.51864772  4.995363e-02
#>  [71,]  -6.20397553  -3.01958169 -1.063835e+01   0.69337369  1.641861e-02
#>  [72,]  -6.76980637  -4.19592926 -1.049657e+01   5.46541122  5.542426e-03
#>  [73,]  -6.37122305   0.62126521 -8.729231e+00   3.68282842  2.505893e-02
#>  [74,]  -6.89707988   0.77233933 -9.253800e+00   2.85166401  3.475889e-02
#>  [75,]  -7.19257560  -3.23945657 -8.857346e+00   5.92739458  6.820234e-03
#>  [76,]  -6.13482674  -0.82657893 -1.191030e+01   1.66127275  8.336465e-03
#>  [77,]  -7.66652046  -2.62974375 -9.092386e+00   6.08141050 -2.106332e-02
#>  [78,]  -6.00569361   0.24674563 -9.740593e+00   5.11935017  6.247264e-04
#>  [79,]  -5.87100233  -1.40016485 -7.127339e+00   3.27742261 -3.697544e-02
#>  [80,]  -4.60296109  -0.56332573 -8.902038e+00   4.44615605 -2.549011e-02
#>  [81,]  -5.45395175   0.31163300 -9.142305e+00   2.41960248 -2.692062e-03
#>  [82,]  -8.26649727  -1.72606961 -1.207979e+01   4.74910113 -4.806504e-02
#>  [83,]  -7.00941337  -0.11106208 -9.761349e+00   1.17698957 -2.019991e-02
#>  [84,]  -9.38452873  -3.30726956 -9.256690e+00   3.68161020 -3.063711e-03
#>  [85,]  -6.02176117   0.25473216 -9.620866e+00   1.70710188  4.393639e-02
#>  [86,]  -5.57717231  -3.06559743 -1.245592e+01   3.90788306  1.201383e-02
#>  [87,]  -6.36985432  -1.51898612 -1.179928e+01   5.63270704 -4.244301e-02
#>  [88,]  -6.51013144  -4.46625293 -1.049223e+01   5.13578858 -3.218874e-02
#>  [89,]  -7.75879987  -3.26582602 -7.634558e+00   4.58466486  8.492386e-03
#>  [90,]  -5.32820399  -2.22046873 -8.055411e+00   5.29078069 -2.604337e-02
#>  [91,]  -1.13757232 -12.53725511  7.306960e+00   3.62374125  1.912451e-02
#>  [92,]  10.21104183   1.74449794  9.470965e+00  -5.29076302 -6.823914e-03
#>  [93,]  -9.36172614  10.77780360 -4.572363e+00   0.53900798  7.189237e-04
#>  [94,]  11.01864665  -6.97994584 -7.313593e+00  -1.17521660 -2.087572e-02
#>  [95,]  12.44893752   3.95591289  1.354874e+00   7.24838075 -4.265772e-02
#>  [96,]  -1.75047799 -10.72929924 -7.577351e+00   7.02863498 -2.790086e-02
#>  [97,]   6.94478266   6.67591470 -9.742390e+00   6.10639008  3.074746e-02
#>  [98,]  10.28374026   8.45738749  6.013975e+00   3.39843917  4.285247e-02
#>  [99,] -11.88490028   4.18037924  6.287805e+00  -5.17079107 -5.767070e-04
#> [100,]  -8.49658628  -4.54280679 -8.849249e+00   7.33905500 -2.300305e-02
#> [101,]  -6.26360824 -12.59625517 -4.813226e+00   1.98353725  2.029363e-02
#> [102,]   0.75686134  -5.88571749 -1.280487e+01  -5.08141556 -4.745385e-02
#> [103,] -12.77566406  -7.31829045  1.576444e+00   2.39579972 -1.705410e-02
#> [104,] -13.12456969   5.99388203  4.101030e+00  -0.02447386  1.890293e-02
#> [105,]  -0.84763804  -9.12420228 -9.710043e+00   6.83706913 -1.236857e-02
#> [106,]  -6.62832735  12.21293348  4.358507e+00  -3.59345958 -3.340483e-02
#> [107,]  -5.87395693   2.72874523  4.605287e+00 -12.72171016  1.148428e-02
#> [108,]  14.49082196  -2.93741871 -1.702450e+00   1.86797059 -1.961427e-02
#> [109,]  -7.90693456   6.68671116 -1.069349e+01   1.84867482 -1.563474e-02
#> [110,]   9.96842716  -5.20696295  1.581934e+00  -9.79874891 -2.543063e-02
#> [111,] -13.35988452  -4.43500095 -2.238417e+00   4.67265873  1.997720e-02
#> [112,]   0.38046368  -4.58903896  1.341647e+01  -4.87794791  2.749464e-02
#> [113,]  -5.16836126   0.63130210 -1.116006e+01   8.56402207  2.690204e-02
#> [114,]  -9.29188912  -8.26356116 -3.982034e+00  -7.38361396  3.704098e-02
#> [115,]  -1.07302562   8.72587483 -1.214307e+01   0.50346164  2.934588e-02
#> [116,]  -1.30712782  -9.72564012 -3.025544e+00 -10.93386593 -2.540732e-02
#> [117,]  -4.72210560  -7.75093643 -3.560737e+00  11.39937953  5.279571e-03
#> [118,]   5.73933577  -0.86628356  1.135398e+01   7.89915260  3.807583e-02
#> [119,]   6.01024161 -12.70362928  1.616493e+00   4.98816100 -4.963515e-02
#> [120,]   4.14174185  -8.41419368 -9.360977e+00  -7.02989588  4.806631e-02
#> [121,]  10.11531312  -2.36981329 -1.132983e+00 -10.76014759  1.766202e-02
#> [122,]  10.78125462  -5.17565879  5.059313e+00   7.50869177  4.904982e-04
#> [123,]  -3.04937162  -5.11734377  9.372407e+00  10.08325865 -3.309283e-02
#> [124,]   3.22312705   0.89301614 -5.150165e+00 -13.68538529  2.918599e-02
#> [125,]   0.70845535  -2.22171235 -1.334197e+01   6.44624087  2.563736e-02
#> [126,]   2.17472270   7.04999256 -1.066612e+01  -7.53671706 -1.180737e-02
#> [127,] -13.14339885  -1.16792466 -1.852156e-01   7.13110888 -1.580594e-02
#> [128,]  12.50384743  -5.86777893 -4.000153e+00  -4.26869416  4.122989e-02
#> [129,] -13.74044295  -3.43009560 -3.649979e+00  -3.33351499  4.302583e-03
#> [130,]  -3.88938213   6.13055655 -1.236464e+01  -4.40506463  2.874878e-02
#> [131,]  11.01672581  -3.91419597  2.966161e+00   8.91699018  3.216237e-02
#> [132,]  -2.82247577  13.19782071 -6.320546e+00  -1.70348426 -1.234710e-02
#> [133,]  -0.86086349  -9.88938039  1.124211e+01   0.27190815  6.071669e-03
#> [134,]   0.16152418  -2.32570482  1.043478e+01  10.52047156 -3.965713e-02
#> [135,]  -8.50461557  -3.89866747  1.169421e+01  -0.84699897  3.225420e-02
#> [136,]  -0.51486613  -9.66516224  9.635763e+00  -6.20254959  2.275280e-02
#> [137,]  -2.38890703  -3.54246840 -7.913201e+00  12.00521927 -4.379803e-02
#> [138,]  -0.21226863  12.11788322 -2.945602e+00   8.33278340 -3.650069e-03
#> [139,]   5.23335465 -10.26639242 -8.274326e+00   4.87326616 -2.254789e-04
#> [140,]  10.59943218  -8.01285551  3.207482e+00  -6.17723611  2.091612e-02
#> [141,]   0.38605629   3.50070698  8.366968e+00  11.94109943 -4.435780e-02
#> [142,]  -6.97244080  -6.65978083 -7.985762e+00   8.26196092  3.647009e-02
#> [143,]  -9.98554951  10.13633691 -2.204142e+00   4.20538118 -3.068151e-02
#> [144,]  -7.99327013   6.73963681  7.099253e+00  -8.07994690  3.726166e-02
#> [145,]  -2.94170433   6.63996509  1.312020e+01   0.34277010 -2.222062e-02
#> [146,] -11.63852549   4.24546480 -7.136823e+00  -4.53723589 -1.180547e-02
#> [147,]  -5.22345227   2.56216488  8.986420e+00 -10.50690806 -4.439201e-02
#> [148,]  -4.63323279   6.59838849 -7.362426e+00  10.28538283 -3.776503e-02
#> [149,]  -2.08661236  12.76755617  5.777541e+00  -4.92499585 -3.226880e-02
#> [150,]   9.56604950   1.29640738  8.375797e+00   7.85213657 -2.295346e-02
#> [151,]  -2.13107419  -5.01270544 -1.353780e+01   3.47264103  6.655193e-03
#> [152,]   1.28320970 -11.28595885 -9.793341e+00  -0.26641629  3.767581e-02
#> [153,]   3.02478547   9.37032730 -2.239073e-04  11.31581367  4.138408e-02
#> [154,]  -8.27084805   0.68481616  2.739150e+00  12.19102761 -3.749861e-04
#> [155,]   8.01746276   4.30020029  2.992915e+00 -11.54430727  2.554590e-02
#> [156,]  -5.66414710  -5.64954633  1.217605e+01  -3.56984742  4.587207e-02
#> [157,]  -0.91470279  -7.68764045  5.723048e+00 -11.50261836  3.865607e-02
#> [158,]  11.56850988  -3.41406492 -7.409557e+00  -4.96106899 -3.888366e-02
#> [159,]   4.49754644  11.90643231 -4.262725e+00   6.69612717  2.585429e-02
#> [160,]  -7.53851513 -12.17978161 -3.295784e+00   2.99357933  3.986343e-02
#> [161,] -13.88674554   4.05381589  5.299728e-01   3.92988599 -1.334919e-02
#> [162,]   8.18729238   5.22917907 -3.481363e-01  11.42377918  1.321781e-02
#> [163,]  -7.46195166  -5.10423482 -4.095515e+00  11.24690269 -3.843992e-02
#> [164,]  -2.36006409   7.92644213 -8.215832e+00  -9.43937066  8.204084e-03
#> [165,] -10.11546337   2.98483898  1.399842e+00 -10.57395758 -4.651237e-02
#> [166,]  -7.82117281  -2.37073759 -1.233448e+01  -2.46364821 -3.535699e-02
#> [167,]  -1.82441857   5.72460984 -1.102545e+01   8.20608268 -2.727869e-02
#> [168,]   4.16300825  -7.53809179 -1.226414e+01  -0.66129500 -7.070095e-03
#> [169,]   1.27653206   4.26522781 -2.571513e+00  14.09133125  1.396984e-02
#> [170,]   3.10089289  -1.10594754  1.270161e+01   7.26845001  4.017218e-02
#> [171,]  -0.20057077  -0.25192421  8.970727e+00  12.01758588 -2.136979e-02
#> [172,]   6.34797841 -13.21769328  2.314233e+00  -2.15408458 -1.134227e-02
#> [173,]  -4.36002518   8.75123452  6.427337e+00  -9.38591554 -4.025717e-02
#> [174,]  11.59750229  -0.21272180  8.963872e+00  -3.17831490 -1.231432e-02
#> [175,]  10.30297913  -0.66576854 -5.096248e+00   9.61424082 -4.593196e-02
#> [176,]   2.32870883  13.85885215  1.122463e+00  -5.12341768 -1.280777e-02
#> [177,] -11.04703950  -3.87629747 -1.377704e-01  -9.37647351  1.659563e-02
#> [178,]   1.26407905  13.07920912 -5.884553e+00   4.20813787 -2.951029e-02
#> [179,]   1.63206310  -5.36261217 -4.971870e+00 -12.99458600 -3.957487e-02
#> [180,] -12.21591574  -7.31188444  4.722025e+00   0.10113794 -4.261148e-02
#> [181,]   8.14603154  10.06377308 -6.777985e+00  -3.37957911 -3.593054e-02
#> [182,]  -1.59545746 -10.56872070 -1.830339e+00  10.36371158  8.075021e-03
#> [183,] -13.86206008   3.54389951  3.354077e+00   3.00569986 -3.040882e-02
#> [184,] -10.87512258  -3.72127462  7.976297e+00  -5.40948402  1.110145e-03
#> [185,]  -2.81541679  11.02731507  7.214633e+00   6.58944786  3.424502e-02
#> [186,]   0.83956290  11.89667236 -1.254160e+00  -9.01062732 -8.308103e-03
#> [187,]  -1.59040504 -13.80013517  5.605295e+00  -0.77945419  1.395520e-02
#> [188,]  -1.16453564  -6.31780410  7.033316e+00 -11.58713392  3.907406e-02
#> [189,]   1.37677424  11.09755123  7.642183e+00   6.44561047 -4.091924e-04
#> [190,]  -3.01710972  -3.87292155  1.338085e+01   4.67444757  3.423203e-02
#> [191,]  -6.66552538   3.87282905 -1.226858e+01  -3.87993679 -4.820397e-02
#> [192,]   5.33093720   3.09575460  1.366281e+01   0.57006702  2.182712e-02
#> [193,]  -1.83924031  -7.21713407  4.654375e+00  12.16005600 -1.844037e-02
#> [194,]  12.33773759   1.70498073  1.169676e+00  -8.27678265  2.035590e-02
#> [195,]   5.11163302   3.03752013  1.320532e+01   3.90692645 -9.199083e-04
#> [196,] -11.02856352  -1.00712709  6.858868e+00   7.43723115 -4.349411e-02
#> [197,]   0.87032125  -9.68537093  4.640909e+00  10.43542480 -1.195387e-02
#> [198,]  13.73697425   4.72138780  2.689278e+00  -2.60227153  1.584331e-02
#> [199,]  -8.15959937  -8.42306059  2.043050e+00  -9.12682509 -3.097637e-02
#> [200,]   9.16706081  -8.47927283  1.990615e+00  -8.06872853  1.821427e-02
#> [201,]  -1.63340927 -10.44141158 -5.438111e+00   9.15072948  1.864395e-03
#> [202,]   8.76167115 -10.27442634 -3.129341e+00   5.73380391  2.359111e-03
#> [203,]  -2.60684860  11.15901651  8.356606e+00   4.88342439  1.646955e-02
#> [204,]  13.60198095   0.32482273 -3.305143e-01   6.30645420  2.074100e-02
#> [205,]   5.61884232  -2.83128345 -1.320488e+01   3.32320232  1.843426e-02
#> [206,]   5.19260629   2.10517047 -1.079866e+01   8.77461963 -1.768110e-02
#> [207,]   1.70778715  -9.41728552  1.144012e+01  -1.58802359  2.695115e-02
#> [208,] -13.33977771   4.23085251 -8.064481e-01  -5.33852594  4.962866e-02
#> [209,]  -1.81595844   7.00998661  8.943064e+00  -9.62205772  3.323244e-02
#> [210,]  -5.42458452   1.86649322 -1.676941e+00  13.75783244  4.900636e-03
#> [211,]  13.73192425  -1.07496360 -8.527088e-01   5.87806068 -7.924833e-03
#> [212,]   4.23363199   6.11443244  1.016479e+01   8.14660723  1.121300e-02
#> [213,]  -4.69247528  -8.11200817  9.490087e+00  -6.86398198  3.590300e-02
#> [214,]   2.08986476  -8.83472063 -7.956161e+00  -8.90391361 -1.361298e-02
#> [215,]  10.86215836  -1.47798672 -1.023514e+01   0.26650558 -2.054805e-02
#> [216,]  -4.53517726   5.21777906  3.596725e-01 -13.30705018  1.574631e-02
#> [217,]  -2.05364346   4.63173240 -1.411534e+01  -0.29459358 -2.745160e-02
#> [218,]  -4.15261503   7.27395781  1.027078e+01   7.02541243  6.646649e-03
#> [219,]  -3.42797877  -3.94348940  8.717602e+00  11.03182940  3.757337e-02
#> [220,]  -0.75256353   5.58300332  1.102547e+01  -8.46774405 -1.374055e-02
#> [221,]  -4.46694747 -10.00543137  8.895967e-01  10.20521146  1.313869e-02
#> [222,]   0.69750394  -2.04826648  1.460449e+01   2.65084635  4.040160e-02
#> [223,]   6.57057654   0.01453703 -5.355378e+00 -12.37526714 -2.814390e-02
#> [224,]  -3.37306212  -3.12247811  1.040091e+01   9.78231600 -2.463870e-02
#> [225,]  -4.13168007   6.20051584 -3.479343e+00 -12.54499878  3.547120e-02
#> [226,]   9.19336778  -0.54967653  1.147686e+01   2.90887528  3.341442e-02
#> [227,]   3.47030262 -12.26968415  7.792369e-01  -7.86159274 -2.429026e-02
#> [228,]  -9.16080762  -2.69821428  1.144969e+01  -1.64434965 -1.993372e-02
#> [229,]  -3.70536039 -12.37104010  7.503187e+00  -1.38919448 -1.338825e-02
#> [230,] -11.22374600   6.21343837  5.751846e+00   5.22847787 -4.514059e-02
#> [231,]   9.77395640  -6.45065548 -7.922196e+00  -5.00975359  3.350452e-02
#> [232,]   1.58753278  -1.54470581 -1.458591e+01   2.71012563  1.250297e-02
#> [233,] -11.41819945   6.58902447 -5.057405e+00  -5.06281832  1.774048e-02
#> [234,]   0.50929954   4.75073871 -1.364125e+01  -4.01092120  4.320934e-02
#> [235,]  -5.98690106  -3.52415653 -3.535141e+00  12.81561993 -2.429277e-02
#> [236,]   3.89367562  11.35834185  6.962200e+00   5.68815672 -3.501206e-02
#> [237,]  -8.81916042   9.21368412 -3.762561e+00   6.94071846 -4.045311e-02
#> [238,]  -9.30799917   0.52263911 -4.367502e-01 -11.74296594  3.129198e-02
#> [239,]   1.28447799  10.31267080  9.789594e+00   4.60030312  1.935798e-02
#> [240,]  -2.23763459 -11.63431702 -6.000364e+00  -6.97361394 -1.037647e-03
#> [241,]   7.29271750  -0.50349815  9.049117e+00   9.46975378 -4.306777e-02
#> [242,]  -7.01045817 -10.64679292  5.561513e+00  -5.61861606 -3.460671e-02
#> [243,]   6.04397149   1.58052132  1.345497e+01   2.22175331 -4.177815e-02
#> [244,]  -1.09907057  12.52558432  4.558260e+00   6.79146852  8.869496e-03
#> [245,] -12.94936617   2.66988298  3.022638e-01  -7.07773108  3.471472e-02
#> [246,]   1.88729617  -6.32547048  2.090944e+00  13.30618241  3.079085e-02
#> [247,]   5.30307779  -7.34520178  7.276942e+00   9.48533049 -2.586799e-02
#> [248,]   3.49622510  -5.32746922  9.412189e+00  -9.78801240 -1.473175e-02
#> [249,]  -0.50967078   5.16421246 -7.571639e+00 -11.86344899  1.238128e-02
#> [250,]   2.09922707  -7.02746882 -7.750697e+00  10.54204110  2.117140e-02
#> [251,]   7.39989517  -9.84225397 -2.729909e+00   8.11906323  1.160142e-02
#> [252,]  -1.99469756   0.98583551 -1.483325e+01  -0.15523470  3.765488e-02
#> [253,]  -3.98029547  -9.32620872  1.915560e+00 -10.88621654 -6.447285e-03
#> [254,]  -4.31129467  13.99354639  3.242430e+00   0.28292364 -3.134232e-02
#> [255,]  -1.90055727 -13.16269488 -2.554915e+00   6.45009740  3.567554e-03
#> [256,]  -2.66137557   8.59171154  1.014495e+01  -6.41713057  4.305765e-02
#> [257,]   0.76193071   2.67322073  7.700910e+00  12.56858519  3.287951e-02
#> [258,]  -8.31231818  -4.69695728 -7.759993e+00  -8.58058711 -2.894529e-02
#> [259,]  -5.38613742  -3.00786271  1.153289e+01   7.34402432  3.999188e-02
#> [260,] -12.50565383   6.89622178 -2.544957e+00   3.81758316  3.823197e-02
#> [261,] -10.37838259   1.06302547 -2.186870e+00  10.55351849  3.094601e-02
#> [262,]   9.30258418   5.98077057  5.808146e+00   8.30408040 -4.190498e-02
#> [263,]  10.35243464   1.04506058 -1.048001e+01  -2.62760766  4.235367e-02
#> [264,] -11.32578704   5.33050777 -2.914748e+00   7.73411150 -3.708997e-02
#> [265,]   9.17456824   5.65234888 -9.422725e+00   4.48224397  2.700184e-03
#> [266,]  -5.55199749   3.86143956 -9.844287e+00  -9.07494426  2.154518e-02
#> [267,]  -0.39954214  12.85819294 -7.314090e+00  -2.45180018 -3.521341e-02
#> [268,]   4.78171111   0.78449454 -5.522357e+00  13.07759077  4.113018e-02
#> [269,]  -4.07750563  13.65091612  1.296361e+00   4.51064117 -2.685297e-02
#> [270,]  -4.12617312   5.95296933  8.229381e+00 -10.23787766  2.965704e-02
#> [271,]  -2.37614624 -12.79151971  7.431163e+00  -0.71328534 -1.825985e-02
#> [272,]   8.17755448   8.80060339 -4.208198e+00   7.93524134  7.389881e-03
#> [273,]  -6.14291776   2.83648421  3.780534e+00 -12.84237045  2.757974e-02
#> [274,]   7.41212064  12.72868979  3.928643e-02  -2.83538010  4.603738e-03
#> [275,]   2.83871962   5.49452588 -1.315711e+01  -3.69355865 -8.885216e-04
#> [276,] -13.11568190   0.61683076 -4.115021e+00   5.97201925  3.778235e-02
#> [277,]   2.50145435  13.22381098  4.536985e+00  -4.82590084  3.089138e-02
#> [278,]  -6.60334924   6.09630492  7.063250e+00  -9.71294695 -3.559450e-03
#> [279,] -12.60168789   7.41479532  2.822219e-01   3.33745762  4.653795e-02
#> [280,]  -5.91957440 -11.18708336  3.040563e+00  -7.45404476 -2.192818e-02
#> [281,]  -3.06252746  -2.57966555  1.384489e+01   4.15756244 -2.271448e-02
#> [282,]   6.08956595  -1.12809818 -1.659215e+00  13.56066323 -1.962436e-02
#> [283,]   1.71192960   8.82284691 -1.436552e+00  11.92321213  3.174975e-02
#> [284,]  -6.59074595  -3.04594244  8.929119e-01  13.09530492  4.505909e-02
#> [285,] -13.13216447  -4.44614794 -4.271484e-01   5.70925294 -1.530852e-02
#> [286,]   7.18171898   3.93151557  3.612704e+00  12.03804264 -4.522111e-02
#> [287,]  13.76789294   2.53140576  5.095735e+00  -1.75231036 -2.363636e-02
#> [288,]   1.55595978  -0.06372118  7.816354e+00 -12.70745970  6.830867e-04
#> [289,]  -5.71075099  -9.63239737 -9.972798e+00   0.38411642 -3.961406e-02
#> [290,]   0.35233610  14.47680719 -9.957046e-02   3.90998702 -1.772474e-02
#> [291,]   1.70647753  12.83641612 -2.657751e+00  -7.08877405  4.463025e-02
#> [292,]  13.77215794   0.49329916 -5.390862e+00  -2.45416554  1.259472e-02
#> [293,]  10.87098745  10.29238280 -6.594943e-01   0.67346518  5.567049e-04
#> [294,] -12.94855719   6.31841268 -1.857975e+00  -3.73636936 -8.778661e-03
#> [295,]   3.34878617   6.04477640 -8.553531e+00 -10.20212824  1.659580e-03
#> [296,]   4.24129138  -5.86019509 -1.066299e+01   7.67920341  1.191862e-02
#> [297,] -13.33012852  -5.57475919 -3.990560e+00   0.55241656  3.202104e-02
#> [298,]  -6.10163770  -2.85421096  8.344030e+00 -10.48811982 -2.628882e-03
#> [299,]   3.28192653  13.55516087 -5.335437e+00   1.42115774 -2.443555e-02
#> [300,] -10.62456157  -5.80934210 -3.482204e+00  -8.13907184 -2.574609e-02
#> [301,]  14.30506447   0.10010075  2.395318e-01  -4.50530074  4.289312e-02
#> [302,]  -8.01927191 -12.48138918  2.121961e+00  -0.63520425 -7.759684e-03
#> [303,] -11.76961908   0.26736569  1.415976e+00  -9.18692523 -3.906299e-02
#> [304,]  -9.07058944   6.44067494 -1.658226e+00   9.92433371  2.608614e-02
#> [305,] -11.72431305  -1.29308835 -5.682732e+00   7.31949167 -4.270295e-02
#> [306,] -14.69487423  -1.63619905 -2.092931e+00  -1.41533229  1.387756e-02
#> [307,]   7.39985297 -11.84934248  5.384207e+00  -0.91955218  1.974681e-02
#> [308,]  -0.51594191  -9.25837764  1.030988e+01   5.72036532  1.468811e-02
#> [309,]  -4.48928115 -11.82592619  7.062573e+00  -3.88765759 -3.571725e-02
#> [310,]  -9.94151342  -2.26685696 -1.029497e+01  -3.87829542  2.849413e-02
#> [311,]  -7.66696116 -10.18607368  6.444799e-01  -7.87694453  1.622426e-02
#> [312,]   9.76074469  -5.81786229 -2.150119e+00  -9.55287023 -4.297363e-02
#> [313,]   6.19433966   5.92655144 -5.045716e+00 -11.22706097  2.848153e-02
#> [314,]   8.52889396   6.29308518  3.746355e+00  -9.93075378  1.093619e-02
#> [315,]  -7.78505353   6.42073844 -1.323173e+00 -11.01890523  2.283338e-02
#> [316,]  12.47625736   3.12656807  8.608648e-01   7.66984265  3.871867e-02
#> [317,]  -5.29199000   1.83023124  5.691246e+00 -12.69861483 -1.886178e-02
#> [318,] -14.36613280   2.07835128 -3.687284e+00   0.83583461 -3.420622e-02
#> [319,]   1.80861470  11.65620203 -9.079021e+00   1.85290067  1.994977e-02
#> [320,]   4.18921797  -4.07688996  4.674746e-01  13.80619023 -3.164221e-02
#> [321,]  -9.78913300   3.59139063  1.323419e+00 -10.70155839  3.328469e-02
#> [322,]   1.39313738   5.00277370 -3.108956e+00  13.72464277  3.784973e-02
#> [323,]  -7.47136139  -3.86080046 -8.387363e+00   9.16106584  2.611218e-02
#> [324,]   4.49437088  14.04777718 -2.720830e+00   0.24014345 -2.860999e-02
#> [325,]  -8.37288154   1.11571634  1.218151e+01  -2.29367212  2.366030e-02
#> [326,]   9.45659648  -6.65234771 -9.487894e+00   1.13970448  2.996860e-02
#> [327,]  -4.08067814  10.51503957 -9.888463e+00   0.01751589 -7.427214e-03
#> [328,]  -1.32241517  -2.90843921  7.838613e+00  12.38338989 -3.970157e-02
#> [329,]  -8.65264629   6.99383934 -1.003541e+01   0.71301032 -2.195209e-02
#> [330,]  -2.65166426  -3.38838578 -9.340324e+00  10.91997560 -3.600596e-02
#> [331,]   4.60424532   0.17500304  1.395880e+01  -2.98700451 -4.273254e-02
#> [332,]   3.70248606   0.76106943  1.031674e+01  10.21161828 -4.676937e-02
#> [333,]  -9.59587774  -1.38481171 -9.812678e+00  -5.89175580  3.858613e-02
#> [334,] -10.82076550   9.65504349 -1.155940e+00   3.65444554 -2.336692e-02
#> [335,]  -2.41145248  13.29547688 -6.185975e+00  -2.03688613 -1.325866e-02
#> [336,]  -1.18212984  11.27245276  8.104133e+00  -5.55494408 -2.264760e-02
#> [337,]  -1.75934654 -10.36202252  1.065880e+01   0.96081191  1.070925e-02
#> [338,]  13.66537416  -3.52890514 -4.054298e+00   3.06056302 -3.631726e-02
#> [339,]  -2.82412871   7.92474752 -8.753087e+00  -8.80943440 -3.484990e-02
#> [340,]   6.26979409  11.14131199 -7.796054e+00   0.88452620 -1.177247e-02
#> [341,]  -2.32242574  10.83741097  3.622914e+00  -9.43564296  3.835105e-02
#> [342,]   1.41687535 -11.26896221  1.375594e+00   9.70106677  1.108216e-02
#> [343,]   0.73985778  -7.24840566 -3.645329e+00  12.59463381 -4.821762e-02
#> [344,]  -7.23751184  -1.39803400 -1.235297e+01  -4.25066546  1.746505e-03
#> [345,]  -7.78086502   5.95024357  5.862850e+00  -9.73035109 -3.477552e-02
#> [346,]   7.13187405 -10.14897614 -6.769641e+00  -5.03056758  3.907354e-02
#> [347,]   1.77726444 -13.92500037  4.637686e+00   2.53526499 -1.879682e-02
#> [348,]  -2.09037320  -9.50075621  1.063399e+01  -4.15742311  3.501518e-02
#> [349,]   1.56931243  10.05809029 -5.383423e+00  -9.61201527 -4.304542e-02
#> [350,]   4.62423481  -4.90061195 -1.317713e+01  -2.44206746 -1.301408e-02
#> [351,] -12.22265406   6.86293650 -4.521742e+00  -2.83913417 -2.313740e-02
#> [352,]  -2.00137789 -13.08619145 -3.031300e+00  -6.36846116 -3.353284e-02
#> [353,]  -5.57122379  -1.31407381  7.613413e+00 -11.58752009 -2.160113e-02
#> [354,]  -0.71350338 -13.70526783  2.545829e+00  -5.49320488 -1.430172e-02
#> [355,]   3.04091229  11.19843874 -6.904349e+00   6.53282428 -7.034689e-03
#> [356,]   3.39307647   5.16637815 -2.102257e+00 -13.50466904 -8.910901e-03
#> [357,]  10.69354756  -2.02278800 -1.373050e+00  10.23088966  2.081789e-02
#> [358,]   8.16353408 -12.15453708  3.069516e+00   1.09636269 -4.947695e-03
#> [359,]  -3.83948681   7.49638813 -1.230483e+01  -1.62900744  1.802991e-02
#> [360,]  -4.01786814   8.48780817  6.845838e+00  -9.48411075  2.044523e-02
#> [361,]  -5.96483448   7.05251193 -1.151510e+01  -2.66180137  2.454755e-02
#> [362,]  13.93560668   1.61898895  1.678792e+00  -5.03581154 -3.253221e-02
#> [363,]   5.88390813   0.23052368  1.153850e+01   7.56236693 -4.553061e-02
#> [364,]  -1.75186799  -3.75175078 -1.388695e+01  -3.87400870  2.788077e-02
#> [365,]   7.30892580 -12.55018644 -2.994752e+00  -2.25917834  4.046501e-02
#> [366,]   5.74774798  -2.59182390 -9.055519e-01  13.58034675  3.395781e-02
#> [367,]   7.78568138   5.19060053 -6.276532e+00   9.90181715  2.557479e-02
#> [368,]  -4.08790748 -12.11130185  7.085985e+00  -3.37552922 -1.000378e-02
#> [369,] -10.37791348  -9.26508565 -4.206290e+00  -3.71001690  4.550585e-02
#> [370,]   1.04049213   2.02619499  1.033433e+01  10.63077873  3.844611e-02
#> [371,]  -9.40494808   4.37842662 -1.024541e+01  -3.52249924  1.537560e-02
#> [372,]   7.21865531   8.31241055  1.017747e+01   0.46249105 -4.000024e-02
#> [373,]  -0.05895474   7.27526124  7.687084e+00  10.62900932 -1.469586e-03
#> [374,] -11.68047845   4.56193147 -5.610119e+00   6.02343496 -3.681870e-02
#> [375,] -12.12588266   8.15180502  3.370557e+00  -0.38780202 -1.466488e-02
#> [376,]   3.09589050   1.45168941 -1.448747e+01  -1.84965066  2.420848e-02
#> [377,]   4.91970842  12.57562561  6.282954e+00  -1.78173863  1.353138e-02
#> [378,]   0.39815338   8.01742309 -8.627331e+00   9.28070934  4.877531e-02
#> [379,]  -9.94439519  -8.05108875 -5.299287e+00   5.76251070 -4.254274e-02
#> [380,]   6.40251585  -4.29318421  1.286407e+01   0.30332950 -4.889923e-02
#> [381,]  -8.09180188   8.50938435 -6.392472e+00   6.80069312  4.762569e-02
#> [382,]   5.95360147   7.35362968  9.375975e+00   6.89709012  3.349131e-02
#> [383,]   6.27665228  -7.70003910 -3.655756e+00 -10.62772226 -4.260858e-02
#> [384,]   0.96646620   4.62482803 -9.662609e+00  10.45518497  3.526805e-02
#> [385,]   3.43690865   1.60929136 -1.155370e+01   8.78122418 -2.304562e-02
#> [386,]   5.11943238   4.76701002 -1.309411e+01  -2.14736788  2.399438e-02
#> [387,]  -8.64880612   3.74836472 -3.459647e+00 -11.14355212  8.116533e-03
#> [388,]  13.68755952  -2.35006286 -5.667617e+00   0.07767460 -2.592272e-02
#> [389,]  -5.47827406 -12.84441766 -3.256698e+00   4.40492539 -1.569113e-02
#> [390,]  -3.97232929   9.48574647  1.089722e+01   0.70131265  3.998552e-02
#>                 [,6]
#>   [1,] -0.0299643832
#>   [2,]  0.0104316566
#>   [3,] -0.0473028292
#>   [4,]  0.0452194851
#>   [5,] -0.0413021792
#>   [6,]  0.0168948047
#>   [7,]  0.0217131028
#>   [8,]  0.0484017081
#>   [9,]  0.0325122841
#>  [10,]  0.0178046789
#>  [11,] -0.0343310878
#>  [12,] -0.0198556353
#>  [13,]  0.0203098150
#>  [14,] -0.0142972067
#>  [15,]  0.0261684653
#>  [16,] -0.0221447836
#>  [17,] -0.0151047887
#>  [18,] -0.0103291712
#>  [19,]  0.0210791810
#>  [20,] -0.0107304781
#>  [21,] -0.0143019429
#>  [22,] -0.0498698028
#>  [23,] -0.0458431272
#>  [24,] -0.0454449338
#>  [25,] -0.0114684443
#>  [26,]  0.0046648704
#>  [27,] -0.0024163717
#>  [28,] -0.0275898947
#>  [29,] -0.0272622008
#>  [30,] -0.0364293487
#>  [31,] -0.0309327350
#>  [32,]  0.0296519995
#>  [33,]  0.0205950031
#>  [34,]  0.0459227742
#>  [35,]  0.0477904299
#>  [36,] -0.0405245173
#>  [37,]  0.0304850290
#>  [38,] -0.0157263169
#>  [39,]  0.0392056732
#>  [40,] -0.0037289331
#>  [41,]  0.0216777676
#>  [42,] -0.0396881675
#>  [43,]  0.0366874143
#>  [44,]  0.0192671251
#>  [45,]  0.0163077996
#>  [46,]  0.0478473094
#>  [47,]  0.0398279809
#>  [48,] -0.0476461635
#>  [49,]  0.0115731091
#>  [50,]  0.0480831801
#>  [51,]  0.0238364849
#>  [52,]  0.0129535138
#>  [53,] -0.0043645946
#>  [54,] -0.0049098072
#>  [55,]  0.0249883467
#>  [56,]  0.0109812774
#>  [57,]  0.0109105099
#>  [58,]  0.0098786488
#>  [59,] -0.0491856167
#>  [60,] -0.0093101903
#>  [61,] -0.0174248512
#>  [62,] -0.0461459515
#>  [63,]  0.0427010586
#>  [64,]  0.0118736496
#>  [65,]  0.0240240101
#>  [66,] -0.0186267743
#>  [67,] -0.0458526994
#>  [68,] -0.0153174600
#>  [69,] -0.0294077890
#>  [70,]  0.0095667423
#>  [71,] -0.0409134390
#>  [72,] -0.0078788311
#>  [73,] -0.0348781692
#>  [74,] -0.0493942986
#>  [75,]  0.0360556579
#>  [76,]  0.0104569928
#>  [77,]  0.0351667408
#>  [78,]  0.0116525801
#>  [79,] -0.0007742667
#>  [80,] -0.0396974565
#>  [81,] -0.0271651650
#>  [82,] -0.0312885590
#>  [83,] -0.0285264278
#>  [84,]  0.0396778687
#>  [85,] -0.0460942727
#>  [86,] -0.0409880836
#>  [87,]  0.0125339646
#>  [88,]  0.0247812803
#>  [89,]  0.0169392463
#>  [90,] -0.0495260336
#>  [91,]  0.0327344286
#>  [92,] -0.0110578050
#>  [93,]  0.0097977272
#>  [94,]  0.0089933715
#>  [95,] -0.0123233595
#>  [96,] -0.0236370778
#>  [97,] -0.0314808176
#>  [98,]  0.0052764100
#>  [99,]  0.0245506608
#> [100,] -0.0295174006
#> [101,] -0.0341121246
#> [102,] -0.0293830940
#> [103,]  0.0429953310
#> [104,] -0.0238097583
#> [105,]  0.0040556213
#> [106,] -0.0290921700
#> [107,]  0.0282746627
#> [108,]  0.0157569766
#> [109,] -0.0193973952
#> [110,] -0.0079421847
#> [111,]  0.0487919415
#> [112,]  0.0305934193
#> [113,] -0.0108824828
#> [114,] -0.0008504821
#> [115,] -0.0170417387
#> [116,] -0.0464184270
#> [117,] -0.0118198337
#> [118,] -0.0223265183
#> [119,] -0.0165689954
#> [120,]  0.0171838681
#> [121,] -0.0042185053
#> [122,]  0.0249359633
#> [123,]  0.0298969204
#> [124,] -0.0109674703
#> [125,]  0.0032705371
#> [126,] -0.0145442782
#> [127,]  0.0160515127
#> [128,] -0.0195470848
#> [129,]  0.0045294957
#> [130,] -0.0480125775
#> [131,]  0.0018926770
#> [132,]  0.0319420340
#> [133,]  0.0183369605
#> [134,] -0.0459261568
#> [135,] -0.0208843993
#> [136,]  0.0149954758
#> [137,]  0.0247239927
#> [138,] -0.0264804634
#> [139,]  0.0139952490
#> [140,]  0.0207657282
#> [141,] -0.0298479932
#> [142,] -0.0468177041
#> [143,]  0.0467784934
#> [144,] -0.0229971079
#> [145,] -0.0290110841
#> [146,]  0.0163030809
#> [147,]  0.0359930183
#> [148,]  0.0391829056
#> [149,] -0.0007797753
#> [150,] -0.0257705497
#> [151,] -0.0003595334
#> [152,] -0.0317866206
#> [153,]  0.0075854585
#> [154,]  0.0369168865
#> [155,]  0.0471848964
#> [156,] -0.0448185394
#> [157,] -0.0243685251
#> [158,]  0.0306072855
#> [159,] -0.0204828955
#> [160,] -0.0286209604
#> [161,]  0.0441898395
#> [162,]  0.0249873501
#> [163,] -0.0445161339
#> [164,]  0.0317941884
#> [165,]  0.0148583225
#> [166,]  0.0082525456
#> [167,]  0.0084217807
#> [168,] -0.0125677991
#> [169,] -0.0442033564
#> [170,] -0.0108832577
#> [171,]  0.0357817216
#> [172,] -0.0427162361
#> [173,]  0.0191645199
#> [174,]  0.0408592896
#> [175,] -0.0461002005
#> [176,] -0.0017143632
#> [177,]  0.0204467894
#> [178,] -0.0362488516
#> [179,]  0.0047268782
#> [180,] -0.0392856859
#> [181,]  0.0316157651
#> [182,] -0.0116076417
#> [183,] -0.0067904739
#> [184,]  0.0373270290
#> [185,] -0.0186952641
#> [186,]  0.0055863998
#> [187,] -0.0267207064
#> [188,]  0.0476969513
#> [189,] -0.0408327662
#> [190,] -0.0300739495
#> [191,]  0.0128309683
#> [192,] -0.0254998236
#> [193,]  0.0368039088
#> [194,]  0.0404451984
#> [195,]  0.0114070552
#> [196,]  0.0360081776
#> [197,] -0.0224193861
#> [198,]  0.0301543643
#> [199,]  0.0140020505
#> [200,] -0.0193725071
#> [201,] -0.0198374213
#> [202,] -0.0299423354
#> [203,]  0.0043765077
#> [204,]  0.0130153314
#> [205,]  0.0445957415
#> [206,]  0.0165921058
#> [207,]  0.0144059786
#> [208,] -0.0286737445
#> [209,] -0.0459265233
#> [210,] -0.0494955947
#> [211,] -0.0204653413
#> [212,] -0.0353885247
#> [213,]  0.0312724704
#> [214,]  0.0357693285
#> [215,]  0.0068128975
#> [216,] -0.0420383352
#> [217,]  0.0426482607
#> [218,] -0.0064203122
#> [219,]  0.0222017498
#> [220,]  0.0466327467
#> [221,] -0.0067406867
#> [222,] -0.0113580145
#> [223,]  0.0248535921
#> [224,] -0.0002419315
#> [225,]  0.0468074596
#> [226,]  0.0074736375
#> [227,] -0.0229902867
#> [228,]  0.0332021134
#> [229,] -0.0387170729
#> [230,]  0.0421645744
#> [231,]  0.0494321901
#> [232,]  0.0222273530
#> [233,] -0.0264308586
#> [234,] -0.0093039250
#> [235,]  0.0273933636
#> [236,] -0.0434736632
#> [237,] -0.0186634808
#> [238,] -0.0220876258
#> [239,] -0.0226503788
#> [240,]  0.0242954168
#> [241,]  0.0355783315
#> [242,]  0.0459465368
#> [243,] -0.0320236968
#> [244,] -0.0205145546
#> [245,]  0.0152705015
#> [246,] -0.0153252888
#> [247,] -0.0225046200
#> [248,] -0.0195495758
#> [249,] -0.0171381534
#> [250,] -0.0072901649
#> [251,]  0.0311489899
#> [252,]  0.0439051131
#> [253,]  0.0053917592
#> [254,]  0.0416760790
#> [255,] -0.0113264317
#> [256,]  0.0259034847
#> [257,]  0.0065685699
#> [258,] -0.0435717741
#> [259,] -0.0070676897
#> [260,]  0.0130117224
#> [261,] -0.0492103347
#> [262,]  0.0424193460
#> [263,] -0.0080156112
#> [264,]  0.0247982748
#> [265,] -0.0494386119
#> [266,] -0.0031001754
#> [267,]  0.0384837404
#> [268,] -0.0133946033
#> [269,] -0.0444344055
#> [270,] -0.0412088149
#> [271,]  0.0482690136
#> [272,]  0.0288730615
#> [273,]  0.0473867274
#> [274,] -0.0318573298
#> [275,] -0.0127151501
#> [276,]  0.0310874645
#> [277,] -0.0415701205
#> [278,]  0.0046033602
#> [279,]  0.0110073412
#> [280,] -0.0042060562
#> [281,]  0.0147514978
#> [282,]  0.0460891930
#> [283,] -0.0059436755
#> [284,]  0.0396971727
#> [285,]  0.0243395832
#> [286,] -0.0275306388
#> [287,] -0.0312191814
#> [288,] -0.0077302772
#> [289,]  0.0090521603
#> [290,] -0.0129959400
#> [291,]  0.0053369763
#> [292,]  0.0006899428
#> [293,]  0.0433534647
#> [294,]  0.0023328541
#> [295,] -0.0130878423
#> [296,]  0.0149118887
#> [297,]  0.0128667695
#> [298,]  0.0430241688
#> [299,] -0.0143187563
#> [300,]  0.0044628767
#> [301,] -0.0243798140
#> [302,] -0.0206240318
#> [303,] -0.0043487728
#> [304,]  0.0313727285
#> [305,] -0.0423568445
#> [306,]  0.0010548269
#> [307,]  0.0006496763
#> [308,]  0.0024959060
#> [309,]  0.0079855336
#> [310,] -0.0123126329
#> [311,]  0.0042613463
#> [312,]  0.0414011640
#> [313,] -0.0006285924
#> [314,]  0.0025430988
#> [315,] -0.0426988430
#> [316,] -0.0421615363
#> [317,]  0.0132318267
#> [318,] -0.0466696334
#> [319,] -0.0381274384
#> [320,]  0.0132470521
#> [321,] -0.0021247955
#> [322,] -0.0453311742
#> [323,]  0.0213985977
#> [324,]  0.0243783228
#> [325,]  0.0044371279
#> [326,]  0.0043813414
#> [327,]  0.0180668992
#> [328,] -0.0206689905
#> [329,] -0.0256352488
#> [330,]  0.0212119147
#> [331,] -0.0097237010
#> [332,] -0.0313594255
#> [333,]  0.0109449895
#> [334,] -0.0033475545
#> [335,]  0.0386246506
#> [336,] -0.0446871840
#> [337,]  0.0402139632
#> [338,] -0.0127152335
#> [339,] -0.0414065541
#> [340,]  0.0302830698
#> [341,] -0.0372908417
#> [342,]  0.0277323624
#> [343,]  0.0431726661
#> [344,] -0.0251979483
#> [345,]  0.0287492317
#> [346,]  0.0066199879
#> [347,] -0.0483327051
#> [348,]  0.0357602461
#> [349,] -0.0237214818
#> [350,] -0.0463125654
#> [351,] -0.0298931180
#> [352,] -0.0498597955
#> [353,]  0.0498459771
#> [354,] -0.0469167612
#> [355,] -0.0229818524
#> [356,]  0.0283170298
#> [357,] -0.0256745207
#> [358,] -0.0342376382
#> [359,] -0.0452259197
#> [360,]  0.0384821494
#> [361,] -0.0076103407
#> [362,]  0.0369772738
#> [363,] -0.0314044803
#> [364,]  0.0197319452
#> [365,] -0.0414765459
#> [366,] -0.0477895960
#> [367,] -0.0459068536
#> [368,] -0.0185632379
#> [369,] -0.0345583691
#> [370,]  0.0376880390
#> [371,] -0.0183102126
#> [372,] -0.0464615169
#> [373,] -0.0308249471
#> [374,]  0.0477240694
#> [375,] -0.0427803583
#> [376,]  0.0302785476
#> [377,]  0.0466766430
#> [378,] -0.0421236453
#> [379,]  0.0232422486
#> [380,]  0.0065203519
#> [381,] -0.0272725250
#> [382,]  0.0246959622
#> [383,]  0.0273174587
#> [384,]  0.0295885132
#> [385,]  0.0476735735
#> [386,] -0.0172092482
#> [387,]  0.0248672731
#> [388,] -0.0302728377
#> [389,]  0.0442896850
#> [390,] -0.0436296382