Generate points sampled from the Dini surface with optional noise.

This function generates points sampled from the Dini surface along with optional noise.

Usage

dini_surface_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
dini_surface_3d(n = 100, num_noise = 2, min_n = -0.05, max_n = 0.05)
#>                   x            y           z                           
#>   [1,]  0.347151898 -0.839324583 11.36041584  0.046811273 -0.0397026339
#>   [2,] -0.648741637 -0.085146601  3.04097530 -0.012496258  0.0290763175
#>   [3,]  0.330471890  0.942706226  1.23365770 -0.039036306  0.0422360181
#>   [4,]  0.253389045 -0.761222512 11.22547072  0.012114244 -0.0214310022
#>   [5,] -0.246514057  0.864733938  8.10003919 -0.007777032 -0.0499433291
#>   [6,]  0.523088728 -0.506932653  5.35988562 -0.043311664 -0.0356851794
#>   [7,]  0.238064969  0.239539384 -0.01913971  0.009618158  0.0015637302
#>   [8,] -0.309640512  0.877870834  8.17538241  0.040446538 -0.0123396746
#>   [9,] -0.504711391 -0.628465947  3.94710645 -0.049745073 -0.0284339315
#>  [10,]  0.960836104 -0.017447131 12.55560826 -0.015265306 -0.0350829806
#>  [11,] -0.198839966  0.953705115  8.06347944  0.036571148 -0.0353439124
#>  [12,]  0.061234707 -0.006712548  3.69283627  0.015919120  0.0423142502
#>  [13,]  0.297093445  0.921524373  7.54752564 -0.031443569 -0.0051962465
#>  [14,]  0.123363267 -0.046773958  4.19780591  0.044940477 -0.0389469245
#>  [15,] -0.744154289 -0.618122361  3.82910207 -0.028938895 -0.0086575911
#>  [16,] -0.084782350  0.284225698  7.21311486  0.013595432  0.0232040592
#>  [17,]  0.006424263 -0.192239598  3.39490002  0.002368991 -0.0245691278
#>  [18,] -0.002655535  0.024989829  4.58287196 -0.030572272 -0.0315889175
#>  [19,] -0.166027330  0.372715415  7.64092767 -0.018436977  0.0227344875
#>  [20,] -0.139770029 -0.265923474  9.59310212  0.045953311 -0.0366494890
#>  [21,]  0.196364538 -0.424593269 10.91956933  0.034806118  0.0157848667
#>  [22,] -0.360768984  0.782508943  1.95094799  0.004200594  0.0070135517
#>  [23,]  0.568905980 -0.535606381  5.42034856 -0.030929957  0.0208199039
#>  [24,]  0.369682516  0.844046466  1.13644660 -0.023326323 -0.0189224581
#>  [25,] -0.125956866  0.909672364  7.96871815  0.038856192  0.0338216087
#>  [26,] -0.240603232  0.930045959  8.11462122  0.036005257 -0.0460992394
#>  [27,]  0.309434329  0.190674646  6.09661582 -0.005547783  0.0447378346
#>  [28,]  0.751391236  0.605385261  6.96767707  0.010797826  0.0272497393
#>  [29,] -0.713505955  0.672832819  2.38807249  0.029635492 -0.0160971409
#>  [30,]  0.230500085 -0.889184617 11.27197392  0.009967849  0.0464448324
#>  [31,]  0.161964438 -0.238815265  4.35128491 -0.006457079  0.0179349351
#>  [32,] -0.207644560  0.219635589  1.41538757 -0.030954550 -0.0207382537
#>  [33,] -0.196956443  0.933042780  8.05239280 -0.025088984  0.0037443788
#>  [34,] -0.950579597  0.025083500  9.40888525  0.031934072 -0.0118583800
#>  [35,]  0.027861544  0.119450107  5.82954655 -0.035403466  0.0063740385
#>  [36,] -0.470847067 -0.871013280  4.21588312 -0.042859625 -0.0297865495
#>  [37,]  0.471775798 -0.879257440  5.20476096 -0.034428535 -0.0433482581
#>  [38,] -0.398186880 -0.886912406 10.57801342  0.034162961 -0.0163532625
#>  [39,] -0.165248625  0.097125167  7.53909915 -0.045406635  0.0359919606
#>  [40,] -0.296276634  0.275903293  1.75253207  0.012260502 -0.0370901504
#>  [41,] -0.217685789 -0.038905499  2.10401887  0.002252682 -0.0496382012
#>  [42,]  0.239945560  0.967977635  1.32767639  0.004865140  0.0114451926
#>  [43,]  0.927942486 -0.344833511 12.21152974  0.037104514 -0.0157686300
#>  [44,]  0.096544105 -0.332565945 10.49405157  0.019189041 -0.0482725061
#>  [45,]  0.266604143  0.921158428  1.28108500  0.016947191  0.0417446218
#>  [46,] -0.759561413 -0.420656930  9.88252661  0.023474528  0.0490760001
#>  [47,]  0.503428273 -0.616886581  5.30086775 -0.046460505 -0.0143416769
#>  [48,] -0.049884198 -0.359081746 10.11631943 -0.014047195 -0.0389693632
#>  [49,] -0.034107378  0.634374263  7.65403261 -0.002774406 -0.0050521928
#>  [50,] -0.222968284 -0.027781133  8.34986287  0.021978159 -0.0338490285
#>  [51,]  0.926140266 -0.068656559 12.47381130  0.001459893  0.0007207744
#>  [52,]  0.325817713  0.835905532  7.44967562 -0.032680244  0.0358561296
#>  [53,] -0.159867243 -0.943375641 10.81907884 -0.006952650 -0.0343432767
#>  [54,] -0.026417070 -0.007611457  0.13491526  0.049056897  0.0380912595
#>  [55,] -0.002278861 -0.900224652 10.96189757 -0.037429136 -0.0017651984
#>  [56,]  0.031801939 -0.963736493  4.73890062  0.040186682  0.0271230818
#>  [57,]  0.441012136  0.529629049  6.96699885  0.037823953 -0.0153674109
#>  [58,]  0.580660577  0.795721731  0.93866224 -0.013801423 -0.0435791666
#>  [59,] -0.269745990  0.603299931  1.76763793  0.014180800 -0.0007038843
#>  [60,]  0.842686227 -0.515966531  5.73501010  0.016110678 -0.0042624283
#>  [61,]  0.101253939  0.225475200  6.32567115  0.045699345  0.0307058573
#>  [62,] -0.611659325  0.383482421  8.70502431 -0.017281139 -0.0172181919
#>  [63,]  0.581112679  0.784110818  7.21974431 -0.022220930  0.0344564827
#>  [64,] -0.159062719 -0.824812214  4.45665921  0.048801365 -0.0145525717
#>  [65,]  0.250373296  0.929896049  1.31460602  0.007458667  0.0024064121
#>  [66,] -0.248982250  0.221961583  7.87738183 -0.033425671  0.0470545179
#>  [67,]  0.206743070 -0.185519380 10.84146120 -0.035560294 -0.0429631366
#>  [68,] -0.902179248  0.417544404  8.99088978  0.046699327  0.0314480813
#>  [69,] -0.263790988 -0.147055282  9.01982902  0.021227714 -0.0469257004
#>  [70,] -0.123016001  0.007107815  7.57633787  0.028738201  0.0067218661
#>  [71,] -0.708259840 -0.658926702  3.88528377 -0.048813058 -0.0365070718
#>  [72,] -0.138086697 -0.979505605  4.57127135 -0.025414831  0.0148092618
#>  [73,] -0.817085427 -0.540843602 10.01220380 -0.044960774 -0.0288158743
#>  [74,]  0.027892732  0.207217578  6.45103515 -0.032921567 -0.0393909458
#>  [75,]  0.037231031 -0.734813680 10.89943494  0.004026652 -0.0436995254
#>  [76,]  0.276868295  0.938332860  1.28691191  0.014284641 -0.0465217978
#>  [77,]  0.110050058  0.502392545  7.21171450 -0.032963758 -0.0291392233
#>  [78,]  0.903075425  0.384291500  6.68310622  0.049464466  0.0114881480
#>  [79,]  0.203913794 -0.493641331 10.99284738 -0.036069776 -0.0194828836
#>  [80,] -0.101160946  0.125700043  7.00777845  0.020289299 -0.0485348954
#>  [81,] -0.521790590 -0.282772514  9.61403054  0.036633093 -0.0097199444
#>  [82,]  0.537640400  0.831553151  0.99593473 -0.029914531 -0.0044783109
#>  [83,]  0.609149362  0.267060310  6.47756273  0.035608689  0.0366461525
#>  [84,] -0.008389469  0.185003897  6.51118830  0.030768446 -0.0395055116
#>  [85,]  0.419341104  0.361058381  6.62931578  0.022653194 -0.0053874930
#>  [86,] -0.612803648  0.374357486  2.42976126  0.014886355  0.0058034629
#>  [87,] -0.620761115 -0.275430442  3.35581853 -0.003475857  0.0060975001
#>  [88,] -0.159272977 -0.973026049  4.54856371  0.021506445 -0.0234173421
#>  [89,]  0.761310900 -0.253022040 12.15404297  0.031937469  0.0187736356
#>  [90,]  0.082504020 -0.057873506  9.96414135 -0.037160310 -0.0442129100
#>  [91,] -0.659966481  0.494136703  2.42332166  0.049801787 -0.0419374371
#>  [92,] -0.454909604  0.471484770  2.10817581  0.012772074 -0.0498900986
#>  [93,] -0.380370892 -0.874857030 10.59495885  0.005877741  0.0109741065
#>  [94,] -0.251025791 -0.476415627  3.83992116  0.025761254 -0.0074751925
#>  [95,] -0.133365763  0.134282790  0.98597769  0.027342288 -0.0305212979
#>  [96,] -0.729239088 -0.558827049  3.77276735 -0.035392174 -0.0172342707
#>  [97,]  0.133839219 -0.723466594  4.74854179  0.018060863 -0.0040695532
#>  [98,]  0.016980852  0.447810417  0.98552143 -0.004448415  0.0392953042
#>  [99,] -0.827555443 -0.555397120  3.73286173 -0.047147391  0.0337261188
#> [100,]  0.019795064 -0.065666173  8.91435430 -0.017987990  0.0190098491