Generate synthetic data with Gaussian clusters

Generate Gaussian Clusters

Usage

gau_clust(
  n,
  num_clust,
  mean_matrix,
  var_vec,
  num_dims,
  num_noise,
  min_n,
  max_n
)

Arguments

n

The total number of data points to be generated.

num_clust

The number of clusters to generate.

mean_matrix

A matrix where each row represents the mean vector for a cluster.

var_vec

A vector specifying the variance for each cluster.

num_dims

The number of effective dimensions for the data points.

num_noise

The number of additional noise dimensions to be generated.

min_n

The minimum value for the noise added to the data points.

max_n

The maximum value for the noise added to the data points.

Value

A matrix containing the generated Gaussian clusters.

Details

This function generates Gaussian clusters with specified parameters.

Examples

gau_clust(n = 300, num_clust = 5,
mean_matrix = rbind(c(1,0,0,0), c(0,1,0,0), c(0,0,1,0),
c(0,0,0,1), c(0,0,0,0)), var_vec = c(0.05, 0.05, 0.05, 0.05, 0.05),
num_dims = 4, num_noise = 2, min_n = -0.05, max_n = 0.05)
#>                 [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,]  0.9280164885  4.325582e-02 -7.414589e-02 -0.0239194943 -0.0335129507
#>   [2,]  1.0757310616 -6.187859e-02  3.229287e-02  0.0423999607 -0.0227104237
#>   [3,]  0.9207541191 -3.943306e-02 -4.298759e-02 -0.0203785936 -0.0145495149
#>   [4,]  0.9497093631  2.736934e-02  3.772809e-02 -0.0339020171  0.0453767560
#>   [5,]  1.0324805367  2.459450e-02  2.006465e-02  0.0140641238  0.0369651715
#>   [6,]  1.0120199996  6.708825e-02 -2.516072e-02  0.0612357471 -0.0211217351
#>   [7,]  0.9872969226 -3.964476e-02 -2.220404e-02  0.0036382880  0.0065015767
#>   [8,]  1.0246431814 -5.655426e-02 -7.443938e-02 -0.0603080322  0.0032997370
#>   [9,]  0.9412673714  6.368643e-02  4.146314e-02 -0.0096053756 -0.0164638333
#>  [10,]  0.9413177257  6.052839e-02  6.507429e-02  0.0737612331  0.0164832536
#>  [11,]  0.9731548082  3.977115e-02  1.133015e-02  0.0635603008 -0.0032750986
#>  [12,]  0.8796990947  4.119096e-02 -2.551764e-02 -0.0726163074 -0.0107397387
#>  [13,]  0.9358371284 -1.308109e-02 -4.789680e-02  0.0189684770  0.0484076125
#>  [14,]  1.0607100348 -2.732900e-02 -1.167070e-01  0.0823396203  0.0077196488
#>  [15,]  0.9373635437  5.900962e-04  8.113703e-02 -0.0469378987 -0.0482682618
#>  [16,]  1.0059304445 -3.753659e-02  4.604847e-02 -0.0295146058  0.0087127417
#>  [17,]  1.0348384045 -8.832603e-03 -7.204562e-02  0.0763612401 -0.0414046281
#>  [18,]  0.9857637790  6.764933e-02  5.830351e-02 -0.0348200758  0.0275565339
#>  [19,]  1.0633754783  3.027038e-02 -7.471144e-02  0.0477279097 -0.0237211133
#>  [20,]  1.1038756726  1.594419e-01 -3.298986e-02 -0.0285193384 -0.0218317830
#>  [21,]  0.9291723370 -3.901084e-02 -6.558772e-02 -0.0517358688 -0.0137451708
#>  [22,]  0.9659444441 -4.952092e-02  4.669362e-02  0.0262020297 -0.0093326652
#>  [23,]  0.9352889817  1.237956e-02  3.258471e-03 -0.0711370731 -0.0308207760
#>  [24,]  0.9627988859 -1.400352e-02 -9.312025e-02  0.0210895926 -0.0407365918
#>  [25,]  1.0567340486 -4.919905e-02 -5.013184e-02  0.0146268975 -0.0084594689
#>  [26,]  1.0179424894 -4.456485e-02 -2.168172e-03 -0.0546980358  0.0015143036
#>  [27,]  0.9961780891  1.521633e-02 -3.336577e-02  0.0518739838  0.0393897939
#>  [28,]  1.0301511576 -2.592352e-02 -1.205135e-02  0.0325277283  0.0097589007
#>  [29,]  1.0010258195 -9.928057e-03  3.230331e-02  0.0043294945 -0.0144068365
#>  [30,]  0.9668843802 -1.464404e-02  3.640205e-02 -0.0827167498 -0.0212392211
#>  [31,]  1.0160737132 -7.453414e-02  6.529392e-03  0.0006801850 -0.0315170588
#>  [32,]  0.9952534673  5.910990e-02 -3.428223e-02 -0.0005243633  0.0494128969
#>  [33,]  1.1132724313 -1.137760e-02 -5.458175e-03  0.0480512821  0.0285161850
#>  [34,]  0.9398566549 -5.308227e-02 -1.042463e-02 -0.0189188719 -0.0397074815
#>  [35,]  1.0139088842 -6.393043e-02  3.398869e-02  0.0469356729 -0.0312670548
#>  [36,]  0.9974141924 -1.519982e-02  1.244150e-02 -0.0629804335  0.0128167836
#>  [37,]  1.0046748738 -4.446161e-03 -1.092969e-01 -0.1109850450 -0.0259874856
#>  [38,]  1.0166328317 -7.057655e-02 -9.601357e-02 -0.0463572265  0.0020631232
#>  [39,]  0.9610313191 -7.749342e-02  3.318082e-02 -0.0078773039  0.0368970601
#>  [40,]  0.9287029311  5.095377e-02  3.316209e-02  0.0538010462  0.0155227508
#>  [41,]  1.0505753591 -5.411379e-02 -2.649385e-03  0.0075325200 -0.0128421906
#>  [42,]  1.0517798065 -1.143167e-02  4.048714e-02 -0.0768382477 -0.0334921471
#>  [43,]  0.9862006195  4.381972e-02 -5.496812e-02 -0.0136453924  0.0449838392
#>  [44,]  0.9682727247  7.394609e-02 -8.048512e-02  0.1132804099  0.0340834167
#>  [45,]  1.0071923497 -6.651883e-02 -6.608052e-02 -0.0198422519 -0.0268903531
#>  [46,]  0.9504726594  4.685305e-02  5.867291e-02 -0.0222870582 -0.0195922420
#>  [47,]  0.9483021689 -4.781533e-02  6.020721e-02 -0.0195793376  0.0181159695
#>  [48,]  1.0018540559  2.860348e-03 -1.907511e-01  0.1116311511 -0.0350119487
#>  [49,]  0.9826537227  3.439777e-02  2.757246e-02  0.0422433138 -0.0023087934
#>  [50,]  1.0557191702 -4.472855e-02 -4.496552e-03  0.0752171431  0.0034276390
#>  [51,]  1.0888413152 -2.367675e-02  1.070262e-03 -0.0708485215 -0.0314073884
#>  [52,]  0.9231363707 -9.253843e-02  6.604505e-02 -0.0597510126  0.0135080494
#>  [53,]  1.0523681726 -5.872496e-02  5.798857e-02  0.0154285945  0.0120050613
#>  [54,]  1.0373752823  1.074472e-03 -1.126748e-02  0.0730562165 -0.0483253102
#>  [55,]  0.9959416264 -3.555089e-05 -5.859290e-03 -0.0358823554 -0.0274884688
#>  [56,]  0.9748782312 -9.791667e-02 -2.174403e-02 -0.0308350835  0.0063555935
#>  [57,]  0.9908854750  1.916799e-03 -1.634493e-02 -0.0151267868 -0.0010147683
#>  [58,]  0.9855804520  2.044469e-02  1.677669e-02  0.0049101701  0.0357338799
#>  [59,]  1.0306251666  4.533867e-03 -1.077260e-02 -0.0439858127 -0.0476997707
#>  [60,]  0.9614847958 -2.418397e-03  5.610083e-02 -0.0576632632  0.0370597162
#>  [61,]  0.0314471423  9.172239e-01  3.885295e-02  0.0129926189  0.0470354715
#>  [62,]  0.0004053517  1.049243e+00 -1.164929e-01 -0.0394583435 -0.0365453299
#>  [63,]  0.0035088667  1.022637e+00 -7.561372e-02  0.0887844447  0.0493190964
#>  [64,] -0.0174242812  1.009175e+00  2.062828e-02  0.0321498102 -0.0135357156
#>  [65,] -0.1023726344  1.037872e+00  5.352750e-03  0.0493502800  0.0482663335
#>  [66,]  0.0512752545  1.044496e+00 -2.182749e-02  0.0205689446 -0.0238936082
#>  [67,]  0.0496982227  1.001097e+00 -1.898345e-02 -0.0756304757  0.0024227666
#>  [68,]  0.0313908043  9.954794e-01  3.329543e-02 -0.0320549307  0.0290466373
#>  [69,]  0.0152831392  1.005517e+00 -1.439615e-02 -0.0131613538 -0.0062587506
#>  [70,]  0.0451968905  9.520416e-01  4.663929e-02  0.0444655384 -0.0116734129
#>  [71,]  0.0180796360  1.008018e+00 -3.314791e-03 -0.1064043517 -0.0367553299
#>  [72,] -0.0455511180  9.666772e-01 -3.505021e-02  0.0279959148  0.0327502299
#>  [73,] -0.0435119940  1.086454e+00  3.897829e-02  0.0759432578  0.0308579141
#>  [74,]  0.0037472282  1.010519e+00 -2.244475e-02 -0.0755549797 -0.0049148934
#>  [75,]  0.0432651286  9.387707e-01 -6.027274e-02 -0.0388091424 -0.0314353233
#>  [76,]  0.0080815540  9.021463e-01 -3.247750e-04 -0.0911978044  0.0394896102
#>  [77,] -0.0383928381  1.056578e+00 -8.560150e-03 -0.0404229975  0.0003021493
#>  [78,] -0.0175322007  9.627371e-01 -6.494033e-02 -0.1409108239 -0.0444248850
#>  [79,] -0.0301071697  1.042312e+00 -1.495579e-01  0.0220597047  0.0094964331
#>  [80,]  0.0179635929  9.914734e-01  7.941264e-02  0.0347438677 -0.0167258734
#>  [81,] -0.0656607211  1.039430e+00  1.873113e-03 -0.0175316373  0.0279281422
#>  [82,]  0.0070798877  9.989103e-01  8.878713e-02 -0.0467549115  0.0004792311
#>  [83,] -0.0359744260  1.012863e+00 -1.809806e-02 -0.0089460481 -0.0459862355
#>  [84,] -0.0120344728  9.866974e-01 -4.178070e-02  0.0480144748  0.0154677605
#>  [85,] -0.0200178475  1.029515e+00  6.095893e-02  0.0110108956 -0.0122127509
#>  [86,] -0.0387841729  1.004539e+00  7.564338e-03  0.0148477075 -0.0139178443
#>  [87,] -0.0235115871  1.030131e+00 -6.428501e-02  0.0723915748  0.0236699918
#>  [88,] -0.0329259893  9.645473e-01  4.361313e-03 -0.0609012066 -0.0231690730
#>  [89,]  0.0193813570  9.903084e-01 -2.821057e-02 -0.0375394364  0.0340812165
#>  [90,] -0.0802913284  9.553207e-01  2.437492e-02 -0.0164416283 -0.0190609426
#>  [91,] -0.0002340421  1.024313e+00  7.879939e-02  0.1445636413 -0.0114888639
#>  [92,] -0.0084983236  9.571175e-01  2.705119e-03  0.0300741296 -0.0299448569
#>  [93,] -0.0503926997  1.003230e+00  3.625569e-02  0.0009549090  0.0176922428
#>  [94,]  0.0205812579  1.049734e+00  2.599320e-02  0.1003334090 -0.0464076540
#>  [95,]  0.0125625357  9.449563e-01  2.395486e-02  0.0481381512  0.0181245120
#>  [96,] -0.0725268622  9.499159e-01  4.779848e-02 -0.0358874023  0.0022451034
#>  [97,] -0.0076847392  8.974502e-01 -3.273170e-02  0.0331881123 -0.0406797413
#>  [98,] -0.0115420783  1.010987e+00  3.869620e-02  0.0684679133 -0.0358975386
#>  [99,] -0.0588116831  8.743303e-01 -1.155646e-02 -0.0191124902  0.0161839053
#> [100,] -0.0008222242  1.038242e+00  1.468941e-02  0.0012079049 -0.0334079372
#> [101,]  0.0299929544  9.818917e-01  2.593600e-03  0.0401501921 -0.0318741915
#> [102,]  0.0182449778  9.382141e-01  1.930741e-02 -0.0762816045  0.0480578174
#> [103,]  0.0255711402  9.784430e-01 -6.688215e-02  0.0316238996  0.0458056278
#> [104,] -0.0199413331  9.883319e-01  1.115981e-02 -0.0046263626  0.0067189096
#> [105,] -0.0267063816  9.593530e-01  5.248756e-02 -0.0012014333 -0.0255607039
#> [106,]  0.0507221907  1.006828e+00  5.022774e-02 -0.0116750082  0.0195360137
#> [107,]  0.0535963273  1.058522e+00  7.507209e-03 -0.0847659602  0.0115832688
#> [108,]  0.0042148805  9.058209e-01  2.241274e-02 -0.0107069951  0.0478555646
#> [109,]  0.0607021469  1.026628e+00  5.232515e-02 -0.0332881664  0.0045113476
#> [110,] -0.0502286036  9.314501e-01 -4.189490e-04  0.0783083686  0.0225193002
#> [111,]  0.0472858901  1.033560e+00 -2.599986e-02  0.1205018757  0.0449829327
#> [112,]  0.0403508895  9.387340e-01 -5.259448e-03 -0.0535030142 -0.0294302615
#> [113,]  0.1084280459  1.023817e+00 -5.053290e-03 -0.0497777649 -0.0017663328
#> [114,]  0.0249844703  1.051962e+00  3.990810e-02 -0.0528922430 -0.0241870475
#> [115,]  0.0391645266  1.006481e+00 -4.130807e-02 -0.0227221087 -0.0334173137
#> [116,]  0.0666532259  9.182217e-01 -8.332117e-03  0.0057331400 -0.0298237236
#> [117,]  0.0250096640  9.967693e-01  8.556767e-04  0.0227442525 -0.0479430120
#> [118,] -0.0217683332  1.110122e+00 -3.505756e-02 -0.0184024342 -0.0344209968
#> [119,]  0.0582945944  1.098371e+00 -6.409558e-02 -0.0887920104 -0.0379975907
#> [120,]  0.0024605892  9.694477e-01 -6.572609e-02  0.0105241373 -0.0221230407
#> [121,] -0.0588844959 -2.006915e-02  9.638161e-01  0.1025192758 -0.0298768586
#> [122,] -0.0187412248  4.151705e-02  1.077708e+00  0.0462480335  0.0258583854
#> [123,]  0.0688132676 -7.381334e-02  1.056261e+00 -0.0666031548 -0.0262802163
#> [124,]  0.0179494147 -1.866655e-02  9.150053e-01 -0.0453446578  0.0271878499
#> [125,] -0.0386715598 -3.835756e-02  9.230783e-01  0.0595581282 -0.0104632955
#> [126,] -0.0309609001 -3.153721e-02  1.010099e+00 -0.0635281871 -0.0460280895
#> [127,]  0.0741477642 -3.306650e-02  9.927931e-01 -0.0376106090  0.0363943189
#> [128,]  0.0428411819 -8.704408e-02  9.853896e-01  0.0142233557 -0.0204666682
#> [129,]  0.0238049562  2.252586e-02  1.038673e+00  0.0282696497 -0.0042719046
#> [130,] -0.0014180219 -1.455086e-02  1.024224e+00  0.0510449744  0.0065861673
#> [131,]  0.0357477414  4.661200e-02  9.982371e-01  0.0522467971  0.0022973729
#> [132,] -0.0703868189 -1.040858e-01  9.372918e-01 -0.0122798325 -0.0341714122
#> [133,] -0.0332667857 -2.018351e-02  9.359366e-01 -0.0008476542 -0.0458459044
#> [134,]  0.0408914749 -4.663242e-02  1.025284e+00  0.0107660759 -0.0492608209
#> [135,] -0.0226762745 -1.793652e-02  9.824555e-01  0.0098711927 -0.0344803523
#> [136,]  0.0071006681  5.909041e-02  9.744253e-01  0.0330153384 -0.0251906292
#> [137,]  0.1180064052  3.333890e-02  1.030161e+00 -0.0295663678  0.0468415777
#> [138,] -0.0275323704 -9.522402e-02  9.955421e-01 -0.0131675350  0.0236520866
#> [139,]  0.0813240325 -4.533771e-03  9.568872e-01  0.0318452565  0.0233393101
#> [140,] -0.0420803977 -5.344467e-03  1.004140e+00  0.0320970794 -0.0198692791
#> [141,] -0.0230652388 -7.377240e-02  9.292117e-01  0.0865158331  0.0383671386
#> [142,]  0.0260384718 -1.808055e-02  1.038860e+00  0.0167506555  0.0184754258
#> [143,]  0.0261156610  2.188952e-02  1.074035e+00  0.1366815425 -0.0187885791
#> [144,] -0.0436218196 -1.434120e-03  9.808474e-01  0.0508027952  0.0371012360
#> [145,]  0.0709384254 -1.127076e-02  1.020101e+00  0.1192191209 -0.0267639059
#> [146,]  0.1213934769 -2.069400e-02  9.579761e-01  0.0483250556 -0.0405898977
#> [147,] -0.0189810881  9.419412e-02  1.003366e+00 -0.0368223200 -0.0163894184
#> [148,]  0.0082991375  1.169027e-02  9.042732e-01 -0.0168255690  0.0464976184
#> [149,] -0.0015363341  4.426019e-02  9.519939e-01 -0.1003518964  0.0266485065
#> [150,]  0.0265355872  3.760285e-02  9.672308e-01 -0.0351208530  0.0017750734
#> [151,]  0.0158884578  7.268193e-02  9.705572e-01 -0.0643739439  0.0257199176
#> [152,] -0.0662631598  5.654490e-02  9.668595e-01 -0.1203355892 -0.0028112540
#> [153,] -0.0685775879  2.025670e-02  1.024797e+00  0.0004705105 -0.0358714731
#> [154,]  0.1024857688 -1.692310e-02  1.064716e+00  0.0557980509  0.0431336456
#> [155,]  0.0071969183 -7.031979e-02  1.044179e+00 -0.0390737963 -0.0431078509
#> [156,]  0.0481740138  4.310428e-02  1.080978e+00 -0.0819277106  0.0231260299
#> [157,]  0.0003558202  4.433821e-03  9.709099e-01  0.0595554910  0.0421890770
#> [158,] -0.0275718959  9.274957e-02  1.124215e+00  0.0785894964 -0.0465954216
#> [159,] -0.0272602435 -3.036498e-02  9.986387e-01  0.0153546349 -0.0141594660
#> [160,]  0.0902689760  3.794084e-02  1.038859e+00 -0.0115474624  0.0097688890
#> [161,] -0.0826616118 -9.404587e-02  9.774842e-01  0.0109096939  0.0050847116
#> [162,]  0.0246024157 -6.959414e-02  9.955509e-01 -0.0377185149  0.0358095456
#> [163,]  0.0567932229 -1.123062e-02  9.427720e-01  0.0543742606 -0.0279361261
#> [164,] -0.0062815583 -7.444756e-02  1.011945e+00  0.0262562711 -0.0491138017
#> [165,] -0.0555025260 -1.283484e-03  8.887080e-01 -0.0607064560 -0.0088877988
#> [166,]  0.0183313819 -6.433080e-03  1.015070e+00 -0.0363419066 -0.0103972674
#> [167,]  0.0482343433  2.842760e-02  9.360555e-01  0.0028403909 -0.0074727983
#> [168,] -0.0304130409 -3.738906e-02  1.030167e+00 -0.0146127858  0.0410869173
#> [169,] -0.0484042650  8.400665e-02  1.056601e+00  0.0381284456 -0.0257261133
#> [170,]  0.0798920640  7.074272e-02  9.588911e-01  0.0293478073  0.0046803454
#> [171,]  0.0567466173  3.223194e-02  9.519454e-01 -0.0105516000 -0.0247436369
#> [172,]  0.0585380173 -4.299515e-02  9.075619e-01  0.0606769053  0.0095136625
#> [173,] -0.0421660194 -4.523800e-02  1.020157e+00 -0.0836102502  0.0312118863
#> [174,] -0.0700292996  3.009317e-02  9.977280e-01 -0.0309290474 -0.0021993788
#> [175,]  0.0170843760 -1.504726e-02  1.007841e+00  0.0623046704 -0.0142282342
#> [176,]  0.0615362709  5.548426e-02  9.765687e-01 -0.0061471547  0.0167430949
#> [177,] -0.0320296240 -2.307241e-02  1.026758e+00  0.0500513847  0.0417515834
#> [178,]  0.0155475785  1.194608e-02  9.928306e-01 -0.0397775274 -0.0385170716
#> [179,] -0.0337607410 -2.028807e-02  1.058394e+00 -0.0435529472 -0.0182151369
#> [180,] -0.1023846695  4.367596e-02  1.019442e+00 -0.0363096936 -0.0276838471
#> [181,]  0.0101763079  4.109698e-02  3.471407e-02  0.9941177632 -0.0056945411
#> [182,]  0.0051646263 -2.981242e-02 -2.556690e-02  1.0777197532 -0.0379620828
#> [183,]  0.0144427763  1.262526e-02 -2.230307e-03  0.9894020916 -0.0482062794
#> [184,] -0.0159554860 -3.213687e-02  4.792637e-02  1.0272619066 -0.0033677013
#> [185,]  0.0352691237 -6.536613e-02 -7.183666e-02  1.0365901517  0.0448091109
#> [186,] -0.0344784280 -2.894394e-02 -6.603751e-03  1.0145095800 -0.0343776602
#> [187,] -0.0114203434  8.552134e-02  2.281094e-02  0.9870250708  0.0337906945
#> [188,]  0.0182104593  3.375584e-02 -1.169100e-01  1.0784310751  0.0360552783
#> [189,]  0.0408991082  6.318189e-02 -3.959010e-02  1.0342512553  0.0260930113
#> [190,] -0.0177245603 -1.779471e-02  6.009509e-03  0.9134857991 -0.0292790123
#> [191,]  0.0186855781  1.469800e-02 -3.298064e-02  1.0392892120 -0.0337313221
#> [192,]  0.0181071106  3.300865e-02 -1.127853e-01  1.0377129544  0.0466257663
#> [193,]  0.0814045604 -2.246769e-02 -2.066966e-03  0.9667319134 -0.0205003564
#> [194,] -0.0545507313  8.786610e-02 -5.051003e-02  0.9536658667 -0.0486953856
#> [195,] -0.0494693954 -3.305502e-02  1.525061e-02  1.0316965358  0.0202699446
#> [196,]  0.0496283328 -1.409364e-02  3.612645e-03  0.9943688553 -0.0489494331
#> [197,]  0.0181654956  8.830452e-02  2.605421e-02  1.0189184080 -0.0497890328
#> [198,] -0.0247902191 -3.422211e-02  3.558673e-02  1.0266826286  0.0036340442
#> [199,] -0.0076953604 -6.369157e-02  8.276278e-02  0.9134191328  0.0039346478
#> [200,] -0.0742033262  5.458261e-02  1.087426e-03  1.0008752040  0.0045905215
#> [201,]  0.0593334544 -3.621683e-02  9.045034e-03  1.0572096024 -0.0195735375
#> [202,]  0.0510221348  4.079255e-02 -4.494610e-02  1.0717129724 -0.0011711374
#> [203,] -0.0348466498  1.990513e-02  3.179119e-02  1.0234767523 -0.0315791002
#> [204,] -0.0579674548  3.029122e-02 -1.314949e-02  1.0177257269  0.0179879040
#> [205,] -0.0634095658 -2.538004e-02 -1.945457e-02  0.9792349157 -0.0416397863
#> [206,] -0.0855531800 -7.558958e-02  1.926237e-02  0.9932132816  0.0148476969
#> [207,] -0.0198543236 -2.100111e-02 -2.680907e-02  1.0479648373  0.0106872817
#> [208,]  0.0118642349  4.055579e-02 -4.519000e-02  0.9954414626 -0.0335448029
#> [209,] -0.0463404719 -4.416378e-02 -1.718143e-02  0.9729184302 -0.0162004506
#> [210,]  0.0128622118  9.235620e-02  3.687563e-02  1.0172278532  0.0153433550
#> [211,] -0.0614393309 -1.142524e-01  8.237897e-03  1.0503776784 -0.0075180754
#> [212,] -0.0665688074  4.765753e-02  1.617173e-02  0.9584606142 -0.0432969442
#> [213,]  0.0418354056  3.216166e-02  2.499342e-02  1.0389950282  0.0482755577
#> [214,]  0.0629706116 -3.113616e-02 -4.661217e-03  1.0548594792  0.0124723578
#> [215,]  0.0664150073 -7.307589e-02  8.832870e-02  0.9884071972 -0.0258232572
#> [216,] -0.0657313035  5.374571e-02  6.842096e-02  0.9651770137 -0.0186904289
#> [217,]  0.0906508480  5.655736e-02  1.190922e-01  1.0024229598  0.0382656358
#> [218,] -0.0200515922  3.989761e-02  3.490249e-02  1.0067260235 -0.0442257831
#> [219,] -0.0176794801  1.953159e-02 -2.098456e-02  0.9300468534  0.0383949689
#> [220,] -0.0056558890 -4.106922e-02 -1.135802e-03  0.9680419983  0.0305386713
#> [221,] -0.0674325853 -4.357519e-02 -3.861304e-03  1.0488880063 -0.0084166751
#> [222,]  0.0225483093 -3.440776e-02 -1.343849e-02  1.0178981668  0.0459982871
#> [223,]  0.0588849465 -5.369970e-02 -6.890358e-02  1.0425293906 -0.0159762019
#> [224,]  0.0224385986  4.434374e-02 -6.561643e-02  1.0361917710  0.0445920724
#> [225,]  0.0391930078 -3.137654e-02  2.653332e-02  1.0131702805 -0.0269456885
#> [226,]  0.0578043368 -2.598819e-02 -1.064953e-01  1.0280840821 -0.0039294525
#> [227,] -0.0471987990  1.868303e-02  6.839490e-02  1.0187743594 -0.0485232820
#> [228,]  0.0978727608  5.739664e-02  4.531123e-02  0.9651270400  0.0145281254
#> [229,]  0.0780986441 -6.177438e-03 -8.609664e-03  0.9660488787  0.0055031747
#> [230,]  0.0624266971  2.638017e-02  6.356080e-02  0.9814309979 -0.0254448388
#> [231,] -0.0157485810 -1.594076e-02  1.315081e-02  1.0837004414 -0.0143058365
#> [232,] -0.0961194531 -3.076117e-02  1.342681e-02  0.9622903077 -0.0200839619
#> [233,] -0.0354287541  4.540045e-03  3.552309e-02  0.9760500815  0.0316136293
#> [234,]  0.0381269717 -5.659138e-02 -2.295201e-02  0.9526051398  0.0408197138
#> [235,]  0.0522935301 -1.019959e-01 -9.551791e-02  1.0185895701 -0.0477582282
#> [236,]  0.0224302247 -3.241537e-02 -5.263049e-02  1.0748998370  0.0394027881
#> [237,]  0.0229842545 -4.437411e-02  3.091204e-02  1.0580717550  0.0477805450
#> [238,]  0.0010619599 -5.028936e-02  9.715312e-02  0.9912301678  0.0221653899
#> [239,] -0.0198484286  1.920640e-02  5.439138e-02  1.0390577462 -0.0304659744
#> [240,]  0.0344773794  2.781446e-02 -5.724851e-06  0.9935659885  0.0492662383
#> [241,] -0.0071384541 -2.979967e-02  3.473325e-02  0.0720379734 -0.0443670713
#> [242,] -0.0651570680  9.708880e-02  6.502526e-02 -0.0253271233 -0.0366280209
#> [243,] -0.0778086184  8.958432e-03 -8.899029e-03  0.0291364299  0.0280639814
#> [244,] -0.0524744378 -9.081518e-03 -2.306106e-02  0.0147438036 -0.0323613488
#> [245,] -0.0148948261 -4.305253e-02  2.113523e-02  0.0180424756 -0.0048598156
#> [246,]  0.0528104921  1.235996e-03 -1.013347e-02  0.0028686774  0.0443913407
#> [247,]  0.0057767551  4.009136e-02  1.429893e-03 -0.0331300202 -0.0223842234
#> [248,] -0.0749632415 -8.889464e-02 -9.411370e-02  0.0517044847 -0.0110152914
#> [249,]  0.0072570884 -3.234861e-03  3.939737e-03 -0.0326770224  0.0170107631
#> [250,] -0.0475271103  2.165538e-02 -7.798351e-02  0.0384304466 -0.0039099276
#> [251,] -0.0078228502 -6.037944e-02 -5.172354e-02 -0.0404312985 -0.0399823405
#> [252,] -0.0021243972  3.456915e-02  2.110009e-02  0.1099300707 -0.0397211892
#> [253,]  0.0289686965 -7.066912e-02 -8.098304e-03  0.0551280858 -0.0335099643
#> [254,]  0.0259349760  3.901045e-02 -1.541185e-02 -0.0190059683 -0.0155101968
#> [255,]  0.0531591139  1.754258e-03 -2.138399e-02  0.0061663728 -0.0479575049
#> [256,]  0.0467457597 -2.694102e-02  3.635137e-02 -0.0503096870  0.0109781611
#> [257,]  0.0321167016 -3.393610e-02  2.017380e-03  0.0330170489 -0.0219537414
#> [258,]  0.0016499792  1.229334e-02  7.800049e-03  0.0264075460 -0.0037841405
#> [259,] -0.0796925342  4.285880e-02 -3.853635e-02 -0.0345223342  0.0389382321
#> [260,] -0.0806298082 -9.454678e-03 -1.780100e-02 -0.0433871824  0.0107888733
#> [261,]  0.0060838915  5.830170e-03 -4.927610e-02 -0.0027044866  0.0475900964
#> [262,] -0.0381472588  3.721889e-02 -7.198254e-03 -0.0564192240 -0.0362075793
#> [263,]  0.0236147357  3.769384e-02  2.339857e-02 -0.0673455178  0.0132063152
#> [264,] -0.0279632888  3.938616e-02  1.865379e-02 -0.0354169733  0.0274617570
#> [265,] -0.0245218589 -5.770139e-02 -7.132156e-02 -0.0554231352 -0.0313454578
#> [266,]  0.0230938591 -1.930200e-02  8.264145e-03  0.0480651167 -0.0431466313
#> [267,]  0.0111369474  2.600865e-02 -4.841756e-02  0.0729160129  0.0144451345
#> [268,]  0.0502912090  3.510804e-02  2.609448e-02  0.0392314940 -0.0360345155
#> [269,] -0.0099709460 -1.960128e-02 -4.360019e-02 -0.0004335347  0.0245018917
#> [270,] -0.0581317505 -1.207508e-02 -3.840130e-03  0.0227201623 -0.0323107038
#> [271,] -0.0627576187  3.241878e-02  6.295342e-02  0.0136616451 -0.0394650104
#> [272,]  0.0453677110 -2.152434e-02 -1.263083e-02 -0.0050985846  0.0387229109
#> [273,]  0.0511926722  9.414053e-04  6.678080e-02 -0.1080880018 -0.0027328381
#> [274,] -0.0376773525  6.660042e-02 -6.072783e-02 -0.0331316297 -0.0401556439
#> [275,] -0.0452987395 -1.363810e-02  2.075559e-02  0.0393485449  0.0485450992
#> [276,]  0.0030925491  5.294312e-02  4.997785e-02  0.0343808777  0.0474903853
#> [277,] -0.0518361216 -4.879549e-02  8.706260e-02  0.0235807266 -0.0484236226
#> [278,] -0.0092834635 -4.440815e-02 -6.069397e-02  0.1233305240  0.0353532954
#> [279,] -0.0005522207 -2.946347e-02 -6.309853e-02  0.0545422330 -0.0104928504
#> [280,] -0.0227669676  1.041717e-01 -3.357455e-02  0.0283063442  0.0231267287
#> [281,] -0.0302987904  7.073844e-02 -6.198611e-02 -0.0199734291 -0.0218063864
#> [282,] -0.0709724006 -4.925214e-02 -6.288856e-02  0.0298937638 -0.0430200680
#> [283,]  0.0404153232  6.533392e-02  5.912094e-02  0.0192482221  0.0346429918
#> [284,]  0.0141746063 -8.941066e-02 -2.730779e-02 -0.0437685473 -0.0447516432
#> [285,] -0.0342273302 -1.017835e-02 -1.931835e-02 -0.0567157504 -0.0255020827
#> [286,]  0.0187990848  2.202026e-03  3.440157e-02 -0.0137182308  0.0475480513
#> [287,]  0.0577982405 -4.095953e-03  8.272546e-02  0.0090223933  0.0424182639
#> [288,]  0.0680929419  7.369546e-02  2.003854e-03  0.0161309193 -0.0168923484
#> [289,] -0.0076615064  8.075072e-02  3.839897e-02 -0.0122436784 -0.0092278562
#> [290,]  0.0310188513  3.237385e-02 -7.261942e-02  0.0530277298 -0.0186258808
#> [291,]  0.0205893992 -1.441983e-01 -9.593262e-02  0.0302679106  0.0168366771
#> [292,] -0.0138173572  8.270807e-02  1.904266e-04 -0.0417722976 -0.0400909165
#> [293,] -0.0492942419  7.415863e-02 -1.719263e-01  0.0804064448  0.0089217677
#> [294,]  0.0211594205  1.901702e-02 -4.396991e-03  0.0022668343  0.0414992077
#> [295,]  0.0069151187  4.131840e-02 -4.038478e-02 -0.0293760667  0.0382307793
#> [296,]  0.0795561622  3.855190e-02  1.146987e-02 -0.0613835820  0.0133700853
#> [297,]  0.0607635078 -2.259708e-02 -1.869590e-02  0.0765721352  0.0079122259
#> [298,]  0.0489632214 -4.732139e-02  1.131152e-02  0.0590684497  0.0408982386
#> [299,]  0.0446737966  1.051909e-01  1.261553e-02 -0.0305699202 -0.0048856552
#> [300,]  0.0057146327 -1.057092e-02 -2.800429e-03 -0.0400875426  0.0107167482
#>                 [,6]
#>   [1,]  4.170031e-02
#>   [2,]  2.931139e-02
#>   [3,]  3.683148e-02
#>   [4,]  3.059514e-02
#>   [5,] -2.427545e-02
#>   [6,]  1.971803e-03
#>   [7,] -4.389804e-02
#>   [8,] -2.042662e-02
#>   [9,]  1.422400e-02
#>  [10,]  3.788155e-02
#>  [11,] -1.184369e-03
#>  [12,] -2.433468e-02
#>  [13,]  1.186147e-02
#>  [14,] -9.542299e-04
#>  [15,]  2.355417e-02
#>  [16,]  2.781313e-02
#>  [17,]  1.652885e-02
#>  [18,]  3.451066e-02
#>  [19,] -3.771809e-02
#>  [20,] -1.743023e-02
#>  [21,] -4.877004e-02
#>  [22,]  1.957432e-02
#>  [23,] -1.812262e-02
#>  [24,] -1.997900e-02
#>  [25,] -4.347485e-02
#>  [26,]  4.976748e-02
#>  [27,]  3.520306e-02
#>  [28,] -2.568291e-02
#>  [29,] -3.700170e-02
#>  [30,]  4.082612e-02
#>  [31,] -2.290931e-02
#>  [32,]  1.078730e-02
#>  [33,]  1.552356e-02
#>  [34,]  1.993202e-02
#>  [35,]  2.253841e-02
#>  [36,] -3.592471e-02
#>  [37,]  1.774518e-02
#>  [38,]  2.969422e-02
#>  [39,]  1.720102e-02
#>  [40,] -2.768736e-03
#>  [41,]  2.053617e-02
#>  [42,]  2.448044e-02
#>  [43,]  4.150173e-02
#>  [44,] -4.350289e-03
#>  [45,]  4.240633e-02
#>  [46,] -3.862054e-03
#>  [47,] -3.754765e-02
#>  [48,] -2.755982e-02
#>  [49,]  7.816351e-03
#>  [50,]  1.188055e-03
#>  [51,]  3.225209e-02
#>  [52,]  4.139099e-02
#>  [53,] -2.528602e-02
#>  [54,] -3.683380e-02
#>  [55,]  9.981048e-04
#>  [56,] -4.333262e-02
#>  [57,] -3.550264e-02
#>  [58,] -9.188397e-03
#>  [59,]  3.273575e-02
#>  [60,]  6.908395e-03
#>  [61,]  3.460353e-03
#>  [62,] -1.214385e-02
#>  [63,]  4.091615e-02
#>  [64,]  4.957481e-02
#>  [65,]  1.192004e-02
#>  [66,]  4.729167e-02
#>  [67,]  4.609774e-02
#>  [68,]  3.423213e-02
#>  [69,]  3.796943e-02
#>  [70,]  4.915905e-02
#>  [71,] -3.671333e-03
#>  [72,] -2.604835e-02
#>  [73,]  4.350584e-02
#>  [74,] -2.398824e-02
#>  [75,] -1.471053e-02
#>  [76,]  1.517502e-03
#>  [77,] -7.586698e-04
#>  [78,]  4.769884e-02
#>  [79,] -3.589216e-02
#>  [80,]  3.996386e-02
#>  [81,]  3.800871e-02
#>  [82,]  1.290704e-02
#>  [83,]  4.012701e-02
#>  [84,] -1.195618e-02
#>  [85,] -1.910311e-02
#>  [86,]  1.773181e-02
#>  [87,]  5.965866e-03
#>  [88,]  1.632462e-05
#>  [89,]  3.984735e-02
#>  [90,] -3.327592e-02
#>  [91,] -4.928121e-02
#>  [92,] -3.547999e-02
#>  [93,]  4.034566e-02
#>  [94,] -1.255991e-03
#>  [95,]  6.717845e-03
#>  [96,]  3.698962e-02
#>  [97,]  2.540728e-02
#>  [98,]  4.931073e-04
#>  [99,]  9.545196e-03
#> [100,]  4.791655e-02
#> [101,]  3.618644e-03
#> [102,]  1.428866e-02
#> [103,] -2.732810e-02
#> [104,] -5.986790e-03
#> [105,] -2.388130e-02
#> [106,]  6.699400e-03
#> [107,]  1.633731e-02
#> [108,]  4.711635e-02
#> [109,] -4.068794e-03
#> [110,]  5.303390e-03
#> [111,]  2.146915e-02
#> [112,]  4.629135e-02
#> [113,]  1.022218e-02
#> [114,]  7.679531e-03
#> [115,]  3.687274e-02
#> [116,] -2.112487e-02
#> [117,]  1.553189e-02
#> [118,] -9.687334e-03
#> [119,]  2.735116e-02
#> [120,]  2.805554e-02
#> [121,]  2.473059e-02
#> [122,] -5.138909e-03
#> [123,]  5.798060e-03
#> [124,] -4.570844e-02
#> [125,]  1.483800e-02
#> [126,]  2.795363e-02
#> [127,]  3.528730e-03
#> [128,]  1.538313e-02
#> [129,] -4.254627e-02
#> [130,] -8.812034e-03
#> [131,]  1.665997e-02
#> [132,] -2.370223e-02
#> [133,] -3.664694e-02
#> [134,] -7.254663e-03
#> [135,] -2.662971e-02
#> [136,]  9.605217e-03
#> [137,]  1.691141e-02
#> [138,]  3.138721e-02
#> [139,] -2.129845e-02
#> [140,]  1.827852e-02
#> [141,]  3.086078e-02
#> [142,] -9.209580e-03
#> [143,]  3.535347e-04
#> [144,]  3.886066e-02
#> [145,] -3.494428e-02
#> [146,] -4.301872e-02
#> [147,] -1.843362e-02
#> [148,]  1.692450e-02
#> [149,] -4.437626e-02
#> [150,] -2.736081e-02
#> [151,]  1.811500e-03
#> [152,] -4.514844e-02
#> [153,]  3.755940e-02
#> [154,] -1.442789e-02
#> [155,] -8.361923e-03
#> [156,] -3.539813e-02
#> [157,]  4.098421e-02
#> [158,]  1.056707e-03
#> [159,]  1.438508e-02
#> [160,] -1.934492e-02
#> [161,] -2.182761e-02
#> [162,]  2.829927e-02
#> [163,]  3.640622e-02
#> [164,]  4.304305e-02
#> [165,]  4.143590e-02
#> [166,]  3.854417e-02
#> [167,] -1.562034e-02
#> [168,]  2.371965e-02
#> [169,] -1.106202e-02
#> [170,]  2.292063e-03
#> [171,] -1.902271e-02
#> [172,]  3.044690e-02
#> [173,] -6.349070e-03
#> [174,] -2.862139e-02
#> [175,]  4.278958e-02
#> [176,] -2.573674e-02
#> [177,] -2.961052e-02
#> [178,]  3.086667e-02
#> [179,] -4.483364e-02
#> [180,]  2.911366e-02
#> [181,] -2.537971e-02
#> [182,]  4.720916e-02
#> [183,] -1.632137e-02
#> [184,] -4.226086e-02
#> [185,] -2.472486e-02
#> [186,] -3.545619e-02
#> [187,]  6.263575e-03
#> [188,] -4.737862e-02
#> [189,]  4.003272e-02
#> [190,]  2.176399e-02
#> [191,]  4.878464e-02
#> [192,] -4.736570e-02
#> [193,]  3.723686e-02
#> [194,] -2.861491e-02
#> [195,]  4.474612e-02
#> [196,] -3.947237e-02
#> [197,]  4.978822e-02
#> [198,]  2.034996e-02
#> [199,]  3.089781e-02
#> [200,]  3.002211e-02
#> [201,]  3.967309e-02
#> [202,]  2.903844e-02
#> [203,] -3.938450e-02
#> [204,] -1.782169e-02
#> [205,]  1.745397e-02
#> [206,] -3.703651e-02
#> [207,]  1.402024e-02
#> [208,] -1.755235e-02
#> [209,]  1.546504e-02
#> [210,]  3.569260e-02
#> [211,] -7.079936e-03
#> [212,]  2.415316e-02
#> [213,] -4.318372e-02
#> [214,]  4.744266e-02
#> [215,] -3.285588e-02
#> [216,] -1.601250e-02
#> [217,] -2.891279e-02
#> [218,] -4.615502e-02
#> [219,] -7.674146e-03
#> [220,]  4.221944e-02
#> [221,]  3.602493e-02
#> [222,] -2.152802e-02
#> [223,] -3.063262e-02
#> [224,] -6.661502e-03
#> [225,] -3.987123e-02
#> [226,] -4.292018e-02
#> [227,]  1.950430e-02
#> [228,]  1.568766e-02
#> [229,] -1.927206e-02
#> [230,]  1.431463e-02
#> [231,]  3.581241e-02
#> [232,] -4.569650e-02
#> [233,]  2.470983e-02
#> [234,]  2.235792e-02
#> [235,]  2.810654e-02
#> [236,]  4.032526e-02
#> [237,] -4.086379e-02
#> [238,]  8.896635e-03
#> [239,] -4.409569e-02
#> [240,]  3.853797e-03
#> [241,] -3.946033e-02
#> [242,]  3.805131e-02
#> [243,]  1.875138e-02
#> [244,] -1.301387e-02
#> [245,] -7.650242e-03
#> [246,] -3.136376e-02
#> [247,] -3.212666e-02
#> [248,]  2.186836e-03
#> [249,]  2.526589e-03
#> [250,] -4.891272e-02
#> [251,]  1.896192e-02
#> [252,] -4.522076e-03
#> [253,]  2.331952e-02
#> [254,] -2.776884e-02
#> [255,]  2.853045e-02
#> [256,] -2.884202e-02
#> [257,] -3.656561e-02
#> [258,] -3.870351e-02
#> [259,] -3.497783e-02
#> [260,] -1.234848e-02
#> [261,] -2.212458e-02
#> [262,] -3.222789e-02
#> [263,]  3.248357e-02
#> [264,]  3.566588e-02
#> [265,] -2.164523e-02
#> [266,] -2.571727e-02
#> [267,] -4.658801e-02
#> [268,]  7.538942e-04
#> [269,] -3.367734e-02
#> [270,]  7.124091e-03
#> [271,] -4.613403e-02
#> [272,]  6.005648e-03
#> [273,] -4.393422e-02
#> [274,]  1.487254e-02
#> [275,] -4.189037e-02
#> [276,]  2.924958e-02
#> [277,]  3.448282e-02
#> [278,] -3.188087e-02
#> [279,] -2.777633e-02
#> [280,] -3.768772e-02
#> [281,]  1.795989e-02
#> [282,] -1.989364e-02
#> [283,] -4.356128e-02
#> [284,]  4.290471e-02
#> [285,] -4.040500e-02
#> [286,]  2.061358e-02
#> [287,] -3.668830e-02
#> [288,]  1.008019e-02
#> [289,]  4.231227e-02
#> [290,]  4.079620e-02
#> [291,] -3.376870e-02
#> [292,] -1.519013e-02
#> [293,] -3.619677e-02
#> [294,]  2.285319e-02
#> [295,] -2.201511e-02
#> [296,] -4.739333e-02
#> [297,] -1.344900e-02
#> [298,]  1.825578e-02
#> [299,]  3.116117e-02
#> [300,]  4.126162e-02