Generate Background Noise Data

This function generates background noise data with specified parameters such as the number of samples, number of dimensions, mean, and standard deviation.

Usage

gen_bkg_noise(n, num_dims, mean, sd)

Arguments

n

Number of samples to generate.

num_dims

Number of dimensions (columns) of the data.

mean

Mean of the normal distribution used to generate noise (default is 0).

sd

Standard deviation of the normal distribution used to generate noise (default is 1).

Value

A matrix containing the generated background noise data, with n rows and num_dims columns.

Examples

# Generate background noise with custom mean and standard deviation
gen_bkg_noise(n = 100, num_dims = 10, mean = 5, sd = 2)
#>            [,1]      [,2]      [,3]       [,4]       [,5]        [,6]
#>   [1,] 6.648211 2.4300358  6.814872  5.9070958  3.1810458  6.01717722
#>   [2,] 6.406993 5.2866236  3.098829  7.5447480  3.1669498  3.73819667
#>   [3,] 3.486093 5.5075415  4.523542  6.9219181  6.6527221  5.87228741
#>   [4,] 4.109632 6.0045113  5.641307  6.5850689  8.4687171  3.45846240
#>   [5,] 5.378956 2.9995058  7.072878  4.3278103  6.9240317  1.10267709
#>   [6,] 8.048119 5.6863314  4.342785  6.6834922  5.0405039  4.42785161
#>   [7,] 4.865736 6.3123138  5.749059  5.2408659  3.2633822  6.15997360
#>   [8,] 3.121881 4.8100625  2.198076  0.9908777  0.8697836  4.21342389
#>   [9,] 8.307257 5.8778024  5.095476  5.2989053  6.8409503  6.18307039
#>  [10,] 3.523207 2.7798005  3.032422  4.9948222  4.5072812  5.95814745
#>  [11,] 6.625099 3.5735283  5.862235  4.4648656  6.6377743  4.76186073
#>  [12,] 2.966110 7.2090804  7.920129  6.4810789  6.1646064  3.97720238
#>  [13,] 7.662324 1.2730993  5.240839  6.3570903  4.9819328  3.89364812
#>  [14,] 1.667611 4.6587024  5.132545  6.4032587  2.2154276  4.87329700
#>  [15,] 5.430137 3.6365922  8.446239  0.5661255  7.1909470  1.40652917
#>  [16,] 2.977602 2.1235584  1.034071  6.6152395  5.4103779  5.86695539
#>  [17,] 7.488902 2.8653169  7.518823  6.4493872  3.6966280  4.64948378
#>  [18,] 5.078088 2.2524378  4.441191  6.1761639  3.7838126  8.53710298
#>  [19,] 8.374038 0.6081470  5.145046  4.8496091  3.5381426  7.72768123
#>  [20,] 6.222861 1.0707594 10.523180  6.3062306  8.1273970  8.08662490
#>  [21,] 3.979999 4.6451642  7.012177  5.5407379  2.7872118  7.39035597
#>  [22,] 4.745656 8.0164023  3.055730  5.5126526  4.5792186  3.39023467
#>  [23,] 5.211086 8.6807064  2.725025  1.7938614  5.8784025  4.18038755
#>  [24,] 3.332300 7.1166967  2.959499  6.5212585  6.4492154  8.72188152
#>  [25,] 3.122506 5.4767486  5.095496  6.9722515  6.3076938 10.71102547
#>  [26,] 5.946124 4.3350505  8.679672  5.2541635  1.1446990  2.92019680
#>  [27,] 5.309037 4.1349017  2.641225  3.4116092  5.2835196  6.15806082
#>  [28,] 2.752732 5.1359403  8.332295  6.6030422  1.7087391  4.06634234
#>  [29,] 6.511965 8.4575988  5.293018  3.5834062  4.4296622  2.41953262
#>  [30,] 8.136778 2.5956528  3.798938  4.3933006  4.0108590  5.13977181
#>  [31,] 3.552622 3.1207610  3.912439  2.7816671  8.5454234  6.26991582
#>  [32,] 2.477127 3.9476595  3.800945  2.7807206  2.1202910  6.13145990
#>  [33,] 4.974999 3.6574409  8.681888  2.7622389  5.3928826  3.86651711
#>  [34,] 5.173286 5.8361812  6.920968  6.4649076  5.7115928  7.29184297
#>  [35,] 4.493222 6.1144763  3.059424  5.3935691  5.3575662  6.51139156
#>  [36,] 4.480272 2.4527855  8.614101  2.8983625  9.9926412  3.78539350
#>  [37,] 7.588228 3.4506453  4.102509  4.0851587  7.0452137  5.61298925
#>  [38,] 5.487137 5.0440504  2.117287  8.9488005  7.5751140  8.47603909
#>  [39,] 4.469216 2.6017936  7.886392  5.5740237  2.3642307  3.64849612
#>  [40,] 3.731753 1.3253764  6.094724  5.6173703  4.7033725  3.69532005
#>  [41,] 4.110360 5.4251299  5.923836  6.4512353  5.0022246  5.19237892
#>  [42,] 7.261802 2.1119378  6.614392  7.7088717  4.7275699  4.58080233
#>  [43,] 6.551244 6.8356240  1.678139  1.6626313  6.7771638  2.96668632
#>  [44,] 6.701522 6.6304822  3.115003  5.5783064  3.8769572  2.24078861
#>  [45,] 6.764485 8.0393013  6.152491  6.6203826  6.1679443  4.84302837
#>  [46,] 5.890924 3.2555539  3.194305  4.2602206  7.0136845  5.87586048
#>  [47,] 7.977589 7.5037714  5.765170  5.2607849  6.5318941  5.42894406
#>  [48,] 3.644589 4.5701899  7.950511  1.5185327  9.4675915  4.64719462
#>  [49,] 9.017344 7.1738586  5.074820  6.5708272  2.6318408  3.18342438
#>  [50,] 2.586332 4.7276944  6.268025  3.7593668  6.7791760  3.60372872
#>  [51,] 3.041930 3.7129914  2.277519  4.8519954  4.0945958  6.50591138
#>  [52,] 4.491088 8.1540460  7.229028  4.1383551  2.3229981  2.35996160
#>  [53,] 5.984563 5.5004734  2.561927  6.3368800  8.6609001  6.03468302
#>  [54,] 7.809098 5.5005886  4.634763  5.5508789  4.2675842  6.60009022
#>  [55,] 3.374947 6.3615596  3.985319  7.1190818  5.7665933  4.36836898
#>  [56,] 4.037366 2.7893016  2.127130  4.6974834  4.5153843  3.12570258
#>  [57,] 7.887768 1.9914960  3.281070  1.8672826  4.7680336  1.46654755
#>  [58,] 4.528045 5.4859782  2.762655  5.9976228  7.2915479  3.23004183
#>  [59,] 2.732889 4.0927404  5.572542  3.8797672  1.3968053  5.93385479
#>  [60,] 5.769080 5.1293806  4.156685  9.4577633  8.9451558  4.62357769
#>  [61,] 5.544797 5.7474942  5.707387  7.0483977  6.0327061  5.99509214
#>  [62,] 2.764199 2.0284688  4.657815  9.1175827  5.7621201  6.68702900
#>  [63,] 6.048863 1.4987117  2.960578  6.6780739  5.9438095  6.83237136
#>  [64,] 5.184842 8.2354914  5.865680  4.7163904  6.5645889  4.71257919
#>  [65,] 3.209157 1.4896119  7.658017  4.2060753  5.3769744  6.34011402
#>  [66,] 2.724656 7.6634647  3.963425  0.9960061  5.4468287  3.01938175
#>  [67,] 9.144326 3.3058096  0.721821  4.1750894  6.5216431  5.94685527
#>  [68,] 4.114523 5.7564528  4.677421  5.2029469  5.6479765  2.44013359
#>  [69,] 2.335352 5.7064219  4.599823  4.5205801  6.3688169  4.41378784
#>  [70,] 6.931136 5.3234878  3.839692  5.6873221  6.3977193  6.33140721
#>  [71,] 6.326567 4.6880124  6.650992  5.2983032  5.1327617  3.55329700
#>  [72,] 6.242367 4.6078395  8.112938  8.0520345  6.7240765 -0.08738184
#>  [73,] 2.202334 5.7183967  6.943543  3.9443345  5.8283876  7.90840613
#>  [74,] 4.113477 5.5745904  6.433728  2.6748272  4.2690256  5.65535082
#>  [75,] 4.756876 0.9894448  9.376303  3.5535230  3.3139536  3.21599097
#>  [76,] 4.662084 4.9608918  3.525693  3.7118033  3.1431957  1.09661246
#>  [77,] 4.469641 7.8368807  8.609823  2.8708682  7.2535594  6.30025010
#>  [78,] 8.286962 1.5941435  6.808727 11.1269903  4.0255013  5.30746990
#>  [79,] 5.095921 6.9209791  1.726106  3.8385406  6.0195236  6.28920346
#>  [80,] 6.165254 3.8606455  6.201110  6.5896267  2.7368223  5.23805941
#>  [81,] 6.313287 4.7795821  3.180560  1.6220331  7.4398142  5.50136730
#>  [82,] 7.105030 5.4584283  2.320093  3.8951229  6.6155105  7.44409831
#>  [83,] 6.588351 2.9861589  2.337842  3.6558364  5.5444570 -0.27745612
#>  [84,] 4.194220 7.0741223  7.111184  3.2286801  6.3733253  7.54317256
#>  [85,] 4.259660 4.0161704  4.667787  9.3482695  2.1371368  1.20726900
#>  [86,] 5.036260 5.4568457  5.894865  4.6069034  5.4489257  6.27536128
#>  [87,] 4.261799 6.7806853  4.461395  7.3345014  3.7231947  7.75565379
#>  [88,] 6.963364 2.5146282  5.585507  6.0970432  0.8109434  6.06914528
#>  [89,] 3.850315 1.5373396  4.851417  6.5689724  4.8223502  6.25881790
#>  [90,] 4.471099 6.6359878  7.649519  6.0673117  3.5697933  5.81218996
#>  [91,] 7.240936 7.8819021  6.599058  3.9321010  5.6484914  3.94554349
#>  [92,] 3.505731 2.5966448 12.796501  3.7821093  6.0904947  3.62356251
#>  [93,] 4.207514 4.6639103  5.268669  7.6958977  5.0310965  5.66271109
#>  [94,] 4.805674 5.8360726  3.632550  4.6921632  7.2395480  2.75562473
#>  [95,] 3.574618 5.5862166  4.208981  6.7172646  4.4099069  3.24711493
#>  [96,] 5.517917 5.9797732  5.479245  2.3017768  6.0882860  7.81437905
#>  [97,] 3.674287 5.0773023  4.830303  7.1068146  2.7868184  5.37988420
#>  [98,] 3.487575 2.3242456  7.041874  3.4431645  5.5753392  5.53764563
#>  [99,] 6.845022 4.6904152  1.509526  3.9015958 11.9690886  2.63935725
#> [100,] 7.315255 0.8116459  7.204359  4.6306431  2.4275282  7.79550416
#>              [,7]      [,8]       [,9]      [,10]
#>   [1,]  3.9406092  4.701549  6.3531194  4.8943884
#>   [2,]  4.7223007  7.195227  6.1206348  7.8099571
#>   [3,]  3.9243711  6.074028  3.1556712  6.2861872
#>   [4,]  6.8901249  6.357176  7.5853876  4.3004791
#>   [5,]  4.1209051  2.748163  6.2952797  6.1737569
#>   [6,]  5.7893912  4.688816  6.7897800  6.6598910
#>   [7,]  5.0909854  3.951017  4.1313014  8.4228262
#>   [8,]  6.4006378  4.667123  5.5779491  4.1792329
#>   [9,]  6.5751197  4.625601  7.7965388  7.8234989
#>  [10,]  5.0906900  2.714641  3.9676303  5.2306446
#>  [11,]  4.7269848  3.480116  4.6218802  6.1743348
#>  [12,]  4.5307167  5.975876  4.8271190  4.0468402
#>  [13,]  3.0185205  3.494952 10.5919985  6.6927123
#>  [14,]  6.8949110  4.717209  3.4913321  2.0667955
#>  [15,]  8.4490380  4.076195  5.9902357  5.6305373
#>  [16,]  5.8925314  6.587545  7.1486271  5.5146137
#>  [17,]  5.6670057  6.757823  4.3039155  6.1496687
#>  [18,]  5.2641645  6.885707  3.7405946  3.7458134
#>  [19,]  4.4248674  1.173453  2.7437347  5.3092852
#>  [20,]  7.9064764  6.400756  6.1612591  0.1634273
#>  [21,]  5.5119355  7.927379  6.5375413  7.7506894
#>  [22,]  4.1319628  2.553699  6.6365227  9.5337740
#>  [23,]  6.3718870  4.363860  1.0151672  8.3127789
#>  [24,]  4.5781227  3.585788  3.8779918  6.5430319
#>  [25,]  3.3142806  5.044536  4.2105038  4.6820763
#>  [26,]  2.6771387  4.430906  7.1174192  6.7094627
#>  [27,]  5.2765213  8.028038  6.3436430  5.8506879
#>  [28,]  4.6179701  4.506939 10.2177661  4.2459990
#>  [29,]  4.8770009  2.911198  5.0281874  2.1355094
#>  [30,]  3.0481680  3.245390  5.0405614  1.8849630
#>  [31,]  5.8163839  2.354965  2.9155144  3.0296418
#>  [32,]  2.4376485  5.677168  6.8264589  4.1637007
#>  [33,]  4.3261364  3.358099  6.6511502  5.7350345
#>  [34,]  6.0675774 10.811089  3.5930922  5.7551877
#>  [35,]  4.8162485  6.272312  2.6342829  7.2612792
#>  [36,]  4.6641873  1.830992  6.4697507  4.3156218
#>  [37,]  5.2062443  4.920214  3.3838242  5.9269296
#>  [38,]  6.6972957  5.240120  3.9686536  4.3828230
#>  [39,]  4.8819537  4.021401  6.6577008  7.6656248
#>  [40,]  4.3555041  6.730889  9.6350822  6.9575538
#>  [41,]  4.7048784  5.113979  7.6486282  8.9174313
#>  [42,]  6.4281169  5.709368  3.5942911  5.8325383
#>  [43,]  1.4334079  1.197988  3.1636746  5.2024433
#>  [44,]  6.3755565  3.605509  6.4834167  6.2495341
#>  [45,]  8.3125340  3.605208  5.3048827 -0.2670826
#>  [46,]  6.3801074  5.683977  6.0553924  4.6075010
#>  [47,]  5.7325217  3.074764  5.4218347  3.4099088
#>  [48,]  5.0527395  5.104057  5.9140556  6.4328087
#>  [49,]  5.5189433  4.990537  5.5769277  5.6504432
#>  [50,]  4.6697615  5.506059  1.0331444  4.2332263
#>  [51,]  4.6666009  4.464086  5.8697168  6.8100159
#>  [52,] 10.2425100  2.462346  4.2176176  7.7099098
#>  [53,]  3.8027626  5.351308  4.0547126  1.1641999
#>  [54,]  5.2401244  5.462103  3.0861196  4.0069624
#>  [55,]  3.5741654  2.786178  6.3479007  4.1665414
#>  [56,]  4.2832585  4.450177  3.5322108  6.8925828
#>  [57,]  6.6978493  6.990929  4.3144621  5.4329213
#>  [58,]  3.5315792  2.552870  4.0422758  4.2450719
#>  [59,]  8.7529100  4.726652  5.6354584  8.4343030
#>  [60,]  2.5417298  5.780370  5.7836655  1.0381728
#>  [61,]  3.1821091  5.134268  6.0194759  4.2609685
#>  [62,]  0.2406656  4.557786  7.6217907  2.8247746
#>  [63,]  3.5898360  2.456077  4.5371229  2.9098470
#>  [64,]  5.7987578 10.388088  7.5982864  4.8380405
#>  [65,]  3.8254722  8.508466  1.7832847  2.6029568
#>  [66,]  7.2693529  3.675009  3.1654151  6.2813986
#>  [67,]  3.2273039  5.282580  6.3771845  6.3619869
#>  [68,]  6.7352604  6.081450  6.0450169  5.1516675
#>  [69,]  5.5591097  3.967910  6.3198819  3.6309078
#>  [70,]  4.4553102  3.147485  5.5957769  1.1057272
#>  [71,]  5.0746732  3.735984  5.8381069  4.6467181
#>  [72,]  6.6802964  4.839631  5.4848655  4.5073561
#>  [73,]  2.6112994  6.199475  6.4291676  5.4985763
#>  [74,]  6.1655935  8.017461  7.8171270  4.6309830
#>  [75,]  7.1010753  2.320011  3.2994611  4.2049330
#>  [76,]  8.4791620  4.850686  6.9925568  4.3869566
#>  [77,]  3.5515166  3.210242  4.1964385  5.5544717
#>  [78,]  6.8143581  6.086579  4.0660028  6.0377418
#>  [79,]  4.9061254  5.741126  4.8152030  7.9464332
#>  [80,]  1.9159667  5.520886  0.7954269  6.6587646
#>  [81,]  7.1278678  2.260339  2.4719896  5.6799493
#>  [82,]  5.4553179  5.460164  2.9242754  3.6783155
#>  [83,]  8.2416140  2.989113  3.8798079  4.2092506
#>  [84,]  4.4628897  4.599926  7.1643419  7.5873660
#>  [85,]  6.1040571  6.920914  5.0950224  2.5854521
#>  [86,]  6.5647325  8.877831  4.4637767  5.9207772
#>  [87,]  4.9589224  5.784598  7.5621188  6.2397836
#>  [88,]  4.3056729  3.335868  3.2262198  5.4706917
#>  [89,]  7.8392025  7.380325  4.6312224  4.8468852
#>  [90,]  8.2091447  2.330448  4.5292680  8.5217116
#>  [91,]  1.9592225  2.198561  2.8537112  4.3572130
#>  [92,]  7.7174903  3.700891  5.8075988  6.5019586
#>  [93,]  8.0055664  6.937180  5.3582579  7.3605026
#>  [94,]  4.9133893  2.504174  7.6327550  3.4257831
#>  [95,]  2.5977106  7.562038  8.5252572  3.7791884
#>  [96,]  6.9573047  4.769706  5.1102031  5.8910840
#>  [97,]  2.8305800  5.504692  5.6873848  4.7036278
#>  [98,]  4.5637980  3.385075  5.0874681  0.9821058
#>  [99,]  7.5568485  7.695662  5.1315320  3.5518963
#> [100,]  4.6238613  1.640624  6.5669478  2.7329719