Generate Random Noise Dimensions

This function generates random noise dimensions to be added to the coordinates of a sphere.

Usage

gen_noise_dims(n, num_noise_dims, min_noise, max_noise)

Arguments

n

The number of observations for which to generate noise dimensions.

num_noise_dims

The number of noise dimensions to generate.

min_noise

The minimum value for the random noise.

max_noise

The maximum value for the random noise.

Value

A matrix containing the generated random noise dimensions.

Examples

# Generate random noise dimensions with 3 dimensions, minimum value -1, and maximum value 1
gen_noise_dims(n = 50, num_noise_dims = 3, min_noise = -1, max_noise = 1)
#>              [,1]         [,2]        [,3]
#>  [1,] -0.59012032  0.363830975  0.67243725
#>  [2,]  0.15085678 -0.569498198 -0.47603450
#>  [3,] -0.51825364  0.122799122  0.20085680
#>  [4,] -0.73776092 -0.199959251  0.72637035
#>  [5,]  0.55092223 -0.738831747 -0.37747965
#>  [6,] -0.03188710  0.373898425 -0.96006755
#>  [7,]  0.34381330  0.550340277 -0.83984463
#>  [8,] -0.62829691  0.469262345 -0.72320982
#>  [9,] -0.15028177  0.883209653  0.58351042
#> [10,]  0.46899435  0.002156275  0.65409597
#> [11,] -0.87250620 -0.530743439  0.36970871
#> [12,] -0.90136433 -0.453510460 -0.28695205
#> [13,]  0.05352350 -0.792848323 -0.85122833
#> [14,]  0.29166354  0.515118668 -0.35363831
#> [15,]  0.65229982 -0.009973590  0.18654969
#> [16,]  0.28036450 -0.666656007  0.82297886
#> [17,] -0.90179694  0.069044840  0.74301895
#> [18,]  0.01693300 -0.776538014 -0.76426848
#> [19,]  0.53972651 -0.200032434 -0.29942194
#> [20,]  0.07537042  0.366558300 -0.04504109
#> [21,] -0.58352349 -0.205031262 -0.21468430
#> [22,]  0.49500559 -0.800781888 -0.69490630
#> [23,]  0.69082032  0.804373992  0.73639058
#> [24,] -0.17789062  0.224200455 -0.46036368
#> [25,] -0.81686435  0.192434929 -0.40002141
#> [26,] -0.21694361  0.427338323 -0.32959335
#> [27,]  0.90432526 -0.209292524 -0.07363631
#> [28,]  0.38190866  0.352938158  0.34515240
#> [29,] -0.17028621 -0.473999693  0.62943813
#> [30,]  0.93755513  0.436667517  0.83695196
#> [31,]  0.68807876 -0.343550300  0.74474248
#> [32,] -0.81112769 -0.319701432 -0.73746754
#> [33,] -0.78666642  0.741903305 -0.96174974
#> [34,] -0.45595643  0.760927265  0.38663321
#> [35,] -0.03114460  0.562027744  0.34204514
#> [36,] -0.41708418 -0.509012773  0.77782187
#> [37,] -0.90840033  0.605097950  0.81724627
#> [38,]  0.95705301  0.415756793  0.78566852
#> [39,] -0.45908522  0.622354390 -0.66573966
#> [40,] -0.16752270 -0.709258038  0.16766893
#> [41,]  0.38994899  0.344013950 -0.49285098
#> [42,]  0.20909609  0.844812593  0.03700774
#> [43,]  0.10119584  0.863458502 -0.46552117
#> [44,]  0.93421514 -0.681335458  0.59170593
#> [45,] -0.08405502 -0.502041401  0.83813688
#> [46,] -0.50257580 -0.585517632  0.84400212
#> [47,]  0.59563458  0.955428061  0.34242347
#> [48,] -0.29152431 -0.708821371 -0.03195703
#> [49,]  0.65213942 -0.772503925  0.09675412
#> [50,] -0.97736451  0.317767636 -0.87900505