Generate Random Noise Dimensions

This function generates random noise dimensions to be added to the coordinates of a sphere.

Usage

gen_noise_dims(n, num_noise_dims, min_noise, max_noise)

Arguments

n

The number of observations for which to generate noise dimensions.

num_noise_dims

The number of noise dimensions to generate.

min_noise

The minimum value for the random noise.

max_noise

The maximum value for the random noise.

Value

A matrix containing the generated random noise dimensions.

Examples

# Generate random noise dimensions with 3 dimensions, minimum value -1, and maximum value 1
gen_noise_dims(n = 50, num_noise_dims = 3, min_noise = -1, max_noise = 1)
#>              [,1]        [,2]         [,3]
#>  [1,] -0.74957441 -0.78392677 -0.800142200
#>  [2,] -0.09452098 -0.95728732 -0.710722029
#>  [3,]  0.08219937  0.37547788 -0.365257301
#>  [4,]  0.04151365 -0.89011912  0.059897573
#>  [5,]  0.03763962 -0.58585217  0.690639886
#>  [6,]  0.80821372 -0.13752705  0.288115236
#>  [7,] -0.94787458 -0.54963726 -0.346201787
#>  [8,]  0.67716134 -0.72107692 -0.606837196
#>  [9,] -0.89124392  0.66372089  0.368044839
#> [10,]  0.07067866 -0.81243096 -0.251459291
#> [11,] -0.64998606  0.83687327  0.597897368
#> [12,] -0.14135958 -0.64263356 -0.584218061
#> [13,]  0.07181137  0.70335707 -0.371622616
#> [14,] -0.22035382 -0.79348913 -0.848535076
#> [15,] -0.93274127 -0.21619455  0.997053021
#> [16,]  0.22685453 -0.85578433 -0.114279465
#> [17,] -0.29191829 -0.73657537 -0.871159972
#> [18,]  0.76295918  0.66720426 -0.994786982
#> [19,] -0.41649862  0.43540777  0.927490392
#> [20,]  0.14615585 -0.53781238 -0.625043496
#> [21,]  0.96594492  0.04890343 -0.007203219
#> [22,]  0.83112723 -0.61705411  0.823740205
#> [23,]  0.18819408 -0.86053206  0.297094038
#> [24,] -0.13346967 -0.95056571 -0.223901995
#> [25,] -0.54703442  0.56356564  0.568515607
#> [26,] -0.45595977 -0.41996837  0.090030250
#> [27,] -0.78563331 -0.49271286  0.294385937
#> [28,] -0.70143331 -0.10909372  0.250645414
#> [29,]  0.21635430 -0.11865389 -0.718208554
#> [30,] -0.26877824 -0.13109885 -0.139461921
#> [31,]  0.13245815  0.47600446 -0.923568353
#> [32,] -0.02009874  0.38498235  0.542756170
#> [33,] -0.35548971  0.65165349  0.279878825
#> [34,]  0.21278203 -0.70202368 -0.666443065
#> [35,] -0.09276298 -0.09486518  0.608938080
#> [36,] -0.93476688  0.41940940 -0.206576977
#> [37,] -0.03723532  0.77254234  0.688224187
#> [38,] -0.40427107  0.55150818 -0.660056769
#> [39,]  0.08868415  0.84557315 -0.475042265
#> [40,]  0.29050577 -0.37822010 -0.824360061
#> [41,]  0.55213310  0.60085900 -0.619727341
#> [42,] -0.60365094 -0.61779933  0.176698082
#> [43,] -0.26967078  0.39159947 -0.403316037
#> [44,] -0.63561423  0.81609190 -0.242036278
#> [45,] -0.12946516 -0.83406260  0.413068699
#> [46,] -0.51584086 -0.86836457 -0.763343190
#> [47,] -0.95296463 -0.11110862 -0.523654773
#> [48,] -0.89351079 -0.54912274 -0.394389472
#> [49,]  0.24422093  0.53951827  0.677094616
#> [50,]  0.62851561  0.59464999 -0.786929845