Generate data points on a Roman surface with optional noise.

This function generates data points on a Roman surface with optional noise.

Usage

roman_surface_3d(n, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

roman_surface_3d(n = 100, num_noise = 8,
min_n = -0.05, max_n = 0.05)
#>                 [,1]         [,2]          [,3]          [,4]          [,5]
#>   [1,] -6.775282e-03 -0.042423238  0.1554466829 -2.122289e-02 -0.0409846716
#>   [2,] -2.809241e-01 -0.085876391  0.2528987514  8.654703e-03  0.0132722384
#>   [3,] -9.516681e-05  0.007023123 -0.0135473405 -4.538239e-02  0.0325356329
#>   [4,] -2.761767e-04  0.009912275 -0.0278377390 -4.455922e-02 -0.0284519097
#>   [5,] -4.012322e-01  0.280642965 -0.2823922486 -1.810834e-02 -0.0382636261
#>   [6,] -9.178303e-02  0.359260843 -0.2003544193 -4.819563e-04  0.0270293918
#>   [7,] -5.386226e-02  0.449353553 -0.0339667092 -3.804512e-03 -0.0172637092
#>   [8,]  4.057222e-02 -0.065495741 -0.4450016604  4.281772e-02 -0.0370280490
#>   [9,]  5.772671e-03  0.071528638  0.0797670283  1.444257e-02 -0.0010803847
#>  [10,]  3.464712e-01 -0.114623744 -0.2510167073 -4.425170e-02 -0.0437022017
#>  [11,] -1.293142e-01 -0.222253034  0.4037376215  4.079691e-02  0.0392507628
#>  [12,]  1.635649e-01 -0.423442832 -0.0976013478  3.536295e-02  0.0451034958
#>  [13,] -7.620874e-04  0.004011476 -0.1833561264 -4.676637e-02 -0.0485105699
#>  [14,] -2.933524e-02 -0.490821533  0.0351814346 -3.213907e-02 -0.0068018055
#>  [15,] -1.894564e-01 -0.460703656  0.1910374752 -1.997321e-02  0.0333789374
#>  [16,] -2.577468e-01 -0.248654090  0.4241810886  1.883535e-02  0.0094716441
#>  [17,]  2.707036e-01  0.375945377  0.1478979451  1.648622e-02 -0.0210112104
#>  [18,] -3.609002e-01  0.329662983 -0.3013791986  1.289622e-02 -0.0126246184
#>  [19,] -1.889433e-01 -0.110149780  0.4131866204  3.736124e-02 -0.0021581711
#>  [20,]  2.517936e-02  0.002098401  0.0827104804 -9.996359e-03  0.0492259581
#>  [21,] -4.212386e-01  0.211353260 -0.1334791670  4.234721e-02 -0.0379708976
#>  [22,]  8.574853e-02  0.484801157  0.1007205548  3.441005e-02  0.0256937347
#>  [23,]  1.430553e-03 -0.035999484 -0.0396239263 -1.557852e-02  0.0468761779
#>  [24,]  3.705305e-01 -0.073035906 -0.1570434638 -3.058587e-02 -0.0458333121
#>  [25,] -2.595275e-01 -0.364432180  0.3417367945  2.131510e-02  0.0206831398
#>  [26,]  4.933908e-02 -0.315622739 -0.1349907998  3.822705e-03 -0.0173709187
#>  [27,]  9.728123e-02 -0.096016641 -0.4904352336  3.714973e-02  0.0410482917
#>  [28,] -5.840551e-02  0.328228934 -0.1505116593  1.885837e-02  0.0464267985
#>  [29,]  6.471557e-03 -0.045176139 -0.1400781286 -2.357517e-02 -0.0258226939
#>  [30,]  1.174624e-03 -0.055396093 -0.0211292884 -2.973967e-02  0.0479609052
#>  [31,] -4.840002e-01 -0.101792947  0.0859278538  2.746788e-02  0.0283033232
#>  [32,] -4.537125e-01 -0.146469235  0.1021381257  2.161177e-02  0.0017614835
#>  [33,]  7.980835e-02 -0.420741840 -0.1388027218 -4.950154e-02 -0.0118772414
#>  [34,]  2.952092e-01 -0.271567643 -0.1003997872 -4.943592e-02  0.0182845023
#>  [35,]  1.379211e-01 -0.272895940 -0.3605454797 -4.183173e-02  0.0276897626
#>  [36,] -3.867275e-02 -0.494220869  0.0338165482  4.187626e-02  0.0347447516
#>  [37,]  3.996754e-01  0.108580729  0.1973710779 -2.603506e-02  0.0367929405
#>  [38,]  1.922887e-02 -0.056864012 -0.3023580567 -3.163103e-02 -0.0140300996
#>  [39,]  2.981524e-01  0.033213023  0.0990279413 -2.531118e-02  0.0377885767
#>  [40,] -1.059984e-01  0.436631205 -0.0643382895 -2.918433e-02  0.0272605858
#>  [41,] -5.735873e-02  0.242307553 -0.2092706860 -1.308183e-02 -0.0430801536
#>  [42,] -3.362750e-01  0.325362471 -0.3379330730 -4.621896e-02  0.0015805147
#>  [43,]  4.775609e-02  0.239235088  0.1797699579 -2.089045e-02 -0.0140607491
#>  [44,] -1.406303e-01 -0.239865699  0.3994690168 -7.678888e-03  0.0354013717
#>  [45,] -7.294643e-02  0.081918256 -0.4906056319 -3.887941e-02  0.0274067319
#>  [46,]  1.602393e-01 -0.400324180 -0.0851771880 -2.512440e-02  0.0360308872
#>  [47,]  5.298454e-02 -0.004537913 -0.0847812000 -1.099050e-02 -0.0389322513
#>  [48,]  2.037706e-02  0.020421824  0.4995823126 -4.554400e-02 -0.0047766983
#>  [49,]  3.735335e-01  0.245569718  0.3323288423 -3.801827e-02 -0.0121226063
#>  [50,] -9.708624e-02  0.456661191 -0.0652969732  2.255489e-02  0.0108194286
#>  [51,]  2.265341e-01 -0.441400026 -0.1779109549 -4.504512e-02 -0.0286161324
#>  [52,]  1.347767e-01 -0.113630547 -0.0158227785  1.101960e-02 -0.0304338814
#>  [53,] -2.097799e-01 -0.433126039  0.2493618362 -3.795845e-02  0.0261214983
#>  [54,]  1.482244e-03  0.011432380  0.1274927977 -1.730093e-02  0.0327858326
#>  [55,] -2.024367e-01 -0.213022416  0.4516758785  1.540083e-02  0.0419754604
#>  [56,] -1.198488e-01 -0.311432688  0.2924041144 -4.017915e-02 -0.0425603777
#>  [57,] -4.670064e-01 -0.160613912  0.1297364165  4.816369e-02 -0.0235084359
#>  [58,]  3.113785e-04 -0.034703945 -0.0089608822  4.226709e-02  0.0223936454
#>  [59,]  6.838401e-02  0.240417889  0.2455274780 -1.907785e-02 -0.0071506144
#>  [60,] -1.401888e-02  0.259416133 -0.0499605344  1.221186e-03 -0.0109671088
#>  [61,] -4.805480e-01  0.094172717 -0.1134256425  3.539278e-02  0.0064272638
#>  [62,]  2.058740e-01 -0.431620391 -0.1375505385 -1.869261e-02 -0.0470498977
#>  [63,] -3.784784e-01 -0.326730158  0.2479352585 -2.295105e-02  0.0489539401
#>  [64,]  5.143674e-02  0.470498864  0.0366015885 -1.906750e-02 -0.0297324308
#>  [65,] -2.365659e-01 -0.191800117  0.4385879319  2.157296e-03  0.0419777144
#>  [66,]  3.053602e-02 -0.109601817 -0.2551488161 -2.297152e-02  0.0115164994
#>  [67,] -6.306949e-02  0.265027622 -0.2070356532  1.346350e-02 -0.0391720870
#>  [68,]  2.457522e-03 -0.060796711 -0.0402060565  1.786862e-02  0.0009205266
#>  [69,] -4.118893e-05  0.043363322 -0.0009480661  2.660021e-02 -0.0065964586
#>  [70,]  7.959240e-02  0.366838444  0.1720485842 -1.127088e-02  0.0428596318
#>  [71,]  1.807453e-01 -0.292596424 -0.3858290002  6.324651e-03  0.0371440677
#>  [72,] -1.159856e-01  0.477242165 -0.1362508459 -1.737827e-04  0.0460846022
#>  [73,]  1.241804e-01 -0.329937645 -0.2817299774 -8.754926e-05  0.0225111873
#>  [74,] -1.766030e-02  0.083322211 -0.2013572941 -1.576318e-03  0.0343880526
#>  [75,]  2.059940e-01  0.219668889  0.0503264512 -8.744025e-03  0.0356741199
#>  [76,]  2.199862e-02  0.133879317  0.1569954430  4.237367e-02  0.0419562812
#>  [77,] -1.379779e-02  0.212624621 -0.0615399730  3.436238e-02  0.0189736202
#>  [78,]  1.362398e-01  0.321813630  0.3078828328  4.161446e-02  0.0410217005
#>  [79,]  7.153677e-02 -0.490087426 -0.0812560168  3.937608e-02  0.0135789534
#>  [80,] -3.566465e-02  0.420649068 -0.0646434561 -2.959415e-02 -0.0374767438
#>  [81,] -7.513999e-02  0.134385774 -0.4156162080  1.086969e-02 -0.0383876778
#>  [82,] -2.832891e-03 -0.043552175  0.0646482468  4.528240e-03 -0.0405857127
#>  [83,]  7.176349e-02  0.294440148  0.2065237597 -4.987060e-02 -0.0071278363
#>  [84,]  3.901119e-01  0.295882281  0.1919552985  2.194735e-02  0.0246126576
#>  [85,] -3.266892e-01  0.359866325 -0.1904348460  3.317970e-02 -0.0347781395
#>  [86,]  4.786898e-01  0.057493291  0.0435136386  3.887020e-02  0.0043314905
#>  [87,] -1.313470e-01  0.032169648 -0.2267549764 -1.921623e-02 -0.0298917597
#>  [88,]  1.082728e-01 -0.184541836 -0.4154762120  3.032382e-02 -0.0496032486
#>  [89,] -2.006253e-03 -0.071168257  0.0280243834  4.897828e-02  0.0384484261
#>  [90,]  3.706404e-01  0.233188572  0.3341518715  3.076901e-02  0.0232940348
#>  [91,]  4.432661e-01  0.231327605  0.2057535454  5.099417e-03 -0.0256177188
#>  [92,]  9.157631e-02  0.172606597  0.3975509331  1.084563e-02 -0.0334240521
#>  [93,]  6.855854e-02  0.069460822  0.4951488874 -5.274749e-03  0.0155697784
#>  [94,] -5.501408e-02 -0.496892894  0.0537667421 -3.968429e-02 -0.0068328597
#>  [95,] -5.079566e-02  0.107249333 -0.3813189595  3.874378e-02 -0.0233134592
#>  [96,]  1.264964e-01  0.387528628  0.0620897933 -3.735936e-02  0.0336599418
#>  [97,] -9.783297e-02  0.432020210 -0.1576616836 -4.502559e-02 -0.0162945946
#>  [98,] -9.656731e-04  0.338697379 -0.0024742223  1.340022e-03 -0.0419883305
#>  [99,] -4.049739e-02  0.248417281 -0.1480073964  1.123429e-02 -0.0031171251
#> [100,]  2.174650e-01  0.149452724  0.4316354077  2.104798e-02  0.0071959540
#>                 [,6]          [,7]          [,8]          [,9]         [,10]
#>   [1,] -2.445545e-02  0.0433163609  0.0442896327 -0.0074606275  0.0216131426
#>   [2,] -1.713964e-02 -0.0271911550  0.0063633443  0.0138390964  0.0195462710
#>   [3,] -1.908673e-02  0.0357879906 -0.0267206956  0.0241253724 -0.0284186496
#>   [4,]  1.383062e-02 -0.0145061703  0.0368153896  0.0409164861 -0.0386915315
#>   [5,] -2.815120e-02  0.0210568843 -0.0146264939 -0.0185878391 -0.0459210987
#>   [6,]  1.128679e-02  0.0380992513 -0.0424700768  0.0171436717  0.0460873031
#>   [7,] -4.326073e-02  0.0363090078  0.0178925464 -0.0241327165  0.0253418096
#>   [8,]  4.021729e-02 -0.0211095950  0.0073440419 -0.0070315686  0.0085353440
#>   [9,]  1.600990e-02  0.0402454065 -0.0311158936  0.0156992736 -0.0307900076
#>  [10,] -2.636015e-02 -0.0335239758 -0.0474344181  0.0347195798  0.0469534871
#>  [11,] -4.206589e-02  0.0396805124 -0.0232392969  0.0244759848  0.0163375882
#>  [12,]  4.144988e-02 -0.0407104879  0.0156027002 -0.0340756334 -0.0239952798
#>  [13,]  4.334580e-02  0.0471357677  0.0171846927  0.0412143918 -0.0388280467
#>  [14,]  2.825326e-02  0.0079000390 -0.0381772376  0.0091738202 -0.0280978021
#>  [15,] -3.252448e-02  0.0441574948  0.0435139775 -0.0495074975  0.0148177539
#>  [16,] -3.799927e-02  0.0352762166 -0.0258826403  0.0294796787 -0.0259473208
#>  [17,]  1.675060e-02  0.0139297968 -0.0134433061  0.0210169895  0.0165952451
#>  [18,]  4.092707e-02  0.0067866249 -0.0169973451  0.0188552435  0.0028494145
#>  [19,] -6.489908e-04  0.0136117625 -0.0428563133  0.0353263823  0.0227259549
#>  [20,]  9.324657e-03 -0.0002405228  0.0354614842  0.0333354607 -0.0463121660
#>  [21,] -1.267636e-02  0.0490986322  0.0361828258 -0.0469660298 -0.0133380719
#>  [22,] -4.015052e-02 -0.0262959340  0.0299343970  0.0433029271  0.0331625406
#>  [23,]  1.963264e-04  0.0086104242  0.0419908367  0.0415557840 -0.0428700109
#>  [24,]  2.582318e-02  0.0056247443 -0.0289580954 -0.0370256638 -0.0138360330
#>  [25,]  1.340590e-02 -0.0253040771 -0.0336261392 -0.0367632700  0.0070837569
#>  [26,] -4.224997e-02 -0.0229332069 -0.0143133345 -0.0184526358  0.0156154521
#>  [27,] -4.107419e-02 -0.0482906847  0.0168624699 -0.0444686700 -0.0022942668
#>  [28,]  2.568144e-02 -0.0248958943  0.0185596225 -0.0115633949 -0.0045842411
#>  [29,]  4.170418e-02 -0.0123194597  0.0283408355  0.0157634551 -0.0117807796
#>  [30,] -4.803106e-02 -0.0337320379 -0.0155424506  0.0240182868  0.0205344067
#>  [31,]  1.892757e-02 -0.0202820333  0.0078990056  0.0282700221  0.0392575591
#>  [32,]  1.223369e-02 -0.0429948093  0.0389050137 -0.0173535326 -0.0032986777
#>  [33,]  3.569313e-02 -0.0152926281  0.0393282842 -0.0464151531  0.0059147818
#>  [34,] -4.310835e-02 -0.0224206841 -0.0270081083  0.0175588437 -0.0211108564
#>  [35,] -1.327926e-02 -0.0189337148  0.0342164542  0.0419447914 -0.0026276361
#>  [36,] -3.730075e-03 -0.0329604652  0.0278535092  0.0429412813  0.0033157368
#>  [37,]  2.880956e-02  0.0166309154  0.0034390963  0.0232101000  0.0095035827
#>  [38,]  1.403993e-02 -0.0179923125  0.0102707196  0.0160370285 -0.0009629260
#>  [39,] -6.773888e-03  0.0415241050  0.0462536138 -0.0300367596 -0.0384218578
#>  [40,] -4.995216e-03 -0.0133599635  0.0208506366 -0.0370825181 -0.0081344341
#>  [41,] -4.500958e-02  0.0357274000  0.0300771132 -0.0381656251 -0.0465914276
#>  [42,] -2.605458e-02 -0.0013668176  0.0040198287 -0.0137281560 -0.0127549055
#>  [43,] -3.976499e-02  0.0454734828 -0.0009411381 -0.0251597907  0.0420259516
#>  [44,]  3.245916e-02 -0.0434382091  0.0241386232  0.0211132850  0.0332060132
#>  [45,]  1.807270e-02  0.0338644130  0.0462775002  0.0029209384  0.0048058668
#>  [46,]  3.776908e-02 -0.0060828313  0.0114114023  0.0384433763 -0.0371581825
#>  [47,] -3.543198e-03  0.0382445128 -0.0002874756 -0.0366361377 -0.0029651299
#>  [48,] -2.082108e-02 -0.0316029094 -0.0187196399 -0.0450994452  0.0218857854
#>  [49,]  4.340589e-02  0.0319600076 -0.0169161956 -0.0049369015 -0.0480088063
#>  [50,] -1.191147e-02  0.0402221324 -0.0236343998 -0.0221250155 -0.0213965237
#>  [51,]  2.768928e-02  0.0230128248 -0.0237382195  0.0427354569  0.0360540837
#>  [52,]  1.804140e-02 -0.0207837848 -0.0413819076  0.0217138006 -0.0010926298
#>  [53,] -2.903818e-02 -0.0141777369  0.0003530915  0.0069523322 -0.0395389074
#>  [54,] -4.466376e-02 -0.0060010354 -0.0092663842  0.0152474643 -0.0164348182
#>  [55,] -2.957417e-03 -0.0456615692 -0.0292030083 -0.0057981252  0.0064327515
#>  [56,]  1.849934e-02  0.0457734970 -0.0151751673  0.0439960615 -0.0345864706
#>  [57,]  1.609581e-03 -0.0320483438  0.0272947988  0.0064159404  0.0367117811
#>  [58,]  9.521616e-05 -0.0172073594  0.0307516940  0.0234618294 -0.0401038633
#>  [59,] -1.893956e-02 -0.0223002825  0.0026822003 -0.0040666848 -0.0132228706
#>  [60,]  4.361391e-02 -0.0231287358 -0.0047595450  0.0347266830  0.0385298261
#>  [61,] -1.992240e-02 -0.0101917212  0.0174088389  0.0231895196  0.0400158201
#>  [62,]  1.764004e-02  0.0057753126 -0.0495410092  0.0231923949 -0.0430111898
#>  [63,]  4.132948e-02  0.0197244972  0.0411356451  0.0084781711  0.0017101304
#>  [64,] -5.802554e-03 -0.0307689531  0.0117141113  0.0332241725 -0.0481662693
#>  [65,] -2.166620e-02 -0.0363289408 -0.0331977856 -0.0099579564  0.0142342242
#>  [66,] -1.522336e-02 -0.0228172658  0.0243638615  0.0290661146 -0.0250347239
#>  [67,]  4.651794e-02 -0.0297810500 -0.0277867216  0.0354406696 -0.0021663966
#>  [68,]  1.456180e-02 -0.0477103471 -0.0411975005 -0.0157318188  0.0257814487
#>  [69,] -3.574610e-02 -0.0344292084  0.0353954606 -0.0354974853 -0.0151992977
#>  [70,]  3.959473e-02  0.0204119854  0.0003660906  0.0311661947 -0.0024110409
#>  [71,] -4.783713e-02 -0.0264584337 -0.0369230833 -0.0483883592  0.0499900384
#>  [72,] -2.103648e-02  0.0028616752 -0.0457143753  0.0445671861  0.0145019288
#>  [73,] -8.665278e-04  0.0330531284 -0.0180086509  0.0010109721  0.0083652628
#>  [74,] -1.530778e-02  0.0378845478 -0.0157960910 -0.0425921815  0.0294790995
#>  [75,] -6.782656e-03 -0.0177045155  0.0483799116 -0.0464208638 -0.0035670719
#>  [76,]  2.324877e-02 -0.0337753196  0.0138730677 -0.0351645889  0.0373078809
#>  [77,]  1.053327e-02 -0.0189064131  0.0429092847  0.0037568555  0.0341527650
#>  [78,] -3.786034e-02  0.0387770105 -0.0418528705 -0.0076332285 -0.0431773406
#>  [79,]  2.771146e-02 -0.0207285941 -0.0148885951 -0.0420274923 -0.0376229295
#>  [80,] -4.074711e-02 -0.0079270235 -0.0388904460  0.0028938612 -0.0355245161
#>  [81,] -4.319066e-02  0.0015550741  0.0083328215  0.0008152405  0.0326189054
#>  [82,] -1.122707e-02  0.0178151881  0.0202727830 -0.0444145594  0.0399738909
#>  [83,] -4.268610e-02 -0.0384385654 -0.0323776991 -0.0245736291 -0.0253501086
#>  [84,]  2.283705e-02  0.0011409885  0.0353496553 -0.0366529968 -0.0037875813
#>  [85,]  4.889322e-02 -0.0444714993  0.0479728599 -0.0494076489 -0.0002357048
#>  [86,]  3.582931e-02  0.0219036604 -0.0368332097 -0.0192902023  0.0308989903
#>  [87,]  2.579276e-02 -0.0043650464 -0.0348450779 -0.0398451325 -0.0261858245
#>  [88,] -3.661453e-02  0.0034961618 -0.0325548516  0.0326397044 -0.0093042359
#>  [89,]  2.352589e-02 -0.0490032254  0.0200867542 -0.0422794077  0.0248394853
#>  [90,]  4.164040e-02 -0.0427478649  0.0160743825  0.0302672950 -0.0343459307
#>  [91,]  3.933539e-02  0.0299405253  0.0467719716  0.0194971071  0.0077628783
#>  [92,] -4.495990e-02  0.0321191312  0.0103122525 -0.0136373684 -0.0094798177
#>  [93,] -2.231202e-02 -0.0161784599  0.0115688592  0.0382314543  0.0190921098
#>  [94,]  7.432453e-03 -0.0082737287  0.0463264088 -0.0016079857  0.0315385019
#>  [95,] -4.171506e-02 -0.0407160239  0.0253210499 -0.0305146767 -0.0062556417
#>  [96,] -1.469917e-02 -0.0046219188 -0.0307449190 -0.0222338091  0.0280707784
#>  [97,] -4.183133e-02  0.0216060758 -0.0071683184  0.0330714962 -0.0094769391
#>  [98,]  1.254448e-02  0.0226711328  0.0360957523  0.0115472567 -0.0357489867
#>  [99,] -2.620543e-02 -0.0412582674 -0.0271281819 -0.0483312341  0.0230836401
#> [100,] -2.503487e-02  0.0318188905  0.0125189194 -0.0195997853 -0.0090781627
#>                [,11]
#>   [1,]  0.0282419795
#>   [2,] -0.0432821232
#>   [3,] -0.0286807350
#>   [4,]  0.0440817680
#>   [5,]  0.0143894460
#>   [6,]  0.0198992772
#>   [7,]  0.0195603144
#>   [8,]  0.0412865476
#>   [9,]  0.0192870727
#>  [10,] -0.0312743022
#>  [11,] -0.0193799291
#>  [12,]  0.0015812704
#>  [13,]  0.0285622859
#>  [14,] -0.0396219824
#>  [15,]  0.0352514074
#>  [16,] -0.0279465789
#>  [17,] -0.0214841192
#>  [18,]  0.0095472151
#>  [19,]  0.0464742140
#>  [20,]  0.0050839094
#>  [21,] -0.0056709421
#>  [22,] -0.0265413817
#>  [23,]  0.0325969992
#>  [24,]  0.0089809836
#>  [25,]  0.0122859175
#>  [26,] -0.0270166170
#>  [27,] -0.0348576669
#>  [28,] -0.0298342893
#>  [29,]  0.0185022906
#>  [30,]  0.0366329697
#>  [31,] -0.0120214755
#>  [32,] -0.0176842314
#>  [33,] -0.0158300434
#>  [34,]  0.0265946368
#>  [35,]  0.0041189420
#>  [36,]  0.0236673514
#>  [37,] -0.0475180184
#>  [38,] -0.0143179732
#>  [39,]  0.0025411272
#>  [40,] -0.0399965454
#>  [41,]  0.0001760205
#>  [42,]  0.0353530779
#>  [43,]  0.0435326782
#>  [44,] -0.0213847220
#>  [45,]  0.0144955840
#>  [46,] -0.0444819321
#>  [47,] -0.0462432905
#>  [48,]  0.0343918403
#>  [49,] -0.0110005761
#>  [50,]  0.0080495462
#>  [51,]  0.0317908633
#>  [52,]  0.0043027752
#>  [53,] -0.0235061679
#>  [54,]  0.0424865043
#>  [55,] -0.0179289135
#>  [56,] -0.0320648373
#>  [57,]  0.0435004009
#>  [58,]  0.0009348501
#>  [59,] -0.0054105347
#>  [60,]  0.0264372552
#>  [61,]  0.0180847583
#>  [62,]  0.0457994778
#>  [63,] -0.0055335815
#>  [64,] -0.0186649161
#>  [65,]  0.0130545829
#>  [66,]  0.0460881645
#>  [67,] -0.0043422894
#>  [68,] -0.0237134074
#>  [69,]  0.0425501091
#>  [70,]  0.0338479123
#>  [71,]  0.0127845748
#>  [72,]  0.0294204812
#>  [73,]  0.0145647840
#>  [74,] -0.0032317718
#>  [75,]  0.0184520660
#>  [76,] -0.0497543893
#>  [77,] -0.0413771939
#>  [78,]  0.0436538680
#>  [79,] -0.0387757104
#>  [80,] -0.0370841580
#>  [81,]  0.0220069072
#>  [82,]  0.0235231907
#>  [83,]  0.0262789725
#>  [84,]  0.0291862053
#>  [85,]  0.0020364895
#>  [86,]  0.0409814705
#>  [87,] -0.0048118972
#>  [88,]  0.0061563396
#>  [89,] -0.0384689309
#>  [90,] -0.0024159052
#>  [91,]  0.0403071044
#>  [92,]  0.0217459274
#>  [93,] -0.0143887536
#>  [94,]  0.0257435456
#>  [95,] -0.0441222281
#>  [96,] -0.0201434591
#>  [97,] -0.0431365239
#>  [98,]  0.0195870253
#>  [99,]  0.0103439874
#> [100,] -0.0042058643