Generate data points on a Roman surface with optional noise.

This function generates data points on a Roman surface with optional noise.

Usage

roman_surface_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

roman_surface_with_noise(sample_size = 100, num_noise_dims = 8,
min_noise = -0.05, max_noise = 0.05)
#>                 [,1]          [,2]         [,3]         [,4]          [,5]
#>   [1,]  0.0112670943 -0.2503800176 -0.041884248 -0.028826040 -0.0345702759
#>   [2,]  0.1108014887 -0.4008082626 -0.057109968  0.016171250 -0.0146347397
#>   [3,]  0.0589282897  0.3319419974  0.022503174  0.017817638 -0.0214843362
#>   [4,]  0.0323392274 -0.1271059920 -0.234774920 -0.025239864  0.0317104267
#>   [5,] -0.3055865962  0.3786581352 -0.188137880 -0.011190268  0.0109884521
#>   [6,]  0.4699051226  0.0785391688  0.105968164 -0.022084021 -0.0259994029
#>   [7,]  0.0055780947 -0.0607952898 -0.090645729  0.041057258  0.0165233966
#>   [8,]  0.0359103273  0.2075145765  0.160202702  0.013413203  0.0108954034
#>   [9,] -0.4764421255 -0.1364251911  0.149858159 -0.018007969  0.0478668613
#>  [10,] -0.0064733396 -0.0134065602  0.391472591 -0.010978389 -0.0380863543
#>  [11,]  0.0025359855  0.0229956745  0.108897519  0.033291107  0.0405433818
#>  [12,]  0.0704911133 -0.3683281197 -0.153364051 -0.027623199  0.0394046014
#>  [13,] -0.1279786415  0.0243510326 -0.180454964  0.035262649  0.0223742301
#>  [14,] -0.0693017199 -0.1535147467  0.364050463 -0.013798027 -0.0220673154
#>  [15,]  0.0130881508  0.1267277994  0.100501235  0.028633110  0.0282689727
#>  [16,]  0.0814437954  0.0979916900  0.483452057 -0.008688408  0.0317223734
#>  [17,]  0.3812981685 -0.2160149613 -0.111196386 -0.032311760 -0.0287424986
#>  [18,] -0.1055251584  0.0447938902 -0.354885444 -0.040291956 -0.0388093836
#>  [19,]  0.0031562104 -0.4879247236 -0.003940380 -0.028189330  0.0364848986
#>  [20,] -0.0251845406 -0.0048057385  0.183995681 -0.033411203  0.0006753244
#>  [21,]  0.2315712000  0.1624175692  0.428886244  0.031517941 -0.0413087434
#>  [22,] -0.0022775627  0.0326284566 -0.069390101 -0.049973621  0.0024714026
#>  [23,] -0.2661463477 -0.2356997757  0.422669334  0.015083747 -0.0081236064
#>  [24,]  0.1671723211 -0.2899737101 -0.377069696 -0.045228714  0.0016612251
#>  [25,] -0.1042755545  0.1884515199 -0.402967533 -0.019581464 -0.0371618611
#>  [26,] -0.1721982115  0.3740145970 -0.297695861  0.040761493  0.0012303544
#>  [27,]  0.2790894902 -0.3933328105 -0.173721415  0.006128581 -0.0217470058
#>  [28,]  0.0636352532  0.4438907265  0.101302038 -0.031985841 -0.0390290844
#>  [29,]  0.1834862377  0.2520447357  0.423910234  0.041463608  0.0483815329
#>  [30,]  0.1928249950 -0.2083094431 -0.454450402 -0.003826172 -0.0275830796
#>  [31,]  0.0207242669 -0.0376099804 -0.421907184  0.001632080  0.0416619187
#>  [32,] -0.0036433631 -0.4204553030  0.006676639 -0.014273140  0.0255655205
#>  [33,] -0.2837035628  0.3432028613 -0.357220614  0.020633360  0.0490143400
#>  [34,]  0.4295463756 -0.2144076077 -0.255622918  0.006074703  0.0067901140
#>  [35,]  0.2322457403  0.0785443299  0.283989995 -0.018044647  0.0263335739
#>  [36,]  0.0276007687 -0.1420577846 -0.183218236  0.029210575 -0.0454915548
#>  [37,] -0.0073437217 -0.2205604687  0.031551473 -0.034232671 -0.0082151871
#>  [38,] -0.1322922599  0.3023308089 -0.321578495  0.049241700  0.0005874417
#>  [39,]  0.1392679773  0.1319372307  0.480164140 -0.029746261 -0.0460463832
#>  [40,]  0.3755396907 -0.0872230201 -0.180347137  0.015489113 -0.0046162232
#>  [41,] -0.2177853696  0.2570944087 -0.428793849 -0.028378752 -0.0433190885
#>  [42,] -0.2395392805 -0.3187300250  0.095231267 -0.046597696 -0.0021035687
#>  [43,]  0.4977110039  0.0356343528  0.033347480 -0.030152144  0.0068760196
#>  [44,] -0.1406521885  0.2725950177 -0.364642815  0.049735682 -0.0087564819
#>  [45,] -0.0095667870  0.0523915103 -0.176207098  0.016951006  0.0019797613
#>  [46,] -0.0028097455  0.1073286065 -0.025855852 -0.039642237  0.0002425236
#>  [47,]  0.4114335089 -0.2793673515 -0.256412377 -0.040791548 -0.0134390029
#>  [48,] -0.0888570768  0.2977439742 -0.244359764  0.018332420 -0.0125788491
#>  [49,] -0.1123627790  0.1168287798 -0.486128919  0.010323881  0.0128764333
#>  [50,]  0.0798432457 -0.1365086091 -0.424618811 -0.042231025  0.0396782458
#>  [51,] -0.0443062249  0.2517195697 -0.158729407 -0.022218062 -0.0189339438
#>  [52,]  0.1312589662  0.4271039356  0.203411803 -0.033376797 -0.0486160808
#>  [53,] -0.2488679624  0.4295950954 -0.191164705  0.020976390  0.0214220177
#>  [54,] -0.0843231591 -0.4927856349  0.084321685 -0.022064497 -0.0403084384
#>  [55,] -0.0148359895 -0.0865155378  0.165291244 -0.033196967  0.0117365611
#>  [56,] -0.1225556905 -0.3266031689  0.282287382  0.033908374 -0.0390883403
#>  [57,]  0.0120188914 -0.1706100302 -0.067984374  0.025083315 -0.0032461168
#>  [58,] -0.3591593966  0.2036739604 -0.093544553  0.006302504 -0.0415146510
#>  [59,]  0.0114469283  0.4027442486  0.022608310 -0.031491108  0.0406259893
#>  [60,] -0.0788754270 -0.3376234825  0.190572979 -0.002384016 -0.0442244321
#>  [61,]  0.0026709605 -0.0195669573 -0.133954359 -0.036071204  0.0400306634
#>  [62,] -0.2781223841  0.1397506152 -0.357577690 -0.046343946 -0.0150248472
#>  [63,]  0.1393576783  0.4799274597  0.138122858  0.010938109  0.0317940413
#>  [64,]  0.2925470598  0.0006843498  0.002118166 -0.045328125 -0.0134852554
#>  [65,]  0.0835645952 -0.1553830715 -0.403731439  0.039920105 -0.0084975499
#>  [66,]  0.2453329144  0.4101214568  0.155509467 -0.011522938  0.0216591329
#>  [67,] -0.0284934187 -0.2658497510  0.097750161  0.012642769 -0.0358550353
#>  [68,] -0.0002213637 -0.0254320741  0.008697822  0.019208418  0.0251137473
#>  [69,] -0.3943884794  0.2972128219 -0.277927440 -0.003169749 -0.0229185686
#>  [70,] -0.0837156879 -0.2964593544  0.233764094  0.025893283 -0.0491505664
#>  [71,]  0.0191467862 -0.2868699017 -0.060405756 -0.005935082 -0.0370880269
#>  [72,]  0.4660541172 -0.1631247183 -0.180426547 -0.049759064 -0.0252517673
#>  [73,] -0.0164908718  0.0024705832 -0.146485716  0.030303869  0.0174546445
#>  [74,] -0.0423481944  0.0110811024 -0.244428266  0.035277172  0.0147487219
#>  [75,]  0.0003690768 -0.0182590602 -0.020198353  0.019641091 -0.0191235702
#>  [76,] -0.1144262975  0.2843986199 -0.309926737  0.045531239 -0.0413123664
#>  [77,]  0.4569945986 -0.1884712606 -0.202689901  0.036039175 -0.0457727534
#>  [78,] -0.2012532238  0.2640834336 -0.421275017 -0.016480170  0.0202990745
#>  [79,] -0.0005941679 -0.0501548654  0.011815127  0.008040294 -0.0286905462
#>  [80,] -0.0626966802  0.2228870984 -0.245851389  0.033701098  0.0315420593
#>  [81,] -0.0354635860  0.0477350090 -0.477009378 -0.014886814 -0.0421814614
#>  [82,] -0.0079743468  0.1583166437 -0.048946239 -0.047041044 -0.0027744681
#>  [83,] -0.0006307676  0.0198883151 -0.031671068 -0.030336055 -0.0091505090
#>  [84,] -0.0075529185  0.3946345949 -0.015438207 -0.047022241 -0.0029401336
#>  [85,]  0.3136671129  0.3841149065  0.276204506  0.049719100  0.0116912649
#>  [86,] -0.4880458605 -0.0227564853  0.018318905  0.031049244  0.0449625654
#>  [87,]  0.0002944860 -0.0008160929 -0.319275021 -0.043092542 -0.0179133869
#>  [88,] -0.3754063473  0.0872911386 -0.180561643 -0.020199814 -0.0074676022
#>  [89,]  0.0818004783  0.4734963779  0.107065628  0.002056769  0.0116233622
#>  [90,] -0.0188432846  0.4328658738 -0.032567020 -0.009655667 -0.0245214925
#>  [91,] -0.3159436055 -0.2364747852  0.387146111 -0.018393898 -0.0205318506
#>  [92,] -0.0703562437  0.1563767136 -0.362827412 -0.005643091 -0.0222357355
#>  [93,]  0.0887844964  0.2982483803  0.243746610 -0.019719863 -0.0186048803
#>  [94,] -0.3132903828 -0.0072988532  0.020714359 -0.028720872  0.0437605470
#>  [95,]  0.4518111231  0.0410848304  0.026137585 -0.042246420  0.0092998645
#>  [96,]  0.0441110303 -0.1034653864 -0.356139731 -0.018174541 -0.0466514921
#>  [97,]  0.2198372972 -0.4163004437 -0.136917328  0.031198340 -0.0205889446
#>  [98,] -0.0738374989 -0.4157049108  0.039975349  0.030910518  0.0497711945
#>  [99,] -0.0837798696  0.2380279994 -0.021404935  0.008089608 -0.0161344004
#> [100,]  0.0025352580  0.4847840691  0.003254773  0.033149441 -0.0404089978
#>                 [,6]          [,7]          [,8]         [,9]         [,10]
#>   [1,]  0.0086717994 -0.0180867594  1.007316e-02 -0.007879345  0.0466847681
#>   [2,] -0.0096600732 -0.0440294546  7.602486e-03 -0.021776608 -0.0189186354
#>   [3,]  0.0154986767  0.0203350304  4.836226e-02 -0.006253471 -0.0457536923
#>   [4,] -0.0276037465 -0.0048207698  7.485907e-03  0.027014294  0.0322595718
#>   [5,] -0.0124045740 -0.0478485932  4.361382e-02  0.037316216 -0.0374140178
#>   [6,] -0.0077476360  0.0200410618  2.447306e-02 -0.048453007  0.0140886150
#>   [7,]  0.0274174136 -0.0322228189 -3.043182e-02  0.010991402 -0.0159181720
#>   [8,]  0.0040270238 -0.0440464906 -4.280913e-02  0.022521642  0.0404572810
#>   [9,]  0.0105518984  0.0220143986  2.858468e-02 -0.048279414 -0.0294070049
#>  [10,]  0.0217045739 -0.0409585367 -1.233272e-02 -0.008354228 -0.0187916041
#>  [11,]  0.0148561134 -0.0096781398  3.739452e-02 -0.046620980  0.0420218433
#>  [12,] -0.0179137947 -0.0165694851 -1.412247e-02 -0.023341010  0.0113759214
#>  [13,]  0.0185965254  0.0102269463 -4.883551e-02 -0.038795065  0.0194914091
#>  [14,] -0.0237088536  0.0253074061  2.567480e-02 -0.042263743  0.0376473900
#>  [15,]  0.0073001024  0.0440773290  1.311810e-02 -0.027341352 -0.0453477377
#>  [16,] -0.0083404441 -0.0204635165 -4.601929e-03  0.048407354  0.0107788051
#>  [17,] -0.0109154138 -0.0411655299 -1.019276e-02  0.011000763  0.0286695989
#>  [18,]  0.0081855058  0.0121023259  2.616188e-02  0.014096746  0.0099088734
#>  [19,] -0.0273360620 -0.0001927412  2.081466e-02 -0.036756135 -0.0356704202
#>  [20,]  0.0311880877 -0.0464467233  2.124058e-02 -0.045037186  0.0332562773
#>  [21,]  0.0241170527  0.0241207640  3.856885e-02 -0.021212743  0.0325258526
#>  [22,]  0.0206282760  0.0126271059 -1.545231e-02  0.036652267 -0.0199665500
#>  [23,] -0.0436452698 -0.0085726117  1.281148e-02  0.010178040  0.0095872082
#>  [24,]  0.0120903074 -0.0464269341  2.656583e-02  0.001869091  0.0151614635
#>  [25,] -0.0391244607  0.0003895283  1.365955e-02 -0.044560653  0.0297297847
#>  [26,]  0.0077296100 -0.0263277642 -1.187419e-02  0.028149508  0.0017148448
#>  [27,]  0.0214351638  0.0276430749 -3.874524e-02  0.011869375 -0.0019121046
#>  [28,]  0.0295738984 -0.0214901014 -4.272807e-02  0.022116184 -0.0104666633
#>  [29,] -0.0246277100  0.0288995129  3.452309e-02  0.040551778 -0.0134267062
#>  [30,]  0.0129884623  0.0048681622 -4.785787e-02  0.005744344  0.0499538681
#>  [31,] -0.0136698482  0.0197152982 -4.965464e-02  0.026740877 -0.0070373449
#>  [32,] -0.0237356144 -0.0405895164  4.777698e-02  0.041552737  0.0197527064
#>  [33,] -0.0460449710 -0.0366849301 -2.605731e-02 -0.039061269  0.0446013640
#>  [34,]  0.0399247497  0.0165566009  1.061932e-02  0.013193320 -0.0074176979
#>  [35,] -0.0095360322  0.0348286097  3.683637e-02  0.016769596 -0.0240388274
#>  [36,]  0.0237809121  0.0391348526  2.770514e-02 -0.042498769  0.0017490658
#>  [37,] -0.0265897569 -0.0047738070 -3.071743e-03  0.046796027  0.0117035700
#>  [38,]  0.0446329717  0.0434700022  4.924003e-02 -0.016242414  0.0412391958
#>  [39,]  0.0465796639 -0.0444036051 -3.989924e-03  0.032348741  0.0466216319
#>  [40,]  0.0088932418 -0.0116262897  2.035457e-02  0.012686717 -0.0364769112
#>  [41,]  0.0359691297  0.0116728980 -4.327191e-02 -0.008435643 -0.0393408199
#>  [42,]  0.0377992609 -0.0117148395 -2.564482e-04 -0.038320840 -0.0419950508
#>  [43,]  0.0226479868  0.0420067900  2.526306e-06 -0.040998043  0.0076525709
#>  [44,]  0.0471406835 -0.0134444994 -3.743916e-02 -0.043079134  0.0101918599
#>  [45,]  0.0356043358 -0.0127557569 -4.055558e-02  0.015361770 -0.0145364964
#>  [46,] -0.0083269444 -0.0438728270 -2.114311e-02  0.009495271 -0.0324155938
#>  [47,]  0.0420652370  0.0106518093  2.616079e-02 -0.046453813 -0.0086261299
#>  [48,] -0.0463698385 -0.0271777286  6.630231e-03 -0.032597850  0.0044844524
#>  [49,]  0.0249782663  0.0213805594  2.412111e-02  0.038254977 -0.0378923130
#>  [50,] -0.0184477598  0.0409851438  2.192715e-02 -0.007064762  0.0401962903
#>  [51,]  0.0127361750 -0.0214932914  3.486875e-02  0.038515907 -0.0278019578
#>  [52,] -0.0189811438 -0.0446535315  4.521091e-02 -0.001066247 -0.0411472857
#>  [53,]  0.0111943397 -0.0018806768 -4.629014e-02  0.040240459  0.0184574851
#>  [54,] -0.0089402044  0.0344108854 -2.050238e-02  0.002709195 -0.0482999544
#>  [55,]  0.0219312144 -0.0429982035 -1.649687e-02  0.014555320 -0.0147495893
#>  [56,]  0.0065006412  0.0259807939  2.612444e-03 -0.024008043 -0.0237072109
#>  [57,]  0.0359971470 -0.0468872605  8.905704e-03  0.041804165 -0.0447006607
#>  [58,] -0.0322065346 -0.0445443711  4.107293e-02  0.046631034  0.0053583174
#>  [59,] -0.0106821229  0.0255828719 -1.085374e-02  0.025850175  0.0149870262
#>  [60,]  0.0491066004 -0.0274983546  4.624655e-02 -0.018441679 -0.0194252739
#>  [61,]  0.0022285686 -0.0451228415  9.179250e-03  0.038848126 -0.0329721940
#>  [62,] -0.0208186478  0.0255118455 -1.937183e-02  0.006612506  0.0011094775
#>  [63,]  0.0410319719 -0.0383269426  2.872208e-02 -0.026721130 -0.0472631751
#>  [64,] -0.0391432331  0.0480801035 -3.429668e-03  0.015768161  0.0034039608
#>  [65,] -0.0124481008  0.0072103263  4.639444e-02  0.048813371  0.0057868574
#>  [66,] -0.0223432432  0.0464571290  3.079388e-02  0.023098534  0.0026351465
#>  [67,] -0.0308542207  0.0055095966 -2.392111e-02  0.025698603 -0.0345325483
#>  [68,] -0.0064586700 -0.0084661917 -1.858524e-02 -0.032563970 -0.0106092751
#>  [69,]  0.0139473788  0.0383355768  4.453178e-02  0.039355427 -0.0074511725
#>  [70,] -0.0349745444 -0.0472831185 -4.373840e-02 -0.020161546  0.0379824675
#>  [71,] -0.0080926448 -0.0272920318  2.741082e-02  0.037732959  0.0010086544
#>  [72,] -0.0453974129  0.0276836755  3.680147e-02  0.032860585  0.0195156724
#>  [73,]  0.0319150536  0.0475209947  3.016436e-02 -0.018351691  0.0357966234
#>  [74,] -0.0368271134 -0.0330939712  1.546929e-03  0.009926033  0.0410892200
#>  [75,] -0.0008664426  0.0260262657  4.702930e-03  0.024556085  0.0212416339
#>  [76,]  0.0345087101 -0.0276171440 -1.051104e-02 -0.017629060 -0.0214431878
#>  [77,]  0.0167151886  0.0362959911  3.349667e-02 -0.024353430 -0.0275968639
#>  [78,] -0.0385699542  0.0404319805 -6.473463e-03 -0.036661414  0.0038865911
#>  [79,] -0.0311343115 -0.0098752119 -1.159156e-03 -0.049251285  0.0395272136
#>  [80,] -0.0483070843 -0.0124441562  6.946413e-03 -0.025133566 -0.0367604721
#>  [81,]  0.0093214929 -0.0146904925  4.047586e-02 -0.036483532 -0.0462253720
#>  [82,] -0.0265211389  0.0376282437  1.205973e-02  0.047112598  0.0066573617
#>  [83,] -0.0494939435 -0.0009647355  3.659714e-02 -0.024932071  0.0456295161
#>  [84,] -0.0254374065  0.0133930647 -4.275474e-03 -0.026466577  0.0465525918
#>  [85,] -0.0055351067 -0.0076462828 -1.760346e-02  0.003599008  0.0034458604
#>  [86,] -0.0319345207 -0.0408399834  3.418689e-02  0.044070127 -0.0354715583
#>  [87,] -0.0214401117  0.0438319675 -4.388362e-04  0.024408111 -0.0007203238
#>  [88,] -0.0175544841 -0.0272564620 -4.094549e-03 -0.048926892 -0.0313109347
#>  [89,] -0.0460089596  0.0341886993 -5.761860e-03  0.011787652  0.0208889419
#>  [90,]  0.0035911330 -0.0177121108  4.924419e-02  0.003125632  0.0186344353
#>  [91,]  0.0441038593 -0.0262666935 -1.250563e-02  0.013318046  0.0304781804
#>  [92,]  0.0111912511 -0.0298964594 -7.915830e-03 -0.034620880  0.0100462520
#>  [93,] -0.0473354339  0.0210790344  4.048820e-02  0.016842412  0.0230813032
#>  [94,]  0.0366435696 -0.0080923515 -1.964941e-02 -0.015598666 -0.0270949428
#>  [95,]  0.0456695389  0.0010575586  1.100112e-02 -0.043950538  0.0246321565
#>  [96,] -0.0481847727  0.0217221439  4.155372e-02 -0.001885981 -0.0399403666
#>  [97,]  0.0058095998 -0.0316973461  3.507448e-02 -0.014619852  0.0316685400
#>  [98,]  0.0012660705  0.0115414259 -5.790612e-03 -0.010586411 -0.0297381790
#>  [99,] -0.0077317495 -0.0151142128 -1.495180e-02 -0.043114460  0.0087783528
#> [100,] -0.0158986821  0.0190606681 -3.424923e-02 -0.048378203  0.0090645304
#>                [,11]
#>   [1,] -0.0385476567
#>   [2,] -0.0406498368
#>   [3,] -0.0200607848
#>   [4,]  0.0172267342
#>   [5,]  0.0330983246
#>   [6,] -0.0215472674
#>   [7,] -0.0439373853
#>   [8,]  0.0246854331
#>   [9,]  0.0180129585
#>  [10,] -0.0411398747
#>  [11,] -0.0280898720
#>  [12,] -0.0164136645
#>  [13,] -0.0047535486
#>  [14,]  0.0331181310
#>  [15,] -0.0273228670
#>  [16,] -0.0400744475
#>  [17,] -0.0249269130
#>  [18,]  0.0125764568
#>  [19,] -0.0072356102
#>  [20,] -0.0056882872
#>  [21,] -0.0165847495
#>  [22,] -0.0368099707
#>  [23,]  0.0303939795
#>  [24,] -0.0216553068
#>  [25,] -0.0175301095
#>  [26,] -0.0387954956
#>  [27,] -0.0406635953
#>  [28,]  0.0119414818
#>  [29,] -0.0159614631
#>  [30,] -0.0395850434
#>  [31,] -0.0004252076
#>  [32,] -0.0073889060
#>  [33,]  0.0129034097
#>  [34,]  0.0397990522
#>  [35,]  0.0347135378
#>  [36,]  0.0345581067
#>  [37,]  0.0442037697
#>  [38,]  0.0215627823
#>  [39,]  0.0458120537
#>  [40,]  0.0331213964
#>  [41,] -0.0384675396
#>  [42,] -0.0068945265
#>  [43,] -0.0435688308
#>  [44,] -0.0345581794
#>  [45,]  0.0318632875
#>  [46,] -0.0427309742
#>  [47,] -0.0335404708
#>  [48,] -0.0267851235
#>  [49,] -0.0270596502
#>  [50,]  0.0450807667
#>  [51,] -0.0213964997
#>  [52,]  0.0314357933
#>  [53,]  0.0392193651
#>  [54,] -0.0455143865
#>  [55,] -0.0384210555
#>  [56,]  0.0376279234
#>  [57,] -0.0035268292
#>  [58,]  0.0186325667
#>  [59,] -0.0298709111
#>  [60,]  0.0119173280
#>  [61,]  0.0199845805
#>  [62,]  0.0351337904
#>  [63,] -0.0207638512
#>  [64,]  0.0310470119
#>  [65,] -0.0316704961
#>  [66,] -0.0403732269
#>  [67,] -0.0054029447
#>  [68,] -0.0054374511
#>  [69,]  0.0353352583
#>  [70,]  0.0238171520
#>  [71,]  0.0423776192
#>  [72,] -0.0003772799
#>  [73,]  0.0415619799
#>  [74,]  0.0362006874
#>  [75,] -0.0199759330
#>  [76,]  0.0405837991
#>  [77,] -0.0158166261
#>  [78,]  0.0377023071
#>  [79,]  0.0088444529
#>  [80,]  0.0121073170
#>  [81,]  0.0169951685
#>  [82,]  0.0318651332
#>  [83,] -0.0080241325
#>  [84,]  0.0297827245
#>  [85,]  0.0362757991
#>  [86,] -0.0405948739
#>  [87,]  0.0094786922
#>  [88,] -0.0286599018
#>  [89,] -0.0149476970
#>  [90,]  0.0119228740
#>  [91,] -0.0102677293
#>  [92,]  0.0176897420
#>  [93,] -0.0289334347
#>  [94,] -0.0446933800
#>  [95,] -0.0077594112
#>  [96,] -0.0288171692
#>  [97,] -0.0480092557
#>  [98,] -0.0388135414
#>  [99,]  0.0447368561
#> [100,] -0.0311190218