Generate data representing small spheres within a larger encompassing sphere with added noise.

This function generates data points representing small spheres within a larger encompassing sphere and adds noise to the data if specified.

Usage

small_big_sphere_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate, should be a multiple of 13.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

small_big_sphere_with_noise(sample_size = 390, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>                [,1]         [,2]         [,3]          [,4]          [,5]
#>   [1,]  -9.03493657  -1.58063748  -3.56100181   8.857854437 -0.0142654950
#>   [2,]  -6.08565121   0.84922534  -3.57017547  12.581455148 -0.0354762347
#>   [3,] -10.35859329  -1.95081778  -2.54030445  13.213074734  0.0366188484
#>   [4,]  -7.95793899  -3.59947657  -4.48898970  10.463622795 -0.0448334731
#>   [5,] -10.11939634  -2.23192028  -4.81843746  12.360747754  0.0409425216
#>   [6,]  -7.14774204   0.35752631  -2.10826580   9.412855855 -0.0004098609
#>   [7,]  -8.44525667  -3.77881848  -1.73684503  11.149960054 -0.0466561069
#>   [8,]  -8.90336935  -2.62847969  -5.45002673  10.783159090  0.0227826453
#>   [9,]  -9.99317232  -3.35624726  -3.70327584  11.450394714 -0.0292143598
#>  [10,]  -6.39071783  -2.90854727  -2.41430560  13.331942900  0.0174030380
#>  [11,]  -8.29953569   0.36417449  -4.75759072  13.562308259 -0.0270248232
#>  [12,]  -8.50177447  -0.07702160  -5.32816916  13.231961271  0.0245905064
#>  [13,]  -8.46227042  -0.07760567  -5.20541218  13.398487345 -0.0374791581
#>  [14,]  -7.27970131  -1.47843047  -5.81229101  10.552819812 -0.0378769194
#>  [15,]  -6.74202277  -0.55967455  -2.68741756   9.002562128  0.0288486092
#>  [16,]  -6.12662340  -1.78424491  -1.62832038   9.909940282 -0.0053544863
#>  [17,]  -6.50763077   1.37571478  -3.11462794  12.010815580 -0.0025114908
#>  [18,]  -9.22650489  -1.81722649  -0.44846202  11.590953528 -0.0179689034
#>  [19,]  -5.40412636  -2.10283126  -2.78512484  10.477066780 -0.0287334990
#>  [20,]  -7.29587842  -1.34413472  -0.71469195   9.969565714  0.0135804157
#>  [21,] -10.84253506  -1.13278064  -3.33011574  10.661974084 -0.0285090832
#>  [22,]  -8.34435035  -2.15061359  -5.18740118  13.532001710 -0.0022062201
#>  [23,]  -9.97654104  -2.48475450  -1.27343515  11.390124560 -0.0363343889
#>  [24,] -10.38412255  -0.28415843  -4.59568331  11.037011147 -0.0110172283
#>  [25,]  -9.26020608  -3.31550904  -3.85947789  10.074254693 -0.0350364899
#>  [26,]  -6.20724524  -0.26065384  -1.33320035  10.304221746 -0.0099672439
#>  [27,]  -6.74282748  -0.64442517  -5.18851616   9.936737170 -0.0360482751
#>  [28,]  -6.03115538  -1.01154021  -0.91197167  11.160098859  0.0305348763
#>  [29,]  -6.01103953  -0.95871786  -3.93648425  13.685511776 -0.0032009949
#>  [30,]  -7.30621173  -0.75794943  -0.65310723  10.113050136 -0.0005195178
#>  [31,] -11.63924737  -1.62077785   3.71907827  -9.280985209  0.0378959184
#>  [32,]  -8.68996201   0.80908497   3.70990461  -5.557384498  0.0028261860
#>  [33,] -12.96290409  -1.99095815   4.73977563  -4.925764912 -0.0431825544
#>  [34,] -10.56224978  -3.63961694   2.79109038  -7.675216851 -0.0265758651
#>  [35,] -12.72370713  -2.27206065   2.46164262  -5.778091892 -0.0248394041
#>  [36,]  -9.75205283   0.31738594   5.17181428  -8.725983791  0.0138229773
#>  [37,] -11.04956747  -3.81895885   5.54323505  -6.988879593  0.0135592397
#>  [38,] -11.50768014  -2.66862006   1.83005336  -7.355680556 -0.0344451946
#>  [39,] -12.59748312  -3.39638763   3.57680424  -6.688444932  0.0084846190
#>  [40,]  -8.99502863  -2.94868764   4.86577449  -4.806896746  0.0394349312
#>  [41,] -10.90384649   0.32403412   2.52248936  -4.576531387  0.0480947239
#>  [42,] -11.10608527  -0.11716197   1.95191092  -4.906878376 -0.0137550354
#>  [43,] -11.06658122  -0.11774604   2.07466790  -4.740352301  0.0354823546
#>  [44,]  -9.88401211  -1.51857084   1.46778907  -7.586019835 -0.0251701997
#>  [45,]  -9.34633356  -0.59981492   4.59266252  -9.136277518 -0.0090644483
#>  [46,]  -8.73093420  -1.82438528   5.65175970  -8.228899364 -0.0441642432
#>  [47,]  -9.11194157   1.33557441   4.16545214  -6.128024066 -0.0274463836
#>  [48,] -11.83081569  -1.85736686   6.83161807  -6.547886118  0.0398637896
#>  [49,]  -8.00843716  -2.14297163   4.49495524  -7.661772866  0.0241649840
#>  [50,]  -9.90018922  -1.38427509   6.56538814  -8.169273932 -0.0363729670
#>  [51,] -13.44684586  -1.17292101   3.94996434  -7.476865562 -0.0350432127
#>  [52,] -10.94866115  -2.19075396   2.09267891  -4.606837937 -0.0313538057
#>  [53,] -12.58085184  -2.52489487   6.00664493  -6.748715086 -0.0418471035
#>  [54,] -12.98843335  -0.32429880   2.68439677  -7.101828499 -0.0279612498
#>  [55,] -11.86451688  -3.35564941   3.42060219  -8.064584953 -0.0237663522
#>  [56,]  -8.81155604  -0.30079421   5.94687973  -7.834617900 -0.0020727467
#>  [57,]  -9.34713828  -0.68456554   2.09156392  -8.202102476  0.0223322747
#>  [58,]  -8.63546618  -1.05168058   6.36810842  -6.978740787  0.0498677238
#>  [59,]  -8.61535033  -0.99885823   3.34359584  -4.453327870  0.0217301269
#>  [60,]  -9.91052253  -0.79808980   6.62697285  -8.025789511 -0.0013216135
#>  [61,]  -3.09523544   1.34432147  -4.27020465  -4.235014367 -0.0313574036
#>  [62,]  -0.14595008   3.77418428  -4.27937831  -0.511413655  0.0030314952
#>  [63,]  -4.41889216   0.97414117  -3.24950729   0.120205931  0.0113520778
#>  [64,]  -2.01823785  -0.67451762  -5.19819254  -2.629246009 -0.0259533292
#>  [65,]  -4.17969520   0.69303867  -5.52764030  -0.732121049 -0.0498913975
#>  [66,]  -1.20804090   3.28248526  -2.81746864  -3.680012948 -0.0379498681
#>  [67,]  -2.50555554  -0.85385953  -2.44604787  -1.942908750 -0.0459555196
#>  [68,]  -2.96366821   0.29647926  -6.15922956  -2.309709713 -0.0001875577
#>  [69,]  -4.05347119  -0.43128831  -4.41247868  -1.642474089  0.0050635456
#>  [70,]  -0.45101670   0.01641168  -3.12350843   0.239074097 -0.0130623515
#>  [71,]  -2.35983456   3.28913343  -5.46679356   0.469439456  0.0168525227
#>  [72,]  -2.56207334   2.84793735  -6.03737200   0.139092467 -0.0421744693
#>  [73,]  -2.52256929   2.84735328  -5.91461502   0.305618542 -0.0359327714
#>  [74,]  -1.34000018   1.44652848  -6.52149385  -2.540048992  0.0209681712
#>  [75,]  -0.80232163   2.36528439  -3.39662040  -4.090306675  0.0187252526
#>  [76,]  -0.18692227   1.14071404  -2.33752322  -3.182928521  0.0430619818
#>  [77,]  -0.56792964   4.30067373  -3.82383077  -1.082053223  0.0134724254
#>  [78,]  -3.28680376   1.10773246  -1.15766485  -1.501915275 -0.0459288234
#>  [79,]   0.53557477   0.82212768  -3.49432768  -2.615802023  0.0251172174
#>  [80,]  -1.35617729   1.58082423  -1.42389478  -3.123303089  0.0123699499
#>  [81,]  -4.90283393   1.79217830  -4.03931858  -2.430894719  0.0424742981
#>  [82,]  -2.40464922   0.77434535  -5.89660401   0.439132906  0.0275251617
#>  [83,]  -4.03683991   0.44020445  -1.98263798  -1.702744244  0.0035650667
#>  [84,]  -4.44442142   2.64080052  -5.30488614  -2.055857656  0.0429790955
#>  [85,]  -3.32050495  -0.39055010  -4.56868073  -3.018614110 -0.0390591999
#>  [86,]  -0.26754411   2.66430510  -2.04240319  -2.788647057 -0.0489502125
#>  [87,]  -0.80312635   2.28053378  -5.89771900  -3.156131633 -0.0404855302
#>  [88,]  -0.09145425   1.91341873  -1.62117450  -1.932769944  0.0375176969
#>  [89,]  -0.07133840   1.96624108  -4.64568708   0.592642973  0.0254410411
#>  [90,]  -1.36651060   2.16700952  -1.36231006  -2.979818668  0.0317330339
#>  [91,]   5.77879675  -9.84990329  -4.78626464   8.466202466  0.0238184077
#>  [92,]  -2.32513107  -1.05872948  14.58200506   2.416192422 -0.0137804546
#>  [93,]   7.16377092   8.72946297  -8.25165635   5.420980528 -0.0378008378
#>  [94,]  -1.88421243   4.23826850   0.43220930 -14.258331557 -0.0433688580
#>  [95,]   5.99744370  -1.05265316 -13.56114290   2.004493361  0.0109652535
#>  [96,]   8.96693425   1.09520942  11.29595915  -3.974407311 -0.0008206311
#>  [97,]   0.89591173  -0.06525386 -14.67463441   2.974590578  0.0085780721
#>  [98,] -11.92076580  -8.52297493   2.32357479  -2.203461191 -0.0008441911
#>  [99,] -12.15532688  -2.25702983  -7.46444661   4.054119078 -0.0296755642
#> [100,]  -2.15107320  -7.31306596   8.62375727   9.618875239 -0.0448833886
#> [101,]   1.63252352  -5.37589205   0.09643606 -13.907744307  0.0136083371
#> [102,]  -7.70513946   6.17896824  10.61093746  -3.854761166 -0.0301518152
#> [103,] -11.39566464  -7.95825694   4.86887633   2.845877210 -0.0270571755
#> [104,]   1.99070058  -5.99827251  -7.60131901  11.281745757 -0.0061548268
#> [105,]   8.20382865  10.01803263   6.47511791   3.925438294  0.0118858161
#> [106,]  -3.22153485   9.06459572   3.04194974 -11.099610781 -0.0166470921
#> [107,]  -8.63128007   4.75881326 -10.86714727   3.124069598 -0.0074348239
#> [108,]   5.15414214   7.98062552  10.58260422  -4.770002424  0.0017691261
#> [109,]  -1.04951207   9.08554372  -9.89497336   6.590972753 -0.0027092509
#> [110,]  10.95762595   6.85458268  -6.72049910   3.574915602  0.0423002261
#> [111,] -12.87081784   0.26642417  -5.95717208   4.876798859  0.0224883411
#> [112,]   5.28918882 -12.28713718  -4.78616312   4.810757127  0.0146759835
#> [113,]  -4.06819260  -7.93816152  11.99377487  -1.258874471 -0.0422295440
#> [114,]  -0.35824119  12.34955852  -7.07070712  -4.729182644  0.0258545735
#> [115,]  -1.68312128  10.62229848   7.45677978  -7.330096391  0.0477003537
#> [116,]   7.75052011  -8.10523998   3.33105609  -9.388215396  0.0030060765
#> [117,]  10.29608768  -7.05827029   6.97038375  -4.537086014 -0.0265506459
#> [118,]  -5.74020371  -4.94722539 -12.93512370  -0.507540341  0.0078081354
#> [119,]  10.35753465  -4.98055198   6.49342342  -7.123975731 -0.0019698547
#> [120,]  12.04568845  -4.45826158   7.42061547   2.227051721  0.0443911214
#> [121,]  -0.57223006  -4.70471030  12.74040516   6.341950023 -0.0091306838
#> [122,]  -0.53698769  -1.80978474  -8.32107669 -12.336774540 -0.0214349334
#> [123,]  -5.54962679  -2.42154951 -13.16085630   3.889678887  0.0486329470
#> [124,]  -0.44868464 -14.48236203  -1.37702086  -3.628179398 -0.0322591547
#> [125,] -11.40752912   5.03564060   4.83523906  -6.791985452 -0.0401585845
#> [126,] -10.88928030   5.64219593  -8.54007664   1.286969615 -0.0152198908
#> [127,]  -6.70397769   6.76691925 -11.56636851   0.696136907 -0.0395091036
#> [128,]  12.99750333   5.65901619  -4.80038614  -0.998366593 -0.0002048393
#> [129,]  -7.44767158   3.11367169   4.26438619  11.901774956 -0.0185578629
#> [130,]   5.29932140  -1.61848039  -3.42342193  13.512138847 -0.0142491821
#> [131,]   2.07261623  13.39516483  -2.38483085   5.965434017 -0.0464488313
#> [132,] -13.62576674   3.56969407  -3.72786098   3.563540040  0.0350199315
#> [133,]  13.36581497  -1.13295317   4.84746025  -4.644732107  0.0302801593
#> [134,]   7.49112598 -12.69219133  -1.64036200   2.258433812 -0.0058469320
#> [135,] -12.25565637   2.49336075  -6.73437041  -4.819781554 -0.0041982303
#> [136,]  -9.94877402  -9.46017743   2.14350690   5.650868663  0.0039280682
#> [137,]   2.29746920  -2.98512540 -11.55790115   8.787808751 -0.0283623309
#> [138,]   3.10364161  13.85904281  -4.63067667  -1.360578843 -0.0396664866
#> [139,]  -0.30653012   7.68230724  10.56173461   7.371428403  0.0389704145
#> [140,]   5.40037909  -2.85785645  13.52373818  -2.185650462 -0.0398126647
#> [141,]  12.60214673  -3.62620673   1.27622028  -7.169922195  0.0051808101
#> [142,]   3.17379143 -11.49733452   9.08091289  -0.524755187 -0.0081318312
#> [143,]  -2.72385224  -8.59060493  -2.80666886 -11.657819085 -0.0366609986
#> [144,]  -2.78374568  12.64868632   0.11381907   7.566276463  0.0167757102
#> [145,]   3.62317336 -11.00538097  -4.63258461  -8.324263594  0.0376606197
#> [146,] -13.02107871   5.40626911  -5.08008147   0.645395911 -0.0272550233
#> [147,]  -2.26888580   9.76715680  -5.23492623   9.851413742 -0.0264171684
#> [148,]   7.69225375 -10.31462407  -4.63485593   6.160833804 -0.0118820909
#> [149,]   7.86910605  -1.38073912  -4.39702361 -11.909530339 -0.0411726092
#> [150,]   8.13407566   1.67007306  -8.35444479   9.287137411  0.0276079002
#> [151,]   7.17375063  -9.46180059   2.18874026  -8.900620631  0.0010622081
#> [152,]  13.46763269  -5.27421182   2.04388370   3.409999823  0.0479064680
#> [153,] -10.35231538  -1.46032286  -8.29768520  -6.844373145  0.0040199017
#> [154,]   0.40770501 -14.46965580  -2.52668443   3.013088698 -0.0245137195
#> [155,]  -5.05817627  11.86656625   1.33545407   7.537640256 -0.0441416510
#> [156,]   3.23912778  -7.09708695   5.22567354 -11.697510165 -0.0121284076
#> [157,]   9.33496530  10.30172195  -4.15892206   3.799515115 -0.0199902450
#> [158,]  -3.73999532  -4.04112057  13.94367713   0.505616207 -0.0337024454
#> [159,]   8.97709411 -11.20126547  -3.43373942  -2.674484419 -0.0400298420
#> [160,]   4.29906966  11.46913640   1.28391413   8.563204708 -0.0106446010
#> [161,]  -2.83756057   3.34860634  -9.94602524 -10.334973025 -0.0051058598
#> [162,] -12.29039604   5.14280504  -2.76244923  -6.313999975 -0.0292815140
#> [163,]  12.88246818   3.72040553   3.67097342  -5.632455087  0.0258847094
#> [164,]   3.69759495   5.65405087 -12.79766549  -3.947056927 -0.0478834353
#> [165,]  -2.08062126   4.47020184   2.24546527 -13.987358451 -0.0407180830
#> [166,] -13.60637403  -5.07183850  -1.77609731   3.314893406 -0.0459258026
#> [167,]   1.59460169  -6.29174307  -8.62929178  10.411846047  0.0062494586
#> [168,]  -5.09775872   5.50098783  12.50341494   3.523720162  0.0109179499
#> [169,]   1.58559803  -2.86489220  -9.54016056 -11.102414518 -0.0346075289
#> [170,]  14.38261648   4.07932790  -0.81660639   0.912458822 -0.0066255765
#> [171,]  -7.46812840  -4.15071751  10.39746486  -6.625052968 -0.0016393725
#> [172,]   7.89128977   8.53596190   5.79192219  -7.504567776 -0.0026284708
#> [173,]  11.01947292  -7.23953985  -1.72864612   6.940609623 -0.0104732386
#> [174,]  -0.64817764  10.35766869   7.20700685   8.084405813  0.0264232928
#> [175,]   2.25265053  -0.76798881  11.21622263  -9.671199966 -0.0170200895
#> [176,]   8.38478247  -5.25763283  -3.48800204  10.718514904  0.0400309947
#> [177,]   1.34199569   7.99724197  12.48917135  -1.806590005 -0.0281232769
#> [178,]  -6.64212549   5.16332715   0.86885046 -12.388192789  0.0350281331
#> [179,]   4.58307045   1.13110695  10.13094022 -10.004004830  0.0190916865
#> [180,]   5.18779153   8.50950027  10.94575971   2.421893605 -0.0492332134
#> [181,]   4.52742896 -13.06291356   1.88180827  -5.506493801 -0.0387078005
#> [182,]  -2.91726418  -2.48594259  -9.81719546 -10.673908955  0.0367262162
#> [183,] -12.14752426  -3.05946400  -5.00190426   6.561881460  0.0425536354
#> [184,]   6.47302875  12.45715515   1.09431630   5.169299385 -0.0339544867
#> [185,]   4.42143221   4.97393927  -2.37188247  13.231970330  0.0029737594
#> [186,]   8.69112195   0.69906942   5.35737127 -10.966962854  0.0283511394
#> [187,]  -7.69670366   9.51999373   8.04863311  -3.217138052  0.0194406254
#> [188,]   3.83304104   7.64909189   4.71929239  11.381013530  0.0096749337
#> [189,]   1.68889371   5.39235605   6.40660511 -12.329863955 -0.0040005547
#> [190,] -13.42036627   1.23651164  -4.80349608   4.504579172  0.0153989097
#> [191,]  14.49463182  -3.84804056  -0.19127301   0.248288035  0.0277423693
#> [192,]  -6.55398010  -1.76034169   8.24879589 -10.531092459  0.0024594139
#> [193,]  -7.89440368   8.55313982  -0.26998135  -9.457764000  0.0434584982
#> [194,]  -4.40080632  -7.02314641 -12.37770725   1.760875220  0.0491912927
#> [195,] -12.90332416  -3.48061338   1.28553272  -6.688569488  0.0039507252
#> [196,] -10.69496590   8.61848462  -5.72443958   1.889502208 -0.0339367461
#> [197,]   5.23013090   2.00306605   8.26131006  11.197509244 -0.0351696699
#> [198,] -12.06907550   8.16208795  -0.88174553   3.455468367  0.0346562735
#> [199,]   4.98500663  -7.76476719  11.54353105   2.570017550 -0.0353590870
#> [200,]  11.45660532   9.54629604   0.87460835  -1.359958296 -0.0341653890
#> [201,]   4.02491575  -2.03641281  -8.42023578 -11.565150474 -0.0486927935
#> [202,]   5.59529255   9.08609025   7.79996330  -7.091984048  0.0186409093
#> [203,]   1.26154617  -3.75105582 -11.58885982   8.662355897 -0.0046425507
#> [204,]   9.21013650  -4.06059801   3.51645265  10.550805187 -0.0059940758
#> [205,]   3.52310302   3.94372547 -12.88209792  -5.575511426 -0.0254501086
#> [206,] -11.32633042   2.56828958   9.48955487   0.257830196 -0.0138193422
#> [207,]   0.67882240  -1.97201800   4.11921851 -14.271733743 -0.0498462459
#> [208,]   5.53763774  11.82189625   4.55926056  -5.812957977 -0.0371268279
#> [209,]  -3.85331341 -12.72297416  -6.93978974   0.342377879 -0.0095759856
#> [210,]  -8.60948248  -5.88391428   9.41016900   5.263561877  0.0237916254
#> [211,]   6.16327618   3.80645057  -4.26133540  12.424410708  0.0027620777
#> [212,]  12.47859905  -3.40213232   2.75071991  -7.081214684  0.0466033558
#> [213,] -11.52340917   6.40547931   0.02209895   7.154046931  0.0309553012
#> [214,]   1.84853438 -13.11215463  -6.89844088   1.437301329 -0.0308160886
#> [215,]  10.45325720   6.75354724   6.48008972   5.303531919  0.0250533162
#> [216,]   6.39015616   0.21006109   4.95140329  12.633502445 -0.0242647322
#> [217,]   6.90267089  -4.76705450 -12.13452934   2.716895972 -0.0290650770
#> [218,]   4.10143979  -0.60964709   9.57458310  10.776543068 -0.0302366883
#> [219,]   0.19359786  -2.09154779   9.22860090  11.637047440  0.0317684560
#> [220,]  -4.75265690  -1.10428288  14.16710070  -0.697186803  0.0017692390
#> [221,]  -0.67273213  -2.37654517  -8.50474685  12.106557959 -0.0237067814
#> [222,]  10.78945478  10.23421269  -0.98484722   1.696653229  0.0135680469
#> [223,]   8.13363646  11.92277329  -4.05082468   0.531276217 -0.0100358631
#> [224,]  -2.58998123   4.27393627  13.94001089   2.387794477  0.0023742913
#> [225,]  -3.53292878  -5.75261718   7.20752646 -11.290587767  0.0404959604
#> [226,]  13.06257469  -5.86057628  -1.53533206  -4.203039797 -0.0449451298
#> [227,]  -2.25236318  10.73060304   9.54621692   3.694693637 -0.0032331870
#> [228,]  -4.47746169   8.32296793   8.42859732  -8.039856273 -0.0189601859
#> [229,] -13.11797898   6.11987167   3.54530856  -1.701935805 -0.0108713281
#> [230,]   4.11593328   9.63705452   8.45639880  -6.608751231 -0.0378699395
#> [231,]  -2.37593083  -2.20454832  -8.90307915 -11.628847799  0.0052085525
#> [232,]  -9.91261276   0.62620930   6.05741632  -9.468668208 -0.0381767166
#> [233,]  -9.35130013   7.31146272   2.03621590  -8.941449743  0.0304293933
#> [234,]  -2.21767713  -2.60729722  13.81830777  -4.726338933 -0.0080539073
#> [235,]   4.44536641 -11.20738620   6.18705446   6.430674091 -0.0203363427
#> [236,]  -6.84764699  -4.38979388  -9.40580711   8.388696741  0.0178254703
#> [237,]  -0.97528862  -2.11366208   4.54917520  14.102703632  0.0281832897
#> [238,] -14.26229868  -0.59336187   4.40174226   1.363606709 -0.0237727214
#> [239,] -10.83301352   6.03911584  -6.44527386   5.443651610  0.0367811327
#> [240,]  -4.18107659 -12.65393208   5.99478704   3.385133623  0.0167824472
#> [241,]  -5.73825531 -13.06891253   4.30065053   1.667440019  0.0095084150
#> [242,]  -0.73244291  13.89964633  -5.47725643  -1.123842172 -0.0044107034
#> [243,]   8.06408385  -8.45588058   0.38757718  -9.397787995  0.0374312584
#> [244,]  10.99580877  -1.93032877   1.96072403   9.824539771  0.0376135420
#> [245,]  -0.58282808  -9.11070682  11.80273364   1.533235564 -0.0465826655
#> [246,] -13.01068316  -2.11754196   3.06052506   6.470805687 -0.0496049841
#> [247,]  -8.88573118 -11.88158651   0.57090581  -2.132076483 -0.0028885404
#> [248,]  -2.09045168   3.69687576  10.31723728  10.025853401 -0.0078669869
#> [249,]   7.22321741  12.38789577  -0.36528324   4.385400417  0.0438675902
#> [250,]   5.96986868  -9.27009716   9.03865469  -4.661404077  0.0208439457
#> [251,]  10.67079330   1.91824082   3.87532078   9.614385649 -0.0376516746
#> [252,] -10.18077884  -0.12306607  10.99996913   0.580754715  0.0286852656
#> [253,]   3.08832368  -3.00616558   5.47634164 -13.282880245 -0.0055488282
#> [254,]   7.68806094 -11.44355987  -3.41530141  -4.824352057  0.0177392087
#> [255,]   2.48637831  -8.28705476   8.13781157   9.160713356  0.0289964651
#> [256,]   7.21024113   6.20899459  11.06465338   3.469042296  0.0113598105
#> [257,]   1.07308038   1.11683402   4.73134328  14.149755160 -0.0010038246
#> [258,]  -6.37342319  -4.68958817  10.72113274   6.888000616 -0.0140679195
#> [259,] -10.16906712  -6.87750059  -2.21079026   8.330814241 -0.0083117526
#> [260,]  -1.74814616  -2.05791849  -4.24120741  14.132272152 -0.0190250016
#> [261,]  -4.04839565  -4.53648824   9.98062428  -9.403079611  0.0377528639
#> [262,]   4.45724234 -13.13119753   5.45575871   1.714450039  0.0200291937
#> [263,] -13.51310138  -0.31635671   4.55427058  -4.642696301 -0.0407539491
#> [264,]  -6.52440658  10.10463480  -1.14696618  -8.888922483 -0.0396335109
#> [265,]   1.67518021   1.47937283   7.11671136 -13.013748371 -0.0103357211
#> [266,]   0.72161757  10.35999618 -10.20333977  -3.611316201 -0.0322253232
#> [267,]   5.95246603  -7.38722641  -9.96926657  -5.967475030 -0.0107373370
#> [268,]  -4.90546248  -0.70684679  12.08297148  -7.378252207 -0.0182523262
#> [269,]   6.54580429  12.87857845   2.61837009   3.072263231  0.0095809901
#> [270,] -13.66902108   4.95223886  -1.05422385   3.538616254 -0.0046864697
#> [271,]   1.24467510   1.26361015  -1.91953662  14.770560329  0.0068546325
#> [272,]   8.27398959   3.91126538   6.26698645 -10.097919602  0.0481068782
#> [273,]   4.74643372  -5.51040659   7.88425187  10.485483233 -0.0104372159
#> [274,]   7.81475558  -9.13067661  -2.59039421   8.593613773 -0.0093141185
#> [275,] -11.33870018   1.41467645   9.48415251   2.117408775  0.0322931050
#> [276,]  -7.11708432  11.24900491   2.85227172   6.298535170 -0.0276243568
#> [277,]   2.47382488 -14.22916279  -2.40735839  -3.258180843  0.0156757499
#> [278,]  -6.96202500   3.49210312  11.47981646  -5.705193933  0.0220252863
#> [279,]  -8.82388656  -5.70082457  -4.85685755  -9.542041706 -0.0144242952
#> [280,]   2.16500703  -8.65926798 -11.72073758  -2.820307309 -0.0436351998
#> [281,]  -6.04159595 -12.21643623  -6.25393650   0.382207478 -0.0278154743
#> [282,]  -1.77708995 -12.77586750  -4.11859518   6.454171883 -0.0304685674
#> [283,]  -0.26036539 -10.89610237  -8.53891920   5.770097212  0.0032621220
#> [284,]   5.61019238  -0.32111998   2.54119717  13.673512365 -0.0213185854
#> [285,]  -8.65962995   1.74013729  12.11956829   0.314318017 -0.0272414592
#> [286,]  -8.69796978   1.87604888  -5.41898304 -10.791681295  0.0119070909
#> [287,]  11.75040809  -3.41412050  -8.28222804  -2.583871050 -0.0394971843
#> [288,] -14.11660016   2.00606479   4.65803639   0.001048346 -0.0157238287
#> [289,]   0.58164545   3.59244477  13.86666836  -4.412656544 -0.0073775683
#> [290,]  10.48187131  -9.80495268   3.81198859   2.112349349  0.0459219912
#> [291,]   8.73552037  -4.46594673  10.87150812  -3.249048330  0.0226348604
#> [292,]  -0.06467827   5.97498088 -13.30686502  -3.496106904 -0.0118887408
#> [293,]   8.50594831  -8.03510811  -3.93720214  -8.519643199  0.0202879455
#> [294,]   3.45071414  14.19804022   2.00841827  -2.733949869 -0.0085334935
#> [295,] -11.37452793  -1.65407791   5.10711377  -8.173220275 -0.0189687931
#> [296,]  -2.08035186 -11.91757623  -7.01367507   5.426958159 -0.0439416778
#> [297,]   1.94144932  -2.50987958  -0.53347239 -14.650825447  0.0255372126
#> [298,]  -6.76106115 -13.14917617   2.44108545   0.654461568 -0.0078683428
#> [299,]  -6.32939579   2.04184354 -12.89812655  -3.795781239 -0.0301495977
#> [300,]  -3.27603534  -4.99751990   0.52824770 -13.748212308 -0.0005122316
#> [301,]   1.63363062  14.55466164  -1.76098551  -2.718824320 -0.0325653430
#> [302,] -13.56647550  -3.65093987   2.34546547   4.703208714  0.0405002255
#> [303,]   9.49019952 -10.05411794   5.14473412   2.717082309 -0.0318704513
#> [304,]  -5.79941640  -6.53046321   8.36833941  -8.870778728  0.0301355670
#> [305,] -10.49512305  -9.84014146   3.91641618  -1.638808303 -0.0292440554
#> [306,]  -9.05394737  -7.63123843  -1.08574365  -9.143926826 -0.0236358887
#> [307,]   0.14693618  11.28284006   8.69890978  -4.690937921  0.0342380172
#> [308,]   4.69774169  12.74487391   4.42725471  -4.571523553 -0.0451731122
#> [309,]  -5.92186832  -6.66717218   3.49084645 -11.545314281  0.0208861045
#> [310,]   8.39588147  -6.10388177 -10.65172601  -1.947443105  0.0479550461
#> [311,] -12.49008305  -5.67102799  -2.00011197  -5.730341958 -0.0238272365
#> [312,]  -2.09915587   2.57744156   9.69619858  10.951441586 -0.0468909115
#> [313,]   0.16086563  13.66135183   3.84972391  -4.849867441  0.0235861135
#> [314,] -11.31554081  -2.27500160  -9.56618138   0.520651318  0.0177708502
#> [315,]   8.97001855   0.73142671   1.08407702 -11.951090295 -0.0121002652
#> [316,]   5.16049228  -7.80559262   2.01791184  11.548596239 -0.0058167164
#> [317,]  -2.79324210   4.46099528 -11.72669015   7.731885786 -0.0087907006
#> [318,]   2.29453595   6.41171576  -9.00793301   9.873304858  0.0155490278
#> [319,]  13.30183609  -1.32935124  -0.91369563  -6.742339528  0.0419516760
#> [320,]   0.30765368   3.54492718  14.25027051   3.044442636 -0.0230809814
#> [321,] -13.88822243  -1.83799783  -5.09881093   1.655647581  0.0226636962
#> [322,] -11.85167708   3.98735552  -8.27824801  -0.330690640 -0.0375408654
#> [323,] -12.13543854   3.52735454   2.14211370   7.791036539 -0.0211497325
#> [324,]   2.86668515  -5.40202626 -12.02746234  -6.552890827  0.0144789500
#> [325,]   4.28566304  -8.84587092 -10.79147943   3.453640363  0.0404549389
#> [326,] -12.92437421  -4.05299193  -0.43442286  -6.430014343 -0.0257775217
#> [327,]   2.31296110  -0.35707218 -14.54103432  -2.842715463  0.0257210549
#> [328,]   0.57643915  -5.91149464 -10.78473226   8.568051065  0.0336559628
#> [329,]  -4.62676272  -2.85774439   7.07858888  12.054872174  0.0225065942
#> [330,]  -5.27180725   2.85227580 -13.07238733  -4.264417951  0.0003932666
#> [331,]  -6.56981270   2.02921138 -13.20102003   1.858206738  0.0049031496
#> [332,]   7.07320017   4.96597250   6.61876875  10.319925230 -0.0040581789
#> [333,]  -0.32958207 -11.70187589   9.30293048   1.188680326 -0.0157605133
#> [334,]  -9.89728678   4.57232985   6.53457591  -7.964724216 -0.0346560197
#> [335,]  -4.73480521  -1.65543478 -10.55941313  -9.398933431 -0.0299988588
#> [336,]  -7.08492345  -2.27878512  12.94111157   1.462405391  0.0375437240
#> [337,]  -8.09911355  -8.46587097   8.96870927   2.701044702 -0.0456036761
#> [338,] -13.23773448  -1.25000938   3.50855295  -5.990819521 -0.0035671920
#> [339,]   2.81253075  -9.47445271  -9.58578522  -5.952910064 -0.0129344609
#> [340,] -10.29699953   0.51618516   9.40374700  -5.502262807  0.0052698983
#> [341,]  -1.91559121  -3.24645133   2.43627144 -14.312779100 -0.0295750235
#> [342,]   5.74908934   7.11019759   9.26462928  -7.453838358  0.0169425233
#> [343,]   6.17175639   3.67766004  12.86153388  -2.822266097 -0.0081031180
#> [344,]  -1.37538089   1.59938077  -8.67417141 -12.054420718 -0.0005213174
#> [345,]  11.56860444   5.32138070   5.07011858  -6.094603864  0.0186029508
#> [346,]   9.62029937  -1.99990355   5.88440590   9.686278590 -0.0193202520
#> [347,] -12.03348116  -6.36439950  -6.01620679  -1.869493533  0.0117970928
#> [348,]   6.92979039   5.80125640  -0.39041679 -11.965408648 -0.0384177311
#> [349,]   9.52720850 -11.48540714  -1.11605105  -1.035447253 -0.0433963194
#> [350,]  14.04962403  -3.48395297   0.29320561   3.922265505  0.0390575929
#> [351,]  -3.19797955   6.47537456   1.46495438  13.065081697 -0.0171533582
#> [352,]   5.55417657  13.04834265  -3.61937371   3.285119573  0.0465612458
#> [353,]   2.50844446  11.83672005   8.07345179  -3.663214559  0.0346247125
#> [354,]   5.23787864  -7.37418086   4.29606847  11.168253206 -0.0416392289
#> [355,] -11.00119512   5.82265397   4.30604438  -7.178327687  0.0049372797
#> [356,] -11.01623292   9.40463000   0.53686422  -3.861000327 -0.0185919262
#> [357,]  -0.49302055   5.13935131 -14.08253990   0.161458326  0.0045117962
#> [358,] -11.76395056  -7.82318214  -0.32079977   5.030345518 -0.0461717237
#> [359,]  11.21629032   8.23156884   5.56974555  -0.643459649  0.0122553712
#> [360,]   3.08231997  12.49576612   2.68441103  -7.221431307  0.0162304267
#> [361,] -13.95783254   3.64346095  -4.06654449  -0.606068419  0.0264930890
#> [362,]   5.24456587   4.58442144 -13.28354591   0.158167333 -0.0172058788
#> [363,] -13.03096881  -5.28067575  -0.18418086  -5.222489152 -0.0004430643
#> [364,]   4.43834109  -8.63684525 -11.24621176   2.056393337  0.0347882810
#> [365,]  -4.61832737  -3.54129247 -11.58776986  -7.540151828  0.0333774019
#> [366,]  14.41336310  -3.83747387   0.93167395   1.288697796 -0.0260965556
#> [367,] -13.94527725   4.29043189  -1.02351340   3.327740518  0.0074521042
#> [368,]   6.06616101 -12.42189168  -3.11367563  -4.919687164  0.0394366712
#> [369,]   5.57801174  -1.77480550  13.69958044   1.748526832 -0.0301168667
#> [370,]  -5.87668306 -11.81216666  -6.66019901   2.564968681  0.0419587971
#> [371,]  11.22103710   5.33701448  -5.81560491  -6.064927231  0.0178800270
#> [372,]  -2.74232597  -8.64812906  -4.96758737 -10.863359871  0.0242777936
#> [373,]  -3.32146085   1.10779010   2.95276318  14.283623087  0.0480781602
#> [374,] -10.51763654  -4.99060921   9.36908160   1.301326801  0.0044925978
#> [375,]  -4.20119846   1.53774764 -14.26830141  -1.183570229 -0.0268020860
#> [376,]   8.87537246  11.15546650   4.44789491  -1.414058618 -0.0154708300
#> [377,]  10.59343885  -9.94355079   3.70420006   0.428664127 -0.0422128927
#> [378,]  -1.18071223 -10.01062899   0.18187643  11.106761311  0.0053710290
#> [379,]   4.71546823  -5.85347529 -12.77078834   2.325543279  0.0415850116
#> [380,]  -1.99381240  -6.86159343  -6.00433361  11.742709468 -0.0147140309
#> [381,] -10.46493581   5.08650360   3.73468266  -8.698548449  0.0090381307
#> [382,]  -5.24565558   7.52234238   4.66490934 -10.914947714  0.0208063293
#> [383,]   7.65114048  12.60191986   2.07397799   1.830377154  0.0371321189
#> [384,]  14.84551516   1.75619988   0.53731657   1.112534238  0.0022840990
#> [385,] -12.20526975  -3.55557599   7.44594466   2.819073921  0.0159999489
#> [386,]  10.34817093   5.59508089  -8.46474676  -3.867620790 -0.0035788741
#> [387,] -14.34391375  -1.40894292  -3.79430545  -1.694185424 -0.0112141618
#> [388,] -10.93812638   1.45673863  -6.66162426  -7.671901078  0.0371199941
#> [389,]   6.34363216  -0.07333389 -11.21271467  -7.682967061 -0.0062326387
#> [390,]  -4.18779186  -5.68394361  -0.05132903  13.234521136 -0.0117852898
#>                 [,6]          [,7]          [,8]
#>   [1,] -0.0316545064 -0.0302327739 -0.0149086856
#>   [2,]  0.0224097124  0.0383540310 -0.0114872495
#>   [3,] -0.0058730685 -0.0337559121 -0.0113976363
#>   [4,] -0.0384471205 -0.0158351770 -0.0371611770
#>   [5,] -0.0066567696  0.0078958530 -0.0390014886
#>   [6,]  0.0256157019  0.0116285037 -0.0094174077
#>   [7,] -0.0338826125 -0.0217218023  0.0141096750
#>   [8,] -0.0412903447  0.0417551709 -0.0489971149
#>   [9,] -0.0403237645  0.0431604248  0.0476158777
#>  [10,]  0.0432364848 -0.0258266344 -0.0070654613
#>  [11,]  0.0427404279  0.0116242698 -0.0105233203
#>  [12,] -0.0391346660  0.0385444032  0.0290634308
#>  [13,] -0.0330254006  0.0430160034  0.0497325133
#>  [14,] -0.0030522383  0.0499364473 -0.0078397391
#>  [15,]  0.0263212564  0.0053977718 -0.0169103026
#>  [16,] -0.0303858228 -0.0228205521  0.0335864885
#>  [17,]  0.0071982476 -0.0246368508 -0.0176756780
#>  [18,] -0.0012970382 -0.0383061840 -0.0208382873
#>  [19,] -0.0453215297 -0.0469407187 -0.0139480855
#>  [20,]  0.0155376290 -0.0024277724  0.0236997608
#>  [21,] -0.0387788581  0.0113538870  0.0046305377
#>  [22,] -0.0166450149  0.0393251977  0.0259026846
#>  [23,]  0.0098181709 -0.0385824237 -0.0481396592
#>  [24,] -0.0075139229  0.0148444821  0.0236935982
#>  [25,]  0.0331760231 -0.0098327809  0.0104190209
#>  [26,] -0.0198006235  0.0104611821 -0.0394405735
#>  [27,]  0.0396713876  0.0421522771 -0.0380121459
#>  [28,]  0.0223306950  0.0062620084  0.0419023976
#>  [29,] -0.0475047150 -0.0103966010  0.0150044224
#>  [30,]  0.0336251184 -0.0005781362  0.0272523994
#>  [31,]  0.0084340744 -0.0240687672 -0.0047822804
#>  [32,]  0.0388732226  0.0310181517  0.0287551134
#>  [33,]  0.0096390683 -0.0326596267 -0.0118214930
#>  [34,] -0.0354503244 -0.0069660363  0.0028810099
#>  [35,]  0.0090557174  0.0052602923  0.0286048706
#>  [36,]  0.0363325094 -0.0284439760 -0.0311320245
#>  [37,]  0.0135676930 -0.0417668967 -0.0122259753
#>  [38,]  0.0378405037  0.0246119145 -0.0318777867
#>  [39,] -0.0376625641  0.0190879883 -0.0011737513
#>  [40,] -0.0498988348  0.0218347915 -0.0369153217
#>  [41,]  0.0105055742  0.0433812128  0.0419871304
#>  [42,] -0.0119267519 -0.0229025315 -0.0034047455
#>  [43,]  0.0281139098  0.0394714022  0.0425362405
#>  [44,]  0.0171310939 -0.0438539736 -0.0319545166
#>  [45,] -0.0122536972 -0.0421961800  0.0414120165
#>  [46,]  0.0224939816  0.0016976571 -0.0152989134
#>  [47,]  0.0473302682  0.0451252137 -0.0395568010
#>  [48,] -0.0481403759  0.0097444409 -0.0289941242
#>  [49,]  0.0174051022 -0.0410480285  0.0473556313
#>  [50,]  0.0003205740  0.0445589403 -0.0104748802
#>  [51,]  0.0363566860  0.0080856413  0.0294528896
#>  [52,] -0.0003904358 -0.0472589904 -0.0326594592
#>  [53,]  0.0482579468  0.0418125868 -0.0491418596
#>  [54,]  0.0030323254  0.0007240079  0.0130012729
#>  [55,]  0.0153248419 -0.0377865727 -0.0447945905
#>  [56,]  0.0283819866  0.0459176495 -0.0183883592
#>  [57,]  0.0429583367 -0.0195566343 -0.0085528377
#>  [58,] -0.0448328620 -0.0309077916  0.0092152898
#>  [59,]  0.0183201506 -0.0047750326  0.0054035998
#>  [60,]  0.0454515950  0.0244993923 -0.0249077035
#>  [61,] -0.0422209830  0.0301493833  0.0335447852
#>  [62,] -0.0139031098  0.0171286077  0.0225848536
#>  [63,] -0.0141489498  0.0256814180 -0.0363274698
#>  [64,]  0.0036226860 -0.0472189985 -0.0309964480
#>  [65,]  0.0035760734  0.0478545144 -0.0264314575
#>  [66,] -0.0181178080 -0.0014935386 -0.0184312303
#>  [67,] -0.0078491741  0.0140462848  0.0154507402
#>  [68,] -0.0274168907 -0.0095167778 -0.0013594597
#>  [69,]  0.0343960205  0.0259115696  0.0488755484
#>  [70,]  0.0430605633  0.0234930093  0.0225479410
#>  [71,]  0.0460080283 -0.0246242637  0.0119685331
#>  [72,]  0.0155184703  0.0078423203  0.0381696011
#>  [73,] -0.0219805820  0.0254326866  0.0059889935
#>  [74,] -0.0460303989 -0.0137431319 -0.0074502126
#>  [75,] -0.0349399659 -0.0220145521  0.0128912321
#>  [76,] -0.0180103326  0.0296562372 -0.0316742466
#>  [77,]  0.0269992773 -0.0286058591 -0.0207969493
#>  [78,]  0.0053544548  0.0463396062  0.0116174912
#>  [79,]  0.0347711010  0.0206147999 -0.0410736010
#>  [80,] -0.0014136756 -0.0230128239  0.0358389253
#>  [81,]  0.0298901162 -0.0150694005 -0.0108867611
#>  [82,]  0.0276710545  0.0229197843  0.0264094083
#>  [83,] -0.0224227906  0.0215438268  0.0494566324
#>  [84,]  0.0135186209  0.0279299944 -0.0340290352
#>  [85,] -0.0186513433  0.0064795838 -0.0335202447
#>  [86,]  0.0079166902 -0.0205054846 -0.0065991141
#>  [87,]  0.0187565769 -0.0265401067  0.0143277417
#>  [88,] -0.0483486468  0.0030105946 -0.0191993365
#>  [89,]  0.0474801549  0.0222560055 -0.0347360489
#>  [90,] -0.0178648174  0.0409699755  0.0414403909
#>  [91,]  0.0324134449 -0.0403272454 -0.0110333550
#>  [92,] -0.0365930069 -0.0165764160 -0.0197400030
#>  [93,] -0.0348708935 -0.0160275303  0.0017181385
#>  [94,]  0.0355114629 -0.0211344184 -0.0165427397
#>  [95,]  0.0271630161  0.0329967468  0.0232373388
#>  [96,] -0.0476140357 -0.0382191589  0.0106966854
#>  [97,]  0.0071402431 -0.0373138068 -0.0430178287
#>  [98,]  0.0408832290 -0.0153693186 -0.0271816092
#>  [99,] -0.0209972294 -0.0205062877  0.0096133667
#> [100,]  0.0066626939  0.0194401738 -0.0383510709
#> [101,]  0.0068892051 -0.0069639835 -0.0347243825
#> [102,]  0.0164028188  0.0184785311 -0.0466620282
#> [103,]  0.0235047219  0.0305473779  0.0008147368
#> [104,] -0.0293101677  0.0411514597 -0.0178322429
#> [105,]  0.0271239813 -0.0277842713 -0.0013549264
#> [106,]  0.0320154624  0.0280903398 -0.0139003184
#> [107,]  0.0260786190  0.0351871758 -0.0281517614
#> [108,]  0.0181429178 -0.0178148480 -0.0043045224
#> [109,] -0.0045807536  0.0387644498  0.0126707347
#> [110,] -0.0386862828  0.0100997081  0.0206500478
#> [111,] -0.0170380435  0.0034424982 -0.0239284310
#> [112,]  0.0234720815 -0.0423739981  0.0212146054
#> [113,] -0.0255245471  0.0360534161 -0.0236133712
#> [114,]  0.0385013859 -0.0406324059 -0.0449493300
#> [115,]  0.0107553932 -0.0352223045  0.0062411134
#> [116,]  0.0471267601 -0.0255592588  0.0027902156
#> [117,]  0.0085367720  0.0157639137  0.0197057987
#> [118,]  0.0445342469  0.0210674550  0.0021704347
#> [119,]  0.0230817826 -0.0435816205 -0.0310464339
#> [120,] -0.0364498982 -0.0233161476  0.0317625453
#> [121,]  0.0237685415 -0.0186572021  0.0074610829
#> [122,]  0.0114170825  0.0230429891  0.0463640575
#> [123,] -0.0251514799  0.0078257111 -0.0346063169
#> [124,] -0.0028231731 -0.0144340040 -0.0192886919
#> [125,] -0.0320612674  0.0097798355  0.0192359051
#> [126,]  0.0288956664  0.0290009541 -0.0122028151
#> [127,]  0.0488141491  0.0113489529  0.0272383918
#> [128,]  0.0018438478 -0.0306449142 -0.0028257519
#> [129,] -0.0328250262  0.0103233630  0.0055435486
#> [130,] -0.0163606977  0.0344990999 -0.0223857072
#> [131,]  0.0286887900 -0.0404158640 -0.0445842380
#> [132,]  0.0103585480 -0.0158153146  0.0083574404
#> [133,]  0.0071313513 -0.0381807014 -0.0328582674
#> [134,]  0.0404556623  0.0415786709  0.0383206135
#> [135,] -0.0076960147  0.0334932922 -0.0331901205
#> [136,] -0.0474338481 -0.0153042593  0.0036256341
#> [137,]  0.0279450347 -0.0032833889 -0.0321667577
#> [138,]  0.0234312846  0.0045865014 -0.0116812765
#> [139,] -0.0485053924 -0.0349629747  0.0220442594
#> [140,]  0.0399691103 -0.0425756249 -0.0159343561
#> [141,]  0.0289001654 -0.0375505047 -0.0040150815
#> [142,]  0.0194087601  0.0332036016 -0.0191652512
#> [143,] -0.0452673686  0.0148408001 -0.0245285378
#> [144,] -0.0331066190  0.0461352635  0.0323220486
#> [145,]  0.0083416061 -0.0340987876  0.0364552748
#> [146,]  0.0311471867 -0.0122685367  0.0379448323
#> [147,]  0.0441733215 -0.0372341619 -0.0394697050
#> [148,]  0.0119291255 -0.0102585910 -0.0301833126
#> [149,]  0.0237154981  0.0049165524 -0.0093545400
#> [150,]  0.0299821134 -0.0036641160  0.0450636965
#> [151,]  0.0074312360 -0.0406942968  0.0071262427
#> [152,] -0.0230554801 -0.0219196065  0.0329413436
#> [153,] -0.0098318679  0.0339932150 -0.0381858915
#> [154,]  0.0245330961  0.0129231562  0.0393598691
#> [155,]  0.0390898018  0.0114566615  0.0403923963
#> [156,]  0.0129817799 -0.0342428840  0.0193439980
#> [157,]  0.0407671100 -0.0036162026  0.0114362437
#> [158,] -0.0079379001  0.0416055099 -0.0048219892
#> [159,] -0.0463330709 -0.0359464856  0.0182031013
#> [160,]  0.0278409448  0.0054651178  0.0232140526
#> [161,] -0.0370877966  0.0451698213 -0.0018199977
#> [162,] -0.0169036374  0.0353342006 -0.0062258195
#> [163,] -0.0438895473 -0.0292908858  0.0114650424
#> [164,]  0.0246802411 -0.0290128477 -0.0007927845
#> [165,] -0.0475824768 -0.0097823656  0.0069374965
#> [166,] -0.0163820421 -0.0144400492  0.0102521857
#> [167,]  0.0143898793  0.0270458351  0.0325230071
#> [168,] -0.0030039197 -0.0434133501  0.0288148529
#> [169,] -0.0081166394  0.0146367602 -0.0473322053
#> [170,]  0.0396844822  0.0469097281 -0.0089604579
#> [171,]  0.0265566208  0.0268055704  0.0005285702
#> [172,]  0.0204097725 -0.0122563329 -0.0115498524
#> [173,] -0.0207407020  0.0379149776 -0.0331652863
#> [174,] -0.0136008093  0.0094532959  0.0411520274
#> [175,]  0.0404925953  0.0065051951  0.0046276898
#> [176,]  0.0140859347  0.0024463860 -0.0380669188
#> [177,] -0.0415320981  0.0193318980 -0.0070798049
#> [178,] -0.0046214467  0.0344214538  0.0029644613
#> [179,]  0.0461561202 -0.0428117999  0.0231182813
#> [180,] -0.0237664109 -0.0199536927  0.0495428747
#> [181,] -0.0341202633 -0.0475057746 -0.0228720848
#> [182,]  0.0107587299  0.0276388028 -0.0383616984
#> [183,] -0.0031358037  0.0357076274  0.0390257563
#> [184,]  0.0244422555 -0.0231309286 -0.0289204262
#> [185,]  0.0419026258  0.0415386613 -0.0340317618
#> [186,]  0.0329581960 -0.0348869239  0.0259028425
#> [187,] -0.0175840472  0.0253965649  0.0453359775
#> [188,] -0.0232373429 -0.0479325875  0.0433819879
#> [189,] -0.0347976334 -0.0001826605  0.0386302504
#> [190,] -0.0263744748 -0.0109943578 -0.0490263169
#> [191,]  0.0026231199 -0.0307794619 -0.0261526270
#> [192,]  0.0498535967 -0.0404083670  0.0423356635
#> [193,]  0.0010101914 -0.0127275979 -0.0450677096
#> [194,] -0.0279497629 -0.0359342281 -0.0026827348
#> [195,]  0.0454600266 -0.0487924160 -0.0335770231
#> [196,] -0.0021767576  0.0319013681 -0.0187298667
#> [197,] -0.0321382353  0.0306152622 -0.0431739729
#> [198,]  0.0413611765 -0.0040673974  0.0493360000
#> [199,] -0.0252109422 -0.0157591008  0.0304371737
#> [200,]  0.0047064543 -0.0274727595  0.0091813516
#> [201,] -0.0379222085 -0.0017752889 -0.0188542147
#> [202,] -0.0231339094 -0.0045398859  0.0048094413
#> [203,]  0.0254172823  0.0099924972  0.0134799651
#> [204,]  0.0385850925 -0.0158017890  0.0235288130
#> [205,]  0.0350221756  0.0313287776 -0.0417520077
#> [206,] -0.0248769978 -0.0383533302  0.0156208195
#> [207,] -0.0168705793 -0.0440198813  0.0359137836
#> [208,] -0.0247189383 -0.0117220946 -0.0413343727
#> [209,] -0.0364377812 -0.0081496359 -0.0468243916
#> [210,]  0.0125814772  0.0130012767 -0.0329038674
#> [211,] -0.0418471181  0.0424385834  0.0259437825
#> [212,] -0.0428638628 -0.0005641295  0.0270026472
#> [213,]  0.0272539393 -0.0012561879  0.0480811094
#> [214,]  0.0161944947 -0.0173678077  0.0391703640
#> [215,]  0.0217740091 -0.0453709402  0.0179304949
#> [216,]  0.0492675133  0.0311284318  0.0180268181
#> [217,]  0.0048696535  0.0049550921  0.0020301469
#> [218,]  0.0427472331 -0.0080957928 -0.0476200600
#> [219,]  0.0456530863  0.0266173788  0.0092027888
#> [220,] -0.0470452822 -0.0215818671 -0.0431037921
#> [221,] -0.0142377665 -0.0019121361  0.0066661950
#> [222,]  0.0205767895 -0.0049170736 -0.0349623552
#> [223,]  0.0295530638 -0.0075277892  0.0294469152
#> [224,]  0.0164014023  0.0293110288  0.0396066113
#> [225,]  0.0203098571 -0.0334458059  0.0303875550
#> [226,] -0.0080661156 -0.0072258456 -0.0154733755
#> [227,]  0.0070536014 -0.0076404400 -0.0468657455
#> [228,] -0.0250502022 -0.0189800955  0.0058628176
#> [229,] -0.0248653408 -0.0088371196  0.0327329073
#> [230,]  0.0100916413 -0.0112050680 -0.0174631941
#> [231,] -0.0407043314  0.0371579429 -0.0070323732
#> [232,]  0.0067056266  0.0133850800  0.0145969413
#> [233,] -0.0139697193 -0.0121491541  0.0155393832
#> [234,] -0.0205438121  0.0268756625  0.0111729950
#> [235,]  0.0418444135  0.0492738383 -0.0039527362
#> [236,] -0.0492599612 -0.0410459159  0.0126904557
#> [237,] -0.0302602059 -0.0258638427  0.0269182754
#> [238,] -0.0328657944  0.0245568391  0.0001532637
#> [239,]  0.0287647745 -0.0308871629  0.0241908319
#> [240,]  0.0208798466  0.0058567726  0.0399177215
#> [241,] -0.0464145993  0.0030661104  0.0227982456
#> [242,] -0.0287747040 -0.0431488342 -0.0260093724
#> [243,]  0.0047427858 -0.0352013747 -0.0151672343
#> [244,] -0.0409220008 -0.0324264881 -0.0284037789
#> [245,] -0.0042775594  0.0029234624 -0.0233733977
#> [246,] -0.0262099978 -0.0153300464  0.0049687954
#> [247,] -0.0107163528  0.0138539880 -0.0476104878
#> [248,] -0.0286980778 -0.0396912986 -0.0395139797
#> [249,] -0.0235257468 -0.0123364045  0.0440950876
#> [250,] -0.0185531181  0.0215964957  0.0116458436
#> [251,] -0.0294697858 -0.0486775603 -0.0128810877
#> [252,]  0.0468884161  0.0196659999 -0.0187333947
#> [253,]  0.0393205576 -0.0434134995 -0.0305176767
#> [254,]  0.0118330416  0.0009032456 -0.0090449434
#> [255,]  0.0029590431  0.0223315342 -0.0343601053
#> [256,] -0.0323901244  0.0229795506  0.0179170560
#> [257,] -0.0379452252  0.0162010657 -0.0427495371
#> [258,] -0.0444275221 -0.0021083007 -0.0095840676
#> [259,] -0.0355318096 -0.0127213856  0.0495702430
#> [260,] -0.0426709957  0.0038775178  0.0367094113
#> [261,]  0.0087971702  0.0392241852  0.0495957595
#> [262,]  0.0311517806 -0.0338103729 -0.0134879293
#> [263,]  0.0011031435 -0.0195843794  0.0356879605
#> [264,] -0.0193374486  0.0458790026 -0.0307381348
#> [265,]  0.0191149206  0.0332694022 -0.0357666128
#> [266,]  0.0142531339  0.0398796815  0.0338485814
#> [267,]  0.0325184840 -0.0442972293  0.0441368766
#> [268,]  0.0433118920  0.0164187283 -0.0426903406
#> [269,] -0.0091552068  0.0089219590 -0.0390983260
#> [270,]  0.0116142937 -0.0481089856 -0.0458088576
#> [271,] -0.0128278553  0.0306078838  0.0440005152
#> [272,] -0.0211744872 -0.0203989227 -0.0062061598
#> [273,] -0.0241856493  0.0425984581 -0.0167176527
#> [274,] -0.0189844236  0.0059796841  0.0447055486
#> [275,]  0.0134978752 -0.0395742948  0.0200110932
#> [276,] -0.0434817782 -0.0443061970  0.0027733525
#> [277,]  0.0403850951 -0.0050130781 -0.0053124114
#> [278,] -0.0245464425 -0.0377495332 -0.0329814512
#> [279,] -0.0182316082 -0.0417571197 -0.0103377071
#> [280,]  0.0249041270 -0.0418732466 -0.0164410450
#> [281,]  0.0494712179  0.0368594302 -0.0441301983
#> [282,] -0.0177415921  0.0085194062  0.0494535308
#> [283,] -0.0256914099 -0.0229601533 -0.0204477844
#> [284,] -0.0058509766  0.0479159137  0.0030238139
#> [285,] -0.0201534934  0.0221352559 -0.0103932291
#> [286,] -0.0452117470  0.0448330908  0.0494023299
#> [287,]  0.0279468767 -0.0290720587  0.0253224324
#> [288,]  0.0279926708 -0.0487698218 -0.0105053961
#> [289,] -0.0082922763  0.0490405730 -0.0272465348
#> [290,] -0.0200072185 -0.0340992453 -0.0298917885
#> [291,] -0.0387308071 -0.0225772991 -0.0055182279
#> [292,]  0.0161308155  0.0291443344 -0.0050255993
#> [293,] -0.0109509172  0.0009079289 -0.0092257385
#> [294,] -0.0176600055  0.0412632473  0.0039713532
#> [295,] -0.0397794622  0.0006198183  0.0243791298
#> [296,]  0.0166495348  0.0295031329 -0.0238809919
#> [297,]  0.0420082424  0.0404827507  0.0317750475
#> [298,] -0.0105494620  0.0214823572 -0.0499297275
#> [299,] -0.0322157142  0.0369237602 -0.0191354012
#> [300,] -0.0113894432  0.0237149512 -0.0227588244
#> [301,]  0.0337994490 -0.0227769018  0.0329037501
#> [302,]  0.0491327587 -0.0351648150  0.0157380030
#> [303,] -0.0298733779  0.0009740921  0.0151585222
#> [304,]  0.0353214066 -0.0368753227  0.0048422640
#> [305,]  0.0383287284 -0.0190020639 -0.0116933167
#> [306,] -0.0267304244  0.0059064964 -0.0334717999
#> [307,] -0.0355022175 -0.0262138775 -0.0304232606
#> [308,] -0.0339183060 -0.0469231449  0.0156231274
#> [309,] -0.0026046358  0.0045200395  0.0260888319
#> [310,]  0.0211841339  0.0140521320 -0.0454493327
#> [311,] -0.0391550695  0.0423046128  0.0329700295
#> [312,]  0.0007669843  0.0091252178  0.0265394378
#> [313,] -0.0350284525  0.0092301981 -0.0413291015
#> [314,]  0.0144637923  0.0036800496  0.0168000175
#> [315,]  0.0435446614 -0.0360688586  0.0287815872
#> [316,]  0.0278114101 -0.0057308759  0.0391854902
#> [317,] -0.0260843266  0.0101547586  0.0385324092
#> [318,]  0.0243631189  0.0410961330 -0.0231897200
#> [319,]  0.0285567578 -0.0290510327 -0.0362200917
#> [320,] -0.0489091729 -0.0490528451  0.0415998235
#> [321,] -0.0026081893  0.0494639928 -0.0379253580
#> [322,] -0.0067090560 -0.0183225093  0.0449130281
#> [323,]  0.0121331936 -0.0327123639 -0.0389443323
#> [324,] -0.0054102787  0.0087276313  0.0383248250
#> [325,] -0.0106938969  0.0417561858  0.0481705754
#> [326,]  0.0076204948  0.0461031248 -0.0301106699
#> [327,] -0.0009310514  0.0124056027  0.0435586116
#> [328,] -0.0214030946  0.0335735665  0.0172298039
#> [329,] -0.0279471935  0.0388051862  0.0127947508
#> [330,] -0.0239026508  0.0084567099 -0.0498185815
#> [331,]  0.0388526369  0.0331592450 -0.0226662355
#> [332,]  0.0399831240  0.0204065072 -0.0154390681
#> [333,]  0.0158668309  0.0264029578 -0.0242189443
#> [334,] -0.0212632748  0.0086961580  0.0363620916
#> [335,] -0.0134260325 -0.0187863916  0.0146102099
#> [336,]  0.0306095738 -0.0066169483  0.0197341886
#> [337,] -0.0119397532  0.0093893756 -0.0011979436
#> [338,] -0.0203225523  0.0347544599 -0.0280715828
#> [339,] -0.0295053924 -0.0051550907 -0.0480619177
#> [340,]  0.0050001372 -0.0332466002 -0.0319660774
#> [341,] -0.0158011169  0.0299844993  0.0235830605
#> [342,] -0.0342345787 -0.0172785020 -0.0196515303
#> [343,] -0.0153676139 -0.0359686786  0.0254246377
#> [344,]  0.0098876848 -0.0242655254 -0.0477262721
#> [345,] -0.0413664310  0.0042715583 -0.0387976553
#> [346,] -0.0366867488  0.0490989746 -0.0315941399
#> [347,] -0.0156652802  0.0084996708  0.0407678162
#> [348,]  0.0092806607 -0.0310152890 -0.0189744465
#> [349,] -0.0426080174 -0.0130581846 -0.0227609293
#> [350,] -0.0457930321 -0.0200493246  0.0336617382
#> [351,]  0.0300707191 -0.0480809159 -0.0232406306
#> [352,]  0.0046635683 -0.0155636368  0.0086032801
#> [353,] -0.0205307506  0.0469505150  0.0485104228
#> [354,]  0.0359196191  0.0326154164 -0.0100988997
#> [355,] -0.0171758374 -0.0120236619  0.0054004765
#> [356,] -0.0073782693  0.0461016713 -0.0408739449
#> [357,]  0.0288933859 -0.0119648514  0.0034214900
#> [358,] -0.0445737168  0.0400496316 -0.0164645739
#> [359,] -0.0099332970  0.0084598232  0.0275908021
#> [360,]  0.0236614419  0.0321507854 -0.0025889314
#> [361,]  0.0092817232  0.0075937037  0.0125224166
#> [362,]  0.0132293402  0.0081267946  0.0035484371
#> [363,] -0.0481698418 -0.0373653023  0.0226039279
#> [364,]  0.0220092569 -0.0181768802 -0.0277800274
#> [365,] -0.0407650940  0.0401670145  0.0452536525
#> [366,]  0.0082799604 -0.0152914612  0.0063735938
#> [367,] -0.0414082336  0.0432955703 -0.0115866584
#> [368,] -0.0270289054 -0.0434515311 -0.0457020956
#> [369,] -0.0040653758  0.0080979745 -0.0158088442
#> [370,]  0.0342927033 -0.0417423550  0.0344704322
#> [371,] -0.0084706679  0.0171512978  0.0309476265
#> [372,]  0.0016720979  0.0181265652  0.0354059956
#> [373,] -0.0400159159  0.0056588952 -0.0427748651
#> [374,]  0.0136091234 -0.0050716357 -0.0441804412
#> [375,] -0.0099424049  0.0492808096  0.0419127827
#> [376,] -0.0189795232 -0.0342421053 -0.0078485871
#> [377,] -0.0471670046 -0.0322532265 -0.0399139748
#> [378,] -0.0355190796 -0.0319866575 -0.0342574345
#> [379,]  0.0479021518  0.0281491781 -0.0439685211
#> [380,]  0.0089147941  0.0273425550  0.0225502720
#> [381,]  0.0401660026  0.0361639470  0.0364167112
#> [382,] -0.0165191547  0.0159443068  0.0093808613
#> [383,] -0.0404188616  0.0050950608  0.0388089723
#> [384,]  0.0110777525  0.0283140062 -0.0003703217
#> [385,]  0.0127219753  0.0153961153  0.0374102678
#> [386,] -0.0331947756  0.0476061869  0.0218911866
#> [387,] -0.0321540025  0.0156704694 -0.0250614667
#> [388,]  0.0184107002 -0.0494962411  0.0317618236
#> [389,] -0.0328720801 -0.0387652563  0.0400820676
#> [390,]  0.0007916308 -0.0333472060 -0.0415978581