Generate Coordinates for a Sphere

This function generates the coordinates for a sphere in three-dimensional space.

Usage

sphere(radius, resolution, num_noise_dims, min_noise, max_noise)

Arguments

radius

The radius of the sphere.

resolution

The number of points used to approximate the surface of the sphere.

num_noise_dims

The number of additional noise dimensions to add to the coordinates.

min_noise

The minimum value for the random noise added to the coordinates.

max_noise

The maximum value for the random noise added to the coordinates.

Value

A matrix containing the Cartesian coordinates of the points on the sphere.

Examples

# Generate coordinates for a sphere with radius 1 and resolution 100
sphere(radius = 1, resolution = 20, num_noise_dims = 3, min_noise = -0.05,
max_noise = 0.05)
#>                 [,1]          [,2]        [,3]          [,4]          [,5]
#>   [1,]  0.000000e+00  0.000000e+00  1.00000000 -1.386247e-02 -0.0186018393
#>   [2,]  0.000000e+00  0.000000e+00  1.00000000  4.760939e-02  0.0452406343
#>   [3,]  0.000000e+00  0.000000e+00  1.00000000  2.233020e-02  0.0350906009
#>   [4,]  0.000000e+00  0.000000e+00  1.00000000  3.175139e-02  0.0452554211
#>   [5,]  0.000000e+00  0.000000e+00  1.00000000 -2.057932e-02  0.0375329020
#>   [6,]  0.000000e+00  0.000000e+00  1.00000000 -5.709907e-03  0.0375595760
#>   [7,]  0.000000e+00  0.000000e+00  1.00000000  1.867164e-02  0.0297837505
#>   [8,]  0.000000e+00  0.000000e+00  1.00000000 -2.629574e-02  0.0252599934
#>   [9,]  0.000000e+00  0.000000e+00  1.00000000 -4.348548e-02 -0.0450706477
#>  [10,]  0.000000e+00  0.000000e+00  1.00000000  2.640402e-02  0.0228430551
#>  [11,]  0.000000e+00  0.000000e+00  1.00000000 -4.437135e-02  0.0348884711
#>  [12,]  0.000000e+00  0.000000e+00  1.00000000 -3.645516e-02 -0.0333798466
#>  [13,]  0.000000e+00  0.000000e+00  1.00000000  3.047476e-02  0.0409425334
#>  [14,]  0.000000e+00  0.000000e+00  1.00000000  1.585633e-03  0.0439088183
#>  [15,]  0.000000e+00  0.000000e+00  1.00000000  4.126461e-02  0.0007083255
#>  [16,]  0.000000e+00  0.000000e+00  1.00000000  9.485889e-03 -0.0168874462
#>  [17,]  0.000000e+00  0.000000e+00  1.00000000  4.111532e-02  0.0453905771
#>  [18,]  0.000000e+00  0.000000e+00  1.00000000 -3.859372e-02  0.0189700126
#>  [19,]  0.000000e+00  0.000000e+00  1.00000000  1.920787e-02 -0.0482794129
#>  [20,]  0.000000e+00  0.000000e+00  1.00000000  5.156871e-03 -0.0034299629
#>  [21,]  1.645946e-01  0.000000e+00  0.98636130  1.622048e-03 -0.0248722789
#>  [22,]  1.556764e-01  5.344378e-02  0.98636130 -4.667698e-02 -0.0189129030
#>  [23,]  1.298883e-01  1.010961e-01  0.98636130 -1.263905e-02 -0.0112062284
#>  [24,]  9.002471e-02  1.377931e-01  0.98636130  3.845850e-02  0.0242076505
#>  [25,]  4.040558e-02  1.595580e-01  0.98636130 -2.476763e-02 -0.0367019178
#>  [26,] -1.359211e-02  1.640324e-01  0.98636130  1.275536e-02 -0.0493981170
#>  [27,] -6.611689e-02  1.507313e-01  0.98636130  1.964568e-02 -0.0401109972
#>  [28,] -1.114769e-01  1.210962e-01  0.98636130  4.482151e-02  0.0103243665
#>  [29,] -1.447566e-01  7.833837e-02  0.98636130  1.221546e-02 -0.0331859975
#>  [30,] -1.623497e-01  2.709138e-02  0.98636130  3.538718e-02 -0.0458399597
#>  [31,] -1.623497e-01 -2.709138e-02  0.98636130 -6.910419e-03  0.0232456831
#>  [32,] -1.447566e-01 -7.833837e-02  0.98636130 -4.307637e-02 -0.0415960403
#>  [33,] -1.114769e-01 -1.210962e-01  0.98636130 -2.451366e-02  0.0446957028
#>  [34,] -6.611689e-02 -1.507313e-01  0.98636130 -4.652521e-03 -0.0472165178
#>  [35,] -1.359211e-02 -1.640324e-01  0.98636130 -5.118949e-03  0.0214626596
#>  [36,]  4.040558e-02 -1.595580e-01  0.98636130  3.035579e-02 -0.0161007500
#>  [37,]  9.002471e-02 -1.377931e-01  0.98636130  2.854251e-02  0.0043655990
#>  [38,]  1.298883e-01 -1.010961e-01  0.98636130 -3.133106e-02  0.0397559307
#>  [39,]  1.556764e-01 -5.344378e-02  0.98636130  3.272070e-02 -0.0322039917
#>  [40,]  1.645946e-01 -4.031405e-17  0.98636130 -2.810299e-02  0.0462590254
#>  [41,]  3.246995e-01  0.000000e+00  0.94581724  4.665706e-02  0.0455590545
#>  [42,]  3.071064e-01  1.054297e-01  0.94581724  4.617676e-02 -0.0213679903
#>  [43,]  2.562335e-01  1.994345e-01  0.94581724  3.413626e-02  0.0002930008
#>  [44,]  1.775938e-01  2.718275e-01  0.94581724  3.187764e-02 -0.0280045754
#>  [45,]  7.970901e-02  3.147638e-01  0.94581724  1.470225e-02 -0.0470309700
#>  [46,] -2.681347e-02  3.235905e-01  0.94581724  7.466727e-03  0.0398404628
#>  [47,] -1.304303e-01  2.973511e-01  0.94581724 -3.266318e-04 -0.0452869376
#>  [48,] -2.199130e-01  2.388892e-01  0.94581724  2.797348e-02  0.0474695073
#>  [49,] -2.855647e-01  1.545399e-01  0.94581724 -2.641405e-02 -0.0275247914
#>  [50,] -3.202710e-01  5.344378e-02  0.94581724 -2.055200e-02 -0.0163977928
#>  [51,] -3.202710e-01 -5.344378e-02  0.94581724 -2.543993e-02 -0.0496179454
#>  [52,] -2.855647e-01 -1.545399e-01  0.94581724  2.240476e-02  0.0137704114
#>  [53,] -2.199130e-01 -2.388892e-01  0.94581724 -3.749161e-02  0.0233116664
#>  [54,] -1.304303e-01 -2.973511e-01  0.94581724  2.316667e-02 -0.0125317842
#>  [55,] -2.681347e-02 -3.235905e-01  0.94581724  2.545465e-02  0.0192050181
#>  [56,]  7.970901e-02 -3.147638e-01  0.94581724  4.875636e-02  0.0395290690
#>  [57,]  1.775938e-01 -2.718275e-01  0.94581724  1.526494e-02  0.0339101754
#>  [58,]  2.562335e-01 -1.994345e-01  0.94581724 -4.745540e-02  0.0377118826
#>  [59,]  3.071064e-01 -1.054297e-01  0.94581724  3.710594e-02 -0.0361131360
#>  [60,]  3.246995e-01 -7.952843e-17  0.94581724 -3.536419e-02  0.0273993538
#>  [61,]  4.759474e-01  0.000000e+00  0.87947375 -1.558507e-02 -0.0369455430
#>  [62,]  4.501593e-01  1.545399e-01  0.87947375 -2.674084e-02 -0.0163885360
#>  [63,]  3.755894e-01  2.923329e-01  0.87947375 -3.544466e-02 -0.0211418650
#>  [64,]  2.603185e-01  3.984472e-01  0.87947375  3.559218e-02  0.0302121091
#>  [65,]  1.168382e-01  4.613835e-01  0.87947375  2.544549e-02 -0.0030777742
#>  [66,] -3.930342e-02  4.743218e-01  0.87947375  3.267513e-02 -0.0009345370
#>  [67,] -1.911859e-01  4.358599e-01  0.87947375 -1.367791e-02 -0.0235869695
#>  [68,] -3.223504e-01  3.501659e-01  0.87947375  2.797851e-02 -0.0101061118
#>  [69,] -4.185832e-01  2.265259e-01  0.87947375 -1.692971e-02  0.0073153331
#>  [70,] -4.694561e-01  7.833837e-02  0.87947375  4.229851e-02  0.0166344834
#>  [71,] -4.694561e-01 -7.833837e-02  0.87947375 -4.064555e-02  0.0231393583
#>  [72,] -4.185832e-01 -2.265259e-01  0.87947375  2.103837e-02  0.0279252531
#>  [73,] -3.223504e-01 -3.501659e-01  0.87947375  1.342141e-02  0.0390898457
#>  [74,] -1.911859e-01 -4.358599e-01  0.87947375  4.837250e-02 -0.0053081617
#>  [75,] -3.930342e-02 -4.743218e-01  0.87947375  1.220231e-02 -0.0436405862
#>  [76,]  1.168382e-01 -4.613835e-01  0.87947375  1.311327e-03  0.0028156905
#>  [77,]  2.603185e-01 -3.984472e-01  0.87947375  4.662341e-02  0.0076821065
#>  [78,]  3.755894e-01 -2.923329e-01  0.87947375 -3.211216e-03  0.0418541325
#>  [79,]  4.501593e-01 -1.545399e-01  0.87947375 -3.708657e-02  0.0223146545
#>  [80,]  4.759474e-01 -1.165735e-16  0.87947375 -4.648162e-02  0.0083071899
#>  [81,]  6.142127e-01  0.000000e+00  0.78914051 -2.504196e-02 -0.0179322477
#>  [82,]  5.809330e-01  1.994345e-01  0.78914051  1.484018e-02  0.0349876859
#>  [83,]  4.847001e-01  3.772573e-01  0.78914051  1.164625e-02 -0.0034494386
#>  [84,]  3.359425e-01  5.141983e-01  0.78914051  3.871694e-05 -0.0261097406
#>  [85,]  1.507803e-01  5.954180e-01  0.78914051 -1.377134e-02 -0.0462074856
#>  [86,] -5.072128e-02  6.121149e-01  0.78914051 -1.846019e-02  0.0273812843
#>  [87,] -2.467264e-01  5.624796e-01  0.78914051  1.237465e-02  0.0186937040
#>  [88,] -4.159950e-01  4.518910e-01  0.78914051  4.147861e-02 -0.0326097892
#>  [89,] -5.401840e-01  2.923329e-01  0.78914051  9.710992e-03 -0.0425857586
#>  [90,] -6.058357e-01  1.010961e-01  0.78914051  1.431111e-02 -0.0226892096
#>  [91,] -6.058357e-01 -1.010961e-01  0.78914051 -1.256721e-02  0.0136312039
#>  [92,] -5.401840e-01 -2.923329e-01  0.78914051 -3.770248e-02 -0.0339925562
#>  [93,] -4.159950e-01 -4.518910e-01  0.78914051 -1.450967e-02 -0.0318546873
#>  [94,] -2.467264e-01 -5.624796e-01  0.78914051 -4.991925e-02 -0.0067028888
#>  [95,] -5.072128e-02 -6.121149e-01  0.78914051 -1.969192e-02  0.0373170787
#>  [96,]  1.507803e-01 -5.954180e-01  0.78914051 -2.974448e-02  0.0259618335
#>  [97,]  3.359425e-01 -5.141983e-01  0.78914051 -3.681899e-02  0.0358531493
#>  [98,]  4.847001e-01 -3.772573e-01  0.78914051 -4.043449e-02 -0.0491625635
#>  [99,]  5.809330e-01 -1.994345e-01  0.78914051  2.140616e-02  0.0270620452
#> [100,]  6.142127e-01 -1.504387e-16  0.78914051  3.575449e-02  0.0494684589
#> [101,]  7.357239e-01  0.000000e+00  0.67728157  2.010213e-02 -0.0142752257
#> [102,]  6.958604e-01  2.388892e-01  0.67728157 -4.655977e-02 -0.0248073300
#> [103,]  5.805895e-01  4.518910e-01  0.67728157  4.153789e-04 -0.0157495608
#> [104,]  4.024028e-01  6.159234e-01  0.67728157 -7.388127e-03  0.0345724812
#> [105,]  1.806095e-01  7.132110e-01  0.67728157  4.050325e-03  0.0029446100
#> [106,] -6.075560e-02  7.332110e-01  0.67728157 -2.424885e-02  0.0187594104
#> [107,] -2.955369e-01  6.737563e-01  0.67728157 -4.884187e-02 -0.0457758490
#> [108,] -4.982922e-01  5.412897e-01  0.67728157  2.171827e-02 -0.0026600885
#> [109,] -6.470499e-01  3.501659e-01  0.67728157 -2.872447e-02 -0.0414039600
#> [110,] -7.256896e-01  1.210962e-01  0.67728157  4.413612e-02 -0.0307138895
#> [111,] -7.256896e-01 -1.210962e-01  0.67728157  4.847650e-02 -0.0138189984
#> [112,] -6.470499e-01 -3.501659e-01  0.67728157 -1.285016e-02 -0.0498558839
#> [113,] -4.982922e-01 -5.412897e-01  0.67728157  1.832099e-02  0.0280338406
#> [114,] -2.955369e-01 -6.737563e-01  0.67728157 -3.382916e-02  0.0435488504
#> [115,] -6.075560e-02 -7.332110e-01  0.67728157  4.864992e-02  0.0482647167
#> [116,]  1.806095e-01 -7.132110e-01  0.67728157 -4.653226e-02  0.0127853509
#> [117,]  4.024028e-01 -6.159234e-01  0.67728157 -3.477368e-03 -0.0319731487
#> [118,]  5.805895e-01 -4.518910e-01  0.67728157 -5.098254e-03  0.0348275465
#> [119,]  6.958604e-01 -2.388892e-01  0.67728157 -1.320561e-02 -0.0139882761
#> [120,]  7.357239e-01 -1.802004e-16  0.67728157 -4.661531e-04 -0.0474228820
#> [121,]  8.371665e-01  0.000000e+00  0.54694816  3.149360e-02  0.0441705484
#> [122,]  7.918065e-01  2.718275e-01  0.54694816 -2.745434e-02 -0.0203955581
#> [123,]  6.606420e-01  5.141983e-01  0.54694816 -4.892582e-02 -0.0080970241
#> [124,]  4.578867e-01  7.008477e-01  0.54694816 -4.631791e-02  0.0115755656
#> [125,]  2.055122e-01  8.115494e-01  0.54694816 -7.494114e-03 -0.0478689952
#> [126,] -6.913266e-02  8.343071e-01  0.54694816  2.861160e-02  0.0200023154
#> [127,] -3.362859e-01  7.666547e-01  0.54694816 -4.358824e-02 -0.0332733955
#> [128,] -5.669974e-01  6.159234e-01  0.54694816 -3.196353e-02 -0.0333539583
#> [129,] -7.362659e-01  3.984472e-01  0.54694816 -3.663724e-02 -0.0204066110
#> [130,] -8.257486e-01  1.377931e-01  0.54694816 -1.975613e-02  0.0189010124
#> [131,] -8.257486e-01 -1.377931e-01  0.54694816  3.364112e-02  0.0379742808
#> [132,] -7.362659e-01 -3.984472e-01  0.54694816  3.062627e-02  0.0334562535
#> [133,] -5.669974e-01 -6.159234e-01  0.54694816 -1.084980e-02  0.0159564751
#> [134,] -3.362859e-01 -7.666547e-01  0.54694816  2.962196e-02 -0.0418507333
#> [135,] -6.913266e-02 -8.343071e-01  0.54694816 -2.440594e-02 -0.0365537456
#> [136,]  2.055122e-01 -8.115494e-01  0.54694816 -1.329895e-02 -0.0278687609
#> [137,]  4.578867e-01 -7.008477e-01  0.54694816 -3.803862e-02 -0.0097203438
#> [138,]  6.606420e-01 -5.141983e-01  0.54694816  8.926194e-03 -0.0400998932
#> [139,]  7.918065e-01 -2.718275e-01  0.54694816 -1.538305e-02  0.0154731375
#> [140,]  8.371665e-01 -2.050466e-16  0.54694816 -3.501523e-02  0.0174571527
#> [141,]  9.157733e-01  0.000000e+00  0.40169542  3.727269e-02 -0.0411400271
#> [142,]  8.661542e-01  2.973511e-01  0.40169542  4.798341e-02  0.0120417417
#> [143,]  7.226738e-01  5.624796e-01  0.40169542  3.630871e-02 -0.0120943295
#> [144,]  5.008805e-01  7.666547e-01  0.40169542 -1.467962e-02  0.0028938754
#> [145,]  2.248091e-01  8.877509e-01  0.40169542  4.667778e-03  0.0216059040
#> [146,] -7.562396e-02  9.126455e-01  0.40169542 -3.941027e-02  0.0257809480
#> [147,] -3.678620e-01  8.386408e-01  0.40169542  3.566019e-02  0.0070101255
#> [148,] -6.202364e-01  6.737563e-01  0.40169542 -1.423162e-02 -0.0080849630
#> [149,] -8.053986e-01  4.358599e-01  0.40169542 -1.028858e-02 -0.0259315589
#> [150,] -9.032834e-01  1.507313e-01  0.40169542  2.808479e-02 -0.0181442028
#> [151,] -9.032834e-01 -1.507313e-01  0.40169542 -4.355619e-03  0.0355401649
#> [152,] -8.053986e-01 -4.358599e-01  0.40169542 -3.327374e-02  0.0231951406
#> [153,] -6.202364e-01 -6.737563e-01  0.40169542  4.346168e-02 -0.0479888653
#> [154,] -3.678620e-01 -8.386408e-01  0.40169542  4.533953e-03  0.0007952142
#> [155,] -7.562396e-02 -9.126455e-01  0.40169542  2.850857e-02 -0.0099837501
#> [156,]  2.248091e-01 -8.877509e-01  0.40169542  4.122771e-02  0.0056869607
#> [157,]  5.008805e-01 -7.666547e-01  0.40169542 -1.722048e-02  0.0009412200
#> [158,]  7.226738e-01 -5.624796e-01  0.40169542 -4.827358e-02 -0.0134702494
#> [159,]  8.661542e-01 -2.973511e-01  0.40169542  3.850037e-03 -0.0256164785
#> [160,]  9.157733e-01 -2.242998e-16  0.40169542 -2.250911e-02 -0.0468201660
#> [161,]  9.694003e-01  0.000000e+00  0.24548549 -1.429300e-02 -0.0135487690
#> [162,]  9.168755e-01  3.147638e-01  0.24548549 -1.526603e-02  0.0219497899
#> [163,]  7.649930e-01  5.954180e-01  0.24548549  3.679693e-04  0.0117803611
#> [164,]  5.302117e-01  8.115494e-01  0.24548549 -2.525221e-02 -0.0086506287
#> [165,]  2.379737e-01  9.397369e-01  0.24548549 -3.469703e-02  0.0184315171
#> [166,] -8.005244e-02  9.660893e-01  0.24548549 -1.906814e-02 -0.0360715476
#> [167,] -3.894037e-01  8.877509e-01  0.24548549 -3.610104e-02 -0.0461215101
#> [168,] -6.565569e-01  7.132110e-01  0.24548549  2.250877e-02  0.0160865578
#> [169,] -8.525621e-01  4.613835e-01  0.24548549  1.982792e-02 -0.0285978725
#> [170,] -9.561789e-01  1.595580e-01  0.24548549 -2.766924e-02  0.0345246856
#> [171,] -9.561789e-01 -1.595580e-01  0.24548549 -1.669196e-02  0.0353035867
#> [172,] -8.525621e-01 -4.613835e-01  0.24548549 -2.661127e-02  0.0271172372
#> [173,] -6.565569e-01 -7.132110e-01  0.24548549  3.690593e-02 -0.0053837291
#> [174,] -3.894037e-01 -8.877509e-01  0.24548549  3.447108e-02 -0.0417743478
#> [175,] -8.005244e-02 -9.660893e-01  0.24548549  4.407497e-02  0.0321266160
#> [176,]  2.379737e-01 -9.397369e-01  0.24548549 -2.289297e-02 -0.0010581523
#> [177,]  5.302117e-01 -8.115494e-01  0.24548549  3.195787e-02 -0.0227824431
#> [178,]  7.649930e-01 -5.954180e-01  0.24548549 -3.383919e-02 -0.0238702872
#> [179,]  9.168755e-01 -3.147638e-01  0.24548549 -2.019575e-02  0.0326573602
#> [180,]  9.694003e-01 -2.374346e-16  0.24548549  1.546098e-02  0.0190673835
#> [181,]  9.965845e-01  0.000000e+00  0.08257935 -1.336718e-02  0.0419352430
#> [182,]  9.425868e-01  3.235905e-01  0.08257935  4.000855e-03  0.0273729427
#> [183,]  7.864452e-01  6.121149e-01  0.08257935 -2.275035e-02 -0.0329492582
#> [184,]  5.450801e-01  8.343071e-01  0.08257935  4.977679e-02 -0.0071118049
#> [185,]  2.446470e-01  9.660893e-01  0.08257935 -4.614536e-02  0.0125517385
#> [186,] -8.229730e-02  9.931807e-01  0.08257935  2.359132e-02 -0.0372355393
#> [187,] -4.003234e-01  9.126455e-01  0.08257935 -5.026203e-03 -0.0242566349
#> [188,] -6.749683e-01  7.332110e-01  0.08257935  4.924082e-02  0.0219561845
#> [189,] -8.764699e-01  4.743218e-01  0.08257935  3.183935e-02 -0.0432288951
#> [190,] -9.829924e-01  1.640324e-01  0.08257935 -4.182037e-02 -0.0252900630
#> [191,] -9.829924e-01 -1.640324e-01  0.08257935  1.573607e-02  0.0429220151
#> [192,] -8.764699e-01 -4.743218e-01  0.08257935 -1.530281e-02  0.0179613811
#> [193,] -6.749683e-01 -7.332110e-01  0.08257935 -9.891095e-03  0.0166643756
#> [194,] -4.003234e-01 -9.126455e-01  0.08257935  6.453046e-03  0.0161850222
#> [195,] -8.229730e-02 -9.931807e-01  0.08257935 -8.333704e-03  0.0453793376
#> [196,]  2.446470e-01 -9.660893e-01  0.08257935  4.835035e-02  0.0119703040
#> [197,]  5.450801e-01 -8.343071e-01  0.08257935  4.001286e-02 -0.0290068778
#> [198,]  7.864452e-01 -6.121149e-01  0.08257935 -3.564835e-02  0.0104970251
#> [199,]  9.425868e-01 -3.235905e-01  0.08257935  3.991483e-02  0.0282884538
#> [200,]  9.965845e-01 -2.440928e-16  0.08257935  2.323870e-02 -0.0163027727
#> [201,]  9.965845e-01  0.000000e+00 -0.08257935 -4.615866e-02 -0.0380071864
#> [202,]  9.425868e-01  3.235905e-01 -0.08257935 -2.955813e-02  0.0424040244
#> [203,]  7.864452e-01  6.121149e-01 -0.08257935 -4.460265e-02 -0.0185994519
#> [204,]  5.450801e-01  8.343071e-01 -0.08257935  4.308776e-02 -0.0207426412
#> [205,]  2.446470e-01  9.660893e-01 -0.08257935 -3.873241e-02 -0.0190330383
#> [206,] -8.229730e-02  9.931807e-01 -0.08257935 -1.827751e-02 -0.0091702867
#> [207,] -4.003234e-01  9.126455e-01 -0.08257935  3.668533e-02  0.0149097691
#> [208,] -6.749683e-01  7.332110e-01 -0.08257935 -1.217878e-02  0.0154284348
#> [209,] -8.764699e-01  4.743218e-01 -0.08257935 -1.982870e-02 -0.0168893659
#> [210,] -9.829924e-01  1.640324e-01 -0.08257935  4.151171e-02 -0.0031429331
#> [211,] -9.829924e-01 -1.640324e-01 -0.08257935 -4.417235e-02 -0.0289007221
#> [212,] -8.764699e-01 -4.743218e-01 -0.08257935  3.772102e-02  0.0245576286
#> [213,] -6.749683e-01 -7.332110e-01 -0.08257935  3.449673e-02  0.0266774244
#> [214,] -4.003234e-01 -9.126455e-01 -0.08257935  4.187780e-02  0.0136960627
#> [215,] -8.229730e-02 -9.931807e-01 -0.08257935  7.470660e-03  0.0261469203
#> [216,]  2.446470e-01 -9.660893e-01 -0.08257935 -2.386852e-02 -0.0394998899
#> [217,]  5.450801e-01 -8.343071e-01 -0.08257935 -1.382045e-02 -0.0355951726
#> [218,]  7.864452e-01 -6.121149e-01 -0.08257935  1.870621e-02 -0.0167218604
#> [219,]  9.425868e-01 -3.235905e-01 -0.08257935  3.528506e-02  0.0358124912
#> [220,]  9.965845e-01 -2.440928e-16 -0.08257935  1.231969e-02 -0.0428234379
#> [221,]  9.694003e-01  0.000000e+00 -0.24548549 -9.647154e-04  0.0140179200
#> [222,]  9.168755e-01  3.147638e-01 -0.24548549 -1.745733e-02  0.0157510552
#> [223,]  7.649930e-01  5.954180e-01 -0.24548549 -3.392341e-02 -0.0214371409
#> [224,]  5.302117e-01  8.115494e-01 -0.24548549  5.617756e-03 -0.0393815453
#> [225,]  2.379737e-01  9.397369e-01 -0.24548549  3.241084e-02 -0.0145615052
#> [226,] -8.005244e-02  9.660893e-01 -0.24548549 -4.077092e-02  0.0378515478
#> [227,] -3.894037e-01  8.877509e-01 -0.24548549 -4.706302e-02 -0.0186402297
#> [228,] -6.565569e-01  7.132110e-01 -0.24548549  2.608010e-02 -0.0401645005
#> [229,] -8.525621e-01  4.613835e-01 -0.24548549 -4.825318e-02 -0.0395903957
#> [230,] -9.561789e-01  1.595580e-01 -0.24548549  2.255943e-02 -0.0110641956
#> [231,] -9.561789e-01 -1.595580e-01 -0.24548549 -8.967100e-03 -0.0440654797
#> [232,] -8.525621e-01 -4.613835e-01 -0.24548549 -2.788474e-03 -0.0170914530
#> [233,] -6.565569e-01 -7.132110e-01 -0.24548549 -3.273690e-02  0.0416006673
#> [234,] -3.894037e-01 -8.877509e-01 -0.24548549 -5.849681e-03  0.0135452446
#> [235,] -8.005244e-02 -9.660893e-01 -0.24548549 -2.197788e-02  0.0049922454
#> [236,]  2.379737e-01 -9.397369e-01 -0.24548549  2.406905e-02 -0.0088914151
#> [237,]  5.302117e-01 -8.115494e-01 -0.24548549  3.629414e-02 -0.0488401050
#> [238,]  7.649930e-01 -5.954180e-01 -0.24548549  4.984348e-02  0.0280523448
#> [239,]  9.168755e-01 -3.147638e-01 -0.24548549  1.785800e-02  0.0304146923
#> [240,]  9.694003e-01 -2.374346e-16 -0.24548549  2.376368e-02 -0.0330301560
#> [241,]  9.157733e-01  0.000000e+00 -0.40169542 -2.363932e-02 -0.0378807141
#> [242,]  8.661542e-01  2.973511e-01 -0.40169542 -3.144070e-02 -0.0436832453
#> [243,]  7.226738e-01  5.624796e-01 -0.40169542 -1.344896e-02  0.0114049602
#> [244,]  5.008805e-01  7.666547e-01 -0.40169542  3.298782e-02 -0.0213487461
#> [245,]  2.248091e-01  8.877509e-01 -0.40169542  9.410643e-03  0.0042389313
#> [246,] -7.562396e-02  9.126455e-01 -0.40169542  2.765249e-02 -0.0207808492
#> [247,] -3.678620e-01  8.386408e-01 -0.40169542  2.614364e-02 -0.0441213807
#> [248,] -6.202364e-01  6.737563e-01 -0.40169542  2.775669e-02  0.0409693898
#> [249,] -8.053986e-01  4.358599e-01 -0.40169542  3.402187e-02  0.0458091147
#> [250,] -9.032834e-01  1.507313e-01 -0.40169542  9.749733e-03  0.0359725750
#> [251,] -9.032834e-01 -1.507313e-01 -0.40169542  4.986264e-02 -0.0325394158
#> [252,] -8.053986e-01 -4.358599e-01 -0.40169542 -5.689704e-03 -0.0215588731
#> [253,] -6.202364e-01 -6.737563e-01 -0.40169542  4.731525e-03  0.0249050695
#> [254,] -3.678620e-01 -8.386408e-01 -0.40169542 -4.281487e-02 -0.0028277011
#> [255,] -7.562396e-02 -9.126455e-01 -0.40169542 -2.918447e-02  0.0223142910
#> [256,]  2.248091e-01 -8.877509e-01 -0.40169542 -4.884492e-03  0.0325163283
#> [257,]  5.008805e-01 -7.666547e-01 -0.40169542 -4.780268e-02  0.0259833605
#> [258,]  7.226738e-01 -5.624796e-01 -0.40169542 -2.214287e-02  0.0187410582
#> [259,]  8.661542e-01 -2.973511e-01 -0.40169542 -6.080527e-03 -0.0305400811
#> [260,]  9.157733e-01 -2.242998e-16 -0.40169542  4.103525e-02  0.0042170736
#> [261,]  8.371665e-01  0.000000e+00 -0.54694816  6.039320e-03  0.0332121525
#> [262,]  7.918065e-01  2.718275e-01 -0.54694816 -1.564643e-02  0.0329608525
#> [263,]  6.606420e-01  5.141983e-01 -0.54694816  4.351170e-02  0.0413685252
#> [264,]  4.578867e-01  7.008477e-01 -0.54694816  3.297970e-02  0.0265417153
#> [265,]  2.055122e-01  8.115494e-01 -0.54694816  2.733009e-02  0.0016249159
#> [266,] -6.913266e-02  8.343071e-01 -0.54694816 -4.615736e-02 -0.0091616368
#> [267,] -3.362859e-01  7.666547e-01 -0.54694816 -1.988585e-02 -0.0154283055
#> [268,] -5.669974e-01  6.159234e-01 -0.54694816 -3.406692e-02  0.0092969455
#> [269,] -7.362659e-01  3.984472e-01 -0.54694816 -4.675491e-02  0.0078367776
#> [270,] -8.257486e-01  1.377931e-01 -0.54694816  3.198878e-02 -0.0339836503
#> [271,] -8.257486e-01 -1.377931e-01 -0.54694816  4.717143e-02 -0.0234401830
#> [272,] -7.362659e-01 -3.984472e-01 -0.54694816 -1.322317e-02  0.0289493584
#> [273,] -5.669974e-01 -6.159234e-01 -0.54694816 -2.633505e-02  0.0217553166
#> [274,] -3.362859e-01 -7.666547e-01 -0.54694816 -2.903464e-02 -0.0334817440
#> [275,] -6.913266e-02 -8.343071e-01 -0.54694816  1.637944e-02 -0.0154701410
#> [276,]  2.055122e-01 -8.115494e-01 -0.54694816  8.810025e-03 -0.0177106471
#> [277,]  4.578867e-01 -7.008477e-01 -0.54694816 -3.746246e-03  0.0023404620
#> [278,]  6.606420e-01 -5.141983e-01 -0.54694816 -3.423887e-02 -0.0126785869
#> [279,]  7.918065e-01 -2.718275e-01 -0.54694816  2.285647e-02 -0.0423836458
#> [280,]  8.371665e-01 -2.050466e-16 -0.54694816 -4.857378e-02  0.0290258873
#> [281,]  7.357239e-01  0.000000e+00 -0.67728157  3.474031e-02 -0.0238284791
#> [282,]  6.958604e-01  2.388892e-01 -0.67728157  3.033973e-02 -0.0065352196
#> [283,]  5.805895e-01  4.518910e-01 -0.67728157  2.515528e-02 -0.0351819196
#> [284,]  4.024028e-01  6.159234e-01 -0.67728157  3.258067e-02 -0.0404684342
#> [285,]  1.806095e-01  7.132110e-01 -0.67728157 -4.382060e-02 -0.0405836351
#> [286,] -6.075560e-02  7.332110e-01 -0.67728157 -3.325462e-02  0.0128343012
#> [287,] -2.955369e-01  6.737563e-01 -0.67728157  7.270221e-03  0.0187457161
#> [288,] -4.982922e-01  5.412897e-01 -0.67728157  1.250899e-02  0.0292764872
#> [289,] -6.470499e-01  3.501659e-01 -0.67728157  1.861847e-02  0.0424894955
#> [290,] -7.256896e-01  1.210962e-01 -0.67728157  4.247432e-02 -0.0250321991
#> [291,] -7.256896e-01 -1.210962e-01 -0.67728157 -1.594968e-02 -0.0145822989
#> [292,] -6.470499e-01 -3.501659e-01 -0.67728157 -3.194949e-02  0.0318413210
#> [293,] -4.982922e-01 -5.412897e-01 -0.67728157 -1.900761e-02  0.0231287681
#> [294,] -2.955369e-01 -6.737563e-01 -0.67728157  4.712041e-02  0.0066390378
#> [295,] -6.075560e-02 -7.332110e-01 -0.67728157 -5.053640e-03 -0.0351758547
#> [296,]  1.806095e-01 -7.132110e-01 -0.67728157 -2.470935e-02  0.0172915439
#> [297,]  4.024028e-01 -6.159234e-01 -0.67728157 -4.016104e-02  0.0187803801
#> [298,]  5.805895e-01 -4.518910e-01 -0.67728157  2.955073e-02  0.0179914030
#> [299,]  6.958604e-01 -2.388892e-01 -0.67728157 -4.778154e-02  0.0388279600
#> [300,]  7.357239e-01 -1.802004e-16 -0.67728157  3.233857e-02 -0.0051391185
#> [301,]  6.142127e-01  0.000000e+00 -0.78914051 -4.486809e-02  0.0184453821
#> [302,]  5.809330e-01  1.994345e-01 -0.78914051 -4.631775e-02 -0.0428399980
#> [303,]  4.847001e-01  3.772573e-01 -0.78914051  3.109507e-03  0.0370683109
#> [304,]  3.359425e-01  5.141983e-01 -0.78914051 -1.038167e-02  0.0325215872
#> [305,]  1.507803e-01  5.954180e-01 -0.78914051 -1.406535e-02  0.0342661202
#> [306,] -5.072128e-02  6.121149e-01 -0.78914051  1.841757e-02 -0.0190882406
#> [307,] -2.467264e-01  5.624796e-01 -0.78914051 -3.501227e-02 -0.0237409763
#> [308,] -4.159950e-01  4.518910e-01 -0.78914051 -2.820250e-02  0.0432651362
#> [309,] -5.401840e-01  2.923329e-01 -0.78914051  4.359630e-02 -0.0081278855
#> [310,] -6.058357e-01  1.010961e-01 -0.78914051  3.760547e-02 -0.0215141515
#> [311,] -6.058357e-01 -1.010961e-01 -0.78914051 -4.345147e-02  0.0194656136
#> [312,] -5.401840e-01 -2.923329e-01 -0.78914051  7.964710e-03  0.0408579351
#> [313,] -4.159950e-01 -4.518910e-01 -0.78914051 -2.881330e-02 -0.0354893491
#> [314,] -2.467264e-01 -5.624796e-01 -0.78914051 -2.838622e-02  0.0272115505
#> [315,] -5.072128e-02 -6.121149e-01 -0.78914051  1.553604e-03  0.0096642551
#> [316,]  1.507803e-01 -5.954180e-01 -0.78914051  2.709294e-02  0.0284285932
#> [317,]  3.359425e-01 -5.141983e-01 -0.78914051 -2.582841e-02  0.0279427908
#> [318,]  4.847001e-01 -3.772573e-01 -0.78914051  1.842675e-02  0.0106357523
#> [319,]  5.809330e-01 -1.994345e-01 -0.78914051  4.200818e-02 -0.0315552889
#> [320,]  6.142127e-01 -1.504387e-16 -0.78914051 -5.994582e-03 -0.0025648329
#> [321,]  4.759474e-01  0.000000e+00 -0.87947375  2.193789e-02  0.0064179965
#> [322,]  4.501593e-01  1.545399e-01 -0.87947375 -2.054620e-03  0.0472810680
#> [323,]  3.755894e-01  2.923329e-01 -0.87947375  1.925337e-02 -0.0207815134
#> [324,]  2.603185e-01  3.984472e-01 -0.87947375 -3.247059e-02 -0.0116014455
#> [325,]  1.168382e-01  4.613835e-01 -0.87947375  4.261971e-02  0.0401021755
#> [326,] -3.930342e-02  4.743218e-01 -0.87947375 -4.279998e-02 -0.0410951136
#> [327,] -1.911859e-01  4.358599e-01 -0.87947375 -2.655641e-02 -0.0137846122
#> [328,] -3.223504e-01  3.501659e-01 -0.87947375 -4.083009e-02  0.0287460793
#> [329,] -4.185832e-01  2.265259e-01 -0.87947375 -5.713772e-03  0.0119035573
#> [330,] -4.694561e-01  7.833837e-02 -0.87947375 -1.477730e-03 -0.0444554345
#> [331,] -4.694561e-01 -7.833837e-02 -0.87947375  2.107090e-02 -0.0074048647
#> [332,] -4.185832e-01 -2.265259e-01 -0.87947375  4.074728e-02 -0.0432008458
#> [333,] -3.223504e-01 -3.501659e-01 -0.87947375 -2.791427e-02 -0.0276024753
#> [334,] -1.911859e-01 -4.358599e-01 -0.87947375 -4.939518e-02 -0.0177683791
#> [335,] -3.930342e-02 -4.743218e-01 -0.87947375  1.263657e-02  0.0082569377
#> [336,]  1.168382e-01 -4.613835e-01 -0.87947375 -2.005284e-02  0.0207680961
#> [337,]  2.603185e-01 -3.984472e-01 -0.87947375  2.606288e-02  0.0094878329
#> [338,]  3.755894e-01 -2.923329e-01 -0.87947375  8.065626e-04 -0.0046935826
#> [339,]  4.501593e-01 -1.545399e-01 -0.87947375  4.016190e-02  0.0123492697
#> [340,]  4.759474e-01 -1.165735e-16 -0.87947375  2.397653e-02 -0.0319148199
#> [341,]  3.246995e-01  0.000000e+00 -0.94581724 -3.688701e-02  0.0357243336
#> [342,]  3.071064e-01  1.054297e-01 -0.94581724 -2.610489e-02 -0.0250900889
#> [343,]  2.562335e-01  1.994345e-01 -0.94581724  2.530692e-02  0.0305797825
#> [344,]  1.775938e-01  2.718275e-01 -0.94581724 -4.746623e-02  0.0008068894
#> [345,]  7.970901e-02  3.147638e-01 -0.94581724  2.831136e-02 -0.0299398511
#> [346,] -2.681347e-02  3.235905e-01 -0.94581724 -3.708595e-02 -0.0167171007
#> [347,] -1.304303e-01  2.973511e-01 -0.94581724 -4.577893e-02  0.0117217810
#> [348,] -2.199130e-01  2.388892e-01 -0.94581724  3.741720e-02  0.0040350181
#> [349,] -2.855647e-01  1.545399e-01 -0.94581724  2.304289e-02  0.0232406309
#> [350,] -3.202710e-01  5.344378e-02 -0.94581724 -2.774930e-02  0.0200998565
#> [351,] -3.202710e-01 -5.344378e-02 -0.94581724 -9.653758e-03 -0.0122408672
#> [352,] -2.855647e-01 -1.545399e-01 -0.94581724  4.948460e-02 -0.0139943485
#> [353,] -2.199130e-01 -2.388892e-01 -0.94581724 -6.211845e-03  0.0388301017
#> [354,] -1.304303e-01 -2.973511e-01 -0.94581724 -2.626562e-02  0.0092671094
#> [355,] -2.681347e-02 -3.235905e-01 -0.94581724  3.507734e-02  0.0328538819
#> [356,]  7.970901e-02 -3.147638e-01 -0.94581724 -1.593362e-02  0.0322694414
#> [357,]  1.775938e-01 -2.718275e-01 -0.94581724  4.198787e-03 -0.0205651568
#> [358,]  2.562335e-01 -1.994345e-01 -0.94581724 -2.810138e-02 -0.0498711132
#> [359,]  3.071064e-01 -1.054297e-01 -0.94581724  1.726668e-02  0.0470405370
#> [360,]  3.246995e-01 -7.952843e-17 -0.94581724 -2.510992e-02 -0.0313419879
#> [361,]  1.645946e-01  0.000000e+00 -0.98636130 -3.435566e-02  0.0477830785
#> [362,]  1.556764e-01  5.344378e-02 -0.98636130 -1.369361e-02  0.0087820548
#> [363,]  1.298883e-01  1.010961e-01 -0.98636130  4.371308e-02  0.0409046817
#> [364,]  9.002471e-02  1.377931e-01 -0.98636130 -6.163400e-03  0.0105919167
#> [365,]  4.040558e-02  1.595580e-01 -0.98636130  9.500874e-03 -0.0309713586
#> [366,] -1.359211e-02  1.640324e-01 -0.98636130  4.354907e-03 -0.0208237002
#> [367,] -6.611689e-02  1.507313e-01 -0.98636130 -1.150747e-02  0.0487091085
#> [368,] -1.114769e-01  1.210962e-01 -0.98636130 -3.413685e-02  0.0204033481
#> [369,] -1.447566e-01  7.833837e-02 -0.98636130  3.592309e-02 -0.0282089352
#> [370,] -1.623497e-01  2.709138e-02 -0.98636130 -9.643194e-03  0.0023281282
#> [371,] -1.623497e-01 -2.709138e-02 -0.98636130  1.218064e-02  0.0206690365
#> [372,] -1.447566e-01 -7.833837e-02 -0.98636130 -2.991909e-02  0.0335687824
#> [373,] -1.114769e-01 -1.210962e-01 -0.98636130 -3.381288e-03 -0.0337689344
#> [374,] -6.611689e-02 -1.507313e-01 -0.98636130  1.073743e-02  0.0413793686
#> [375,] -1.359211e-02 -1.640324e-01 -0.98636130  1.841355e-04 -0.0360326619
#> [376,]  4.040558e-02 -1.595580e-01 -0.98636130 -6.304969e-03 -0.0273940365
#> [377,]  9.002471e-02 -1.377931e-01 -0.98636130 -3.276833e-03 -0.0199279093
#> [378,]  1.298883e-01 -1.010961e-01 -0.98636130 -2.627968e-02  0.0031689772
#> [379,]  1.556764e-01 -5.344378e-02 -0.98636130  1.843509e-02 -0.0063964503
#> [380,]  1.645946e-01 -4.031405e-17 -0.98636130 -3.954099e-02  0.0352735527
#> [381,]  1.224647e-16  0.000000e+00 -1.00000000 -3.530609e-03  0.0334915186
#> [382,]  1.158292e-16  3.976422e-17 -1.00000000  3.136026e-03 -0.0134660115
#> [383,]  9.664184e-17  7.521936e-17 -1.00000000 -2.232496e-02 -0.0266729430
#> [384,]  6.698183e-17  1.025233e-16 -1.00000000  4.986355e-03  0.0214869546
#> [385,]  3.006330e-17  1.187173e-16 -1.00000000 -2.214946e-02  0.0362943965
#> [386,] -1.011305e-17  1.220464e-16 -1.00000000  1.499387e-02 -0.0049533740
#> [387,] -4.919350e-17  1.121499e-16 -1.00000000 -1.037501e-02 -0.0468848157
#> [388,] -8.294307e-17  9.010019e-17 -1.00000000  4.770996e-02 -0.0406807824
#> [389,] -1.077045e-16  5.828675e-17 -1.00000000 -3.595874e-02 -0.0178826248
#> [390,] -1.207944e-16  2.015702e-17 -1.00000000  2.388100e-02  0.0474003338
#> [391,] -1.207944e-16 -2.015702e-17 -1.00000000 -2.710626e-02  0.0077575058
#> [392,] -1.077045e-16 -5.828675e-17 -1.00000000  4.730719e-02  0.0129874283
#> [393,] -8.294307e-17 -9.010019e-17 -1.00000000  3.482662e-02  0.0274112441
#> [394,] -4.919350e-17 -1.121499e-16 -1.00000000  2.461111e-03  0.0479897551
#> [395,] -1.011305e-17 -1.220464e-16 -1.00000000  4.694575e-03 -0.0288173655
#> [396,]  3.006330e-17 -1.187173e-16 -1.00000000  1.353626e-02  0.0199320918
#> [397,]  6.698183e-17 -1.025233e-16 -1.00000000 -4.955571e-02  0.0340592087
#> [398,]  9.664184e-17 -7.521936e-17 -1.00000000  1.665852e-02  0.0273357358
#> [399,]  1.158292e-16 -3.976422e-17 -1.00000000  1.874820e-02 -0.0349686657
#> [400,]  1.224647e-16 -2.999520e-32 -1.00000000 -2.295107e-02  0.0209931267
#>                 [,6]
#>   [1,] -3.553979e-02
#>   [2,]  3.486927e-02
#>   [3,]  4.392672e-02
#>   [4,]  1.002101e-02
#>   [5,] -3.916878e-02
#>   [6,]  2.329629e-04
#>   [7,] -2.967061e-02
#>   [8,]  4.262086e-02
#>   [9,] -6.873088e-03
#>  [10,]  3.355089e-02
#>  [11,]  4.836161e-02
#>  [12,] -4.298102e-02
#>  [13,] -4.025655e-03
#>  [14,]  3.844458e-02
#>  [15,]  1.271128e-02
#>  [16,] -1.913043e-02
#>  [17,] -4.310435e-02
#>  [18,]  1.343779e-02
#>  [19,]  3.913285e-02
#>  [20,]  3.185013e-02
#>  [21,] -1.772342e-02
#>  [22,] -1.683174e-02
#>  [23,]  4.904842e-02
#>  [24,] -2.412347e-02
#>  [25,]  4.731039e-02
#>  [26,] -4.800023e-02
#>  [27,]  2.060722e-02
#>  [28,]  7.987511e-03
#>  [29,] -3.732614e-02
#>  [30,]  3.265659e-02
#>  [31,] -3.934535e-02
#>  [32,]  2.782055e-02
#>  [33,] -8.957195e-03
#>  [34,] -4.187444e-02
#>  [35,] -4.096410e-02
#>  [36,] -1.833752e-02
#>  [37,]  5.291158e-03
#>  [38,]  1.827919e-02
#>  [39,] -2.202988e-02
#>  [40,]  6.319229e-03
#>  [41,]  1.277819e-02
#>  [42,]  1.888811e-02
#>  [43,] -3.668927e-02
#>  [44,]  4.240006e-02
#>  [45,]  2.131579e-02
#>  [46,]  3.328063e-02
#>  [47,] -7.787780e-03
#>  [48,] -4.449518e-02
#>  [49,] -3.738750e-02
#>  [50,]  4.752554e-02
#>  [51,]  4.832704e-02
#>  [52,]  9.116688e-03
#>  [53,] -3.877643e-02
#>  [54,]  4.410609e-02
#>  [55,]  4.577667e-02
#>  [56,] -4.547840e-03
#>  [57,] -2.173033e-02
#>  [58,] -1.693482e-02
#>  [59,] -7.294555e-03
#>  [60,] -4.056674e-02
#>  [61,]  2.913322e-02
#>  [62,] -3.249532e-02
#>  [63,]  4.152010e-02
#>  [64,]  1.955347e-04
#>  [65,] -3.876300e-02
#>  [66,]  4.280878e-02
#>  [67,]  1.154107e-02
#>  [68,] -3.122757e-02
#>  [69,]  8.746056e-03
#>  [70,] -3.537531e-02
#>  [71,] -1.768345e-02
#>  [72,] -1.588652e-02
#>  [73,]  8.498256e-03
#>  [74,] -1.321531e-02
#>  [75,] -4.903558e-02
#>  [76,]  3.066472e-02
#>  [77,]  1.736061e-02
#>  [78,]  2.119991e-02
#>  [79,] -3.749492e-02
#>  [80,]  2.446735e-02
#>  [81,]  3.942205e-02
#>  [82,] -4.632825e-02
#>  [83,]  4.743593e-02
#>  [84,] -2.835349e-02
#>  [85,]  3.883066e-02
#>  [86,]  4.164490e-02
#>  [87,] -2.074188e-02
#>  [88,]  4.880444e-02
#>  [89,]  2.170039e-03
#>  [90,]  1.584529e-02
#>  [91,]  3.565217e-03
#>  [92,] -4.152824e-02
#>  [93,]  4.340219e-02
#>  [94,] -4.180511e-02
#>  [95,] -3.742636e-02
#>  [96,]  9.951603e-03
#>  [97,]  4.218868e-03
#>  [98,] -3.676329e-02
#>  [99,] -4.286616e-02
#> [100,]  2.163317e-02
#> [101,] -4.273169e-04
#> [102,] -2.485376e-02
#> [103,]  4.966003e-02
#> [104,]  3.865578e-02
#> [105,] -2.484886e-02
#> [106,]  4.903650e-02
#> [107,]  1.055146e-02
#> [108,] -3.316055e-02
#> [109,]  3.700407e-02
#> [110,]  4.275281e-02
#> [111,]  1.372308e-02
#> [112,] -1.418066e-03
#> [113,] -3.532169e-02
#> [114,] -4.880886e-02
#> [115,] -2.072637e-02
#> [116,] -5.363351e-03
#> [117,]  2.024295e-02
#> [118,] -3.709343e-04
#> [119,]  3.933489e-02
#> [120,]  4.646030e-02
#> [121,] -4.109467e-02
#> [122,]  4.100209e-03
#> [123,] -1.496179e-02
#> [124,] -2.343874e-02
#> [125,]  4.093590e-02
#> [126,] -3.593191e-03
#> [127,]  2.019247e-02
#> [128,] -4.074347e-02
#> [129,] -6.652787e-03
#> [130,] -4.859287e-03
#> [131,] -3.896349e-02
#> [132,]  4.974859e-02
#> [133,] -2.354634e-02
#> [134,]  3.999261e-02
#> [135,]  1.409953e-02
#> [136,]  5.449354e-03
#> [137,]  1.782738e-02
#> [138,]  2.980763e-02
#> [139,]  1.626547e-02
#> [140,]  6.659579e-03
#> [141,] -4.403694e-02
#> [142,]  3.252864e-02
#> [143,] -3.321784e-02
#> [144,] -1.122086e-02
#> [145,]  3.304717e-02
#> [146,]  1.089300e-02
#> [147,] -4.902497e-02
#> [148,] -3.300998e-03
#> [149,] -4.297408e-02
#> [150,] -4.190472e-02
#> [151,]  6.451108e-03
#> [152,] -3.674762e-02
#> [153,]  3.346675e-03
#> [154,] -9.567526e-03
#> [155,]  4.946089e-02
#> [156,] -2.818782e-02
#> [157,]  3.308436e-02
#> [158,]  5.050050e-03
#> [159,] -4.877238e-02
#> [160,]  1.524336e-02
#> [161,]  1.066774e-02
#> [162,] -4.345645e-02
#> [163,]  4.931665e-02
#> [164,]  1.313899e-02
#> [165,] -2.703630e-02
#> [166,] -3.368004e-03
#> [167,]  3.857730e-02
#> [168,]  8.638209e-03
#> [169,]  1.253418e-02
#> [170,] -3.388104e-02
#> [171,]  1.852393e-03
#> [172,] -1.675746e-02
#> [173,] -1.592640e-02
#> [174,]  7.902568e-03
#> [175,]  3.825354e-02
#> [176,]  1.816412e-02
#> [177,] -7.255588e-03
#> [178,] -4.556995e-02
#> [179,] -3.030339e-02
#> [180,] -1.071172e-02
#> [181,] -2.098923e-02
#> [182,] -4.579722e-02
#> [183,] -2.053530e-02
#> [184,]  4.588622e-02
#> [185,] -4.651342e-02
#> [186,]  2.874710e-02
#> [187,]  1.194211e-02
#> [188,] -4.431160e-02
#> [189,]  7.801388e-04
#> [190,] -1.431920e-02
#> [191,]  2.140042e-02
#> [192,] -2.821631e-02
#> [193,]  8.497185e-03
#> [194,]  1.059573e-02
#> [195,]  4.383950e-02
#> [196,] -3.990808e-02
#> [197,]  1.236345e-02
#> [198,]  4.939561e-02
#> [199,] -4.344370e-02
#> [200,] -9.394440e-03
#> [201,]  4.912787e-02
#> [202,] -4.043181e-02
#> [203,]  3.419088e-02
#> [204,] -1.169354e-02
#> [205,] -4.626618e-02
#> [206,] -3.318116e-03
#> [207,]  2.328794e-02
#> [208,]  4.669672e-02
#> [209,]  2.495377e-02
#> [210,] -5.950069e-03
#> [211,] -1.381599e-02
#> [212,] -2.012377e-03
#> [213,]  1.258721e-02
#> [214,] -9.598364e-03
#> [215,]  4.607161e-02
#> [216,]  2.270988e-02
#> [217,] -4.813962e-03
#> [218,]  9.119736e-03
#> [219,]  2.944444e-02
#> [220,]  2.456639e-02
#> [221,] -4.529509e-02
#> [222,] -2.311637e-02
#> [223,]  5.023058e-03
#> [224,] -3.129312e-03
#> [225,] -4.966105e-02
#> [226,]  2.742952e-02
#> [227,]  2.814299e-02
#> [228,] -1.144606e-02
#> [229,]  2.385839e-02
#> [230,] -2.328518e-02
#> [231,]  3.888834e-02
#> [232,]  4.935520e-02
#> [233,] -1.136107e-02
#> [234,] -2.956465e-02
#> [235,]  2.663928e-02
#> [236,] -7.001049e-03
#> [237,] -3.502445e-02
#> [238,]  4.638341e-02
#> [239,] -8.875826e-03
#> [240,] -1.909150e-02
#> [241,] -2.509672e-02
#> [242,] -2.549262e-02
#> [243,] -6.047804e-03
#> [244,] -8.516249e-03
#> [245,] -1.383191e-02
#> [246,]  1.101680e-02
#> [247,]  1.950497e-02
#> [248,]  4.275377e-02
#> [249,]  3.603305e-02
#> [250,] -1.601203e-02
#> [251,]  1.363236e-02
#> [252,] -4.683588e-02
#> [253,] -3.658550e-02
#> [254,]  1.396039e-02
#> [255,]  1.997473e-04
#> [256,] -1.836233e-02
#> [257,]  4.769237e-02
#> [258,]  1.193674e-02
#> [259,] -3.639576e-03
#> [260,] -1.515541e-03
#> [261,]  4.812398e-02
#> [262,] -3.799880e-02
#> [263,]  2.515106e-02
#> [264,] -2.724168e-02
#> [265,]  9.466326e-04
#> [266,] -4.397889e-02
#> [267,]  5.014475e-04
#> [268,]  3.450598e-02
#> [269,]  3.062576e-02
#> [270,]  1.187620e-03
#> [271,] -3.318955e-02
#> [272,]  2.794614e-02
#> [273,] -7.510905e-03
#> [274,]  1.218251e-02
#> [275,] -7.201036e-03
#> [276,] -1.307523e-05
#> [277,]  4.692097e-02
#> [278,] -1.965580e-04
#> [279,] -4.994702e-02
#> [280,] -4.027079e-02
#> [281,]  9.515750e-03
#> [282,] -2.197708e-02
#> [283,]  6.981247e-03
#> [284,] -1.029408e-02
#> [285,] -4.469373e-02
#> [286,] -2.557657e-02
#> [287,] -2.223585e-02
#> [288,]  2.098280e-02
#> [289,]  7.045902e-03
#> [290,]  2.968307e-02
#> [291,]  3.181108e-02
#> [292,] -1.244016e-02
#> [293,] -4.979910e-02
#> [294,] -9.040796e-03
#> [295,]  5.213414e-03
#> [296,]  7.803112e-03
#> [297,] -3.603631e-02
#> [298,]  4.211404e-02
#> [299,]  2.500165e-03
#> [300,]  5.457975e-04
#> [301,] -2.746613e-02
#> [302,] -4.331419e-02
#> [303,] -4.005978e-02
#> [304,] -4.454726e-02
#> [305,] -3.259852e-02
#> [306,]  3.880369e-04
#> [307,]  4.398361e-02
#> [308,] -2.571239e-02
#> [309,]  1.923221e-02
#> [310,] -8.796513e-03
#> [311,] -5.718624e-03
#> [312,]  1.084470e-02
#> [313,]  1.021095e-02
#> [314,]  4.339419e-02
#> [315,] -4.008293e-02
#> [316,]  2.889140e-02
#> [317,]  2.987334e-02
#> [318,] -2.989474e-02
#> [319,] -4.381740e-02
#> [320,]  4.939072e-02
#> [321,]  4.731177e-02
#> [322,] -3.254360e-02
#> [323,] -1.611407e-02
#> [324,] -2.801099e-02
#> [325,]  3.132470e-02
#> [326,] -7.054869e-03
#> [327,] -4.187926e-03
#> [328,]  1.849088e-02
#> [329,]  1.937804e-03
#> [330,] -4.470994e-02
#> [331,] -4.851268e-02
#> [332,] -2.326234e-04
#> [333,]  3.823711e-02
#> [334,] -4.976897e-03
#> [335,] -2.133923e-02
#> [336,]  4.496145e-02
#> [337,]  3.674765e-02
#> [338,]  1.252075e-02
#> [339,] -3.867749e-02
#> [340,] -2.194174e-02
#> [341,] -2.253080e-02
#> [342,]  3.580010e-02
#> [343,] -2.823668e-02
#> [344,] -2.676787e-02
#> [345,] -4.473307e-02
#> [346,] -2.708332e-02
#> [347,] -1.060460e-02
#> [348,] -2.498430e-02
#> [349,]  2.483848e-02
#> [350,]  4.976586e-02
#> [351,] -4.784005e-02
#> [352,]  5.039727e-03
#> [353,]  1.559569e-02
#> [354,]  3.096275e-02
#> [355,]  4.929284e-02
#> [356,] -1.355089e-02
#> [357,]  2.673950e-03
#> [358,] -1.012933e-02
#> [359,]  4.926663e-02
#> [360,]  1.428520e-02
#> [361,]  3.659567e-03
#> [362,]  8.614360e-03
#> [363,] -2.518553e-02
#> [364,]  1.825244e-02
#> [365,] -2.649119e-02
#> [366,]  2.219956e-02
#> [367,]  1.086444e-03
#> [368,] -1.029547e-02
#> [369,] -3.483957e-02
#> [370,] -2.000913e-04
#> [371,]  4.624560e-03
#> [372,]  3.341464e-03
#> [373,] -2.913179e-03
#> [374,]  1.018064e-02
#> [375,] -2.572617e-02
#> [376,] -5.993936e-03
#> [377,] -3.731198e-02
#> [378,] -3.149270e-02
#> [379,] -1.771785e-02
#> [380,] -2.811470e-02
#> [381,]  3.086541e-02
#> [382,]  3.452644e-02
#> [383,] -4.012871e-02
#> [384,] -3.470827e-03
#> [385,] -1.834466e-02
#> [386,]  1.536887e-03
#> [387,] -3.316502e-02
#> [388,] -4.083264e-04
#> [389,]  2.668378e-02
#> [390,] -9.410109e-03
#> [391,]  2.745830e-02
#> [392,]  9.995364e-03
#> [393,]  3.900574e-02
#> [394,]  1.150508e-02
#> [395,]  3.301401e-02
#> [396,]  3.034526e-03
#> [397,]  4.283202e-03
#> [398,]  1.745348e-02
#> [399,]  3.768712e-02
#> [400,] -7.121745e-04