Generate Coordinates for a Sphere

This function generates the coordinates for a sphere in three-dimensional space.

Usage

sphere(radius, resolution, num_noise, min_n, max_n)

Arguments

radius

The radius of the sphere.

resolution

The number of points used to approximate the surface of the sphere.

num_noise

The number of additional noise dimensions to add to the coordinates.

min_n

The minimum value for the random noise added to the coordinates.

max_n

The maximum value for the random noise added to the coordinates.

Value

A matrix containing the Cartesian coordinates of the points on the sphere.

Examples

# Generate coordinates for a sphere with radius 1 and resolution 20
set.seed(20240412)
sphere(radius = 1, resolution = 20, num_noise = 3, min_n = -0.05,
max_n = 0.05)
#>                 [,1]          [,2]        [,3]          [,4]          [,5]
#>   [1,]  0.000000e+00  0.000000e+00  1.00000000 -0.0406208903  1.458899e-02
#>   [2,]  0.000000e+00  0.000000e+00  1.00000000 -0.0069582161 -2.865623e-02
#>   [3,]  0.000000e+00  0.000000e+00  1.00000000  0.0239614908 -2.609781e-02
#>   [4,]  0.000000e+00  0.000000e+00  1.00000000  0.0143368074 -4.110739e-02
#>   [5,]  0.000000e+00  0.000000e+00  1.00000000  0.0401831183 -4.965122e-02
#>   [6,]  0.000000e+00  0.000000e+00  1.00000000 -0.0308282462 -3.251858e-02
#>   [7,]  0.000000e+00  0.000000e+00  1.00000000 -0.0400570993  2.281042e-02
#>   [8,]  0.000000e+00  0.000000e+00  1.00000000  0.0034441012  4.688578e-02
#>   [9,]  0.000000e+00  0.000000e+00  1.00000000 -0.0147099345 -3.585914e-02
#>  [10,]  0.000000e+00  0.000000e+00  1.00000000 -0.0058951876  4.046627e-02
#>  [11,]  0.000000e+00  0.000000e+00  1.00000000  0.0061251915 -1.089824e-02
#>  [12,]  0.000000e+00  0.000000e+00  1.00000000  0.0091990083 -3.413554e-02
#>  [13,]  0.000000e+00  0.000000e+00  1.00000000  0.0437254337 -3.089505e-02
#>  [14,]  0.000000e+00  0.000000e+00  1.00000000  0.0327753831  4.024651e-02
#>  [15,]  0.000000e+00  0.000000e+00  1.00000000 -0.0151984228 -2.819042e-02
#>  [16,]  0.000000e+00  0.000000e+00  1.00000000 -0.0098404858 -3.771454e-02
#>  [17,]  0.000000e+00  0.000000e+00  1.00000000  0.0178843632 -4.630252e-03
#>  [18,]  0.000000e+00  0.000000e+00  1.00000000  0.0031283391 -4.492376e-02
#>  [19,]  0.000000e+00  0.000000e+00  1.00000000 -0.0498555169  1.361151e-02
#>  [20,]  0.000000e+00  0.000000e+00  1.00000000 -0.0425508382 -2.761497e-02
#>  [21,]  1.645946e-01  0.000000e+00  0.98636130 -0.0141356953  7.016805e-03
#>  [22,]  1.556764e-01  5.344378e-02  0.98636130 -0.0399193364  7.230506e-03
#>  [23,]  1.298883e-01  1.010961e-01  0.98636130  0.0008688590  1.352021e-02
#>  [24,]  9.002471e-02  1.377931e-01  0.98636130  0.0469179725 -4.457019e-02
#>  [25,]  4.040558e-02  1.595580e-01  0.98636130 -0.0100181968 -1.310034e-03
#>  [26,] -1.359211e-02  1.640324e-01  0.98636130 -0.0411767531 -9.513795e-03
#>  [27,] -6.611689e-02  1.507313e-01  0.98636130  0.0028839635  3.003845e-02
#>  [28,] -1.114769e-01  1.210962e-01  0.98636130  0.0433840647  6.382821e-03
#>  [29,] -1.447566e-01  7.833837e-02  0.98636130  0.0194841265 -2.444980e-02
#>  [30,] -1.623497e-01  2.709138e-02  0.98636130 -0.0157354954 -2.677100e-02
#>  [31,] -1.623497e-01 -2.709138e-02  0.98636130 -0.0148068666  4.704865e-02
#>  [32,] -1.447566e-01 -7.833837e-02  0.98636130  0.0349434115 -2.961780e-02
#>  [33,] -1.114769e-01 -1.210962e-01  0.98636130  0.0122341669  3.370171e-02
#>  [34,] -6.611689e-02 -1.507313e-01  0.98636130  0.0403223421  1.654683e-02
#>  [35,] -1.359211e-02 -1.640324e-01  0.98636130 -0.0133424654  1.833225e-02
#>  [36,]  4.040558e-02 -1.595580e-01  0.98636130  0.0487433413 -3.866709e-03
#>  [37,]  9.002471e-02 -1.377931e-01  0.98636130 -0.0158348253 -4.063287e-02
#>  [38,]  1.298883e-01 -1.010961e-01  0.98636130  0.0289856836  3.401228e-02
#>  [39,]  1.556764e-01 -5.344378e-02  0.98636130 -0.0336490937  4.455656e-03
#>  [40,]  1.645946e-01 -4.031405e-17  0.98636130  0.0347434584  2.337754e-02
#>  [41,]  3.246995e-01  0.000000e+00  0.94581724 -0.0409471453  6.600727e-03
#>  [42,]  3.071064e-01  1.054297e-01  0.94581724  0.0256599021  7.430371e-03
#>  [43,]  2.562335e-01  1.994345e-01  0.94581724  0.0340622700 -4.363294e-02
#>  [44,]  1.775938e-01  2.718275e-01  0.94581724 -0.0019275298  4.190935e-02
#>  [45,]  7.970901e-02  3.147638e-01  0.94581724  0.0060101570  3.425425e-02
#>  [46,] -2.681347e-02  3.235905e-01  0.94581724  0.0051576530 -3.860183e-02
#>  [47,] -1.304303e-01  2.973511e-01  0.94581724  0.0407851192 -7.435258e-03
#>  [48,] -2.199130e-01  2.388892e-01  0.94581724 -0.0085904858  4.040416e-02
#>  [49,] -2.855647e-01  1.545399e-01  0.94581724 -0.0135948996 -4.276738e-02
#>  [50,] -3.202710e-01  5.344378e-02  0.94581724 -0.0081947034  2.994692e-02
#>  [51,] -3.202710e-01 -5.344378e-02  0.94581724 -0.0145145064 -4.456985e-02
#>  [52,] -2.855647e-01 -1.545399e-01  0.94581724 -0.0426104567 -1.082111e-02
#>  [53,] -2.199130e-01 -2.388892e-01  0.94581724 -0.0043946581 -1.241861e-02
#>  [54,] -1.304303e-01 -2.973511e-01  0.94581724  0.0314007990 -1.631652e-02
#>  [55,] -2.681347e-02 -3.235905e-01  0.94581724 -0.0053969306  3.040836e-03
#>  [56,]  7.970901e-02 -3.147638e-01  0.94581724 -0.0418217706 -4.027878e-02
#>  [57,]  1.775938e-01 -2.718275e-01  0.94581724  0.0310165984 -9.351035e-03
#>  [58,]  2.562335e-01 -1.994345e-01  0.94581724 -0.0383762233 -1.717753e-02
#>  [59,]  3.071064e-01 -1.054297e-01  0.94581724 -0.0395184289 -7.135615e-03
#>  [60,]  3.246995e-01 -7.952843e-17  0.94581724 -0.0488566692 -3.335375e-02
#>  [61,]  4.759474e-01  0.000000e+00  0.87947375  0.0077576239  3.431115e-03
#>  [62,]  4.501593e-01  1.545399e-01  0.87947375  0.0353640246  1.866506e-02
#>  [63,]  3.755894e-01  2.923329e-01  0.87947375  0.0314732646 -2.138503e-02
#>  [64,]  2.603185e-01  3.984472e-01  0.87947375  0.0346473052  4.667769e-02
#>  [65,]  1.168382e-01  4.613835e-01  0.87947375 -0.0141555009 -3.253958e-02
#>  [66,] -3.930342e-02  4.743218e-01  0.87947375 -0.0132494227  2.804652e-02
#>  [67,] -1.911859e-01  4.358599e-01  0.87947375 -0.0247900630  9.009592e-03
#>  [68,] -3.223504e-01  3.501659e-01  0.87947375 -0.0442713907 -2.988130e-02
#>  [69,] -4.185832e-01  2.265259e-01  0.87947375 -0.0106764714 -7.058739e-03
#>  [70,] -4.694561e-01  7.833837e-02  0.87947375  0.0438516971  1.708714e-02
#>  [71,] -4.694561e-01 -7.833837e-02  0.87947375  0.0164436805  2.566097e-02
#>  [72,] -4.185832e-01 -2.265259e-01  0.87947375 -0.0215097323  3.205063e-02
#>  [73,] -3.223504e-01 -3.501659e-01  0.87947375  0.0085810695  2.594132e-02
#>  [74,] -1.911859e-01 -4.358599e-01  0.87947375 -0.0252457969 -3.803132e-02
#>  [75,] -3.930342e-02 -4.743218e-01  0.87947375 -0.0341419302 -4.379304e-02
#>  [76,]  1.168382e-01 -4.613835e-01  0.87947375 -0.0403578342  1.434154e-02
#>  [77,]  2.603185e-01 -3.984472e-01  0.87947375 -0.0207715337  4.449161e-03
#>  [78,]  3.755894e-01 -2.923329e-01  0.87947375  0.0403564284  7.503827e-03
#>  [79,]  4.501593e-01 -1.545399e-01  0.87947375  0.0309667719  4.656778e-02
#>  [80,]  4.759474e-01 -1.165735e-16  0.87947375  0.0291593320  2.710841e-02
#>  [81,]  6.142127e-01  0.000000e+00  0.78914051  0.0235926253 -2.600811e-02
#>  [82,]  5.809330e-01  1.994345e-01  0.78914051  0.0388510874  3.284313e-02
#>  [83,]  4.847001e-01  3.772573e-01  0.78914051  0.0394336176  4.479017e-02
#>  [84,]  3.359425e-01  5.141983e-01  0.78914051 -0.0248468031  3.177405e-02
#>  [85,]  1.507803e-01  5.954180e-01  0.78914051 -0.0471686648 -1.214109e-02
#>  [86,] -5.072128e-02  6.121149e-01  0.78914051 -0.0356369173  2.545728e-03
#>  [87,] -2.467264e-01  5.624796e-01  0.78914051 -0.0397483398  3.369623e-02
#>  [88,] -4.159950e-01  4.518910e-01  0.78914051  0.0323190081  3.942831e-02
#>  [89,] -5.401840e-01  2.923329e-01  0.78914051  0.0397418948 -1.072001e-02
#>  [90,] -6.058357e-01  1.010961e-01  0.78914051 -0.0141661465  4.184796e-02
#>  [91,] -6.058357e-01 -1.010961e-01  0.78914051 -0.0290247731  4.625760e-02
#>  [92,] -5.401840e-01 -2.923329e-01  0.78914051 -0.0025848112  2.384936e-02
#>  [93,] -4.159950e-01 -4.518910e-01  0.78914051  0.0070531712  2.029594e-02
#>  [94,] -2.467264e-01 -5.624796e-01  0.78914051  0.0039477463  3.089043e-05
#>  [95,] -5.072128e-02 -6.121149e-01  0.78914051 -0.0364015279  2.164460e-02
#>  [96,]  1.507803e-01 -5.954180e-01  0.78914051  0.0314509333 -7.280862e-03
#>  [97,]  3.359425e-01 -5.141983e-01  0.78914051 -0.0129274398  5.159603e-03
#>  [98,]  4.847001e-01 -3.772573e-01  0.78914051  0.0155807688 -4.571919e-02
#>  [99,]  5.809330e-01 -1.994345e-01  0.78914051 -0.0259864262 -4.371184e-02
#> [100,]  6.142127e-01 -1.504387e-16  0.78914051  0.0386686355  8.465952e-03
#> [101,]  7.357239e-01  0.000000e+00  0.67728157 -0.0494111540 -9.797570e-03
#> [102,]  6.958604e-01  2.388892e-01  0.67728157 -0.0095421034  2.422641e-02
#> [103,]  5.805895e-01  4.518910e-01  0.67728157 -0.0095423958 -3.277202e-02
#> [104,]  4.024028e-01  6.159234e-01  0.67728157 -0.0056648134  9.997107e-03
#> [105,]  1.806095e-01  7.132110e-01  0.67728157 -0.0361641475  3.069320e-02
#> [106,] -6.075560e-02  7.332110e-01  0.67728157 -0.0137938427 -1.476614e-02
#> [107,] -2.955369e-01  6.737563e-01  0.67728157  0.0227676261  1.839283e-02
#> [108,] -4.982922e-01  5.412897e-01  0.67728157  0.0486252397 -2.645821e-02
#> [109,] -6.470499e-01  3.501659e-01  0.67728157 -0.0374798555 -1.278195e-02
#> [110,] -7.256896e-01  1.210962e-01  0.67728157 -0.0060145900  4.227686e-02
#> [111,] -7.256896e-01 -1.210962e-01  0.67728157  0.0122375009  9.606105e-03
#> [112,] -6.470499e-01 -3.501659e-01  0.67728157 -0.0151320067  2.830282e-02
#> [113,] -4.982922e-01 -5.412897e-01  0.67728157 -0.0069745122  4.181241e-02
#> [114,] -2.955369e-01 -6.737563e-01  0.67728157  0.0119806817 -1.311135e-02
#> [115,] -6.075560e-02 -7.332110e-01  0.67728157  0.0425165671 -1.880957e-02
#> [116,]  1.806095e-01 -7.132110e-01  0.67728157 -0.0198855986 -1.712295e-02
#> [117,]  4.024028e-01 -6.159234e-01  0.67728157  0.0341541310 -7.884755e-03
#> [118,]  5.805895e-01 -4.518910e-01  0.67728157  0.0139019090 -4.200629e-02
#> [119,]  6.958604e-01 -2.388892e-01  0.67728157  0.0043720318 -2.877043e-02
#> [120,]  7.357239e-01 -1.802004e-16  0.67728157 -0.0362611223 -3.228527e-02
#> [121,]  8.371665e-01  0.000000e+00  0.54694816 -0.0091411981 -2.788319e-02
#> [122,]  7.918065e-01  2.718275e-01  0.54694816  0.0375122381  1.741983e-02
#> [123,]  6.606420e-01  5.141983e-01  0.54694816 -0.0205436346 -5.182130e-03
#> [124,]  4.578867e-01  7.008477e-01  0.54694816  0.0096704798  2.974713e-02
#> [125,]  2.055122e-01  8.115494e-01  0.54694816 -0.0074246488 -2.027644e-02
#> [126,] -6.913266e-02  8.343071e-01  0.54694816 -0.0395226781 -6.945097e-03
#> [127,] -3.362859e-01  7.666547e-01  0.54694816  0.0140160184  2.858230e-03
#> [128,] -5.669974e-01  6.159234e-01  0.54694816  0.0001349537 -4.814338e-02
#> [129,] -7.362659e-01  3.984472e-01  0.54694816  0.0395929786  1.638557e-02
#> [130,] -8.257486e-01  1.377931e-01  0.54694816 -0.0421812224  4.239687e-02
#> [131,] -8.257486e-01 -1.377931e-01  0.54694816 -0.0192060802 -2.032662e-02
#> [132,] -7.362659e-01 -3.984472e-01  0.54694816  0.0329963851 -4.524067e-03
#> [133,] -5.669974e-01 -6.159234e-01  0.54694816 -0.0441801412 -1.712918e-02
#> [134,] -3.362859e-01 -7.666547e-01  0.54694816  0.0359260298 -4.000003e-02
#> [135,] -6.913266e-02 -8.343071e-01  0.54694816 -0.0215506210  3.129462e-02
#> [136,]  2.055122e-01 -8.115494e-01  0.54694816 -0.0231140854  1.696542e-02
#> [137,]  4.578867e-01 -7.008477e-01  0.54694816 -0.0290469873  2.599024e-02
#> [138,]  6.606420e-01 -5.141983e-01  0.54694816  0.0346599014 -2.685868e-02
#> [139,]  7.918065e-01 -2.718275e-01  0.54694816 -0.0245407152 -3.175927e-02
#> [140,]  8.371665e-01 -2.050466e-16  0.54694816  0.0438232428  2.017552e-02
#> [141,]  9.157733e-01  0.000000e+00  0.40169542  0.0190370204 -1.430562e-02
#> [142,]  8.661542e-01  2.973511e-01  0.40169542 -0.0157330000  4.388218e-02
#> [143,]  7.226738e-01  5.624796e-01  0.40169542  0.0136145061 -4.704672e-02
#> [144,]  5.008805e-01  7.666547e-01  0.40169542 -0.0211817049  4.432870e-03
#> [145,]  2.248091e-01  8.877509e-01  0.40169542 -0.0296529576  3.661782e-02
#> [146,] -7.562396e-02  9.126455e-01  0.40169542 -0.0385897858 -8.528756e-03
#> [147,] -3.678620e-01  8.386408e-01  0.40169542 -0.0114352393  7.837044e-03
#> [148,] -6.202364e-01  6.737563e-01  0.40169542  0.0394679684  3.644674e-02
#> [149,] -8.053986e-01  4.358599e-01  0.40169542 -0.0379028531 -3.336327e-02
#> [150,] -9.032834e-01  1.507313e-01  0.40169542  0.0086608654  1.907975e-02
#> [151,] -9.032834e-01 -1.507313e-01  0.40169542  0.0397832549  1.471168e-03
#> [152,] -8.053986e-01 -4.358599e-01  0.40169542 -0.0389684510  1.342903e-02
#> [153,] -6.202364e-01 -6.737563e-01  0.40169542 -0.0107839450 -3.520064e-02
#> [154,] -3.678620e-01 -8.386408e-01  0.40169542  0.0229901668  2.381260e-02
#> [155,] -7.562396e-02 -9.126455e-01  0.40169542 -0.0032016000  2.661871e-02
#> [156,]  2.248091e-01 -8.877509e-01  0.40169542 -0.0188916723 -3.555762e-02
#> [157,]  5.008805e-01 -7.666547e-01  0.40169542 -0.0406173033 -3.595247e-02
#> [158,]  7.226738e-01 -5.624796e-01  0.40169542  0.0218278701  1.364102e-02
#> [159,]  8.661542e-01 -2.973511e-01  0.40169542 -0.0178925602 -4.875386e-02
#> [160,]  9.157733e-01 -2.242998e-16  0.40169542  0.0418970450 -2.493871e-02
#> [161,]  9.694003e-01  0.000000e+00  0.24548549 -0.0289520095  4.896922e-03
#> [162,]  9.168755e-01  3.147638e-01  0.24548549  0.0182308341  3.970057e-02
#> [163,]  7.649930e-01  5.954180e-01  0.24548549  0.0420673159 -8.724967e-03
#> [164,]  5.302117e-01  8.115494e-01  0.24548549 -0.0215416480 -1.597159e-02
#> [165,]  2.379737e-01  9.397369e-01  0.24548549 -0.0032879685  2.019766e-02
#> [166,] -8.005244e-02  9.660893e-01  0.24548549  0.0136172979 -9.139048e-03
#> [167,] -3.894037e-01  8.877509e-01  0.24548549 -0.0128606171 -3.028368e-02
#> [168,] -6.565569e-01  7.132110e-01  0.24548549  0.0406865353 -1.788580e-02
#> [169,] -8.525621e-01  4.613835e-01  0.24548549 -0.0056567913  4.290069e-02
#> [170,] -9.561789e-01  1.595580e-01  0.24548549  0.0206802091  2.202454e-02
#> [171,] -9.561789e-01 -1.595580e-01  0.24548549  0.0293639539 -6.666061e-03
#> [172,] -8.525621e-01 -4.613835e-01  0.24548549  0.0099463716 -4.398104e-02
#> [173,] -6.565569e-01 -7.132110e-01  0.24548549  0.0216768384 -2.984306e-02
#> [174,] -3.894037e-01 -8.877509e-01  0.24548549  0.0126717287 -4.668091e-03
#> [175,] -8.005244e-02 -9.660893e-01  0.24548549  0.0137911398  1.147877e-02
#> [176,]  2.379737e-01 -9.397369e-01  0.24548549 -0.0201560635  2.931105e-02
#> [177,]  5.302117e-01 -8.115494e-01  0.24548549 -0.0474466317  3.496157e-02
#> [178,]  7.649930e-01 -5.954180e-01  0.24548549  0.0034467781 -5.097572e-04
#> [179,]  9.168755e-01 -3.147638e-01  0.24548549 -0.0451321898  1.107564e-02
#> [180,]  9.694003e-01 -2.374346e-16  0.24548549  0.0449525188 -3.444004e-02
#> [181,]  9.965845e-01  0.000000e+00  0.08257935  0.0301143471 -1.721028e-03
#> [182,]  9.425868e-01  3.235905e-01  0.08257935  0.0015380326 -5.106072e-03
#> [183,]  7.864452e-01  6.121149e-01  0.08257935  0.0313923658  9.471303e-03
#> [184,]  5.450801e-01  8.343071e-01  0.08257935  0.0142799772 -2.284583e-02
#> [185,]  2.446470e-01  9.660893e-01  0.08257935 -0.0342363236  3.711272e-02
#> [186,] -8.229730e-02  9.931807e-01  0.08257935 -0.0437694831 -4.431580e-02
#> [187,] -4.003234e-01  9.126455e-01  0.08257935  0.0163590239 -1.383906e-02
#> [188,] -6.749683e-01  7.332110e-01  0.08257935  0.0215670073  3.990231e-02
#> [189,] -8.764699e-01  4.743218e-01  0.08257935  0.0312772650 -1.393104e-02
#> [190,] -9.829924e-01  1.640324e-01  0.08257935  0.0404797181  4.668946e-03
#> [191,] -9.829924e-01 -1.640324e-01  0.08257935  0.0197967365 -1.337010e-02
#> [192,] -8.764699e-01 -4.743218e-01  0.08257935 -0.0082434738  3.080920e-02
#> [193,] -6.749683e-01 -7.332110e-01  0.08257935  0.0110442976 -1.190440e-02
#> [194,] -4.003234e-01 -9.126455e-01  0.08257935  0.0086618911 -2.176235e-02
#> [195,] -8.229730e-02 -9.931807e-01  0.08257935  0.0378016111 -3.808644e-03
#> [196,]  2.446470e-01 -9.660893e-01  0.08257935  0.0267662517  2.022534e-02
#> [197,]  5.450801e-01 -8.343071e-01  0.08257935  0.0202962862  3.642897e-03
#> [198,]  7.864452e-01 -6.121149e-01  0.08257935 -0.0326327926 -5.561239e-03
#> [199,]  9.425868e-01 -3.235905e-01  0.08257935 -0.0398299208 -2.585751e-02
#> [200,]  9.965845e-01 -2.440928e-16  0.08257935  0.0465680600  1.580620e-02
#> [201,]  9.965845e-01  0.000000e+00 -0.08257935  0.0468112729 -1.216545e-02
#> [202,]  9.425868e-01  3.235905e-01 -0.08257935 -0.0124962581  2.438039e-02
#> [203,]  7.864452e-01  6.121149e-01 -0.08257935 -0.0390363060 -3.692567e-02
#> [204,]  5.450801e-01  8.343071e-01 -0.08257935  0.0121142437  1.229900e-02
#> [205,]  2.446470e-01  9.660893e-01 -0.08257935 -0.0077770323 -4.542733e-02
#> [206,] -8.229730e-02  9.931807e-01 -0.08257935 -0.0433116644 -1.067751e-02
#> [207,] -4.003234e-01  9.126455e-01 -0.08257935  0.0096181576  3.728084e-02
#> [208,] -6.749683e-01  7.332110e-01 -0.08257935  0.0404465381  4.634579e-02
#> [209,] -8.764699e-01  4.743218e-01 -0.08257935 -0.0497450735  3.505162e-02
#> [210,] -9.829924e-01  1.640324e-01 -0.08257935 -0.0152653055 -4.885832e-02
#> [211,] -9.829924e-01 -1.640324e-01 -0.08257935  0.0365711479 -4.737772e-02
#> [212,] -8.764699e-01 -4.743218e-01 -0.08257935  0.0159191196  1.044061e-02
#> [213,] -6.749683e-01 -7.332110e-01 -0.08257935 -0.0314435686  4.555401e-02
#> [214,] -4.003234e-01 -9.126455e-01 -0.08257935  0.0449404774  1.465659e-02
#> [215,] -8.229730e-02 -9.931807e-01 -0.08257935 -0.0289388946 -1.518818e-02
#> [216,]  2.446470e-01 -9.660893e-01 -0.08257935  0.0135954321 -9.794076e-03
#> [217,]  5.450801e-01 -8.343071e-01 -0.08257935  0.0023689911 -2.090481e-03
#> [218,]  7.864452e-01 -6.121149e-01 -0.08257935 -0.0305722723 -1.672968e-02
#> [219,]  9.425868e-01 -3.235905e-01 -0.08257935 -0.0184369768  1.053083e-02
#> [220,]  9.965845e-01 -2.440928e-16 -0.08257935  0.0459533107 -2.569068e-02
#> [221,]  9.694003e-01  0.000000e+00 -0.24548549  0.0348061180 -1.951988e-02
#> [222,]  9.168755e-01  3.147638e-01 -0.24548549  0.0042005939 -2.975835e-02
#> [223,]  7.649930e-01  5.954180e-01 -0.24548549 -0.0309299566 -1.109935e-02
#> [224,]  5.302117e-01  8.115494e-01 -0.24548549 -0.0233263229  1.491552e-02
#> [225,]  2.379737e-01  9.397369e-01 -0.24548549  0.0388561925 -1.618895e-02
#> [226,] -8.005244e-02  9.660893e-01 -0.24548549  0.0360052565 -4.736626e-02
#> [227,] -3.894037e-01  8.877509e-01 -0.24548549 -0.0055477833  2.745094e-02
#> [228,] -6.565569e-01  7.132110e-01 -0.24548549  0.0107978261  4.158819e-03
#> [229,] -8.525621e-01  4.613835e-01 -0.24548549  0.0296354921 -3.366253e-02
#> [230,] -9.561789e-01  1.595580e-01 -0.24548549  0.0099678486  3.562754e-02
#> [231,] -9.561789e-01 -1.595580e-01 -0.24548549 -0.0064570792  2.016836e-02
#> [232,] -8.525621e-01 -4.613835e-01 -0.24548549 -0.0309545504  3.974189e-03
#> [233,] -6.565569e-01 -7.132110e-01 -0.24548549 -0.0250889844  1.890066e-02
#> [234,] -3.894037e-01 -8.877509e-01 -0.24548549  0.0319340720  3.394537e-02
#> [235,] -8.005244e-02 -9.660893e-01 -0.24548549 -0.0354034658 -3.954843e-02
#> [236,]  2.379737e-01 -9.397369e-01 -0.24548549 -0.0428596250  3.515784e-02
#> [237,]  5.302117e-01 -8.115494e-01 -0.24548549 -0.0344285349  1.184551e-02
#> [238,]  7.649930e-01 -5.954180e-01 -0.24548549  0.0341629610 -2.892848e-03
#> [239,]  9.168755e-01 -3.147638e-01 -0.24548549 -0.0454066350 -4.160459e-02
#> [240,]  9.694003e-01 -2.374346e-16 -0.24548549  0.0122605025  1.335511e-02
#> [241,]  9.157733e-01  0.000000e+00 -0.40169542  0.0022526822 -3.629767e-02
#> [242,]  8.661542e-01  2.973511e-01 -0.40169542  0.0048651400 -3.141726e-02
#> [243,]  7.226738e-01  5.624796e-01 -0.40169542  0.0371045142 -1.774895e-02
#> [244,]  5.008805e-01  7.666547e-01 -0.40169542  0.0191890407  8.120899e-03
#> [245,]  2.248091e-01  8.877509e-01 -0.40169542  0.0169471913 -1.586619e-02
#> [246,] -7.562396e-02  9.126455e-01 -0.40169542  0.0234745285 -9.185939e-03
#> [247,] -3.678620e-01  8.386408e-01 -0.40169542 -0.0464605047  2.909539e-02
#> [248,] -6.202364e-01  6.737563e-01 -0.40169542 -0.0140471952 -2.053609e-02
#> [249,] -8.053986e-01  4.358599e-01 -0.40169542 -0.0027744060  8.063634e-03
#> [250,] -9.032834e-01  1.507313e-01 -0.40169542  0.0219781587 -1.553382e-02
#> [251,] -9.032834e-01 -1.507313e-01 -0.40169542  0.0014598926 -2.602863e-02
#> [252,] -8.053986e-01 -4.358599e-01 -0.40169542 -0.0326802443  4.460973e-02
#> [253,] -6.202364e-01 -6.737563e-01 -0.40169542 -0.0069526496 -4.520893e-02
#> [254,] -3.678620e-01 -8.386408e-01 -0.40169542  0.0490568970 -2.327186e-03
#> [255,] -7.562396e-02 -9.126455e-01 -0.40169542 -0.0374291364  3.853751e-02
#> [256,]  2.248091e-01 -8.877509e-01 -0.40169542  0.0401866821  7.380333e-03
#> [257,]  5.008805e-01 -7.666547e-01 -0.40169542  0.0378239532  2.232958e-02
#> [258,]  7.226738e-01 -5.624796e-01 -0.40169542 -0.0138014233  6.065673e-03
#> [259,]  8.661542e-01 -2.973511e-01 -0.40169542  0.0141808003  1.259044e-02
#> [260,]  9.157733e-01 -2.242998e-16 -0.40169542  0.0161106778 -4.332718e-02
#> [261,]  8.371665e-01  0.000000e+00 -0.54694816  0.0456993452 -2.363263e-02
#> [262,]  7.918065e-01  2.718275e-01 -0.54694816 -0.0172811388  2.158810e-02
#> [263,]  6.606420e-01  5.141983e-01 -0.54694816 -0.0222209299  1.234432e-03
#> [264,]  4.578867e-01  7.008477e-01 -0.54694816  0.0488013647 -3.786748e-02
#> [265,]  2.055122e-01  8.115494e-01 -0.54694816  0.0074586672  5.989409e-03
#> [266,] -6.913266e-02  8.343071e-01 -0.54694816 -0.0334256714 -4.697274e-02
#> [267,] -3.362859e-01  7.666547e-01 -0.54694816 -0.0355602939  4.611238e-02
#> [268,] -5.669974e-01  6.159234e-01 -0.54694816  0.0466993265 -5.639224e-04
#> [269,] -7.362659e-01  3.984472e-01 -0.54694816  0.0212277140  3.845493e-02
#> [270,] -8.257486e-01  1.377931e-01 -0.54694816  0.0287382015  8.162953e-03
#> [271,] -8.257486e-01 -1.377931e-01 -0.54694816 -0.0488130579 -4.392665e-02
#> [272,] -7.362659e-01 -3.984472e-01 -0.54694816 -0.0254148313 -2.709454e-02
#> [273,] -5.669974e-01 -6.159234e-01 -0.54694816 -0.0449607738  1.836716e-02
#> [274,] -3.362859e-01 -7.666547e-01 -0.54694816 -0.0329215669 -6.381873e-03
#> [275,] -6.913266e-02 -8.343071e-01 -0.54694816  0.0040266522  1.990309e-02
#> [276,]  2.055122e-01 -8.115494e-01 -0.54694816  0.0142846406 -2.730781e-02
#> [277,]  4.578867e-01 -7.008477e-01 -0.54694816 -0.0329637585 -4.226218e-02
#> [278,]  6.606420e-01 -5.141983e-01 -0.54694816  0.0494644662 -4.664999e-02
#> [279,]  7.918065e-01 -2.718275e-01 -0.54694816 -0.0360697758 -2.116798e-02
#> [280,]  8.371665e-01 -2.050466e-16 -0.54694816  0.0202892990  4.309416e-02
#> [281,]  7.357239e-01  0.000000e+00 -0.67728157  0.0366330932  1.920677e-02
#> [282,]  6.958604e-01  2.388892e-01 -0.67728157 -0.0299145314  2.242898e-02
#> [283,]  5.805895e-01  4.518910e-01 -0.67728157  0.0356086889 -4.191774e-02
#> [284,]  4.024028e-01  6.159234e-01 -0.67728157  0.0307684459 -3.674067e-03
#> [285,]  1.806095e-01  7.132110e-01 -0.67728157  0.0226531942  4.443457e-02
#> [286,] -6.075560e-02  7.332110e-01 -0.67728157  0.0148863554  1.860181e-02
#> [287,] -2.955369e-01  6.737563e-01 -0.67728157 -0.0034758566 -2.175452e-02
#> [288,] -4.982922e-01  5.412897e-01 -0.67728157  0.0215064450  2.180254e-02
#> [289,] -6.470499e-01  3.501659e-01 -0.67728157  0.0319374689 -4.092405e-02
#> [290,] -7.256896e-01  1.210962e-01 -0.67728157 -0.0371603101  1.156094e-02
#> [291,] -7.256896e-01 -1.210962e-01 -0.67728157  0.0498017874 -4.391131e-02
#> [292,] -6.470499e-01 -3.501659e-01 -0.67728157  0.0127720742 -3.421685e-02
#> [293,] -4.982922e-01 -5.412897e-01 -0.67728157  0.0058777411 -4.220810e-02
#> [294,] -2.955369e-01 -6.737563e-01 -0.67728157  0.0257612543  4.022145e-02
#> [295,] -6.075560e-02 -7.332110e-01 -0.67728157  0.0273422876 -4.296581e-02
#> [296,]  1.806095e-01 -7.132110e-01 -0.67728157 -0.0353921735  4.524026e-02
#> [297,]  4.024028e-01 -6.159234e-01 -0.67728157  0.0180608628  2.511152e-02
#> [298,]  5.805895e-01 -4.518910e-01 -0.67728157 -0.0044484145  2.906693e-02
#> [299,]  6.958604e-01 -2.388892e-01 -0.67728157 -0.0471473908 -2.489058e-03
#> [300,]  7.357239e-01 -1.802004e-16 -0.67728157 -0.0179879901 -1.890632e-02
#> [301,]  6.142127e-01  0.000000e+00 -0.78914051  0.0397026339 -2.927220e-02
#> [302,]  5.809330e-01  1.994345e-01 -0.78914051 -0.0290763175 -4.092765e-02
#> [303,]  4.847001e-01  3.772573e-01 -0.78914051 -0.0422360181 -3.048572e-02
#> [304,]  3.359425e-01  5.141983e-01 -0.78914051  0.0214310022 -3.307067e-02
#> [305,]  1.507803e-01  5.954180e-01 -0.78914051  0.0499433291 -3.749385e-02
#> [306,] -5.072128e-02  6.121149e-01 -0.78914051  0.0356851794  3.411267e-02
#> [307,] -2.467264e-01  5.624796e-01 -0.78914051 -0.0015637302  3.205123e-02
#> [308,] -4.159950e-01  4.518910e-01 -0.78914051  0.0123396746 -3.857820e-02
#> [309,] -5.401840e-01  2.923329e-01 -0.78914051  0.0284339315  2.676152e-02
#> [310,] -6.058357e-01  1.010961e-01 -0.78914051  0.0350829806 -2.005807e-02
#> [311,] -6.058357e-01 -1.010961e-01 -0.78914051  0.0353439124  3.963579e-02
#> [312,] -5.401840e-01 -2.923329e-01 -0.78914051 -0.0423142502  4.464460e-02
#> [313,] -4.159950e-01 -4.518910e-01 -0.78914051  0.0051962465 -3.259540e-02
#> [314,] -2.467264e-01 -5.624796e-01 -0.78914051  0.0389469245 -4.315758e-02
#> [315,] -5.072128e-02 -6.121149e-01 -0.78914051  0.0086575911  4.323491e-02
#> [316,]  1.507803e-01 -5.954180e-01 -0.78914051 -0.0232040592  1.448548e-02
#> [317,]  3.359425e-01 -5.141983e-01 -0.78914051  0.0245691278 -2.217300e-02
#> [318,]  4.847001e-01 -3.772573e-01 -0.78914051  0.0315889175  1.027064e-02
#> [319,]  5.809330e-01 -1.994345e-01 -0.78914051 -0.0227344875 -4.496821e-02
#> [320,]  6.142127e-01 -1.504387e-16 -0.78914051  0.0366494890 -2.655916e-02
#> [321,]  4.759474e-01  0.000000e+00 -0.87947375 -0.0157848667  8.054416e-03
#> [322,]  4.501593e-01  1.545399e-01 -0.87947375 -0.0070135517  4.392374e-02
#> [323,]  3.755894e-01  2.923329e-01 -0.87947375 -0.0208199039  2.359525e-02
#> [324,]  2.603185e-01  3.984472e-01 -0.87947375  0.0189224581  1.092390e-02
#> [325,]  1.168382e-01  4.613835e-01 -0.87947375 -0.0338216087 -4.630102e-02
#> [326,] -3.930342e-02  4.743218e-01 -0.87947375  0.0460992394 -2.662198e-02
#> [327,] -1.911859e-01  4.358599e-01 -0.87947375 -0.0447378346  4.263518e-03
#> [328,] -3.223504e-01  3.501659e-01 -0.87947375 -0.0272497393  2.611584e-02
#> [329,] -4.185832e-01  2.265259e-01 -0.87947375  0.0160971409 -4.355451e-02
#> [330,] -4.694561e-01  7.833837e-02 -0.87947375 -0.0464448324  1.124803e-02
#> [331,] -4.694561e-01 -7.833837e-02 -0.87947375 -0.0179349351  2.867090e-02
#> [332,] -4.185832e-01 -2.265259e-01 -0.87947375  0.0207382537 -1.485437e-02
#> [333,] -3.223504e-01 -3.501659e-01 -0.87947375 -0.0037443788 -4.124495e-02
#> [334,] -1.911859e-01 -4.358599e-01 -0.87947375  0.0118583800  3.849647e-02
#> [335,] -3.930342e-02 -4.743218e-01 -0.87947375 -0.0063740385  4.035356e-02
#> [336,]  1.168382e-01 -4.613835e-01 -0.87947375  0.0297865495 -4.438101e-02
#> [337,]  2.603185e-01 -3.984472e-01 -0.87947375  0.0433482581 -1.914029e-02
#> [338,]  3.755894e-01 -2.923329e-01 -0.87947375  0.0163532625  8.827151e-03
#> [339,]  4.501593e-01 -1.545399e-01 -0.87947375 -0.0359919606 -6.906394e-03
#> [340,]  4.759474e-01 -1.165735e-16 -0.87947375  0.0370901504  2.500496e-02
#> [341,]  3.246995e-01  0.000000e+00 -0.94581724  0.0496382012 -3.486415e-02
#> [342,]  3.071064e-01  1.054297e-01 -0.94581724 -0.0114451926 -3.207366e-02
#> [343,]  2.562335e-01  1.994345e-01 -0.94581724  0.0157686300 -1.476420e-02
#> [344,]  1.775938e-01  2.718275e-01 -0.94581724  0.0482725061  1.487132e-03
#> [345,]  7.970901e-02  3.147638e-01 -0.94581724 -0.0417446218  3.114883e-02
#> [346,] -2.681347e-02  3.235905e-01 -0.94581724 -0.0490760001 -3.930624e-02
#> [347,] -1.304303e-01  2.973511e-01 -0.94581724  0.0143416769  1.292091e-02
#> [348,] -2.199130e-01  2.388892e-01 -0.94581724  0.0389693632  3.438110e-02
#> [349,] -2.855647e-01  1.545399e-01 -0.94581724  0.0050521928 -3.673330e-02
#> [350,] -3.202710e-01  5.344378e-02 -0.94581724  0.0338490285  7.785790e-03
#> [351,] -3.202710e-01 -5.344378e-02 -0.94581724 -0.0007207744 -3.775987e-02
#> [352,] -2.855647e-01 -1.545399e-01 -0.94581724 -0.0358561296  2.817780e-02
#> [353,] -2.199130e-01 -2.388892e-01 -0.94581724  0.0343432767 -1.064854e-02
#> [354,] -1.304303e-01 -2.973511e-01 -0.94581724 -0.0380912595 -4.422468e-02
#> [355,] -2.681347e-02 -3.235905e-01 -0.94581724  0.0017651984 -4.409485e-03
#> [356,]  7.970901e-02 -3.147638e-01 -0.94581724 -0.0271230818 -2.765841e-02
#> [357,]  1.775938e-01 -2.718275e-01 -0.94581724  0.0153674109 -6.641598e-03
#> [358,]  2.562335e-01 -1.994345e-01 -0.94581724  0.0435791666  2.512449e-02
#> [359,]  3.071064e-01 -1.054297e-01 -0.94581724  0.0007038843  4.955630e-02
#> [360,]  3.246995e-01 -7.952843e-17 -0.94581724  0.0042624283  3.146839e-03
#> [361,]  1.645946e-01  0.000000e+00 -0.98636130 -0.0307058573 -9.831291e-03
#> [362,]  1.556764e-01  5.344378e-02 -0.98636130  0.0172181919  2.714389e-02
#> [363,]  1.298883e-01  1.010961e-01 -0.98636130 -0.0344564827  4.002474e-02
#> [364,]  9.002471e-02  1.377931e-01 -0.98636130  0.0145525717 -3.417797e-02
#> [365,]  4.040558e-02  1.595580e-01 -0.98636130 -0.0024064121 -5.691283e-03
#> [366,] -1.359211e-02  1.640324e-01 -0.98636130 -0.0470545179 -2.159807e-02
#> [367,] -6.611689e-02  1.507313e-01 -0.98636130  0.0429631366  8.071225e-03
#> [368,] -1.114769e-01  1.210962e-01 -0.98636130 -0.0314480813 -4.391317e-03
#> [369,] -1.447566e-01  7.833837e-02 -0.98636130  0.0469257004  4.771512e-03
#> [370,] -1.623497e-01  2.709138e-02 -0.98636130 -0.0067218661 -1.587071e-02
#> [371,] -1.623497e-01 -2.709138e-02 -0.98636130  0.0365070718 -3.635973e-02
#> [372,] -1.447566e-01 -7.833837e-02 -0.98636130 -0.0148092618  1.133370e-02
#> [373,] -1.114769e-01 -1.210962e-01 -0.98636130  0.0288158743 -2.288790e-02
#> [374,] -6.611689e-02 -1.507313e-01 -0.98636130  0.0393909458  6.661088e-03
#> [375,] -1.359211e-02 -1.640324e-01 -0.98636130  0.0436995254  3.594665e-02
#> [376,]  4.040558e-02 -1.595580e-01 -0.98636130  0.0465217978  3.459761e-02
#> [377,]  9.002471e-02 -1.377931e-01 -0.98636130  0.0291392233  5.795360e-03
#> [378,]  1.298883e-01 -1.010961e-01 -0.98636130 -0.0114881480 -1.656536e-02
#> [379,]  1.556764e-01 -5.344378e-02 -0.98636130  0.0194828836 -1.015206e-02
#> [380,]  1.645946e-01 -4.031405e-17 -0.98636130  0.0485348954  4.870286e-02
#> [381,]  1.224647e-16  0.000000e+00 -1.00000000  0.0097199444  1.152507e-02
#> [382,]  1.158292e-16  3.976422e-17 -1.00000000  0.0044783109  4.444306e-02
#> [383,]  9.664184e-17  7.521936e-17 -1.00000000 -0.0366461525  8.271111e-03
#> [384,]  6.698183e-17  1.025233e-16 -1.00000000  0.0395055116 -3.604335e-02
#> [385,]  3.006330e-17  1.187173e-16 -1.00000000  0.0053874930 -4.192450e-02
#> [386,] -1.011305e-17  1.220464e-16 -1.00000000 -0.0058034629 -3.463579e-02
#> [387,] -4.919350e-17  1.121499e-16 -1.00000000 -0.0060975001 -5.629439e-03
#> [388,] -8.294307e-17  9.010019e-17 -1.00000000  0.0234173421 -3.736544e-03
#> [389,] -1.077045e-16  5.828675e-17 -1.00000000 -0.0187736356 -1.467698e-02
#> [390,] -1.207944e-16  2.015702e-17 -1.00000000  0.0442129100 -2.188860e-03
#> [391,] -1.207944e-16 -2.015702e-17 -1.00000000  0.0419374371 -2.250117e-03
#> [392,] -1.077045e-16 -5.828675e-17 -1.00000000  0.0498900986  8.365604e-03
#> [393,] -8.294307e-17 -9.010019e-17 -1.00000000 -0.0109741065  2.348013e-02
#> [394,] -4.919350e-17 -1.121499e-16 -1.00000000  0.0074751925  1.505425e-02
#> [395,] -1.011305e-17 -1.220464e-16 -1.00000000  0.0305212979  1.848556e-02
#> [396,]  3.006330e-17 -1.187173e-16 -1.00000000  0.0172342707 -3.956056e-02
#> [397,]  6.698183e-17 -1.025233e-16 -1.00000000  0.0040695532  3.476416e-02
#> [398,]  9.664184e-17 -7.521936e-17 -1.00000000 -0.0392953042 -4.016556e-02
#> [399,]  1.158292e-16 -3.976422e-17 -1.00000000 -0.0337261188  3.392490e-02
#> [400,]  1.224647e-16 -2.999520e-32 -1.00000000 -0.0190098491 -9.393090e-03
#>                 [,6]
#>   [1,]  1.393803e-02
#>   [2,]  2.225547e-02
#>   [3,]  2.738212e-02
#>   [4,] -3.073480e-02
#>   [5,]  2.495993e-02
#>   [6,]  4.024293e-02
#>   [7,] -4.621974e-02
#>   [8,]  3.932236e-02
#>   [9,] -2.399021e-02
#>  [10,]  4.876298e-02
#>  [11,]  5.570777e-03
#>  [12,] -3.719361e-02
#>  [13,] -4.364063e-02
#>  [14,] -3.907752e-02
#>  [15,] -2.274678e-02
#>  [16,]  2.764111e-02
#>  [17,] -4.468815e-02
#>  [18,] -1.634977e-03
#>  [19,]  2.401393e-02
#>  [20,] -1.812660e-02
#>  [21,]  4.922323e-02
#>  [22,]  3.299203e-02
#>  [23,] -6.474367e-03
#>  [24,] -4.223610e-02
#>  [25,]  3.448482e-02
#>  [26,] -3.052665e-02
#>  [27,] -2.311675e-02
#>  [28,]  1.942860e-02
#>  [29,]  1.718890e-04
#>  [30,] -3.404807e-02
#>  [31,]  2.371948e-02
#>  [32,] -4.342598e-04
#>  [33,]  4.823518e-02
#>  [34,]  1.964920e-02
#>  [35,] -1.948357e-02
#>  [36,]  1.592498e-02
#>  [37,] -2.456669e-02
#>  [38,]  4.787787e-02
#>  [39,] -2.193019e-03
#>  [40,] -4.167884e-02
#>  [41,]  2.316105e-02
#>  [42,]  2.623009e-02
#>  [43,] -3.703178e-02
#>  [44,] -3.463169e-02
#>  [45,]  4.101954e-02
#>  [46,]  1.660521e-02
#>  [47,] -3.055208e-02
#>  [48,] -2.456755e-02
#>  [49,] -3.454551e-02
#>  [50,] -3.508595e-04
#>  [51,] -3.405702e-03
#>  [52,] -1.186254e-03
#>  [53,] -1.096035e-02
#>  [54,]  2.121938e-02
#>  [55,] -2.175540e-02
#>  [56,] -1.917075e-02
#>  [57,]  4.423345e-02
#>  [58,] -2.195092e-02
#>  [59,] -4.953112e-02
#>  [60,]  1.138406e-02
#>  [61,]  1.550899e-02
#>  [62,] -4.010219e-02
#>  [63,]  1.634947e-02
#>  [64,] -4.253979e-02
#>  [65,]  4.808981e-02
#>  [66,]  3.081084e-02
#>  [67,] -7.190514e-03
#>  [68,]  4.127902e-02
#>  [69,] -2.404599e-02
#>  [70,] -1.933766e-02
#>  [71,]  4.239964e-02
#>  [72,] -1.849773e-02
#>  [73,] -1.989845e-02
#>  [74,] -1.169297e-02
#>  [75,]  2.322755e-02
#>  [76,] -1.927021e-02
#>  [77,] -3.201269e-02
#>  [78,] -8.195556e-03
#>  [79,] -3.135465e-02
#>  [80,]  2.442750e-02
#>  [81,] -1.517884e-02
#>  [82,] -3.514090e-02
#>  [83,]  2.150824e-02
#>  [84,]  2.728485e-03
#>  [85,]  3.871202e-02
#>  [86,]  2.278382e-02
#>  [87,]  2.512407e-02
#>  [88,] -4.649326e-02
#>  [89,]  1.995078e-02
#>  [90,] -2.565149e-02
#>  [91,] -4.058669e-02
#>  [92,]  2.873166e-03
#>  [93,]  4.442756e-02
#>  [94,] -4.911115e-02
#>  [95,] -4.165463e-02
#>  [96,] -3.419861e-02
#>  [97,] -8.630894e-03
#>  [98,]  1.077493e-02
#>  [99,]  1.501543e-02
#> [100,]  3.249849e-02
#> [101,] -3.576007e-02
#> [102,] -1.850679e-02
#> [103,]  2.710531e-02
#> [104,]  2.235682e-02
#> [105,]  3.600592e-02
#> [106,] -1.070241e-02
#> [107,]  1.818586e-02
#> [108,]  2.815670e-02
#> [109,] -3.828808e-03
#> [110,] -4.094278e-02
#> [111,]  3.250933e-02
#> [112,] -2.053256e-02
#> [113,] -3.559308e-02
#> [114,]  4.878065e-02
#> [115,]  4.481919e-02
#> [116,] -4.397552e-02
#> [117,]  3.610309e-02
#> [118,] -2.131414e-02
#> [119,]  3.636798e-03
#> [120,]  3.402189e-02
#> [121,]  3.049437e-02
#> [122,] -8.520503e-03
#> [123,]  3.672661e-02
#> [124,]  2.642170e-02
#> [125,]  9.144743e-03
#> [126,] -5.949574e-03
#> [127,]  2.816194e-02
#> [128,] -1.726960e-03
#> [129,]  2.569834e-02
#> [130,]  4.967059e-02
#> [131,] -5.135249e-03
#> [132,]  2.735784e-02
#> [133,]  1.159155e-02
#> [134,] -2.640640e-02
#> [135,] -1.985293e-02
#> [136,]  2.438598e-04
#> [137,]  5.309682e-03
#> [138,]  3.622295e-02
#> [139,]  1.938526e-02
#> [140,] -9.325156e-03
#> [141,]  2.487195e-02
#> [142,] -4.730527e-02
#> [143,] -2.389774e-02
#> [144,]  2.647772e-02
#> [145,]  1.890590e-02
#> [146,]  3.497542e-02
#> [147,] -4.571855e-02
#> [148,] -1.080535e-02
#> [149,]  4.276782e-02
#> [150,] -1.553904e-02
#> [151,]  1.026275e-02
#> [152,]  2.543897e-02
#> [153,]  1.780995e-02
#> [154,]  1.053365e-02
#> [155,]  2.969802e-02
#> [156,]  1.318234e-02
#> [157,]  2.211034e-02
#> [158,]  1.949560e-02
#> [159,] -3.431024e-02
#> [160,]  3.547910e-02
#> [161,]  4.555309e-02
#> [162,] -4.446336e-02
#> [163,]  2.781911e-02
#> [164,]  1.002056e-02
#> [165,] -2.079151e-02
#> [166,] -1.940213e-02
#> [167,]  4.202857e-02
#> [168,] -1.437366e-02
#> [169,] -4.711666e-02
#> [170,] -1.110588e-03
#> [171,] -3.239017e-02
#> [172,] -1.339229e-02
#> [173,] -4.645296e-02
#> [174,] -1.192731e-02
#> [175,] -2.116185e-02
#> [176,] -1.631644e-02
#> [177,] -1.546603e-02
#> [178,]  6.941579e-04
#> [179,] -2.659954e-02
#> [180,]  1.646715e-02
#> [181,] -4.339491e-02
#> [182,]  2.145939e-02
#> [183,] -5.977730e-03
#> [184,]  3.318530e-02
#> [185,]  1.249937e-02
#> [186,] -6.530065e-03
#> [187,]  4.114174e-02
#> [188,]  2.478450e-02
#> [189,] -4.352686e-02
#> [190,]  4.589231e-02
#> [191,] -4.617009e-03
#> [192,] -4.507646e-02
#> [193,] -5.940151e-03
#> [194,]  8.188367e-03
#> [195,]  3.046013e-02
#> [196,]  2.082963e-03
#> [197,]  3.047816e-02
#> [198,] -4.257658e-02
#> [199,]  1.141599e-02
#> [200,]  2.337407e-02
#> [201,] -3.390966e-02
#> [202,]  4.143774e-02
#> [203,]  3.741529e-02
#> [204,] -4.039443e-02
#> [205,] -2.845354e-02
#> [206,] -1.434015e-02
#> [207,] -1.508975e-02
#> [208,]  2.946663e-02
#> [209,]  2.687273e-02
#> [210,] -2.450101e-02
#> [211,] -3.236095e-02
#> [212,]  2.834071e-02
#> [213,] -1.017609e-02
#> [214,]  4.951010e-03
#> [215,]  9.329137e-03
#> [216,] -9.235702e-03
#> [217,] -1.424448e-02
#> [218,] -2.765936e-02
#> [219,]  2.357522e-02
#> [220,]  9.247106e-04
#> [221,] -2.226533e-02
#> [222,] -1.087636e-02
#> [223,]  4.905825e-03
#> [224,]  2.455052e-02
#> [225,]  1.645280e-03
#> [226,]  2.338030e-02
#> [227,]  4.603940e-02
#> [228,] -3.328674e-02
#> [229,] -4.064574e-02
#> [230,]  1.298184e-02
#> [231,]  1.381245e-03
#> [232,]  9.914179e-03
#> [233,]  2.530526e-03
#> [234,]  2.729615e-02
#> [235,] -4.977361e-02
#> [236,]  3.738483e-02
#> [237,]  1.781895e-02
#> [238,] -2.909667e-02
#> [239,] -3.765714e-02
#> [240,]  1.222253e-02
#> [241,]  2.075209e-02
#> [242,] -4.898746e-02
#> [243,]  3.192730e-02
#> [244,] -2.347345e-02
#> [245,] -3.927989e-02
#> [246,] -4.121712e-02
#> [247,]  3.928497e-02
#> [248,] -3.702536e-02
#> [249,] -1.453956e-02
#> [250,] -1.511670e-02
#> [251,]  2.085086e-02
#> [252,] -2.631276e-02
#> [253,] -3.924439e-02
#> [254,] -3.962608e-02
#> [255,] -3.601795e-02
#> [256,]  2.132409e-02
#> [257,]  4.799574e-02
#> [258,]  3.955065e-02
#> [259,]  3.353678e-02
#> [260,] -3.749058e-03
#> [261,] -2.561923e-02
#> [262,]  4.328054e-02
#> [263,]  2.657243e-02
#> [264,]  2.482937e-03
#> [265,] -2.067488e-02
#> [266,]  2.909830e-02
#> [267,] -5.261940e-03
#> [268,] -1.225310e-03
#> [269,]  2.573979e-02
#> [270,]  4.920211e-02
#> [271,] -3.828112e-03
#> [272,]  4.840133e-02
#> [273,] -1.622421e-02
#> [274,] -4.705459e-02
#> [275,]  5.583441e-03
#> [276,]  1.856189e-02
#> [277,]  3.931910e-02
#> [278,]  2.130964e-02
#> [279,] -2.342741e-02
#> [280,] -2.933642e-02
#> [281,]  2.671146e-02
#> [282,]  2.874434e-03
#> [283,]  4.335716e-02
#> [284,] -3.560783e-02
#> [285,]  4.759485e-02
#> [286,]  2.414846e-02
#> [287,]  3.752585e-02
#> [288,]  4.680808e-02
#> [289,] -3.778814e-02
#> [290,] -4.903727e-02
#> [291,] -1.696616e-03
#> [292,] -1.757738e-02
#> [293,]  1.058579e-02
#> [294,] -3.779881e-02
#> [295,] -9.174128e-03
#> [296,]  4.528655e-02
#> [297,]  2.627506e-02
#> [298,] -2.480230e-02
#> [299,] -1.863678e-02
#> [300,] -4.428247e-02
#> [301,] -3.309084e-02
#> [302,] -6.205570e-04
#> [303,] -1.274938e-02
#> [304,] -2.556577e-02
#> [305,]  2.704463e-03
#> [306,]  1.715678e-02
#> [307,] -2.553369e-02
#> [308,]  1.342221e-02
#> [309,]  2.135242e-02
#> [310,]  4.435638e-02
#> [311,]  3.480150e-02
#> [312,] -7.342764e-03
#> [313,]  7.514273e-03
#> [314,] -1.997895e-02
#> [315,] -2.137183e-02
#> [316,]  8.705977e-03
#> [317,] -1.731751e-02
#> [318,]  4.477979e-02
#> [319,] -4.190433e-02
#> [320,] -2.007988e-02
#> [321,]  3.115944e-02
#> [322,]  2.308626e-02
#> [323,] -3.330740e-02
#> [324,] -4.415502e-02
#> [325,]  3.834558e-02
#> [326,] -2.309174e-02
#> [327,]  8.338904e-03
#> [328,] -3.812055e-02
#> [329,] -3.733654e-05
#> [330,]  4.724280e-02
#> [331,]  1.894036e-02
#> [332,] -7.934875e-03
#> [333,] -2.471494e-02
#> [334,]  9.015889e-03
#> [335,]  2.119069e-03
#> [336,] -2.862516e-02
#> [337,]  4.937218e-02
#> [338,] -1.189991e-03
#> [339,]  1.052345e-02
#> [340,] -1.419472e-02
#> [341,]  3.895420e-02
#> [342,] -1.607776e-02
#> [343,] -3.519951e-02
#> [344,]  4.306235e-02
#> [345,]  2.677924e-02
#> [346,] -3.370115e-02
#> [347,]  7.628256e-03
#> [348,] -1.474725e-03
#> [349,] -2.119039e-02
#> [350,] -1.268381e-02
#> [351,] -1.346007e-02
#> [352,]  1.548746e-02
#> [353,]  2.307818e-02
#> [354,] -4.217972e-02
#> [355,] -4.829483e-02
#> [356,] -2.072083e-02
#> [357,] -3.700059e-02
#> [358,]  2.682958e-02
#> [359,] -5.010260e-03
#> [360,] -2.138814e-02
#> [361,] -3.384213e-02
#> [362,]  2.381642e-02
#> [363,]  3.741723e-02
#> [364,]  2.485847e-02
#> [365,]  1.459736e-03
#> [366,]  4.685443e-03
#> [367,]  4.021027e-02
#> [368,] -4.312391e-02
#> [369,] -1.495112e-02
#> [370,]  1.122955e-02
#> [371,] -4.721409e-02
#> [372,]  4.704446e-02
#> [373,] -6.842718e-03
#> [374,] -1.286648e-02
#> [375,]  3.359787e-02
#> [376,]  4.484048e-02
#> [377,]  4.241089e-02
#> [378,]  3.371412e-02
#> [379,]  1.659019e-02
#> [380,]  2.000044e-02
#> [381,] -1.690339e-02
#> [382,]  9.961860e-03
#> [383,]  2.888342e-02
#> [384,]  2.930204e-02
#> [385,]  4.338725e-02
#> [386,] -1.025363e-03
#> [387,] -6.947431e-03
#> [388,]  3.107307e-02
#> [389,]  2.983059e-02
#> [390,]  3.587464e-03
#> [391,] -3.662836e-02
#> [392,]  2.557130e-02
#> [393,]  4.820601e-02
#> [394,]  4.970583e-02
#> [395,] -1.212139e-02
#> [396,] -3.294325e-02
#> [397,] -1.258265e-02
#> [398,]  1.569915e-04
#> [399,]  1.541384e-02
#> [400,] -3.356151e-02