Generate Coordinates for a Sphere

This function generates the coordinates for a sphere in three-dimensional space.

Usage

sphere(radius, resolution, num_noise_dims, min_noise, max_noise)

Arguments

radius

The radius of the sphere.

resolution

The number of points used to approximate the surface of the sphere.

num_noise_dims

The number of additional noise dimensions to add to the coordinates.

min_noise

The minimum value for the random noise added to the coordinates.

max_noise

The maximum value for the random noise added to the coordinates.

Value

A matrix containing the Cartesian coordinates of the points on the sphere.

Examples

# Generate coordinates for a sphere with radius 1 and resolution 100
sphere(radius = 1, resolution = 20, num_noise_dims = 3, min_noise = -0.05,
max_noise = 0.05)
#>                 [,1]          [,2]        [,3]          [,4]          [,5]
#>   [1,]  0.000000e+00  0.000000e+00  1.00000000 -2.099807e-02  0.0280797721
#>   [2,]  0.000000e+00  0.000000e+00  1.00000000  1.613195e-02 -0.0121924175
#>   [3,]  0.000000e+00  0.000000e+00  1.00000000 -3.913759e-02 -0.0330291233
#>   [4,]  0.000000e+00  0.000000e+00  1.00000000 -4.824815e-02 -0.0417178093
#>   [5,]  0.000000e+00  0.000000e+00  1.00000000  3.391568e-02 -0.0480593848
#>   [6,]  0.000000e+00  0.000000e+00  1.00000000  3.214650e-02 -0.0453294148
#>   [7,]  0.000000e+00  0.000000e+00  1.00000000  3.760383e-02  0.0111828555
#>   [8,]  0.000000e+00  0.000000e+00  1.00000000 -1.995349e-02  0.0488918754
#>   [9,]  0.000000e+00  0.000000e+00  1.00000000 -1.495153e-02  0.0225352554
#>  [10,]  0.000000e+00  0.000000e+00  1.00000000 -1.867661e-03  0.0142654950
#>  [11,]  0.000000e+00  0.000000e+00  1.00000000 -3.196475e-02  0.0354762347
#>  [12,]  0.000000e+00  0.000000e+00  1.00000000  3.527836e-02 -0.0366188484
#>  [13,]  0.000000e+00  0.000000e+00  1.00000000  1.439069e-02  0.0448334731
#>  [14,]  0.000000e+00  0.000000e+00  1.00000000  4.063819e-02 -0.0409425216
#>  [15,]  0.000000e+00  0.000000e+00  1.00000000 -1.781973e-02  0.0004098609
#>  [16,]  0.000000e+00  0.000000e+00  1.00000000 -9.991877e-03  0.0466561069
#>  [17,]  0.000000e+00  0.000000e+00  1.00000000 -2.113383e-02 -0.0227826453
#>  [18,]  0.000000e+00  0.000000e+00  1.00000000  3.586941e-02  0.0292143598
#>  [19,]  0.000000e+00  0.000000e+00  1.00000000  2.959686e-02 -0.0174030380
#>  [20,]  0.000000e+00  0.000000e+00  1.00000000 -4.514771e-03  0.0270248232
#>  [21,]  1.645946e-01  0.000000e+00  0.98636130  5.790182e-03 -0.0245905064
#>  [22,]  1.556764e-01  5.344378e-02  0.98636130 -3.248872e-02  0.0374791581
#>  [23,]  1.298883e-01  1.010961e-01  0.98636130  4.831964e-02  0.0378769194
#>  [24,]  9.002471e-02  1.377931e-01  0.98636130 -1.425967e-02 -0.0288486092
#>  [25,]  4.040558e-02  1.595580e-01  0.98636130  1.630627e-02  0.0053544863
#>  [26,] -1.359211e-02  1.640324e-01  0.98636130 -1.170086e-02  0.0025114908
#>  [27,] -6.611689e-02  1.507313e-01  0.98636130 -1.013226e-02  0.0179689034
#>  [28,] -1.114769e-01  1.210962e-01  0.98636130  7.622374e-03  0.0287334990
#>  [29,] -1.447566e-01  7.833837e-02  0.98636130  7.560358e-03 -0.0135804157
#>  [30,] -1.623497e-01  2.709138e-02  0.98636130  4.796044e-02  0.0285090832
#>  [31,] -1.623497e-01 -2.709138e-02  0.98636130  1.840153e-02  0.0022062201
#>  [32,] -1.447566e-01 -7.833837e-02  0.98636130  4.024499e-02  0.0363343889
#>  [33,] -1.114769e-01 -1.210962e-01  0.98636130 -3.349629e-02  0.0110172283
#>  [34,] -6.611689e-02 -1.507313e-01  0.98636130 -1.285116e-02  0.0350364899
#>  [35,] -1.359211e-02 -1.640324e-01  0.98636130 -2.579287e-02  0.0099672439
#>  [36,]  4.040558e-02 -1.595580e-01  0.98636130 -4.225003e-02  0.0360482751
#>  [37,]  9.002471e-02 -1.377931e-01  0.98636130  4.538793e-02 -0.0305348763
#>  [38,]  1.298883e-01 -1.010961e-01  0.98636130  1.727609e-02  0.0032009949
#>  [39,]  1.556764e-01 -5.344378e-02  0.98636130 -1.929411e-02  0.0005195178
#>  [40,]  1.645946e-01 -4.031405e-17  0.98636130 -9.397555e-04 -0.0378959184
#>  [41,]  3.246995e-01  0.000000e+00  0.94581724 -3.038556e-02 -0.0028261860
#>  [42,]  3.071064e-01  1.054297e-01  0.94581724  2.764673e-02  0.0431825544
#>  [43,]  2.562335e-01  1.994345e-01  0.94581724 -1.372276e-02  0.0265758651
#>  [44,]  1.775938e-01  2.718275e-01  0.94581724  1.545207e-02  0.0248394041
#>  [45,]  7.970901e-02  3.147638e-01  0.94581724  2.456389e-02 -0.0138229773
#>  [46,] -2.681347e-02  3.235905e-01  0.94581724 -2.235940e-03 -0.0135592397
#>  [47,] -1.304303e-01  2.973511e-01  0.94581724 -2.622800e-02  0.0344451946
#>  [48,] -2.199130e-01  2.388892e-01  0.94581724 -2.280766e-02 -0.0084846190
#>  [49,] -2.855647e-01  1.545399e-01  0.94581724 -1.940838e-02 -0.0394349312
#>  [50,] -3.202710e-01  5.344378e-02  0.94581724 -4.755792e-02 -0.0480947239
#>  [51,] -3.202710e-01 -5.344378e-02  0.94581724 -2.520550e-02  0.0137550354
#>  [52,] -2.855647e-01 -1.545399e-01  0.94581724  3.518140e-02 -0.0354823546
#>  [53,] -2.199130e-01 -2.388892e-01  0.94581724 -1.064635e-02  0.0251701997
#>  [54,] -1.304303e-01 -2.973511e-01  0.94581724 -1.871709e-02  0.0090644483
#>  [55,] -2.681347e-02 -3.235905e-01  0.94581724 -1.241584e-03  0.0441642432
#>  [56,]  7.970901e-02 -3.147638e-01  0.94581724 -3.620774e-03  0.0274463836
#>  [57,]  1.775938e-01 -2.718275e-01  0.94581724  2.281883e-02 -0.0398637896
#>  [58,]  2.562335e-01 -1.994345e-01  0.94581724 -2.209607e-02 -0.0241649840
#>  [59,]  3.071064e-01 -1.054297e-01  0.94581724 -4.388664e-02  0.0363729670
#>  [60,]  3.246995e-01 -7.952843e-17  0.94581724 -4.281530e-02  0.0350432127
#>  [61,]  4.759474e-01  0.000000e+00  0.87947375 -3.661060e-02  0.0313538057
#>  [62,]  4.501593e-01  1.545399e-01  0.87947375 -2.223650e-02  0.0418471035
#>  [63,]  3.755894e-01  2.923329e-01  0.87947375 -4.125096e-02  0.0279612498
#>  [64,]  2.603185e-01  3.984472e-01  0.87947375 -4.688179e-02  0.0237663522
#>  [65,]  1.168382e-01  4.613835e-01  0.87947375 -2.623198e-02  0.0020727467
#>  [66,] -3.930342e-02  4.743218e-01  0.87947375 -4.070112e-02 -0.0223322747
#>  [67,] -1.911859e-01  4.358599e-01  0.87947375  4.412034e-02 -0.0498677238
#>  [68,] -3.223504e-01  3.501659e-01  0.87947375  1.414885e-03 -0.0217301269
#>  [69,] -4.185832e-01  2.265259e-01  0.87947375 -1.894294e-02  0.0013216135
#>  [70,] -4.694561e-01  7.833837e-02  0.87947375 -3.722722e-02  0.0313574036
#>  [71,] -4.694561e-01 -7.833837e-02  0.87947375  6.057399e-03 -0.0030314952
#>  [72,] -4.185832e-01 -2.265259e-01  0.87947375 -1.173406e-02 -0.0113520778
#>  [73,] -3.223504e-01 -3.501659e-01  0.87947375  8.456683e-03  0.0259533292
#>  [74,] -1.911859e-01 -4.358599e-01  0.87947375 -3.458896e-03  0.0498913975
#>  [75,] -3.930342e-02 -4.743218e-01  0.87947375  1.265525e-02  0.0379498681
#>  [76,]  1.168382e-01 -4.613835e-01  0.87947375 -1.022387e-02  0.0459555196
#>  [77,]  2.603185e-01 -3.984472e-01  0.87947375 -2.665104e-02  0.0001875577
#>  [78,]  3.755894e-01 -2.923329e-01  0.87947375  3.598179e-02 -0.0050635456
#>  [79,]  4.501593e-01 -1.545399e-01  0.87947375 -4.704363e-02  0.0130623515
#>  [80,]  4.759474e-01 -1.165735e-16  0.87947375  1.688001e-02 -0.0168525227
#>  [81,]  6.142127e-01  0.000000e+00  0.78914051 -1.821828e-02  0.0421744693
#>  [82,]  5.809330e-01  1.994345e-01  0.78914051 -2.292980e-02  0.0359327714
#>  [83,]  4.847001e-01  3.772573e-01  0.78914051 -3.586433e-02 -0.0209681712
#>  [84,]  3.359425e-01  5.141983e-01  0.78914051  1.105405e-02 -0.0187252526
#>  [85,]  1.507803e-01  5.954180e-01  0.78914051  2.475721e-02 -0.0430619818
#>  [86,] -5.072128e-02  6.121149e-01  0.78914051  1.364248e-02 -0.0134724254
#>  [87,] -2.467264e-01  5.624796e-01  0.78914051 -3.149454e-02  0.0459288234
#>  [88,] -4.159950e-01  4.518910e-01  0.78914051  2.705457e-02 -0.0251172174
#>  [89,] -5.401840e-01  2.923329e-01  0.78914051  2.094090e-02 -0.0123699499
#>  [90,] -6.058357e-01  1.010961e-01  0.78914051  3.442369e-02 -0.0424742981
#>  [91,] -6.058357e-01 -1.010961e-01  0.78914051 -3.187770e-02 -0.0275251617
#>  [92,] -5.401840e-01 -2.923329e-01  0.78914051  4.541255e-02 -0.0035650667
#>  [93,] -4.159950e-01 -4.518910e-01  0.78914051 -4.690885e-02 -0.0429790955
#>  [94,] -2.467264e-01 -5.624796e-01  0.78914051  2.711215e-02  0.0390591999
#>  [95,] -5.072128e-02 -6.121149e-01  0.78914051 -2.472351e-02  0.0489502125
#>  [96,]  1.507803e-01 -5.954180e-01  0.78914051  1.218859e-02  0.0404855302
#>  [97,]  3.359425e-01 -5.141983e-01  0.78914051 -4.152191e-02 -0.0375176969
#>  [98,]  4.847001e-01 -3.772573e-01  0.78914051 -1.804215e-02 -0.0254410411
#>  [99,]  5.809330e-01 -1.994345e-01  0.78914051  3.710539e-02 -0.0317330339
#> [100,]  6.142127e-01 -1.504387e-16  0.78914051 -3.078041e-02 -0.0238184077
#> [101,]  7.357239e-01  0.000000e+00  0.67728157  2.244261e-02  0.0137804546
#> [102,]  6.958604e-01  2.388892e-01  0.67728157 -4.725582e-02  0.0378008378
#> [103,]  5.805895e-01  4.518910e-01  0.67728157  2.571498e-02  0.0433688580
#> [104,]  4.024028e-01  6.159234e-01  0.67728157 -3.528439e-03 -0.0109652535
#> [105,]  1.806095e-01  7.132110e-01  0.67728157  3.206485e-02  0.0008206311
#> [106,] -6.075560e-02  7.332110e-01  0.67728157 -8.738265e-03 -0.0085780721
#> [107,] -2.955369e-01  6.737563e-01  0.67728157 -2.291081e-02  0.0008441911
#> [108,] -4.982922e-01  5.412897e-01  0.67728157 -1.376853e-02  0.0296755642
#> [109,] -6.470499e-01  3.501659e-01  0.67728157  4.182947e-03  0.0448833886
#> [110,] -7.256896e-01  1.210962e-01  0.67728157 -1.044516e-02 -0.0136083371
#> [111,] -7.256896e-01 -1.210962e-01  0.67728157 -8.396360e-03  0.0301518152
#> [112,] -6.470499e-01 -3.501659e-01  0.67728157  5.215994e-03  0.0270571755
#> [113,] -4.982922e-01 -5.412897e-01  0.67728157  1.607129e-03  0.0061548268
#> [114,] -2.955369e-01 -6.737563e-01  0.67728157  3.557785e-02 -0.0118858161
#> [115,] -6.075560e-02 -7.332110e-01  0.67728157 -1.339833e-02  0.0166470921
#> [116,]  1.806095e-01 -7.132110e-01  0.67728157 -2.824950e-02  0.0074348239
#> [117,]  4.024028e-01 -6.159234e-01  0.67728157 -1.283247e-03 -0.0017691261
#> [118,]  5.805895e-01 -4.518910e-01  0.67728157 -1.005237e-02  0.0027092509
#> [119,]  6.958604e-01 -2.388892e-01  0.67728157 -3.412011e-02 -0.0423002261
#> [120,]  7.357239e-01 -1.802004e-16  0.67728157 -3.609985e-02 -0.0224883411
#> [121,]  8.371665e-01  0.000000e+00  0.54694816  2.975808e-02 -0.0146759835
#> [122,]  7.918065e-01  2.718275e-01  0.54694816  1.063481e-02  0.0422295440
#> [123,]  6.606420e-01  5.141983e-01  0.54694816  1.622648e-02 -0.0258545735
#> [124,]  4.578867e-01  7.008477e-01  0.54694816 -4.164142e-02 -0.0477003537
#> [125,]  2.055122e-01  8.115494e-01  0.54694816 -2.803801e-02 -0.0030060765
#> [126,] -6.913266e-02  8.343071e-01  0.54694816 -1.398614e-02  0.0265506459
#> [127,] -3.362859e-01  7.666547e-01  0.54694816  6.640408e-03 -0.0078081354
#> [128,] -5.669974e-01  6.159234e-01  0.54694816  2.090000e-02  0.0019698547
#> [129,] -7.362659e-01  3.984472e-01  0.54694816 -1.493015e-02 -0.0443911214
#> [130,] -8.257486e-01  1.377931e-01  0.54694816 -3.065316e-02  0.0091306838
#> [131,] -8.257486e-01 -1.377931e-01  0.54694816 -2.076540e-02  0.0214349334
#> [132,] -7.362659e-01 -3.984472e-01  0.54694816 -1.967559e-03 -0.0486329470
#> [133,] -5.669974e-01 -6.159234e-01  0.54694816 -1.897951e-02  0.0322591547
#> [134,] -3.362859e-01 -7.666547e-01  0.54694816  3.352277e-02  0.0401585845
#> [135,] -6.913266e-02 -8.343071e-01  0.54694816 -2.913357e-02  0.0152198908
#> [136,]  2.055122e-01 -8.115494e-01  0.54694816  1.018786e-02  0.0395091036
#> [137,]  4.578867e-01 -7.008477e-01  0.54694816 -3.809335e-02  0.0002048393
#> [138,]  6.606420e-01 -5.141983e-01  0.54694816 -4.270168e-02  0.0185578629
#> [139,]  7.918065e-01 -2.718275e-01  0.54694816  2.503381e-02  0.0142491821
#> [140,]  8.371665e-01 -2.050466e-16  0.54694816 -2.009889e-02  0.0464488313
#> [141,]  9.157733e-01  0.000000e+00  0.40169542  4.631907e-02 -0.0350199315
#> [142,]  8.661542e-01  2.973511e-01  0.40169542 -4.460922e-02 -0.0302801593
#> [143,]  7.226738e-01  5.624796e-01  0.40169542  4.727203e-02  0.0058469320
#> [144,]  5.008805e-01  7.666547e-01  0.40169542 -3.123575e-02  0.0041982303
#> [145,]  2.248091e-01  8.877509e-01  0.40169542 -3.287764e-02 -0.0039280682
#> [146,] -7.562396e-02  9.126455e-01  0.40169542 -4.616897e-02  0.0283623309
#> [147,] -3.678620e-01  8.386408e-01  0.40169542  3.662017e-02  0.0396664866
#> [148,] -6.202364e-01  6.737563e-01  0.40169542 -4.527129e-02 -0.0389704145
#> [149,] -8.053986e-01  4.358599e-01  0.40169542  4.233678e-02  0.0398126647
#> [150,] -9.032834e-01  1.507313e-01  0.40169542  3.484225e-02 -0.0051808101
#> [151,] -9.032834e-01 -1.507313e-01  0.40169542  1.476177e-03  0.0081318312
#> [152,] -8.053986e-01 -4.358599e-01  0.40169542 -7.141919e-03  0.0366609986
#> [153,] -6.202364e-01 -6.737563e-01  0.40169542 -5.994058e-03 -0.0167757102
#> [154,] -3.678620e-01 -8.386408e-01  0.40169542  1.278608e-02 -0.0376606197
#> [155,] -7.562396e-02 -9.126455e-01  0.40169542  2.373553e-02  0.0272550233
#> [156,]  2.248091e-01 -8.877509e-01  0.40169542  3.552742e-02  0.0264171684
#> [157,]  5.008805e-01 -7.666547e-01  0.40169542  2.582292e-02  0.0118820909
#> [158,]  7.226738e-01 -5.624796e-01  0.40169542  3.927075e-02  0.0411726092
#> [159,]  8.661542e-01 -2.973511e-01  0.40169542  4.792449e-04 -0.0276079002
#> [160,]  9.157733e-01 -2.242998e-16  0.40169542  1.735247e-02 -0.0010622081
#> [161,]  9.694003e-01  0.000000e+00  0.24548549 -2.700335e-02 -0.0479064680
#> [162,]  9.168755e-01  3.147638e-01  0.24548549  3.081018e-02 -0.0040199017
#> [163,]  7.649930e-01  5.954180e-01  0.24548549 -3.437978e-02  0.0245137195
#> [164,]  5.302117e-01  8.115494e-01  0.24548549  2.427493e-02  0.0441416510
#> [165,]  2.379737e-01  9.397369e-01  0.24548549 -4.724592e-02  0.0121284076
#> [166,] -8.005244e-02  9.660893e-01  0.24548549 -1.606065e-02  0.0199902450
#> [167,] -3.894037e-01  8.877509e-01  0.24548549  2.510918e-02  0.0337024454
#> [168,] -6.565569e-01  7.132110e-01  0.24548549 -1.421421e-02  0.0400298420
#> [169,] -8.525621e-01  4.613835e-01  0.24548549  1.608046e-02  0.0106446010
#> [170,] -9.561789e-01  1.595580e-01  0.24548549 -4.660045e-02  0.0051058598
#> [171,] -9.561789e-01 -1.595580e-01  0.24548549  3.924726e-02  0.0292815140
#> [172,] -8.525621e-01 -4.613835e-01  0.24548549  2.484573e-02 -0.0258847094
#> [173,] -6.565569e-01 -7.132110e-01  0.24548549  2.926683e-02  0.0478834353
#> [174,] -3.894037e-01 -8.877509e-01  0.24548549 -4.891862e-02  0.0407180830
#> [175,] -8.005244e-02 -9.660893e-01  0.24548549 -1.354476e-02  0.0459258026
#> [176,]  2.379737e-01 -9.397369e-01  0.24548549 -2.875110e-02 -0.0062494586
#> [177,]  5.302117e-01 -8.115494e-01  0.24548549  5.290703e-03 -0.0109179499
#> [178,]  7.649930e-01 -5.954180e-01  0.24548549 -9.127117e-03  0.0346075289
#> [179,]  9.168755e-01 -3.147638e-01  0.24548549  3.671608e-02  0.0066255765
#> [180,]  9.694003e-01 -2.374346e-16  0.24548549 -3.625103e-02  0.0016393725
#> [181,]  9.965845e-01  0.000000e+00  0.08257935 -4.571816e-02  0.0026284708
#> [182,]  9.425868e-01  3.235905e-01  0.08257935  1.720678e-02  0.0104732386
#> [183,]  7.864452e-01  6.121149e-01  0.08257935 -2.905942e-02 -0.0264232928
#> [184,]  5.450801e-01  8.343071e-01  0.08257935  2.907211e-02  0.0170200895
#> [185,]  2.446470e-01  9.660893e-01  0.08257935  4.659122e-02 -0.0400309947
#> [186,] -8.229730e-02  9.931807e-01  0.08257935  3.951857e-02  0.0281232769
#> [187,] -4.003234e-01  9.126455e-01  0.08257935 -3.559896e-02 -0.0350281331
#> [188,] -6.749683e-01  7.332110e-01  0.08257935  1.793591e-02 -0.0190916865
#> [189,] -8.764699e-01  4.743218e-01  0.08257935 -7.551380e-03  0.0492332134
#> [190,] -9.829924e-01  1.640324e-01  0.08257935 -1.094093e-03  0.0387078005
#> [191,] -9.829924e-01 -1.640324e-01  0.08257935  3.492902e-02 -0.0367262162
#> [192,] -8.764699e-01 -4.743218e-01  0.08257935 -1.758037e-02 -0.0425536354
#> [193,] -6.749683e-01 -7.332110e-01  0.08257935  7.119553e-03  0.0339544867
#> [194,] -4.003234e-01 -9.126455e-01  0.08257935 -1.837746e-02 -0.0029737594
#> [195,] -8.229730e-02 -9.931807e-01  0.08257935  3.158287e-04 -0.0283511394
#> [196,]  2.446470e-01 -9.660893e-01  0.08257935 -4.807500e-02 -0.0194406254
#> [197,]  5.450801e-01 -8.343071e-01  0.08257935  2.339810e-02 -0.0096749337
#> [198,]  7.864452e-01 -6.121149e-01  0.08257935 -1.495110e-03  0.0040005547
#> [199,]  9.425868e-01 -3.235905e-01  0.08257935 -2.254169e-02 -0.0153989097
#> [200,]  9.965845e-01 -2.440928e-16  0.08257935  3.576802e-02 -0.0277423693
#> [201,]  9.965845e-01  0.000000e+00 -0.08257935 -9.320607e-03 -0.0024594139
#> [202,]  9.425868e-01  3.235905e-01 -0.08257935  8.885773e-03 -0.0434584982
#> [203,]  7.864452e-01  6.121149e-01 -0.08257935 -8.807094e-03 -0.0491912927
#> [204,]  5.450801e-01  8.343071e-01 -0.08257935 -2.369820e-06 -0.0039507252
#> [205,]  2.446470e-01  9.660893e-01 -0.08257935 -2.187817e-02  0.0339367461
#> [206,] -8.229730e-02  9.931807e-01 -0.08257935  1.617901e-02  0.0351696699
#> [207,] -4.003234e-01  9.126455e-01 -0.08257935  2.396105e-02 -0.0346562735
#> [208,] -6.749683e-01  7.332110e-01 -0.08257935 -1.396856e-02  0.0353590870
#> [209,] -8.764699e-01  4.743218e-01 -0.08257935 -3.882461e-02  0.0341653890
#> [210,] -9.829924e-01  1.640324e-01 -0.08257935  8.309654e-03  0.0486927935
#> [211,] -9.829924e-01 -1.640324e-01 -0.08257935  3.594191e-02 -0.0186409093
#> [212,] -8.764699e-01 -4.743218e-01 -0.08257935  1.535813e-02  0.0046425507
#> [213,] -6.749683e-01 -7.332110e-01 -0.08257935 -3.055588e-02  0.0059940758
#> [214,] -4.003234e-01 -9.126455e-01 -0.08257935  3.756269e-02  0.0254501086
#> [215,] -8.229730e-02 -9.931807e-01 -0.08257935  2.099368e-02  0.0138193422
#> [216,]  2.446470e-01 -9.660893e-01 -0.08257935  1.577430e-02  0.0498462459
#> [217,]  5.450801e-01 -8.343071e-01 -0.08257935 -2.850424e-02  0.0371268279
#> [218,]  7.864452e-01 -6.121149e-01 -0.08257935  2.602364e-02  0.0095759856
#> [219,]  9.425868e-01 -3.235905e-01 -0.08257935 -2.892651e-02 -0.0237916254
#> [220,]  9.965845e-01 -2.440928e-16 -0.08257935 -2.526338e-02 -0.0027620777
#> [221,]  9.694003e-01  0.000000e+00 -0.24548549 -5.968637e-03 -0.0466033558
#> [222,]  9.168755e-01  3.147638e-01 -0.24548549  4.335062e-02 -0.0309553012
#> [223,]  7.649930e-01  5.954180e-01 -0.24548549 -5.219566e-03  0.0308160886
#> [224,]  5.302117e-01  8.115494e-01 -0.24548549 -3.497633e-03 -0.0250533162
#> [225,]  2.379737e-01  9.397369e-01 -0.24548549  3.290750e-04  0.0242647322
#> [226,] -8.005244e-02  9.660893e-01 -0.24548549  1.658156e-02  0.0290650770
#> [227,] -3.894037e-01  8.877509e-01 -0.24548549  1.689639e-02  0.0302366883
#> [228,] -6.565569e-01  7.132110e-01 -0.24548549  4.747614e-02 -0.0317684560
#> [229,] -8.525621e-01  4.613835e-01 -0.24548549  4.842098e-02 -0.0017692390
#> [230,] -9.561789e-01  1.595580e-01 -0.24548549  1.546469e-02  0.0237067814
#> [231,] -9.561789e-01 -1.595580e-01 -0.24548549  2.307396e-02 -0.0135680469
#> [232,] -8.525621e-01 -4.613835e-01 -0.24548549 -2.990679e-02  0.0100358631
#> [233,] -6.565569e-01 -7.132110e-01 -0.24548549  2.042425e-02 -0.0023742913
#> [234,] -3.894037e-01 -8.877509e-01 -0.24548549 -1.374456e-02 -0.0404959604
#> [235,] -8.005244e-02 -9.660893e-01 -0.24548549 -3.528680e-02  0.0449451298
#> [236,]  2.379737e-01 -9.397369e-01 -0.24548549  3.851849e-02  0.0032331870
#> [237,]  5.302117e-01 -8.115494e-01 -0.24548549 -4.833921e-02  0.0189601859
#> [238,]  7.649930e-01 -5.954180e-01 -0.24548549  4.447897e-03  0.0108713281
#> [239,]  9.168755e-01 -3.147638e-01 -0.24548549 -4.580795e-02  0.0378699395
#> [240,]  9.694003e-01 -2.374346e-16 -0.24548549  4.712559e-02 -0.0052085525
#> [241,]  9.157733e-01  0.000000e+00 -0.40169542 -9.925555e-03  0.0381767166
#> [242,]  8.661542e-01  2.973511e-01 -0.40169542  2.840028e-02 -0.0304293933
#> [243,]  7.226738e-01  5.624796e-01 -0.40169542  3.997728e-02  0.0080539073
#> [244,]  5.008805e-01  7.666547e-01 -0.40169542  2.822383e-02  0.0203363427
#> [245,]  2.248091e-01  8.877509e-01 -0.40169542 -1.862615e-02 -0.0178254703
#> [246,] -7.562396e-02  9.126455e-01 -0.40169542  4.691692e-02 -0.0281832897
#> [247,] -3.678620e-01  8.386408e-01 -0.40169542  3.840268e-02  0.0237727214
#> [248,] -6.202364e-01  6.737563e-01 -0.40169542  1.032877e-02 -0.0367811327
#> [249,] -8.053986e-01  4.358599e-01 -0.40169542  8.116701e-03 -0.0167824472
#> [250,] -9.032834e-01  1.507313e-01 -0.40169542  1.446084e-02 -0.0095084150
#> [251,] -9.032834e-01 -1.507313e-01 -0.40169542  7.370091e-04  0.0044107034
#> [252,] -8.053986e-01 -4.358599e-01 -0.40169542 -3.593937e-02 -0.0374312584
#> [253,] -6.202364e-01 -6.737563e-01 -0.40169542 -2.589176e-02 -0.0376135420
#> [254,] -3.678620e-01 -8.386408e-01 -0.40169542  2.379276e-02  0.0465826655
#> [255,] -7.562396e-02 -9.126455e-01 -0.40169542  2.767747e-02  0.0496049841
#> [256,]  2.248091e-01 -8.877509e-01 -0.40169542 -2.651068e-03  0.0028885404
#> [257,]  5.008805e-01 -7.666547e-01 -0.40169542  3.622647e-02  0.0078669869
#> [258,]  7.226738e-01 -5.624796e-01 -0.40169542  4.699369e-02 -0.0438675902
#> [259,]  8.661542e-01 -2.973511e-01 -0.40169542 -7.581202e-03 -0.0208439457
#> [260,]  9.157733e-01 -2.242998e-16 -0.40169542 -3.301410e-02  0.0376516746
#> [261,]  8.371665e-01  0.000000e+00 -0.54694816  7.923903e-03 -0.0286852656
#> [262,]  7.918065e-01  2.718275e-01 -0.54694816 -2.329295e-02  0.0055488282
#> [263,]  6.606420e-01  5.141983e-01 -0.54694816  1.736769e-02 -0.0177392087
#> [264,]  4.578867e-01  7.008477e-01 -0.54694816 -4.115406e-02 -0.0289964651
#> [265,]  2.055122e-01  8.115494e-01 -0.54694816 -3.049952e-02 -0.0113598105
#> [266,] -6.913266e-02  8.343071e-01 -0.54694816  3.355744e-02  0.0010038246
#> [267,] -3.362859e-01  7.666547e-01 -0.54694816 -4.913579e-02  0.0140679195
#> [268,] -5.669974e-01  6.159234e-01 -0.54694816 -1.417915e-02  0.0083117526
#> [269,] -7.362659e-01  3.984472e-01 -0.54694816 -6.221605e-03  0.0190250016
#> [270,] -8.257486e-01  1.377931e-01 -0.54694816 -7.042451e-03 -0.0377528639
#> [271,] -8.257486e-01 -1.377931e-01 -0.54694816 -2.106900e-02 -0.0200291937
#> [272,] -7.362659e-01 -3.984472e-01 -0.54694816  9.207022e-04  0.0407539491
#> [273,] -5.669974e-01 -6.159234e-01 -0.54694816 -2.243683e-02  0.0396335109
#> [274,] -3.362859e-01 -7.666547e-01 -0.54694816 -3.878158e-02  0.0103357211
#> [275,] -6.913266e-02 -8.343071e-01 -0.54694816  4.445817e-02  0.0322253232
#> [276,]  2.055122e-01 -8.115494e-01 -0.54694816  2.532280e-02  0.0107373370
#> [277,]  4.578867e-01 -7.008477e-01 -0.54694816  4.759618e-02  0.0182523262
#> [278,]  6.606420e-01 -5.141983e-01 -0.54694816 -1.209271e-02 -0.0095809901
#> [279,]  7.918065e-01 -2.718275e-01 -0.54694816 -3.342596e-02  0.0046864697
#> [280,]  8.371665e-01 -2.050466e-16 -0.54694816  4.841230e-02 -0.0068546325
#> [281,]  7.357239e-01  0.000000e+00 -0.67728157 -4.354098e-02 -0.0481068782
#> [282,]  6.958604e-01  2.388892e-01 -0.67728157  2.017294e-02  0.0104372159
#> [283,]  5.805895e-01  4.518910e-01 -0.67728157  2.603287e-02  0.0093141185
#> [284,]  4.024028e-01  6.159234e-01 -0.67728157 -3.991330e-02 -0.0322931050
#> [285,]  1.806095e-01  7.132110e-01 -0.67728157 -2.804533e-02  0.0276243568
#> [286,] -6.075560e-02  7.332110e-01 -0.67728157 -4.062016e-02 -0.0156757499
#> [287,] -2.955369e-01  6.737563e-01 -0.67728157 -4.483700e-02 -0.0220252863
#> [288,] -4.982922e-01  5.412897e-01 -0.67728157  1.925468e-02  0.0144242952
#> [289,] -6.470499e-01  3.501659e-01 -0.67728157  6.002385e-04  0.0436351998
#> [290,] -7.256896e-01  1.210962e-01 -0.67728157 -8.938643e-03  0.0278154743
#> [291,] -7.256896e-01 -1.210962e-01 -0.67728157  3.478389e-02  0.0304685674
#> [292,] -6.470499e-01 -3.501659e-01 -0.67728157 -4.124806e-02 -0.0032621220
#> [293,] -4.982922e-01 -5.412897e-01 -0.67728157  2.494363e-02  0.0213185854
#> [294,] -2.955369e-01 -6.737563e-01 -0.67728157 -6.416604e-03  0.0272414592
#> [295,] -6.075560e-02 -7.332110e-01 -0.67728157 -4.458165e-02 -0.0119070909
#> [296,]  1.806095e-01 -7.132110e-01 -0.67728157  3.483997e-02  0.0394971843
#> [297,]  4.024028e-01 -6.159234e-01 -0.67728157  1.774499e-02  0.0157238287
#> [298,]  5.805895e-01 -4.518910e-01 -0.67728157  3.123099e-02  0.0073775683
#> [299,]  6.958604e-01 -2.388892e-01 -0.67728157  3.353982e-02 -0.0459219912
#> [300,]  7.357239e-01 -1.802004e-16 -0.67728157 -1.069783e-03 -0.0226348604
#> [301,]  6.142127e-01  0.000000e+00 -0.78914051  3.889259e-02  0.0118887408
#> [302,]  5.809330e-01  1.994345e-01 -0.78914051 -1.520787e-02 -0.0202879455
#> [303,]  4.847001e-01  3.772573e-01 -0.78914051 -3.292522e-02  0.0085334935
#> [304,]  3.359425e-01  5.141983e-01 -0.78914051  1.839855e-02  0.0189687931
#> [305,]  1.507803e-01  5.954180e-01 -0.78914051 -4.215862e-02  0.0439416778
#> [306,] -5.072128e-02  6.121149e-01 -0.78914051  1.656574e-02 -0.0255372126
#> [307,] -2.467264e-01  5.624796e-01 -0.78914051  1.633445e-02  0.0078683428
#> [308,] -4.159950e-01  4.518910e-01 -0.78914051  3.362518e-02  0.0301495977
#> [309,] -5.401840e-01  2.923329e-01 -0.78914051 -4.571140e-02  0.0005122316
#> [310,] -6.058357e-01  1.010961e-01 -0.78914051  4.595713e-02  0.0325653430
#> [311,] -6.058357e-01 -1.010961e-01 -0.78914051 -1.712658e-02 -0.0405002255
#> [312,] -5.401840e-01 -2.923329e-01 -0.78914051  4.374033e-02  0.0318704513
#> [313,] -4.159950e-01 -4.518910e-01 -0.78914051 -2.611038e-03 -0.0301355670
#> [314,] -2.467264e-01 -5.624796e-01 -0.78914051  2.269044e-02  0.0292440554
#> [315,] -5.072128e-02 -6.121149e-01 -0.78914051  1.636613e-02  0.0236358887
#> [316,]  1.507803e-01 -5.954180e-01 -0.78914051  3.999788e-02 -0.0342380172
#> [317,]  3.359425e-01 -5.141983e-01 -0.78914051 -1.949816e-02  0.0451731122
#> [318,]  4.847001e-01 -3.772573e-01 -0.78914051  3.885122e-02 -0.0208861045
#> [319,]  5.809330e-01 -1.994345e-01 -0.78914051  2.003334e-03 -0.0479550461
#> [320,]  6.142127e-01 -1.504387e-16 -0.78914051 -3.430801e-02  0.0238272365
#> [321,]  4.759474e-01  0.000000e+00 -0.87947375 -3.413659e-02  0.0468909115
#> [322,]  4.501593e-01  1.545399e-01 -0.87947375  9.335427e-03 -0.0235861135
#> [323,]  3.755894e-01  2.923329e-01 -0.87947375  3.766894e-02 -0.0177708502
#> [324,]  2.603185e-01  3.984472e-01 -0.87947375  4.145389e-02  0.0121002652
#> [325,]  1.168382e-01  4.613835e-01 -0.87947375  2.446222e-02  0.0058167164
#> [326,] -3.930342e-02  4.743218e-01 -0.87947375  3.743208e-03  0.0087907006
#> [327,] -1.911859e-01  4.358599e-01 -0.87947375  4.517774e-02 -0.0155490278
#> [328,] -3.223504e-01  3.501659e-01 -0.87947375 -1.070952e-02 -0.0419516760
#> [329,] -4.185832e-01  2.265259e-01 -0.87947375  4.002746e-02  0.0230809814
#> [330,] -4.694561e-01  7.833837e-02 -0.87947375 -4.081297e-02 -0.0226636962
#> [331,] -4.694561e-01 -7.833837e-02 -0.87947375  3.968569e-02  0.0375408654
#> [332,] -4.185832e-01 -2.265259e-01 -0.87947375 -1.157779e-02  0.0211497325
#> [333,] -3.223504e-01 -3.501659e-01 -0.87947375  1.170740e-02 -0.0144789500
#> [334,] -1.911859e-01 -4.358599e-01 -0.87947375  1.831581e-02 -0.0404549389
#> [335,] -3.930342e-02 -4.743218e-01 -0.87947375  9.588638e-03  0.0257775217
#> [336,]  1.168382e-01 -4.613835e-01 -0.87947375 -4.574251e-02 -0.0257210549
#> [337,]  2.603185e-01 -3.984472e-01 -0.87947375 -2.341394e-02 -0.0336559628
#> [338,]  3.755894e-01 -2.923329e-01 -0.87947375  3.973561e-02 -0.0225065942
#> [339,]  4.501593e-01 -1.545399e-01 -0.87947375 -1.540693e-02 -0.0003932666
#> [340,]  4.759474e-01 -1.165735e-16 -0.87947375  3.312627e-02 -0.0049031496
#> [341,]  3.246995e-01  0.000000e+00 -0.94581724 -4.982232e-02  0.0040581789
#> [342,]  3.071064e-01  1.054297e-01 -0.94581724 -8.464566e-04  0.0157605133
#> [343,]  2.562335e-01  1.994345e-01 -0.94581724 -1.266126e-02  0.0346560197
#> [344,]  1.775938e-01  2.718275e-01 -0.94581724 -2.533797e-02  0.0299988588
#> [345,]  7.970901e-02  3.147638e-01 -0.94581724 -4.118392e-02 -0.0375437240
#> [346,] -2.681347e-02  3.235905e-01 -0.94581724  4.155776e-03  0.0456036761
#> [347,] -1.304303e-01  2.973511e-01 -0.94581724  1.402479e-02  0.0035671920
#> [348,] -2.199130e-01  2.388892e-01 -0.94581724  2.530704e-02  0.0129344609
#> [349,] -2.855647e-01  1.545399e-01 -0.94581724 -2.116601e-02 -0.0052698983
#> [350,] -3.202710e-01  5.344378e-02 -0.94581724  9.327792e-04  0.0295750235
#> [351,] -3.202710e-01 -5.344378e-02 -0.94581724 -3.258958e-02 -0.0169425233
#> [352,] -2.855647e-01 -1.545399e-01 -0.94581724 -1.009954e-03  0.0081031180
#> [353,] -2.199130e-01 -2.388892e-01 -0.94581724  1.965206e-02  0.0005213174
#> [354,] -1.304303e-01 -2.973511e-01 -0.94581724  1.579541e-02 -0.0186029508
#> [355,] -2.681347e-02 -3.235905e-01 -0.94581724 -3.144355e-02  0.0193202520
#> [356,]  7.970901e-02 -3.147638e-01 -0.94581724 -7.118557e-03 -0.0117970928
#> [357,]  1.775938e-01 -2.718275e-01 -0.94581724 -2.984907e-02  0.0384177311
#> [358,]  2.562335e-01 -1.994345e-01 -0.94581724  3.363478e-02  0.0433963194
#> [359,]  3.071064e-01 -1.054297e-01 -0.94581724 -3.837031e-02 -0.0390575929
#> [360,]  3.246995e-01 -7.952843e-17 -0.94581724 -1.007004e-02  0.0171533582
#> [361,]  1.645946e-01  0.000000e+00 -0.98636130 -4.066879e-02 -0.0465612458
#> [362,]  1.556764e-01  5.344378e-02 -0.98636130 -9.733680e-03 -0.0346247125
#> [363,]  1.298883e-01  1.010961e-01 -0.98636130 -4.813094e-03  0.0416392289
#> [364,]  9.002471e-02  1.377931e-01 -0.98636130  1.783845e-02 -0.0049372797
#> [365,]  4.040558e-02  1.595580e-01 -0.98636130 -1.307484e-02  0.0185919262
#> [366,] -1.359211e-02  1.640324e-01 -0.98636130 -6.667444e-03 -0.0045117962
#> [367,] -6.611689e-02  1.507313e-01 -0.98636130 -3.290617e-02  0.0461717237
#> [368,] -1.114769e-01  1.210962e-01 -0.98636130 -8.257724e-03 -0.0122553712
#> [369,] -1.447566e-01  7.833837e-02 -0.98636130 -2.379192e-02 -0.0162304267
#> [370,] -1.623497e-01  2.709138e-02 -0.98636130  2.736267e-02 -0.0264930890
#> [371,] -1.623497e-01 -2.709138e-02 -0.98636130  2.927325e-02  0.0172058788
#> [372,] -1.447566e-01 -7.833837e-02 -0.98636130  1.527919e-02  0.0004430643
#> [373,] -1.114769e-01 -1.210962e-01 -0.98636130  4.065855e-02 -0.0347882810
#> [374,] -6.611689e-02 -1.507313e-01 -0.98636130 -4.764542e-02 -0.0333774019
#> [375,] -1.359211e-02 -1.640324e-01 -0.98636130 -1.713199e-02  0.0260965556
#> [376,]  4.040558e-02 -1.595580e-01 -0.98636130 -5.897846e-03 -0.0074521042
#> [377,]  9.002471e-02 -1.377931e-01 -0.98636130  1.804426e-02 -0.0394366712
#> [378,]  1.298883e-01 -1.010961e-01 -0.98636130  3.669566e-03  0.0301168667
#> [379,]  1.556764e-01 -5.344378e-02 -0.98636130 -3.647639e-02 -0.0419587971
#> [380,]  1.645946e-01 -4.031405e-17 -0.98636130  9.567656e-03 -0.0178800270
#> [381,]  1.224647e-16  0.000000e+00 -1.00000000  2.803267e-02 -0.0242777936
#> [382,]  1.158292e-16  3.976422e-17 -1.00000000 -2.222794e-03 -0.0480781602
#> [383,]  9.664184e-17  7.521936e-17 -1.00000000 -1.373098e-02 -0.0044925978
#> [384,]  6.698183e-17  1.025233e-16 -1.00000000 -4.923526e-02  0.0268020860
#> [385,]  3.006330e-17  1.187173e-16 -1.00000000 -3.362425e-02  0.0154708300
#> [386,] -1.011305e-17  1.220464e-16 -1.00000000 -1.111713e-02  0.0422128927
#> [387,] -4.919350e-17  1.121499e-16 -1.00000000  3.094393e-02 -0.0053710290
#> [388,] -8.294307e-17  9.010019e-17 -1.00000000 -4.189133e-02 -0.0415850116
#> [389,] -1.077045e-16  5.828675e-17 -1.00000000 -1.425399e-02  0.0147140309
#> [390,] -1.207944e-16  2.015702e-17 -1.00000000  3.304404e-02 -0.0090381307
#> [391,] -1.207944e-16 -2.015702e-17 -1.00000000  3.638922e-02 -0.0208063293
#> [392,] -1.077045e-16 -5.828675e-17 -1.00000000  3.358489e-02 -0.0371321189
#> [393,] -8.294307e-17 -9.010019e-17 -1.00000000 -2.726210e-02 -0.0022840990
#> [394,] -4.919350e-17 -1.121499e-16 -1.00000000 -6.798036e-03 -0.0159999489
#> [395,] -1.011305e-17 -1.220464e-16 -1.00000000 -2.147680e-02  0.0035788741
#> [396,]  3.006330e-17 -1.187173e-16 -1.00000000 -7.835588e-03  0.0112141618
#> [397,]  6.698183e-17 -1.025233e-16 -1.00000000  1.377198e-02 -0.0371199941
#> [398,]  9.664184e-17 -7.521936e-17 -1.00000000 -1.840246e-02  0.0062326387
#> [399,]  1.158292e-16 -3.976422e-17 -1.00000000 -4.141806e-02  0.0117852898
#> [400,]  1.224647e-16 -2.999520e-32 -1.00000000  2.342941e-02 -0.0316545064
#>                 [,6]
#>   [1,] -0.0224097124
#>   [2,]  0.0058730685
#>   [3,]  0.0384471205
#>   [4,]  0.0066567696
#>   [5,] -0.0256157019
#>   [6,]  0.0338826125
#>   [7,]  0.0412903447
#>   [8,]  0.0403237645
#>   [9,] -0.0432364848
#>  [10,] -0.0427404279
#>  [11,]  0.0391346660
#>  [12,]  0.0330254006
#>  [13,]  0.0030522383
#>  [14,] -0.0263212564
#>  [15,]  0.0303858228
#>  [16,] -0.0071982476
#>  [17,]  0.0012970382
#>  [18,]  0.0453215297
#>  [19,] -0.0155376290
#>  [20,]  0.0387788581
#>  [21,]  0.0166450149
#>  [22,] -0.0098181709
#>  [23,]  0.0075139229
#>  [24,] -0.0331760231
#>  [25,]  0.0198006235
#>  [26,] -0.0396713876
#>  [27,] -0.0223306950
#>  [28,]  0.0475047150
#>  [29,] -0.0336251184
#>  [30,] -0.0084340744
#>  [31,] -0.0388732226
#>  [32,] -0.0096390683
#>  [33,]  0.0354503244
#>  [34,] -0.0090557174
#>  [35,] -0.0363325094
#>  [36,] -0.0135676930
#>  [37,] -0.0378405037
#>  [38,]  0.0376625641
#>  [39,]  0.0498988348
#>  [40,] -0.0105055742
#>  [41,]  0.0119267519
#>  [42,] -0.0281139098
#>  [43,] -0.0171310939
#>  [44,]  0.0122536972
#>  [45,] -0.0224939816
#>  [46,] -0.0473302682
#>  [47,]  0.0481403759
#>  [48,] -0.0174051022
#>  [49,] -0.0003205740
#>  [50,] -0.0363566860
#>  [51,]  0.0003904358
#>  [52,] -0.0482579468
#>  [53,] -0.0030323254
#>  [54,] -0.0153248419
#>  [55,] -0.0283819866
#>  [56,] -0.0429583367
#>  [57,]  0.0448328620
#>  [58,] -0.0183201506
#>  [59,] -0.0454515950
#>  [60,]  0.0422209830
#>  [61,]  0.0139031098
#>  [62,]  0.0141489498
#>  [63,] -0.0036226860
#>  [64,] -0.0035760734
#>  [65,]  0.0181178080
#>  [66,]  0.0078491741
#>  [67,]  0.0274168907
#>  [68,] -0.0343960205
#>  [69,] -0.0430605633
#>  [70,] -0.0460080283
#>  [71,] -0.0155184703
#>  [72,]  0.0219805820
#>  [73,]  0.0460303989
#>  [74,]  0.0349399659
#>  [75,]  0.0180103326
#>  [76,] -0.0269992773
#>  [77,] -0.0053544548
#>  [78,] -0.0347711010
#>  [79,]  0.0014136756
#>  [80,] -0.0298901162
#>  [81,] -0.0276710545
#>  [82,]  0.0224227906
#>  [83,] -0.0135186209
#>  [84,]  0.0186513433
#>  [85,] -0.0079166902
#>  [86,] -0.0187565769
#>  [87,]  0.0483486468
#>  [88,] -0.0474801549
#>  [89,]  0.0178648174
#>  [90,] -0.0324134449
#>  [91,]  0.0365930069
#>  [92,]  0.0348708935
#>  [93,] -0.0355114629
#>  [94,] -0.0271630161
#>  [95,]  0.0476140357
#>  [96,] -0.0071402431
#>  [97,] -0.0408832290
#>  [98,]  0.0209972294
#>  [99,] -0.0066626939
#> [100,] -0.0068892051
#> [101,] -0.0164028188
#> [102,] -0.0235047219
#> [103,]  0.0293101677
#> [104,] -0.0271239813
#> [105,] -0.0320154624
#> [106,] -0.0260786190
#> [107,] -0.0181429178
#> [108,]  0.0045807536
#> [109,]  0.0386862828
#> [110,]  0.0170380435
#> [111,] -0.0234720815
#> [112,]  0.0255245471
#> [113,] -0.0385013859
#> [114,] -0.0107553932
#> [115,] -0.0471267601
#> [116,] -0.0085367720
#> [117,] -0.0445342469
#> [118,] -0.0230817826
#> [119,]  0.0364498982
#> [120,] -0.0237685415
#> [121,] -0.0114170825
#> [122,]  0.0251514799
#> [123,]  0.0028231731
#> [124,]  0.0320612674
#> [125,] -0.0288956664
#> [126,] -0.0488141491
#> [127,] -0.0018438478
#> [128,]  0.0328250262
#> [129,]  0.0163606977
#> [130,] -0.0286887900
#> [131,] -0.0103585480
#> [132,] -0.0071313513
#> [133,] -0.0404556623
#> [134,]  0.0076960147
#> [135,]  0.0474338481
#> [136,] -0.0279450347
#> [137,] -0.0234312846
#> [138,]  0.0485053924
#> [139,] -0.0399691103
#> [140,] -0.0289001654
#> [141,] -0.0194087601
#> [142,]  0.0452673686
#> [143,]  0.0331066190
#> [144,] -0.0083416061
#> [145,] -0.0311471867
#> [146,] -0.0441733215
#> [147,] -0.0119291255
#> [148,] -0.0237154981
#> [149,] -0.0299821134
#> [150,] -0.0074312360
#> [151,]  0.0230554801
#> [152,]  0.0098318679
#> [153,] -0.0245330961
#> [154,] -0.0390898018
#> [155,] -0.0129817799
#> [156,] -0.0407671100
#> [157,]  0.0079379001
#> [158,]  0.0463330709
#> [159,] -0.0278409448
#> [160,]  0.0370877966
#> [161,]  0.0169036374
#> [162,]  0.0438895473
#> [163,] -0.0246802411
#> [164,]  0.0475824768
#> [165,]  0.0163820421
#> [166,] -0.0143898793
#> [167,]  0.0030039197
#> [168,]  0.0081166394
#> [169,] -0.0396844822
#> [170,] -0.0265566208
#> [171,] -0.0204097725
#> [172,]  0.0207407020
#> [173,]  0.0136008093
#> [174,] -0.0404925953
#> [175,] -0.0140859347
#> [176,]  0.0415320981
#> [177,]  0.0046214467
#> [178,] -0.0461561202
#> [179,]  0.0237664109
#> [180,]  0.0341202633
#> [181,] -0.0107587299
#> [182,]  0.0031358037
#> [183,] -0.0244422555
#> [184,] -0.0419026258
#> [185,] -0.0329581960
#> [186,]  0.0175840472
#> [187,]  0.0232373429
#> [188,]  0.0347976334
#> [189,]  0.0263744748
#> [190,] -0.0026231199
#> [191,] -0.0498535967
#> [192,] -0.0010101914
#> [193,]  0.0279497629
#> [194,] -0.0454600266
#> [195,]  0.0021767576
#> [196,]  0.0321382353
#> [197,] -0.0413611765
#> [198,]  0.0252109422
#> [199,] -0.0047064543
#> [200,]  0.0379222085
#> [201,]  0.0231339094
#> [202,] -0.0254172823
#> [203,] -0.0385850925
#> [204,] -0.0350221756
#> [205,]  0.0248769978
#> [206,]  0.0168705793
#> [207,]  0.0247189383
#> [208,]  0.0364377812
#> [209,] -0.0125814772
#> [210,]  0.0418471181
#> [211,]  0.0428638628
#> [212,] -0.0272539393
#> [213,] -0.0161944947
#> [214,] -0.0217740091
#> [215,] -0.0492675133
#> [216,] -0.0048696535
#> [217,] -0.0427472331
#> [218,] -0.0456530863
#> [219,]  0.0470452822
#> [220,]  0.0142377665
#> [221,] -0.0205767895
#> [222,] -0.0295530638
#> [223,] -0.0164014023
#> [224,] -0.0203098571
#> [225,]  0.0080661156
#> [226,] -0.0070536014
#> [227,]  0.0250502022
#> [228,]  0.0248653408
#> [229,] -0.0100916413
#> [230,]  0.0407043314
#> [231,] -0.0067056266
#> [232,]  0.0139697193
#> [233,]  0.0205438121
#> [234,] -0.0418444135
#> [235,]  0.0492599612
#> [236,]  0.0302602059
#> [237,]  0.0328657944
#> [238,] -0.0287647745
#> [239,] -0.0208798466
#> [240,]  0.0464145993
#> [241,]  0.0287747040
#> [242,] -0.0047427858
#> [243,]  0.0409220008
#> [244,]  0.0042775594
#> [245,]  0.0262099978
#> [246,]  0.0107163528
#> [247,]  0.0286980778
#> [248,]  0.0235257468
#> [249,]  0.0185531181
#> [250,]  0.0294697858
#> [251,] -0.0468884161
#> [252,] -0.0393205576
#> [253,] -0.0118330416
#> [254,] -0.0029590431
#> [255,]  0.0323901244
#> [256,]  0.0379452252
#> [257,]  0.0444275221
#> [258,]  0.0355318096
#> [259,]  0.0426709957
#> [260,] -0.0087971702
#> [261,] -0.0311517806
#> [262,] -0.0011031435
#> [263,]  0.0193374486
#> [264,] -0.0191149206
#> [265,] -0.0142531339
#> [266,] -0.0325184840
#> [267,] -0.0433118920
#> [268,]  0.0091552068
#> [269,] -0.0116142937
#> [270,]  0.0128278553
#> [271,]  0.0211744872
#> [272,]  0.0241856493
#> [273,]  0.0189844236
#> [274,] -0.0134978752
#> [275,]  0.0434817782
#> [276,] -0.0403850951
#> [277,]  0.0245464425
#> [278,]  0.0182316082
#> [279,] -0.0249041270
#> [280,] -0.0494712179
#> [281,]  0.0177415921
#> [282,]  0.0256914099
#> [283,]  0.0058509766
#> [284,]  0.0201534934
#> [285,]  0.0452117470
#> [286,] -0.0279468767
#> [287,] -0.0279926708
#> [288,]  0.0082922763
#> [289,]  0.0200072185
#> [290,]  0.0387308071
#> [291,] -0.0161308155
#> [292,]  0.0109509172
#> [293,]  0.0176600055
#> [294,]  0.0397794622
#> [295,] -0.0166495348
#> [296,] -0.0420082424
#> [297,]  0.0105494620
#> [298,]  0.0322157142
#> [299,]  0.0113894432
#> [300,] -0.0337994490
#> [301,] -0.0491327587
#> [302,]  0.0298733779
#> [303,] -0.0353214066
#> [304,] -0.0383287284
#> [305,]  0.0267304244
#> [306,]  0.0355022175
#> [307,]  0.0339183060
#> [308,]  0.0026046358
#> [309,] -0.0211841339
#> [310,]  0.0391550695
#> [311,] -0.0007669843
#> [312,]  0.0350284525
#> [313,] -0.0144637923
#> [314,] -0.0435446614
#> [315,] -0.0278114101
#> [316,]  0.0260843266
#> [317,] -0.0243631189
#> [318,] -0.0285567578
#> [319,]  0.0489091729
#> [320,]  0.0026081893
#> [321,]  0.0067090560
#> [322,] -0.0121331936
#> [323,]  0.0054102787
#> [324,]  0.0106938969
#> [325,] -0.0076204948
#> [326,]  0.0009310514
#> [327,]  0.0214030946
#> [328,]  0.0279471935
#> [329,]  0.0239026508
#> [330,] -0.0388526369
#> [331,] -0.0399831240
#> [332,] -0.0158668309
#> [333,]  0.0212632748
#> [334,]  0.0134260325
#> [335,] -0.0306095738
#> [336,]  0.0119397532
#> [337,]  0.0203225523
#> [338,]  0.0295053924
#> [339,] -0.0050001372
#> [340,]  0.0158011169
#> [341,]  0.0342345787
#> [342,]  0.0153676139
#> [343,] -0.0098876848
#> [344,]  0.0413664310
#> [345,]  0.0366867488
#> [346,]  0.0156652802
#> [347,] -0.0092806607
#> [348,]  0.0426080174
#> [349,]  0.0457930321
#> [350,] -0.0300707191
#> [351,] -0.0046635683
#> [352,]  0.0205307506
#> [353,] -0.0359196191
#> [354,]  0.0171758374
#> [355,]  0.0073782693
#> [356,] -0.0288933859
#> [357,]  0.0445737168
#> [358,]  0.0099332970
#> [359,] -0.0236614419
#> [360,] -0.0092817232
#> [361,] -0.0132293402
#> [362,]  0.0481698418
#> [363,] -0.0220092569
#> [364,]  0.0407650940
#> [365,] -0.0082799604
#> [366,]  0.0414082336
#> [367,]  0.0270289054
#> [368,]  0.0040653758
#> [369,] -0.0342927033
#> [370,]  0.0084706679
#> [371,] -0.0016720979
#> [372,]  0.0400159159
#> [373,] -0.0136091234
#> [374,]  0.0099424049
#> [375,]  0.0189795232
#> [376,]  0.0471670046
#> [377,]  0.0355190796
#> [378,] -0.0479021518
#> [379,] -0.0089147941
#> [380,] -0.0401660026
#> [381,]  0.0165191547
#> [382,]  0.0404188616
#> [383,] -0.0110777525
#> [384,] -0.0127219753
#> [385,]  0.0331947756
#> [386,]  0.0321540025
#> [387,] -0.0184107002
#> [388,]  0.0328720801
#> [389,] -0.0007916308
#> [390,] -0.0302327739
#> [391,]  0.0383540310
#> [392,] -0.0337559121
#> [393,] -0.0158351770
#> [394,]  0.0078958530
#> [395,]  0.0116285037
#> [396,] -0.0217218023
#> [397,]  0.0417551709
#> [398,]  0.0431604248
#> [399,] -0.0258266344
#> [400,]  0.0116242698