Generate a spiral dataset with optional noise.

This function generates a dataset arranged in a spiral pattern with optional noise.

Usage

spiral_3d(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate.

num_dims

Number of effective dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
spiral_3d(n = 100, num_dims = 10, num_noise = 2, min_n = -0.05, max_n = 0.05)
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,] -0.439112258  0.0427584121 -4.163586e-03  4.054279e-04 -3.947840e-05
#>   [2,] -1.162891415  1.0326564554 -9.170068e-01  8.143091e-01 -7.231127e-01
#>   [3,]  0.616623480 -0.3594179731  2.094978e-01 -1.221123e-01  7.117689e-02
#>   [4,] -0.872298453 -0.6776035480 -5.263641e-01 -4.088809e-01 -3.176197e-01
#>   [5,]  0.014110930  0.0004058330  1.167184e-05  3.356842e-07  9.654342e-09
#>   [6,]  0.519877205  0.5154836421  5.111272e-01  5.068076e-01  5.025245e-01
#>   [7,] -0.040384530  0.0003622093 -3.248659e-06  2.913726e-08 -2.613324e-10
#>   [8,]  1.027601537  0.5292064218  2.725370e-01  1.403543e-01  7.228134e-02
#>   [9,]  1.611036044 -1.1898789351  8.788207e-01 -6.490794e-01  4.793970e-01
#>  [10,] -1.342482766  1.0729532471 -8.575370e-01  6.853698e-01 -5.477685e-01
#>  [11,]  1.029051287  0.8442528612  6.926408e-01  5.682554e-01  4.662073e-01
#>  [12,]  0.253977652  0.2519700010  2.499782e-01  2.480022e-01  2.460418e-01
#>  [13,]  0.144455983  0.0797913559  4.407336e-02  2.434425e-02  1.344673e-02
#>  [14,] -0.329662221  0.2988278239 -2.708775e-01  2.455414e-01 -2.225751e-01
#>  [15,]  1.626794727 -1.1139113892  7.627260e-01 -5.222596e-01  3.576056e-01
#>  [16,] -0.074938088  0.0749145653 -7.489105e-02  7.486754e-02 -7.484404e-02
#>  [17,] -0.495208057 -0.1615567939 -5.270633e-02 -1.719492e-02 -5.609676e-03
#>  [18,]  0.974991432  0.4600587254  2.170830e-01  1.024326e-01  4.833377e-02
#>  [19,] -0.113273921  0.1132447495 -1.132156e-01  1.131864e-01 -1.131573e-01
#>  [20,] -1.662003733  0.6482627786 -2.528542e-01  9.862551e-02 -3.846878e-02
#>  [21,]  1.544647046 -1.2300110147  9.794646e-01 -7.799531e-01  6.210810e-01
#>  [22,]  0.056960525  0.0007217050  9.144195e-06  1.158594e-07  1.467969e-09
#>  [23,]  0.331123158  0.0450515394  6.129566e-03  8.339688e-04  1.134671e-04
#>  [24,]  0.063481581  0.0561864761  4.972970e-02  4.401492e-02  3.895686e-02
#>  [25,]  0.008577614 -0.0085775785  8.577543e-03 -8.577508e-03  8.577473e-03
#>  [26,] -0.823609694  0.1513738103 -2.782147e-02  5.113395e-03 -9.398069e-04
#>  [27,]  0.927075359  0.4057583052  1.775905e-01  7.772705e-02  3.401924e-02
#>  [28,]  0.144538991  0.0732646885  3.713679e-02  1.882409e-02  9.541656e-03
#>  [29,] -0.123099435 -0.0099642300 -8.065502e-04 -6.528586e-05 -5.284535e-06
#>  [30,]  1.599712026 -0.9932317507  6.166793e-01 -3.828848e-01  2.377261e-01
#>  [31,]  1.617190318 -1.1776862653  8.576263e-01 -6.245491e-01  4.548153e-01
#>  [32,] -0.376355231  0.2684783340 -1.915228e-01  1.366255e-01 -9.746375e-02
#>  [33,] -0.609595011 -0.5724397350 -5.375491e-01 -5.047851e-01 -4.740180e-01
#>  [34,]  0.024458773  0.0012379013  6.265235e-05  3.170945e-06  1.604871e-07
#>  [35,]  1.373670507 -1.1886470967  1.028545e+00 -8.900075e-01  7.701300e-01
#>  [36,]  0.012083252  0.0118486040  1.161851e-02  1.139289e-02  1.117165e-02
#>  [37,]  1.589589117 -0.9673849752  5.887268e-01 -3.582847e-01  2.180433e-01
#>  [38,]  0.164590099 -0.1625054095  1.604471e-01 -1.584149e-01  1.564084e-01
#>  [39,]  1.905706346  1.6011187672  1.345213e+00  1.130209e+00  9.495684e-01
#>  [40,] -0.380175467  0.2794364310 -2.053913e-01  1.509666e-01 -1.109634e-01
#>  [41,] -0.666604002  0.0988100402 -1.464651e-02  2.171038e-03 -3.218108e-04
#>  [42,]  0.595472727 -0.4623707716  3.590202e-01 -2.787709e-01  2.164591e-01
#>  [43,] -0.381278059  0.3062536134 -2.459918e-01  1.975878e-01 -1.587082e-01
#>  [44,] -0.738948832  0.2203329382 -6.569684e-02  1.958888e-02 -5.840830e-03
#>  [45,]  1.044830945  0.8462647441  6.854353e-01  5.551709e-01  4.496628e-01
#>  [46,]  1.119683956  0.8110979708  5.875586e-01  4.256269e-01  3.083237e-01
#>  [47,]  0.056247432  0.0069192137  8.511592e-04  1.047044e-04  1.288009e-05
#>  [48,] -0.627624059  0.6123035419 -5.973570e-01  5.827753e-01 -5.685496e-01
#>  [49,]  1.437473116 -1.2143153932  1.025801e+00 -8.665528e-01  7.320265e-01
#>  [50,] -0.781599976  0.7503838378 -7.204144e-01  6.916420e-01 -6.640186e-01
#>  [51,]  1.594138938 -1.2098305007  9.181696e-01 -6.968210e-01  5.288343e-01
#>  [52,] -1.692950346  0.6749027786 -2.690532e-01  1.072594e-01 -4.275946e-02
#>  [53,] -1.335349968  0.8503204461 -5.414647e-01  3.447924e-01 -2.195559e-01
#>  [54,] -0.198084471  0.1933085305 -1.886477e-01  1.840993e-01 -1.796606e-01
#>  [55,] -1.374169470  1.0303414629 -7.725419e-01  5.792459e-01 -4.343140e-01
#>  [56,] -1.243241382  0.3509339703 -9.905932e-02  2.796181e-02 -7.892876e-03
#>  [57,] -0.149006457  0.1471106652 -1.452390e-01  1.433911e-01 -1.415668e-01
#>  [58,]  1.078821558  0.2721927598  6.867577e-02  1.732728e-02  4.371771e-03
#>  [59,]  0.337290965  0.0254900507  1.926357e-03  1.455803e-04  1.100192e-05
#>  [60,] -0.868737950  0.8547654287 -8.410176e-01  8.274910e-01 -8.141818e-01
#>  [61,]  0.683930400  0.6419406005  6.025288e-01  5.655366e-01  5.308156e-01
#>  [62,] -0.363509267  0.2419300181 -1.610141e-01  1.071614e-01 -7.132020e-02
#>  [63,] -0.206800450  0.2012875048 -1.959215e-01  1.906986e-01 -1.856149e-01
#>  [64,] -0.381463686  0.2842577383 -2.118222e-01  1.578449e-01 -1.176223e-01
#>  [65,]  1.547770696 -1.2295797387  9.768025e-01 -7.759913e-01  6.164629e-01
#>  [66,]  1.348075084 -1.1762498578  1.026325e+00 -8.955103e-01  7.813688e-01
#>  [67,] -1.865686447 -1.2750258336 -8.713634e-01 -5.954971e-01 -4.069677e-01
#>  [68,] -2.295312095  1.4270624209 -8.872463e-01  5.516269e-01 -3.429625e-01
#>  [69,]  0.319952550 -0.3181479313  3.163535e-01 -3.145692e-01  3.127949e-01
#>  [70,]  0.143813778  0.0817985629  4.652548e-02  2.646281e-02  1.505155e-02
#>  [71,] -0.754093743 -0.3996855389 -2.118417e-01 -1.122806e-01 -5.951107e-02
#>  [72,] -0.816484943 -0.1918181154 -4.506414e-02 -1.058699e-02 -2.487218e-03
#>  [73,]  0.446447308  0.4354038889  4.246336e-01  4.141298e-01  4.038858e-01
#>  [74,] -1.875539234 -1.3764098350 -1.010112e+00 -7.412949e-01 -5.440172e-01
#>  [75,]  2.027579349  1.6133707238  1.283780e+00  1.021520e+00  8.128363e-01
#>  [76,] -0.253409218  0.0142362438 -7.997761e-04  4.493052e-05 -2.524146e-06
#>  [77,] -0.424661174 -0.0513397443 -6.206758e-03 -7.503709e-04 -9.071668e-05
#>  [78,]  0.026939663  0.0015075004  8.435731e-05  4.720500e-06  2.641516e-07
#>  [79,] -0.142307802  0.1406700366 -1.390511e-01  1.374508e-01 -1.358690e-01
#>  [80,]  0.136008325 -0.1348242125  1.336504e-01 -1.324868e-01  1.313334e-01
#>  [81,]  0.596699825 -0.3191439156  1.706936e-01 -9.129519e-02  4.882908e-02
#>  [82,] -0.098432424  0.0176679025 -3.171260e-03  5.692180e-04 -1.021705e-04
#>  [83,] -0.046755410  0.0041542099 -3.691008e-04  3.279454e-05 -2.913789e-06
#>  [84,] -1.869003374 -1.2894026193 -8.895431e-01 -6.136850e-01 -4.233738e-01
#>  [85,] -1.886895845  1.7033393069 -1.537639e+00  1.388058e+00 -1.253028e+00
#>  [86,]  2.098207247  1.3280500500  8.405828e-01  5.320428e-01  3.367539e-01
#>  [87,]  0.177205955  0.0070032498  2.767712e-04  1.093811e-05  4.322784e-07
#>  [88,] -0.294290997  0.2749809356 -2.569379e-01  2.400788e-01 -2.243259e-01
#>  [89,] -0.020771975  0.0008419290 -3.412504e-05  1.383155e-06 -5.606198e-08
#>  [90,]  1.549425586 -1.2293190269  9.753455e-01 -7.738421e-01  6.139687e-01
#>  [91,] -0.595858498 -0.5888807578 -5.819847e-01 -5.751695e-01 -5.684340e-01
#>  [92,] -0.954011894  0.3767938037 -1.488174e-01  5.877649e-02 -2.321420e-02
#>  [93,]  0.857955325  0.7676702712  6.868862e-01  6.146032e-01  5.499268e-01
#>  [94,]  1.088969761  0.6328276871  3.677521e-01  2.137100e-01  1.241922e-01
#>  [95,]  1.954019321  1.0466199259  5.605949e-01  3.002682e-01  1.608309e-01
#>  [96,] -0.204133617  0.1988547775 -1.937124e-01  1.887031e-01 -1.838233e-01
#>  [97,]  1.251326512 -1.1213424657  1.004861e+00 -9.004788e-01  8.069398e-01
#>  [98,] -0.846198010 -0.5413540454 -3.463305e-01 -2.215645e-01 -1.417456e-01
#>  [99,] -1.809171494 -1.4613798006 -1.180447e+00 -9.535200e-01 -7.702172e-01
#> [100,] -0.115061938  0.0238879142 -4.959350e-03  1.029607e-03 -2.137558e-04
#>                 [,6]          [,7]          [,8]          [,9]       [,10]
#>   [1,]  3.844197e-06 -3.743274e-07  3.645001e-08 -3.549307e-09 -0.44120897
#>   [2,]  6.421296e-01 -5.702160e-01  5.063561e-01 -4.496481e-01 -2.52896654
#>   [3,] -4.148764e-02  2.418235e-02 -1.409543e-02  8.215953e-03 -0.75886732
#>   [4,] -2.467277e-01 -1.916587e-01 -1.488809e-01 -1.156511e-01  1.38516271
#>   [5,]  2.776607e-10  7.985575e-12  2.296667e-13  6.605257e-15  0.01411677
#>   [6,]  4.982776e-01  4.940666e-01  4.898911e-01  4.857510e-01  4.00725771
#>   [7,]  2.343893e-12 -2.102240e-14  1.885501e-16 -1.691108e-18 -0.04038615
#>   [8,]  3.722430e-02  1.917021e-02  9.872503e-03  5.084259e-03  1.19879543
#>   [9,] -3.540730e-01  2.615112e-01 -1.931470e-01  1.426545e-01 -2.38967304
#>  [10,]  4.377933e-01 -3.498978e-01  2.796490e-01 -2.235040e-01 -2.23365709
#>  [11,]  3.824852e-01  3.137980e-01  2.574457e-01  2.112133e-01  1.79978554
#>  [12,]  2.440968e-01  2.421673e-01  2.402530e-01  2.383538e-01  2.02392333
#>  [13,]  7.427405e-03  4.102583e-03  2.266093e-03  1.251694e-03  0.17329021
#>  [14,]  2.017569e-01 -1.828859e-01  1.657800e-01 -1.502741e-01 -0.78067708
#>  [15,] -2.448625e-01  1.676641e-01 -1.148043e-01  7.860965e-02 -2.23215826
#>  [16,]  7.482055e-02 -7.479706e-02  7.477358e-02 -7.475011e-02 -2.99108510
#>  [17,] -1.830102e-03 -5.970529e-04 -1.947827e-04 -6.354595e-05  0.52387065
#>  [18,]  2.280674e-02  1.076157e-02  5.077946e-03  2.396076e-03  1.10584134
#>  [19,]  1.131281e-01 -1.130990e-01  1.130699e-01 -1.130408e-01 -4.99149006
#>  [20,]  1.500470e-02 -5.852569e-03  2.282788e-03 -8.903991e-04 -1.80496777
#>  [21,] -4.945702e-01  3.938290e-01 -3.136082e-01  2.497280e-01 -2.55358050
#>  [22,]  1.859956e-11  2.356614e-13  2.985893e-15  3.783206e-17  0.05696510
#>  [23,]  1.543796e-05  2.100439e-06  2.857789e-07  3.888214e-08  0.33423116
#>  [24,]  3.448006e-02  3.051772e-02  2.701072e-02  2.390673e-02  0.13639253
#>  [25,] -8.577438e-03  8.577403e-03 -8.577368e-03  8.577333e-03 -3.00089758
#>  [26,]  1.727301e-04 -3.174660e-05  5.834807e-06 -1.072398e-06 -0.83788308
#>  [27,]  1.488939e-02  6.516724e-03  2.852211e-03  1.248343e-03  1.03107708
#>  [28,]  4.836525e-03  2.451564e-03  1.242662e-03  6.298868e-04  0.16767601
#>  [29,] -4.277544e-07 -3.462439e-08 -2.802656e-09 -2.268598e-10  0.12350470
#>  [30,] -1.475998e-01  9.164199e-02 -5.689883e-02  3.532743e-02 -2.04069795
#>  [31,] -3.312101e-01  2.411971e-01 -1.756469e-01  1.279113e-01 -2.35971474
#>  [32,]  6.952715e-02 -4.959817e-02  3.538156e-02 -2.523994e-02 -0.53704153
#>  [33,] -4.451263e-01 -4.179955e-01 -3.925184e-01 -3.685941e-01  1.77319885
#>  [34,]  8.122532e-09  4.110955e-10  2.080626e-11  1.053041e-12  0.02449016
#>  [35,] -6.663991e-01  5.766400e-01 -4.989708e-01  4.317630e-01 -2.74053484
#>  [36,]  1.095470e-02  1.074197e-02  1.053337e-02  1.032882e-02  0.06161276
#>  [37,] -1.326958e-01  8.075541e-02 -4.914576e-02  2.990891e-02 -2.00327305
#>  [38,] -1.544274e-01  1.524714e-01 -1.505402e-01  1.486335e-01 -1.03740751
#>  [39,]  7.977996e-01  6.702879e-01  5.631563e-01  4.731475e-01  3.51397617
#>  [40,]  8.156029e-02 -5.994842e-02  4.406326e-02 -3.238736e-02 -0.56069288
#>  [41,]  4.770168e-05 -7.070773e-06  1.048094e-06 -1.553578e-07 -0.67405019
#>  [42,] -1.680755e-01  1.305067e-01 -1.013354e-01  7.868462e-02 -0.94497609
#>  [43,]  1.274791e-01 -1.023949e-01  8.224655e-02 -6.606281e-02 -0.64008239
#>  [44,]  1.741565e-03 -5.192837e-04  1.548352e-04 -4.616734e-05 -0.77416356
#>  [45,]  3.642060e-01  2.949900e-01  2.389283e-01  1.935209e-01  1.78148692
#>  [46,]  2.233494e-01  1.617941e-01  1.172035e-01  8.490208e-02  1.62418763
#>  [47,]  1.584430e-06  1.949068e-07  2.397624e-08  2.949410e-09  0.05667790
#>  [48,]  5.546711e-01 -5.411314e-01  5.279222e-01 -5.150354e-01 -2.85801360
#>  [49,] -6.183845e-01  5.223846e-01 -4.412880e-01  3.727811e-01 -2.68611165
#>  [50,]  6.374986e-01 -6.120377e-01  5.875936e-01 -5.641259e-01 -2.79352394
#>  [51,] -4.013451e-01  3.045905e-01 -2.311611e-01  1.754337e-01 -2.44808077
#>  [52,]  1.704626e-02 -6.795575e-03  2.709088e-03 -1.079991e-03 -1.84598015
#>  [53,]  1.398082e-01 -8.902669e-02  5.669017e-02 -3.609901e-02 -1.73186397
#>  [54,]  1.753288e-01 -1.711016e-01  1.669762e-01 -1.629503e-01 -0.90753813
#>  [55,]  3.256452e-01 -2.441662e-01  1.830739e-01 -1.372674e-01 -2.07680915
#>  [56,]  2.227949e-03 -6.288907e-04  1.775191e-04 -5.010892e-05 -1.29594216
#>  [57,]  1.397656e-01 -1.379874e-01  1.362318e-01 -1.344986e-01 -0.93709390
#>  [58,]  1.103023e-03  2.782988e-04  7.021636e-05  1.771598e-05  1.11489096
#>  [59,]  8.314472e-07  6.283486e-08  4.748612e-09  3.588663e-10  0.33825829
#>  [60,]  8.010868e-01 -7.882023e-01  7.755251e-01 -7.630518e-01 -4.86333440
#>  [61,]  4.982262e-01  4.676377e-01  4.389272e-01  4.119793e-01  1.98244560
#>  [62,]  4.746646e-02 -3.159084e-02  2.102497e-02 -1.399296e-02 -0.48704148
#>  [63,]  1.806667e-01 -1.758505e-01  1.711626e-01 -1.665997e-01 -0.90164222
#>  [64,]  8.764935e-02 -6.531422e-02  4.867062e-02 -3.626820e-02 -0.57202330
#>  [65,] -4.897303e-01  3.890515e-01 -3.090702e-01  2.455315e-01 -2.54832012
#>  [66,] -6.817758e-01  5.948769e-01 -5.190541e-01  4.528956e-01 -2.75938356
#>  [67,] -2.781252e-01 -1.900731e-01 -1.298975e-01 -8.877308e-02  2.55560795
#>  [68,]  2.132298e-01 -1.325712e-01  8.242336e-02 -5.124501e-02 -2.93056354
#>  [69,] -3.110307e-01  3.092764e-01 -3.075320e-01  3.057974e-01 -3.01671198
#>  [70,]  8.561037e-03  4.869356e-03  2.769598e-03  1.575295e-03  0.17485193
#>  [71,] -3.154212e-02 -1.671798e-02 -8.860883e-03 -4.696455e-03  0.88927774
#>  [72,] -5.843262e-04 -1.372767e-04 -3.225063e-05 -7.576693e-06  0.83999480
#>  [73,]  3.938952e-01  3.841517e-01  3.746493e-01  3.653818e-01  2.01971913
#>  [74,] -3.992402e-01 -2.929920e-01 -2.150193e-01 -1.577971e-01  2.76104741
#>  [75,]  6.467842e-01  5.146545e-01  4.095171e-01  3.258580e-01  3.34764030
#>  [76,]  1.418037e-07 -7.966370e-09  4.475416e-10 -2.514238e-11 -0.25381006
#>  [77,] -1.096726e-05 -1.325896e-06 -1.602952e-07 -1.937902e-08  0.42779899
#>  [78,]  1.478150e-08  8.271491e-10  4.628594e-11  2.590087e-12  0.02698194
#>  [79,]  1.343053e-01 -1.327596e-01  1.312318e-01 -1.297215e-01 -0.94070910
#>  [80,] -1.301900e-01  1.290565e-01 -1.279329e-01  1.268191e-01 -1.03296127
#>  [81,] -2.611615e-02  1.396818e-02 -7.470859e-03  3.995776e-03 -0.70619704
#>  [82,]  1.833886e-05 -3.291691e-06  5.908345e-07 -1.060505e-07 -0.10005743
#>  [83,]  2.588897e-07 -2.300230e-08  2.043750e-09 -1.815869e-10 -0.04694106
#>  [84,] -2.920804e-01 -2.015027e-01 -1.390143e-01 -9.590424e-02  2.58179480
#>  [85,]  1.131134e+00 -1.021098e+00  9.217658e-01 -8.320968e-01 -4.38580663
#>  [86,]  2.131467e-01  1.349102e-01  8.539074e-02  5.404766e-02  2.71017347
#>  [87,]  1.708381e-08  6.751591e-10  2.668256e-11  1.054505e-12  0.17734450
#>  [88,]  2.096066e-01 -1.958532e-01  1.830022e-01 -1.709944e-01 -0.82604221
#>  [89,]  2.272302e-09 -9.210088e-11  3.733030e-12 -1.513070e-13 -0.02078906
#>  [90,] -4.871247e-01  3.864862e-01 -3.066393e-01  2.432886e-01 -2.54548090
#>  [91,] -5.617774e-01 -5.551988e-01 -5.486972e-01 -5.422717e-01  3.90496808
#>  [92,]  9.168612e-03 -3.621209e-03  1.430222e-03 -5.648765e-04 -1.03843732
#>  [93,]  4.920564e-01  4.402760e-01  3.939445e-01  3.524886e-01  1.92137065
#>  [94,]  7.217123e-02  4.194051e-02  2.437269e-02  1.416358e-02  1.33810608
#>  [95,]  8.614490e-02  4.614129e-02  2.471439e-02  1.323762e-02  2.31393722
#>  [96,]  1.790697e-01 -1.744390e-01  1.699280e-01 -1.655337e-01 -0.90346965
#>  [97,] -7.231173e-01  6.480021e-01 -5.806895e-01  5.203693e-01 -2.81953630
#>  [98,] -9.068155e-02 -5.801340e-02 -3.711400e-02 -2.374363e-02  1.10098286
#>  [99,] -6.221521e-01 -5.025507e-01 -4.059413e-01 -3.279039e-01  3.06894699
#> [100,]  4.437766e-05 -9.213210e-06  1.912747e-06 -3.971038e-07 -0.11762476
#>                [,11]        [,12]
#>   [1,] -0.0494111540 -0.046811273
#>   [2,] -0.0095421034  0.012496258
#>   [3,] -0.0095423958  0.039036306
#>   [4,] -0.0056648134 -0.012114244
#>   [5,] -0.0361641475  0.007777032
#>   [6,] -0.0137938427  0.043311664
#>   [7,]  0.0227676261 -0.009618158
#>   [8,]  0.0486252397 -0.040446538
#>   [9,] -0.0374798555  0.049745073
#>  [10,] -0.0060145900  0.015265306
#>  [11,]  0.0122375009 -0.036571148
#>  [12,] -0.0151320067 -0.015919120
#>  [13,] -0.0069745122  0.031443569
#>  [14,]  0.0119806817 -0.044940477
#>  [15,]  0.0425165671  0.028938895
#>  [16,] -0.0198855986 -0.013595432
#>  [17,]  0.0341541310 -0.002368991
#>  [18,]  0.0139019090  0.030572272
#>  [19,]  0.0043720318  0.018436977
#>  [20,] -0.0362611223 -0.045953311
#>  [21,] -0.0091411981 -0.034806118
#>  [22,]  0.0375122381 -0.004200594
#>  [23,] -0.0205436346  0.030929957
#>  [24,]  0.0096704798  0.023326323
#>  [25,] -0.0074246488 -0.038856192
#>  [26,] -0.0395226781 -0.036005257
#>  [27,]  0.0140160184  0.005547783
#>  [28,]  0.0001349537 -0.010797826
#>  [29,]  0.0395929786 -0.029635492
#>  [30,] -0.0421812224 -0.009967849
#>  [31,] -0.0192060802  0.006457079
#>  [32,]  0.0329963851  0.030954550
#>  [33,] -0.0441801412  0.025088984
#>  [34,]  0.0359260298 -0.031934072
#>  [35,] -0.0215506210  0.035403466
#>  [36,] -0.0231140854  0.042859625
#>  [37,] -0.0290469873  0.034428535
#>  [38,]  0.0346599014 -0.034162961
#>  [39,] -0.0245407152  0.045406635
#>  [40,]  0.0438232428 -0.012260502
#>  [41,]  0.0190370204 -0.002252682
#>  [42,] -0.0157330000 -0.004865140
#>  [43,]  0.0136145061 -0.037104514
#>  [44,] -0.0211817049 -0.019189041
#>  [45,] -0.0296529576 -0.016947191
#>  [46,] -0.0385897858 -0.023474528
#>  [47,] -0.0114352393  0.046460505
#>  [48,]  0.0394679684  0.014047195
#>  [49,] -0.0379028531  0.002774406
#>  [50,]  0.0086608654 -0.021978159
#>  [51,]  0.0397832549 -0.001459893
#>  [52,] -0.0389684510  0.032680244
#>  [53,] -0.0107839450  0.006952650
#>  [54,]  0.0229901668 -0.049056897
#>  [55,] -0.0032016000  0.037429136
#>  [56,] -0.0188916723 -0.040186682
#>  [57,] -0.0406173033 -0.037823953
#>  [58,]  0.0218278701  0.013801423
#>  [59,] -0.0178925602 -0.014180800
#>  [60,]  0.0418970450 -0.016110678
#>  [61,] -0.0289520095 -0.045699345
#>  [62,]  0.0182308341  0.017281139
#>  [63,]  0.0420673159  0.022220930
#>  [64,] -0.0215416480 -0.048801365
#>  [65,] -0.0032879685 -0.007458667
#>  [66,]  0.0136172979  0.033425671
#>  [67,] -0.0128606171  0.035560294
#>  [68,]  0.0406865353 -0.046699327
#>  [69,] -0.0056567913 -0.021227714
#>  [70,]  0.0206802091 -0.028738201
#>  [71,]  0.0293639539  0.048813058
#>  [72,]  0.0099463716  0.025414831
#>  [73,]  0.0216768384  0.044960774
#>  [74,]  0.0126717287  0.032921567
#>  [75,]  0.0137911398 -0.004026652
#>  [76,] -0.0201560635 -0.014284641
#>  [77,] -0.0474466317  0.032963758
#>  [78,]  0.0034467781 -0.049464466
#>  [79,] -0.0451321898  0.036069776
#>  [80,]  0.0449525188 -0.020289299
#>  [81,]  0.0301143471 -0.036633093
#>  [82,]  0.0015380326  0.029914531
#>  [83,]  0.0313923658 -0.035608689
#>  [84,]  0.0142799772 -0.030768446
#>  [85,] -0.0342363236 -0.022653194
#>  [86,] -0.0437694831 -0.014886355
#>  [87,]  0.0163590239  0.003475857
#>  [88,]  0.0215670073 -0.021506445
#>  [89,]  0.0312772650 -0.031937469
#>  [90,]  0.0404797181  0.037160310
#>  [91,]  0.0197967365 -0.049801787
#>  [92,] -0.0082434738 -0.012772074
#>  [93,]  0.0110442976 -0.005877741
#>  [94,]  0.0086618911 -0.025761254
#>  [95,]  0.0378016111 -0.027342288
#>  [96,]  0.0267662517  0.035392174
#>  [97,]  0.0202962862 -0.018060863
#>  [98,] -0.0326327926  0.004448415
#>  [99,] -0.0398299208  0.047147391
#> [100,]  0.0465680600  0.017987990