Generate a spiral dataset with optional noise.

This function generates a dataset arranged in a spiral pattern with optional noise.

Usage

spiral_with_noise(sample_size, num_dims, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_dims

Number of effective dimensions for each data point.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

spiral_with_noise(sample_size = 100, num_dims = 10, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>                 [,1]          [,2]          [,3]          [,4]          [,5]
#>   [1,] -1.264462e-01  0.0286599852 -6.496004e-03  1.472369e-03 -3.337236e-04
#>   [2,]  1.417685e-01  0.0646788770  2.950837e-02  1.346257e-02  6.142014e-03
#>   [3,]  3.171968e-05  0.0000317181  3.171652e-05  3.171494e-05  3.171336e-05
#>   [4,]  1.106361e+00  0.8296372887  6.221280e-01  4.665211e-01  3.498346e-01
#>   [5,] -1.410193e+00 -0.6045472142 -2.591683e-01 -1.111050e-01 -4.763049e-02
#>   [6,] -1.853791e+00 -1.2339627657 -8.213784e-01 -5.467446e-01 -3.639366e-01
#>   [7,]  1.120094e+00  0.2945129474  7.743805e-02  2.036125e-02  5.353704e-03
#>   [8,] -1.986855e+00  1.7618250282 -1.562282e+00  1.385338e+00 -1.228435e+00
#>   [9,] -9.055340e-01  0.3370190972 -1.254308e-01  4.668248e-02 -1.737416e-02
#>  [10,] -3.986607e-01 -0.3896794099 -3.809004e-01 -3.723192e-01 -3.639313e-01
#>  [11,] -5.615241e-02  0.0059408861 -6.285417e-04  6.649928e-05 -7.035579e-06
#>  [12,]  9.539640e-01  0.2106703548  4.652376e-02  1.027416e-02  2.268912e-03
#>  [13,] -8.778392e-01 -0.6357028585 -4.603555e-01 -3.333746e-01 -2.414192e-01
#>  [14,] -7.854437e-02  0.0785172744 -7.849019e-02  7.846311e-02 -7.843605e-02
#>  [15,] -1.427137e-01 -0.1423393576 -1.419660e-01 -1.415937e-01 -1.412223e-01
#>  [16,]  1.227778e-01  0.0407223645  1.350660e-02  4.479808e-03  1.485842e-03
#>  [17,]  1.008630e+00  0.8397007780  6.990644e-01  5.819824e-01  4.845097e-01
#>  [18,]  1.876428e-01 -0.1872788165  1.869155e-01 -1.865529e-01  1.861910e-01
#>  [19,] -2.283981e-01  0.0207131111 -1.878444e-03  1.703535e-04 -1.544913e-05
#>  [20,] -1.773040e+00 -1.0575999465 -6.308475e-01 -3.762941e-01 -2.244555e-01
#>  [21,] -1.492141e-01  0.1473098783 -1.454300e-01  1.435740e-01 -1.417418e-01
#>  [22,]  1.532640e+00  1.4008247288  1.280346e+00  1.170230e+00  1.069584e+00
#>  [23,] -1.161028e+00  1.0316570547 -9.167014e-01  8.145550e-01 -7.237906e-01
#>  [24,]  1.114755e+00  0.8198121050  6.029054e-01  4.433881e-01  3.260760e-01
#>  [25,] -9.209352e-01 -0.8935631060 -8.670046e-01 -8.412354e-01 -8.162321e-01
#>  [26,] -1.209068e+00  0.3312774478 -9.076802e-02  2.486989e-02 -6.814200e-03
#>  [27,]  6.299907e-01 -0.4175039868  2.766860e-01 -1.833638e-01  1.215179e-01
#>  [28,] -2.184033e-01 -0.0311812102 -4.451709e-03 -6.355658e-04 -9.073906e-05
#>  [29,]  3.837673e-01 -0.1090800699  3.100437e-02 -8.812525e-03  2.504828e-03
#>  [30,]  1.445292e-01  0.0732038911  3.707770e-02  1.877982e-02  9.511963e-03
#>  [31,] -1.440995e+00 -0.6344648629 -2.793525e-01 -1.229978e-01 -5.415549e-02
#>  [32,] -4.350988e-02  0.0036085604 -2.992817e-04  2.482140e-05 -2.058602e-06
#>  [33,] -2.195912e+00  1.2496391258 -7.111386e-01  4.046914e-01 -2.302998e-01
#>  [34,] -1.467069e+00  0.4961235669 -1.677757e-01  5.673727e-02 -1.918703e-02
#>  [35,]  6.139321e-01  0.1617822041  4.263253e-02  1.123444e-02  2.960478e-03
#>  [36,]  1.217754e-01  0.0877850528  6.328221e-02  4.561868e-02  3.288545e-02
#>  [37,]  8.071244e-02  0.0806474198  8.058245e-02  8.051753e-02  8.045266e-02
#>  [38,] -7.359671e-01  0.1206112889 -1.976594e-02  3.239269e-03 -5.308558e-04
#>  [39,]  1.347227e-01  0.0884401205  5.805742e-02  3.811238e-02  2.501926e-02
#>  [40,]  5.928876e-01  0.5662831421  5.408725e-01  5.166022e-01  4.934209e-01
#>  [41,] -1.844440e+00  1.6761048334 -1.523133e+00  1.384122e+00 -1.257798e+00
#>  [42,]  1.103584e-01  0.0309984837  8.707139e-03  2.445741e-03  6.869823e-04
#>  [43,]  2.777905e-01  0.0315241888  3.577425e-03  4.059729e-04  4.607058e-05
#>  [44,]  1.406052e+00  0.4790724903  1.632304e-01  5.561611e-02  1.894961e-02
#>  [45,]  9.784700e-02  0.0784093195  6.283301e-02  5.035099e-02  4.034857e-02
#>  [46,]  2.414267e-01 -0.0168002672  1.169088e-03 -8.135381e-05  5.661203e-06
#>  [47,]  5.690667e-01  0.5632909507  5.575738e-01  5.519147e-01  5.463130e-01
#>  [48,] -1.365957e+00  0.9311496986 -6.347489e-01  4.326975e-01 -2.949625e-01
#>  [49,]  6.296473e-01 -0.4088774896  2.655150e-01 -1.724189e-01  1.119646e-01
#>  [50,] -2.328127e-02  0.0232748569 -2.326845e-02  2.326204e-02 -2.325564e-02
#>  [51,] -6.447845e-01 -0.2810610994 -1.225143e-01 -5.340391e-02 -2.327873e-02
#>  [52,]  5.940869e-01 -0.4625573020  3.601481e-01 -2.804121e-01  2.183294e-01
#>  [53,] -1.863433e+00  1.6884449146 -1.529889e+00  1.386223e+00 -1.256048e+00
#>  [54,]  3.347683e-02  0.0315936515  2.981641e-02  2.813914e-02  2.655623e-02
#>  [55,] -5.821099e-01  0.1353162882 -3.145540e-02  7.312069e-03 -1.699751e-03
#>  [56,] -8.587987e-01 -0.6910943796 -5.561390e-01 -4.475374e-01 -3.601433e-01
#>  [57,] -2.250125e-01  0.2243646544 -2.237186e-01  2.230745e-01 -2.224322e-01
#>  [58,]  5.776847e-01 -0.4625951831  3.704344e-01 -2.966345e-01  2.375373e-01
#>  [59,]  5.897854e-01 -0.3076096349  1.604375e-01 -8.367809e-02  4.364331e-02
#>  [60,] -1.852624e+00 -1.2304500111 -8.172230e-01 -5.427717e-01 -3.604905e-01
#>  [61,]  6.609521e-01  0.6233516618  5.878903e-01  5.544463e-01  5.229048e-01
#>  [62,]  6.085172e-01 -0.4585156769  3.454900e-01 -2.603256e-01  1.961544e-01
#>  [63,]  1.470955e+00 -1.2239354144  1.018398e+00 -8.473770e-01  7.050757e-01
#>  [64,] -1.064875e+00 -0.3313785983 -1.031218e-01 -3.209047e-02 -9.986239e-03
#>  [65,]  5.481955e-02 -0.0547396527  5.465987e-02 -5.458020e-02  5.450064e-02
#>  [66,] -8.334443e-01 -0.8135392187 -7.941096e-01 -7.751439e-01 -7.566313e-01
#>  [67,]  8.352489e-02  0.0700946573  5.882391e-02  4.936543e-02  4.142781e-02
#>  [68,]  1.410096e-01 -0.0135965012  1.311009e-03 -1.264107e-04  1.218884e-05
#>  [69,] -1.481026e+00 -0.6750284559 -3.076674e-01 -1.402300e-01 -6.391465e-02
#>  [70,]  1.626348e+00 -1.1373972494  7.954461e-01 -5.563004e-01  3.890523e-01
#>  [71,]  5.375079e-01 -0.2379697189  1.053558e-01 -4.664396e-02  2.065058e-02
#>  [72,]  3.316674e-01 -0.0796445684  1.912536e-02 -4.592646e-03  1.102850e-03
#>  [73,]  3.340485e-01 -0.3165446567  2.999580e-01 -2.842405e-01  2.693466e-01
#>  [74,]  9.725725e-01  0.8276204751  7.042720e-01  5.993074e-01  5.099868e-01
#>  [75,] -2.787365e-01 -0.0221088038 -1.753625e-03 -1.390939e-04 -1.103264e-05
#>  [76,] -1.693863e+00 -1.4498023304 -1.240907e+00 -1.062111e+00 -9.090769e-01
#>  [77,]  9.526736e-01  0.4339168115  1.976373e-01  9.001837e-02  4.100091e-02
#>  [78,]  4.514513e-01 -0.1566544953  5.435941e-02 -1.886282e-02  6.545436e-03
#>  [79,]  6.773254e-02  0.0102800764  1.560254e-03  2.368069e-04  3.594127e-05
#>  [80,] -2.317485e-01  0.0119074555 -6.118164e-04  3.143570e-05 -1.615196e-06
#>  [81,]  1.464435e+00 -0.7501224243  3.842327e-01 -1.968142e-01  1.008135e-01
#>  [82,]  1.565427e+00 -0.9146498561  5.344127e-01 -3.122473e-01  1.824402e-01
#>  [83,] -6.204549e-01 -0.1099771676 -1.949373e-02 -3.455311e-03 -6.124625e-04
#>  [84,] -3.436425e-01  0.3062673748 -2.729572e-01  2.432700e-01 -2.168116e-01
#>  [85,]  1.170525e+00  0.3231818504  8.923046e-02  2.463652e-02  6.802140e-03
#>  [86,]  1.631100e+00  0.6679952157  2.735686e-01  1.120364e-01  4.588302e-02
#>  [87,]  1.623245e+00 -1.0793271993  7.176655e-01 -4.771896e-01  3.172926e-01
#>  [88,] -2.791788e-01 -0.0221789870 -1.761980e-03 -1.399782e-04 -1.112038e-05
#>  [89,] -1.336758e-01 -0.0117399444 -1.031049e-03 -9.055085e-05 -7.952539e-06
#>  [90,] -1.161518e+00  1.0319204130 -9.167827e-01  8.144916e-01 -7.236138e-01
#>  [91,] -3.624887e-01 -0.0858094840 -2.031309e-02 -4.808580e-03 -1.138302e-03
#>  [92,] -8.321712e-02  0.0829097064 -8.260342e-02  8.229827e-02 -8.199425e-02
#>  [93,]  7.828830e-02  0.0140829927  2.533338e-03  4.557127e-04  8.197648e-05
#>  [94,] -1.246928e+00 -1.1721631118 -1.101881e+00 -1.035814e+00 -9.737073e-01
#>  [95,]  3.092498e-01 -0.2954405125  2.822478e-01 -2.696443e-01  2.576035e-01
#>  [96,] -3.696805e-01  0.2536060864 -1.739774e-01  1.193509e-01 -8.187643e-02
#>  [97,]  1.074565e+00 -0.3616368858  1.217062e-01 -4.095932e-02  1.378455e-02
#>  [98,] -1.063123e-01  0.0205037835 -3.954435e-03  7.626667e-04 -1.470907e-04
#>  [99,] -3.124103e-02 -0.0312372011 -3.123337e-02 -3.122954e-02 -3.122571e-02
#> [100,]  1.027362e+00  0.5288610838  2.722450e-01  1.401452e-01  7.214338e-02
#>                 [,6]          [,7]          [,8]          [,9]        [,10]
#>   [1,]  7.564100e-05 -1.714461e-05  3.885956e-06 -8.807816e-07 -0.129824916
#>   [2,]  2.802164e-03  1.278428e-03  5.832559e-04  2.660982e-04  0.159315049
#>   [3,]  3.171178e-05  3.171020e-05  3.170862e-05  3.170704e-05  0.003177476
#>   [4,]  2.623338e-01  1.967188e-01  1.475154e-01  1.106188e-01  1.672314968
#>   [5,] -2.041910e-02 -8.753630e-03 -3.752665e-03 -1.608761e-03  1.560901753
#>   [6,] -2.422518e-01 -1.612532e-01 -1.073370e-01 -7.144810e-02  2.484074075
#>   [7,]  1.407681e-03  3.701301e-04  9.732051e-05  2.558906e-05  1.160943594
#>   [8,]  1.089303e+00 -9.659294e-01  8.565287e-01 -7.595186e-01 -4.298062830
#>   [9,]  6.466264e-03 -2.406596e-03  8.956800e-04 -3.333516e-04 -0.975620993
#>  [10,] -3.557324e-01 -3.477182e-01 -3.398845e-01 -3.322274e-01  1.888773217
#>  [11,]  7.443594e-07 -7.875272e-08  8.331984e-09 -8.815182e-10 -0.056469339
#>  [12,]  5.010593e-04  1.106523e-04  2.443610e-05  5.396390e-06  0.978112892
#>  [13,] -1.748280e-01 -1.266048e-01 -9.168312e-02 -6.639396e-02  1.272924679
#>  [14,]  7.840899e-02 -7.838195e-02  7.835491e-02 -7.832788e-02 -2.990607081
#>  [15,] -1.408519e-01 -1.404825e-01 -1.401140e-01 -1.397465e-01  1.971755777
#>  [16,]  4.928171e-04  1.634553e-04  5.421409e-05  1.798148e-05  0.130144813
#>  [17,]  4.033621e-01  3.358055e-01  2.795634e-01  2.327411e-01  1.820629097
#>  [18,] -1.858298e-01  1.854693e-01 -1.851096e-01  1.847505e-01 -3.013971915
#>  [19,]  1.401061e-06 -1.270603e-07  1.152293e-08 -1.044998e-09 -0.229343126
#>  [20,] -1.338854e-01 -7.986128e-02 -4.763643e-02 -2.841464e-02  2.209062938
#>  [21,]  1.399329e-01 -1.381471e-01  1.363841e-01 -1.346436e-01 -0.936980376
#>  [22,]  9.775939e-01  8.935156e-01  8.166685e-01  7.464307e-01  3.777522958
#>  [23,]  6.431400e-01 -5.714760e-01  5.077975e-01 -4.512145e-01 -2.530920722
#>  [24,]  2.398025e-01  1.763553e-01  1.296950e-01  9.538018e-02  1.645121496
#>  [25,] -7.919720e-01 -7.684329e-01 -7.455935e-01 -7.234329e-01  3.805622903
#>  [26,]  1.867050e-03 -5.115604e-04  1.401645e-04 -3.840422e-05 -1.257178835
#>  [27,] -8.053166e-02  5.336951e-02 -3.536875e-02  2.343938e-02 -0.841252694
#>  [28,] -1.295472e-05 -1.849532e-06 -2.640557e-07 -3.769896e-08  0.220663773
#>  [29,] -7.119595e-04  2.023638e-04 -5.751886e-05  1.634887e-05 -0.400276827
#>  [30,]  4.817801e-03  2.440212e-03  1.235965e-03  6.260151e-04  0.167620482
#>  [31,] -2.384446e-02 -1.049862e-02 -4.622505e-03 -2.035271e-03  1.604935489
#>  [32,]  1.707334e-07 -1.416004e-08  1.174386e-09 -9.739952e-11 -0.043660295
#>  [33,]  1.310579e-01 -7.458181e-02  4.244266e-02 -2.415306e-02 -2.670498450
#>  [34,]  6.488541e-03 -2.194251e-03  7.420372e-04 -2.509372e-04 -1.558914270
#>  [35,]  7.801395e-04  2.055809e-04  5.417427e-05  1.427590e-05  0.636426984
#>  [36,]  2.370636e-02  1.708937e-02  1.231933e-02  8.880720e-03  0.175706032
#>  [37,]  8.038785e-02  8.032308e-02  8.025837e-02  8.019372e-02  2.011156408
#>  [38,]  8.699738e-05 -1.425725e-05  2.336497e-06 -3.829083e-07 -0.746053719
#>  [39,]  1.642415e-02  1.078180e-02  7.077823e-03  4.646309e-03  0.178591956
#>  [40,]  4.712798e-01  4.501322e-01  4.299336e-01  4.106414e-01  2.001677597
#>  [41,]  1.143003e+00 -1.038685e+00  9.438879e-01 -8.577426e-01 -4.419125160
#>  [42,]  1.929659e-04  5.420203e-05  1.522477e-05  4.276472e-06  0.114987792
#>  [43,]  5.228176e-06  5.933034e-07  6.732919e-08  7.640644e-09  0.279596647
#>  [44,]  6.456543e-03  2.199884e-03  7.495480e-04  2.553872e-04  1.495539214
#>  [45,]  3.233318e-02  2.591007e-02  2.076293e-02  1.663830e-02  0.163568797
#>  [46,] -3.939486e-07  2.741388e-08 -1.907661e-09  1.327493e-10 -0.242013399
#>  [47,]  5.407682e-01  5.352797e-01  5.298468e-01  5.244691e-01  4.004330766
#>  [48,]  2.010709e-01 -1.370666e-01  9.343595e-02 -6.369369e-02 -1.866960227
#>  [49,] -7.270708e-02  4.721419e-02 -3.065974e-02  1.990968e-02 -0.827973447
#>  [50,]  2.324924e-02 -2.324284e-02  2.323644e-02 -2.323004e-02 -0.992257597
#>  [51,] -1.014718e-02 -4.423149e-03 -1.928047e-03 -8.404345e-04  0.716430829
#>  [52,] -1.699917e-01  1.323559e-01 -1.030526e-01  8.023697e-02 -0.946724944
#>  [53,]  1.138097e+00 -1.031222e+00  9.343842e-01 -8.466396e-01 -4.404475949
#>  [54,]  2.506236e-02  2.365252e-02  2.232200e-02  2.106631e-02  0.101239769
#>  [55,]  3.951214e-04 -9.184926e-05  2.135113e-05 -4.963247e-06 -0.598505102
#>  [56,] -2.898153e-01 -2.332208e-01 -1.876780e-01 -1.510286e-01  1.446637662
#>  [57,]  2.217917e-01 -2.211531e-01  2.205163e-01 -2.198814e-01 -2.967270401
#>  [58,] -1.902138e-01  1.523184e-01 -1.219727e-01  9.767261e-02 -0.964470056
#>  [59,] -2.276269e-02  1.187215e-02 -6.192064e-03  3.229545e-03 -0.691251915
#>  [60,] -2.394255e-01 -1.590183e-01 -1.056145e-01 -7.014558e-02  2.478144836
#>  [61,]  4.931577e-01  4.651028e-01  4.386439e-01  4.136903e-01  1.987970208
#>  [62,] -1.478017e-01  1.113681e-01 -8.391547e-02  6.323002e-02 -0.925569368
#>  [63,] -5.866713e-01  4.881507e-01 -4.061748e-01  3.379653e-01 -2.651932612
#>  [64,] -3.107619e-03 -9.670605e-04 -3.009397e-04 -9.364947e-05  1.120510951
#>  [65,] -5.442121e-02  5.434189e-02 -5.426268e-02  5.418359e-02 -1.015705542
#>  [66,] -7.385607e-01 -7.209218e-01 -7.037041e-01 -6.868976e-01  3.836426231
#>  [67,]  3.476650e-02  2.917628e-02  2.448493e-02  2.054792e-02  0.153591370
#>  [68,] -1.175279e-06  1.133233e-07 -1.092692e-08  1.053601e-09 -0.141669732
#>  [69,] -2.913130e-02 -1.327759e-02 -6.051720e-03 -2.758279e-03  1.663905292
#>  [70,] -2.720862e-01  1.902853e-01 -1.330772e-01  9.306841e-02 -2.275335282
#>  [71,] -9.142587e-03  4.047678e-03 -1.792020e-03  7.933771e-04 -0.599458293
#>  [72,] -2.648316e-04  6.359504e-05 -1.527132e-05  3.667162e-06 -0.341664548
#>  [73,] -2.552331e-01  2.418591e-01 -2.291860e-01  2.171769e-01 -1.045677960
#>  [74,]  4.339784e-01  3.692983e-01  3.142582e-01  2.674212e-01  1.851706992
#>  [75,] -8.750862e-07 -6.941003e-08 -5.505460e-09 -4.366818e-10  0.279617473
#>  [76,] -7.780925e-01 -6.659810e-01 -5.700231e-01 -4.878912e-01  3.275591021
#>  [77,]  1.867479e-02  8.505858e-03  3.874186e-03  1.764586e-03  1.070119509
#>  [78,] -2.271279e-03  7.881382e-04 -2.734855e-04  9.490000e-05 -0.481360999
#>  [79,]  5.454970e-06  8.279258e-07  1.256581e-07  1.907170e-08  0.068526404
#>  [80,]  8.299032e-08 -4.264121e-09  2.190946e-10 -1.125729e-11 -0.232054974
#>  [81,] -5.163933e-02  2.645104e-02 -1.354893e-02  6.940120e-03 -1.705110522
#>  [82,] -1.065964e-01  6.228228e-02 -3.639037e-02  2.126221e-02 -1.928932294
#>  [83,] -1.085605e-04 -1.924262e-05 -3.410802e-06 -6.045730e-07  0.630437605
#>  [84,]  1.932308e-01 -1.722147e-01  1.534843e-01 -1.367911e-01 -0.757697615
#>  [85,]  1.878070e-03  5.185348e-04  1.431674e-04  3.952850e-05  1.217864889
#>  [86,]  1.879079e-02  7.695517e-03  3.151597e-03  1.290695e-03  1.787910696
#>  [87,] -2.109739e-01  1.402806e-01 -9.327530e-02  6.202055e-02 -2.173265997
#>  [88,] -8.834441e-07 -7.018405e-08 -5.575679e-09 -4.429524e-10  0.280063939
#>  [89,] -6.984239e-07 -6.133838e-08 -5.386983e-09 -4.731064e-10  0.134194323
#>  [90,]  6.428757e-01 -5.711461e-01  5.074198e-01 -4.508039e-01 -2.530408056
#>  [91,] -2.694625e-04 -6.378803e-05 -1.510010e-05 -3.574545e-06  0.373093108
#>  [92,]  8.169135e-02 -8.138957e-02  8.108890e-02 -8.078934e-02 -0.969038043
#>  [93,]  1.474645e-05  2.652684e-06  4.771815e-07  8.583841e-08  0.079586561
#>  [94,] -9.153248e-01 -8.604428e-01 -8.088515e-01 -7.603536e-01  3.656020175
#>  [95,] -2.461004e-01  2.351110e-01 -2.246123e-01  2.145824e-01 -1.046565018
#>  [96,]  5.616839e-02 -3.853231e-02  2.643371e-02 -1.813390e-02 -0.508090356
#>  [97,] -4.639089e-03  1.561251e-03 -5.254274e-04  1.768287e-04 -1.141129135
#>  [98,]  2.836845e-05 -5.471243e-06  1.055204e-06 -2.035105e-07 -0.108346466
#>  [99,] -3.122188e-02 -3.121805e-02 -3.121422e-02 -3.121039e-02  1.994769910
#> [100,]  3.713768e-02  1.911759e-02  9.841276e-03  5.066053e-03  1.198334246
#>                [,11]         [,12]        [,13]         [,14]
#>   [1,] -0.0423739981  0.0005641295  0.009125218  0.0259026846
#>   [2,]  0.0360534161  0.0012561879  0.009230198 -0.0481396592
#>   [3,] -0.0406324059  0.0173678077  0.003680050  0.0236935982
#>   [4,] -0.0352223045  0.0453709402 -0.036068859  0.0104190209
#>   [5,] -0.0255592588 -0.0311284318 -0.005730876 -0.0394405735
#>   [6,]  0.0157639137 -0.0049550921  0.010154759 -0.0380121459
#>   [7,]  0.0210674550  0.0080957928  0.041096133  0.0419023976
#>   [8,] -0.0435816205 -0.0266173788 -0.029051033  0.0150044224
#>   [9,] -0.0233161476  0.0215818671 -0.049052845  0.0272523994
#>  [10,] -0.0186572021  0.0019121361  0.049463993 -0.0047822804
#>  [11,]  0.0230429891  0.0049170736 -0.018322509  0.0287551134
#>  [12,]  0.0078257111  0.0075277892 -0.032712364 -0.0118214930
#>  [13,] -0.0144340040 -0.0293110288  0.008727631  0.0028810099
#>  [14,]  0.0097798355  0.0334458059  0.041756186  0.0286048706
#>  [15,]  0.0290009541  0.0072258456  0.046103125 -0.0311320245
#>  [16,]  0.0113489529  0.0076404400  0.012405603 -0.0122259753
#>  [17,] -0.0306449142  0.0189800955  0.033573566 -0.0318777867
#>  [18,]  0.0103233630  0.0088371196  0.038805186 -0.0011737513
#>  [19,]  0.0344990999  0.0112050680  0.008456710 -0.0369153217
#>  [20,] -0.0404158640 -0.0371579429  0.033159245  0.0419871304
#>  [21,] -0.0158153146 -0.0133850800  0.020406507 -0.0034047455
#>  [22,] -0.0381807014  0.0121491541  0.026402958  0.0425362405
#>  [23,]  0.0415786709 -0.0268756625  0.008696158 -0.0319545166
#>  [24,]  0.0334932922 -0.0492738383 -0.018786392  0.0414120165
#>  [25,] -0.0153042593  0.0410459159 -0.006616948 -0.0152989134
#>  [26,] -0.0032833889  0.0258638427  0.009389376 -0.0395568010
#>  [27,]  0.0045865014 -0.0245568391  0.034754460 -0.0289941242
#>  [28,] -0.0349629747  0.0308871629 -0.005155091  0.0473556313
#>  [29,] -0.0425756249 -0.0058567726 -0.033246600 -0.0104748802
#>  [30,] -0.0375505047 -0.0030661104  0.029984499  0.0294528896
#>  [31,]  0.0332036016  0.0431488342 -0.017278502 -0.0326594592
#>  [32,]  0.0148408001  0.0352013747 -0.035968679 -0.0491418596
#>  [33,]  0.0461352635  0.0324264881 -0.024265525  0.0130012729
#>  [34,] -0.0340987876 -0.0029234624  0.004271558 -0.0447945905
#>  [35,] -0.0122685367  0.0153300464  0.049098975 -0.0183883592
#>  [36,] -0.0372341619 -0.0138539880  0.008499671 -0.0085528377
#>  [37,] -0.0102585910  0.0396912986 -0.031015289  0.0092152898
#>  [38,]  0.0049165524  0.0123364045 -0.013058185  0.0054035998
#>  [39,] -0.0036641160 -0.0215964957 -0.020049325 -0.0249077035
#>  [40,] -0.0406942968  0.0486775603 -0.048080916  0.0335447852
#>  [41,] -0.0219196065 -0.0196659999 -0.015563637  0.0225848536
#>  [42,]  0.0339932150  0.0434134995  0.046950515 -0.0363274698
#>  [43,]  0.0129231562 -0.0009032456  0.032615416 -0.0309964480
#>  [44,]  0.0114566615 -0.0223315342 -0.012023662 -0.0264314575
#>  [45,] -0.0342428840 -0.0229795506  0.046101671 -0.0184312303
#>  [46,] -0.0036162026 -0.0162010657 -0.011964851  0.0154507402
#>  [47,]  0.0416055099  0.0021083007  0.040049632 -0.0013594597
#>  [48,] -0.0359464856  0.0127213856  0.008459823  0.0488755484
#>  [49,]  0.0054651178 -0.0038775178  0.032150785  0.0225479410
#>  [50,]  0.0451698213 -0.0392241852  0.007593704  0.0119685331
#>  [51,]  0.0353342006  0.0338103729  0.008126795  0.0381696011
#>  [52,] -0.0292908858  0.0195843794 -0.037365302  0.0059889935
#>  [53,] -0.0290128477 -0.0458790026 -0.018176880 -0.0074502126
#>  [54,] -0.0097823656 -0.0332694022  0.040167015  0.0128912321
#>  [55,] -0.0144400492 -0.0398796815 -0.015291461 -0.0316742466
#>  [56,]  0.0270458351  0.0442972293  0.043295570 -0.0207969493
#>  [57,] -0.0434133501 -0.0164187283 -0.043451531  0.0116174912
#>  [58,]  0.0146367602 -0.0089219590  0.008097974 -0.0410736010
#>  [59,]  0.0469097281  0.0481089856 -0.041742355  0.0358389253
#>  [60,]  0.0268055704 -0.0306078838  0.017151298 -0.0108867611
#>  [61,] -0.0122563329  0.0203989227  0.018126565  0.0264094083
#>  [62,]  0.0379149776 -0.0425984581  0.005658895  0.0494566324
#>  [63,]  0.0094532959 -0.0059796841 -0.005071636 -0.0340290352
#>  [64,]  0.0065051951  0.0395742948  0.049280810 -0.0335202447
#>  [65,]  0.0024463860  0.0443061970 -0.034242105 -0.0065991141
#>  [66,]  0.0193318980  0.0050130781 -0.032253226  0.0143277417
#>  [67,]  0.0344214538  0.0377495332 -0.031986658 -0.0191993365
#>  [68,] -0.0428117999  0.0417571197  0.028149178 -0.0347360489
#>  [69,] -0.0199536927  0.0418732466  0.027342555  0.0414403909
#>  [70,] -0.0475057746 -0.0368594302  0.036163947 -0.0110333550
#>  [71,]  0.0276388028 -0.0085194062  0.015944307 -0.0197400030
#>  [72,]  0.0357076274  0.0229601533  0.005095061  0.0017181385
#>  [73,] -0.0231309286 -0.0479159137  0.028314006 -0.0165427397
#>  [74,]  0.0415386613 -0.0221352559  0.015396115  0.0232373388
#>  [75,] -0.0348869239 -0.0448330908  0.047606187  0.0106966854
#>  [76,]  0.0253965649  0.0290720587  0.015670469 -0.0430178287
#>  [77,] -0.0479325875  0.0487698218 -0.049496241 -0.0271816092
#>  [78,] -0.0001826605 -0.0490405730 -0.038765256  0.0096133667
#>  [79,] -0.0109943578  0.0340992453 -0.033347206 -0.0383510709
#>  [80,] -0.0307794619  0.0225772991  0.014908686 -0.0347243825
#>  [81,] -0.0404083670 -0.0291443344  0.011487250 -0.0466620282
#>  [82,] -0.0127275979 -0.0009079289  0.011397636  0.0008147368
#>  [83,] -0.0359342281 -0.0412632473  0.037161177 -0.0178322429
#>  [84,] -0.0487924160 -0.0006198183  0.039001489 -0.0013549264
#>  [85,]  0.0319013681 -0.0295031329  0.009417408 -0.0139003184
#>  [86,]  0.0306152622 -0.0404827507 -0.014109675 -0.0281517614
#>  [87,] -0.0040673974 -0.0214823572  0.048997115 -0.0043045224
#>  [88,] -0.0157591008 -0.0369237602 -0.047615878  0.0126707347
#>  [89,] -0.0274727595 -0.0237149512  0.007065461  0.0206500478
#>  [90,] -0.0017752889  0.0227769018  0.010523320 -0.0239284310
#>  [91,] -0.0045398859  0.0351648150 -0.029063431  0.0212146054
#>  [92,]  0.0099924972 -0.0009740921 -0.049732513 -0.0236133712
#>  [93,] -0.0158017890  0.0368753227  0.007839739 -0.0449493300
#>  [94,]  0.0313287776  0.0190020639  0.016910303  0.0062411134
#>  [95,] -0.0383533302 -0.0059064964 -0.033586488  0.0027902156
#>  [96,] -0.0440198813  0.0262138775  0.017675678  0.0197057987
#>  [97,] -0.0117220946  0.0469231449  0.020838287  0.0021704347
#>  [98,] -0.0081496359 -0.0045200395  0.013948086 -0.0310464339
#>  [99,]  0.0130012767 -0.0140521320 -0.023699761  0.0317625453
#> [100,]  0.0424385834 -0.0423046128 -0.004630538  0.0074610829