Generate three clusters of data points with optional noise.

This function generates three clusters of data points along with optional noise.

Usage

three_clust_diff_dist(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate, should be a multiple of three.

num_dims

Number of dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
three_clust_diff_dist(
  n = 150, num_dims = 7, num_noise = 4, min_n = -0.05,
  max_n = 0.05
)
#>               [,1]          [,2]        [,3]        [,4]          [,5]
#>   [1,]  1.31776542  2.518759e+00 -1.85375269 -1.26478804  0.3742474090
#>   [2,] -0.64215864  2.415200e-01  1.22846119  1.42112767 -0.7094525428
#>   [3,] -1.29205633  1.087724e+00  0.19619270 -3.25512761 -2.6980107273
#>   [4,]  1.28481192 -6.058001e-01 -0.24347578  0.03920695  0.6070894325
#>   [5,]  0.37750101  1.149371e+00  2.80074196 -0.78693252 -1.0740123669
#>   [6,] -0.15414403 -3.117244e-01 -1.10634541 -1.05129798 -0.2766678320
#>   [7,] -1.53212609  1.757253e-01  0.89436126 -0.13062107 -0.8740353806
#>   [8,]  0.39068303 -1.440703e+00  0.80430281 -0.21874566 -0.7786403967
#>   [9,] -0.46446764 -1.000813e+00 -0.05941675 -0.66099202 -0.1163250336
#>  [10,]  2.97923939 -1.098109e-01  0.47995348  0.60480246  0.3480936761
#>  [11,]  0.36208825  2.311786e-01 -1.02815347  0.40659882  0.1768020291
#>  [12,] -0.02178079  5.401011e-01  0.87531808  0.54813103  0.3456633048
#>  [13,]  0.25381815  1.871959e-01 -1.21891581  0.98715260 -0.0328435813
#>  [14,] -0.07235334 -3.588870e-01  0.13951356  1.61994869  0.8429954414
#>  [15,] -0.50962033 -1.258695e+00 -0.82867179 -0.41511572 -0.6572750459
#>  [16,]  0.38011146  5.017002e-01  0.16256823  0.46588014  1.8879919733
#>  [17,] -0.31163667  1.570077e+00  0.67729261  0.09399589  0.9822722288
#>  [18,]  0.34093740  5.695432e-01  1.05389569  0.16045935  0.4770100225
#>  [19,]  0.40795937  8.080527e-01  1.01222747 -1.50224119 -1.3184811640
#>  [20,]  0.98013776  6.601062e-01  1.68562886  1.07995819  0.1119199568
#>  [21,]  1.33750733 -4.969000e-01 -0.05649640 -2.68578408  0.1662179123
#>  [22,] -0.99702040 -3.481735e-01 -1.13134545 -0.40615679 -1.5246710544
#>  [23,] -0.15122674  8.292891e-01 -0.43845544  1.38809682  1.0049624937
#>  [24,] -1.32763833  2.906811e-01  1.80682282 -0.36760678 -0.1874665552
#>  [25,]  0.34765139  1.170144e+00  0.06960019 -0.12698011 -1.4586827245
#>  [26,]  0.37224577 -1.269298e+00 -0.03660224  0.01806813 -1.6045039864
#>  [27,]  0.11038148  2.736923e-01  0.17516878 -1.00866677 -0.3164936390
#>  [28,]  0.13569629  8.033855e-02  1.14691373 -0.04426141  0.0762963957
#>  [29,] -0.87850811  1.317551e+00 -1.16623130 -0.39525897 -0.2365843199
#>  [30,]  1.25457960  4.646965e-01 -0.36329571 -0.01764468 -0.1798278646
#>  [31,] -0.19569676  8.047572e-01 -1.71681432  0.86710798  0.0861116234
#>  [32,] -0.89547223 -1.409614e+00  0.58941725  1.01339748 -0.5646681601
#>  [33,]  0.36261837  8.251067e-02 -0.18806377  0.06035644 -0.9361258474
#>  [34,]  0.66789796  3.281640e-01  1.06077068 -1.47305103  0.2277917159
#>  [35,]  0.27089658  1.422732e-01 -0.56004944 -1.86998429 -0.1778697963
#>  [36,] -0.42460261 -8.191148e-01  2.26132992 -1.10338833  0.6954385209
#>  [37,] -0.21678152 -5.732680e-01  1.64106216 -0.80004882  0.7044163235
#>  [38,]  1.00030819 -3.528816e-01 -0.10110520 -1.53002929 -1.5376298701
#>  [39,]  0.54672259  1.950913e+00  0.95273404 -0.81126146  0.1117561164
#>  [40,] -0.87667248  1.657808e+00  1.08345957 -0.50958487  1.8207524374
#>  [41,] -0.63083642 -8.457127e-01 -1.10921339 -0.24610479 -0.7065635329
#>  [42,] -1.24992306 -8.924483e-01 -1.06290240  1.10981918  1.6248401399
#>  [43,]  1.90618125  1.004218e+00 -0.60235762 -0.13545753 -0.3091883899
#>  [44,]  1.26734149 -4.222816e-01  0.08723732  0.15344166  0.9820494324
#>  [45,] -1.26698095 -8.881600e-01 -0.91298464  0.48944414 -0.2720288435
#>  [46,]  0.80728113 -5.185633e-01 -2.88099249 -1.40087570  1.7813951155
#>  [47,] -0.17772803 -2.804738e-01 -0.14787024  0.27864422  0.5329313378
#>  [48,]  1.09853842 -1.165127e+00 -0.75016680 -0.86039010  0.5723157398
#>  [49,]  0.32993219 -5.329412e-01 -0.46940140 -0.10218612  0.1296947477
#>  [50,]  0.70586598  1.271920e+00  1.90291080  0.98326380 -1.5310246290
#>  [51,]  9.18409764 -3.568029e-01  1.06960190  0.99075162  0.9577609963
#>  [52,]  9.14090037 -7.514903e-01 -0.74231963 -1.14624391  0.3252229615
#>  [53,]  8.84994945 -6.732295e-01 -1.08058524  0.78760273 -0.0678428576
#>  [54,] 10.91731904  1.776775e+00 -0.47290249  0.38774472  0.6913809719
#>  [55,] 10.73101557  6.430436e-01  0.09612179 -0.73466187 -0.5637098676
#>  [56,] 11.26106838 -1.400956e-01 -0.93495145  0.92920848 -1.0279571113
#>  [57,]  9.06170351  1.525287e+00  1.06221428  0.25790760 -0.1894826617
#>  [58,] 11.49351734  6.051725e-01 -1.62756968 -0.23601985  0.5642801869
#>  [59,]  9.41201151  1.615340e+00 -1.08496242  0.36500145  0.4486977171
#>  [60,]  8.35822094 -6.437751e-01 -0.09128764 -0.63030432  1.3986649687
#>  [61,] 10.20328578 -2.419653e+00 -0.85941315  0.59074187 -0.8837868527
#>  [62,] 10.63091669  1.630073e-01 -1.11356064 -0.12328239  0.9663841764
#>  [63,]  8.21326080 -1.014585e+00 -0.23126989 -0.04125274 -1.1924407387
#>  [64,] 10.10707481  6.163477e-01 -0.77767403 -1.75527616 -0.2105711704
#>  [65,]  8.48158503 -4.308625e-03 -0.69663200  1.31925111  0.0009358841
#>  [66,] 10.79505423 -6.347211e-01  0.12907914 -0.03462962 -0.4941604747
#>  [67,]  8.64400551 -2.104941e+00 -0.29477063 -0.06347348  0.6655462871
#>  [68,] 11.30196363  5.096046e-01  0.52017540  2.83883124 -0.0531421842
#>  [69,]  9.50016902  6.609159e-01 -0.13348941 -0.46264207 -2.4961099700
#>  [70,]  9.82600847  5.499856e-02 -0.50680048  1.15801659 -0.2669197209
#>  [71,]  8.96937563 -6.176911e-01 -0.67046573 -0.54615670 -1.2240962028
#>  [72,]  9.62103777  1.127895e+00  0.64019608 -0.91259786  1.0450283491
#>  [73,] 10.88339398 -1.341959e+00 -0.49318492  1.24155111 -0.7315958295
#>  [74,] 10.32975938  8.615079e-01  1.71891939 -1.24182669 -0.1923923478
#>  [75,]  8.88612791  1.017134e+00 -1.45871520  0.37291896  0.5589554187
#>  [76,]  8.83693374  8.547226e-02 -0.26015390 -0.54903289  0.3440633341
#>  [77,]  9.72982975  2.782857e-01 -0.46239084  1.23962998 -0.6151790883
#>  [78,]  9.88924460  5.755904e-01 -0.83088338  1.08112593  2.1188470158
#>  [79,]  9.83274327 -1.574677e+00 -0.58612260 -2.05287050  1.1264190543
#>  [80,] 12.61687313  2.597974e+00  1.00729041 -0.97559678  0.1259205617
#>  [81,]  9.75101732 -3.990989e-01 -1.70102570  0.69410665  0.9879903361
#>  [82,] 11.28296252 -4.220198e-01 -0.76609835 -0.72483790 -1.1463377392
#>  [83,]  9.85685332 -2.072662e+00  0.54730413  0.54391142 -0.0365983249
#>  [84,] 10.20371593  1.812266e-01 -1.40699610  0.13228204 -1.2936261616
#>  [85,] 10.11989067  6.447639e-01  1.89822210 -0.69795674  0.3840011799
#>  [86,]  8.90337525 -1.432477e+00  0.93033721  0.09610427  1.9132326731
#>  [87,]  9.39057379  5.214822e-01  1.80585662  0.41859006  0.1723712919
#>  [88,] 11.07792551 -6.197096e-01  0.55811928 -0.14041617 -0.9780641500
#>  [89,] 10.14578287  9.158488e-01  0.39793331 -1.24367806 -1.4332648175
#>  [90,]  9.74271493  8.910413e-01  0.72572200  0.62579125 -0.4286249220
#>  [91,] 10.29303095  3.901533e-01  1.50586513 -0.72937741  0.4372471487
#>  [92,] 10.20883406 -5.682941e-01  0.15040459 -1.50293132 -0.8023827994
#>  [93,]  8.59998860 -1.211355e+00 -0.31862273 -1.97645703 -1.5052685199
#>  [94,]  9.85841932 -6.783993e-01 -1.34953487 -1.15160617  0.1750359599
#>  [95,]  9.62338584 -5.229855e-01  1.51622384  1.16446132 -0.8355854632
#>  [96,]  9.94356798  1.315725e+00  0.11599080  0.04254069  1.1089939192
#>  [97,] 10.62739951 -1.591715e+00  0.14945202 -0.26853925 -2.0982864312
#>  [98,] 10.48132049  1.382210e+00 -0.85817223  0.23202649  0.3086705127
#>  [99,] 11.02637067  2.180603e-01 -0.85882555 -0.71517839  0.3208197103
#> [100,] 10.99137620 -3.857370e-01 -0.29017795  0.48558076 -0.3965175012
#> [101,] 50.69663813 -1.710523e+00 -0.35470760  1.22187360  0.6634110908
#> [102,] 51.58753482  2.255251e-01 -0.31016750 -0.29219493 -0.2285120407
#> [103,] 50.79599173 -5.453890e-01  0.10296669 -0.05381596 -0.3556388089
#> [104,] 50.20641936  1.382573e+00 -0.02581438  1.14936904  0.7116233387
#> [105,] 50.02519359 -1.699227e+00 -0.58947499  0.65947365 -1.0514494571
#> [106,] 49.69997529 -4.505332e-01 -0.96313334 -0.92573299 -0.4991291317
#> [107,] 47.38092298 -1.478195e+00  0.12095458 -1.31217766 -1.7471356779
#> [108,] 49.72711044 -5.063198e-01 -1.22796433  0.03889702 -0.6922318198
#> [109,] 51.07692394 -4.984257e-01 -0.70298333 -1.36741940 -0.9596941788
#> [110,] 48.09267844  1.196466e+00 -1.02547741  1.97490548 -0.3084846082
#> [111,] 47.83980712  1.093573e-01 -0.35615932  1.95372719  0.1623481510
#> [112,] 50.31743120 -1.238861e+00  1.41587657  1.90844265  0.5622499111
#> [113,] 50.47233500 -1.105677e-01  1.46896176  0.10602155  1.3060820759
#> [114,] 50.70973536  7.601924e-01  0.30732427  1.50140835 -0.3878979787
#> [115,] 50.53170236  1.275880e+00  0.96892293 -0.34477555  0.6035411834
#> [116,] 50.24789689 -7.444199e-01  1.45925548  1.64603792  0.0028051758
#> [117,] 51.41507790 -1.770175e+00 -0.48586949 -0.37346310 -2.2851614624
#> [118,] 50.23895792 -7.869870e-01  0.22992404  1.10957139 -0.3941407679
#> [119,] 51.68941547 -2.116944e-02 -0.01644311  0.15852576  0.5467078562
#> [120,] 51.10747957 -3.454935e-01 -0.52101090  1.13265562  1.4749959860
#> [121,] 49.49991065 -7.134455e-01 -0.50636150  0.74242927  0.0244960757
#> [122,] 49.24369435 -1.291203e+00  0.58935694  0.96739381  0.2366314179
#> [123,] 51.41121163  3.737334e-01 -0.83618890 -0.89848940  1.8481408976
#> [124,] 51.06409707 -8.232048e-01 -0.21506517 -1.81666822 -0.2206273680
#> [125,] 48.75220593  1.658248e+00  0.35294427 -0.42902620 -0.1707835837
#> [126,] 50.40901016  1.167237e+00  1.17681429 -0.98794168  0.0958144168
#> [127,] 50.13030393  1.386960e+00  0.81102927 -0.61709655 -0.0541721669
#> [128,] 50.53596672  8.283560e-01 -0.11397341  1.29876681  0.6048036327
#> [129,] 50.68245983 -1.309286e+00  0.83382841 -0.45999188  0.2193538507
#> [130,] 49.55394509 -2.448510e-01  1.08608631 -0.10522716  1.6126686063
#> [131,] 49.06426852 -1.370963e+00  0.47798741  0.93787772 -1.1037794335
#> [132,] 51.00776448  3.066630e-01  2.50429682 -0.59122024 -0.0087458078
#> [133,] 51.20798861  1.308100e+00 -1.72303521  0.65776316 -0.2347085366
#> [134,] 50.98388805 -1.687133e+00  0.58381393  0.67574519 -0.6904383968
#> [135,] 49.72646611 -4.127677e-01  0.45011670  3.26899970 -0.0726797508
#> [136,] 48.49904911  9.799784e-01  0.06593084  0.69625594 -0.2222135910
#> [137,] 50.17860423  3.568273e+00  1.81977978  0.01951456 -1.2208925318
#> [138,] 49.56234259 -9.382244e-01  0.95628055  0.32330831 -1.3113881272
#> [139,] 48.66347058  1.395693e+00  0.59592939  1.41994529 -0.4259739955
#> [140,] 50.63361175 -8.097545e-01  0.35644751  0.75726362  0.7634773523
#> [141,] 50.58683856  6.814301e-01  0.17691364 -0.64108086 -0.8156474184
#> [142,] 48.89298040  9.717303e-01  0.35283187 -0.88495954 -0.4587061735
#> [143,] 49.53712247 -8.605036e-01 -1.02938114 -0.41751060  0.0266629837
#> [144,] 48.52408556 -1.171053e+00 -0.35075130 -2.63365600  0.1522987628
#> [145,] 50.11265711 -3.623071e-05  0.87202523  0.17162533  0.1607421974
#> [146,] 50.90022285  4.639439e-01 -0.04328609  0.25696382 -0.2828802881
#> [147,] 48.30608553  1.040177e-01 -1.25244530 -0.40728307 -0.6878854907
#> [148,] 50.49327785  1.849556e+00 -1.19020105  0.36447300 -0.2098288464
#> [149,] 50.36801908  6.393509e-01  0.86879893  0.92420494  0.4933220134
#> [150,] 49.77643708 -6.297742e-01 -0.09255716 -0.65509917  1.2475554032
#>               [,6]         [,7]          [,8]          [,9]         [,10]
#>   [1,] -0.24811091 -0.309828911 -0.0116275290 -0.0287596528 -1.636529e-02
#>   [2,] -0.94519467 -1.122884387  0.0112091000  0.0148003248  2.224537e-03
#>   [3,]  0.86664607 -1.687780308  0.0418783873  0.0270577303 -1.341185e-02
#>   [4,]  0.47871207  1.139767500 -0.0026603673  0.0345774974 -1.857520e-02
#>   [5,] -0.32608382  1.038649878 -0.0372542120 -0.0293564787 -4.569429e-02
#>   [6,]  0.24316460 -1.939457832 -0.0433450281  0.0333989013  4.582746e-02
#>   [7,]  1.39256372  1.701123210  0.0192403497 -0.0046153712 -3.455254e-02
#>   [8,] -0.49045976 -0.390406045 -0.0338090247  0.0422258466  2.582430e-02
#>   [9,] -0.19894611 -0.052424602 -0.0403875436 -0.0463215775  1.039657e-02
#>  [10,] -0.79848101  0.267111358  0.0028063231  0.0228497851 -1.376949e-02
#>  [11,] -0.76825427 -0.510641137  0.0298251742  0.0156579840 -4.073165e-02
#>  [12,] -0.13026420 -0.281909472  0.0032823595  0.0339446843  4.847071e-02
#>  [13,] -0.53236786 -0.417625372 -0.0024434743  0.0309171630 -2.970373e-02
#>  [14,]  0.07170660  0.753780388 -0.0375077707  0.0162746326  6.991699e-05
#>  [15,]  0.42300917 -0.980682498  0.0466357397 -0.0207092173  2.383065e-02
#>  [16,] -0.53381767  0.529248919  0.0332118956 -0.0239421536 -4.223552e-02
#>  [17,] -0.44348316  0.493036378 -0.0444718203  0.0009716894  1.261618e-02
#>  [18,]  0.88880572 -1.256233083 -0.0162042385 -0.0274106946  4.623033e-02
#>  [19,]  0.70598859  0.301425910 -0.0297408582  0.0380198223 -2.674662e-02
#>  [20,] -0.90622907 -1.378956290 -0.0499447828  0.0485247135  4.767000e-02
#>  [21,] -0.36663995 -1.093791351 -0.0132112680 -0.0179583907 -2.139238e-02
#>  [22,] -1.88770405 -0.460690146  0.0102981293 -0.0171249152  6.776200e-03
#>  [23,]  1.10850529 -0.408813868  0.0323907212  0.0204849507  2.578724e-02
#>  [24,]  0.19772643  0.809735359 -0.0481277417 -0.0179980048  4.627391e-02
#>  [25,] -0.96862054  0.203521657  0.0353159780  0.0350112810  4.988278e-03
#>  [26,]  0.03688506 -0.707223727 -0.0290311078  0.0322723470  1.933860e-02
#>  [27,] -1.04507740 -1.665458068 -0.0370387824  0.0409875852 -2.677462e-02
#>  [28,]  0.72634748  1.202293405 -0.0270279932 -0.0472463219 -1.975191e-02
#>  [29,] -1.07818655  0.592660374 -0.0415069370 -0.0134920967  4.798282e-02
#>  [30,] -2.24259619  0.321025296 -0.0251418293 -0.0032567904 -7.112537e-03
#>  [31,]  0.12305753 -0.632060284  0.0476042713 -0.0285329087  8.021953e-03
#>  [32,] -0.22047577  0.030947571 -0.0041391176 -0.0335137849  4.239841e-03
#>  [33,]  0.53009379  0.150700689 -0.0303798981 -0.0423664921  2.394034e-02
#>  [34,] -0.85179755  1.763878698  0.0060540041  0.0134145463  3.145418e-03
#>  [35,]  1.46843459  1.198040087  0.0210924581  0.0458375993 -2.163410e-02
#>  [36,] -0.16787861 -1.548645722 -0.0330376656  0.0167461503 -2.126563e-03
#>  [37,] -0.83602877  0.477990684 -0.0292480736  0.0239781460  2.195558e-02
#>  [38,]  0.29181967  0.521615248  0.0435781101 -0.0402660282  3.284489e-03
#>  [39,]  1.03478639 -1.422930125  0.0143819636  0.0194831758  4.787364e-02
#>  [40,]  0.28129101 -0.558298819  0.0321920014 -0.0450241640  4.729554e-02
#>  [41,] -0.04315316  0.501719943  0.0242666359  0.0295578158 -1.139702e-02
#>  [42,]  0.23968569 -1.399559545 -0.0249976477 -0.0347286321 -1.521704e-02
#>  [43,]  1.13173560  1.592338358  0.0177244761  0.0123693719  2.190798e-02
#>  [44,] -0.35416030 -0.575564447 -0.0044021047  0.0065094367 -1.470256e-02
#>  [45,] -0.35661619 -1.336092456 -0.0369394224  0.0333256996 -2.105489e-02
#>  [46,] -0.34167176 -1.547370932 -0.0144854619 -0.0418847026  4.948568e-02
#>  [47,] -0.30297090 -1.419208973  0.0087791027 -0.0047995879 -4.643196e-02
#>  [48,] -0.09561403 -1.473249289  0.0249628162 -0.0098169956  2.906671e-02
#>  [49,]  0.09144115  0.678003376 -0.0015996212 -0.0229965936 -4.218292e-02
#>  [50,] -0.70172655 -0.062431958 -0.0415833517  0.0450571919  4.991574e-02
#>  [51,] -0.17302248 -2.149297571 -0.0380366813 -0.0405685086 -4.200190e-02
#>  [52,]  0.61921271 -0.195891297 -0.0465834866 -0.0084990631 -1.333662e-02
#>  [53,]  0.17287409 -1.527919524  0.0040008845  0.0491739614 -2.654866e-02
#>  [54,]  0.73277358 -0.546765233 -0.0076150128  0.0305278162 -2.783990e-02
#>  [55,]  0.10484993  0.699412952  0.0203057111  0.0144671857 -4.780630e-02
#>  [56,]  0.91281030  0.930485330  0.0198581789 -0.0268178170  4.970827e-02
#>  [57,]  1.04804298  0.773861899 -0.0476501956  0.0321518205  3.006407e-03
#>  [58,]  0.27152918  0.374277597  0.0125901363 -0.0151065914  8.176925e-03
#>  [59,]  1.68714529  0.203355429 -0.0255737485 -0.0183313074  4.956267e-02
#>  [60,]  0.04382268  0.237516502  0.0171372454 -0.0431845203 -1.627479e-02
#>  [61,]  1.70955480  1.310140928 -0.0290295028 -0.0349900791 -1.695537e-03
#>  [62,]  0.40672032 -0.613999543  0.0059841295 -0.0414125434 -4.294849e-02
#>  [63,] -2.24820471  0.655510396  0.0222604845  0.0013627960 -1.343644e-02
#>  [64,]  0.81121101  0.384506988  0.0109771356  0.0225866848 -2.834701e-02
#>  [65,]  0.45282886 -0.784635745 -0.0469041090 -0.0424938873  2.343479e-02
#>  [66,]  0.90886882  2.485051842  0.0255797171  0.0220835434 -3.167909e-02
#>  [67,]  0.95430388  0.180691104 -0.0459549752 -0.0011383740  1.233606e-02
#>  [68,]  0.05908066 -0.915880758  0.0267173836  0.0408576883 -1.595814e-02
#>  [69,]  1.59347272  1.287958114 -0.0284516421  0.0411632049 -2.958210e-03
#>  [70,]  1.27307185 -0.382666354  0.0375702585  0.0255098755  3.158711e-02
#>  [71,]  0.64358355  0.833042404  0.0367650446  0.0493289073  1.725153e-02
#>  [72,] -0.87063338 -0.484250677 -0.0309639698  0.0192735269  3.596430e-03
#>  [73,] -0.34435709  0.033576976 -0.0037136571 -0.0428637079  4.273322e-02
#>  [74,] -0.15000428  0.067784126  0.0317855162 -0.0428088173 -4.458161e-02
#>  [75,]  1.04889852 -0.160746433  0.0224571433  0.0388055221 -3.275242e-02
#>  [76,] -0.32684658  1.037769029  0.0359659464  0.0120599197  1.450788e-02
#>  [77,] -0.56387300  1.489576216  0.0448643628  0.0107388559 -2.351118e-03
#>  [78,]  0.65817602  1.056131896  0.0039709064  0.0257744417 -1.977617e-02
#>  [79,]  0.10896830 -0.985091877 -0.0371021925 -0.0287190594  3.313871e-02
#>  [80,] -0.01314694 -0.675773451  0.0211910621 -0.0481452628 -2.347922e-02
#>  [81,] -1.06610424  0.698666754  0.0219338316  0.0439205733 -8.599617e-03
#>  [82,]  1.27157569  2.279801457  0.0058851186 -0.0456252234 -4.361080e-02
#>  [83,]  1.13359638  0.667982342 -0.0318846274 -0.0410936688  3.581735e-02
#>  [84,]  0.06317963 -1.620232257  0.0169525247 -0.0189565605 -2.593632e-02
#>  [85,]  1.13427481 -1.174571984 -0.0087349280  0.0110731922  1.919897e-02
#>  [86,]  0.75568132 -0.211829525 -0.0121261802 -0.0148284667 -3.999030e-02
#>  [87,]  0.86280421  0.466045148  0.0012706998 -0.0445127488  1.818160e-03
#>  [88,] -0.72071613 -1.533936267  0.0192272061 -0.0210358739  3.022244e-02
#>  [89,] -0.59507999 -1.142219699 -0.0466451617 -0.0064609530  3.515036e-02
#>  [90,] -0.26115982 -1.169361888  0.0256251766  0.0427381722  4.743709e-02
#>  [91,]  2.07222632  1.878103848 -0.0480531712 -0.0054665905 -4.547735e-02
#>  [92,] -0.76525701 -0.221315467  0.0178853331  0.0412468209 -4.286151e-02
#>  [93,]  0.01114251  0.786096810  0.0155870373  0.0310399489 -3.121946e-02
#>  [94,] -0.89582246 -0.925738156  0.0362235341  0.0106673543 -4.569175e-03
#>  [95,]  0.03035147 -0.088831255  0.0206181577 -0.0215042649  5.641874e-03
#>  [96,]  1.09872829 -1.083969481 -0.0357071559  0.0088676500  4.250327e-02
#>  [97,] -0.15652624  2.310456300  0.0400606687  0.0132146884  3.357234e-03
#>  [98,] -0.16403877  0.732869920  0.0219971110 -0.0252191990 -5.549375e-04
#>  [99,] -1.08568295 -0.194147944  0.0163365343 -0.0166214987 -2.555933e-02
#> [100,] -1.12031728 -1.476041945 -0.0195872578 -0.0473705463  2.165986e-02
#> [101,]  1.28358419 -1.143485846 -0.0211302730 -0.0114276793  2.600802e-02
#> [102,] -0.88964408  1.280425431 -0.0396419236  0.0258483029 -4.266945e-02
#> [103,]  0.09667252 -1.299668583 -0.0110923532  0.0487360196  1.312419e-02
#> [104,] -0.20206912  0.214044215 -0.0391812961 -0.0050808378 -3.457974e-02
#> [105,] -0.59546245  1.261026159 -0.0277424756  0.0355677718  3.813075e-02
#> [106,]  0.55392995 -1.110529014 -0.0489850755 -0.0281348281  4.583308e-02
#> [107,] -0.15947592 -0.624969602 -0.0301523507  0.0301930686  3.153494e-02
#> [108,]  0.10305306 -0.095344307  0.0294701772 -0.0264757206 -1.656454e-02
#> [109,]  0.63234944 -1.302228343 -0.0447253486 -0.0497667200  4.074052e-02
#> [110,] -0.42138731  1.115920647  0.0282002091  0.0138383291 -5.229591e-03
#> [111,] -1.02046938 -0.980713800  0.0365759518  0.0170294064  4.289947e-02
#> [112,]  0.07040396  1.619014956 -0.0092774051 -0.0105322713 -4.693646e-02
#> [113,]  0.24155395 -1.105251724  0.0319550125 -0.0253994956  4.079205e-02
#> [114,] -0.23895047  0.868979235  0.0334051919  0.0039561825  4.593775e-02
#> [115,] -1.70457514  0.561799843  0.0096876160  0.0254380152 -3.207291e-02
#> [116,] -0.48131490 -0.304855760 -0.0370069354 -0.0002756319 -4.309148e-02
#> [117,]  1.97280418  0.585097493 -0.0199589521 -0.0252596536  4.208677e-02
#> [118,] -1.30627256 -2.384736182  0.0018917130  0.0132897190  3.084040e-02
#> [119,] -1.21121049  0.844055069  0.0022048189  0.0205330441 -3.804917e-03
#> [120,]  1.03133442 -1.191728611  0.0168870137 -0.0344612013 -1.082145e-02
#> [121,] -1.19532709 -0.329162571 -0.0474511615  0.0350218165 -8.248837e-03
#> [122,] -1.33487108  2.266189829  0.0250236757 -0.0409939812 -2.394960e-02
#> [123,]  2.67820657 -1.060139010 -0.0291095996 -0.0176145782 -4.812042e-02
#> [124,]  0.91885378  1.356737837 -0.0456319631 -0.0448173560  4.655695e-02
#> [125,] -1.13903782 -0.592416593 -0.0475459471  0.0480178217  4.746066e-02
#> [126,] -0.12424580 -0.854541023  0.0208746986  0.0479148789 -1.690880e-02
#> [127,] -0.69305383 -0.935500150  0.0112582754  0.0467905645 -3.542551e-02
#> [128,]  0.75928759 -1.472203692  0.0341073811  0.0036784293 -2.490102e-02
#> [129,]  1.01330583 -0.629893458  0.0421455409 -0.0284505607 -4.686855e-02
#> [130,] -0.77638440 -1.075745067  0.0081570214  0.0237940408  4.568508e-02
#> [131,] -0.77169540 -0.237254405  0.0197564095 -0.0342422199 -1.694252e-03
#> [132,] -1.59527495  1.157719103 -0.0484505952 -0.0245812318 -2.421517e-02
#> [133,]  0.22330448  0.599077898 -0.0320326174  0.0385641397  1.448532e-02
#> [134,]  0.19778054 -0.549629155  0.0075467648 -0.0231698694  3.271302e-02
#> [135,] -0.59070942  0.032245242  0.0353020210  0.0392830931 -6.234291e-03
#> [136,] -1.87171621 -0.241186830  0.0152449868 -0.0384910566 -4.804818e-02
#> [137,]  0.45920015  0.955506712  0.0368450529 -0.0010994278 -8.559206e-05
#> [138,] -1.02317696  0.405964297 -0.0484752014 -0.0483961901 -2.051229e-03
#> [139,]  0.38041333  0.837717461 -0.0145270658 -0.0335307898 -4.483748e-02
#> [140,]  1.30884171  0.358047883 -0.0123740624 -0.0166782717  1.087038e-02
#> [141,] -1.00522333 -0.482588035  0.0407154984  0.0008944256 -3.384040e-03
#> [142,] -0.59508950  0.918098039 -0.0082684050  0.0390550125 -4.725878e-02
#> [143,] -0.51793087  0.190752475 -0.0005035731  0.0348076932 -4.406941e-02
#> [144,] -0.21443905 -0.585693906 -0.0273028755 -0.0330922237  1.295361e-02
#> [145,]  0.51466592  0.171210539  0.0161636332 -0.0091165979  2.167122e-02
#> [146,]  0.20135333 -0.005372291 -0.0258473551 -0.0027610870 -4.919681e-02
#> [147,] -1.42450176  1.192724968 -0.0388338244  0.0434272171 -3.184544e-02
#> [148,] -0.29821291 -0.406692963  0.0229592581 -0.0172642431 -1.948746e-02
#> [149,] -0.13178584 -1.016621100 -0.0124106927 -0.0265581872  2.529822e-02
#> [150,]  0.37276075  0.245298812  0.0136160975  0.0253309568  9.170931e-03
#>                [,11]
#>   [1,]  0.0152765223
#>   [2,] -0.0223713805
#>   [3,] -0.0090256640
#>   [4,] -0.0384177636
#>   [5,] -0.0198625185
#>   [6,]  0.0020243809
#>   [7,]  0.0036389584
#>   [8,] -0.0416026886
#>   [9,]  0.0221091650
#>  [10,] -0.0097982891
#>  [11,] -0.0281321220
#>  [12,] -0.0116173150
#>  [13,]  0.0071815043
#>  [14,] -0.0327414995
#>  [15,] -0.0219575853
#>  [16,] -0.0101757864
#>  [17,]  0.0418462364
#>  [18,] -0.0324064380
#>  [19,]  0.0479369845
#>  [20,] -0.0469670376
#>  [21,] -0.0240231430
#>  [22,]  0.0200164374
#>  [23,] -0.0277004966
#>  [24,]  0.0039307385
#>  [25,] -0.0276471300
#>  [26,]  0.0494497998
#>  [27,]  0.0082369526
#>  [28,] -0.0302305267
#>  [29,] -0.0195569334
#>  [30,]  0.0411729548
#>  [31,] -0.0463658023
#>  [32,] -0.0321239714
#>  [33,] -0.0043908593
#>  [34,]  0.0112816812
#>  [35,] -0.0127485556
#>  [36,]  0.0387298077
#>  [37,] -0.0173830483
#>  [38,] -0.0025416805
#>  [39,] -0.0283278972
#>  [40,]  0.0393380380
#>  [41,] -0.0168024790
#>  [42,]  0.0008446755
#>  [43,]  0.0411146535
#>  [44,] -0.0264288456
#>  [45,] -0.0472297551
#>  [46,] -0.0114359361
#>  [47,]  0.0336941352
#>  [48,]  0.0444863969
#>  [49,]  0.0052959394
#>  [50,] -0.0085872512
#>  [51,] -0.0060163360
#>  [52,]  0.0433165835
#>  [53,] -0.0461958522
#>  [54,] -0.0466172259
#>  [55,] -0.0118074633
#>  [56,]  0.0100013499
#>  [57,] -0.0239583077
#>  [58,] -0.0079262669
#>  [59,]  0.0295734048
#>  [60,] -0.0158890818
#>  [61,] -0.0121644075
#>  [62,]  0.0140660760
#>  [63,] -0.0181345317
#>  [64,]  0.0118097182
#>  [65,]  0.0423620816
#>  [66,]  0.0479755789
#>  [67,]  0.0480030891
#>  [68,]  0.0032401488
#>  [69,] -0.0202494221
#>  [70,] -0.0015670205
#>  [71,] -0.0306482177
#>  [72,]  0.0308181562
#>  [73,]  0.0452089204
#>  [74,]  0.0443659193
#>  [75,] -0.0148027087
#>  [76,]  0.0239503613
#>  [77,]  0.0015749611
#>  [78,]  0.0305062093
#>  [79,]  0.0281363494
#>  [80,]  0.0063759152
#>  [81,]  0.0128285358
#>  [82,]  0.0032365235
#>  [83,] -0.0026836061
#>  [84,]  0.0318031608
#>  [85,] -0.0119613190
#>  [86,] -0.0033496301
#>  [87,] -0.0094586632
#>  [88,]  0.0478561279
#>  [89,]  0.0400075605
#>  [90,]  0.0234232664
#>  [91,]  0.0139804197
#>  [92,]  0.0076303265
#>  [93,]  0.0021307665
#>  [94,]  0.0142457839
#>  [95,]  0.0370306083
#>  [96,] -0.0436818916
#>  [97,]  0.0340868593
#>  [98,]  0.0017820277
#>  [99,] -0.0457485725
#> [100,]  0.0375819848
#> [101,]  0.0322422037
#> [102,]  0.0332951861
#> [103,]  0.0261462097
#> [104,] -0.0104718531
#> [105,] -0.0095390568
#> [106,] -0.0358294725
#> [107,] -0.0478765350
#> [108,]  0.0392359449
#> [109,]  0.0005881421
#> [110,]  0.0373804413
#> [111,]  0.0125858496
#> [112,]  0.0115549479
#> [113,]  0.0029696661
#> [114,] -0.0345047883
#> [115,] -0.0170131781
#> [116,] -0.0016062550
#> [117,]  0.0151519471
#> [118,]  0.0150455180
#> [119,]  0.0039169698
#> [120,]  0.0310115105
#> [121,] -0.0139238091
#> [122,]  0.0098086946
#> [123,]  0.0013765651
#> [124,] -0.0472162024
#> [125,] -0.0413761807
#> [126,] -0.0139807249
#> [127,] -0.0024320933
#> [128,]  0.0218235584
#> [129,]  0.0002312348
#> [130,]  0.0467690276
#> [131,]  0.0466423045
#> [132,]  0.0169794254
#> [133,] -0.0124759461
#> [134,]  0.0127030541
#> [135,] -0.0018315836
#> [136,]  0.0229571232
#> [137,] -0.0074759367
#> [138,]  0.0295434827
#> [139,] -0.0475616841
#> [140,]  0.0139095037
#> [141,] -0.0057813829
#> [142,] -0.0210469098
#> [143,] -0.0428817054
#> [144,] -0.0226559085
#> [145,] -0.0385465592
#> [146,] -0.0353889463
#> [147,] -0.0101217329
#> [148,] -0.0330728494
#> [149,]  0.0394446606
#> [150,] -0.0212162870