Generate three clusters of data points with optional noise.

This function generates three clusters of data points along with optional noise.

Usage

three_clust_diff_dist(n, num_dims, num_noise, min_n, max_n)

Arguments

n

Total number of data points to generate, should be a multiple of three.

num_dims

Number of dimensions for each data point.

num_noise

Number of additional noise dimensions to add to the data.

min_n

Minimum value for the noise added to the data.

max_n

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

set.seed(20240412)
three_clust_diff_dist(
  n = 100, num_dims = 7, num_noise = 4, min_n = -0.05,
  max_n = 0.05
)
#> Warning: The sample size should be a product of three.
#>               [,1]        [,2]        [,3]          [,4]        [,5]
#>   [1,]  1.31776542  0.66789796  1.57007706  1.2719198597 -0.18806377
#>   [2,] -0.64215864  0.27089658  0.56954317 -1.8537526860  1.06077068
#>   [3,] -1.29205633 -0.42460261  0.80805268  1.2284611910 -0.56004944
#>   [4,]  1.28481192 -0.21678152  0.66010617  0.1961927008  2.26132992
#>   [5,]  0.37750101  1.00030819 -0.49689997 -0.2434757824  1.64106216
#>   [6,] -0.15414403  0.54672259 -0.34817350  2.8007419646 -0.10110520
#>   [7,] -1.53212609 -0.87667248  0.82928910 -1.1063454092  0.95273404
#>   [8,]  0.39068303 -0.63083642  0.29068111  0.8943612602  1.08345957
#>   [9,] -0.46446764 -1.24992306  1.17014422  0.8043028102 -1.10921339
#>  [10,]  2.97923939  1.90618125 -1.26929768 -0.0594167456 -1.06290240
#>  [11,]  0.36208825  1.26734149  0.27369230  0.4799534845 -0.60235762
#>  [12,] -0.02178079 -1.26698095  0.08033855 -1.0281534664  0.08723732
#>  [13,]  0.25381815  0.80728113  1.31755117  0.8753180835 -0.91298464
#>  [14,] -0.07235334 -0.17772803  0.46469651 -1.2189158095 -2.88099249
#>  [15,] -0.50962033  1.09853842  0.80475719  0.1395135636 -0.14787024
#>  [16,]  0.38011146  0.32993219 -1.40961401 -0.8286717919 -0.75016680
#>  [17,] -0.31163667  0.70586598  0.08251067  0.1625682288 -0.46940140
#>  [18,]  0.34093740  2.51875924  0.32816401  0.6772926051  1.90291080
#>  [19,]  0.40795937  0.24152004  0.14227324  1.0538956896 -1.26478804
#>  [20,]  0.98013776  1.08772383 -0.81911479  1.0122274704  1.42112767
#>  [21,]  1.33750733 -0.60580015 -0.57326802  1.6856288586 -3.25512761
#>  [22,] -0.99702040  1.14937135 -0.35288160 -0.0564964032  0.03920695
#>  [23,] -0.15122674 -0.31172439  1.95091330 -1.1313454545 -0.78693252
#>  [24,] -1.32763833  0.17572529  1.65780806 -0.4384554404 -1.05129798
#>  [25,]  0.34765139 -1.44070282 -0.84571266  1.8068228243 -0.13062107
#>  [26,]  0.37224577 -1.00081268 -0.89244833  0.0696001893 -0.21874566
#>  [27,]  0.11038148 -0.10981087  1.00421805 -0.0366022424 -0.66099202
#>  [28,]  0.13569629  0.23117862 -0.42228156  0.1751687828  0.60480246
#>  [29,] -0.87850811  0.54010110 -0.88816000  1.1469137295  0.40659882
#>  [30,]  1.25457960  0.18719592 -0.51856334 -1.1662313022  0.54813103
#>  [31,] -0.19569676 -0.35888699 -0.28047377 -0.3632957087  0.98715260
#>  [32,] -0.89547223 -1.25869524 -1.16512654 -1.7168143201  1.61994869
#>  [33,]  0.36261837  0.50170023 -0.53294121  0.5894172531 -0.41511572
#>  [34,]  9.43533184  0.42300917 -0.09561403 -0.6320602840  0.10707481
#>  [35,]  9.06387415 -0.53381767  0.09144115  0.0309475709 -1.51841497
#>  [36,] 10.22779172 -0.44348316 -0.70172655  0.1507006890  0.79505423
#>  [37,]  9.82213020  0.88880572 -0.30982891  1.7638786981 -1.35599449
#>  [38,] 10.69543852  0.70598859 -1.12288439  1.1980400866  1.30196363
#>  [39,] 10.70441632 -0.90622907 -1.68778031 -1.5486457225 -0.49983098
#>  [40,]  8.46237013 -0.36663995  1.13976750  0.4779906836 -0.17399153
#>  [41,] 10.11175612 -1.88770405  1.03864988  0.5216152484 -1.03062437
#>  [42,] 11.82075244  1.10850529 -1.93945783 -1.4229301245 -0.37896223
#>  [43,]  9.29343647  0.19772643  1.70112321 -0.5582988185  0.88339398
#>  [44,] 11.62484014 -0.96862054 -0.39040605  0.5017199426  0.32975938
#>  [45,]  9.69081161  0.03688506 -0.05242460 -1.3995595454 -1.11387209
#>  [46,] 10.98204943 -1.04507740  0.26711136  1.5923383581 -1.16306626
#>  [47,]  9.72797116  0.72634748 -0.51064114 -0.5755644470 -0.27017025
#>  [48,] 11.78139512 -1.07818655 -0.28190947 -1.3360924557 -0.11075540
#>  [49,] 10.53293134 -2.24259619 -0.41762537 -1.5473709318 -0.16725673
#>  [50,] 10.57231574  0.12305753  0.75378039 -1.4192089733  2.61687313
#>  [51,] 10.12969475 -0.22047577 -0.98068250 -1.4732492889 -0.24898268
#>  [52,]  8.46897537  0.53009379  0.52924892  0.6780033761  1.28296252
#>  [53,]  9.75188909 -0.85179755  0.49303638 -0.0624319576 -0.14314668
#>  [54,]  9.05480533  1.46843459 -1.25623308 -0.8159023585  0.20371593
#>  [55,] 10.86664607 -0.16787861  0.30142591 -0.8590996252  0.11989067
#>  [56,] 10.47871207 -0.83602877 -1.37895629 -1.1500505507 -1.09662475
#>  [57,]  9.67391618  0.29181967 -1.09379135  0.9173190395 -0.60942621
#>  [58,] 10.24316460  1.03478639 -0.46069015  0.7310155710  1.07792551
#>  [59,] 11.39256372  0.28129101 -0.40881387  1.2610683767  0.14578287
#>  [60,]  9.50954024 -0.04315316  0.80973536 -0.9382964857 -0.25728507
#>  [61,]  9.80105389  0.23968569  0.20352166  1.4935173398  0.29303095
#>  [62,]  9.20151899  1.13173560 -0.70722373 -0.5879884858  0.20883406
#>  [63,]  9.23174573 -0.35416030 -1.66545807 -1.6417790596 -1.40001140
#>  [64,]  9.86973580 -0.35661619  1.20229341  0.2032857813 -0.14158068
#>  [65,]  9.46763214 -0.34167176  0.59266037  0.6309166908 -0.37661416
#>  [66,] 10.07170660 -0.30297090  0.32102530 -1.7867392047 -0.05643202
#>  [67,] 49.76873011  0.11599080 -2.05287050  0.9663841764 -0.83558546
#>  [68,] 49.22232597  0.14945202 -0.97559678 -1.1924407387  1.10899392
#>  [69,] 49.30336800 -0.85817223  0.69410665 -0.2105711704 -2.09828643
#>  [70,] 50.12907914 -0.85882555 -0.72483790  0.0009358841  0.30867051
#>  [71,] 49.70522937 -0.29017795  0.54391142 -0.4941604747  0.32081971
#>  [72,] 50.52017540  0.99075162  0.13228204  0.6655462871 -0.39651750
#>  [73,] 49.86651059 -1.14624391 -0.69795674 -0.0531421842 -0.17302248
#>  [74,] 49.49319952  0.78760273  0.09610427 -2.4961099700  0.61921271
#>  [75,] 49.32953427  0.38774472  0.41859006 -0.2669197209  0.17287409
#>  [76,] 50.64019608 -0.73466187 -0.14041617 -1.2240962028  0.73277358
#>  [77,] 49.50681508  0.92920848 -1.24367806  1.0450283491  0.10484993
#>  [78,] 51.71891939  0.25790760  0.62579125 -0.7315958295  0.91281030
#>  [79,] 48.54128480 -0.23601985 -0.72937741 -0.1923923478  1.04804298
#>  [80,] 49.73984610  0.36500145 -1.50293132  0.5589554187  0.27152918
#>  [81,] 49.53760916 -0.63030432 -1.97645703  0.3440633341  1.68714529
#>  [82,] 49.16911662  0.59074187 -1.15160617 -0.6151790883  0.04382268
#>  [83,] 49.41387740 -0.12328239  1.16446132  2.1188470158  1.70955480
#>  [84,] 51.00729041 -0.04125274  0.04254069  1.1264190543  0.40672032
#>  [85,] 48.29897430 -1.75527616 -0.26853925  0.1259205617 -2.24820471
#>  [86,] 49.23390165  1.31925111  0.23202649  0.9879903361  0.81121101
#>  [87,] 50.54730413 -0.03462962 -0.71517839 -1.1463377392  0.45282886
#>  [88,] 48.59300390 -0.06347348  0.48558076 -0.0365983249  0.90886882
#>  [89,] 51.89822210  2.83883124  0.95776100 -1.2936261616  0.95430388
#>  [90,] 50.93033721 -0.46264207  0.32522296  0.3840011799  0.05908066
#>  [91,] 51.80585662  1.15801659 -0.06784286  1.9132326731  1.59347272
#>  [92,] 50.55811928 -0.54615670  0.69138097  0.1723712919  1.27307185
#>  [93,] 50.39793331 -0.91259786 -0.56370987 -0.9780641500  0.64358355
#>  [94,] 50.72572200  1.24155111 -1.02795711 -1.4332648175 -0.87063338
#>  [95,] 51.50586513 -1.24182669 -0.18948266 -0.4286249220 -0.34435709
#>  [96,] 50.15040459  0.37291896  0.56428019  0.4372471487 -0.15000428
#>  [97,] 49.68137727 -0.54903289  0.44869772 -0.8023827994  1.04889852
#>  [98,] 48.65046513  1.23962998  1.39866497 -1.5052685199 -0.32684658
#>  [99,] 51.51622384  1.08112593 -0.88378685  0.1750359599 -0.56387300
#>                [,6]        [,7]         [,8]          [,9]         [,10]
#>   [1,]  0.465880137 -0.10218612  0.032270896  0.0408611069  3.052443e-02
#>   [2,]  0.093995891  0.98326380  0.035427231 -0.0430016565 -1.160597e-02
#>   [3,]  0.160459355  0.37424741  0.003539199 -0.0277144931  3.792113e-02
#>   [4,] -1.502241191 -0.70945254  0.013988467  0.0044848796 -3.158116e-02
#>   [5,]  1.079958193 -2.69801073  0.036081084  0.0013245016  1.444737e-06
#>   [6,] -2.685784084  0.60708943 -0.004937051  0.0315999172  4.780702e-02
#>   [7,] -0.406156792 -1.07401237 -0.048956855 -0.0275080922 -1.786560e-02
#>   [8,]  1.388096825 -0.27666783  0.036973973 -0.0454859233  8.289509e-03
#>   [9,] -0.367606775 -0.87403538 -0.026818111 -0.0078129898 -4.142237e-03
#>  [10,] -0.126980115 -0.77864040 -0.038079275  0.0189091858 -2.878036e-02
#>  [11,]  0.018068134 -0.11632503  0.007696998  0.0333838260 -4.678112e-02
#>  [12,] -1.008666767  0.34809368  0.023053223  0.0143570505 -7.248781e-03
#>  [13,] -0.044261411  0.17680203  0.043003368 -0.0480747631 -2.387026e-02
#>  [14,] -0.395258966  0.34566330 -0.048456126 -0.0088451295 -2.788312e-02
#>  [15,] -0.017644679 -0.03284358 -0.025698535  0.0088314157  2.355788e-02
#>  [16,]  0.867107978  0.84299544 -0.010428709 -0.0456415419  9.073966e-03
#>  [17,]  1.013397477 -0.65727505 -0.044380421 -0.0365808491  1.385957e-02
#>  [18,]  0.060356438  1.88799197 -0.007530993  0.0089214609  4.566643e-02
#>  [19,] -1.473051026  0.98227223 -0.028698158 -0.0335134245  1.217832e-02
#>  [20,] -1.869984291  0.47701002  0.001014189 -0.0207257007 -2.136589e-02
#>  [21,] -1.103388327 -1.31848116 -0.008176832 -0.0080427809 -4.100529e-03
#>  [22,] -0.800048820  0.11191996 -0.046918919  0.0416602065 -3.363362e-02
#>  [23,] -1.530029288  0.16621791 -0.001004972  0.0249443349  1.029730e-03
#>  [24,] -0.811261464 -1.52467105  0.007590666 -0.0455361755  7.640567e-03
#>  [25,] -0.509584873  1.00496249  0.011792085  0.0171547973  2.222287e-02
#>  [26,] -0.246104794 -0.18746656  0.031183350 -0.0173837004  5.950782e-03
#>  [27,]  1.109819180 -1.45868272  0.049559160  0.0336281423  3.322597e-02
#>  [28,] -0.135457526 -1.60450399 -0.019730603 -0.0430322238 -2.577741e-02
#>  [29,]  0.153441658 -0.31649364  0.010753095  0.0335027961 -4.813649e-03
#>  [30,]  0.489444144  0.07629640 -0.036035923 -0.0193683902 -1.810236e-02
#>  [31,] -1.400875699 -0.23658432 -0.035924287 -0.0194425357  3.902698e-02
#>  [32,]  0.278644225 -0.17982786 -0.027052627 -0.0190907998  1.436196e-02
#>  [33,] -0.860390101  0.08611162  0.047176052  0.0304273498  2.589669e-02
#>  [34,]  0.627399507  2.59797441 -0.038546310  0.0384242548  3.991658e-02
#>  [35,]  0.481320494 -0.39909894  0.048462113 -0.0346597465  3.474310e-02
#>  [36,]  1.026370668 -0.42201975  0.040882778  0.0043540464  2.178738e-02
#>  [37,]  0.991376195 -2.07266245 -0.012454179  0.0101461249  1.391394e-02
#>  [38,] -0.356802884  0.18122659  0.030690455 -0.0392301483 -1.622641e-02
#>  [39,] -0.751490307  0.64476388 -0.018165616 -0.0017683804 -4.215942e-02
#>  [40,] -0.673229476 -1.43247722  0.006543203 -0.0044020407 -1.544534e-02
#>  [41,]  1.776774970  0.52148217 -0.026106587  0.0386063067 -4.290784e-02
#>  [42,]  0.643043642 -0.61970958 -0.035683260  0.0276430219 -4.012456e-03
#>  [43,] -0.140095574  0.91584885 -0.020253392  0.0231796788 -1.207017e-02
#>  [44,]  1.525287267  0.89104131 -0.004983460  0.0399001111  2.392500e-02
#>  [45,]  0.605172455  0.39015326 -0.009789291 -0.0252182457 -3.337082e-02
#>  [46,]  1.615340390 -0.56829415 -0.025330059 -0.0271688743 -1.450514e-02
#>  [47,] -0.643775058 -1.21135454 -0.042147717 -0.0405868134 -4.277526e-02
#>  [48,] -2.419653265 -0.67839932  0.013366779 -0.0461651006 -2.957142e-02
#>  [49,]  0.163007346 -0.52298548 -0.009443090 -0.0239686144  1.864701e-02
#>  [50,] -1.014585047  1.31572536  0.036815157 -0.0284355255  5.865818e-03
#>  [51,]  0.616347675 -1.59171476 -0.045443008 -0.0432236770 -9.092460e-03
#>  [52,] -0.004308625  1.38220954  0.046873171 -0.0008444753 -1.464067e-02
#>  [53,] -0.634721137  0.21806030 -0.036595668 -0.0431302498  6.559555e-04
#>  [54,] -2.104940995 -0.38573697  0.040763914 -0.0135138288  9.123974e-04
#>  [55,]  0.509604572  1.06960190  0.019149391 -0.0240350844  1.988204e-02
#>  [56,]  0.660915949 -0.74231963 -0.007463101 -0.0262214925  2.317451e-02
#>  [57,]  0.054998561 -1.08058524  0.027526701  0.0253784381  1.936986e-02
#>  [58,] -0.617691130 -0.47290249  0.035032165 -0.0401683430  1.021051e-02
#>  [59,]  1.127894509  0.09612179 -0.042090889  0.0148201867 -2.221891e-02
#>  [60,] -1.341959003 -0.93495145 -0.031858538  0.0145698657 -1.553250e-02
#>  [61,]  0.861507879  1.06221428 -0.035635763 -0.0427757076  2.984757e-02
#>  [62,]  1.017133728 -1.62756968  0.010518194 -0.0294804221  3.411384e-03
#>  [63,]  0.085472257 -1.08496242  0.039394676 -0.0106754863  8.514175e-03
#>  [64,]  0.278285662 -0.09128764  0.048665499  0.0451366269  2.402575e-02
#>  [65,]  0.575590442 -0.85941315 -0.015873389 -0.0453712832 -1.379349e-02
#>  [66,] -1.574676740 -1.11356064 -0.033968407  0.0378442727 -1.567895e-02
#>  [67,]  0.658176024  1.31014093 -0.005183702 -0.0337859626 -3.803652e-02
#>  [68,]  0.108968300 -0.61399954  0.005894272  0.0417272977  1.626677e-02
#>  [69,] -0.013146940  0.65551040 -0.020400923 -0.0372528013 -2.913256e-02
#>  [70,] -1.066104236  0.38450699 -0.030290221  0.0296265551 -3.116542e-02
#>  [71,]  1.271575686 -0.78463574 -0.025252589 -0.0478828177  4.537057e-03
#>  [72,]  1.133596379  2.48505184  0.038092777 -0.0404781338 -2.753074e-02
#>  [73,]  0.063179632  0.18069110  0.017222121  0.0207082593 -2.978112e-02
#>  [74,]  1.134274806 -0.91588076 -0.005295695 -0.0096714110  3.389820e-02
#>  [75,]  0.755681318  1.28795811  0.032529426  0.0004271431 -3.612796e-02
#>  [76,]  0.862804210 -0.38266635  0.038810037 -0.0414806698 -5.702680e-03
#>  [77,] -0.720716131  0.83304240 -0.034321623  0.0026098275 -1.836704e-02
#>  [78,] -0.595079993 -0.48425068  0.003307733  0.0120450056 -4.548412e-02
#>  [79,] -0.261159820  0.03357698 -0.038647418  0.0151853368 -4.938652e-02
#>  [80,]  2.072226323  0.06778413  0.029351479  0.0404580371  6.767121e-03
#>  [81,] -0.765257007 -0.16074643 -0.033741472 -0.0406473272  4.575589e-02
#>  [82,]  0.011142506  1.03776903  0.044586502 -0.0454211103  3.026242e-02
#>  [83,] -0.895822456  1.48957622  0.010777857  0.0232835801 -2.203272e-02
#>  [84,]  0.030351473  1.05613190  0.025577739 -0.0160111573 -2.572107e-02
#>  [85,]  1.098728285 -0.98509188  0.043331587 -0.0096129851 -1.736868e-02
#>  [86,] -0.156526236 -0.67577345 -0.020010959  0.0336451620 -8.254937e-03
#>  [87,] -0.164038770  0.69866675 -0.007087577  0.0367506423 -2.628355e-03
#>  [88,] -1.085682951  2.27980146  0.039218491  0.0499820325  1.052407e-02
#>  [89,] -1.120317283  0.66798234  0.016918268  0.0450245543 -4.656037e-02
#>  [90,] -2.149297571 -1.62023226 -0.026308006 -0.0325935442  2.746602e-02
#>  [91,] -0.195891297 -1.17457198  0.040931186 -0.0142342126 -3.305347e-02
#>  [92,] -1.527919524 -0.21182952 -0.045673970  0.0418596522  2.058071e-02
#>  [93,] -0.546765233  0.46604515 -0.023683289  0.0273946436 -2.243888e-02
#>  [94,]  0.699412952 -1.53393627 -0.028064542 -0.0290959365  4.207739e-02
#>  [95,]  0.930485330 -1.14221970 -0.022134393  0.0440761212 -1.392473e-02
#>  [96,]  0.773861899 -1.16936189  0.004185543  0.0252200304 -2.141981e-02
#>  [97,]  0.374277597  1.87810385  0.036585727 -0.0032015956 -7.021189e-03
#>  [98,]  0.203355429 -0.22131547  0.043765987  0.0334407635  4.598387e-02
#>  [99,]  0.237516502  0.78609681  0.017827392  0.0100947189 -1.378928e-02
#>                [,11]
#>   [1,] -1.969493e-02
#>   [2,] -3.483497e-02
#>   [3,] -2.907384e-02
#>   [4,] -1.371125e-02
#>   [5,]  2.013265e-02
#>   [6,]  3.084027e-02
#>   [7,]  3.868453e-02
#>   [8,] -1.726326e-03
#>   [9,] -3.016054e-02
#>  [10,] -3.947962e-02
#>  [11,] -2.612400e-02
#>  [12,] -3.830163e-02
#>  [13,] -4.030351e-02
#>  [14,]  3.075214e-02
#>  [15,] -2.825564e-02
#>  [16,] -3.687231e-03
#>  [17,] -5.937318e-03
#>  [18,]  3.891223e-02
#>  [19,]  4.915326e-02
#>  [20,] -1.149312e-02
#>  [21,]  1.708466e-02
#>  [22,] -2.145910e-03
#>  [23,]  2.224431e-02
#>  [24,]  3.747981e-02
#>  [25,]  2.936580e-02
#>  [26,]  2.452042e-02
#>  [27,]  1.861751e-02
#>  [28,] -3.227076e-02
#>  [29,] -1.698146e-02
#>  [30,] -4.052699e-02
#>  [31,]  9.298904e-03
#>  [32,]  1.551375e-03
#>  [33,]  1.099064e-02
#>  [34,] -4.142531e-02
#>  [35,]  2.446827e-02
#>  [36,]  4.758606e-02
#>  [37,]  4.651402e-04
#>  [38,]  4.746333e-02
#>  [39,] -3.040590e-02
#>  [40,]  4.718330e-02
#>  [41,] -1.494812e-02
#>  [42,]  4.221738e-03
#>  [43,] -3.731033e-02
#>  [44,]  4.333750e-02
#>  [45,]  5.679760e-03
#>  [46,] -1.348684e-02
#>  [47,] -1.485019e-02
#>  [48,]  4.501220e-02
#>  [49,] -1.433110e-02
#>  [50,] -1.455981e-02
#>  [51,]  3.729943e-02
#>  [52,]  3.664081e-02
#>  [53,]  4.065603e-02
#>  [54,]  6.297873e-03
#>  [55,]  4.625811e-02
#>  [56,]  3.713206e-02
#>  [57,] -1.392627e-02
#>  [58,]  2.710864e-02
#>  [59,]  1.909521e-02
#>  [60,]  3.333264e-02
#>  [61,]  2.350889e-02
#>  [62,] -3.155377e-02
#>  [63,] -2.224099e-02
#>  [64,] -4.653660e-02
#>  [65,] -3.230525e-02
#>  [66,] -1.660479e-02
#>  [67,] -2.903426e-03
#>  [68,] -3.384094e-02
#>  [69,] -3.966680e-02
#>  [70,] -2.314145e-02
#>  [71,]  8.618596e-05
#>  [72,]  4.029880e-02
#>  [73,] -4.980999e-02
#>  [74,] -1.772390e-02
#>  [75,]  4.220014e-04
#>  [76,] -4.190222e-03
#>  [77,] -3.202595e-02
#>  [78,]  3.258464e-02
#>  [79,]  2.502427e-02
#>  [80,] -2.228136e-02
#>  [81,] -2.469313e-02
#>  [82,]  2.446548e-02
#>  [83,] -4.090000e-02
#>  [84,]  2.503988e-02
#>  [85,] -3.567281e-03
#>  [86,]  4.994604e-02
#>  [87,] -3.451284e-03
#>  [88,]  2.568657e-02
#>  [89,] -1.665835e-02
#>  [90,]  7.784689e-04
#>  [91,] -4.800170e-03
#>  [92,]  1.267691e-02
#>  [93,] -2.163002e-02
#>  [94,]  4.221882e-02
#>  [95,] -6.387561e-03
#>  [96,]  2.755540e-02
#>  [97,]  1.837432e-02
#>  [98,] -2.392649e-02
#>  [99,] -3.514697e-03