Generate three clusters of data points with optional noise.

This function generates three clusters of data points along with optional noise.

Usage

three_clusters_data_with_noise(
  sample_size,
  num_dims,
  num_noise_dims,
  min_noise,
  max_noise
)

Arguments

sample_size

Total number of data points to generate, should be a multiple of three.

num_dims

Number of dimensions for each data point.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated data points with or without added noise.

Examples

three_clusters_data_with_noise(sample_size = 100, num_dims = 7,
num_noise_dims = 4, min_noise = -0.05, max_noise = 0.05)
#> Warning: The sample size should be a product of three.
#>               [,1]         [,2]         [,3]         [,4]          [,5]
#>   [1,] -0.67058844 -0.138782262 -0.001506127  0.526601585  2.4889820703
#>   [2,] -0.62623667  0.353656841 -0.890746733 -0.501286450  0.5135743050
#>   [3,]  1.44289559  0.291826111 -0.276886240 -0.450135590 -1.3992029986
#>   [4,] -0.88432998 -0.654287548 -0.247939416  0.160308210  0.0885838940
#>   [5,] -0.53046593  1.186960371 -0.408584534  0.421386214 -0.5096380900
#>   [6,]  0.81241722  1.510238241 -0.283804095  0.415903795  1.0599653888
#>   [7,] -0.16047959 -0.881378040 -1.786067367  1.161141571  0.5315143547
#>   [8,]  0.09416887  0.064886245 -2.791309486  0.731757823  0.8088224064
#>   [9,] -0.33474208  0.203497516  0.695400324 -2.084344261 -0.7633234498
#>  [10,] -0.72468069  0.546127235 -0.915397486  0.717871974 -0.7385729928
#>  [11,]  0.10315580  0.390758578 -0.946105746  1.350418099 -0.4435610614
#>  [12,]  0.43981512  0.221413175  0.615919419 -1.949944857  0.1413197785
#>  [13,] -0.55645314 -0.473106909 -0.356803150  1.023281926 -0.3754522264
#>  [14,] -0.86532266 -0.024840970  0.315541984  1.227132927 -0.9782285349
#>  [15,] -0.13947943  1.037368700 -0.739398876  0.324996506  0.7028657749
#>  [16,] -1.12867276 -0.798910229  0.848334016 -0.341281563 -1.2494252346
#>  [17,]  1.31966235 -0.164329881 -2.764671384  1.513129440 -0.2536209487
#>  [18,] -0.25904202  1.375170993 -0.190140616  1.507860374  0.2108259502
#>  [19,]  2.33590108  0.281176600  0.456241283  0.840588656 -0.4487235379
#>  [20,]  0.62073511 -0.441558417 -1.001132833  1.869551395 -0.8614149312
#>  [21,] -1.05919792  0.762838307  0.366111076 -0.337301590 -1.0161789586
#>  [22,] -1.95956908 -2.112336022  0.708034618 -1.515664936  0.2101812228
#>  [23,] -0.35020423  0.308899017 -0.472920433  0.246351860 -1.0348767511
#>  [24,] -0.96152893 -1.053813240 -0.341282825 -2.235151527  1.4137379874
#>  [25,] -1.73077557  1.269455008  0.996189071 -0.466149146 -0.1364174866
#>  [26,] -0.62104382  1.670073616  1.852646187 -0.694359705  1.0369112311
#>  [27,]  0.75254291  0.751044072  0.829879619 -2.211987900 -0.1721154478
#>  [28,] -3.05343942  0.306733674  0.361551757  0.119167289 -0.3147469642
#>  [29,]  1.20107543  0.242220281 -1.216818639 -0.719790599 -0.8960989756
#>  [30,]  2.05265559  1.243964774 -0.879286645 -0.072701138 -0.0805749226
#>  [31,]  0.93045452 -0.938571257  0.282227025  1.931189718  0.0411937230
#>  [32,] -1.29412107  0.503155100  0.569020421  0.139232516  0.1449169106
#>  [33,]  0.12995106  1.723292457 -0.497879530  0.355860474 -0.3646242004
#>  [34,] 10.34747628  1.003954743  0.155967687  1.287984189 -1.6017907949
#>  [35,]  7.92970956  0.005522643 -0.096490971  0.558824966  0.9708986129
#>  [36,] 10.92983359 -0.309976534 -1.708336780 -0.144844737  0.5476374063
#>  [37,] 10.68499283  1.118178041 -0.282490376  1.262209106 -0.2473901742
#>  [38,]  8.91421414 -0.886618661  1.444469094 -0.488716314  0.0038721053
#>  [39,]  8.65483495 -0.450296508 -0.568264888 -1.264442917  0.4475650135
#>  [40,]  9.31285757 -1.632145831  0.213956756  0.964921992  1.4339119126
#>  [41,] 10.70642926 -0.369128600 -0.659670608 -2.254998708  1.5201623307
#>  [42,]  9.73080551 -0.840780521 -1.154387787  0.768093639  1.6650488879
#>  [43,]  8.88553436 -0.568438661  0.758521481  0.811512982  1.2603459403
#>  [44,]  9.72690947 -0.958605795  0.155425444  0.517795578 -1.9007517069
#>  [45,] 10.93750250 -0.317262247 -0.778755346  0.081404358 -1.1525738411
#>  [46,] 10.32944936 -2.087974716 -0.179376309 -0.513693367  0.5027139293
#>  [47,] 10.57358654  1.586590190 -0.524411304  1.342854485  0.4973480345
#>  [48,] 10.83052672 -1.687858124  0.033458855 -0.864884051  0.2866500885
#>  [49,] 12.26129854 -1.243758516 -0.236845595 -1.346682760  0.4908237155
#>  [50,] 10.49339772  0.053836970  0.683973355 -1.600495822 -1.1610221053
#>  [51,]  8.85479728 -0.621855778  0.990197678 -0.127984399  1.3879688845
#>  [52,]  9.13313777  0.899464408 -0.503179861 -0.408080636  1.9552486102
#>  [53,]  9.89905957  0.112851915  0.729246074 -0.640625797 -0.9509175063
#>  [54,]  9.31663943 -1.640366605  0.018368160 -1.305853652 -1.9166418486
#>  [55,]  9.42508237  0.784291953  1.566172133 -0.736024184 -0.3562910911
#>  [56,]  9.26110967 -0.053684601  0.677337036 -1.657173556  0.9550917084
#>  [57,]  9.01129311  0.296784014  1.908911200  0.994263903 -0.1850023993
#>  [58,]  8.47972245  2.056704265  0.630108012  0.471710727  1.3614134768
#>  [59,]  9.26322426 -0.330251990 -0.828971689 -1.737070354  0.6255813807
#>  [60,]  9.08086236  2.581818711  0.972174317  1.759647052  1.0384561338
#>  [61,]  9.48492437 -0.600994667  0.046890773 -0.091534201  0.9768531962
#>  [62,] 10.72403669  0.134644831  0.283637415 -0.359497277 -1.0195240597
#>  [63,] 10.62655074 -0.882892332  0.382343519 -1.995254134 -0.5899610354
#>  [64,]  9.56854266  0.973111437  0.063774128  0.492581599  0.2776914682
#>  [65,]  9.87105060 -0.893557883  0.669839879  1.409586864  0.0009068391
#>  [66,]  9.10137724  0.210178521 -1.436128353  0.492566153  0.2249930164
#>  [67,] 49.79480925  0.977976151 -0.930759212  0.521227336  0.6193327857
#>  [68,] 50.79781152 -0.019908846  0.337862161 -1.328971117  0.5587077094
#>  [69,] 50.81927374  1.475578130  0.522782335  1.816607718 -0.0941062548
#>  [70,] 52.15966486  0.109779665 -1.735851759  0.227068092 -0.0805277099
#>  [71,] 50.49034234  0.012386562  0.939183196 -1.267317608  0.0195499723
#>  [72,] 51.14294279 -0.104813708  0.545023872  0.603874808 -0.2140639623
#>  [73,] 50.07044037  0.750888021 -1.524834411 -1.118178219  0.7348413168
#>  [74,] 49.37169395 -0.222881889 -0.574177146  0.322338208 -0.0158435919
#>  [75,] 49.90935023 -0.620374524  1.123966973  0.298355288 -1.1765210384
#>  [76,] 50.51051997  1.313999944 -1.330066640  1.461273902 -0.2206305958
#>  [77,] 49.89902637  1.623621068 -0.262061981  1.920752132  1.5645164501
#>  [78,] 48.17522316 -1.189772265 -0.341716840 -1.075694881  1.1220406616
#>  [79,] 51.27772162  0.115880775  1.108042032 -0.687520529  0.4630763187
#>  [80,] 48.88133452  1.618139506  1.425323846  0.615612326  0.6333566575
#>  [81,] 51.02661692 -0.009937086 -0.044740652 -0.003956603 -0.4873692217
#>  [82,] 50.68385339 -0.733299762 -0.120750163  0.060333316 -1.0396775057
#>  [83,] 49.00909312  0.642069265  0.657129977 -0.227867211 -2.4312069386
#>  [84,] 49.33649659  0.895565016  1.035033451 -0.890782870  0.2593732190
#>  [85,] 49.85280203 -0.864834733 -1.138008974 -0.070900923 -0.3637415731
#>  [86,] 49.98911178  0.035779960 -0.691706753 -1.217835315  1.0899766858
#>  [87,] 51.84293212  0.702140130  0.083571455 -0.396826129  0.7882513538
#>  [88,] 50.32479277  1.087774942  0.475664569  1.678253574  0.5745494844
#>  [89,] 50.21877346  1.032846943  0.237171833  1.972330385 -0.1757937479
#>  [90,] 50.69479682  1.230770581 -0.774876031 -0.158474927  0.0429632477
#>  [91,] 49.53040138 -0.996483870  0.324164624 -0.361596606  0.1622217408
#>  [92,] 50.00825016 -0.917445457  0.840477937 -0.706524518 -0.2756896339
#>  [93,] 48.07887637 -0.564285361  1.661976556  0.829596532 -0.0178250498
#>  [94,] 50.05179579  0.397653218  1.101599611 -0.778878702  0.2244857177
#>  [95,] 49.48474749 -0.057095721  0.800191004 -0.350058974 -1.6948945387
#>  [96,] 49.59328478 -0.525854563  0.489789121 -0.596009865 -0.1813324225
#>  [97,] 49.28052787  0.428255840  0.815966163 -0.453112302  0.8511553223
#>  [98,] 49.48676279 -1.046364085 -2.473557782 -1.267318495  0.5746835216
#>  [99,] 50.59792959  1.611411534 -1.080850836 -0.634832847  0.7530307227
#>               [,6]        [,7]         [,8]          [,9]         [,10]
#>   [1,]  0.09930323  1.13485350  0.003401300  4.950622e-02 -3.267329e-03
#>   [2,]  0.88087127  0.09146188  0.002590946  4.915037e-02 -3.623147e-02
#>   [3,] -0.76056556  1.17592896 -0.013263313  4.277479e-03 -3.954315e-02
#>   [4,] -1.22476597  0.18218080 -0.049227766 -4.061770e-02 -1.724989e-02
#>   [5,] -0.50380192  0.26843069  0.012781920  2.240658e-02  3.687436e-02
#>   [6,]  2.34334404  1.58179736 -0.002867659 -3.457006e-02 -7.543329e-03
#>   [7,] -1.30426121  1.17948584 -0.005237569  2.672784e-02  3.972700e-02
#>   [8,]  0.90581905  0.38508497  0.007331511 -4.858391e-02  3.728534e-02
#>   [9,]  0.28848522 -2.39959276  0.004665379 -1.788598e-03 -2.234722e-03
#>  [10,]  0.44904539  0.03341267  0.019012263  3.178011e-02 -2.665584e-02
#>  [11,]  1.14708792  0.54690339 -0.031562791  3.222550e-02 -4.170585e-02
#>  [12,]  1.19142501  0.32119708 -0.032660111  2.227654e-02 -4.221463e-02
#>  [13,]  1.50623877  1.55925044 -0.035928557  1.863988e-02  4.471483e-03
#>  [14,]  1.19230804  1.12880549  0.002157623 -1.720825e-02 -4.848081e-02
#>  [15,]  0.76144694 -0.21534401 -0.027910035 -4.314391e-03  4.792650e-02
#>  [16,] -0.14380793  0.66828598 -0.023282894  3.263116e-02  3.908964e-02
#>  [17,]  0.02289461 -1.28054095  0.042578303  1.819058e-02  2.519451e-02
#>  [18,]  0.12004023  0.17912750  0.038101873 -4.788207e-02 -3.593141e-03
#>  [19,] -0.65654471 -0.33664049  0.042340960  4.776362e-02  2.852735e-02
#>  [20,]  0.25375959  0.07686961  0.004503750 -1.683655e-02  1.127711e-02
#>  [21,] -0.63877747  0.47832193  0.010933377 -5.938731e-03  2.521832e-02
#>  [22,] -1.09812601  0.97582473 -0.020344824 -2.214924e-02 -4.440902e-02
#>  [23,] -2.32941393  0.95228405 -0.006701338  1.006501e-02 -1.488218e-02
#>  [24,]  1.10968086  1.53713442 -0.012554776  2.741717e-02 -8.525402e-03
#>  [25,] -0.19297322 -0.73101736 -0.033589396  1.372309e-02 -8.040710e-03
#>  [26,] -1.59804599 -1.41613120  0.007035280  2.531013e-02  4.259253e-02
#>  [27,] -0.24934434  1.25282862 -0.011672867 -4.102899e-02 -3.869467e-02
#>  [28,]  0.14410629  2.01139232 -0.047064239 -3.234673e-02 -3.292042e-02
#>  [29,]  0.83598311  0.25445222  0.008247358  3.660267e-02  4.274646e-02
#>  [30,] -0.98101671 -1.26123874  0.036447318  1.239281e-02  4.597302e-03
#>  [31,]  0.72836812  0.97089002  0.012008332 -2.308853e-02 -2.662024e-02
#>  [32,] -0.91300452  0.94580784  0.015922169  6.704685e-03 -4.058339e-02
#>  [33,]  2.33082168  1.07142014 -0.007986756  1.960521e-02  3.808500e-02
#>  [34,] -0.67520396 -2.64686433  0.013921663 -3.605981e-02  3.025132e-02
#>  [35,] -0.28988767  0.02327287  0.035178242 -5.947057e-03 -3.049119e-02
#>  [36,] -0.23792711  1.14736240  0.040541839 -1.920391e-02 -1.949013e-02
#>  [37,] -0.14478603  1.56104935 -0.036132105 -1.243606e-02 -4.712041e-05
#>  [38,] -0.39677388 -1.26401671 -0.045505840  2.838283e-02 -4.352197e-02
#>  [39,]  0.74238918 -1.16681868  0.042461237  1.443750e-02  3.278855e-02
#>  [40,] -0.58108347 -0.82444525  0.040968930  1.792356e-02 -4.120348e-03
#>  [41,]  0.01171851 -0.52382820 -0.016068687  2.449536e-02 -3.696755e-02
#>  [42,] -0.97883396 -0.11002393 -0.009438941  2.261596e-02 -5.323506e-03
#>  [43,]  1.17796765  2.19925998  0.010742136 -3.563256e-02 -4.505897e-02
#>  [44,]  0.59564475 -0.46512056  0.008367303 -6.356906e-03 -3.654808e-02
#>  [45,] -0.24082757  2.03698446 -0.033566379 -3.459441e-02  1.124566e-03
#>  [46,]  1.06814971  0.03900430 -0.019892153 -7.968933e-03 -2.388513e-02
#>  [47,] -0.99491639 -1.20309544  0.024926617 -7.002503e-03  1.670342e-02
#>  [48,]  0.36601269  1.98348411  0.009349628  1.483485e-02  2.298204e-02
#>  [49,]  0.45857178 -0.97033013 -0.026861360 -1.405079e-02 -3.322188e-02
#>  [50,]  0.23439775 -1.43256190 -0.016224542  1.740973e-02  4.860754e-02
#>  [51,] -1.17494693  0.48329953 -0.013932815  1.708562e-03 -5.674636e-03
#>  [52,]  0.84336225  0.12882417 -0.047265757 -1.185008e-02  3.892319e-02
#>  [53,]  2.23819694 -1.52829042 -0.005615839  1.451419e-02  2.803715e-02
#>  [54,]  0.04503217  0.49869349 -0.005790850 -4.225657e-02  4.023711e-03
#>  [55,] -0.03614213 -1.34907330 -0.013675934  4.760165e-02  6.421198e-03
#>  [56,] -0.44782194 -0.25833755  0.004803130 -3.663773e-02  1.538639e-02
#>  [57,] -0.24254512 -1.64482776 -0.033903892 -5.068642e-03 -1.075086e-02
#>  [58,]  1.26527431 -0.10173264 -0.043728934 -1.344551e-04  4.018917e-02
#>  [59,] -0.64023258 -0.23615696  0.018975607 -3.737504e-02 -3.806573e-02
#>  [60,] -1.45056294  0.60273090 -0.005420260  4.825890e-02 -1.042489e-02
#>  [61,] -0.55712935  0.96667031 -0.006456056 -4.928158e-02  2.619779e-03
#>  [62,]  0.15444626  0.73556977 -0.010783111  3.645439e-02  2.823097e-02
#>  [63,] -0.20033086  0.91883128 -0.043648472  2.195884e-02  1.279108e-02
#>  [64,] -1.24901593  0.28152229  0.047127890 -4.027703e-02 -6.164526e-03
#>  [65,]  0.29525230  0.96419456 -0.009180066  2.904753e-02 -2.313746e-02
#>  [66,] -0.97389676 -0.57686103  0.047663086 -3.887247e-02 -9.890199e-03
#>  [67,] -0.73245512 -1.10601370  0.021979346  2.919794e-02 -4.262715e-02
#>  [68,] -1.49079601 -0.78588903  0.026487218 -1.560338e-02 -4.380382e-02
#>  [69,] -0.15174096  0.53073988  0.004878641  6.297586e-03 -1.558964e-02
#>  [70,] -0.32492588  1.21812044  0.009508482  4.934957e-02 -3.543222e-02
#>  [71,]  1.15552046 -1.52982217 -0.020165519 -1.477702e-03 -2.693341e-02
#>  [72,] -0.15654042  0.55111366  0.015274463  4.673667e-04 -1.279266e-02
#>  [73,]  1.06332611 -0.51645765 -0.034391617  2.315788e-02 -2.502222e-02
#>  [74,]  0.23548886 -0.39822071 -0.002857349 -3.754765e-02  3.686172e-02
#>  [75,]  2.09259509  0.32480773 -0.032164985 -5.373857e-05 -3.681864e-03
#>  [76,] -1.69274866  0.72685149  0.013370231  3.932745e-02  2.021575e-02
#>  [77,] -1.00535948 -0.34229429  0.003974842  2.428877e-02 -2.163084e-02
#>  [78,]  1.67447758 -0.55891969  0.033188129  4.757632e-02 -3.115888e-02
#>  [79,] -0.43774473 -0.39758655  0.022563261  4.407943e-02  1.689626e-02
#>  [80,]  1.32715796 -0.08922549  0.015549457 -3.487123e-02 -3.835214e-02
#>  [81,]  1.42754421  1.04095216  0.034183988 -4.939691e-02 -4.360948e-02
#>  [82,] -1.53957946  0.24091850 -0.023698009  2.455375e-02  1.276499e-02
#>  [83,]  1.18267687 -0.34464959  0.013518371 -1.954660e-02  3.389586e-02
#>  [84,]  0.65495938 -0.89775118 -0.017118932  2.505887e-02  1.952225e-02
#>  [85,]  1.31929669 -0.10166170  0.012280953  3.074620e-02 -4.819031e-03
#>  [86,] -0.40120177 -0.78606287  0.004544277  3.862457e-02 -2.757129e-02
#>  [87,] -0.57401016 -0.90709029 -0.046717028 -3.967521e-02 -2.074433e-02
#>  [88,] -1.80330418  1.40847600  0.045630728 -3.166735e-02 -2.423152e-02
#>  [89,]  0.04415093  0.25038848  0.046616045 -3.609492e-02 -2.962644e-02
#>  [90,]  0.02434522 -0.59770939 -0.038278397  1.737880e-02  4.330101e-03
#>  [91,]  1.14911187 -0.52593488 -0.003054495 -5.543192e-03  2.244072e-02
#>  [92,] -0.27331021 -0.50899392 -0.045765396  2.662101e-02  3.555318e-02
#>  [93,] -0.56484152 -0.74419359 -0.002204493 -1.563256e-02  9.417013e-03
#>  [94,]  0.70407516 -0.88602139 -0.040906914  3.858482e-02 -7.889513e-06
#>  [95,]  0.52037194  1.81608521 -0.013774332  3.147875e-02 -1.821598e-02
#>  [96,]  0.74337154 -1.66914619 -0.018079979  4.151892e-02 -8.202764e-03
#>  [97,]  0.17305276  1.12261694  0.001361593  2.871351e-02 -1.334175e-02
#>  [98,]  0.91004402  0.71735709 -0.041308750 -2.339730e-02  1.272601e-02
#>  [99,] -0.64166965 -0.34072011 -0.019086133  3.832444e-02  1.195107e-02
#>                [,11]
#>   [1,] -0.0407013789
#>   [2,]  0.0268525331
#>   [3,]  0.0177928318
#>   [4,]  0.0123152165
#>   [5,] -0.0360843137
#>   [6,]  0.0370757646
#>   [7,] -0.0083398750
#>   [8,]  0.0148626652
#>   [9,]  0.0385009889
#>  [10,]  0.0143807329
#>  [11,]  0.0118162926
#>  [12,]  0.0311816217
#>  [13,]  0.0335351573
#>  [14,]  0.0321517655
#>  [15,] -0.0245868751
#>  [16,]  0.0462344634
#>  [17,]  0.0249259567
#>  [18,] -0.0258999829
#>  [19,]  0.0313133982
#>  [20,] -0.0491943880
#>  [21,]  0.0430724923
#>  [22,] -0.0012664595
#>  [23,]  0.0127195019
#>  [24,] -0.0116280132
#>  [25,]  0.0082862699
#>  [26,]  0.0225958756
#>  [27,] -0.0181954966
#>  [28,]  0.0377049603
#>  [29,] -0.0078049220
#>  [30,]  0.0022543276
#>  [31,] -0.0216374389
#>  [32,] -0.0170077915
#>  [33,] -0.0466033931
#>  [34,] -0.0065372409
#>  [35,]  0.0139334952
#>  [36,] -0.0340669622
#>  [37,] -0.0299035606
#>  [38,] -0.0490657143
#>  [39,] -0.0239209336
#>  [40,] -0.0117288258
#>  [41,] -0.0221880441
#>  [42,] -0.0484066116
#>  [43,] -0.0073983550
#>  [44,] -0.0114803229
#>  [45,] -0.0194652983
#>  [46,] -0.0254894859
#>  [47,] -0.0351374996
#>  [48,] -0.0192333885
#>  [49,] -0.0275533803
#>  [50,]  0.0227705990
#>  [51,]  0.0360165568
#>  [52,] -0.0103490839
#>  [53,]  0.0391795520
#>  [54,] -0.0121489386
#>  [55,] -0.0219939725
#>  [56,]  0.0476250201
#>  [57,]  0.0441380933
#>  [58,] -0.0208328515
#>  [59,] -0.0423630060
#>  [60,] -0.0266238635
#>  [61,]  0.0386360822
#>  [62,]  0.0042336951
#>  [63,]  0.0087551628
#>  [64,]  0.0427828757
#>  [65,]  0.0238749421
#>  [66,] -0.0254783164
#>  [67,] -0.0190301529
#>  [68,]  0.0421578234
#>  [69,] -0.0329973655
#>  [70,]  0.0001952656
#>  [71,]  0.0323641187
#>  [72,]  0.0042779815
#>  [73,]  0.0232587267
#>  [74,] -0.0064407191
#>  [75,]  0.0220592903
#>  [76,] -0.0358415196
#>  [77,]  0.0455403575
#>  [78,]  0.0001227995
#>  [79,]  0.0173907470
#>  [80,]  0.0351899527
#>  [81,] -0.0208063371
#>  [82,] -0.0042233906
#>  [83,] -0.0408500287
#>  [84,]  0.0263331969
#>  [85,] -0.0400008208
#>  [86,] -0.0397274859
#>  [87,] -0.0070783759
#>  [88,] -0.0498675498
#>  [89,] -0.0282140205
#>  [90,]  0.0309396625
#>  [91,] -0.0134882119
#>  [92,]  0.0256037896
#>  [93,]  0.0071955899
#>  [94,]  0.0400404189
#>  [95,] -0.0366707267
#>  [96,]  0.0341137409
#>  [97,] -0.0413224573
#>  [98,] -0.0407644194
#>  [99,] -0.0337856365