Generate a torus-shaped dataset with optional noise.

This function generates a torus-shaped dataset along with optional noise.

Usage

torus_with_noise(sample_size, num_noise_dims, min_noise, max_noise)

Arguments

sample_size

Total number of data points to generate.

num_noise_dims

Number of additional noise dimensions to add to the data.

min_noise

Minimum value for the noise added to the data.

max_noise

Maximum value for the noise added to the data.

Value

A matrix containing the generated torus-shaped data points with or without added noise.

Examples

torus_with_noise(sample_size = 100, num_noise_dims = 4,
min_noise = -0.05, max_noise = 0.05)
#>               [,1]        [,2]         [,3]         [,4]          [,5]
#>   [1,]  1.75596439  2.32702846 -0.402968256 -0.035220649 -0.0413890014
#>   [2,] -0.31627536 -0.99117170 -0.281399410 -0.047708134 -0.0070030712
#>   [3,] -2.41033011 -1.51358369  0.532928923 -0.026895262  0.0361978746
#>   [4,] -1.64041720  1.06934926  0.999125241 -0.008843861  0.0032281447
#>   [5,] -1.39108948  0.74806168 -0.907278741  0.039767782 -0.0369721360
#>   [6,]  0.64159533 -1.60282597 -0.961863224 -0.018150367  0.0411472063
#>   [7,]  2.53925523  1.50557020  0.305963812 -0.042268784  0.0399192747
#>   [8,]  0.20512912  1.00917460  0.242350767 -0.049610272  0.0063302057
#>   [9,]  1.15256725 -0.48146815  0.660403199  0.019296562  0.0318057873
#>  [10,]  0.53827074  1.06314845 -0.588696607  0.028215118 -0.0249199912
#>  [11,]  2.76356813 -0.84556923  0.455894287  0.008979859 -0.0180578328
#>  [12,]  1.91116384 -0.30047659 -0.997861774  0.025402755 -0.0078388975
#>  [13,] -0.15473809  2.87152435  0.482872765 -0.031466889  0.0135346089
#>  [14,]  2.41431680  1.27023909  0.685489741 -0.026466946  0.0009568886
#>  [15,]  0.04950259  1.14970751  0.528027496  0.011576106 -0.0142152189
#>  [16,] -0.46466245  1.12398383 -0.621069578 -0.015868054 -0.0386087419
#>  [17,]  1.24139525  0.41288269 -0.722143004 -0.036141318  0.0383671022
#>  [18,]  1.94372104 -0.99364881 -0.983117157 -0.028705032  0.0411669702
#>  [19,] -0.70151756 -1.48292784  0.933140577  0.030453105  0.0201116817
#>  [20,]  0.75748078 -0.65289175 -0.006690021 -0.041348666  0.0375252078
#>  [21,] -0.61968814 -0.94759875  0.496975283 -0.031346469  0.0468601100
#>  [22,]  2.65242640 -1.15705113  0.448446964 -0.008711458 -0.0427922674
#>  [23,] -0.71247310  1.55030201  0.955861134 -0.035708845 -0.0079684225
#>  [24,]  0.98817753 -0.28199108  0.233425546 -0.040389311  0.0422511406
#>  [25,] -1.32183502  0.82030963  0.895870609 -0.047890890 -0.0408170913
#>  [26,] -0.12587438  1.10005186 -0.450512899  0.039022108  0.0485213517
#>  [27,] -2.65131035 -0.62428514  0.689992093  0.043298258  0.0213965312
#>  [28,]  2.50532575  1.61335606 -0.199678081 -0.027863876 -0.0168641825
#>  [29,] -1.10925771  0.58935609  0.668293636 -0.001364039  0.0067950562
#>  [30,] -2.35215797 -1.40394333  0.673388325 -0.035106546 -0.0039661721
#>  [31,] -0.15673893 -1.03654590 -0.307120546  0.017399470  0.0347263255
#>  [32,]  0.12523397 -2.53120030 -0.845297169 -0.030794455 -0.0266020129
#>  [33,] -1.04689149 -1.98916233  0.968803032  0.023072892  0.0443006858
#>  [34,] -1.68226587 -0.09412788 -0.949057540 -0.032548941 -0.0351835602
#>  [35,]  1.66620496 -1.78313532 -0.897775701  0.039135945 -0.0254672030
#>  [36,] -2.50032264 -1.19040228  0.638964142  0.001362512  0.0372412618
#>  [37,] -0.94018279 -1.31860094 -0.924764642 -0.031502645 -0.0263276917
#>  [38,] -0.89050497  0.55834676  0.315489641 -0.036047550 -0.0260977332
#>  [39,]  1.22814268 -1.90193379 -0.964523250  0.028044433  0.0392256111
#>  [40,] -1.14992630  1.02538431 -0.888279029 -0.049002150 -0.0325477761
#>  [41,]  0.41139831 -1.06243324  0.509118763 -0.047470851 -0.0142261531
#>  [42,]  1.42667161 -0.40761761  0.856444015  0.044633782  0.0359497275
#>  [43,]  1.15652034 -1.91601826  0.971263989  0.020616366 -0.0373747984
#>  [44,] -0.65521908 -1.14665820  0.733821639 -0.004581891  0.0124696270
#>  [45,] -0.77895984 -2.89223362 -0.096888248  0.018971746 -0.0104161012
#>  [46,]  2.65888296 -1.36597940  0.146298960  0.009015472 -0.0139370589
#>  [47,]  1.99032131  0.24513349 -0.999985634 -0.007582631  0.0441316154
#>  [48,] -1.95409534 -1.10985042  0.968944681 -0.039681346 -0.0490629904
#>  [49,] -2.15394557  1.90357012  0.484928131  0.016611275 -0.0137041819
#>  [50,]  0.98795167  0.33454289 -0.290275177 -0.037618171 -0.0195972586
#>  [51,]  1.20697100 -0.75467828 -0.817088737 -0.047370342 -0.0225775151
#>  [52,] -1.24904699  0.37569825  0.718358096 -0.028656463 -0.0369493379
#>  [53,]  2.27298058 -1.67478700  0.567521873  0.040269858 -0.0138401185
#>  [54,] -0.87956641  1.66383197 -0.993015170 -0.037872026  0.0464528964
#>  [55,] -1.07855023  0.33037351  0.489532136 -0.008815536 -0.0075613866
#>  [56,] -2.14637609 -0.37649328  0.983822491  0.029240300 -0.0309259859
#>  [57,] -0.03242686 -1.00119325 -0.058596312  0.044173142  0.0380746909
#>  [58,] -0.46606933 -0.95592183  0.350636071  0.034253678  0.0167759117
#>  [59,] -1.54603909  2.34124511  0.592393372  0.023766878  0.0237321745
#>  [60,] -2.59127117  1.12763047 -0.563680297 -0.019719704  0.0303163804
#>  [61,]  1.09954579  0.17134489 -0.461416292  0.002931907  0.0350885266
#>  [62,]  1.82414685 -1.04069735 -0.994973964 -0.045572575  0.0234708515
#>  [63,]  0.25127350  1.38480829  0.805511945 -0.000767401 -0.0487990947
#>  [64,]  0.87775609  1.43099533  0.946994606 -0.010241389 -0.0413011562
#>  [65,] -1.98276091 -0.59553626 -0.997528240 -0.030745791 -0.0288850958
#>  [66,]  1.46066673 -0.50122944  0.890119459 -0.010138801  0.0025170439
#>  [67,] -1.71473604  2.17701373 -0.636558168 -0.045365448  0.0325163940
#>  [68,] -0.81144954  0.86830259  0.584275000 -0.015533447 -0.0332528251
#>  [69,] -1.94746260 -0.39750781  0.999923333 -0.023409305 -0.0104095605
#>  [70,] -1.40576128 -2.32468953 -0.697402928  0.048599400 -0.0177573336
#>  [71,] -2.19787665  0.16840131  0.978904431 -0.011586869 -0.0039292349
#>  [72,] -0.20547603  2.98084522 -0.154972529  0.023697786 -0.0280672709
#>  [73,]  2.02368182  2.01855960  0.513150555 -0.032803816 -0.0363123827
#>  [74,]  2.82289072 -1.01427585  0.029052423 -0.015033471 -0.0481486798
#>  [75,] -0.39600978 -2.31992382  0.935441908  0.018560811  0.0071866081
#>  [76,] -0.65249959 -0.80967586  0.279555707  0.040005419 -0.0273290583
#>  [77,]  2.63532689  0.51115971 -0.729066749 -0.036950103  0.0200688085
#>  [78,] -1.29246119  0.64592527  0.831769616 -0.028027540 -0.0428461904
#>  [79,] -0.99281688 -2.22706429  0.898809363  0.009074518 -0.0324983501
#>  [80,] -1.23517893 -0.10034570 -0.649042981  0.028130975 -0.0349277206
#>  [81,]  1.16062872 -1.06606816  0.905630155  0.048502879 -0.0111150932
#>  [82,]  0.20624924  1.36731683  0.786794457 -0.017939755  0.0257509095
#>  [83,]  2.54730128  0.53630509  0.797631114  0.012253383  0.0162511526
#>  [84,]  1.57284282 -1.91612789 -0.877821544 -0.019601148 -0.0394674942
#>  [85,] -0.40020379  2.21835679  0.967160285 -0.020980399 -0.0049157295
#>  [86,] -1.95978005  0.24037878 -0.999673979  0.031964127  0.0253965042
#>  [87,] -2.22088151  0.23816421 -0.972329131 -0.032292374  0.0099579119
#>  [88,] -0.57884100 -1.51800447 -0.926871541 -0.005871122 -0.0301681072
#>  [89,]  1.74244403  1.42420607  0.968132470 -0.026190765  0.0495881533
#>  [90,] -1.71680881  1.46251863  0.966861000  0.047064107  0.0215314501
#>  [91,]  1.28802319  0.81580621  0.879794282  0.029292961  0.0055103527
#>  [92,]  0.95990980  0.48310190  0.379047653  0.019510498 -0.0014198164
#>  [93,] -1.69011896  2.11675197 -0.705494146 -0.006819493  0.0353692735
#>  [94,]  0.90247955  0.59268316 -0.391203117 -0.022548103 -0.0244210305
#>  [95,] -0.71496556  1.73740808 -0.992623985 -0.048144334 -0.0284167386
#>  [96,]  2.63838589  0.24250565 -0.760355329 -0.039582591 -0.0414769596
#>  [97,] -1.54547503 -2.57092321  0.024903670 -0.030479953  0.0192226702
#>  [98,] -0.17749615 -1.88895383  0.994709767  0.045122221  0.0147487297
#>  [99,]  1.85118786 -1.36141233 -0.954597381  0.024581253 -0.0179904100
#> [100,] -2.25652294 -1.81576197 -0.443331770 -0.049885435 -0.0423380179
#>                 [,6]          [,7]
#>   [1,]  0.0095832565  0.0002922800
#>   [2,] -0.0266406630 -0.0024799030
#>   [3,]  0.0329666909  0.0289243373
#>   [4,] -0.0147078566  0.0174967367
#>   [5,] -0.0069697323 -0.0458306788
#>   [6,]  0.0314624313  0.0104631993
#>   [7,]  0.0081860261  0.0451189126
#>   [8,] -0.0194040084  0.0094473721
#>   [9,] -0.0428688820  0.0370479102
#>  [10,]  0.0026849007  0.0486879260
#>  [11,]  0.0022872865 -0.0136065118
#>  [12,]  0.0013940846  0.0224043879
#>  [13,]  0.0313281161  0.0310346203
#>  [14,] -0.0210632421  0.0252130944
#>  [15,] -0.0144269704 -0.0139956920
#>  [16,] -0.0093651362 -0.0069703208
#>  [17,] -0.0108168368  0.0094861534
#>  [18,]  0.0172680418  0.0339568058
#>  [19,]  0.0344833083 -0.0358987421
#>  [20,] -0.0454159579 -0.0327478505
#>  [21,] -0.0093561981 -0.0189579947
#>  [22,]  0.0396547277  0.0264252754
#>  [23,] -0.0451344814  0.0472637191
#>  [24,]  0.0487099020  0.0323779521
#>  [25,] -0.0240822078  0.0110516789
#>  [26,] -0.0376884447  0.0137200678
#>  [27,]  0.0200525602 -0.0224207042
#>  [28,]  0.0272840366 -0.0141355241
#>  [29,]  0.0025653793 -0.0083042403
#>  [30,]  0.0008814695  0.0469945305
#>  [31,] -0.0285198053  0.0024664085
#>  [32,] -0.0056778962  0.0166698525
#>  [33,] -0.0049028260 -0.0411085621
#>  [34,]  0.0187988844 -0.0271453298
#>  [35,] -0.0379018271 -0.0140130572
#>  [36,]  0.0417903086 -0.0256205071
#>  [37,] -0.0424125727 -0.0236385710
#>  [38,] -0.0438426733 -0.0094260246
#>  [39,] -0.0313470324  0.0174269036
#>  [40,]  0.0031317025 -0.0452166670
#>  [41,] -0.0291986905 -0.0085330202
#>  [42,]  0.0126839608  0.0425648325
#>  [43,]  0.0089078305 -0.0314434839
#>  [44,]  0.0014842902 -0.0383992604
#>  [45,] -0.0296623494  0.0015955400
#>  [46,]  0.0209008188 -0.0485900112
#>  [47,]  0.0223604991  0.0099147135
#>  [48,] -0.0494228308 -0.0493410359
#>  [49,]  0.0320848967 -0.0044842294
#>  [50,]  0.0482571659  0.0139203837
#>  [51,] -0.0448505170 -0.0090072411
#>  [52,]  0.0219973698  0.0106927139
#>  [53,]  0.0109987007  0.0305214964
#>  [54,]  0.0159363716 -0.0419314152
#>  [55,] -0.0441858407  0.0446353598
#>  [56,] -0.0258936247  0.0176150748
#>  [57,] -0.0262809399  0.0108909499
#>  [58,]  0.0277324335 -0.0488770320
#>  [59,]  0.0251754210 -0.0066879806
#>  [60,]  0.0188167986  0.0008472490
#>  [61,]  0.0420880673  0.0113221152
#>  [62,]  0.0073119656  0.0014994823
#>  [63,]  0.0212186538 -0.0347111745
#>  [64,]  0.0315057692  0.0368487634
#>  [65,]  0.0374848023  0.0196946633
#>  [66,] -0.0315754419 -0.0116633489
#>  [67,] -0.0338012822 -0.0031975573
#>  [68,] -0.0026918047 -0.0103791422
#>  [69,]  0.0048027311 -0.0264657836
#>  [70,] -0.0295907385 -0.0029352558
#>  [71,]  0.0234476721  0.0497341457
#>  [72,] -0.0259494630 -0.0469626904
#>  [73,] -0.0296446106  0.0038461851
#>  [74,]  0.0069129028 -0.0425382219
#>  [75,]  0.0092031661 -0.0064976265
#>  [76,]  0.0266492500  0.0474321380
#>  [77,]  0.0224620257  0.0323446197
#>  [78,]  0.0424475306 -0.0385212377
#>  [79,]  0.0054160010 -0.0173011404
#>  [80,] -0.0426761724  0.0358089990
#>  [81,] -0.0109921372  0.0463989817
#>  [82,]  0.0081394651  0.0463904848
#>  [83,]  0.0381643004  0.0021304497
#>  [84,]  0.0229201293 -0.0387863542
#>  [85,]  0.0070497429  0.0001248337
#>  [86,]  0.0414602482  0.0132561871
#>  [87,]  0.0467431856 -0.0038118397
#>  [88,] -0.0340264374  0.0458660896
#>  [89,] -0.0425061390  0.0054312696
#>  [90,]  0.0346774267 -0.0471662242
#>  [91,] -0.0184196660  0.0073096257
#>  [92,] -0.0215751942 -0.0324232042
#>  [93,]  0.0398827155  0.0474069657
#>  [94,] -0.0280261239 -0.0341347246
#>  [95,] -0.0041962688  0.0013918406
#>  [96,] -0.0000148938  0.0140629166
#>  [97,]  0.0475174043  0.0056495602
#>  [98,] -0.0198791114 -0.0390863102
#>  [99,] -0.0266712060  0.0355017966
#> [100,] -0.0211487415  0.0123463956